Basic use

Basic use

Compute a lazy Kronecker products between two matrices A and B by either

K = kronecker(A, B)

or, by using the binary operator:

K = A ⊗ B

Note, ⊗ can be formed by typing \otimes<tab>.

The Kronecker product K behaves like a matrix, for which size(K), eltype(K) works as one would expect. Elements can be accessed via K[i,j]; every element is computed on the fly. The function collect can be used to turn K in a regular, dense matrix.

julia> using Kronecker

julia> A = randn(4, 4)
4×4 Array{Float64,2}:
 -0.093381   0.319488  -0.281411   1.01075
 -0.751063  -0.468556   0.578736  -0.395136
 -0.834637  -1.20992    1.3865     1.09267
 -0.214442  -0.652075   0.151511   0.528215

julia> B = rand(1:10, 5, 7)
5×7 Array{Int64,2}:
  5   5  2  7   5  8   8
  6   3  3  8   7  8  10
  5   9  6  7   6  9   9
  4  10  3  5   9  1   3
 10   3  3  7  10  8   1

julia> K = A ⊗ B
20×28 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Int64,2}}:
 -0.466905  -0.466905  -0.186762  …   5.05375   8.08599    8.08599
 -0.560286  -0.280143  -0.280143      7.07524   8.08599   10.1075
 -0.466905  -0.840429  -0.560286      6.06449   9.09674    9.09674
 -0.373524  -0.93381   -0.280143      9.09674   1.01075    3.03225
 -0.93381   -0.280143  -0.280143     10.1075    8.08599    1.01075
 -3.75531   -3.75531   -1.50213   …  -1.97568  -3.16109   -3.16109
 -4.50638   -2.25319   -2.25319      -2.76595  -3.16109   -3.95136
 -3.75531   -6.75957   -4.50638      -2.37082  -3.55623   -3.55623
 -3.00425   -7.51063   -2.25319      -3.55623  -0.395136  -1.18541
 -7.51063   -2.25319   -2.25319      -3.95136  -3.16109   -0.395136
 -4.17319   -4.17319   -1.66927   …   5.46334   8.74134    8.74134
 -5.00782   -2.50391   -2.50391       7.64868   8.74134   10.9267
 -4.17319   -7.51174   -5.00782       6.55601   9.83401    9.83401
 -3.33855   -8.34637   -2.50391       9.83401   1.09267    3.278
 -8.34637   -2.50391   -2.50391      10.9267    8.74134    1.09267
 -1.07221   -1.07221   -0.428885  …   2.64107   4.22572    4.22572
 -1.28665   -0.643327  -0.643327      3.6975    4.22572    5.28215
 -1.07221   -1.92998   -1.28665       3.16929   4.75393    4.75393
 -0.85777   -2.14442   -0.643327      4.75393   0.528215   1.58464
 -2.14442   -0.643327  -0.643327      5.28215   4.22572    0.528215

julia> K[4, 5]
-0.8404286086394058

julia> eltype(K)  # promotion
Float64

julia> collect(K)
20×28 Array{Float64,2}:
 -0.466905  -0.466905  -0.186762  …   5.05375   8.08599    8.08599
 -0.560286  -0.280143  -0.280143      7.07524   8.08599   10.1075
 -0.466905  -0.840429  -0.560286      6.06449   9.09674    9.09674
 -0.373524  -0.93381   -0.280143      9.09674   1.01075    3.03225
 -0.93381   -0.280143  -0.280143     10.1075    8.08599    1.01075
 -3.75531   -3.75531   -1.50213   …  -1.97568  -3.16109   -3.16109
 -4.50638   -2.25319   -2.25319      -2.76595  -3.16109   -3.95136
 -3.75531   -6.75957   -4.50638      -2.37082  -3.55623   -3.55623
 -3.00425   -7.51063   -2.25319      -3.55623  -0.395136  -1.18541
 -7.51063   -2.25319   -2.25319      -3.95136  -3.16109   -0.395136
 -4.17319   -4.17319   -1.66927   …   5.46334   8.74134    8.74134
 -5.00782   -2.50391   -2.50391       7.64868   8.74134   10.9267
 -4.17319   -7.51174   -5.00782       6.55601   9.83401    9.83401
 -3.33855   -8.34637   -2.50391       9.83401   1.09267    3.278
 -8.34637   -2.50391   -2.50391      10.9267    8.74134    1.09267
 -1.07221   -1.07221   -0.428885  …   2.64107   4.22572    4.22572
 -1.28665   -0.643327  -0.643327      3.6975    4.22572    5.28215
 -1.07221   -1.92998   -1.28665       3.16929   4.75393    4.75393
 -0.85777   -2.14442   -0.643327      4.75393   0.528215   1.58464
 -2.14442   -0.643327  -0.643327      5.28215   4.22572    0.528215

Constructing Kronecker products

Kronecker.kronecker — Function.

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Construct a Kronecker product object between two arrays. Does not evaluate the Kronecker product explictly.

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Higher-order Kronecker lazy kronecker product, e.g.

kronecker(A, B, C, D)

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Function.

⊗(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kronecker, computes as Lazy Kronecker product. See kronecker for documentation.

Base.collect — Method.

collect(K::AbstractKroneckerProduct)

Collects a lazy instance of the AbstractKroneckerProduct type into a full, native matrix. Equivalent with Matrix(K::AbstractKroneckerProduct).

Basic properties of Kronecker products

Missing docstring.

Missing docstring for getindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eltype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for size(::AbstractKroneckerProduct). Check Documenter's build log for details.

Kronecker.order — Function.

order(M::AbstractMatrix)

Returns the order of a matrix, i.e. how many matrices are involved in the Kronecker product (default to 1 for general matrices).

Kronecker.getmatrices — Function.

getmatrices(K::AbstractKroneckerProduct)

Obtain the two matrices of an AbstractKroneckerProduct object.

getmatrices(A::AbstractArray)

Returns a matrix itself. Needed for recursion.

getmatrices(K::T) where T <: AbstractKroneckerSum

Obtain the two matrices of an AbstractKroneckerSum object.

Kronecker.issquare — Function.

issquare(A::AbstractMatrix)

Checks if an array is a square matrix.

issquare(K::AbstractKroneckerProduct)

Checks if all matrices of a Kronecker product are square.

LinearAlgebra.issymmetric — Function.

issymmetric(K::AbstractKroneckerProduct)

Checks if a Kronecker product is symmetric.

Missing docstring.

Missing docstring for sum. Check Documenter's build log for details.

Linear algebra

Many functions of the LinearAlgebra module are overloaded to work with subtypes of GeneralizedKroneckerProduct.

LinearAlgebra.det — Method.

tr(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method.

logdet(K::AbstractKroneckerProduct)

Compute the logarithm of the determinant of a Kronecker product.

LinearAlgebra.tr — Method.

tr(K::AbstractKroneckerProduct)

Compute the trace of a Kronecker product.

Base.inv — Method.

inv(K::AbstractKroneckerProduct)

Compute the inverse of a Kronecker product.

Base.adjoint — Method.

adjoint(K::AbstractKroneckerProduct)

Compute the adjoint of a Kronecker product.

Base.transpose — Method.

transpose(K::AbstractKroneckerProduct)

Compute the transpose of a Kronecker product.

Missing docstring.

Missing docstring for conj(K::AbstractKroneckerProduct). Check Documenter's build log for details.