Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 25.2333      2.67388    5.10711  …    1.60949    3.07413  -0.610808
  2.67388    33.8388   -21.7896       20.3686   -13.1159    4.59691
  5.10711   -21.7896    20.6148      -13.1159    12.4087   -1.51338
 -1.01475     7.63693   -2.5142        4.59691   -1.51338   2.72175
 19.6357      2.08072    3.97417       1.96928    3.76132  -0.747347
  2.08072    26.3321   -16.9559   …   24.9218   -16.0478    5.6245
  3.97417   -16.9559    16.0417      -16.0478    15.1825   -1.85168
 -0.789639    5.94278   -1.95646       5.6245    -1.85168   3.33016
 15.3952      1.63137    3.11592       1.70251    3.2518   -0.646109
  1.63137    20.6455   -13.2942       21.5458   -13.8739    4.86259
  ⋮                               ⋱
 -0.520923    3.92044   -1.29067       5.63311   -1.85451   3.33526
 15.2453      1.61549    3.08558       1.17453    2.24336  -0.445739
  1.61549    20.4445   -13.1647       14.8641    -9.57134   3.35461
  3.08558   -13.1647    12.4549       -9.57134    9.05528  -1.10439
 -0.613083    4.61403   -1.51902  …    3.35461   -1.10439   1.9862
 15.1887      1.60949    3.07413       1.99668    3.81366  -0.757747
  1.60949    20.3686   -13.1159       25.2686   -16.2711    5.70276
  3.07413   -13.1159    12.4087      -16.2711    15.3938   -1.87744
 -0.610808    4.59691   -1.51338       5.70276   -1.87744   3.3765

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0001428181605909896
   0.001659588536374598
   0.0076739141439206215
   0.014762655488355855
   0.0011384665831055732
   0.013229312592664556
   0.06117215610621738
   0.11767964159351851
   0.01439375039067581
   0.16725956310426
   ⋮
  25.425098137776633
   0.2946212775161121
   3.423584876842522
  15.830608511376495
  30.454057113686485
   3.6744713294636466
  42.69842483883471
 197.436918315701
 379.8183236238589
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0654743  -0.0577742   0.341148     …  -0.298858      0.0121294
 -0.176047    0.104352    0.0628217       -0.0550342    -0.245243
 -0.177331    0.21565     0.0621998       -0.0544894     0.180341
  0.239734    0.251925   -0.00102973       0.000902083  -0.0500078
 -0.103348    0.0911942  -0.538488        -0.324769      0.0131811
  0.277884   -0.164715   -0.0991615    …  -0.0598057    -0.266506
  0.279909   -0.340395   -0.0981798       -0.0592136     0.195976
 -0.378411   -0.397653    0.00162539       0.000980294  -0.0543435
 -0.0218596   0.0192888  -0.113897        -0.30549       0.0123986
  0.0587761  -0.0348394  -0.020974        -0.0562553    -0.250685
  ⋮                                    ⋱
  0.166526    0.174994   -0.000715279      0.000938119  -0.0520055
 -0.0322245   0.0284347  -0.167903        -0.248928      0.010103
  0.0866453  -0.0513588  -0.030919        -0.0458397    -0.20427
  0.0872769  -0.106137   -0.0306129       -0.0453859     0.150211
 -0.11799    -0.12399     0.000506803  …   0.000751373  -0.041653
  0.0584529  -0.0515786   0.304564        -0.275476      0.0111805
 -0.157168    0.0931612   0.0560848       -0.0507284    -0.226056
 -0.158314    0.192524    0.0555296       -0.0502262     0.166231
  0.214025    0.224909   -0.000919304      0.000831505  -0.0460953

julia> logdet(E)
-9.077228403717367

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  2.9410301180361205
  6.7019724948057045
 -0.19375084663889652
 -3.1764320420622556
 -4.4236406333284695
 -5.535985044638984
  1.9163501269785914
 -4.250227884449619
 -0.8845621268806949
 -0.324419924334683
  ⋮
 -8.451804810665378
  0.5972150628673122
 -6.128994369201065
 -2.6480462904108584
  2.4180594379011784
  3.9581980189377353
 -2.460821079552345
 -3.7674293541964015
 10.719692131080308

Missing docstring.

Missing docstring for eigen. Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for det(E::Eigen{<:Number, <:Number, <:AbstractKroneckerProduct}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for logdet(E::Eigen{<:Number, <:Number, <:AbstractKroneckerProduct}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for inv(E::Eigen{<:Number, <:Number, <:AbstractKroneckerProduct}). Check Documenter's build log for details.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 12.5417     3.77659    -5.21996  …    1.76789    -2.44356   3.78387
  3.77659   13.6673    -23.2451        6.3979    -10.8815   -0.188915
 -5.21996  -23.2451     53.7952      -10.8815     25.1825    6.27178
  8.08315   -0.403562   13.3978       -0.188915    6.27178   7.59577
  8.9659     2.69983    -3.73168       2.29626    -3.17386   4.91475
  2.69983    9.77054   -16.6176   …    8.31003   -14.1336   -0.245376
 -3.73168  -16.6176     38.4575      -14.1336     32.7088    8.14621
  5.77853   -0.288501    9.57793      -0.245376    8.14621   9.86591
 11.3032     3.40363    -4.70447       2.39221    -3.30649   5.12012
  3.40363   12.3176    -20.9495        8.65727   -14.7242   -0.255629
  ⋮                               ⋱
  3.80852   -0.190145    6.31262      -0.153951    5.11101   6.18996
  7.17839    2.16157    -2.9877        1.18664    -1.64016   2.5398
  2.16157    7.82262   -13.3046        4.29438    -7.30382  -0.126803
 -2.9877   -13.3046     30.7903       -7.30382    16.9029    4.20972
  4.62648   -0.230983    7.6684   …   -0.126803    4.20972   5.09841
  5.87101    1.76789    -2.44356       2.14081    -2.95901   4.58205
  1.76789    6.3979    -10.8815        7.74749   -13.1768   -0.228765
 -2.44356  -10.8815     25.1825      -13.1768     30.4946    7.59476
  3.78387   -0.188915    6.27178      -0.228765    7.59476   9.19804

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.54143  1.0664   -1.47397   2.28246   …   0.499203   -0.689993    1.06846
 0.0      3.53978  -6.12276  -0.801625      1.65704    -2.86618    -0.375256
 0.0      0.0       3.75958   3.153         0.0         1.75993     1.47598
 0.0      0.0       0.0       0.657699      0.0         0.0         0.307881
 0.0      0.0      -0.0       0.0           0.467455   -0.646112    1.00051
 0.0      0.0      -0.0      -0.0       …   1.55166    -2.6839     -0.35139
 0.0      0.0       0.0       0.0           0.0         1.648       1.38211
 0.0      0.0       0.0       0.0           0.0         0.0         0.288301
 0.0      0.0      -0.0       0.0           0.0842797  -0.116491    0.180387
 0.0      0.0      -0.0      -0.0           0.279755   -0.483893   -0.0633539
 ⋮                                      ⋱
 0.0      0.0       0.0       0.0           0.0         0.0         0.012117
 0.0      0.0      -0.0       0.0          -0.047796    0.0660632  -0.102299
 0.0      0.0      -0.0      -0.0          -0.158653    0.274421    0.0359287
 0.0      0.0       0.0       0.0          -0.0        -0.168504   -0.141317
 0.0      0.0       0.0       0.0       …  -0.0        -0.0        -0.029478
 0.0      0.0      -0.0       0.0           0.240708   -0.332704    0.515194
 0.0      0.0      -0.0      -0.0           0.798997   -1.38203    -0.180942
 0.0      0.0       0.0       0.0           0.0         0.84861     0.711693
 0.0      0.0       0.0       0.0           0.0         0.0         0.148455

julia> logdet(C)
-31.42422656672916

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  10.9544     19.15      13.3166   …   -3.73281    -2.59574     3.11414
  19.15       41.5336    27.5991       -8.09591    -5.37974     6.10018
  13.3166     27.5991    18.7631       -5.37974    -3.65737     4.17915
 -15.9762    -31.2951   -21.4399        6.10018     4.17915    -4.98315
 -17.0903    -29.8766   -20.7757        0.406717    0.282825   -0.339309
 -29.8766    -64.798    -43.0583   …    0.88211     0.586163   -0.66466
 -20.7757    -43.0583   -29.2729        0.586163    0.398498   -0.45535
  24.9249     48.8246    33.449        -0.66466    -0.45535     0.542951
 -21.506     -37.596    -26.1436       30.3412     21.0988    -25.3125
 -37.596     -81.5401   -54.1835       65.8056     43.7279    -49.5838
   ⋮                               ⋱
 -57.4839   -112.603    -77.143        41.1019     28.1583    -33.5755
 -14.1518    -24.7397   -17.2036       14.4784     10.068     -12.0787
 -24.7397    -53.6568   -35.655        31.4014     20.8663    -23.6606
 -17.2036    -35.655    -24.2398       20.8663     14.1858    -16.2096
  20.6394     40.4298    27.6979   …  -23.6606    -16.2096     19.328
  -2.13527    -3.73281   -2.59574      33.9892     23.6355    -28.3559
  -3.73281    -8.09591   -5.37974      73.7175     48.9854    -55.5453
  -2.59574    -5.37974   -3.65737      48.9854     33.3023    -38.0533
   3.11414     6.10018    4.17915     -55.5453    -38.0533     45.3741

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, `,inv,det, andlogdet` are overloaded to efficiently work with this type.