Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.638382  0.860726
 0.122651  0.0493887

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.407531   0.549471   0.549471   0.740849
 0.0782984  0.0315288  0.105569   0.0425101
 0.0782984  0.105569   0.0315288  0.0425101
 0.0150434  0.0060576  0.0060576  0.00243924

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  0.444956   -7.75451   -7.75451   135.142
 -1.105       5.75135   19.2575   -100.232
 -1.105      19.2575     5.75135  -100.232
  2.74415   -14.2829   -14.2829     74.34

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.00458104   0.00617658   0.00617658   …  0.122629     0.16534
 0.000880149  0.000354414  0.0011867       0.0235605    0.00948725
 0.000880149  0.0011867    0.000354414     0.00703648   0.00948725
 0.000169102  6.80932e-5   6.80932e-5      0.00135191   0.000544381
 0.000880149  0.0011867    0.0011867       0.00703648   0.00948725
 0.000169102  6.80932e-5   0.000227999  …  0.00135191   0.000544381
 0.000169102  0.000227999  6.80932e-5      0.000403756  0.000544381
 3.24894e-5   1.30827e-5   1.30827e-5      7.75731e-5   3.12367e-5
 0.000880149  0.0011867    0.0011867       0.00703648   0.00948725
 0.000169102  6.80932e-5   0.000227999     0.00135191   0.000544381
 ⋮                                      ⋱               ⋮
 6.03227e-11  2.42905e-11  2.42905e-11     9.1162e-15   3.67087e-15
 1.63417e-9   2.20334e-9   2.20334e-9      8.2691e-13   1.11492e-12
 3.1397e-10   1.26428e-10  4.23324e-10     1.58873e-13  6.39743e-14
 3.1397e-10   4.23324e-10  1.26428e-10  …  4.74484e-14  6.39743e-14
 6.03227e-11  2.42905e-11  2.42905e-11     9.1162e-15   3.67087e-15
 3.1397e-10   4.23324e-10  4.23324e-10     4.74484e-14  6.39743e-14
 6.03227e-11  2.42905e-11  8.13328e-11     9.1162e-15   3.67087e-15
 6.03227e-11  8.13328e-11  2.42905e-11     2.7226e-15   3.67087e-15
 1.15897e-11  4.6669e-12   4.6669e-12   …  5.2309e-16   2.10635e-16

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.000332987  0.000384178  0.000384178  …  0.00160543   0.00185224
 5.47445e-5   7.00943e-5   6.31607e-5      0.00026394   0.000337946
 5.47445e-5   6.31607e-5   7.00943e-5      0.000292915  0.000337946
 9.00025e-6   1.15238e-5   1.15238e-5      4.81566e-5   6.16592e-5
 5.47445e-5   6.31607e-5   6.31607e-5      0.000292915  0.000337946
 9.00025e-6   1.15238e-5   1.03839e-5   …  4.81566e-5   6.16592e-5
 9.00025e-6   1.03839e-5   1.15238e-5      5.34431e-5   6.16592e-5
 1.47968e-6   1.89457e-6   1.89457e-6      8.78629e-6   1.12499e-5
 5.47445e-5   6.31607e-5   6.31607e-5      0.000292915  0.000337946
 9.00025e-6   1.15238e-5   1.03839e-5      4.81566e-5   6.16592e-5
 ⋮                                      ⋱               ⋮
 7.89733e-13  1.01117e-12  1.01117e-12     1.07896e-11  1.38149e-11
 2.92181e-11  3.371e-11    3.371e-11       3.597e-10    4.14999e-10
 4.80359e-12  6.15047e-12  5.54207e-12     5.91364e-11  7.57176e-11
 4.80359e-12  5.54207e-12  6.15047e-12  …  6.56283e-11  7.57176e-11
 7.89733e-13  1.01117e-12  1.01117e-12     1.07896e-11  1.38149e-11
 4.80359e-12  5.54207e-12  5.54207e-12     6.56283e-11  7.57176e-11
 7.89733e-13  1.01117e-12  9.11142e-13     1.07896e-11  1.38149e-11
 7.89733e-13  9.11142e-13  1.01117e-12     1.1974e-11   1.38149e-11
 1.29836e-13  1.6624e-13   1.6624e-13   …  1.96859e-12  2.52056e-12

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Missing docstring.

Missing docstring for ⊗(A::AbstractMatrix, pow::Int). Check Documenter's build log for details.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.828559  0.76686
 0.103496  0.849499

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 0.152487     0.141132    0.141132    …  0.075992    0.075992    0.0703332
 0.0190473    0.156341    0.0176289      0.0841811   0.00949223  0.0779126
 0.0190473    0.0176289   0.156341       0.00949223  0.0841811   0.0779126
 0.00237921   0.0195287   0.0195287      0.0105151   0.0105151   0.0863087
 0.0190473    0.0176289   0.0176289      0.0841811   0.0841811   0.0779126
 0.00237921   0.0195287   0.00220205  …  0.0932528   0.0105151   0.0863087
 0.00237921   0.00220205  0.0195287      0.0105151   0.0932528   0.0863087
 0.00029719   0.00243935  0.00243935     0.0116483   0.0116483   0.0956096
 0.0190473    0.0176289   0.0176289      0.0841811   0.0841811   0.0779126
 0.00237921   0.0195287   0.00220205     0.0932528   0.0105151   0.0863087
 ⋮                                    ⋱
 1.12885e-9   9.26566e-9  9.26566e-9  …  0.021525    0.021525    0.176678
 7.23496e-8   6.69621e-8  6.69621e-8     0.155559    0.155559    0.143976
 9.03726e-9   7.41781e-8  8.3643e-9      0.172323    0.0194311   0.159491
 9.03726e-9   8.3643e-9   7.41781e-8     0.0194311   0.172323    0.159491
 1.12885e-9   9.26566e-9  9.26566e-9     0.021525    0.021525    0.176678
 9.03726e-9   8.3643e-9   8.3643e-9   …  0.172323    0.172323    0.159491
 1.12885e-9   9.26566e-9  1.04479e-9     0.190893    0.021525    0.176678
 1.12885e-9   1.04479e-9  9.26566e-9     0.021525    0.190893    0.176678
 1.41006e-10  1.15738e-9  1.15738e-9     0.0238446   0.0238446   0.195717

julia> sum(P10)  # expected number of edges
11553.127812751693

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 11055 stored entries:
  [5   ,    5]  =  true
  [5   ,    6]  =  true
  [6   ,    6]  =  true
  [33  ,    6]  =  true
  [165 ,    6]  =  true
  [5   ,    7]  =  true
  [263 ,    7]  =  true
  [517 ,    7]  =  true
  [3   ,    8]  =  true
  ⋮
  [986 , 1024]  =  true
  [988 , 1024]  =  true
  [999 , 1024]  =  true
  [1001, 1024]  =  true
  [1007, 1024]  =  true
  [1012, 1024]  =  true
  [1015, 1024]  =  true
  [1019, 1024]  =  true
  [1023, 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.