Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.82535   0.723803  0.268702  0.666883  …  0.0       0.0       0.0
 0.579664  1.95661   0.100741  0.166943     1.0       0.0       0.0
 0.906225  0.262575  1.50273   0.769562     0.0       1.0       0.0
 0.236601  0.304551  0.278931  1.17076      0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.723803  0.268702  0.666883
 0.0       1.0       0.0       0.0          1.95661   0.100741  0.166943
 0.0       0.0       1.0       0.0          0.262575  1.50273   0.769562
 0.0       0.0       0.0       1.0          0.304551  0.278931  1.17076

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 23.0875    16.1171     6.12178   …  13.7193     5.21104    9.83492
 12.2305    22.6941     3.486        19.3179     2.96739    5.18402
 17.0517    11.8268    14.9459       10.0674    12.7224    11.0352
  6.28228    6.52066    3.80247       5.55058    3.23678    9.37404
 14.8342    10.3556     3.93338      10.3556     3.93338    7.42355
  7.85835   14.5815     2.23983   …  14.5815     2.23983    3.91298
 10.9561     7.599      9.60307       7.599      9.60307    8.32954
  4.0365     4.18967    2.44317       4.18967    2.44317    7.07567
  1.45003    1.01225    0.384483      1.01225    0.384483   0.725644
  0.768146   1.42532    0.218941      1.42532    0.218941   0.38249
  1.07095    0.742794   0.93869   …   0.742794   0.93869    0.814204
  0.394564   0.409536   0.238817      0.409536   0.238817   0.691639
  4.52111    3.15613    1.1988        3.15613    1.1988     2.26252
  2.39504    4.44408    0.682646      4.44408    0.682646   1.19258
  3.33916    2.31599    2.92679       2.31599    2.92679    2.53865
  1.23023    1.27691    0.74462   …   1.27691    0.74462    2.15649
 14.8342    10.3556     3.93338      12.7533     4.84411    9.1424
  7.85835   14.5815     2.23983      17.9576     2.75844    4.81899
 10.9561     7.599      9.60307       9.35847   11.8266    10.2582
  4.0365     4.18967    2.44317       5.15974    3.00886    8.71397

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 124 stored entries:
  [1 ,  1]  =  1.82535
  [2 ,  1]  =  0.579664
  [3 ,  1]  =  0.906225
  [4 ,  1]  =  0.236601
  [5 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.723803
  [2 ,  2]  =  1.95661
  [3 ,  2]  =  0.262575
  ⋮
  [18, 19]  =  0.100741
  [19, 19]  =  1.50273
  [20, 19]  =  0.278931
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [17, 20]  =  0.666883
  [18, 20]  =  0.166943
  [19, 20]  =  0.769562
  [20, 20]  =  1.17076

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes `A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A)'.

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Missing docstring.

Missing docstring for exp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.