Basic use

Compute a lazy Kronecker products between two matrices A and B by either

K = kronecker(A, B)

or, by using the binary operator:

K = A ⊗ B

Note, ⊗ can be formed by typing \otimes<tab>.

The Kronecker product K behaves like a matrix, for which size(K), eltype(K) works as one would expect. Elements can be accessed via K[i,j]; every element is computed on the fly. The function collect can be used to turn K in a regular, dense matrix.

julia> using Kronecker
julia> A = randn(4, 4)4×4 Array{Float64,2}:
  1.06293     1.41958   -1.52325    0.200124
 -0.828415    0.631868   0.168658  -0.0124556
  0.0094161  -1.44251    0.729273   0.530106
  1.21855    -0.245958  -0.589076  -0.532207
julia> B = rand(1:10, 5, 7)5×7 Array{Int64,2}:
 8  9   1  10   5  2  10
 7  5   9   5  10  9   4
 7  1   2   9   5  6   9
 2  5  10   5   7  1   5
 7  4  10   5   9  8   3
julia> K = A ⊗ B20×28 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Int64,2}}:
  8.50341     9.56633     1.06293    …   1.00062     0.400249    2.00124
  7.44048     5.31463     9.56633        2.00124     1.80112     0.800497
  7.44048     1.06293     2.12585        1.00062     1.20075     1.80112
  2.12585     5.31463    10.6293         1.40087     0.200124    1.00062
  7.44048     4.2517     10.6293         1.80112     1.60099     0.600373
 -6.62732    -7.45573    -0.828415   …  -0.062278   -0.0249112  -0.124556
 -5.7989     -4.14207    -7.45573       -0.124556   -0.1121     -0.0498224
 -5.7989     -0.828415   -1.65683       -0.062278   -0.0747336  -0.1121
 -1.65683    -4.14207    -8.28415       -0.0871893  -0.0124556  -0.062278
 -5.7989     -3.31366    -8.28415       -0.1121     -0.0996449  -0.0373668
  0.0753288   0.0847449   0.0094161  …   2.65053     1.06021     5.30106
  0.0659127   0.0470805   0.0847449      5.30106     4.77095     2.12042
  0.0659127   0.0094161   0.0188322      2.65053     3.18063     4.77095
  0.0188322   0.0470805   0.094161       3.71074     0.530106    2.65053
  0.0659127   0.0376644   0.094161       4.77095     4.24085     1.59032
  9.74842    10.967       1.21855    …  -2.66103    -1.06441    -5.32207
  8.52987     6.09276    10.967         -5.32207    -4.78986    -2.12883
  8.52987     1.21855     2.43711       -2.66103    -3.19324    -4.78986
  2.43711     6.09276    12.1855        -3.72545    -0.532207   -2.66103
  8.52987     4.87421    12.1855        -4.78986    -4.25765    -1.59662
julia> K[4, 5]7.440480584383761
julia> eltype(K)  # promotionFloat64
julia> collect(K)20×28 Array{Float64,2}:
  8.50341     9.56633     1.06293    …   1.00062     0.400249    2.00124
  7.44048     5.31463     9.56633        2.00124     1.80112     0.800497
  7.44048     1.06293     2.12585        1.00062     1.20075     1.80112
  2.12585     5.31463    10.6293         1.40087     0.200124    1.00062
  7.44048     4.2517     10.6293         1.80112     1.60099     0.600373
 -6.62732    -7.45573    -0.828415   …  -0.062278   -0.0249112  -0.124556
 -5.7989     -4.14207    -7.45573       -0.124556   -0.1121     -0.0498224
 -5.7989     -0.828415   -1.65683       -0.062278   -0.0747336  -0.1121
 -1.65683    -4.14207    -8.28415       -0.0871893  -0.0124556  -0.062278
 -5.7989     -3.31366    -8.28415       -0.1121     -0.0996449  -0.0373668
  0.0753288   0.0847449   0.0094161  …   2.65053     1.06021     5.30106
  0.0659127   0.0470805   0.0847449      5.30106     4.77095     2.12042
  0.0659127   0.0094161   0.0188322      2.65053     3.18063     4.77095
  0.0188322   0.0470805   0.094161       3.71074     0.530106    2.65053
  0.0659127   0.0376644   0.094161       4.77095     4.24085     1.59032
  9.74842    10.967       1.21855    …  -2.66103    -1.06441    -5.32207
  8.52987     6.09276    10.967         -5.32207    -4.78986    -2.12883
  8.52987     1.21855     2.43711       -2.66103    -3.19324    -4.78986
  2.43711     6.09276    12.1855        -3.72545    -0.532207   -2.66103
  8.52987     4.87421    12.1855        -4.78986    -4.25765    -1.59662

Constructing Kronecker products

Kronecker.kronecker — Function

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Construct a Kronecker product object between two arrays. Does not evaluate the Kronecker product explictly.

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Higher-order Kronecker lazy kronecker product, e.g.

kronecker(A, B, C, D)

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Function

⊗(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kronecker, computes as Lazy Kronecker product. See kronecker for documentation.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Base.collect — Method

collect(K::GeneralizedKroneckerProduct)

Collects a lazy instance of the GeneralizedKroneckerProduct type into a dense, native matrix. Falls back to the element-wise case when not specialized method is defined.

Basic properties of Kronecker products

Base.getindex — Function

getindex(K::AbstractKroneckerProduct, i1::Integer, i2::Integer)

Computes and returns the (i,j)-th element of an AbstractKroneckerProduct K. Uses recursion if K is of an order greater than two.

Missing docstring.

Missing docstring for eltype. Check Documenter's build log for details.

Base.size — Method

size(K::AbstractKroneckerProduct)

Returns a the size of an AbstractKroneckerProduct instance.

Base.collect — Function

collect(K::GeneralizedKroneckerProduct)

Collects a lazy instance of the GeneralizedKroneckerProduct type into a dense, native matrix. Falls back to the element-wise case when not specialized method is defined.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

function collect(E::Eigen{<:Number, <:Number, <:AbstractKroneckerProduct})

Collects eigenvalue decomposition of a AbstractKroneckerProduct type into a matrix.

Kronecker.collect! — Function

collect!(C::AbstractMatrix, K::GeneralizedKroneckerProduct)

In-place collection of K in C. If possible, specialized routines are used to speed up the computation. The fallback is an element-wise iteration. In this case, this function might be slow.

collect!(C::AbstractMatrix, K::AbstractKroneckerProduct)

In-place collection of K in C where K is an AbstractKroneckerProduct, i.e., K = A ⊗ B. This is equivalent to the broadcasted assignment C .= K.

collect!(f, C::AbstractMatrix, K1::AbstractKroneckerProduct, Ks::AbstractKroneckerProduct...)

Evaluate f.(K1, Ks...) and assign it in-place to C. This is equivalent to the broadcasted operation C .= f.(K1, Ks...).

collect!(C::AbstractMatrix, K::AbstractKroneckerSum)

In-place collection of K in C where K is an AbstractKroneckerSum, i.e., K = A ⊗ B.

Kronecker.order — Function

order(M::AbstractMatrix)

Returns the order of a matrix, i.e. how many matrices are involved in the Kronecker product (default to 1 for general matrices).

Kronecker.getmatrices — Function

getmatrices(K::AbstractKroneckerProduct)

Obtain the two matrices of an AbstractKroneckerProduct object.

getmatrices(A::AbstractArray)

Returns a matrix itself. Needed for recursion.

getmatrices(K::T) where T <: AbstractKroneckerSum

Obtain the two matrices of an AbstractKroneckerSum object.

Kronecker.issquare — Function

issquare(A::AbstractMatrix)

Checks if an array is a square matrix.

issquare(A::Factorization)

Checks if a Factorization struct represents a square matrix.

LinearAlgebra.issymmetric — Function

issymmetric(K::AbstractKroneckerProduct)

Checks if a Kronecker product is symmetric.

Missing docstring.

Missing docstring for sum. Check Documenter's build log for details.

Linear algebra

Many functions of the LinearAlgebra module are overloaded to work with subtypes of GeneralizedKroneckerProduct.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::AbstractKroneckerProduct)

Compute the logarithm of the determinant of a Kronecker product.

LinearAlgebra.tr — Method

tr(K::AbstractKroneckerProduct)

Compute the trace of a Kronecker product.

Base.inv — Method

inv(K::AbstractKroneckerProduct)

Compute the inverse of a Kronecker product.

Base.adjoint — Method

adjoint(K::AbstractKroneckerProduct)

Compute the adjoint of a Kronecker product.

Base.transpose — Method

transpose(K::AbstractKroneckerProduct)

Compute the transpose of a Kronecker product.

Missing docstring.

Missing docstring for conj(K::AbstractKroneckerProduct). Check Documenter's build log for details.