Basic use

Basic use

Compute a lazy Kronecker products between two matrices A and B by either

K = kronecker(A, B)

or, by using the binary operator:

K = A ⊗ B

Note, ⊗ can be formed by typing \otimes<tab>.

The Kronecker product K behaves like a matrix, for which size(K), eltype(K) works as one would expect. Elements can be accessed via K[i,j]; every element is computed on the fly. The function collect can be used to turn K in a regular, dense matrix.

julia> using Kronecker

julia> A = randn(4, 4)
4×4 Array{Float64,2}:
 -0.086713   1.55361    -1.3854    -0.433724
 -0.851836  -0.0376959   0.236003   0.41851
 -2.13986   -0.169267    0.359041   0.172387
 -0.448097  -1.97147     0.910317   0.529855

julia> B = rand(1:10, 5, 7)
5×7 Array{Int64,2}:
 4  7  2  5  8  1  2
 6  9  6  9  7  4  3
 1  5  8  1  2  3  8
 4  9  4  3  9  8  5
 1  1  9  1  7  6  8

julia> K = A ⊗ B
20×28 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Int64,2}}:
  -0.346852   -0.606991   -0.173426  …  -3.46979   -0.433724  -0.867448
  -0.520278   -0.780417   -0.520278     -3.03607   -1.7349    -1.30117
  -0.086713   -0.433565   -0.693704     -0.867448  -1.30117   -3.46979
  -0.346852   -0.780417   -0.346852     -3.90351   -3.46979   -2.16862
  -0.086713   -0.086713   -0.780417     -3.03607   -2.60234   -3.46979
  -3.40734    -5.96285    -1.70367   …   3.34808    0.41851    0.83702
  -5.11102    -7.66652    -5.11102       2.92957    1.67404    1.25553
  -0.851836   -4.25918    -6.81469       0.83702    1.25553    3.34808
  -3.40734    -7.66652    -3.40734       3.76659    3.34808    2.09255
  -0.851836   -0.851836   -7.66652       2.92957    2.51106    3.34808
  -8.55945   -14.979      -4.27973   …   1.3791     0.172387   0.344775
 -12.8392    -19.2588    -12.8392        1.20671    0.689549   0.517162
  -2.13986   -10.6993    -17.1189        0.344775   0.517162   1.3791
  -8.55945   -19.2588     -8.55945       1.55149    1.3791     0.861937
  -2.13986    -2.13986   -19.2588        1.20671    1.03432    1.3791
  -1.79239    -3.13668    -0.896195  …   4.23884    0.529855   1.05971
  -2.68858    -4.03288    -2.68858       3.70898    2.11942    1.58956
  -0.448097   -2.24049    -3.58478       1.05971    1.58956    4.23884
  -1.79239    -4.03288    -1.79239       4.76869    4.23884    2.64927
  -0.448097   -0.448097   -4.03288       3.70898    3.17913    4.23884

julia> K[4, 5]
-0.7804171399052441

julia> eltype(K)  # promotion
Float64

julia> collect(K)
20×28 Array{Float64,2}:
  -0.346852   -0.606991   -0.173426  …  -3.46979   -0.433724  -0.867448
  -0.520278   -0.780417   -0.520278     -3.03607   -1.7349    -1.30117
  -0.086713   -0.433565   -0.693704     -0.867448  -1.30117   -3.46979
  -0.346852   -0.780417   -0.346852     -3.90351   -3.46979   -2.16862
  -0.086713   -0.086713   -0.780417     -3.03607   -2.60234   -3.46979
  -3.40734    -5.96285    -1.70367   …   3.34808    0.41851    0.83702
  -5.11102    -7.66652    -5.11102       2.92957    1.67404    1.25553
  -0.851836   -4.25918    -6.81469       0.83702    1.25553    3.34808
  -3.40734    -7.66652    -3.40734       3.76659    3.34808    2.09255
  -0.851836   -0.851836   -7.66652       2.92957    2.51106    3.34808
  -8.55945   -14.979      -4.27973   …   1.3791     0.172387   0.344775
 -12.8392    -19.2588    -12.8392        1.20671    0.689549   0.517162
  -2.13986   -10.6993    -17.1189        0.344775   0.517162   1.3791
  -8.55945   -19.2588     -8.55945       1.55149    1.3791     0.861937
  -2.13986    -2.13986   -19.2588        1.20671    1.03432    1.3791
  -1.79239    -3.13668    -0.896195  …   4.23884    0.529855   1.05971
  -2.68858    -4.03288    -2.68858       3.70898    2.11942    1.58956
  -0.448097   -2.24049    -3.58478       1.05971    1.58956    4.23884
  -1.79239    -4.03288    -1.79239       4.76869    4.23884    2.64927
  -0.448097   -0.448097   -4.03288       3.70898    3.17913    4.23884

Constructing Kronecker products

Kronecker.kronecker — Function.

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Construct a Kronecker product object between two arrays. Does not evaluate the Kronecker product explictly.

source

kronecker(A::AbstractMatrix, B::AbstractMatrix)

Higher-order Kronecker lazy kronecker product, e.g.

kronecker(A, B, C, D)

source

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Function.

⊗(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kronecker, computes as Lazy Kronecker product. See kronecker for documentation.

source

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Base.collect — Method.

collect(K::AbstractKroneckerProduct)

Collects a lazy instance of the AbstractKroneckerProduct type into a full, native matrix. Equivalent with Matrix(K::AbstractKroneckerProduct).

source

Basic properties of Kronecker products

Base.getindex — Function.

getindex(K::AbstractKroneckerProduct, i1::Integer, i2::Integer)

Computes and returns the (i,j)-th element of an AbstractKroneckerProduct K. Uses recursion if K is of an order greater than two.

source

Missing docstring.

Missing docstring for eltype. Check Documenter's build log for details.

Base.size — Method.

size(K::AbstractKroneckerProduct)

Returns a the size of an AbstractKroneckerProduct instance.

source

Kronecker.order — Function.

order(M::AbstractMatrix)

Returns the order of a matrix, i.e. how many matrices are involved in the Kronecker product (default to 1 for general matrices).

source

Kronecker.getmatrices — Function.

getmatrices(K::AbstractKroneckerProduct)

Obtain the two matrices of an AbstractKroneckerProduct object.

source

getmatrices(A::AbstractArray)

Returns a matrix itself. Needed for recursion.

source

getmatrices(K::T) where T <: AbstractKroneckerSum

Obtain the two matrices of an AbstractKroneckerSum object.

source

Kronecker.issquare — Function.

issquare(A::AbstractMatrix)

Checks if an array is a square matrix.

source

LinearAlgebra.issymmetric — Function.

issymmetric(K::AbstractKroneckerProduct)

Checks if a Kronecker product is symmetric.

source

Missing docstring.

Missing docstring for sum. Check Documenter's build log for details.

Linear algebra

Many functions of the LinearAlgebra module are overloaded to work with subtypes of GeneralizedKroneckerProduct.

LinearAlgebra.det — Method.

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::AbstractKroneckerProduct)

Compute the logarithm of the determinant of a Kronecker product.

source

LinearAlgebra.tr — Method.

tr(K::AbstractKroneckerProduct)

Compute the trace of a Kronecker product.

source

Base.inv — Method.

inv(K::AbstractKroneckerProduct)

Compute the inverse of a Kronecker product.

source

Base.adjoint — Method.

adjoint(K::AbstractKroneckerProduct)

Compute the adjoint of a Kronecker product.

source

Base.transpose — Method.

transpose(K::AbstractKroneckerProduct)

Compute the transpose of a Kronecker product.

source

Missing docstring.

Missing docstring for conj(K::AbstractKroneckerProduct). Check Documenter's build log for details.