Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  68.9629    -4.54681  -13.1822   …   -4.01874  -11.6512     8.37575
  -4.54681   32.2265     5.7396       28.4837     5.073    -13.2293
 -13.1822     5.7396     5.35335       5.073      4.73161   -1.79693
   9.47633  -14.9676    -2.03305     -13.2293    -1.79693   14.0681
  47.742     -3.14769   -9.12584      -3.48978  -10.1176     7.2733
  -3.14769   22.3099     3.97344  …   24.7346     4.40528  -11.488
  -9.12584    3.97344    3.70604       4.40528    4.10881   -1.56041
   6.56032  -10.3619    -1.40745     -11.488     -1.56041   12.2164
  65.3643    -4.30955  -12.4943       -5.74461  -16.6549    11.9727
  -4.30955   30.5449     5.4401       40.7162     7.25163  -18.9106
   ⋮                              ⋱
   8.11417  -12.8161    -1.74081     -14.4921    -1.96845   15.4109
  42.0635    -2.7733    -8.0404       -3.37307   -9.77928    7.03006
  -2.7733    19.6564     3.50083      23.9074     4.25795  -11.1038
  -8.0404     3.50083    3.26524       4.25795    3.9714    -1.50823
   5.78003   -9.1294    -1.24005  …  -11.1038    -1.50823   11.8078
  60.9535    -4.01874  -11.6512       -5.88536  -17.063     12.2661
  -4.01874   28.4837     5.073        41.7138     7.4293   -19.374
 -11.6512     5.073      4.73161       7.4293     6.92933   -2.63156
   8.37575  -13.2293    -1.79693     -19.374     -2.63156   20.6024
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0012697722614184578
   0.007034357336474403
   0.03664601926976728
   0.07069128162353247
   0.004509822005162944
   0.024983771084094428
   0.13015485463495838
   0.2510726585591353
   0.03363689901452929
   0.18634362597808063
   ⋮
  33.61639031075739
   0.7670096242022658
   4.249123989467591
  22.136134425565434
  42.7012740787645
  10.063007103794137
  55.74762498630476
 290.4214900910612
 560.231854773118
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0885604  -0.00593679   …   0.0040334    0.0892149   0.287043
 -0.124815    0.182957        -0.124299     0.259986   -0.0612842
  0.40431     0.177004        -0.120255     0.0176331  -0.0613672
 -0.129388    0.372546        -0.253104    -0.134635    0.0638276
 -0.0644174   0.00431833       0.00378709   0.0837668   0.269514
  0.0907888  -0.13308      …  -0.116708     0.244109   -0.0575417
 -0.294089   -0.12875         -0.112911     0.0165563  -0.0576197
  0.0941146  -0.270984        -0.237648    -0.126413    0.0599298
 -0.0219936   0.00147437       0.00580285   0.128354    0.412969
  0.0309973  -0.0454365       -0.178829     0.374042   -0.0881696
  ⋮                        ⋱
  0.0229806  -0.0661679       -0.288737    -0.153589    0.0728134
  0.028895   -0.00193702       0.00375136   0.0829764   0.266971
 -0.040724    0.0596941       -0.115607     0.241806   -0.0569987
  0.131916    0.0577519       -0.111846     0.0164001  -0.057076
 -0.0422159   0.121552     …  -0.235405    -0.12522     0.0593643
  0.010129   -0.000679017      0.00509163   0.112622    0.362354
 -0.0142757   0.0209256       -0.156911     0.328197   -0.0773631
  0.0462427   0.0202447       -0.151806     0.0222595  -0.0774679
 -0.0147986   0.0426097       -0.31951     -0.169958    0.0805738
julia> logdet(E)9.038754548544187
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  1.5918047635808747
 -3.255373281607514
 -3.9544491553339287
  3.9884122627716714
 -2.2524954242320643
  3.762353804130988
  4.732740918661407
  0.47872296759079197
  1.0223845114676573
 -0.3711170184331754
  ⋮
  1.240482330034785
  1.1157081214792601
 -2.43370657268012
 -4.107991529115209
 -2.1942516927390927
 -0.0732570050446599
 -0.7012102210509307
  0.11272281564265328
  0.28671165784808783

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.4853   -1.79675     4.16783   …  -1.59449     3.69866   -1.89271
 -1.79675   4.46272     0.98103       3.96035     0.870595  -0.0685515
  4.16783   0.98103     7.37684       0.870595    6.54643   -3.72125
 -2.13279  -0.0772472  -4.19329      -0.0685515  -3.72125    2.31591
  4.71923  -1.54582     3.58576      -1.50616     3.49377   -1.78786
 -1.54582   3.83946     0.84402   …   3.74097     0.822368  -0.064754
  3.58576   0.84402     6.34659       0.822368    6.18379   -3.51511
 -1.83493  -0.0664589  -3.60766      -0.064754   -3.51511    2.18762
  4.33488  -1.41992     3.29372      -1.77739     4.12293   -2.10982
 -1.41992   3.52676     0.775281      4.41464     0.970461  -0.076415
  ⋮                               ⋱
 -2.3565   -0.0853494  -4.63311      -0.11046    -5.9962     3.73172
  6.62215  -2.16913     5.03163      -2.55944     5.93701   -3.03813
 -2.16913   5.38763     1.18435       6.35707     1.39746   -0.110037
  5.03163   1.18435     8.90571       1.39746    10.5082    -5.97327
 -2.57482  -0.0932569  -5.06236   …  -0.110037   -5.97327    3.71745
  4.86782  -1.59449     3.69866      -2.54954     5.91406   -3.02638
 -1.59449   3.96035     0.870595      6.33249     1.39206   -0.109612
  3.69866   0.870595    6.54643       1.39206    10.4676    -5.95018
 -1.89271  -0.0685515  -3.72125      -0.109612   -5.95018    3.70308
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.34207  -0.767162  1.77955  -0.910644  …  -0.680803    1.57923   -0.808133
 0.0       1.96829   1.19201  -0.394178      1.74672     1.05783   -0.349806
 0.0       0.0       1.67007  -1.25916       0.0         1.48207   -1.11742
 0.0       0.0       0.0       0.19888       0.0         0.0        0.176492
 0.0      -0.0       0.0      -0.0          -0.0748484   0.173622  -0.0888472
 0.0       0.0       0.0      -0.0       …   0.192037    0.116299  -0.0384581
 0.0       0.0       0.0      -0.0           0.0         0.162941  -0.12285
 0.0       0.0       0.0       0.0           0.0         0.0        0.0194038
 0.0      -0.0       0.0      -0.0          -0.32916     0.763537  -0.390723
 0.0       0.0       0.0      -0.0           0.84452     0.511448  -0.169127
 ⋮                                       ⋱
 0.0       0.0       0.0       0.0          -0.0        -0.0       -0.0110503
 0.0      -0.0       0.0      -0.0           0.0833327  -0.193303   0.0989183
 0.0       0.0       0.0      -0.0          -0.213805   -0.129482   0.0428175
 0.0       0.0       0.0      -0.0          -0.0        -0.181411   0.136776
 0.0       0.0       0.0       0.0       …  -0.0        -0.0       -0.0216033
 0.0      -0.0       0.0      -0.0          -0.219079    0.508188  -0.260054
 0.0       0.0       0.0      -0.0           0.562088    0.340405  -0.112566
 0.0       0.0       0.0      -0.0           0.0         0.476925  -0.35958
 0.0       0.0       0.0       0.0           0.0         0.0        0.0567945
julia> logdet(C)-50.301462783624785
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  20.9074    18.0161    -48.4003  …   -11.835      31.7947    39.5135
  18.0161    18.2056    -45.3022      -11.9594     29.7596    37.7919
 -48.4003   -45.3022    130.76         29.7596    -85.8975  -111.156
 -60.1504   -57.5297    169.21         37.7919   -111.156   -147.431
 -55.3641   -47.7079    128.167        68.838    -184.933   -229.829
 -47.7079   -48.2094    119.963   …    69.5617   -173.096   -219.816
 128.167    119.963    -346.26       -173.096     499.62     646.537
 159.282    152.342    -448.08       -219.816     646.537    857.528
  27.5341    23.7265    -63.741       -35.7979     96.1709   119.518
  23.7265    23.9759    -59.661       -36.1743     90.0151   114.311
   ⋮                              ⋱
   4.63369    4.43181   -13.0351      -13.4982     39.7019    52.6583
 -17.2385   -14.8546     39.9068        9.48591   -25.4838   -31.6706
 -14.8546   -15.0108     37.3524        9.58564   -23.8527   -30.2907
  39.9068    37.3524   -107.813       -23.8527     68.8479    89.0931
  49.595     47.4341   -139.517   …   -30.2907     89.0931   118.168
 -13.7343   -11.835      31.7947       24.8867    -66.858    -83.0891
 -11.835    -11.9594     29.7596       25.1483    -62.5785   -79.469
  31.7947    29.7596    -85.8975      -62.5785    180.625    233.74
  39.5135    37.7919   -111.156       -79.469     233.74     310.018

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.