Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 15.1769     1.66682  -0.987142  …   1.15024   -0.681206  -4.05048
  1.66682   15.6856    1.86601      10.8243     1.2877    -6.44722
 -0.987142   1.86601   2.36335       1.2877     1.6309     0.415279
 -5.86958   -9.34273   0.601785     -6.44722    0.415279   5.57235
 11.3403     1.24546  -0.737601      1.31928   -0.781318  -4.64575
  1.24546   11.7204    1.3943    …  12.4151     1.47694   -7.39473
 -0.737601   1.3943    1.76592       1.47694    1.87058    0.47631
 -4.3858    -6.98097   0.449659     -7.39473    0.47631    6.39128
 10.7202     1.17736  -0.697269      1.25431   -0.742841  -4.41696
  1.17736   11.0795    1.31806      11.8037     1.40421   -7.03057
  ⋮                              ⋱
 -5.28412   -8.41083   0.541759     -7.7811     0.501197   6.72523
 10.1644     1.11632  -0.661118      0.901862  -0.534112  -3.17585
  1.11632   10.5051    1.24972       8.48698    1.00964   -5.05506
 -0.661118   1.24972   1.5828        1.00964    1.27873    0.325607
 -3.93103   -6.2571    0.403033  …  -5.05506    0.325607   4.3691
 10.4733     1.15024  -0.681206      1.68481   -0.9978    -5.93296
  1.15024   10.8243    1.2877       15.8549     1.88616   -9.4436
 -0.681206   1.2877    1.6309        1.88616    2.38887    0.608282
 -4.05048   -6.44722   0.415279     -9.4436     0.608282   8.16213
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   4.042548445772057e-6
   0.00032661923354956423
   0.0017191178042689629
   0.0029875959970246477
   0.0010253196738763644
   0.08284109158296786
   0.43602329819240304
   0.7577499674858239
   0.0021463091119597283
   0.17341185801790646
   ⋮
  15.135573822967576
   0.02826733620701901
   2.283870140401278
  12.020853084288095
  20.890629174021225
   0.26896616139543955
  21.731222931323245
 114.37946211488682
 198.77615269163064
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0497186  -0.0615319   0.208277    …  0.0774455  -0.262142    -0.149632
  0.117559   -0.0357387  -0.137227       0.0449815   0.172718    -0.217426
 -0.120041   -0.218648   -0.0390453      0.275196    0.0491434   -0.00684658
  0.181849   -0.104406    0.00599442     0.131408   -0.00754472   0.17695
  0.0163477  -0.0202319   0.0684821      0.0745626  -0.252384    -0.144062
  0.0386539  -0.011751   -0.0451209   …  0.0433071   0.166289    -0.209332
 -0.03947    -0.0718922  -0.0128382      0.264952    0.047314    -0.00659172
  0.0597925  -0.0343291   0.00197099     0.126517   -0.00726388   0.170363
 -0.0964424   0.119357   -0.404007       0.0673516  -0.227976    -0.13013
 -0.228037    0.0693245   0.266189       0.0391189   0.150207    -0.189088
  ⋮                                   ⋱
 -0.307666    0.176643   -0.0101418      0.149681   -0.00859384   0.201555
  0.068444   -0.0847064   0.286719       0.0566384  -0.191713    -0.109431
  0.161835   -0.0491988  -0.188911       0.0328965   0.126314    -0.159011
 -0.165252   -0.300997   -0.0537508      0.20126     0.0359402   -0.00500713
  0.250338   -0.143728    0.00825208  …  0.0961031  -0.0055177    0.129409
  0.0412873  -0.0510972   0.172957       0.0707601  -0.239513    -0.136715
  0.0976235  -0.0296781  -0.113956       0.0410986   0.157808    -0.198657
 -0.0996847  -0.18157    -0.032424       0.25144     0.0449011   -0.00625555
  0.151011   -0.0867009   0.00497788     0.120065   -0.00689343   0.161675
julia> logdet(E)-40.18003137895454
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -4.605727191443102
 -5.056504532014271
  3.027601508337808
 -7.282548703733808
 -0.6463355145250703
 -0.2173314823349404
  0.12818731917572587
  1.2057350957876878
  2.4036907540974983
 -2.820782633837217
  ⋮
 -7.221673086421257
 -2.820527651399959
 -1.3639013360036758
  1.2357674455698113
 -2.4582469736798864
 -0.44901830412734073
  3.126145861461582
 -5.100247066350724
  4.754713498686276

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.234     -0.13219    1.5028    …  -0.143775   1.6345    -1.89145
 -0.13219    2.3143    -0.265765      2.51713   -0.289057  -1.01611
  1.5028    -0.265765   4.84797      -0.289057   5.27284    3.50022
 -1.73904   -0.934238   3.21818      -1.01611    3.50022    7.33377
 17.1873    -0.202242   2.29918      -0.289059   3.28616   -3.80275
 -0.202242   3.54073   -0.406604  …   5.06067   -0.581148  -2.04289
  2.29918   -0.406604   7.41707      -0.581148  10.601      7.03718
 -2.66062   -1.42932    4.9236       -2.04289    7.03718   14.7445
 15.421     -0.181458   2.06291      -0.260325   2.9595    -3.42474
 -0.181458   3.17687   -0.364819      4.55762   -0.523379  -1.83982
  ⋮                               ⋱
 -1.54939   -0.832352   2.86721      -0.957482   3.29825    6.9106
 12.0093    -0.141313   1.60651      -0.221223   2.51497   -2.91033
 -0.141313   2.47402   -0.284107      3.87304   -0.444765  -1.56347
  1.60651   -0.284107   5.18254      -0.444765   8.11319    5.3857
 -1.85906   -0.998712   3.44027   …  -1.56347    5.3857    11.2843
 12.2185    -0.143775   1.6345       -0.340543   3.87145   -4.48006
 -0.143775   2.51713   -0.289057      5.96203   -0.684656  -2.40675
  1.6345    -0.289057   5.27284      -0.684656  12.4892     8.29057
 -1.89145   -1.01611    3.50022      -2.40675    8.29057   17.3707
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.35172  -0.0394394   0.448366  …  -0.0428959   0.487661   -0.564322
 0.0       1.52077    -0.163129      1.65405    -0.177426   -0.682791
 0.0       0.0         2.14949       0.0         2.33787     1.69429
 0.0       0.0         0.0           0.0         0.0         2.07876
 0.0      -0.0         0.0          -0.024588    0.279528   -0.323471
 0.0       0.0        -0.0       …   0.948106   -0.101701   -0.391377
 0.0       0.0         0.0           0.0         1.34007     0.971169
 0.0       0.0         0.0           0.0         0.0         1.19155
 0.0      -0.0         0.0          -0.0229213   0.26058    -0.301544
 0.0       0.0        -0.0           0.883838   -0.0948069  -0.364848
 ⋮                               ⋱
 0.0       0.0         0.0          -0.0        -0.0        -0.62327
 0.0      -0.0         0.0          -0.0160996   0.183028   -0.2118
 0.0       0.0        -0.0           0.620795   -0.066591   -0.256264
 0.0       0.0         0.0           0.0         0.877446    0.635896
 0.0       0.0         0.0       …   0.0         0.0         0.780195
 0.0      -0.0         0.0          -0.0105725   0.120193   -0.139087
 0.0       0.0        -0.0           0.407671   -0.0437297  -0.168286
 0.0       0.0         0.0           0.0         0.576211    0.417588
 0.0       0.0         0.0           0.0         0.0         0.512347
julia> logdet(C)2.5217958529805244
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   7.43182      1.70843     -5.32747   …   -0.251598    0.784568  -0.691588
   1.70843     31.238       -2.90218       -4.60037     0.4274    -0.90627
  -5.32747     -2.90218     24.1669         0.4274     -3.55902    1.96019
   4.6961       6.15386    -13.3103        -0.90627     1.96019   -2.70479
  -2.26586     -0.520879     1.62427        0.348689   -1.08733    0.95847
  -0.520879    -9.52405      0.884837  …    6.37564    -0.592333   1.256
   1.62427      0.884837    -7.36817       -0.592333    4.93244   -2.71663
  -1.43178     -1.87623      4.05815        1.256      -2.71663    3.74856
 -16.934       -3.89281     12.1391         1.27517    -3.97641    3.50516
  -3.89281    -71.1785       6.61287       23.3159     -2.16618    4.59322
   ⋮                                   ⋱
  -0.0621629   -0.0814594    0.176191       0.479787   -1.03774    1.43194
   7.04068      1.61852     -5.04708       -0.854391    2.66428   -2.34853
   1.61852     29.5939      -2.74944      -15.6222      1.45139   -3.07756
  -5.04708     -2.74944     22.895          1.45139   -12.0859     6.65653
   4.44895      5.82998    -12.6098    …   -3.07756     6.65653   -9.18506
  -1.09447     -0.251598     0.784568       0.354361   -1.10502    0.97406
  -0.251598    -4.60037      0.4274         6.47934    -0.601967   1.27643
   0.784568     0.4274      -3.55902       -0.601967    5.01267   -2.76081
  -0.691588    -0.90627      1.96019        1.27643    -2.76081    3.80953

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.