Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  20.3836   -27.4681     6.90324  …  -32.8648     8.25953    9.29572
 -27.4681    51.3612    -8.86326      61.4522   -10.6046   -20.0429
   6.90324   -8.86326    5.2493      -10.6046     6.28063    8.90278
   7.76928  -16.7517     7.44087     -20.0429     8.90278   19.3548
  15.5502   -20.9548     5.26633     -35.3758     8.8906    10.006
 -20.9548    39.1823    -6.76159  …   66.1475   -11.4149   -21.5743
   5.26633   -6.76159    4.00458     -11.4149     6.76051    9.58301
   5.92702  -12.7795     5.67648     -21.5743     9.58301   20.8336
  16.5707   -22.33       5.61193     -30.7032     7.71629    8.68433
 -22.33      41.7537    -7.20531      57.4104    -9.90715  -18.7246
   ⋮                              ⋱
   6.24452  -13.4641     5.98056     -22.2098     9.86532   21.4473
  11.2577   -15.1704     3.81262     -27.5299     6.91879    7.78678
 -15.1704    28.3665    -4.89512      51.4769    -8.88322  -16.7894
   3.81262   -4.89512    2.89915      -8.88322    5.26112    7.45762
   4.29093   -9.25184    4.10954  …  -16.7894     7.45762   16.213
  24.3884   -32.8648     8.25953     -53.7992    13.5207    15.217
 -32.8648    61.4522   -10.6046      100.596    -17.3597   -32.81
   8.25953  -10.6046     6.28063     -17.3597    10.2813    14.5737
   9.29572  -20.0429     8.90278     -32.81      14.5737    31.6835
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0020059808017702315
   0.032644356905533745
   0.0729768389060243
   0.46580100790201956
   0.006860692037568775
   0.11164756875823866
   0.2495894363334231
   1.5930946413768015
   0.013132773367256885
   0.21371637284957787
   ⋮
  39.16372823942696
   0.2566687599648018
   4.176902689894976
   9.33751446256654
  59.600055485606
   3.2570474147638713
  53.003607099314785
 118.49017911175964
 756.3063251865825
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.182489    0.319895   -0.128635    …  -0.196661    0.0790805   0.12679
  0.0949668   0.218034    0.105513       -0.13404    -0.0648657  -0.219108
 -0.349066    0.208766    0.151516       -0.128342   -0.0931472   0.0489098
  0.174782   -0.0355318   0.379579        0.0218438  -0.233353    0.0843504
  0.140121    0.245624   -0.0987694      -0.216035    0.0868709   0.13928
  0.0729182   0.167413    0.0810155   …  -0.147245   -0.0712558  -0.240693
 -0.268023    0.160296    0.116338       -0.140986   -0.102323    0.053728
  0.134202   -0.0272823   0.291451        0.0239957  -0.256341    0.09266
  0.0111874   0.019611   -0.00788589     -0.205947    0.0828146   0.132777
  0.0058219   0.0133665   0.0064684      -0.14037    -0.0679286  -0.229454
  ⋮                                   ⋱
  0.0861463  -0.0175129   0.187087        0.0267725  -0.286005    0.103383
  0.0576918   0.101131   -0.0406663      -0.18042     0.0725499   0.11632
  0.0300226   0.0689287   0.0333565      -0.122971   -0.059509   -0.201014
 -0.110353    0.0659987   0.0478999      -0.117744   -0.085455    0.0448707
  0.0552551  -0.0112329   0.119999    …   0.0200399  -0.214082    0.0773847
 -0.236186   -0.414022    0.166485       -0.302245    0.121538    0.194862
 -0.12291    -0.282189   -0.136559       -0.206004   -0.0996911  -0.336744
  0.451777   -0.270194   -0.196099       -0.197247   -0.143156    0.0751686
 -0.22621     0.0459868  -0.491268        0.0335714  -0.358636    0.129637
julia> logdet(E)9.689178674770181
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  1.5789761913033553
  0.8492528129055551
 -0.33575129294843326
 -0.012566621619695571
 -1.2293324313982399
 -0.6293777624093359
  3.518940521500802
 -1.9343245022084694
 -0.39881307624079615
  0.08328535921461913
  ⋮
  1.6771250019129993
  1.6162114299049788
  1.0435673329875228
 -4.342843599973339
  3.311053329534554
 -0.8148764574728228
 -0.32650295241227134
 -0.6241375676550377
  1.0018536409495156

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3.56063   -0.911065     2.84906   …  -0.711968     2.22645    0.328764
 -0.911065   2.15805     -0.518139      1.68645     -0.404909  -0.00201721
  2.84906   -0.518139     4.53049      -0.404909     3.54044    2.95191
  0.420701  -0.00258131   3.77739      -0.00201721   2.95191    7.32076
  3.35229   -0.857756     2.68236      -0.900977     2.81752    0.416043
 -0.857756   2.03178     -0.487822  …   2.13416     -0.512402  -0.00255272
  2.68236   -0.487822     4.26541      -0.512402     4.48033    3.73557
  0.396085  -0.00243027   3.55637      -0.00255272   3.73557    9.26423
  3.17852   -0.813294     2.54331      -0.74519      2.33034    0.344105
 -0.813294   1.92646     -0.462535      1.76514     -0.423803  -0.00211133
  ⋮                                 ⋱
  0.385174  -0.00236333   3.45841      -0.0028839    4.2202    10.4661
  2.48759   -0.636504     1.99046      -0.495553     1.54968    0.228831
 -0.636504   1.5077      -0.361992      1.17382     -0.28183   -0.00140404
  1.99046   -0.361992     3.16518      -0.28183      2.46426    2.05463
  0.293917  -0.0018034    2.63903   …  -0.00140404   2.05463    5.09548
  2.78252   -0.711968     2.22645      -1.01211      3.16506    0.467362
 -0.711968   1.68645     -0.404909      2.39741     -0.575608  -0.00286761
  2.22645   -0.404909     3.54044      -0.575608     5.03299    4.19636
  0.328764  -0.00201721   2.95191      -0.00286761   4.19636   10.407
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.88696  -0.48282  1.50987   0.222951   …   1.17991      0.174229
 0.0       1.38742  0.151976  0.0757261      0.118764     0.0591776
 0.0       0.0      1.49255   2.29758        1.16638      1.79549
 0.0       0.0      0.0       2.00839        0.0          1.56949
 0.0      -0.0      0.0       0.0            0.771302     0.113893
 0.0       0.0      0.0       0.0        …   0.0776356    0.0386841
 0.0       0.0      0.0       0.0            0.762457     1.1737
 0.0       0.0      0.0       0.0            0.0          1.02597
 0.0      -0.0      0.0       0.0           -0.043892    -0.00648122
 0.0       0.0      0.0       0.0           -0.00441796  -0.00220137
 ⋮                                       ⋱
 0.0       0.0      0.0       0.0            0.0          0.388914
 0.0      -0.0      0.0       0.0           -0.339483    -0.0501291
 0.0       0.0      0.0       0.0           -0.0341708   -0.0170265
 0.0       0.0      0.0       0.0           -0.33559     -0.516597
 0.0       0.0      0.0       0.0        …  -0.0         -0.451573
 0.0      -0.0      0.0       0.0            0.147725     0.0218135
 0.0       0.0      0.0       0.0            0.0148693    0.00740905
 0.0       0.0      0.0       0.0            0.146031     0.224796
 0.0       0.0      0.0       0.0            0.0          0.196501
julia> logdet(C)-35.651103950262936
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  25.5998     5.57041   -21.8505    …   -4.66325    18.2921    -6.41469
   5.57041   14.2836     -2.50702      -11.9574      2.09874   -0.640141
 -21.8505    -2.50702    28.0303         2.09874   -23.4655     8.64098
   7.66257    0.76467   -10.3219        -0.640141    8.64098   -5.61333
  -5.36092   -1.16651     4.57577       -7.58964    29.7711   -10.4402
  -1.16651   -2.99115     0.525001  …  -19.4612      3.41579   -1.04186
   4.57577    0.525001   -5.8699         3.41579   -38.1911    14.0636
  -1.60464   -0.160131    2.16154       -1.04186    14.0636    -9.13592
  -9.18158   -1.99787     7.83687        2.95499   -11.5913     4.06483
  -1.99787   -5.12291     0.899162       7.57713    -1.32992    0.405641
   ⋮                                ⋱
  15.5474     1.55152   -20.9432        -5.08463    68.6353   -44.5866
 -18.4957    -4.02458    15.7868        11.8533    -46.496     16.3053
  -4.02458  -10.3198      1.8113        30.3941     -5.33471    1.62715
  15.7868     1.8113    -20.2517        -5.33471    59.646    -21.9642
  -5.53614   -0.552468    7.45752   …    1.62715   -21.9642    14.2683
 -21.4308    -4.66325    18.2921        21.5149    -84.3946    29.5956
  -4.66325  -11.9574      2.09874       55.1683     -9.68301    2.95343
  18.2921     2.09874   -23.4655        -9.68301   108.263    -39.8671
  -6.41469   -0.640141    8.64098        2.95343   -39.8671    25.8983

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.