Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  16.6241   -6.17903   -4.9442    …  -5.42342  -4.33959    -8.76231
  -6.17903  13.6387     2.89762      11.9709    2.54329    -1.54669
  -4.9442    2.89762    4.26095       2.54329   3.7399      0.645213
  -9.9831   -1.76219    0.735106     -1.54669   0.645213    9.43415
   9.00821  -3.34827   -2.67914      -4.63554  -3.70916    -7.48937
  -3.34827   7.39047    1.57015   …  10.2318    2.17381    -1.322
  -2.67914   1.57015    2.30891       2.17381   3.19659     0.55148
  -5.4096   -0.954886   0.398337     -1.322     0.55148     8.06361
  12.3876   -4.60437   -3.68422      -6.02339  -4.81966    -9.73164
  -4.60437  10.163      2.15919      13.2951    2.82464    -1.7178
   ⋮                              ⋱
  -8.81579  -1.55614    0.649151     -1.85145   0.772342   11.293
  16.6927   -6.20451   -4.96459      -6.53905  -5.23227   -10.5648
  -6.20451  13.6949     2.90957      14.4333    3.06645    -1.86486
  -4.96459   2.90957    4.27853       3.06645   4.50922     0.777936
 -10.0243   -1.76945    0.738138  …  -1.86486   0.777936   11.3748
  14.5912   -5.42342   -4.33959      -7.30808  -5.84761   -11.8072
  -5.42342  11.9709     2.54329      16.1308    3.42708    -2.08418
  -4.33959   2.54329    3.7399        3.42708   5.03952     0.869426
  -8.76231  -1.54669    0.645213     -2.08418   0.869426   12.7125

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0004762355288379772
   0.004274840291216995
   0.01837115070801915
   0.03346882848761644
   0.0038563902758843504
   0.0346161752572989
   0.1487632120183834
   0.27101897466536234
   0.02208322705680082
   0.19822601016948188
   ⋮
  11.333485823443398
   0.305665455657345
   2.743749523823894
  11.79127933990596
  21.481523512004298
   3.824929283559985
  34.3337714687407
 147.54958011480616
 268.8079624825383
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.315983    0.0695863    0.0458891   …   0.0358717   0.0236558  -0.210329
 -0.135424    0.199144     0.424783        0.102658    0.218976    0.0962019
 -0.24277    -0.440292     0.0756692      -0.22697     0.0390074   0.0625978
 -0.31007     0.186837    -0.291535        0.0963144  -0.150286    0.123312
 -0.0694515   0.0152947    0.0100862       0.0305994   0.020179   -0.179416
 -0.0297656   0.0437708    0.0933654   …   0.0875701   0.186791    0.0820625
 -0.0533596  -0.096774     0.0166317      -0.193611    0.0332743   0.0533974
 -0.0681519   0.0410659   -0.0640781       0.0821585  -0.128198    0.105188
 -0.0873146   0.0192286    0.0126804       0.0435119   0.0286942  -0.255126
 -0.0374214   0.0550288    0.117379        0.124523    0.265614    0.116692
  ⋮                                    ⋱
  0.0520843  -0.0313841    0.0489709       0.115082   -0.179571    0.147341
  0.0259621  -0.00571741  -0.00377039      0.0447465   0.0295084  -0.262365
  0.0111269  -0.0163622   -0.0349014       0.128057    0.273151    0.120003
  0.0199467   0.0361757   -0.0062172      -0.283124    0.0486581   0.0780848
  0.0254762  -0.0153511    0.0239534   …   0.120143   -0.187468    0.15382
 -0.101863    0.0224324    0.0147932       0.042623    0.028108   -0.249914
 -0.0436565   0.0641976    0.136937        0.12198     0.260188    0.114308
 -0.0782612  -0.141936     0.0243933      -0.269688    0.0463489   0.0743792
 -0.0999567   0.0602303   -0.0939817       0.114442   -0.178571    0.146521

julia> logdet(E)
-20.227705379311118

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  3.0885913701087286
 -1.1936560101667528
  2.418777777282242
  2.1954118317777485
  2.6035946720416896
 -2.437176949215827
  3.450417636579961
  4.061655471842489
 -1.1338151413478066
 -0.6071342575859598
  ⋮
  2.2044687267244547
 -3.3445558101086172
 -1.5606781249027897
 -0.5932756349383158
 -4.529137514103479
  2.9433757801191
  0.7605291982168432
  3.600820293908884
  2.167021532189856

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.80611   2.48454    2.7097   -1.31484   …  1.7813     1.94274  -0.942684
  2.48454   1.19042    1.6691    0.476722     0.853478   1.19667   0.341789
  2.7097    1.6691    10.3533    3.75258      1.19667    7.42287   2.69044
 -1.31484   0.476722   3.75258   4.18799      0.341789   2.69044   3.00261
  6.16462   2.25036    2.4543   -1.19091      4.45773    4.86172  -2.35907
  2.25036   1.07822    1.51178   0.43179   …  2.13584    2.99468   0.85533
  2.4543    1.51178    9.37748   3.39889      2.99468   18.5758    6.73284
 -1.19091   0.43179    3.39889   3.79326      0.85533    6.73284   7.51405
  7.09182   2.58883    2.82345  -1.37004      3.0468     3.32292  -1.6124
  2.58883   1.24039    1.73917   0.496735     1.45982    2.04683   0.584607
  ⋮                                        ⋱
 -1.53071   0.554991   4.36869   4.87558      0.609563   4.79826   5.35499
  7.16015   2.61378    2.85065  -1.38324      4.44184    4.84438  -2.35066
  2.61378   1.25234    1.75592   0.50152      2.12822    2.98401   0.852281
  2.85065   1.75592   10.8919    3.94778      2.98401   18.5096    6.70884
 -1.38324   0.50152    3.94778   4.40584   …  0.852281   6.70884   7.48726
  4.87968   1.7813     1.94274  -0.942684     5.01227    5.46652  -2.65254
  1.7813    0.853478   1.19667   0.341789     2.40154    3.36722   0.961734
  1.94274   1.19667    7.42287   2.69044      3.36722   20.8866    7.57042
 -0.942684  0.341789   2.69044   3.00261      0.961734   7.57042   8.4488

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.60885  0.952349  1.03866  -0.503992  …   0.682792    0.744671   -0.36134
 0.0      0.5324    1.27712   1.79695       0.381708    0.915638    1.28834
 0.0      0.0       2.76468   0.716588      0.0         1.98216     0.513762
 0.0      0.0       0.0       0.437546      0.0         0.0         0.313701
 0.0      0.0       0.0      -0.0           0.850431    0.927502   -0.450056
 0.0      0.0       0.0       0.0       …   0.475425    1.14044     1.60465
 0.0      0.0       0.0       0.0           0.0         2.46881     0.639901
 0.0      0.0       0.0       0.0           0.0         0.0         0.390721
 0.0      0.0       0.0      -0.0           0.206958    0.225714   -0.109524
 0.0      0.0       0.0       0.0           0.115698    0.277534    0.390501
 ⋮                                      ⋱
 0.0      0.0       0.0       0.0          -0.0        -0.0        -0.187293
 0.0      0.0       0.0      -0.0          -0.0455769  -0.0497074   0.0241198
 0.0      0.0       0.0       0.0          -0.0254793  -0.0611196  -0.0859975
 0.0      0.0       0.0       0.0          -0.0        -0.132311   -0.034294
 0.0      0.0       0.0       0.0       …  -0.0        -0.0        -0.0209398
 0.0      0.0       0.0      -0.0           0.281253    0.306742   -0.148842
 0.0      0.0       0.0       0.0           0.157232    0.377166    0.530686
 0.0      0.0       0.0       0.0           0.0         0.816482    0.211627
 0.0      0.0       0.0       0.0           0.0         0.0         0.129219

julia> logdet(C)
-48.84061481257473

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  165.521   -351.378   -29.6283   …  -166.798   -14.0645     56.2571
 -351.378    764.969    59.5142       363.128    28.2512   -119.016
  -29.6283    59.5142    8.39806       28.2512    3.98653   -11.2035
  118.512   -250.721   -23.6015      -119.016   -11.2035     43.2246
  103.198   -219.075   -18.4725      -259.335   -21.8672     87.4677
 -219.075    476.939    37.1056   …   564.586    43.9245   -185.045
  -18.4725    37.1056    5.23598       43.9245    6.1982    -17.4191
   73.8891  -156.318   -14.7149      -185.045   -17.4191     67.2049
 -101.626    215.738    18.1911       -58.7046   -4.94999    19.7997
  215.738   -469.674   -36.5404       127.803     9.94302   -41.8878
    ⋮                             ⋱
   49.2382  -104.167    -9.80572     -132.704   -12.492      48.1957
  169.487   -359.798   -30.3383      -104.774    -8.8346     35.3379
 -359.798    783.3      60.9403       228.099    17.746     -74.7601
  -30.3383    60.9403    8.5993        17.746     2.50414    -7.0375
  121.352   -256.729   -24.167    …   -74.7601   -7.0375     27.1515
   78.5721  -166.798   -14.0645      -231.105   -19.4869     77.9465
 -166.798    363.128    28.2512       503.129    39.1432   -164.902
  -14.0645    28.2512    3.98653       39.1432    5.5235    -15.523
   56.2571  -119.016   -11.2035      -164.902   -15.523      59.8894

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source