Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.19154   -1.93838   -3.67453  -1.71429   …  -1.67257   -3.17063  -1.47921
 -1.93838   12.9694     4.33989  -0.5738       11.1909     3.74475  -0.495114
 -3.67453    4.33989   39.5566    4.97363       3.74475   34.1322    4.29159
 -1.71429   -0.5738     4.97363   1.50981      -0.495114   4.29159   1.30277
  2.63884   -0.985272  -1.86775  -0.871365     -1.00153   -1.89856  -0.885742
 -0.985272   6.59229    2.20594  -0.29166   …   6.70106    2.24234  -0.296472
 -1.86775    2.20594   20.1065    2.52808       2.24234   20.4382    2.56979
 -0.871365  -0.29166    2.52808   0.76743      -0.296472   2.56979   0.780092
  3.87448   -1.44663   -2.74233  -1.27939      -1.44109   -2.73183  -1.27449
 -1.44663    9.67916    3.23889  -0.428231      9.6421     3.22649  -0.426591
  ⋮                                         ⋱
 -1.24638   -0.417185   3.61611   1.09772      -0.493444   4.27711   1.29837
  2.07162   -0.773489  -1.46628  -0.684066     -1.01927   -1.93219  -0.901431
 -0.773489   5.17529    1.73178  -0.228968      6.81975    2.28206  -0.301723
 -1.46628    1.73178   15.7846    1.98467       2.28206   20.8002    2.6153
 -0.684066  -0.228968   1.98467   0.602472  …  -0.301723   2.6153    0.793909
  4.47962   -1.67257   -3.17063  -1.47921      -2.24301   -4.25199  -1.9837
 -1.67257   11.1909     3.74475  -0.495114     15.0076     5.02191  -0.663975
 -3.17063    3.74475   34.1322    4.29159       5.02191   45.7731    5.75526
 -1.47921   -0.495114   4.29159   1.30277      -0.663975   5.75526   1.74708
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0006531994213758364
   0.009881406381512817
   0.025081119960225422
   0.08209595672978777
   0.0032298094656942125
   0.04885959604519931
   0.1240160906559321
   0.40593161822249557
   0.01197417249517143
   0.18114171665660825
   ⋮
  17.016330616790093
   0.23092814318945215
   3.4934122001773886
   8.867026319967009
  29.023704293919348
   2.341192354320703
  35.4168609359846
  89.89555772263017
 294.24769821685646
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.017734    -0.0548942    0.00893123  …   0.317935   -0.0517277  -0.0385346
  0.00733523  -0.0080071   -0.0570243       0.0463753   0.330272    0.0525176
 -0.00614763  -0.00739882   0.00934398      0.0428523  -0.0541182   0.331301
  0.0551889    0.0178794    0.00575012     -0.103554   -0.0333034   0.0423067
 -0.125621     0.38885     -0.0632655       0.210944   -0.0343203  -0.025567
 -0.05196      0.0567192    0.403938    …   0.0307691   0.219129    0.0348444
  0.0435475    0.0524104   -0.0661892       0.0284317  -0.0359064   0.219812
 -0.390937    -0.126651    -0.0407317      -0.0687059  -0.0220962   0.0280697
  0.102976    -0.318754     0.0518609       0.302551   -0.0492247  -0.03667
  0.0425934   -0.0464947   -0.331122        0.0441313   0.314291    0.0499763
  ⋮                                     ⋱
 -0.109934    -0.0356152   -0.0114541      -0.0988078  -0.0317771   0.0403678
  0.036915    -0.114267     0.0185912       0.186195   -0.0302936  -0.0225673
  0.0152689   -0.0166675   -0.118701        0.0271591   0.193419    0.0307562
 -0.0127968   -0.0154013    0.0194503       0.0250959  -0.0316936   0.194022
  0.114881     0.0372176    0.0119694   …  -0.0606448  -0.0195037   0.0247764
  0.0910334   -0.281786     0.0458463       0.348133   -0.0566409  -0.0421947
  0.0376536   -0.0411025   -0.29272         0.0507801   0.361642    0.0575058
 -0.0315574   -0.0379801    0.0479651       0.0469225  -0.0592585   0.362768
  0.283299     0.0917797    0.0295169      -0.113389   -0.0364666   0.0463251
julia> logdet(E)-14.79432463427856
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
   7.42486029296709
   3.8028690407810473
  -2.383866227707482
  12.891393124707541
  -0.13361818015889898
   0.06481776514589639
   3.1800639543322498
  -5.961602113365081
  -1.9578612890289855
  -1.324557832003061
   ⋮
 -14.488887873158374
   0.9402591331865003
   2.272609872805296
  -2.183098937562193
   6.275157145665493
   0.028813386823691143
  -1.726566141627423
  -1.001563340358759
  -4.9921191683343125

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 6.11441   2.89379   0.729166   3.16321  …  1.84536   0.464986   2.01716
 2.89379   7.5644    4.90497    7.82286     4.82379   3.12788    4.98861
 0.729166  4.90497  13.9861     9.68801     3.12788   8.91889    6.17801
 3.16321   7.82286   9.68801   16.5443      4.98861   6.17801   10.5502
 4.18751   1.98184   0.499375   2.16635     1.96586   0.49535    2.14889
 1.98184   5.18055   3.35921    5.35755  …  5.13879   3.33213    5.31437
 0.499375  3.35921   9.57851    6.63492     3.33213   9.50131    6.58144
 2.16635   5.35755   6.63492   11.3305      5.31437   6.58144   11.2392
 2.7545    1.30363   0.328484   1.425       1.73391   0.436903   1.89534
 1.30363   3.40771   2.20965    3.52415     4.53245   2.93897    4.68732
 ⋮                                       ⋱
 2.80543   6.93806   8.59225   14.673       5.6234    6.96414   11.8927
 5.61006   2.65509   0.669019   2.90228     2.07383   0.522555   2.2669
 2.65509   6.94044   4.50037    7.17758     5.42101   3.51513    5.60623
 0.669019  4.50037  12.8324     8.88888     3.51513  10.0231     6.94289
 2.90228   7.17758   8.88888   15.1796   …  5.60623   6.94289   11.8564
 3.89914   1.84536   0.464986   2.01716     2.335     0.588365   2.5524
 1.84536   4.82379   3.12788    4.98861     6.10373   3.95783    6.31228
 0.464986  3.12788   8.91889    6.17801     3.95783  11.2854     7.81728
 2.01716   4.98861   6.17801   10.5502      6.31228   7.81728   13.3496
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.47273  1.17028  0.294882  1.27923  …   0.746282    0.188045    0.815763
 0.0      2.48895  1.83205   2.54156      1.58719     1.16829     1.62074
 0.0      0.0      3.24696   1.4335       0.0         2.07057     0.914139
 0.0      0.0      0.0       2.52852      0.0         0.0         1.61243
 0.0      0.0      0.0       0.0          0.412942    0.104051    0.451388
 0.0      0.0      0.0       0.0      …   0.878243    0.646453    0.896808
 0.0      0.0      0.0       0.0          0.0         1.14571     0.505822
 0.0      0.0      0.0       0.0          0.0         0.0         0.892207
 0.0      0.0      0.0       0.0          0.15057     0.0379401   0.164589
 0.0      0.0      0.0       0.0          0.320232    0.235715    0.327001
 ⋮                                    ⋱
 0.0      0.0      0.0       0.0         -0.0        -0.0        -0.793896
 0.0      0.0      0.0       0.0         -0.0904724  -0.0227969  -0.0988956
 0.0      0.0      0.0       0.0         -0.192416   -0.141633   -0.196484
 0.0      0.0      0.0       0.0         -0.0        -0.251017   -0.110822
 0.0      0.0      0.0       0.0      …  -0.0        -0.0        -0.195476
 0.0      0.0      0.0       0.0          0.187504    0.0472466   0.204961
 0.0      0.0      0.0       0.0          0.398784    0.293535    0.407213
 0.0      0.0      0.0       0.0          0.0         0.520234    0.229679
 0.0      0.0      0.0       0.0          0.0         0.0         0.405124
julia> logdet(C)-0.31552961215606246
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  11.866      -4.44074    1.77629   …    0.829146  -0.331657    0.225758
  -4.44074    16.5015    -1.24323       -3.08104    0.232127    1.16239
   1.77629    -1.24323    7.16606        0.232127  -1.338       0.737158
  -1.20912    -6.22555   -3.94807        1.16239    0.737158   -1.6697
  -8.77862     3.28532   -1.31412        5.91579   -2.36631     1.61074
   3.28532   -12.208      0.919754  …  -21.9826     1.65618     8.29345
  -1.31412     0.919754  -5.30154        1.65618   -9.54636     5.25948
   0.89452     4.60574    2.92083        8.29345    5.25948   -11.913
   3.23829    -1.2119     0.484758      -0.373136   0.149254   -0.101597
  -1.2119      4.50333   -0.339283       1.38655   -0.104463   -0.523106
   ⋮                                ⋱
   0.462987    2.38384    1.51177       -5.03346   -3.19208     7.23024
 -14.0377      5.25349   -2.10139       -2.01618    0.806468   -0.548962
   5.25349   -19.5216     1.47076        7.49198   -0.564448   -2.82652
  -2.10139     1.47076   -8.4776        -0.564448   3.25353    -1.7925
   1.43041     7.36496    4.67065   …   -2.82652   -1.7925      4.06011
  -2.21554     0.829146  -0.331657      -3.02563    1.21025    -0.823814
   0.829146   -3.08104    0.232127      11.243     -0.847054   -4.24169
  -0.331657    0.232127  -1.338         -0.847054   4.88249    -2.68996
   0.225758    1.16239    0.737158      -4.24169   -2.68996     6.0929

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.