Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.3022     -1.23499  -0.0674905  …  -1.35513  -0.0740563   0.260166
 -1.23499    16.2154   -1.12885       17.7929   -1.23867     2.44613
 -0.0674905  -1.12885   0.572693      -1.23867   0.628407   -0.520251
  0.2371      2.22926  -0.474126       2.44613  -0.520251    0.945229
  7.72951    -1.51469  -0.0827757     -2.29082  -0.12519     0.439804
 -1.51469    19.8878   -1.38451    …  30.0785   -2.09393     4.13512
 -0.0827757  -1.38451   0.702395      -2.09393   1.06231    -0.879472
  0.290798    2.73414  -0.581505       4.13512  -0.879472    1.59789
  5.20244    -1.01948  -0.0557131     -1.58107  -0.0864033   0.303542
 -1.01948    13.3857   -0.931859      20.7594   -1.44518     2.85396
  ⋮                                ⋱
  0.214287    2.01477  -0.428509       2.7606   -0.587133    1.06674
  6.27099    -1.22887  -0.0671563     -1.86018  -0.101657    0.357128
 -1.22887    16.1351   -1.12326       24.4242   -1.70031     3.35779
 -0.0671563  -1.12326   0.569856      -1.70031   0.862611   -0.714146
  0.235925    2.21822  -0.471778   …   3.35779  -0.714146    1.29751
  6.91531    -1.35513  -0.0740563     -2.33416  -0.127559    0.448125
 -1.35513    17.7929   -1.23867       30.6476   -2.13355     4.21336
 -0.0740563  -1.23867   0.628407      -2.13355   1.08241    -0.896112
  0.260166    2.44613  -0.520251       4.21336  -0.896112    1.62812
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0020006755434103453
   0.007985120844003072
   0.06228615835336137
   0.1682080345946528
   0.005809285185728513
   0.023186090507332557
   0.1808579398043395
   0.48841924754353094
   0.013148977202887798
   0.052480359589503665
   ⋮
  12.17919053520483
   0.18440142358398195
   0.7359852305758335
   5.74088902509337
  15.503663819793996
   2.7657273625004946
  11.03860507710655
  86.10418267486057
 232.53023979982913
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0121346   -0.0321947   0.428459   …   0.0176991  -0.235547     0.0267221
 -0.0201309   -0.0685426   0.0424403      0.0376815  -0.0233317   -0.233382
  0.313804    -0.295626   -0.0168616      0.162521    0.00926974   0.017133
  0.296812     0.306585    0.0382221     -0.168546   -0.0210127   -0.0328502
 -0.00388759  -0.0103143   0.137266       0.0285431  -0.379862     0.0430942
 -0.00644939  -0.0219591   0.0135967  …   0.0607684  -0.0376266   -0.376372
  0.100534    -0.0947102  -0.005402       0.262095    0.0149492    0.0276301
  0.0950904    0.0982212   0.0122453     -0.271811   -0.0338869   -0.052977
  0.00659713   0.017503   -0.232937       0.0191092  -0.254312     0.028851
  0.0109444    0.037264   -0.0230732      0.0406836  -0.0251905   -0.251976
  ⋮                                   ⋱
 -0.1634      -0.16878    -0.0210419     -0.174532   -0.021759    -0.0340169
 -0.0105199   -0.0279106   0.371445       0.0245629  -0.326892     0.037085
 -0.0174522   -0.0594218   0.0367929      0.0522946  -0.0323798   -0.323889
  0.272047    -0.256287   -0.0146179      0.225547    0.0128646    0.0237772
  0.257316     0.265788    0.033136   …  -0.233909   -0.0291616   -0.0455896
 -0.0134551   -0.0356981   0.475084       0.0260919  -0.347241     0.0393935
 -0.0223216   -0.0760014   0.0470586      0.0555499  -0.0343954   -0.344051
  0.347953    -0.327796   -0.0186965      0.239588    0.0136654    0.0252573
  0.329112     0.339947    0.0423815     -0.248469   -0.0309769   -0.0484276
julia> logdet(E)-27.524054232281504
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -1.1717721969739654
  0.3793788841134772
  6.230329763814805
 -0.40183532841022596
  1.9226147096946031
 -1.7721939251309962
  7.071626336695405
 12.205808333417774
  0.4863458099497463
 -1.1395626330941213
  ⋮
 -0.15520047394704956
  1.2998022783912804
  0.8594573091707597
  8.286383664184829
 -2.0298375326133433
  0.8198383985635341
  0.6462767928312849
  7.978546504932028
 -4.690587854333466

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 1.87447    1.14643    3.97844   …   0.922118   3.20002    1.46566
 1.14643   13.2874     1.07949      10.6876     0.868274  -1.76613
 3.97844    1.07949   12.8159        0.868274  10.3083     5.50847
 1.82219   -2.19575    6.84843      -1.76613    5.50847    8.7798
 1.5887     0.971651   3.37192       0.957376   3.32237    1.5217
 0.971651  11.2617     0.914915  …  11.0962     0.901473  -1.83366
 3.37192    0.914915  10.8621        0.901473  10.7025     5.71909
 1.54439   -1.861      5.80436      -1.83366    5.71909    9.1155
 0.895769   0.547853   1.90121       0.655637   2.27525    1.0421
 0.547853   6.34975    0.515863      7.59899    0.617353  -1.25574
 ⋮                               ⋱
 0.839469  -1.01157    3.15502      -1.07333    3.34765    5.33573
 0.762782   0.466518   1.61896       0.540177   1.87457    0.858584
 0.466518   5.40706    0.439278      6.26079    0.508636  -1.0346
 1.61896    0.439278   5.2152        0.508636   6.03864    3.22686
 0.741506  -0.893521   2.78685   …  -1.0346     3.22686    5.14321
 1.50771    0.922118   3.20002       1.03641    3.59663    1.64731
 0.922118  10.6876     0.868274     12.0122     0.975889  -1.98503
 3.20002    0.868274  10.3083        0.975889  11.586      6.19119
 1.46566   -1.76613    5.50847      -1.98503    6.19119    9.86797
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.36911  0.837351   2.90585   …   0.673515    2.33729      1.07052
 0.0      3.54771   -0.381579      2.85356    -0.306919    -0.750492
 0.0      0.0        2.0558        0.0         1.65356      1.02701
 0.0      0.0        0.0           0.0         0.0          2.07315
 0.0      0.0        0.0           0.183687    0.637446     0.29196
 0.0      0.0       -0.0       …   0.778248   -0.0837057   -0.204681
 0.0      0.0        0.0           0.0         0.450974     0.280095
 0.0      0.0        0.0           0.0         0.0          0.565407
 0.0      0.0        0.0           0.110684    0.384105     0.175926
 0.0      0.0       -0.0           0.468947   -0.0504384   -0.123334
 ⋮                             ⋱
 0.0      0.0        0.0          -0.0        -0.0         -0.118965
 0.0      0.0        0.0           0.0183108   0.0635438    0.029104
 0.0      0.0       -0.0           0.0775795  -0.00834419  -0.0204036
 0.0      0.0        0.0           0.0         0.0449552    0.0279212
 0.0      0.0        0.0       …   0.0         0.0          0.0563625
 0.0      0.0        0.0           0.105976    0.367767     0.168443
 0.0      0.0       -0.0           0.449001   -0.048293    -0.118088
 0.0      0.0        0.0           0.0         0.260184     0.161598
 0.0      0.0        0.0           0.0         0.0          0.326205
julia> logdet(C)-12.675969735945792
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  17.3918     -0.996259    -5.50068     …  -0.783773    -4.32747     0.273327
  -0.996259    0.89807      0.0725348       0.706526     0.0570643   0.237161
  -5.50068     0.0725348    3.00786         0.0570643    2.36633    -0.750756
   0.347427    0.301457    -0.954291        0.237161    -0.750756    1.20877
  -1.96192     0.112385     0.620515        0.792046     4.37315    -0.276212
   0.112385   -0.101309    -0.00818244  …  -0.713984    -0.0576667  -0.239664
   0.620515   -0.00818244  -0.339308       -0.0576667   -2.39131     0.758681
  -0.0391923  -0.0340064    0.107651       -0.239664     0.758681   -1.22153
 -10.7848      0.61779      3.41103         2.0159      11.1305     -0.70301
   0.61779    -0.556902    -0.0449796      -1.81722     -0.146772   -0.609989
   ⋮                                    ⋱
   0.237329    0.205926    -0.651879        0.774837    -2.45283     3.94923
   1.31231    -0.0751732   -0.415056        0.552261     3.04922    -0.192591
  -0.0751732   0.0677643    0.00547315     -0.497832    -0.0402086  -0.167108
  -0.415056    0.00547315   0.22696        -0.0402086   -1.66736     0.528997
   0.0262153   0.0227466   -0.0720065   …  -0.167108     0.528997   -0.851724
  13.6824     -0.783773    -4.32747        -6.09347    -33.6441      2.12499
  -0.783773    0.706526     0.0570643       5.49291      0.443649    1.84382
  -4.32747     0.0570643    2.36633         0.443649    18.3971     -5.83678
   0.273327    0.237161    -0.750756        1.84382     -5.83678     9.39765

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.