Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 16.0366   -6.66474  11.2383   11.8744   …   -5.09866   8.59755    9.08415
 -6.66474  16.7909   -5.78575  -9.05334      12.8454   -4.42622   -6.92599
 11.2383   -5.78575  10.7101    8.39631      -4.42622   8.19347    6.42335
 11.8744   -9.05334   8.39631  11.2367       -6.92599   6.42335    8.5963
 10.0479   -4.17588   7.04152   7.44006      -4.14839   6.99516    7.39108
 -4.17588  10.5206   -3.62514  -5.67249  …   10.4513   -3.60127   -5.63514
  7.04152  -3.62514   6.71058   5.26082      -3.60127   6.6664     5.22618
  7.44006  -5.67249   5.26082   7.0405       -5.63514   5.22618    6.99414
  9.20321  -3.82481   6.44954   6.81457      -3.24172   5.46632    5.7757
 -3.82481   9.6361   -3.32037  -5.1956        8.16709  -2.81419   -4.40354
  ⋮                                      ⋱
  8.98522  -6.85056   6.35339   8.50267      -6.77244   6.28094    8.40572
 10.5095   -4.36769   7.36496   7.7818       -5.46525   9.2157     9.73729
 -4.36769  11.0038   -3.79165  -5.93304      13.769    -4.74446   -7.42396
  7.36496  -3.79165   7.01882   5.50246      -4.74446   8.78257    6.88518
  7.7818   -5.93304   5.50246   7.36389  …   -7.42396   6.88518    9.21436
 12.2683   -5.09866   8.59755   9.08415      -8.12404  13.6991    14.4744
 -5.09866  12.8454   -4.42622  -6.92599      20.4674   -7.05259  -11.0357
  8.59755  -4.42622   8.19347   6.42335      -7.05259  13.0552    10.2348
  9.08415  -6.92599   6.42335   8.5963      -11.0357   10.2348    13.6971

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0006727201450049931
   0.0022550944715180564
   0.010699823030890131
   0.038867891956459635
   0.004600955297814488
   0.015423336037790504
   0.07317962428386328
   0.26583035269536454
   0.024335928301640677
   0.08157897125539829
   ⋮
  22.262955834555346
   0.5613519280246607
   1.88176560322126
   8.928476919430238
  32.4333472841829
   4.863605959004756
  16.303794365653335
  77.3571646987728
 281.0055746613431
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0386946    0.0183315   …  -0.0975004  -0.143966     0.189513
  0.0139583    0.00752097     -0.0400021  -0.279604    -0.152361
  0.0146847   -0.050112        0.266532   -0.0978358    0.147622
  0.0439068    0.0305244      -0.162351   -0.00526604   0.16608
 -0.337148     0.159723       -0.0870588  -0.128548     0.169217
  0.121619     0.0655305   …  -0.0357182  -0.24966     -0.136044
  0.127948    -0.436627        0.237989   -0.0873583    0.131813
  0.382561     0.26596        -0.144965   -0.00470208   0.148294
  0.0116546   -0.00552133     -0.0763937  -0.1128       0.148487
 -0.00420416  -0.00226527     -0.0313425  -0.219076    -0.119378
  ⋮                        ⋱
  0.0323799    0.0225108      -0.157356   -0.00510399   0.16097
 -0.180325     0.0854288      -0.0887081  -0.130983     0.172423
  0.0650489    0.0350494      -0.0363948  -0.25439     -0.138621
  0.0684339   -0.233533        0.242497   -0.0890132    0.13431
  0.204615     0.14225     …  -0.147711   -0.00479116   0.151103
 -0.0801474    0.0379696      -0.104334   -0.154056     0.202796
  0.0289116    0.015578       -0.0428058  -0.299201    -0.16304
  0.0304161   -0.103796        0.285214   -0.104693     0.157969
  0.0909432    0.0632246      -0.173731   -0.00563513   0.177721

julia> logdet(E)
-15.303780686671274

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  1.322683264733594
 -0.8428759998941495
  1.0368940326538858
 -4.48302133739816
 -8.348958440669945
 12.154081201309456
  8.236116430161474
 -0.6547511494283444
  3.200477118629892
  0.2203773782637392
  ⋮
  1.5701288240536098
 -4.394475515572829
  8.473553430214162
  2.3139293277899413
  4.869155508737336
  0.3061783439750265
  1.2418815147023679
 -0.3837605919817209
 -1.8258081521284224

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  9.97666   2.26461  -9.45981  -9.40768  …   1.6493    -6.88952   -6.85155
  2.26461   4.5751   -3.01655  -3.02439      3.33202   -2.19693   -2.20264
 -9.45981  -3.01655   9.41984   8.99086     -2.19693    6.86041    6.54798
 -9.40768  -3.02439   8.99086   9.46992     -2.20264    6.54798    6.89688
  8.28901   1.88153  -7.85959  -7.81627      1.56899   -6.55403   -6.51791
  1.88153   3.80118  -2.50627  -2.51278  …   3.16976   -2.08995   -2.09538
 -7.85959  -2.50627   7.82638   7.46997     -2.08995    6.52634    6.22913
 -7.81627  -2.51278   7.46997   7.86799     -2.09538    6.22913    6.56103
  6.45164   1.46446  -6.1174   -6.08369      1.6501    -6.89287   -6.85488
  1.46446   2.95859  -1.95072  -1.95579      3.33364   -2.198     -2.20372
  ⋮                                      ⋱
 -6.66699  -2.14331   6.3716    6.7111      -1.82823    5.43495    5.72454
  4.36423   0.99064  -4.13814  -4.11533      2.04543   -8.54425   -8.49716
  0.99064   2.00135  -1.31957  -1.323        4.1323    -2.72459   -2.73167
 -4.13814  -1.31957   4.12065   3.933       -2.72459    8.50814    8.12068
 -4.11533  -1.323     3.933     4.14256  …  -2.73167    8.12068    8.55338
  7.26593   1.6493   -6.88952  -6.85155      2.89621  -12.0982   -12.0315
  1.6493    3.33202  -2.19693  -2.20264      5.8511    -3.85787   -3.86789
 -6.88952  -2.19693   6.86041   6.54798     -3.85787   12.047     11.4984
 -6.85155  -2.20264   6.54798   6.89688     -3.86789   11.4984    12.1111

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.15859  0.71697  -2.99495   …   0.522164   -2.1812     -2.16918
 0.0      2.01521  -0.43135       1.46766    -0.31415    -0.32126
 0.0      0.0       0.513844      0.0         0.374229   -0.169681
 0.0      0.0       0.0           0.0         0.0         0.430812
 0.0      0.0      -0.0           0.103485   -0.432281   -0.429898
 0.0      0.0      -0.0       …   0.290868   -0.0622595  -0.0636687
 0.0      0.0       0.0           0.0         0.0741664  -0.0336281
 0.0      0.0       0.0           0.0         0.0         0.0853803
 0.0      0.0      -0.0           0.229024   -0.95669    -0.951417
 0.0      0.0      -0.0           0.643725   -0.137788   -0.140907
 ⋮                            ⋱
 0.0      0.0       0.0          -0.0        -0.0        -0.111178
 0.0      0.0      -0.0           0.110891   -0.463217   -0.460664
 0.0      0.0      -0.0           0.311684   -0.0667152  -0.0682252
 0.0      0.0       0.0           0.0         0.0794742  -0.0360348
 0.0      0.0       0.0       …   0.0         0.0         0.0914906
 0.0      0.0      -0.0           0.0767471  -0.320591   -0.318824
 0.0      0.0      -0.0           0.215715   -0.0461734  -0.0472184
 0.0      0.0       0.0           0.0         0.0550038  -0.0249395
 0.0      0.0       0.0           0.0         0.0         0.0633203

julia> logdet(C)
-55.75043263253008

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  671.143    154.736     426.402    …    11.2221     30.9245     22.5781
  154.736     59.0258    102.145          4.2808      7.40797     5.48228
  426.402    102.145     366.267          7.40797    26.5632      7.86766
  311.317     75.5923    108.483          5.48228     7.86766    17.352
 -731.302   -168.606    -464.624        -11.0753    -30.5201    -22.2828
 -168.606    -64.3167   -111.301    …    -4.22482    -7.31108    -5.41058
 -464.624   -111.301    -399.098         -7.31108   -26.2158     -7.76476
 -339.223    -82.3682   -118.207         -5.41058    -7.76476   -17.1251
  532.698    122.817     338.443         -4.23222   -11.6626     -8.51491
  122.817     46.8498     81.074         -1.61443    -2.79378    -2.06754
    ⋮                               ⋱
 -253.241    -61.4905    -88.2455       -23.2924    -33.4271    -73.7231
 -361.871    -83.4316   -229.911       -127.915    -352.492    -257.355
  -83.4316   -31.8259    -55.0751       -48.7946    -84.4395    -62.4896
 -229.911    -55.0751   -197.486        -84.4395   -302.78      -89.6792
 -167.858    -40.7584    -58.4927   …   -62.4896    -89.6792   -197.786
   48.6742    11.2221     30.9245       161.301     444.495     324.526
   11.2221     4.2808      7.40797       61.5303    106.479      78.7997
   30.9245     7.40797    26.5632       106.479     381.807     113.086
   22.5781     5.48228     7.86766       78.7997    113.086     249.41

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.