Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3.19053    -0.134204   -0.77847    …  -0.153667   -0.891372    0.579663
 -0.134204    0.28176     0.129522       0.322623    0.148306   -1.54308
 -0.77847     0.129522    4.72415        0.148306    5.40929     2.84512
  0.506243   -1.34763     2.48476       -1.54308     2.84512    10.0092
  1.93944    -0.0815788  -0.473211      -0.0978925  -0.567842    0.36927
 -0.0815788   0.171274    0.0787329  …   0.205525    0.0944775  -0.983007
 -0.473211    0.0787329   2.87169        0.0944775   3.44595     1.81246
  0.307731   -0.81919     1.51042       -0.983007    1.81246     6.37626
  3.31606    -0.139484   -0.809098      -0.138269   -0.802054    0.521579
 -0.139484    0.292845    0.134618       0.290295    0.133446   -1.38846
  ⋮                                  ⋱
  0.475047   -1.26459     2.33164       -1.77177     3.26678    11.4925
  3.30162    -0.138876   -0.805574      -0.161841   -0.938784    0.610495
 -0.138876    0.29157     0.134032       0.339784    0.156195   -1.62516
 -0.805574    0.134032    4.88863        0.156195    5.69701     2.99645
  0.523868   -1.39455     2.57127    …  -1.62516     2.99645    10.5415
  3.65325    -0.153667   -0.891372      -0.230601   -1.33764     0.869872
 -0.153667    0.322623    0.148306       0.484145    0.222556   -2.31562
 -0.891372    0.148306    5.40929        0.222556    8.11746     4.26953
  0.579663   -1.54308     2.84512       -2.31562     4.26953    15.0203

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   4.502183803030306e-5
   0.01606283858820627
   0.028461277993907966
   0.06534586222886932
   0.00028998158926967085
   0.10345929144108061
   0.18331651896940082
   0.42088679206233925
   0.0005243061105194624
   0.1870613194071693
   ⋮
   7.979354401487306
   0.007226007302659037
   2.578086413568633
   4.568036571242659
  10.488014224264726
   0.07819703824564075
  27.8990476259564
  49.43351362757443
 113.49720738780042
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.000753014   0.0291019   …  -0.209177   -0.16145    -0.00580362
 -0.0358002     0.00502647     -0.0361291   0.0370512   0.0314805
  0.00475177    0.0199116      -0.14312     0.190502   -0.109285
 -0.006919     -0.00916586      0.0658821  -0.0784492  -0.238572
  0.00428258    0.16551        -0.19177    -0.148015   -0.00532065
 -0.203604      0.0285868   …  -0.0331225   0.0339679   0.0288607
  0.0270245     0.113242       -0.13121     0.174649   -0.100191
 -0.03935      -0.0521285       0.0603995  -0.0719207  -0.218718
  0.00681158    0.263248       -0.263974   -0.203744   -0.00732394
 -0.32384       0.0454682      -0.0455935   0.0467571   0.0397271
  ⋮                         ⋱
 -0.0704319    -0.0933039       0.0932094  -0.110989   -0.337529
 -0.0054556    -0.210844       -0.269716   -0.208176   -0.00748327
  0.259373     -0.0364168      -0.0465854   0.0477744   0.0405914
 -0.0344267    -0.14426        -0.184541    0.245636   -0.140914
  0.0501283     0.0664068   …   0.0849493  -0.101154   -0.307618
  0.0029146     0.112641       -0.260287   -0.200898   -0.00722164
 -0.138567      0.0194554      -0.0449567   0.0461041   0.0391722
  0.0183921     0.0770697      -0.17809     0.237049   -0.135988
 -0.0267805    -0.0354772       0.0819793  -0.097617   -0.296863

julia> logdet(E)
-57.54611533342943

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  -0.25352410823103766
   4.683814180565557
  -1.4798828052395478
  -0.6227581888877363
  -0.31197462494206935
  -6.258273097609101
  -3.2453593861673555
   1.528381169373044
   2.0227346765421146
  -6.611233845102179
   ⋮
  -0.6469838812657756
  -2.758847208538396
   8.191976863248302
   2.6749315018798803
  -0.2871297199575841
   1.0919365678151243
 -12.287857162218284
   0.21913025420045512
  -0.43646162828529805

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.28403  11.5674    4.74692  -5.44684  …   9.81376   4.02729  -4.6211
 11.5674   28.8565   14.8024   -5.89001     24.4818   12.5583   -4.99709
  4.74692  14.8024   22.3253   11.3327      12.5583   18.9408    9.61467
 -5.44684  -5.89001  11.3327   25.3214      -4.99709   9.61467  21.4827
  5.52986  12.1055    4.96776  -5.70024     12.2716    5.03593  -5.77846
 12.1055   30.1989   15.491    -6.16402  …  30.6133   15.7036   -6.24861
  4.96776  15.491    23.3639   11.8599      15.7036   23.6845   12.0227
 -5.70024  -6.16402  11.8599   26.4994      -6.24861  12.0227   26.863
  3.87443   8.48157   3.4806   -3.9938       8.45076   3.46796  -3.9793
  8.48157  21.1585   10.8536   -4.31875     21.0817   10.8142   -4.30306
  ⋮                                      ⋱
 -4.32485  -4.67673   8.99828  20.1054      -5.30076  10.1989   22.7882
  4.11752   9.01373   3.69898  -4.24439      9.25303   3.79718  -4.35707
  9.01373  22.4861   11.5346   -4.58972     23.083    11.8408   -4.71157
  3.69898  11.5346   17.3967    8.83087     11.8408   17.8586    9.06532
 -4.24439  -4.58972   8.83087  19.7314   …  -4.71157   9.06532  20.2552
  4.48298   9.81376   4.02729  -4.6211      11.731     4.81409  -5.52392
  9.81376  24.4818   12.5583   -4.99709     29.2648   15.0118   -5.97336
  4.02729  12.5583   18.9408    9.61467     15.0118   22.6412   11.4931
 -4.6211   -4.99709   9.61467  21.4827      -5.97336  11.4931   25.6797

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.2987  5.03213  2.06504  -2.36953  2.40564  …   1.75198     -2.01031
 0.0     1.87993  2.34626   3.20956  0.0          1.99057      2.72299
 0.0     0.0      3.54344   2.45395  0.0          3.00626      2.08194
 0.0     0.0      0.0       1.83945  0.0          0.0          1.56059
 0.0     0.0      0.0      -0.0      1.24555      0.659367    -0.756589
 0.0     0.0      0.0       0.0      0.0      …   0.749158     1.02481
 0.0     0.0      0.0       0.0      0.0          1.13142      0.783546
 0.0     0.0      0.0       0.0      0.0          0.0          0.587335
 0.0     0.0      0.0      -0.0      0.0          0.00889183  -0.0102029
 0.0     0.0      0.0       0.0      0.0          0.0101027    0.01382
 ⋮                                            ⋱
 0.0     0.0      0.0       0.0      0.0          0.0          0.0929336
 0.0     0.0      0.0      -0.0      0.0         -0.673015     0.772249
 0.0     0.0      0.0       0.0      0.0         -0.764665    -1.04602
 0.0     0.0      0.0       0.0      0.0         -1.15484     -0.799764
 0.0     0.0      0.0       0.0      0.0      …  -0.0         -0.599492
 0.0     0.0      0.0      -0.0      0.0          0.0657679   -0.0754652
 0.0     0.0      0.0       0.0      0.0          0.0747241    0.102219
 0.0     0.0      0.0       0.0      0.0          0.112852     0.078154
 0.0     0.0      0.0       0.0      0.0          0.0          0.0585831

julia> logdet(C)
-38.98709731645563

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
      1.41117e5  -43485.0    …   -8422.05    -2868.87     5204.07
 -43485.0         16035.5         3105.7       368.108   -1253.98
 -14812.6          1900.63         368.108    1114.34    -1030.22
  26869.8         -6474.62       -1253.98    -1030.22     1487.61
  36866.0        -11360.2        -2195.21     -747.772    1356.44
 -11360.2          4189.17   …     809.502      95.9476   -326.852
  -3869.72          496.529         95.9476    290.454    -268.528
   7019.6         -1691.46        -326.852    -268.528     387.747
 -53954.0         16625.9         3212.71     1094.37    -1985.17
  16625.9         -6130.92       -1184.71     -140.42      478.351
      ⋮                      ⋱
 -56984.3         13731.1         2662.5      2187.4     -3158.54
      2.29261e5  -70646.5       -13698.8     -4666.32     8464.6
 -70646.5         26051.5         5051.53      598.741   -2039.65
 -24064.9          3087.8          598.741    1812.52    -1675.69
  43653.2        -10518.8    …   -2039.65    -1675.69     2419.65
  27331.1         -8422.05       -1649.62     -561.922    1019.31
  -8422.05         3105.7          608.309      72.1009   -245.617
  -2868.87          368.108         72.1009    218.265    -201.788
   5204.07        -1253.98        -245.617    -201.788     291.377

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source