Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.79107   -0.897093  -1.913      2.88698  …  -0.784841  -1.67363    2.52574
 -0.897093   2.33784   -1.0514    -1.66845      2.0453    -0.919837  -1.45968
 -1.913     -1.0514     7.76942   -3.53022     -0.919837   6.79724   -3.08849
  2.88698   -1.66845   -3.53022    6.46125     -1.45968   -3.08849    5.65276
  2.48143   -0.797569  -1.70077    2.5667      -0.993707  -2.11902    3.1979
 -0.797569   2.07847   -0.934754  -1.48335  …   2.58961   -1.16463   -1.84813
 -1.70077   -0.934754   6.90747   -3.13858     -1.16463    8.60616   -3.91041
  2.5667    -1.48335   -3.13858    5.74443     -1.84813   -3.91041    7.1571
  1.86508   -0.599467  -1.27833    1.92918     -0.679628  -1.44927    2.18715
 -0.599467   1.56222   -0.702578  -1.11491      1.77112   -0.796527  -1.264
  ⋮                                         ⋱
  1.68881   -0.975994  -2.06508    3.77965     -1.07901   -2.28305    4.17858
  2.32999   -0.748894  -1.59697    2.41006     -0.846467  -1.80504    2.72406
 -0.748894   1.95163   -0.877706  -1.39282      2.2059    -0.992062  -1.57429
 -1.59697   -0.877706   6.48591   -2.94703     -0.992062   7.33096   -3.331
  2.41006   -1.39282   -2.94703    5.39385  …  -1.57429   -3.331      6.09661
  2.44183   -0.784841  -1.67363    2.52574     -1.07235   -2.28673    3.45099
 -0.784841   2.0453    -0.919837  -1.45968      2.79456   -1.2568    -1.99439
 -1.67363   -0.919837   6.79724   -3.08849     -1.2568     9.28726   -4.21989
  2.52574   -1.45968   -3.08849    5.65276     -1.99439   -4.21989    7.72352
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   9.8867505212289e-5
   0.00013631545961651035
   0.0005995436614440315
   0.0013520368258141673
   0.011372993887957001
   0.015680732367260353
   0.06896711293083983
   0.15552841677617268
   0.07370826615688421
   0.10162667862547478
   ⋮
   4.290517090921771
   0.3206473101032303
   0.44209859809516683
   1.9444413199128165
   4.384928803725251
   7.355547889775576
  10.141602027613219
  44.60486895359447
 100.5888800335596
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0565282  -0.345401    -0.100317   …  -0.264225    -0.0767408  -0.104047
 -0.319152   -0.0219956    0.219075      -0.0168262    0.167588    0.0205925
 -0.145199   -0.0092063   -0.244939      -0.00704263  -0.187373    0.202247
 -0.157874    0.176607    -0.181681       0.135101    -0.138982   -0.190384
  0.0149669   0.0914515    0.0265609     -0.340145    -0.098791   -0.133943
  0.0845015   0.00582375  -0.0580043  …  -0.0216609    0.215742    0.0265094
  0.0384443   0.00243754   0.0648522     -0.00906621  -0.241212    0.260359
  0.0418001  -0.04676      0.0481034      0.17392     -0.178916   -0.245087
  0.0140833   0.0860521    0.0249927     -0.239051    -0.0694293  -0.0941339
  0.0795125   0.0054799   -0.0545797     -0.015223     0.151621    0.0186305
  ⋮                                   ⋱
  0.145499   -0.162764     0.16744        0.0973283   -0.100125   -0.137155
  0.0123722   0.0755972    0.0219563     -0.300679    -0.0873284  -0.118402
  0.0698521   0.00481412  -0.0479485     -0.0191476    0.190709    0.0234335
  0.0317794   0.00201496   0.0536092     -0.00801427  -0.213224    0.23015
  0.0345535  -0.0386535    0.0397641  …   0.15374     -0.158157   -0.21665
  0.0241517   0.147573     0.0428606     -0.307332    -0.0892608  -0.121022
  0.136358    0.00939761  -0.0935999     -0.0195713    0.19493     0.023952
  0.0620364   0.00393339   0.10465       -0.00819161  -0.217942    0.235243
  0.0674516  -0.0754552    0.077623       0.157142    -0.161656   -0.221444
julia> logdet(E)-46.44318927664023
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  0.9991448960698565
 -2.4535252156848344
 -1.1165686620348023
 -2.6528299800068984
  4.227521293560716
  6.627301309180828
  3.9945552227534344
  4.464443898444791
  6.230677217587683
  2.956006239528009
  ⋮
 -2.771281762793124
  3.6458954107901445
 -0.07283339760799468
  1.4631207338470367
  0.7458633684802747
 -3.662089343726486
 -4.940952537780513
  0.25502125669220505
  2.7307956788207797

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  20.653      -13.0306    0.0894294  …  -14.1525    0.0971292   12.4529
 -13.0306      18.4097   -3.80722        19.9948   -4.13502     -9.52748
   0.0894294   -3.80722   1.51093        -4.13502   1.64102      1.81535
  11.4657      -8.7722    1.67144        -9.52748   1.81535     16.5074
  13.2985      -8.3904    0.0575836     -10.916     0.0749166    9.60504
  -8.3904      11.854    -2.45147    …   15.4221   -3.18938     -7.34863
   0.0575836   -2.45147   0.972889       -3.18938   1.26573      1.40019
   7.38278     -5.64842   1.07624        -7.34863   1.40019     12.7323
  14.2497      -8.99057   0.0617026      -9.81637   0.06737      8.6375
  -8.99057     12.7019   -2.62683        13.8686   -2.8681      -6.60838
   ⋮                                 ⋱
   7.17561     -5.48992   1.04604        -7.46028   1.42147     12.9258
  24.7284     -15.6019    0.107076      -15.455     0.106068    13.599
 -15.6019      22.0424   -4.55849        21.835    -4.51559    -10.4043
   0.107076    -4.55849   1.80908        -4.51559   1.79205      1.98242
  13.7282     -10.5032    2.00126    …  -10.4043    1.98242     18.0267
  22.4312     -14.1525    0.0971292     -21.2192    0.145628    18.6709
 -14.1525      19.9948   -4.13502        29.9786   -6.19973    -14.2848
   0.0971292   -4.13502   1.64102        -6.19973   2.46042      2.72179
  12.4529      -9.52748   1.81535       -14.2848    2.72179     24.75
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 4.54456  -2.8673    0.0196784   2.52296   …   0.0213726     2.74018
 0.0       3.19191  -1.17509    -0.481883     -1.27627      -0.523372
 0.0       0.0       0.360137    2.93091       0.391145      3.18326
 0.0       0.0       0.0         0.105166      0.0           0.11422
 0.0      -0.0       0.0         0.0           0.00753587    0.966172
 0.0       0.0      -0.0        -0.0       …  -0.450005     -0.184538
 0.0       0.0       0.0         0.0           0.137915      1.1224
 0.0       0.0       0.0         0.0           0.0           0.0402735
 0.0      -0.0       0.0         0.0          -0.0028978    -0.371527
 0.0       0.0      -0.0        -0.0           0.173042      0.0709613
 ⋮                                         ⋱
 0.0       0.0       0.0         0.0          -0.0          -0.00711357
 0.0      -0.0       0.0         0.0          -0.0027589    -0.353718
 0.0       0.0      -0.0        -0.0           0.164748      0.0675599
 0.0       0.0       0.0         0.0          -0.0504912    -0.410913
 0.0       0.0       0.0         0.0       …  -0.0          -0.0147442
 0.0      -0.0       0.0         0.0           0.000763515   0.0978901
 0.0       0.0      -0.0        -0.0          -0.0455933    -0.0186969
 0.0       0.0       0.0         0.0           0.0139732     0.113719
 0.0       0.0       0.0         0.0           0.0           0.0040804
julia> logdet(C)-94.40286428679747
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
      2.0169e5   247922.0        …      -8.63822e5        1.06033e5
 247922.0             3.04809e5         -1.06198e6        1.30315e5
      7.09053e5       8.71707e5         -3.03752e6        3.72757e5
 -87035.4            -1.06966e5          3.72757e5   -45802.6
      2.67456e5       3.28763e5         -1.27566e6        1.56585e5
      3.28763e5       4.04199e5  …      -1.56829e6        1.92443e5
      9.40257e5       1.15595e6         -4.48568e6        5.50471e5
     -1.15415e5      -1.41845e5          5.50471e5   -67639.4
     -3.2762e5       -4.02719e5          1.5308e6   -187904.0
     -4.02719e5      -4.95124e5          1.88196e6       -2.30934e5
      ⋮                          ⋱
  -1755.27        -2157.22            6957.46          -854.901
 -99795.7            -1.22671e5          4.38258e5   -53795.6
     -1.22671e5      -1.50819e5     538793.0         -66114.8
     -3.50838e5      -4.31319e5          1.54108e6       -1.89117e5
  43064.9         52926.7        …      -1.89117e5    23237.9
     -2.45714e5      -3.02037e5          1.13273e6       -1.39042e5
     -3.02037e5      -3.71341e5          1.39258e6       -1.70882e5
     -8.63822e5      -1.06198e6          3.98311e6       -4.88798e5
      1.06033e5       1.30315e5         -4.88798e5    60061.2

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.