Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 16.0176    -0.352725  -2.39822   …  -0.151329  -1.02891   -0.971908
 -0.352725   7.59705    2.1692        3.25935    0.930648  -0.370655
 -2.39822    2.1692     6.41781       0.930648   2.75342    0.752375
 -2.26537   -0.863942   1.75367      -0.370655   0.752375   4.61551
  8.34682   -0.183806  -1.24973      -0.120894  -0.821974  -0.776439
 -0.183806   3.95885    1.13038   …   2.60383    0.743478  -0.29611
 -1.24973    1.13038    3.34435       0.743478   2.19966    0.601059
 -1.18049   -0.450204   0.913847     -0.29611    0.601059   3.68725
 16.3116    -0.3592    -2.44225      -0.224098  -1.52367   -1.43927
 -0.3592     7.73652    2.20902       4.82666    1.37817   -0.548891
  ⋮                               ⋱
 -1.57945   -0.602355   1.22269      -0.305965   0.621063   3.80997
 13.1002    -0.288481  -1.96143      -0.15183   -1.03232   -0.975129
 -0.288481   6.21337    1.77412       3.27015    0.933732  -0.371884
 -1.96143    1.77412    5.24891       0.933732   2.76254    0.754868
 -1.85277   -0.706589   1.43427   …  -0.371884   0.754868   4.63081
  6.87198   -0.151329  -1.02891      -0.160931  -1.09419   -1.03358
 -0.151329   3.25935    0.930648      3.46616    0.989699  -0.394174
 -1.02891    0.930648   2.75342       0.989699   2.92813    0.800114
 -0.971908  -0.370655   0.752375     -0.394174   0.800114   4.90838
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.00036828340209797564
   0.0007929541629812898
   0.0009210507988245178
   0.001584517330379513
   0.012410316297326977
   0.02672075883903665
   0.031037325261476324
   0.05339464427825213
   0.03198657513566396
   0.06887057024250828
   ⋮
   9.528209383712728
   2.414076258975741
   5.19776837186884
   6.037434361924864
  10.386418848675635
  32.30676607995235
  69.55997612063373
  80.79694207196667
 138.9979305355257
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0322799  -0.0909726  -0.111121    …  -0.0874252  -0.106788    -0.28674
  0.181114   -0.257058    0.107026       -0.247034    0.102853     0.0181846
 -0.264746   -0.182184   -0.00475762     -0.17508    -0.0045721    0.0826084
  0.0822111  -0.0561054  -0.294734       -0.0539176  -0.283241     0.113377
 -0.0390559  -0.110069   -0.134447       -0.0594595  -0.0726285   -0.195017
  0.219132   -0.311017    0.129492    …  -0.168012    0.0699522    0.0123677
 -0.32032    -0.220427   -0.0057563      -0.119075   -0.00310957   0.0561835
  0.0994683  -0.0678826  -0.356603       -0.0366704  -0.192638     0.0771097
  0.0404323   0.113948    0.139185       -0.105513   -0.128882    -0.346064
 -0.226854    0.321978   -0.134056       -0.298143    0.124132     0.0219469
  ⋮                                   ⋱
  0.044351   -0.0302676  -0.159002       -0.042863   -0.225169     0.0901315
 -0.0107669  -0.0303438  -0.0370643      -0.0986118  -0.120452    -0.32343
  0.0604102  -0.0857413   0.0356985      -0.278643    0.116013     0.0205115
 -0.0883058  -0.0607672  -0.0015869      -0.197482   -0.00515713   0.0931786
  0.0274214  -0.0187139  -0.0983083   …  -0.0608167  -0.319484     0.127884
 -0.0453439  -0.12779    -0.156093       -0.0533141  -0.065122    -0.174861
  0.254412   -0.361091    0.150341       -0.150647    0.0627223    0.0110895
 -0.371892   -0.255915   -0.00668307     -0.106768   -0.00278818   0.0503767
  0.115483   -0.0788118  -0.414016       -0.0328803  -0.172728     0.0691401
julia> logdet(E)-12.659290348988439
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  2.2810671885289304
  0.8048570468496116
  4.94791974131207
  0.8951889915186695
  5.243581117796657
 -6.700304561270636
 -0.5107023982187463
 -4.74064232207516
 -3.5764388387224693
  3.5589678481886695
  ⋮
 -0.200843702743148
  0.7655763479668506
 -6.313175657750063
 -3.7582249862073427
 -0.04180105029241231
  2.4360180158774054
 -0.7005703806197544
  3.3908378284807856
  2.4261341015113476

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 35.3086   17.216      3.28893   11.58     …  12.5187     2.39157    8.4205
 17.216    27.5248    -0.404744   5.01528     20.0148    -0.294313   3.6469
  3.28893  -0.404744   3.30591   -1.69743     -0.294313   2.40392   -1.2343
 11.58      5.01528   -1.69743    7.60236      3.6469    -1.2343     5.52811
 31.4821   15.3502     2.9325    10.3251      11.6787     2.23109    7.85545
 15.3502   24.5418    -0.360881   4.47176  …  18.6718    -0.274563   3.40218
  2.9325   -0.360881   2.94764   -1.51347     -0.274563   2.2426    -1.15147
 10.3251    4.47176   -1.51347    6.77847      3.40218   -1.15147    5.15715
 14.7525    7.19312    1.37417    4.83832      6.21148    1.18664    4.17804
  7.19312  11.5003    -0.169109   2.09547      9.93086   -0.146031   1.8095
  ⋮                                        ⋱
  7.30185   3.16242   -1.07032    4.79371      2.21339   -0.749124   3.35514
 26.7402   13.0381     2.4908     8.76987     10.2295     1.95424    6.88069
 13.0381   20.8452    -0.306524   3.79821     16.3548    -0.240493   2.98001
  2.4908   -0.306524   2.50365   -1.28551     -0.240493   1.96432   -1.00859
  8.76987   3.79821   -1.28551    5.75747  …   2.98001   -1.00859    4.51721
 25.6749   12.5187     2.39157    8.4205      11.5693     2.21019    7.78186
 12.5187   20.0148    -0.294313   3.6469      18.4968    -0.271991   3.3703
  2.39157  -0.294313   2.40392   -1.2343      -0.271991   2.22159   -1.14068
  8.4205    3.6469    -1.2343     5.52811      3.3703    -1.14068    5.10884
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 5.9421  2.89729   0.553497   1.94881   …  2.10678     0.402479    1.41709
 0.0     4.37384  -0.45918   -0.144261     3.18047    -0.333896   -0.104901
 0.0     0.0       1.66994   -1.70205      0.0         1.21431    -1.23766
 0.0     0.0       0.0        0.94165      0.0         0.0         0.684728
 0.0     0.0       0.0        0.0          0.0993002   0.0189703   0.0667926
 0.0     0.0      -0.0       -0.0       …  0.149907   -0.0157377  -0.00494434
 0.0     0.0       0.0       -0.0          0.0         0.0572348  -0.0583354
 0.0     0.0       0.0        0.0          0.0         0.0         0.0322737
 0.0     0.0       0.0        0.0          0.28152     0.0537814   0.189359
 0.0     0.0      -0.0       -0.0          0.424992   -0.0446171  -0.0140174
 ⋮                                      ⋱
 0.0     0.0       0.0        0.0          0.0         0.0         0.0112426
 0.0     0.0       0.0        0.0          0.556691    0.10635     0.374448
 0.0     0.0      -0.0       -0.0          0.840399   -0.0882279  -0.0277187
 0.0     0.0       0.0       -0.0          0.0         0.320866   -0.327036
 0.0     0.0       0.0        0.0       …  0.0         0.0         0.180931
 0.0     0.0       0.0        0.0          0.521701    0.0996655   0.350913
 0.0     0.0      -0.0       -0.0          0.787577   -0.0826825  -0.0259765
 0.0     0.0       0.0       -0.0          0.0         0.300699   -0.306481
 0.0     0.0       0.0        0.0          0.0         0.0         0.169559
julia> logdet(C)8.271327613688637
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  13.7842    -4.66791   -26.5252    …   1.45692     8.2789      7.44057
  -4.66791    3.3709      8.54525      -1.0521     -2.66709    -2.12063
 -26.5252     8.54525    65.8542       -2.66709   -20.554     -15.4403
 -23.8393     6.7944     49.47         -2.12063   -15.4403    -15.149
   1.16956   -0.396065   -2.25062      -0.151845   -0.862852   -0.775479
  -0.396065   0.286015    0.72505   …   0.109654    0.277972    0.221018
  -2.25062    0.72505     5.58762       0.277972    2.1422      1.60923
  -2.02272    0.576494    4.19745       0.221018    1.60923     1.57887
  -1.17833    0.399031    2.26748       1.43445     8.15119     7.3258
   0.399031  -0.288158   -0.730481     -1.03587    -2.62595    -2.08791
   ⋮                                ⋱
   4.30614   -1.22729    -8.93588      -0.182201   -1.32661    -1.30158
   0.486558  -0.164769   -0.936297      1.95956    11.1352     10.0076
  -0.164769   0.118987    0.301633     -1.41509    -3.58725    -2.85225
  -0.936297   0.301633    2.32454      -3.58725   -27.6452    -20.7672
  -0.841487   0.239831    1.74621   …  -2.85225   -20.7672    -20.3754
  -4.30223    1.45692     8.2789       -3.34513   -19.0086    -17.0838
   1.45692   -1.0521     -2.66709       2.41566     6.12372     4.86902
   8.2789    -2.66709   -20.554         6.12372    47.1926     35.4513
   7.44057   -2.12063   -15.4403        4.86902    35.4513     34.7825

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.