Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 17.0781    7.42443   -3.25546    5.10783  …   7.23316   -3.1716     4.97624
  7.42443   6.58313    0.540533   5.17083      6.41354    0.526608   5.03762
 -3.25546   0.540533   5.11557    4.95707      0.526608   4.98378    4.82936
  5.10783   5.17083    4.95707   10.3112       5.03762    4.82936   10.0456
 28.5182   12.3978    -5.43621    8.52942     16.8643    -7.39464   11.6022
 12.3978   10.993      0.902622   8.63463  …  14.9533     1.2278    11.7453
 -5.43621   0.902622   8.54234    8.27767      1.2278    11.6198    11.2598
  8.52942   8.63463    8.27767   17.2184      11.7453    11.2598    23.4214
 18.5086    8.04631   -3.52814    5.53567      9.09772   -3.98917    6.25901
  8.04631   7.13455    0.58581    5.60395      8.06682    0.662357   6.33622
  ⋮                                        ⋱
  5.03778   5.09992    4.88909   10.1698       5.59655    5.36519   11.1601
  8.21969   3.57338   -1.56685    2.4584       8.38746   -3.67773    5.77037
  3.57338   3.16846    0.260159   2.48872      7.43704    0.610647   5.84155
 -1.56685   0.260159   2.46212    2.38584      0.610647   5.77911    5.60005
  2.4584    2.48872    2.38584    4.96279  …   5.84155    5.60005   11.6487
 16.6381    7.23316   -3.1716     4.97624     13.2789    -5.82254    9.13559
  7.23316   6.41354    0.526608   5.03762     11.7742     0.966769   9.24827
 -3.1716    0.526608   4.98378    4.82936      0.966769   9.14943    8.86594
  4.97624   5.03762    4.82936   10.0456       9.24827    8.86594   18.4421
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   1.0131795858186938e-5
   6.476710166361592e-5
   0.00042415091952541836
   0.0008269980319258861
   0.004040464082282186
   0.025828505789909157
   0.1691473633891484
   0.3297989705758388
   0.027179161502252865
   0.17374170786572204
   ⋮
  24.274899367646164
   0.35287295768156113
   2.2557263336523916
  14.772444250872576
  28.80291367958326
   4.391178919637076
  28.070436425290854
 183.8294614927223
 358.42572977226854
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0299434   -0.0349931    0.0412891   …   0.0922069   0.168223
  0.00249542   0.0864617   -0.00888021     -0.0198313   0.0983513
  0.075438    -0.00829656  -0.0611592      -0.136581   -0.000711028
 -0.0539124   -0.0270426   -0.0630569      -0.140819    0.09699
  0.11652     -0.13617      0.160669        0.211159    0.385241
  0.0097105    0.336451    -0.0345558   …  -0.0454147   0.22523
  0.293554    -0.0322846   -0.23799        -0.312778   -0.00162829
 -0.209791    -0.105232    -0.245375       -0.322483    0.222113
 -0.191114     0.223344    -0.263529        0.140514    0.256356
 -0.0159271   -0.551843     0.0566781      -0.030221    0.149878
  ⋮                                     ⋱
 -0.201384    -0.101014    -0.235542       -0.202521    0.139488
  0.120599    -0.140937     0.166294        0.0903114   0.164765
  0.0100505    0.34823     -0.0357656      -0.0194236   0.0963295
  0.303831    -0.0334149   -0.246322       -0.133773   -0.000696412
 -0.217135    -0.108916    -0.253965    …  -0.137924    0.0949963
 -0.102341     0.1196      -0.141118        0.128381    0.234221
 -0.00852886  -0.295509     0.0303508      -0.0276115   0.136936
 -0.257832     0.028356     0.20903        -0.190164   -0.000989979
  0.184262     0.0924262    0.215516       -0.196065    0.135041
julia> logdet(E)-24.401465133257734
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
   9.404441280571561
  -1.9004381503583907
   9.048982310969558
  -8.766200200838218
   2.6542770347805718
  -0.6716619583084162
  -6.34804005526655
   9.979839262553131
 -12.191469407538685
  -1.8238840826191784
   ⋮
   3.84370886228514
  13.153670705818215
  -1.675580914872656
   3.9954257738562204
  -3.584531299921685
  -8.12757227974238
   2.4456214697513787
   3.011635219108588
  -2.7634065544643804

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 21.652     -6.66475  -3.81903   10.4071   …  -3.42555  -1.96291   5.34904
 -6.66475   24.3826   12.385    -10.3448      12.5322    6.36565  -5.31703
 -3.81903   12.385     7.28785   -3.1213       6.36565   3.74582  -1.60429
 10.4071   -10.3448   -3.1213    15.7878      -5.31703  -1.60429   8.11464
 13.7819    -4.24225  -2.4309     6.62434     -3.74213  -2.14432   5.84339
 -4.24225   15.52      7.88332   -6.58469  …  13.6904    6.95395  -5.80842
 -2.4309     7.88332   4.63887   -1.98678      6.95395   4.092    -1.75255
  6.62434   -6.58469  -1.98678   10.0493      -5.80842  -1.75255   8.86458
 11.6833    -3.59625  -2.06073    5.6156      -2.72803  -1.56322   4.25986
 -3.59625   13.1567    6.68286   -5.58199      9.98034   5.06946  -4.23437
  ⋮                                        ⋱
  6.03143   -5.99534  -1.80895    9.14985     -5.491    -1.65678   8.38015
 12.643     -3.89169  -2.23002    6.07692     -2.50042  -1.43279   3.90444
 -3.89169   14.2375    7.23187   -6.04056      9.14763   4.64649  -3.88107
 -2.23002    7.23187   4.25553   -1.8226       4.64649   2.73419  -1.17102
  6.07692   -6.04056  -1.8226     9.21886  …  -3.88107  -1.17102   5.92313
 11.1287    -3.42555  -1.96291    5.34904     -3.75642  -2.15251   5.8657
 -3.42555   12.5322    6.36565   -5.31703     13.7426    6.98051  -5.8306
 -1.96291    6.36565   3.74582   -1.60429      6.98051   4.10762  -1.75925
  5.34904   -5.31703  -1.60429    8.11464     -5.8306   -1.75925   8.89843
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 4.65317  -1.4323   -0.820739  …  -0.736177   -0.421844    1.14955
 0.0       4.72558   2.37208       2.42886     1.2192     -0.776736
 0.0       0.0       0.993713      0.0         0.51075     1.18915
 0.0       0.0       0.0           0.0         0.0         0.912092
 0.0      -0.0      -0.0          -0.466114   -0.267092    0.727842
 0.0       0.0       0.0       …   1.53784     0.771945   -0.491794
 0.0       0.0       0.0           0.0         0.323383    0.752915
 0.0       0.0       0.0           0.0         0.0         0.577495
 0.0      -0.0      -0.0          -0.0412152  -0.0236171   0.064358
 0.0       0.0       0.0           0.135981    0.0682577  -0.0434859
 ⋮                             ⋱
 0.0       0.0       0.0           0.0         0.0         0.42421
 0.0      -0.0      -0.0          -0.148181   -0.0849105   0.231386
 0.0       0.0       0.0           0.48889     0.245406   -0.156345
 0.0       0.0       0.0           0.0         0.102806    0.239357
 0.0       0.0       0.0       …   0.0         0.0         0.18359
 0.0      -0.0      -0.0          -0.20352    -0.116621    0.3178
 0.0       0.0       0.0           0.671471    0.337056   -0.214733
 0.0       0.0       0.0           0.0         0.1412      0.328747
 0.0       0.0       0.0           0.0         0.0         0.252153
julia> logdet(C)-25.37288002336843
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   71.2524   -130.311     220.8     …   -31.5733     53.4979   -21.4914
 -130.311     647.161   -1037.52        156.802    -251.383     73.856
  220.8     -1037.52     1763.16       -251.383     427.198   -115.522
  -88.7006    304.823    -476.791        73.856    -115.522     49.6179
   76.9594   -140.748     238.485       -34.2593     58.0491   -23.3197
 -140.748     698.996   -1120.62    …   170.141    -272.768     80.1391
  238.485   -1120.62     1904.38       -272.768     463.541   -125.35
  -95.8051    329.238    -514.98         80.1391   -125.35      53.839
   42.2799    -77.3242    131.019       -18.8638     31.963    -12.8403
  -77.3242    384.014    -615.647        93.6831   -150.192     44.1262
    ⋮                               ⋱
   38.0287   -130.687     204.415       -33.6211     52.5888   -22.5873
  -14.3216     26.1922    -44.3802        5.99618   -10.16       4.0815
   26.1922   -130.078     208.539       -29.7787     47.7409   -14.0262
  -44.3802    208.539    -354.39         47.7409    -81.1306    21.9393
   17.8286    -61.2686     95.8339  …   -14.0262     21.9393    -9.4231
   17.2638    -31.5733     53.4979      -10.0081     16.9578    -6.81236
  -31.5733    156.802    -251.383        49.7031    -79.6835    23.4109
   53.4979   -251.383     427.198       -79.6835    135.414    -36.6184
  -21.4914     73.856    -115.522        23.4109    -36.6184    15.7279

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.