Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.07667    -0.860249  -2.934    …  -0.709832   -2.42098    0.0461794
 -0.860249   10.7248    -3.27705      8.84956    -2.70405    3.26228
 -2.934      -3.27705   12.4434      -2.70405    10.2677    -6.72807
  0.055965    3.95358   -8.15378      3.26228    -6.72807   11.6148
  1.23802    -0.512843  -1.74912     -0.940655   -3.20823    0.061196
 -0.512843    6.39367   -1.95364  …  11.7273     -3.58335    4.32311
 -1.74912    -1.95364    7.41823     -3.58335    13.6065    -8.91589
  0.0333639   2.35695   -4.86092      4.32311    -8.91589   15.3917
  1.88689    -0.781632  -2.66586     -1.05143    -3.58606    0.0684029
 -0.781632    9.74469   -2.97757     13.1084     -4.00536    4.83224
  ⋮                               ⋱
  0.0359387   2.53884   -5.23606      3.50256    -7.2236    12.4702
  1.20581    -0.499499  -1.70361     -0.905925   -3.08978    0.0589365
 -0.499499    6.22731   -1.9028      11.2943     -3.45105    4.1635
 -1.70361    -1.9028     7.22521     -3.45105    13.1041    -8.58671
  0.0324958   2.29562   -4.73445  …   4.1635     -8.58671   14.8234
  1.71356    -0.709832  -2.42098     -1.19801    -4.08597    0.0779385
 -0.709832    8.84956   -2.70405     14.9357     -4.56372    5.50586
 -2.42098    -2.70405   10.2677      -4.56372    17.3291   -11.3552
  0.0461794   3.26228   -6.72807      5.50586   -11.3552    19.6027
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.008322292620348075
   0.07376874468701165
   0.11407823426915781
   0.2952499296381759
   0.026404964809874986
   0.23405342690866915
   0.3619473501910825
   0.936768791709229
   0.04002988531150531
   0.3548246287537156
   ⋮
  19.794364997095023
   0.7768143771825677
   6.885677309575792
  10.648221174921446
  27.55903939790619
   6.304807215235018
  55.88576789817906
  86.42345412884255
 223.67561098837533
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.126184    -0.0529027    0.0366535   …   0.0926205  -0.0641718  -0.0180167
 -0.0179334   -0.0347071   -0.12636         0.0607642   0.221227   -0.0905948
 -0.0559535    0.0810305   -0.0507173      -0.141866    0.0887943   0.155627
 -0.0282078    0.0979864    0.0169731      -0.171552   -0.0297161  -0.170513
  0.0282785    0.0118558   -0.00821423      0.119929   -0.0830923  -0.0233287
  0.00401897   0.00777804   0.0283178   …   0.0786799   0.286453   -0.117306
  0.0125394   -0.0181593    0.011366       -0.183693    0.114974    0.201513
  0.00632149  -0.0219592   -0.00380376     -0.222132   -0.0384775  -0.220787
  0.195476     0.0819533   -0.0567811       0.130156   -0.0901783  -0.0253182
  0.0277813    0.0537659    0.195748        0.0853897   0.310881   -0.127309
  ⋮                                     ⋱
  0.0605565   -0.210358    -0.036438       -0.165231   -0.0286212  -0.164231
  0.170635     0.0715387   -0.0495654       0.110473   -0.0765408  -0.0214893
  0.0242508    0.0469334    0.170872        0.0724763   0.263867   -0.108057
  0.0756642   -0.109575     0.0685834      -0.16921     0.105909    0.185624
  0.0381445   -0.132504    -0.0229523   …  -0.204618   -0.0354437  -0.203379
 -0.254215    -0.10658      0.0738434       0.131661   -0.0912211  -0.025611
 -0.0361293   -0.0699222   -0.254568        0.0863771   0.314476   -0.128782
 -0.112726     0.163247    -0.102177       -0.201664    0.126222    0.221226
 -0.0568283    0.197407     0.0341947      -0.243863   -0.0422418  -0.242387
julia> logdet(E)14.933402143717153
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -0.9295657800166547
 -0.614866259593722
  3.128086511094295
  2.4972289522877658
 -3.71637100430146
  0.47785700000188464
 -0.48806209830859887
 -0.25394230073712054
 -2.64809690996642
  0.9819184326004616
  ⋮
 -2.393021326980156
  0.7269202586020441
  0.4383763575461603
 -0.5479125293725053
 -0.06517207986006782
  2.631393331608355
 -1.0194038630532631
  0.9202430852731766
  0.3356656518888601

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.12573    5.04211   -0.733776  1.27225   …   3.18456   -0.463448  0.803548
  5.04211   13.139     -4.12618   0.56546       8.29851   -2.60607   0.357141
 -0.733776  -4.12618   11.4973    3.71519      -2.60607    7.26161   2.34649
  1.27225    0.56546    3.71519   6.59606       0.357141   2.34649   4.16603
  1.96645    4.66431   -0.678795  1.17693       4.57176   -0.665326  1.15357
  4.66431   12.1545    -3.81702   0.523092  …  11.9134    -3.74128   0.512712
 -0.678795  -3.81702   10.6358    3.43682      -3.74128   10.4248    3.36863
  1.17693    0.523092   3.43682   6.10183       0.512712   3.36863   5.98076
  2.36659    5.61341   -0.816917  1.41641       5.25965   -0.765435  1.32715
  5.61341   14.6277    -4.5937    0.629531     13.7059    -4.30421   0.589858
  ⋮                                         ⋱
  0.834009   0.37068    2.43544   4.32395       0.438908   2.88372   5.11983
  1.38178    3.27749   -0.476972  0.826996      3.80414   -0.553616  0.959885
  3.27749    8.54067   -2.68212   0.367563      9.91306   -3.1131    0.426626
 -0.476972  -2.68212    7.47351   2.41496      -3.1131     8.67442   2.80302
  0.826996   0.367563   2.41496   4.2876    …   0.426626   2.80302   4.97656
  1.3426     3.18456   -0.463448  0.803548      5.89976   -0.85859   1.48866
  3.18456    8.29851   -2.60607   0.357141     15.3739    -4.82804   0.661644
 -0.463448  -2.60607    7.26161   2.34649      -4.82804   13.453     4.34714
  0.803548   0.357141   2.34649   4.16603       0.661644   4.34714   7.71804
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.45799  3.45826  -0.503279  …   2.18422    -0.317868    0.551135
 0.0      1.08602  -2.19675       0.685922   -1.38745    -1.42615
 0.0      0.0       2.53344       0.0         1.6001     -0.200923
 0.0      0.0       0.0           0.0         0.0         0.503201
 0.0      0.0      -0.0           1.66057    -0.241662    0.419005
 0.0      0.0      -0.0       …   0.521479   -1.05482    -1.08424
 0.0      0.0       0.0           0.0         1.21649    -0.152753
 0.0      0.0       0.0           0.0         0.0         0.382563
 0.0      0.0      -0.0           0.809647   -0.117828    0.204295
 0.0      0.0      -0.0           0.254258   -0.514301   -0.528647
 ⋮                            ⋱
 0.0      0.0       0.0          -0.0        -0.0        -0.0189054
 0.0      0.0      -0.0          -0.206086    0.0299916  -0.0520009
 0.0      0.0      -0.0          -0.0647184   0.130909    0.134561
 0.0      0.0       0.0          -0.0        -0.150973    0.0189575
 0.0      0.0       0.0       …  -0.0        -0.0        -0.0474783
 0.0      0.0      -0.0           0.947407   -0.137876    0.239056
 0.0      0.0      -0.0           0.29752    -0.601809   -0.618595
 0.0      0.0       0.0           0.0         0.694047   -0.0871505
 0.0      0.0       0.0           0.0         0.0         0.218264
julia> logdet(C)-40.03575130179897
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 2049.66    -792.187    -58.2959    -294.594    …  -31.6265     -159.822
 -792.187    308.972     23.8518     112.876        12.94         61.2371
  -58.2959    23.8518     5.54741      6.07489       3.00956       3.29572
 -294.594    112.876      6.07489     48.379         3.29572      26.2464
 -541.5      209.288     15.4012      77.8288        7.89603      39.9019
  209.288    -81.6275    -6.30141    -29.8208   …   -3.23066     -15.2888
   15.4012    -6.30141   -1.46557     -1.60493      -0.751383     -0.822828
   77.8288   -29.8208    -1.60493    -12.7813       -0.822828     -6.55281
  644.752   -249.195    -18.3379     -92.6689      -15.9491      -80.5977
 -249.195     97.1919     7.50294     35.5069        6.52559      30.8817
    ⋮                                           ⋱
  -31.6817    12.1391     0.653317     5.20286       0.0665245     0.529785
  408.24    -157.783    -11.6111     -58.6755       -4.14033     -20.9228
 -157.783     61.5394     4.75066     22.4821        1.69402       8.01677
  -11.6111     4.75066    1.1049       1.20996       0.393992      0.431455
  -58.6755    22.4821     1.20996      9.63586  …    0.431455      3.43601
 1111.97    -429.774    -31.6265    -159.822       -25.2939     -127.821
 -429.774    167.622     12.94        61.2371       10.349        48.9756
  -31.6265    12.94       3.00956      3.29572       2.40695       2.63582
 -159.822     61.2371     3.29572     26.2464        2.63582      20.991

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.