Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.1705     -0.0478666   6.13688  …  -0.0376034   4.82105  -6.38016
 -0.0478666   7.88174    -4.99058      6.1918     -3.92053   3.06905
  6.13688    -4.99058    17.6995      -3.92053    13.9045   -7.70596
 -8.12152     3.9067     -9.80917      3.06905    -7.70596   7.10042
  6.42088    -0.0302193   3.87436     -0.0281476   3.60875  -4.7758
 -0.0302193   4.97594    -3.15067  …   4.6348     -2.93467   2.2973
  3.87436    -3.15067    11.1741      -2.93467    10.4081   -5.76821
 -5.12731     2.46639    -6.19277      2.2973     -5.76821   5.31494
 10.5271     -0.0495451   6.35208     -0.0410016   5.25673  -6.95674
 -0.0495451   8.15813    -5.16558      6.75135    -4.27483   3.3464
  ⋮                                ⋱
 -6.33372     3.04671    -7.64986      2.93444    -7.36796   6.78898
  9.94436    -0.0468023   6.00043     -0.0414212   5.31053  -7.02793
 -0.0468023   7.7065     -4.87961      6.82044    -4.31858   3.38065
  6.00043    -4.87961    17.306       -4.31858    15.3162   -8.48833
 -7.94094     3.81984    -9.59107  …   3.38065    -8.48833   7.82131
  7.98982    -0.0376034   4.82105     -0.0457624   5.86711  -7.76451
 -0.0376034   6.1918     -3.92053      7.53527    -4.7712    3.73497
  4.82105    -3.92053    13.9045      -4.7712     16.9215   -9.37797
 -6.38016     3.06905    -7.70596      3.73497    -9.37797   8.64105

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.00030608558972668046
   0.026757275951276863
   0.04988094752535628
   0.16414358983285215
   0.0007597534862403277
   0.06641584696760741
   0.12381250555833308
   0.4074306952211328
   0.0014422623225542578
   0.12607914993049377
   ⋮
   9.127285349701914
   0.025659308242677183
   2.243075840261748
   4.181544806251155
  13.76023932175856
   0.4360851063216566
  38.1215252194977
  71.06619532283028
 233.85803587877396
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0489904    0.026302    …   0.096401   -0.235667    0.144532
 -0.0259066   -0.0633231      -0.232089   -0.219012   -0.0843481
  0.016616    -0.0580468      -0.212751    0.106598    0.240057
  0.0737154   -0.0266501      -0.0976769   0.0556238  -0.179808
  0.297538     0.159742        0.0663387  -0.162175    0.09946
 -0.15734     -0.384586    …  -0.159713   -0.150714   -0.0580445
  0.100915    -0.35254        -0.146405    0.0733558   0.165196
  0.447702    -0.161856       -0.0672168   0.0382777  -0.123736
  0.0501711    0.0269359       0.105216   -0.257217    0.157748
 -0.0265309   -0.0648492      -0.253312   -0.239039   -0.0920611
  ⋮                        ⋱
 -0.147831     0.053445       -0.0879335   0.0500752  -0.161872
  0.00215061   0.00115462      0.101661   -0.248527    0.152418
 -0.00113726  -0.0027798      -0.244754   -0.230963   -0.0889508
  0.00072942  -0.00254818     -0.22436     0.112415    0.253157
  0.00323601  -0.0011699   …  -0.103007    0.0586591  -0.18962
 -0.0251291   -0.0134913       0.0878711  -0.214814    0.131743
  0.0132885    0.0324809      -0.211553   -0.199633   -0.0768847
 -0.00852299   0.0297744      -0.193926    0.0971658   0.218816
 -0.0378115    0.0136699      -0.0890341   0.050702   -0.163898

julia> logdet(E)
-40.710607925578316

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  1.4446491188647839
 -4.517900019431869
 -0.5666692046833028
  1.9159992446509089
  3.206609696954369
  2.5421550150902212
  3.898178606357167
  0.6755540987797051
  2.70698328821795
  3.8617386513740546
  ⋮
  0.8858676271145092
 -1.5914387524620597
 -0.609205709706931
 -0.28530361295621226
 -0.7399694052796675
  3.075500309939212
 -2.3852436536585535
  1.0168209620481792
  4.969401606600869

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  37.5617   -2.85553   -3.1018    …  -2.59992   -2.82414   -11.1443
  -2.85553   5.20872    0.508518      4.74246    0.462998    2.60707
  -3.1018    0.508518  21.4694        0.462998  19.5475      3.98107
 -12.24      2.86339    4.37247       2.60707    3.98107    16.9018
  37.3221   -2.83731   -3.082        -2.90787   -3.15865   -12.4643
  -2.83731   5.17548    0.505273  …   5.30419    0.517838    2.91587
  -3.082     0.505273  21.3324        0.517838  21.8629      4.45261
 -12.1619    2.84512    4.34457       2.91587    4.45261    18.9037
  25.9992   -1.97652   -2.14698      -1.92364   -2.08954    -8.24553
  -1.97652   3.60534    0.351983      3.50888    0.342565    1.92893
   ⋮                              ⋱
 -10.4527    2.44527    3.73399       2.33486    3.5654     15.137
  21.0584   -1.60091   -1.73897      -2.02625   -2.201      -8.68536
  -1.60091   2.92019    0.285093      3.69604    0.360838    2.03182
  -1.73897   0.285093  12.0365        0.360838  15.2344      3.10265
  -6.86217   1.60532    2.45136   …   2.03182    3.10265    13.1724
  34.1994   -2.59992   -2.82414      -3.18676   -3.46159   -13.6598
  -2.59992   4.74246    0.462998      5.81292    0.567505    3.19553
  -2.82414   0.462998  19.5475        0.567505  23.9597      4.87967
 -11.1443    2.60707    3.98107       3.19553    4.87967    20.7168

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 6.12876  -0.465923  -0.506105  …  -0.424216   -0.460801    -1.81837
 0.0       2.2342     0.122063      2.0342      0.111136     0.787689
 0.0       0.0        4.60416       0.0         4.19202      0.643904
 0.0       0.0        0.0           0.0         0.0          3.32374
 0.0      -0.0       -0.0          -0.0910771  -0.0989318   -0.390395
 0.0       0.0        0.0       …   0.436734    0.0238604    0.169113
 0.0       0.0        0.0           0.0         0.900007     0.138243
 0.0       0.0        0.0           0.0         0.0          0.713592
 0.0      -0.0       -0.0           0.0490655   0.0532969    0.210315
 0.0       0.0        0.0          -0.235279   -0.0128542   -0.0911054
 ⋮                              ⋱
 0.0       0.0        0.0           0.0         0.0          0.267117
 0.0      -0.0       -0.0          -0.0793104  -0.0861503   -0.339958
 0.0       0.0        0.0           0.38031     0.0207778    0.147265
 0.0       0.0        0.0           0.0         0.783731     0.120383
 0.0       0.0        0.0       …   0.0         0.0          0.6214
 0.0      -0.0       -0.0          -0.0335126  -0.0364028   -0.143649
 0.0       0.0        0.0           0.1607      0.00877965   0.0622266
 0.0       0.0        0.0           0.0         0.331165     0.0508677
 0.0       0.0        0.0           0.0         0.0          0.262572

julia> logdet(C)
15.70710877611127

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.14782      1.25042     0.0759513  …    0.845393    0.0513498    2.59812
  1.25042     38.041       0.322204       25.7191      0.217838    -3.46103
  0.0759513    0.322204    8.81384         0.217838    5.95894     -1.40332
  3.84286     -5.11919    -2.07565        -3.46103    -1.40332      9.13608
 -7.24559     -1.47369    -0.0895133      -1.97952    -0.120238    -6.0836
 -1.47369    -44.8337     -0.379737   …  -60.2224     -0.510078     8.10415
 -0.0895133   -0.379737  -10.3877         -0.510078  -13.9531       3.28594
 -4.52905      6.03328     2.44628         8.10415     3.28594    -21.3925
  8.61077      1.75136     0.106379        2.78507     0.169167     8.55926
  1.75136     53.2811      0.451285       84.7293      0.717648   -11.402
  ⋮                                   ⋱
 -6.60578      8.79976     3.56798         9.80632     3.97611    -25.8858
 -5.01789     -1.0206     -0.0619919      -1.23292    -0.0748883   -3.78908
 -1.0206     -31.0493     -0.262984      -37.5086     -0.317694     5.04755
 -0.0619919   -0.262984   -7.1939         -0.317694   -8.69049      2.0466
 -3.13656      4.17831     1.69415    …    5.04755     2.0466     -13.324
  4.15648      0.845393    0.0513498       1.34215     0.0815233    4.12479
  0.845393    25.7191      0.217838       40.8318      0.345841    -5.49475
  0.0513498    0.217838    5.95894         0.345841    9.46045     -2.22792
  2.59812     -3.46103    -1.40332        -5.49475    -2.22792     14.5045

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.