Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.23264  -9.14196  -1.87385  -7.0035   …  -5.63611  -1.15525   -4.31772
 -9.14196  29.1014   -2.93779  15.8065      17.9413   -1.81117    9.74487
 -1.87385  -2.93779  16.3365    8.71994     -1.81117  10.0716     5.37593
 -7.0035   15.8065    8.71994  15.751        9.74487   5.37593    9.71062
  5.78105  -6.41959  -1.31584  -4.91794     -4.40301  -0.902498  -3.37307
 -6.41959  20.4353   -2.06295  11.0995   …  14.016    -1.41492    7.61285
 -1.31584  -2.06295  11.4717    6.12324     -1.41492   7.86811    4.19976
 -4.91794  11.0995    6.12324  11.0605       7.61285   4.19976    7.58609
  7.40706  -8.22519  -1.68594  -6.30118     -6.9465   -1.42384   -5.32159
 -8.22519  26.183    -2.64318  14.2214      22.1126   -2.23227   12.0105
  ⋮                                      ⋱
 -6.08068  13.7238    7.57096  13.6755       8.68137   4.78923    8.65085
  5.27646  -5.85927  -1.20099  -4.48868     -5.17768  -1.06128   -3.96653
 -5.85927  18.6517   -1.88289  10.1307      16.482    -1.66386    8.95225
 -1.20099  -1.88289  10.4704    5.58879     -1.66386   9.25243    4.93866
 -4.48868  10.1307    5.58879  10.0951   …   8.95225   4.93866    8.92078
  5.0755   -5.63611  -1.15525  -4.31772     -6.26715  -1.2846    -4.80116
 -5.63611  17.9413   -1.81117   9.74487     19.9501   -2.01396   10.836
 -1.15525  -1.81117  10.0716    5.37593     -2.01396  11.1993     5.97784
 -4.31772   9.74487   5.37593   9.71062     10.836     5.97784   10.7979
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0001583069173395446
   0.0023650891827513435
   0.011376139784375544
   0.02358164810664045
   0.0013280233622369436
   0.019840533447637785
   0.09543347605790306
   0.19782445474886534
   0.01769238336123569
   0.2643224011178182
   ⋮
  24.226704825716247
   0.2038730488184061
   3.0458425349837968
  14.650580913343678
  30.369250923835928
   2.076464476340886
  31.022167280459726
 149.21742231565355
 309.3134236325536
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.00338952  -0.133822    0.00514563  …  -0.32609     0.0125386   -0.107307
  0.0637476   -0.0351665   0.0536005      -0.0856917   0.130611     0.263711
  0.0685648   -0.0116933  -0.121615       -0.0284935  -0.296345     0.0391007
 -0.1054      -0.0245724  -0.04686        -0.0598767  -0.114186     0.188383
 -0.00488092  -0.192704    0.00740973     -0.263094    0.0101163   -0.0865765
  0.0917968   -0.0506398   0.0771849   …  -0.0691372   0.105378     0.212765
  0.0987336   -0.0168383  -0.175126       -0.0229889  -0.239095     0.031547
 -0.151776    -0.0353843  -0.0674786      -0.0483093  -0.0921267    0.15199
 -0.00602097  -0.237715    0.00914043     -0.350004    0.0134581   -0.115176
  0.113238    -0.0624679   0.0952131      -0.091976    0.140189     0.28305
  ⋮                                    ⋱
  0.302185     0.0704497   0.134349       -0.0602932  -0.11498      0.189693
 -0.00394868  -0.155898    0.0059945      -0.247478    0.00951584  -0.0814376
  0.0742639   -0.0409678   0.0624429      -0.0650335   0.0991236    0.200136
  0.0798759   -0.0136223  -0.141678       -0.0216244  -0.224903     0.0296745
 -0.122788    -0.0286261  -0.0545904   …  -0.0454418  -0.0866584    0.142968
  0.00821957   0.324518   -0.0124781      -0.236955    0.00911121  -0.0779748
 -0.154588     0.0852786  -0.129981       -0.0622682   0.0949088    0.191626
 -0.166269     0.0283561   0.294916       -0.0207049  -0.21534      0.0284127
  0.255595     0.059588    0.113635       -0.0435096  -0.0829736    0.136889
julia> logdet(E)-12.620016321212791
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -0.4014526885433839
  7.177107137939965
  5.810327758986452
 -5.960256455092849
  4.735113327203116
 -0.031641564634528156
 -1.0398389511353658
  5.048211770015741
  4.799933532945746
  4.56358084577457
  ⋮
 -1.9179833634975318
  2.8410645404838486
  0.1189017896237684
 -1.0986809586566415
  4.784763240393943
 -6.386955861831917
 -6.570180233100666
 -5.032152197439454
  0.20800826070370143

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.85114   6.49443   -2.56955   -2.67318  …    5.348     -2.11596   -2.2013
  6.49443  10.1003    -2.04622   -6.83014       8.3173    -1.68501   -5.62444
 -2.56955  -2.04622   22.0263   -12.1331       -1.68501   18.1381    -9.99127
 -2.67318  -6.83014  -12.1331    14.8553       -5.62444   -9.99127   12.2329
  2.91652   3.90447   -1.54482   -1.60713       5.09149   -2.01447   -2.09571
  3.90447   6.0723    -1.23019   -4.1063   …    7.91837   -1.60419   -5.35467
 -1.54482  -1.23019   13.2423    -7.29444      -1.60419   17.2681    -9.51205
 -1.60713  -4.1063    -7.29444    8.93104      -5.35467   -9.51205   11.6462
  3.84338   5.1453    -2.03576   -2.11787       4.497     -1.77926   -1.85102
  5.1453    8.00206   -1.62114   -5.41126       6.99381   -1.41688   -4.72945
  ⋮                                        ⋱
 -2.23588  -5.71281  -10.1483    12.4251       -8.09794  -14.3852    17.6127
  4.74398   6.35098   -2.51279   -2.61414       4.81759   -1.9061    -1.98298
  6.35098   9.87716   -2.00102   -6.67927       7.4924    -1.51789   -5.06662
 -2.51279  -2.00102   21.5398   -11.8651       -1.51789   16.3392    -9.00035
 -2.61414  -6.67927  -11.8651    14.5272   …   -5.06662   -9.00035   11.0197
  3.99479   5.348     -2.11596   -2.2013       10.7225    -4.2424    -4.4135
  5.348     8.3173    -1.68501   -5.62444      16.6758    -3.37836  -11.2767
 -2.11596  -1.68501   18.1381    -9.99127      -3.37836   36.366    -20.032
 -2.2013   -5.62444   -9.99127   12.2329      -11.2767   -20.032     24.5265
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.20253  2.94862  -1.16664  -1.21369   …   2.42812   -0.960695  -0.999441
 0.0      1.18569   1.17548  -2.74221       0.976388   0.967976  -2.25814
 0.0      0.0       4.3913   -2.35138       0.0        3.61612   -1.9363
 0.0      0.0       0.0       0.577532      0.0        0.0        0.475583
 0.0      0.0      -0.0      -0.0           1.26518   -0.500573  -0.520762
 0.0      0.0       0.0      -0.0       …   0.50875    0.504367  -1.17661
 0.0      0.0       0.0      -0.0           0.0        1.88419   -1.00891
 0.0      0.0       0.0       0.0           0.0        0.0        0.247804
 0.0      0.0      -0.0      -0.0          -0.64478    0.25511    0.265399
 0.0      0.0       0.0      -0.0          -0.259277  -0.257043   0.599644
 ⋮                                      ⋱
 0.0      0.0       0.0       0.0           0.0        0.0        0.277536
 0.0      0.0      -0.0      -0.0          -1.16591    0.461299   0.479904
 0.0      0.0       0.0      -0.0          -0.468835  -0.464795   1.0843
 0.0      0.0       0.0      -0.0          -0.0       -1.73636    0.929756
 0.0      0.0       0.0       0.0       …  -0.0       -0.0       -0.228362
 0.0      0.0      -0.0      -0.0           1.14731   -0.453938  -0.472246
 0.0      0.0       0.0      -0.0           0.461353   0.457378  -1.067
 0.0      0.0       0.0      -0.0           0.0        1.70865   -0.91492
 0.0      0.0       0.0       0.0           0.0        0.0        0.224718
julia> logdet(C)-13.716487538274151
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  128.23    -134.912   -34.598    …    62.712     16.0824    31.2429
 -134.912    148.63     40.6089       -69.0883   -18.8764   -35.8977
  -34.598     40.6089   13.176        -18.8764    -6.12464  -10.7873
  -67.213     77.2269   23.2067       -35.8977   -10.7873   -20.1457
   38.2331   -40.2255  -10.3158        18.9844     4.8685     9.45795
  -40.2255    44.3156   12.108    …   -20.9146    -5.71433  -10.8671
  -10.3158    12.108     3.92855       -5.71433   -1.85407   -3.26556
  -20.0403    23.026     6.91932      -10.8671    -3.26556   -6.09858
   46.5559   -48.982   -12.5613        52.4784    13.458     26.1446
  -48.982     53.9624   14.7437       -57.8143   -15.7961   -30.0398
    ⋮                             ⋱
  -12.6725    14.5606    4.37545      -15.7494    -4.7327    -8.83853
 -143.103    150.56     38.6108      -102.083    -26.1789   -50.8573
  150.56    -165.869   -45.3189       112.462     30.7271    58.4344
   38.6108   -45.3189  -14.7041        30.7271     9.96971   17.5596
   75.0086   -86.1839  -25.8983   …    58.4344    17.5596    32.7933
  -59.6057    62.712    16.0824       -61.6442   -15.8085   -30.711
   62.712    -69.0883  -18.8764        67.912     18.555     35.2865
   16.0824   -18.8764   -6.12464       18.555      6.02036   10.6036
   31.2429   -35.8977  -10.7873        35.2865    10.6036    19.8027

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.