Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  13.0073   -21.0182    11.099    …  -16.0724     8.48729  -5.55797
 -21.0182    50.4313   -18.5924       38.5643   -14.2174    4.63384
  11.099    -18.5924    20.9851      -14.2174    16.0471   -7.9554
  -7.26828    6.05977  -10.4034        4.63384   -7.9554   11.4936
   6.37738  -10.3051     5.44178      -9.89225    5.22377  -3.42083
 -10.3051    24.7262    -9.11575  …   23.7356    -8.75055   2.85204
   5.44178   -9.11575   10.2889       -8.75055    9.87666  -4.8964
  -3.5636     2.97107   -5.10075       2.85204   -4.8964    7.07407
   8.68593  -14.0354     7.41165     -12.4863     6.59359  -4.31787
 -14.0354    33.6768   -12.4156       29.9598   -11.0452    3.59993
   ⋮                              ⋱
  -5.3827     4.48771   -7.70452       3.69403   -6.34193   9.16249
   9.17015  -14.8179     7.82483     -14.3831     7.59522  -4.97379
 -14.8179    35.5542   -13.1077       34.5109   -12.7231    4.14679
   7.82483  -13.1077    14.7945      -12.7231    14.3604   -7.11924
  -5.12416    4.27215   -7.33446  …    4.14679   -7.11924  10.2855
   9.94651  -16.0724     8.48729     -14.744      7.78581  -5.0986
 -16.0724    38.5643   -14.2174       35.3769   -13.0423    4.25085
   8.48729  -14.2174    16.0471      -13.0423    14.7208   -7.29788
  -5.55797    4.63384   -7.9554        4.25085   -7.29788  10.5436
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0025654214145010125
   0.0075692897211995635
   0.02334765305608285
   0.08789393373519287
   0.015614861157700739
   0.04607173207132978
   0.14210934654562857
   0.5349809446983721
   0.05400815713554432
   0.1593513589448698
   ⋮
  26.22678779803119
   1.1460444860173928
   3.381410437682874
  10.430040419529023
  39.26464382897048
  15.08428088732088
  44.50625212165894
 137.28058663849998
 516.802727715242
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.22822      -0.0717862   …  -0.092407    0.0255839   0.129868
 -0.0890058     0.0273161       0.0351627   0.164232   -0.266512
 -0.00215097    0.196066        0.252386    0.170717    0.13969
 -0.0883171     0.153198        0.197205   -0.235782   -0.0704029
 -0.0197895    -0.00622475     -0.0671536   0.0185922   0.0943768
 -0.00771789    0.00236864  …   0.0255532   0.11935    -0.193678
 -0.000186515   0.0170013       0.183413    0.124063    0.101515
 -0.00765817    0.0132842       0.143312   -0.171346   -0.0511629
 -0.0315094    -0.00991124     -0.0840059   0.0232579   0.118061
 -0.0122887     0.00377142      0.0319659   0.149301   -0.242282
  ⋮                         ⋱
 -0.0225376     0.0390946       0.194391   -0.232417   -0.0693983
 -0.126306     -0.0397293      -0.0901813   0.0249677   0.12674
 -0.0492592     0.0151178       0.0343157   0.160276   -0.260093
 -0.00119043    0.10851         0.246307    0.166605    0.136325
 -0.048878      0.0847859   …   0.192455   -0.230103   -0.0687072
  0.491751      0.15468        -0.0801316   0.0221853   0.112616
  0.191783     -0.0588586       0.0304916   0.142415   -0.231108
  0.00463475   -0.422468        0.218859    0.148039    0.121133
  0.190299     -0.330101        0.171008   -0.20446    -0.0610505
julia> logdet(E)11.930816371820884
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -1.6678903495789628
 -3.365958769111387
 -2.602839362170779
  1.4426584292355866
  0.2348550885261613
  0.32620282099134523
 -0.3454091858130247
  0.6154734242336075
 -1.5343261342030348
 -1.2792051758673988
  ⋮
  5.794234870523919
 -2.202417537734457
 -3.1578209978215748
 -5.537595957031796
 -2.271307000686846
  1.7899719476807903
  6.922763332614272
  4.718982994527394
 -4.152923991279292

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.7235   5.78706   9.11791  4.76687  …   4.43858   6.99329   3.65612
  5.78706  8.17611   9.14378  3.29459      6.27094   7.01314   2.5269
  9.11791  9.14378  14.1736   7.00308      7.01314  10.871     5.37125
  4.76687  3.29459   7.00308  4.44911      2.5269    5.37125   3.4124
  7.37822  3.64209   5.73837  3.00004      7.14907  11.2639    5.88878
  3.64209  5.14564   5.75465  2.07345  …  10.1004   11.2958    4.06999
  5.73837  5.75465   8.92019  4.4074      11.2958   17.5095    8.6513
  3.00004  2.07345   4.4074   2.80005      4.06999   8.6513    5.49623
  3.97698  1.96315   3.09308  1.61707      4.77266   7.51965   3.9313
  1.96315  2.77359   3.10185  1.11763      6.74294   7.54099   2.71709
  ⋮                                    ⋱
  3.50245  2.42069   5.14549  3.26897      4.45124   9.46169   6.01108
  5.50374  2.7168    4.28051  2.23786      5.63632   8.88042   4.64271
  2.7168   3.83836   4.29265  1.54668      7.96314   8.90562   3.20878
  4.28051  4.29265   6.65397  3.28767      8.90562  13.8045    6.82067
  2.23786  1.54668   3.28767  2.08868  …   3.20878   6.82067   4.33322
  8.99176  4.43858   6.99329  3.65612      9.54312  15.0358    7.86079
  4.43858  6.27094   7.01314  2.5269      13.4828   15.0785    5.43293
  6.99329  7.01314  10.871    5.37125     15.0785   23.373    11.5484
  3.65612  2.5269    5.37125  3.4124       5.43293  11.5484    7.33678
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.42396  1.69016  2.66297  1.39221   …   1.29633     2.04246     1.0678
 0.0      2.30639  2.01307  0.408225      1.76897     1.54399     0.313103
 0.0      0.0      1.74063  1.42126       0.0         1.33503     1.09009
 0.0      0.0      0.0      0.569412      0.0         0.0         0.43673
 0.0      0.0      0.0      0.0           1.19407     1.88135     0.983575
 0.0      0.0      0.0      0.0       …   1.62943     1.4222      0.288405
 0.0      0.0      0.0      0.0           0.0         1.22973     1.0041
 0.0      0.0      0.0      0.0           0.0         0.0         0.402281
 0.0      0.0      0.0      0.0           0.212702    0.335126    0.175205
 0.0      0.0      0.0      0.0           0.290252    0.253338    0.0513739
 ⋮                                    ⋱
 0.0      0.0      0.0      0.0           0.0         0.0         0.379342
 0.0      0.0      0.0      0.0          -0.0703676  -0.110869   -0.0579627
 0.0      0.0      0.0      0.0          -0.0960236  -0.0838114  -0.0169959
 0.0      0.0      0.0      0.0          -0.0        -0.0724686  -0.0591723
 0.0      0.0      0.0      0.0       …  -0.0        -0.0        -0.0237067
 0.0      0.0      0.0      0.0           0.291029    0.458536    0.239724
 0.0      0.0      0.0      0.0           0.397138    0.34663     0.0702923
 0.0      0.0      0.0      0.0           0.0         0.299718    0.244727
 0.0      0.0      0.0      0.0           0.0         0.0         0.098047
julia> logdet(C)-23.803685736368465
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.6175      -3.47695     -1.58597     -7.37615  …    0.844972      3.92987
 -3.47695     88.017     -108.794      109.794        57.9632      -58.4962
 -1.58597   -108.794      158.584     -167.356       -84.4904       89.1639
 -7.37615    109.794     -167.356      204.961        89.1639     -109.199
  3.52817     -1.05593     -0.48165     -2.2401        0.0600974     0.279506
 -1.05593     26.7303     -33.0401      33.3439   …    4.12255      -4.16045
 -0.48165    -33.0401      48.1611     -50.8251       -6.00925       6.34165
 -2.2401      33.3439     -50.8251      62.2458        6.34165      -7.76665
  9.04922     -2.70831     -1.23536     -5.74551       0.917455      4.26698
 -2.70831     68.5592     -84.7429      85.5221       62.9354      -63.5141
  ⋮                                               ⋱
 -4.54977     67.7234    -103.229      126.425        92.6545     -113.474
 -1.84294      0.551565     0.251589     1.17011      -0.0863467    -0.401589
  0.551565   -13.9626      17.2585     -17.4172       -5.9232        5.97766
  0.251589    17.2585     -25.1569      26.5484        8.63398      -9.11157
  1.17011    -17.4172      26.5484     -32.514    …   -9.11157      11.159
 -6.18957      1.85245      0.844972     3.92987      -0.804921     -3.7436
  1.85245    -46.8938      57.9632     -58.4962      -55.2158       55.7235
  0.844972    57.9632     -84.4904      89.1639       80.4856      -84.9377
  3.92987    -58.4962      89.1639    -109.199       -84.9377      104.024

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.