Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  11.5279    -11.3548    -0.681179  …  -10.2201    -0.613105  -6.09288
 -11.3548     16.1289    -0.854603      14.517     -0.769198   1.39571
  -0.681179   -0.854603   1.7804        -0.769198   1.60247    2.99776
  -6.76938     1.55068    3.3306         1.39571    2.99776    9.85162
   7.89293    -7.7744    -0.466389      -8.30307   -0.498104  -4.95003
  -7.7744     11.0431    -0.585128  …   11.794     -0.624918   1.13392
  -0.466389   -0.585128   1.219         -0.624918   1.30189    2.43546
  -4.63485     1.06172    2.28039        1.13392    2.43546    8.00374
   6.19064    -6.09767   -0.365801      -8.0243    -0.48138   -4.78383
  -6.09767     8.6614    -0.458932      11.3981    -0.603937   1.09585
   ⋮                                ⋱
  -6.36595     1.45827    3.13211        1.66804    3.58266   11.7738
   8.679      -8.54867   -0.512837     -11.2671    -0.67592   -6.71712
  -8.54867    12.1429    -0.643402      16.0043    -0.848005   1.53871
  -0.512837   -0.643402   1.3404        -0.848005   1.76665    3.30489
  -5.09644     1.16746    2.5075    …    1.53871    3.30489   10.861
  10.3759    -10.2201    -0.613105     -16.5687    -0.993962  -9.87773
 -10.2201     14.517     -0.769198      23.5349    -1.24702    2.26272
  -0.613105   -0.769198   1.60247       -1.24702    2.59791    4.85994
  -6.09288     1.39571    2.99776        2.26272    4.85994   15.9714

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   7.865760357621699e-6
   0.00012816589015614754
   0.0016814634288020703
   0.003822939045482496
   0.00016583308589616827
   0.0027021094089942133
   0.0354501353387418
   0.08059866437337053
   0.0006487074126156292
   0.01057013679652864
   ⋮
  13.370524116745484
   0.031284413105262344
   0.5097529636487605
   6.687668341823235
  15.204938739231261
   0.4069949734492136
   6.63163771709009
  87.00330704812399
 197.80884549914757
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.138127   -0.107054      0.020968   …   0.0312014  -0.218073
  0.0804391  -0.090286      0.113445       0.168811    0.235113
 -0.120517   -0.181219     -0.0708265     -0.105393    0.0128798
  0.111275    0.0018835    -0.184744      -0.274908    0.114686
 -0.218256    0.169156     -0.0331317      0.023811   -0.16642
 -0.127102    0.142662     -0.179255   …   0.128826    0.179424
  0.190429    0.286346      0.111914      -0.0804298   0.00982905
 -0.175827   -0.00297613    0.291915      -0.209793    0.0875215
  0.221046   -0.171318      0.0335552      0.0217082  -0.151723
  0.128727   -0.144485      0.181546       0.117449    0.163578
  ⋮                                    ⋱
  0.0210899   0.000356978  -0.0350143     -0.312933    0.130549
 -0.133025    0.103099     -0.0201935      0.0307744  -0.215089
 -0.077468    0.0869512    -0.109254       0.166501    0.231896
  0.116065    0.174526      0.0682105     -0.103951    0.0127035
 -0.107165   -0.00181393    0.17792    …  -0.271146    0.113117
  0.220808   -0.171134      0.0335191      0.037312   -0.260782
  0.128589   -0.14433       0.181351       0.201872    0.281159
 -0.192656   -0.289694     -0.113222      -0.126034    0.0154022
  0.177883    0.00301093   -0.295329      -0.328748    0.137147

julia> logdet(E)
-74.0191624933685

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
   3.6188926463419744
   5.121510688722437
   8.01048321011603
   0.495472765176594
  -0.11233445342082331
  -5.67307152272965
 -11.15497537062587
   2.386207624364456
   3.433275595034983
  -3.725171179786487
   ⋮
   1.2518394144827096
 -11.020187723761955
  -6.02201295135513
  19.954589502530826
 -12.04922178338725
  -0.24226962120489384
   7.094543908027747
   1.7657771757482525
   2.340796748933226

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  25.4759    -9.45283  -11.4418   -16.6137   …  -4.04113  -4.89141   -7.10243
  -9.45283   24.101    -11.4143    18.5797      10.3033   -4.87968    7.94291
 -11.4418   -11.4143    23.5464     3.84387     -4.87968  10.0662     1.64327
 -16.6137    18.5797     3.84387   23.595        7.94291   1.64327   10.087
  10.9086    -4.04765   -4.8993    -7.11389     -3.76977  -4.56295   -6.62551
  -4.04765   10.3199    -4.88756    7.95573  …   9.61142  -4.55201    7.40955
  -4.8993    -4.88756   10.0824     1.64592     -4.55201   9.39025    1.53293
  -7.11389    7.95573    1.64592   10.1033       7.40955   1.53293    9.40965
  15.6639    -5.81208   -7.03499  -10.2149      -4.78667  -5.79383   -8.41276
  -5.81208   14.8185    -7.01812   11.4238      12.2041   -5.77994    9.4083
   ⋮                                         ⋱
  -6.51293    7.28366    1.50688    9.24977      4.1455    0.857642   5.26451
  12.6251    -4.68454   -5.6702    -8.23325     -4.58759  -5.55285   -8.06285
  -4.68454   11.9437    -5.65661    9.20755     11.6965   -5.53953    9.01699
  -5.6702    -5.65661   11.6689     1.90491     -5.53953  11.4274     1.86548
  -8.23325    9.20755    1.90491   11.693    …   9.01699   1.86548   11.451
  10.8911    -4.04113   -4.89141   -7.10243     -4.41582  -5.34495   -7.76097
  -4.04113   10.3033    -4.87968    7.94291     11.2586   -5.33213    8.67939
  -4.89141   -4.87968   10.0662     1.64327     -5.33213  10.9995     1.79564
  -7.10243    7.94291    1.64327   10.087        8.67939   1.79564   11.0223

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 5.04736  -1.87283  -2.26688  …  -0.800641   -0.969102   -1.40716
 0.0       4.53801  -3.45081      1.94002    -1.47524     1.16958
 0.0       0.0       2.54942      0.0         1.08989     0.976462
 0.0       0.0       0.0          0.0         0.0         0.103674
 0.0      -0.0      -0.0         -0.530905   -0.642612   -0.933086
 0.0       0.0      -0.0      …   1.28643    -0.978231    0.775548
 0.0       0.0       0.0          0.0         0.722705    0.647493
 0.0       0.0       0.0          0.0         0.0         0.0687466
 0.0      -0.0      -0.0         -0.149377   -0.180807   -0.262535
 0.0       0.0      -0.0          0.361952   -0.275238    0.21821
 ⋮                            ⋱
 0.0       0.0       0.0         -0.0        -0.0        -0.00470684
 0.0      -0.0      -0.0         -0.198375   -0.240115   -0.348652
 0.0       0.0      -0.0          0.48068    -0.365521    0.289787
 0.0       0.0       0.0          0.0         0.270042    0.241939
 0.0       0.0       0.0      …   0.0         0.0         0.0256875
 0.0      -0.0      -0.0         -0.0114145  -0.0138162  -0.0200614
 0.0       0.0      -0.0          0.0276583  -0.021032    0.0166743
 0.0       0.0       0.0          0.0         0.0155382   0.0139211
 0.0       0.0       0.0          0.0         0.0         0.00147806

julia> logdet(C)
-61.583716996791665

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   2467.59    12443.0        …   -28654.9          30974.5
  12443.0     69368.1                -1.58114e5        1.75431e5
   8780.31    48448.5                -1.10893e5        1.22394e5
  -9491.06   -53754.7                 1.22394e5       -1.36612e5
  -4494.13   -22661.9             53658.4         -58001.9
 -22661.9        -1.26337e5  …   296079.0             -3.28507e5
 -15991.2    -88237.2                 2.07655e5       -2.29192e5
  17285.7     97901.3                -2.29192e5        2.55815e5
   -344.835   -1738.85             4089.44         -4420.48
  -1738.85    -9693.87            22564.9         -25036.3
      ⋮                      ⋱
 -32843.7        -1.86018e5           4.30052e5  -480006.0
   5701.99    28752.7            -67501.0          72965.1
  28752.7         1.60292e5          -3.72461e5        4.13254e5
  20289.1         1.11952e5          -2.61225e5        2.88319e5
 -21931.5        -1.24214e5  …        2.88319e5       -3.2181e5
  -8053.1    -40608.3             96012.8             -1.03785e5
 -40608.3        -2.26386e5           5.29785e5       -5.87809e5
 -28654.9        -1.58114e5           3.71564e5  -410102.0
  30974.5         1.75431e5     -410102.0              4.57739e5

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.