Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  14.6758   -4.35411    8.69787   …  -1.72536    3.44662   -4.01069
  -4.35411   3.52613   -1.93783       1.39727   -0.767886   0.742745
   8.69787  -1.93783    5.5165       -0.767886   2.18597   -2.21399
 -10.1213    1.87439   -5.5872        0.742745  -2.21399    4.72426
  12.343    -3.662      7.31529      -2.33658    4.6676    -5.43148
  -3.662     2.96563   -1.6298    …   1.89225   -1.03991    1.00586
   7.31529  -1.6298     4.63962      -1.03991    2.96036   -2.9983
  -8.51248   1.57644   -4.69908       1.00586   -2.9983     6.39784
  12.9569   -3.84413    7.67912      -2.2864     4.56736   -5.31484
  -3.84413   3.11313   -1.71086       1.85162   -1.01758    0.984264
   ⋮                              ⋱
 -10.2517    1.89852   -5.65914       0.906974  -2.70353    5.76884
   9.99595  -2.96566    5.92427      -1.48723    2.97093   -3.45715
  -2.96566   2.40171   -1.31989       1.20442   -0.661905   0.640234
   5.92427  -1.31989    3.75738      -0.661905   1.88427   -1.90842
  -6.89381   1.27668   -3.80554   …   0.640234  -1.90842    4.07223
   5.81545  -1.72536    3.44662      -1.58225    3.16074   -3.67802
  -1.72536   1.39727   -0.767886      1.28137   -0.704193   0.681138
   3.44662  -0.767886   2.18597      -0.704193   2.00466   -2.03035
  -4.01069   0.742745  -2.21399       0.681138  -2.03035    4.3324

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   7.059409817189738e-5
   0.005073178909303394
   0.01103188238804387
   0.07211183605761068
   0.0003260451263468046
   0.023430928382087327
   0.050951730813167735
   0.3330549338735437
   0.001087595683228247
   0.07815904763833971
   ⋮
  12.993291116301545
   0.023515159632266136
   1.6898949768410751
   3.674761518555773
  24.020723830785695
   0.25317457077513156
  18.19415398863527
  39.56410182666792
 258.6176977179669
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0206209   -0.0060688    0.0115601   …  0.0521485  -0.0993349   0.203043
  0.00955546  -0.0265521   -0.0177747      0.22816     0.152736   -0.0555316
 -0.0248696   -0.0171411    0.00723221     0.147291   -0.0621456   0.118261
  0.00435942  -0.0108796    0.0255371      0.0934868  -0.219438   -0.164057
  0.276274    -0.0813082    0.15488        0.0586091  -0.111641    0.228198
  0.128022    -0.355739    -0.238141    …  0.256426    0.171658   -0.0624113
 -0.333196    -0.229652     0.0968953      0.165539   -0.0698447   0.132913
  0.0584064   -0.145761     0.34214        0.105069   -0.246624   -0.184382
  0.103203    -0.0303729    0.0578557      0.0653247  -0.124433    0.254345
  0.0478229   -0.132887    -0.0889583      0.285808    0.191327   -0.0695626
  ⋮                                     ⋱
 -0.0269184    0.0671788   -0.157686       0.127976   -0.300393   -0.224581
  0.0323899   -0.00953246   0.0181579      0.0515488  -0.0981925   0.200708
  0.0150091   -0.0417063   -0.0279193      0.225536    0.15098    -0.054893
 -0.0390634   -0.026924     0.0113599      0.145597   -0.0614309   0.116901
  0.00684748  -0.0170889    0.040112    …  0.0924116  -0.216914   -0.16217
 -0.376567     0.110825    -0.211104       0.0284734  -0.0542375   0.110863
 -0.174496     0.484879     0.324591       0.124577    0.0833948  -0.0303206
  0.454153     0.31302     -0.13207        0.0804219  -0.0339319   0.0645714
 -0.0796091    0.198676    -0.466344       0.0510444  -0.119814   -0.0895763

julia> logdet(E)
-53.89472525469099

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  0.3698596021980942
  1.2427817659013314
  2.910579104640197
  2.1920616570600764
 -2.8759172337325607
 -4.8562730243675745
 -0.6923530661689541
 -5.309807994741563
 -2.7901690812540894
  2.41962807223644
  ⋮
  3.373202704613318
 -0.466862436839057
 -1.687636114274151
 -0.6548027794817427
 -0.5600452823842796
 -1.7807812553190838
  3.2996430653912534
  6.910102899777014
  5.836825072819716

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3.34047   -1.67831     0.248367   …   -3.93365    0.582125   -4.95865
 -1.67831    2.93157     0.0589129       6.87104    0.13808     6.03834
  0.248367   0.0589129   1.33692         0.13808    3.13348    -0.889111
 -2.11564    2.5763     -0.379344        6.03834   -0.889111    6.3829
  6.28093   -3.15566     0.466994       -7.72733    1.14354    -9.74086
 -3.15566    5.5121      0.110771   …   13.4976     0.271248   11.8618
  0.466994   0.110771    2.51375         0.271248   6.15547    -1.74659
 -3.97794    4.84409    -0.713264       11.8618    -1.74659    12.5387
  5.51175   -2.76921     0.409804       -6.4898     0.9604     -8.18086
 -2.76921    4.83707     0.0972058      11.336      0.227808    9.96217
  ⋮                                 ⋱
 -2.77348    3.37737    -0.497298       11.6057    -1.70888    12.268
  4.56951   -2.29581     0.339747       -7.74016    1.14544    -9.75703
 -2.29581    4.01017     0.0805883      13.52       0.271698   11.8815
  0.339747   0.0805883   1.8288          0.271698   6.1657     -1.74949
 -2.89403    3.52418    -0.518914   …   11.8815    -1.74949    12.5595
  7.82942   -3.93365     0.582125      -11.3927     1.68596   -14.3614
 -3.93365    6.87104     0.13808        19.9001     0.399912   17.4884
  0.582125   0.13808     3.13348         0.399912   9.07527    -2.57506
 -4.95865    6.03834    -0.889111       17.4884    -2.57506    18.4863

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.8277  -0.918268  0.135891  …  -2.15224    0.318502    -2.71306
 0.0      1.44511   0.127116      3.38707    0.297935     2.4545
 0.0      0.0       1.14118       0.0        2.67471     -0.729451
 0.0      0.0       0.0           0.0        0.0          1.02125
 0.0     -0.0       0.0          -0.118017   0.017465    -0.14877
 0.0      0.0       0.0       …   0.185729   0.0163372    0.134591
 0.0      0.0       0.0           0.0        0.146667    -0.0399992
 0.0      0.0       0.0           0.0        0.0          0.0559996
 0.0     -0.0       0.0           0.016117  -0.0023851    0.0203167
 0.0      0.0       0.0          -0.025364  -0.00223108  -0.0183804
 ⋮                            ⋱
 0.0      0.0       0.0           0.0        0.0          0.20179
 0.0     -0.0       0.0          -0.518397   0.0767155   -0.653477
 0.0      0.0       0.0           0.815822   0.0717617    0.591198
 0.0      0.0       0.0           0.0        0.64424     -0.175698
 0.0      0.0       0.0       …   0.0        0.0          0.245981
 0.0     -0.0       0.0          -0.30521    0.0451669   -0.38474
 0.0      0.0       0.0           0.480322   0.0422503    0.348073
 0.0      0.0       0.0           0.0        0.379302    -0.103444
 0.0      0.0       0.0           0.0        0.0          0.144823

julia> logdet(C)
-4.205420958534319

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  52.536      -49.5714     18.0239    …   16.2093     -5.89358    -29.5007
 -49.5714     248.482     -82.6333       -81.2507     27.0201      93.2209
  18.0239     -82.6333     83.8374        27.0201    -27.4138     -33.9587
  90.2195    -285.09      103.853         93.2209    -33.9587    -124.518
 -13.1767      12.4331     -4.52061       -4.64495     1.68888      8.45377
  12.4331     -62.3225     20.7255    …   23.2834     -7.74293    -26.7136
  -4.52061     20.7255    -21.0275        -7.74293     7.85575      9.73126
 -22.6282      71.5042    -26.0477       -26.7136      9.73126     35.682
   0.588988    -0.555751    0.202067       0.610611   -0.222014    -1.11131
  -0.555751     2.78576    -0.926412      -3.06076     1.01786      3.51168
   ⋮                                  ⋱
  13.6105     -43.0086     15.6672        12.7074     -4.62909    -16.9737
  10.535       -9.9405      3.6143         4.06064    -1.47642     -7.39033
  -9.9405      49.8279    -16.5704       -20.3544      6.7689      23.3531
   3.6143     -16.5704     16.8118         6.7689     -6.86753     -8.50711
  18.0916     -57.1688     20.8255    …   23.3531     -8.50711    -31.1934
 -17.1787      16.2093     -5.89358       -6.20663     2.25669     11.296
  16.2093     -81.2507     27.0201        31.1115    -10.3462     -35.6949
  -5.89358     27.0201    -27.4138       -10.3462     10.4969      13.003
 -29.5007      93.2209    -33.9587       -35.6949     13.003       47.6787

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source