Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.82496   -1.24601     1.10805   …  -0.756115    0.672401    4.01919
 -1.24601    3.63871     2.14915       2.20808     1.30417    -2.73997
  1.10805    2.14915    13.1456        1.30417     7.97715   -12.4514
  6.62325   -4.51521   -20.5187       -2.73997   -12.4514     27.1198
  3.938     -0.842371    0.749107     -0.763042    0.678561    4.05601
 -0.842371   2.45997     1.45295   …   2.22831     1.31612    -2.76507
  0.749107   1.45295     8.88716       1.31612     8.05022   -12.5655
  4.47769   -3.05253   -13.8718       -2.76507   -12.5655     27.3682
  4.51748   -0.966326    0.859338     -0.994601    0.884483    5.28688
 -0.966326   2.82196     1.66675       2.90453     1.71552    -3.60418
  ⋮                                ⋱
  4.98462   -3.39812   -15.4423       -3.07151   -13.958      30.4013
  4.52131   -0.967145    0.860067     -0.977232    0.869037    5.19456
 -0.967145   2.82435     1.66816       2.85381     1.68556    -3.54124
  0.860067   1.66816    10.2036        1.68556    10.31      -16.0927
  5.14094   -3.50468   -15.9265    …  -3.54124   -16.0927     35.0506
  3.53477   -0.756115    0.672401     -0.947769    0.842836    5.03794
 -0.756115   2.20808     1.30417       2.76777     1.63474    -3.43447
  0.672401   1.30417     7.97715       1.63474     9.99913   -15.6075
  4.01919   -2.73997   -12.4514       -3.43447   -15.6075     33.9939

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   8.21457460529029e-6
   0.0008273691539091695
   0.002179476401990521
   0.014785586044919585
   0.0002207496038276809
   0.02223382484432844
   0.058568894362629624
   0.3973318666646579
   0.0014397135439867539
   0.1450074573542446
   ⋮
  20.22256721246186
   0.016498780518772852
   1.6617515494422606
   4.3774272594868755
  29.69650294970307
   0.21978290579928136
  22.13645934845389
  58.312411751906154
 395.5918864995072
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.111904    0.00383983   0.146069   …  0.00641532   0.244042   -0.0342661
 -0.0234645   0.181686    -0.020223      0.303549    -0.0337872   0.0300971
  0.126453    0.0204102    0.10751       0.0341       0.17962     0.132911
  0.0723219   0.0292018    0.0314727     0.0487885    0.0525824  -0.275647
 -0.0973083   0.003339     0.127017      0.00649954   0.247245   -0.0347159
 -0.0204041   0.157989    -0.0175853  …  0.307534    -0.0342307   0.0304922
  0.10996     0.0177481    0.0934874     0.0345476    0.181978    0.134655
  0.062889    0.0253931    0.0273677     0.0494289    0.0532727  -0.279265
  0.228838   -0.00785225  -0.298703      0.00826746   0.314498   -0.0441589
  0.0479837  -0.371539     0.0413549     0.391185    -0.0435417   0.0387864
  ⋮                                   ⋱
  0.190233    0.0768117    0.0827848     0.0582234    0.0627511  -0.328953
 -0.0719181   0.00246777   0.0938752     0.00815761   0.310319   -0.0435721
 -0.0150801   0.116766    -0.0129969     0.385988    -0.0429632   0.038271
  0.0812688   0.0131172    0.0690942     0.0433609    0.228402    0.169007
  0.0464797   0.0187674    0.0202268  …  0.0620386    0.0668628  -0.350507
  0.0904528  -0.00310376  -0.118069      0.00579672   0.22051    -0.0309619
  0.0189666  -0.146858     0.0163464     0.274279    -0.0305292   0.027195
 -0.102213   -0.0164977   -0.086901      0.0308118    0.1623      0.120095
 -0.0584584  -0.0236041   -0.0254396     0.044084     0.0475121  -0.249067

julia> logdet(E)
-54.98633543213958

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
   4.124459752389865
  -0.26476179787526843
  -5.611478719753868
  -4.803802565618542
  -1.825795991538653
  -1.1294835222001982
   2.4601676731128443
   0.18195882410434966
  -3.6613049944986598
   2.084245345624615
   ⋮
  -4.688604190157852
  -8.146125540513609
  -3.9993856689025065
   8.89018543061493
   4.195831996711126
   5.191730920002242
   3.3860830295327586
 -10.249009731450196
  -4.020734091187463

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.2512     5.59437   0.29037   -3.37619   …   5.91329   0.306924  -3.56866
  5.59437   14.1016   -3.88646    3.43827      14.9056   -4.10802    3.63428
  0.29037   -3.88646   3.44402    0.949778     -4.10802   3.64036    1.00392
 -3.37619    3.43827   0.949778   7.49218       3.63428   1.00392    7.9193
  9.69221    4.8192    0.250136  -2.90838       6.92307   0.359335  -4.17806
  4.8192    12.1477   -3.34794    2.96185   …  17.4509   -4.80952    4.25488
  0.250136  -3.34794   2.96681    0.818175     -4.80952   4.26201    1.17536
 -2.90838    2.96185   0.818175   6.45404       4.25488   1.17536    9.27163
 12.7224     6.32586   0.328338  -3.81765       9.43765   0.489852  -5.69561
  6.32586   15.9455   -4.39463    3.88784      23.7893   -6.55642    5.80033
  ⋮                                         ⋱
 -2.7003     2.74995   0.759639   5.99229       3.60783   0.996617   7.86166
 10.9594     5.44928   0.28284   -3.28863       7.57939   0.393401  -4.57415
  5.44928   13.7359   -3.78567    3.3491       19.1053   -5.26547    4.65825
  0.28284   -3.78567   3.35471    0.925147     -5.26547   4.66605    1.28678
 -3.28863    3.3491    0.925147   7.29788   …   4.65825   1.28678   10.1506
 11.8926     5.91329   0.306924  -3.56866      10.9267    0.567141  -6.59426
  5.91329   14.9056   -4.10802    3.63428      27.5428   -7.5909     6.71551
  0.306924  -4.10802   3.64036    1.00392      -7.5909    6.72675    1.85507
 -3.56866    3.63428   1.00392    7.9193        6.71551   1.85507   14.6335

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.35428  1.66783   0.086567  -1.00653   …   1.76291    0.0915021  -1.06391
 0.0      3.36452  -1.19804    1.52087       3.55633   -1.26634     1.60757
 0.0      0.0       1.41465    2.02098       0.0        1.49529     2.1362
 0.0      0.0       0.0        0.285747      0.0        0.0         0.302037
 0.0      0.0       0.0       -0.0           0.771401   0.0400388  -0.465539
 0.0      0.0      -0.0        0.0       …   1.55615   -0.554117    0.703429
 0.0      0.0       0.0        0.0           0.0        0.654299    0.934742
 0.0      0.0       0.0        0.0           0.0        0.0         0.132163
 0.0      0.0       0.0       -0.0           0.674996   0.035035   -0.407359
 0.0      0.0      -0.0        0.0           1.36167   -0.484867    0.615519
 ⋮                                       ⋱
 0.0      0.0       0.0        0.0          -0.0       -0.0        -0.0569594
 0.0      0.0       0.0       -0.0           0.070193   0.0036433  -0.0423614
 0.0      0.0      -0.0        0.0           0.141601  -0.0504214   0.064008
 0.0      0.0       0.0        0.0           0.0        0.0595374   0.0850561
 0.0      0.0       0.0        0.0       …   0.0        0.0         0.0120261
 0.0      0.0       0.0       -0.0           0.422874   0.0219489  -0.255204
 0.0      0.0      -0.0        0.0           0.853066  -0.303761    0.385613
 0.0      0.0       0.0        0.0           0.0        0.35868     0.512416
 0.0      0.0       0.0        0.0           0.0        0.0         0.0724505

julia> logdet(C)
-21.476926562690903

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  130.705    -152.96     -227.173    …   -76.2279  -113.212     78.6864
 -152.96      181.298     268.837         90.3501   133.975    -92.7974
 -227.173     268.837     403.481        133.975    201.075   -137.99
  157.894    -186.209    -276.893        -92.7974  -137.99      96.59
    5.60075    -6.55441    -9.73447       10.0814    14.9726   -10.4065
   -6.55441     7.7687     11.5198   …   -11.9491   -17.7186    12.2727
   -9.73447    11.5198     17.2894       -17.7186   -26.5928    18.2496
    6.76581    -7.97913   -11.865         12.2727    18.2496   -12.7743
 -120.93      141.521     210.184        182.862    271.583   -188.76
  141.521    -167.74     -248.732       -216.74    -321.391    222.61
    ⋮                                ⋱
   92.8843   -109.541    -162.888       -107.825   -160.336    112.232
   33.1119    -38.75      -57.5507        20.8521    30.9691   -21.5246
  -38.75       45.929      68.1054       -24.7152   -36.6488    25.3847
  -57.5507     68.1054    102.216        -36.6488   -55.004     37.747
   39.9998    -47.173     -70.1464   …    25.3847    37.747    -26.4221
   65.1368    -76.2279   -113.212       -150.348   -223.294    155.197
  -76.2279     90.3501    133.975        178.202    264.246   -183.029
 -113.212     133.975     201.075        264.246    396.592   -272.165
   78.6864    -92.7974   -137.99        -183.029   -272.165    190.51

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source