Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  9.89384    1.98374     0.0830888  …  1.83095     0.076689   -3.62144
  1.98374    0.632165    0.0725547     0.583473    0.0669663   0.141378
  0.0830888  0.0725547   9.60947       0.0669663   8.86932     1.91921
 -3.92365    0.153176    2.07937       0.141378    1.91921     7.63553
  6.20626    1.24437     0.0521204     1.46759     0.0614699  -2.90275
  1.24437    0.396548    0.0455125  …  0.467681    0.0536766   0.113321
  0.0521204  0.0455125   6.02788       0.0536766   7.10918     1.53834
 -2.46125    0.0960853   1.30436       0.113321    1.53834     6.12024
 10.1943     2.044       0.0856125     2.39683     0.100391   -4.7407
  2.044      0.651365    0.0747584     0.763804    0.0876632   0.185073
  ⋮                                 ⋱
 -2.45063    0.0956709   1.29873       0.131373    1.78338     7.09515
 10.3424     2.07368     0.0868559     2.6105      0.10934    -5.1633
  2.07368    0.660826    0.0758442     0.831894    0.095478    0.201572
  0.0868559  0.0758442  10.0451        0.095478   12.6455      2.73634
 -4.10154    0.160121    2.17365    …  0.201572    2.73634    10.8865
  9.13178    1.83095     0.076689      2.30995     0.096752   -4.56886
  1.83095    0.583473    0.0669663     0.736118    0.0844856   0.178365
  0.076689   0.0669663   8.86932       0.0844856  11.1897      2.4213
 -3.62144    0.141378    1.91921       0.178365    2.4213      9.6331
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.00015829179266351945
   0.008654137942560706
   0.017625940351010984
   0.02444485691754225
   0.001338760322127702
   0.07319278090651286
   0.14907222405575396
   0.20674353338621687
   0.003082027320812128
   0.1685007739709297
   ⋮
   7.2266747501427
   0.06233314323226177
   3.4078811721595286
   6.940854266767528
   9.626050358607568
   0.8077833533042347
  44.163177695884386
  89.9474379707458
 124.74524521854931
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0525716    0.102946    0.0925032  …  -0.142111   -0.127694   -0.203035
 -0.199588     0.0531655   0.0216544     -0.0733914  -0.0298924  -0.02831
 -0.00556682  -0.071442    0.18511        0.0986208  -0.255532    0.0889363
  0.030353     0.158186    0.0161233     -0.218365   -0.0222571   0.181815
 -0.0668351   -0.130877   -0.117601      -0.11986    -0.107701   -0.171245
  0.253739    -0.0675902  -0.0275296  …  -0.0619001  -0.025212   -0.0238774
  0.00707719   0.0908253  -0.235333       0.0831793  -0.215522    0.0750111
 -0.0385883   -0.201105   -0.0204978     -0.184175   -0.0187722   0.153348
 -0.108144    -0.211769   -0.190287      -0.188846   -0.169689   -0.269807
  0.410569    -0.109366   -0.0445448     -0.0975275  -0.0397231  -0.0376203
  ⋮                                   ⋱
  0.0256361    0.133604    0.0136177     -0.194527   -0.0198274   0.161967
  0.153264     0.300123    0.269678      -0.204251   -0.183531   -0.291816
 -0.581866     0.154995    0.0631298     -0.105483   -0.0429633  -0.040689
 -0.0162291   -0.208277    0.539658       0.141744   -0.367267    0.127825
  0.0884892    0.461166    0.0470049  …  -0.313849   -0.0319894   0.261317
 -0.120988    -0.23692    -0.212886      -0.173485   -0.155886   -0.24786
  0.459329    -0.122354   -0.0498351     -0.0895943  -0.0364919  -0.0345602
  0.0128114    0.164416   -0.42601        0.120394   -0.311947    0.108571
 -0.0698541   -0.364048   -0.037106      -0.266575   -0.0271709   0.221955
julia> logdet(E)-36.00837648736635
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
   0.14296540482798023
  -2.0432514724484454
   1.832537220483241
  -0.5354681274147037
   0.24846473660182244
  -6.393622858625096
   0.5198910457588821
  -1.2497744309153467
  -1.4458339206476096
  10.059377945171443
   ⋮
  -1.6089580125703402
   3.355577997832633
 -13.516064628290241
   8.340682451924467
  -2.6463241511171685
   0.8033119854802637
  -4.506937923924675
  -2.6049629335392517
   0.8373117617452481

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  9.65887  -3.19998  -3.62178    5.96723   …  -2.15139  -2.43498    4.01186
 -3.19998  16.1793   -2.71976    7.05512      10.8776   -1.82853    4.74326
 -3.62178  -2.71976   3.00938    0.303123     -1.82853   2.02325    0.203794
  5.96723   7.05512   0.303123  54.9714        4.74326   0.203794  36.9581
  7.28154  -2.41237  -2.73036    4.49852      -1.80325  -2.04094    3.36265
 -2.41237  12.1971   -2.05035    5.31865   …   9.11737  -1.53263    3.97569
 -2.73036  -2.05035   2.26868    0.228515     -1.53263   1.69584    0.170815
  4.49852   5.31865   0.228515  41.4413        3.97569   0.170815  30.9774
  7.57825  -2.51067  -2.84161    4.68182      -1.84247  -2.08533    3.43578
 -2.51067  12.6941   -2.13389    5.53537       9.31565  -1.56597    4.06216
  ⋮                                        ⋱
  4.21771   4.98665   0.214251  38.8545        4.05037   0.174024  31.5593
  5.44597  -1.80425  -2.04207    3.36451      -1.56481  -1.77108    2.91802
 -1.80425   9.12241  -1.53348    3.97789       7.91182  -1.32998    3.45001
 -2.04207  -1.53348   1.69678    0.17091      -1.32998   1.47161    0.148229
  3.36451   3.97789   0.17091   30.9946    …   3.45001   0.148229  26.8814
  6.4938   -2.15139  -2.43498    4.01186      -1.96272  -2.22144    3.66002
 -2.15139  10.8776   -1.82853    4.74326       9.92366  -1.66817    4.32728
 -2.43498  -1.82853   2.02325    0.203794     -1.66817   1.84581    0.185921
  4.01186   4.74326   0.203794  36.9581        4.32728   0.185921  33.7169
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.10787  -1.02964  -1.16536   1.92004  …  -0.69224    -0.783488    1.29087
 0.0       3.88834  -1.00805   2.32286      2.61419    -0.677729    1.56169
 0.0       0.0       0.796962  6.12603      0.0         0.53581     4.11862
 0.0       0.0       0.0       2.89152      0.0         0.0         1.94401
 0.0      -0.0      -0.0       0.0         -0.116187   -0.131503    0.216663
 0.0       0.0      -0.0       0.0      …   0.438772   -0.113752    0.262118
 0.0       0.0       0.0       0.0          0.0         0.0899317   0.691281
 0.0       0.0       0.0       0.0          0.0         0.0         0.326288
 0.0      -0.0      -0.0       0.0         -0.0367774  -0.0416252   0.0685814
 0.0       0.0      -0.0       0.0          0.138887   -0.0360065   0.0829697
 ⋮                                      ⋱
 0.0       0.0       0.0       0.0          0.0         0.0         0.42502
 0.0      -0.0      -0.0       0.0          0.127883    0.14474    -0.238473
 0.0       0.0      -0.0       0.0         -0.482941    0.125202   -0.288504
 0.0       0.0       0.0       0.0         -0.0        -0.0989845  -0.760867
 0.0       0.0       0.0       0.0      …  -0.0        -0.0        -0.359133
 0.0      -0.0      -0.0       0.0         -0.103722   -0.117394    0.193418
 0.0       0.0      -0.0       0.0          0.391698   -0.101548    0.233996
 0.0       0.0       0.0       0.0          0.0         0.0802831   0.617115
 0.0       0.0       0.0       0.0          0.0         0.0         0.291281
julia> logdet(C)-18.35074995189946
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  280.561    161.109    488.68    …   134.42     407.726   -44.9101
  161.109    100.618    288.051        83.9499   240.333   -26.691
  488.68     288.051    892.331       240.333    744.509   -79.2094
  -53.8269   -31.9905   -94.9364      -26.691    -79.2094   10.3047
 -132.674    -76.1866  -231.091       -63.135   -191.503    21.0936
  -76.1866   -47.5812  -136.216   …   -39.43    -112.881    12.5364
 -231.091   -136.216   -421.974      -112.881   -349.685    37.2035
   25.4542    15.1279    44.8944       12.5364    37.2035   -4.83997
  142.779     81.989    248.691        69.6148   211.158   -23.2585
   81.989     51.205    146.591        43.4769   124.466   -13.823
    ⋮                             ⋱
   37.5844    22.3372    66.289        22.2984    66.1737   -8.60883
  185.488    106.514    323.083        88.5052   268.456   -29.5698
  106.514     66.522    190.44         55.2745   158.241   -17.574
  323.083    190.44     589.95        158.241    490.202   -52.1533
  -35.5868   -21.15     -62.7656  …   -17.574    -52.1533    6.78485
  234.083    134.42     407.726       153.746    466.345   -51.3668
  134.42      83.9499   240.333        96.0194   274.886   -30.5284
  407.726    240.333    744.509       274.886    851.548   -90.5975
  -44.9101   -26.691    -79.2094      -30.5284   -90.5975   11.7862

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.