Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.8779   -5.67319   2.51165  -4.19741  …  -4.33321   1.91841  -3.20601
 -5.67319  10.3916   -1.82625   4.63486      7.93719  -1.3949    3.54013
  2.51165  -1.82625   6.66265  -4.90347     -1.3949    5.08897  -3.7453
 -4.19741   4.63486  -4.90347   4.85459      3.54013  -3.7453    3.70796
  4.9541   -4.08635   1.80912  -3.02336     -8.47598   3.75251  -6.27111
 -4.08635   7.48501  -1.31543   3.33845  …  15.5255   -2.72849   6.92468
  1.80912  -1.31543   4.79905  -3.53193     -2.72849   9.95428  -7.32599
 -3.02336   3.33845  -3.53193   3.49672      6.92468  -7.32599   7.25296
  7.37073  -6.07969   2.69162  -4.49817     -8.87189   3.92779  -6.56403
 -6.07969  11.1362   -1.9571    4.96697     16.2507   -2.85593   7.24813
  ⋮                                      ⋱
 -2.74817   3.03458  -3.21044   3.17844      5.31471  -5.62272   5.56667
  4.92835  -4.06511   1.79972  -3.00765     -6.83217   3.02476  -5.05491
 -4.06511   7.44611  -1.30859   3.3211      12.5146   -2.19933   5.58173
  1.79972  -1.30859   4.77411  -3.51358     -2.19933   8.02378  -5.90522
 -3.00765   3.3211   -3.51358   3.47855  …   5.58173  -5.90522   5.84635
  5.25338  -4.33321   1.91841  -3.20601     -7.35063   3.2543   -5.4385
 -4.33321   7.93719  -1.3949    3.54013     13.4642   -2.36623   6.0053
  1.91841  -1.3949    5.08897  -3.7453      -2.36623   8.63266  -6.35333
 -3.20601   3.54013  -3.7453    3.70796      6.0053   -6.35333   6.28999

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   5.169873942314028e-6
   0.001606665441360394
   0.004436824292251839
   0.012734003732129923
   0.00021146287743445557
   0.06571728848624436
   0.18147901515190873
   0.5208577811574281
   0.0004264800467270259
   0.1325391605582126
   ⋮
  16.744925742067014
   0.009550577134949913
   2.96807666859718
   8.196376374005764
  23.524187676036693
   0.09358120911473201
  29.0826616515261
  80.3120901073069
 230.50145504918083
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0251341   0.125048     0.0205776  …  -0.191933   -0.0315842   0.116984
 -0.0287736   0.0767266   -0.0835508     -0.117766    0.128241   -0.144029
  0.0743801  -0.029567    -0.114387       0.0453818   0.17557     0.0842108
  0.124535    0.0101492    0.0448615     -0.0155779  -0.0688572  -0.107184
 -0.0466363  -0.232025    -0.0381817     -0.303813   -0.049995    0.185175
  0.0533894  -0.142366     0.155028   …  -0.186413    0.202993   -0.227986
 -0.138012    0.0548614    0.212244       0.0718354   0.277912    0.133298
 -0.231075   -0.0188318   -0.0832404     -0.0246583  -0.108995   -0.169662
 -0.0257646  -0.128184    -0.0210938     -0.338803   -0.0557529   0.206502
  0.0294954  -0.0786511    0.0856465     -0.207882    0.226372   -0.254243
  ⋮                                   ⋱
 -0.131362   -0.0107055   -0.0473206     -0.0188956  -0.0835224  -0.130012
  0.0134679   0.0670058    0.0110264     -0.262811   -0.0432478   0.160184
 -0.0154181   0.0411133   -0.04477       -0.161255    0.175598   -0.197217
  0.039856   -0.0158432   -0.0612933      0.0621406   0.240406    0.115308
  0.0667312   0.00543837   0.0240387  …  -0.0213305  -0.094285   -0.146765
  0.14586     0.725686     0.119418      -0.27034    -0.0444867   0.164773
 -0.166981    0.445265    -0.484868      -0.165875    0.180628   -0.202867
  0.431648   -0.171585    -0.663819       0.0639207   0.247293    0.118612
  0.722713    0.0588987    0.260344      -0.0219415  -0.096986   -0.150969

julia> logdet(E)
-53.598323713329364

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  1.176197016290541
 -0.7442195660537447
 -0.17552040486833753
  4.419576548941513
 -3.4531256560578947
 -1.7183286522802188
 -1.7508427315092427
 -1.056501957552384
  1.019383125967273
 -0.21876798175382794
  ⋮
 -1.5453425349880792
 -1.9802028284309068
  1.4267485649493117
 -4.489808630322814
 -8.020909531729265
  5.145311271966731
  4.716428392799441
 -0.9994745344213711
 -3.556652993715528

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.09564   4.01347  -5.31983   5.56031  …   2.57278  -3.4102    3.56435
  4.01347   5.40741  -6.70785   4.15906      3.46634  -4.29997   2.66611
 -5.31983  -6.70785   8.81601  -4.94273     -4.29997   5.65137  -3.16846
  5.56031   4.15906  -4.94273   8.6691       2.66611  -3.16846   5.5572
  3.57157   2.35158  -3.11701   3.2579       2.62574  -3.4804    3.63772
  2.35158   3.16832  -3.93028   2.43689  …   3.53769  -4.38848   2.72099
 -3.11701  -3.93028   5.16549  -2.89605     -4.38848   5.7677   -3.23368
  3.2579    2.43689  -2.89605   5.07941      2.72099  -3.23368   5.67159
  5.13204   3.37902  -4.47888   4.68133      3.04375  -4.03448   4.21685
  3.37902   4.55261  -5.64748   3.5016       4.10089  -5.08713   3.15417
  ⋮                                      ⋱
  3.96305   2.96433  -3.52287   6.17881      3.00752  -3.5742    6.26883
  4.50118   2.96366  -3.92831   4.10588      2.18201  -2.89224   3.02298
  2.96366   3.99298  -4.95326   3.07116      2.93986  -3.64687   2.26117
 -3.92831  -4.95326   6.50998  -3.64984     -3.64687   4.79302  -2.68722
  4.10588   3.07116  -3.64984   6.40149  …   2.26117  -2.68722   4.71314
  3.90752   2.57278  -3.4102    3.56435      3.21615  -4.26299   4.45569
  2.57278   3.46634  -4.29997   2.66611      4.33317  -5.37526   3.33282
 -3.4102   -4.29997   5.65137  -3.16846     -5.37526   7.06461  -3.9608
  3.56435   2.66611  -3.16846   5.5572       3.33282  -3.9608    6.94688

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.46893  1.62559  -2.15471  2.25211   …   1.04206   -1.38124    1.44368
 0.0      1.66279  -1.92759  0.299533      1.06591   -1.23566    0.192011
 0.0      0.0       0.67648  0.720327      0.0        0.433648   0.461755
 0.0      0.0       0.0      1.72873       0.0        0.0        1.10818
 0.0      0.0      -0.0      0.0           0.753103  -0.998233   1.04336
 0.0      0.0      -0.0      0.0       …   0.770338  -0.893016   0.138768
 0.0      0.0       0.0      0.0           0.0        0.3134     0.333713
 0.0      0.0       0.0      0.0           0.0        0.0        0.800888
 0.0      0.0      -0.0      0.0           0.115171  -0.152659   0.159559
 0.0      0.0      -0.0      0.0           0.117807  -0.136568   0.0212216
 ⋮                                     ⋱
 0.0      0.0       0.0      0.0           0.0        0.0        0.0489074
 0.0      0.0      -0.0      0.0          -0.3718     0.492819  -0.515096
 0.0      0.0      -0.0      0.0          -0.380309   0.440874  -0.0685084
 0.0      0.0       0.0      0.0          -0.0       -0.154723  -0.164751
 0.0      0.0       0.0      0.0       …  -0.0       -0.0       -0.395391
 0.0      0.0      -0.0      0.0           0.243219  -0.322385   0.336958
 0.0      0.0      -0.0      0.0           0.248785  -0.288405   0.0448159
 0.0      0.0       0.0      0.0           0.0        0.101214   0.107775
 0.0      0.0       0.0      0.0           0.0        0.0        0.258651

julia> logdet(C)
-43.353560345612905

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  10.741        8.08624     9.69594   …     3.91768     4.69756   -2.53894
   8.08624     68.2776     52.2635         33.0796     25.321     -3.94607
   9.69594     52.2635     44.1311         25.321      21.3809    -2.97049
  -5.24047     -8.14485    -6.1312         -3.94607    -2.97049    2.78967
 -20.4362     -15.3851    -18.4478        -47.8896    -57.4229    31.036
 -15.3851    -129.907     -99.4383    …  -404.365    -309.524     48.2368
 -18.4478     -99.4383    -83.9652       -309.524    -261.36      36.3112
   9.97068     15.4966     11.6654         48.2368     36.3112   -34.1009
   0.292698     0.220354    0.264219        5.2957      6.3499    -3.432
   0.220354     1.8606      1.42421        44.7152     34.2275    -5.33408
   ⋮                                  ⋱
  -1.86416     -2.89732    -2.18101        -7.90766    -5.95265    5.59031
  20.2773      15.2655     18.3044         37.6381     45.1306   -24.3922
  15.2655     128.897      98.6653        317.804     243.265    -37.9109
  18.3044      98.6653     83.3125        243.265     205.412    -28.5382
  -9.89317    -15.3762    -11.5747    …   -37.9109    -28.5382    26.8011
   5.20388      3.91768     4.69756        20.9916     25.1704   -13.6041
   3.91768     33.0796     25.321         177.247     135.675    -21.1438
   4.69756     25.321      21.3809        135.675     114.563    -15.9164
  -2.53894     -3.94607    -2.97049       -21.1438    -15.9164    14.9476

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source