Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.5106    -3.54536   1.57171    1.655    …  -2.89142   1.28181   1.34974
 -3.54536    6.04523   3.06866   -3.95595      4.9302    2.50265  -3.22628
  1.57171    3.06866  25.028      2.47498      2.50265  20.4117    2.01847
  1.655     -3.95595   2.47498    4.36274     -3.22628   2.01847   3.55804
  5.32633   -1.79663   0.796475   0.83868     -2.28183   1.01157   1.06517
 -1.79663    3.06346   1.55506   -2.0047   …   3.89077   1.97502  -2.54609
  0.796475   1.55506  12.6831     1.25421      1.97502  16.1083    1.59292
  0.83868   -2.0047    1.25421    2.21084     -2.54609   1.59292   2.8079
  8.6761    -2.92655   1.29739    1.36613     -2.40332   1.06543   1.12189
 -2.92655    4.99009   2.53305   -3.26547      4.09793   2.08018  -2.68165
  ⋮                                        ⋱
  1.31602   -3.14568   1.96804    3.46915     -3.86217   2.41631   4.25933
  8.83533   -2.98026   1.3212     1.39121     -3.0745    1.36297   1.4352
 -2.98026    5.08167   2.57954   -3.3254       5.24236   2.66111  -3.43055
  1.3212     2.57954  21.0388     2.08048      2.66111  21.704     2.14627
  1.39121   -3.3254    2.08048    3.66736  …  -3.43055   2.14627   3.78332
  8.57197   -2.89142   1.28181    1.34974     -3.98775   1.76783   1.86151
 -2.89142    4.9302    2.50265   -3.22628      6.79956   3.45156  -4.44957
  1.28181    2.50265  20.4117     2.01847      3.45156  28.151     2.7838
  1.34974   -3.22628   2.01847    3.55804     -4.44957   2.7838    4.90712
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0005925854826549074
   0.009696688739911757
   0.022064711153804855
   0.04135578274376744
   0.0041541559916462
   0.06797594407403101
   0.15467853115952823
   0.28991323227249377
   0.008630351927150679
   0.14122154322537164
   ⋮
  13.275131427530981
   0.3083702188209144
   5.045972464881009
  11.482056185776402
  21.520763075514655
   3.014637961431969
  49.32960194124632
 112.24897976622857
 210.3877267217795
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0681212   0.295516     0.356834    …   0.190854    0.230455    0.0256729
 -0.315897    0.228952    -0.256101        0.147865   -0.165399    0.035862
  0.0770209  -0.0244781   -0.0165677      -0.0158088  -0.0107      0.298816
 -0.332441   -0.283784     0.166398       -0.183277    0.107465    0.0298924
 -0.0715668   0.310464     0.374883        0.177761    0.214646    0.0239117
 -0.331876    0.240532    -0.269055    …   0.137721   -0.154052    0.0334018
  0.0809166  -0.0257162   -0.0174057      -0.0147243  -0.00996596  0.278317
 -0.349256   -0.298138     0.174815       -0.170704    0.100093    0.0278418
  0.0638292  -0.276897    -0.334352        0.176489    0.213109    0.0237405
  0.295994   -0.214526     0.239965        0.136735   -0.152949    0.0331627
  ⋮                                    ⋱
 -0.0475714  -0.0406086    0.0238111      -0.209398    0.122782    0.0341528
 -0.0209984   0.091093     0.109994        0.196002    0.236671    0.0263653
 -0.0973755   0.0705745   -0.0789433       0.151853   -0.16986     0.0368293
  0.0237417  -0.00754539  -0.00510702     -0.0162352  -0.0109886   0.306875
 -0.102475   -0.0874765    0.0512923   …  -0.18822     0.110364    0.0306987
 -0.0200173   0.086837     0.104855        0.210508    0.254188    0.0283167
 -0.0928259   0.0672771   -0.0752549       0.163092   -0.182431    0.0395551
  0.0226325  -0.00719285  -0.0048684      -0.0174368  -0.0118019   0.329588
 -0.0976874  -0.0833894    0.0488958      -0.202151    0.118532    0.0329707
julia> logdet(E)-13.323261510331474
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -8.065699415632075
 -2.651649227009828
  2.5525403255291135
  0.9313918944842289
 -5.4542610760776045
  1.3788697388295015
  2.449771905438374
  3.0304627782416027
  8.944824616922618
 -0.25334276787860655
  ⋮
 -2.293223151440917
 -0.7430887382867861
  7.855516408008503
 -1.9912266011562205
  7.855576035809567
 -1.844193553064968
  1.122116551909373
  0.34797424807809435
  0.7409333113259706

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.13504   -5.72768   1.43151   0.926222  …   -4.3178    1.07914   0.69823
 -5.72768    9.11905   3.47909  -3.29898        6.87437   2.6227   -2.48693
  1.43151    3.47909   9.10539  -3.27528        2.6227    6.86407  -2.46906
  0.926222  -3.29898  -3.27528   2.80183       -2.48693  -2.46906   2.11215
  6.52299   -6.08988   1.52203   0.984792      -6.71323   1.67783   1.0856
 -6.08988    9.69569   3.69909  -3.50759   …   10.6881    4.07773  -3.86663
  1.52203    3.69909   9.68117  -3.4824         4.07773  10.6721   -3.83885
  0.984792  -3.50759  -3.4824    2.97901       -3.86663  -3.83885   3.28394
  4.9762    -4.64579   1.16112   0.75127       -7.99492   1.99816   1.29286
 -4.64579    7.39657   2.82193  -2.67584       12.7287    4.85625  -4.60485
  ⋮                                        ⋱
  0.894093  -3.18454  -3.16167   2.70464       -5.0223   -4.98622   4.26545
  6.18712   -5.77631   1.44367   0.934086      -8.96257   2.24      1.44934
 -5.77631    9.19647   3.50863  -3.32699       14.2693    5.44401  -5.16218
  1.44367    3.50863   9.1827   -3.30309        5.44401  14.2479   -5.1251
  0.934086  -3.32699  -3.30309   2.82562   …   -5.16218  -5.1251    4.38425
  4.62488   -4.3178    1.07914   0.69823      -11.6478    2.91111   1.88356
 -4.3178     6.87437   2.6227   -2.48693       18.5444    7.07503  -6.70877
  1.07914    2.6227    6.86407  -2.46906        7.07503  18.5166   -6.66058
  0.69823   -2.48693  -2.46906   2.11215       -6.70877  -6.66058   5.69777
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.4769  -2.31244  0.577945   0.373944  …  -1.74323   0.435682    0.281897
 0.0      1.94208  2.47958   -1.25343       1.46403   1.86923    -0.944893
 0.0      0.0      1.61958   -0.236741      0.0       1.22091    -0.178466
 0.0      0.0      0.0        1.01728       0.0       0.0         0.766877
 0.0     -0.0      0.0        0.0          -1.11481   0.278622    0.180275
 0.0      0.0      0.0       -0.0       …   0.93626   1.19539    -0.604268
 0.0      0.0      0.0       -0.0           0.0       0.780785   -0.114131
 0.0      0.0      0.0        0.0           0.0       0.0         0.490424
 0.0     -0.0      0.0        0.0          -2.03073   0.507537    0.328389
 0.0      0.0      0.0       -0.0           1.70549   2.17751    -1.10073
 ⋮                                      ⋱
 0.0      0.0      0.0        0.0           0.0       0.0         0.415997
 0.0     -0.0      0.0        0.0          -0.362082  0.0904947   0.0585523
 0.0      0.0      0.0       -0.0           0.304091  0.388254   -0.196262
 0.0      0.0      0.0       -0.0           0.0       0.253594   -0.037069
 0.0      0.0      0.0        0.0       …   0.0       0.0         0.159287
 0.0     -0.0      0.0        0.0          -0.274686  0.0686518   0.0444194
 0.0      0.0      0.0       -0.0           0.230692  0.29454    -0.14889
 0.0      0.0      0.0       -0.0           0.0       0.192383   -0.0281215
 0.0      0.0      0.0        0.0           0.0       0.0         0.120839
julia> logdet(C)-14.65420545193327
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 173.175    148.221    -71.9155   …  102.421    -49.6935      22.9452
 148.221    153.992    -59.6127      106.408    -41.1923      43.2781
 -71.9155   -59.6127    56.7778      -41.1923    39.2334      13.7893
  33.2058    62.6312    19.9555       43.2781    13.7893      94.3342
   8.92734    7.64095   -3.70731       6.78631   -3.29265      1.52033
   7.64095    7.93844   -3.07309  …    7.05052   -2.72937      2.86757
  -3.70731   -3.07309    2.92695      -2.72937    2.59957      0.913665
   1.71179    3.2287     1.02873       2.86757    0.913665     6.2505
  99.6707    85.3086   -41.3909       59.6643   -28.9485      13.3665
  85.3086    88.6299   -34.31         61.9872   -23.9962      25.2113
   ⋮                              ⋱
 -36.4626   -68.7741   -21.9128      -48.6173   -15.4904    -105.972
 -63.9643   -54.7473    26.5628      -38.345     18.6046      -8.59037
 -54.7473   -56.8788    22.0187      -39.8379    15.4219     -16.2028
  26.5628    22.0187   -20.9716       15.4219   -14.6885      -5.16252
 -12.265    -23.1336    -7.37082  …  -16.2028    -5.16252    -35.3175
 119.664    102.421    -49.6935       74.3538   -36.0757      16.6574
 102.421    106.408    -41.1923       77.2487   -29.9042      31.4184
 -49.6935   -41.1923    39.2334      -29.9042    28.4821      10.0105
  22.9452    43.2781    13.7893       31.4184    10.0105      68.4833

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.