Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 33.4877      8.78831   -0.0316079  …   8.01977  -0.0288438   8.03425
  8.78831     4.22609   -2.10239        3.85651  -1.91854     1.22455
 -0.0316079  -2.10239   10.5717        -1.91854   9.64725    -2.45713
  8.80417     1.3419    -2.6926         1.22455  -2.45713     5.82567
 31.0906      8.15921   -0.0293453      8.98818  -0.0323267   9.00441
  8.15921     3.92357   -1.9519     …   4.3222   -2.15021     1.37242
 -0.0293453  -1.9519     9.81499       -2.15021  10.8122     -2.75384
  8.17394     1.24584   -2.49985        1.37242  -2.75384     6.52914
 27.9859      7.34445   -0.0264149      8.15448  -0.0293283   8.1692
  7.34445     3.53177   -1.75698        3.92129  -1.95076     1.24512
  ⋮                                 ⋱
  4.55851     0.694792  -1.39414        1.10986  -2.227       5.28005
 28.0443      7.35978   -0.0264701     10.3062   -0.037067   10.3248
  7.35978     3.53914   -1.76065        4.95599  -2.46551     1.57367
 -0.0264701  -1.76065    8.85332       -2.46551  12.3976     -3.15765
  7.37306     1.12378   -2.25492    …   1.57367  -3.15765     7.48655
 30.5592      8.01977   -0.0288438     11.8586   -0.0426505  11.88
  8.01977     3.85651   -1.91854        5.70252  -2.83689     1.81071
 -0.0288438  -1.91854    9.64725       -2.83689  14.2651     -3.63329
  8.03425     1.22455   -2.45713        1.81071  -3.63329     8.61426

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0019849403655375335
   0.014739996469405001
   0.041660178750326585
   0.1381641770933298
   0.014962809459506012
   0.11111253639389394
   0.31404133217941294
   1.0415044662649577
   0.022380610697227122
   0.1661964905282781
   ⋮
  16.093113200465947
   0.37325002162038784
   2.771722564773357
   7.8338185313983875
  25.980519607843757
   5.5185736236887
  40.98045318721433
 115.82451926582146
 384.1269978100441
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.178867    -0.0427733   -0.0854986   …   0.0216893   0.0433542   0.256803
 -0.427482    -0.30164      0.070691        0.152954   -0.0358457   0.0696915
 -0.154426     0.0866036   -0.515397       -0.0439146   0.261345   -0.0127557
 -0.242966     0.444182     0.140261       -0.225234   -0.0711228   0.0745432
 -0.172268     0.0411953    0.0823443       0.0239263   0.0478257   0.28329
  0.411711     0.290512    -0.068083    …   0.16873    -0.0395427   0.0768794
  0.148729    -0.0834086    0.496382       -0.0484439   0.2883     -0.0140713
  0.234002    -0.427795    -0.135086       -0.248464   -0.0784584   0.0822315
 -0.0246542    0.00589568   0.0117847       0.0232057   0.0463853   0.274758
  0.0589222    0.0415767   -0.00974372      0.163648   -0.0383518   0.074564
  ⋮                                     ⋱
 -0.150733     0.275566     0.0870162      -0.192505   -0.0607878   0.0637111
  0.0894634   -0.0213939   -0.0427637       0.0269956   0.0539608   0.31963
 -0.213813    -0.150871     0.0353574       0.190374   -0.0446153   0.0867415
 -0.0772389    0.0433164   -0.257785       -0.0546583   0.325283   -0.0158764
 -0.121524     0.222166     0.070154    …  -0.280337   -0.0885231   0.0927802
  0.00646773  -0.00154666  -0.00309159      0.0269473   0.0538643   0.319058
 -0.0154576   -0.0109072    0.00255615      0.190034   -0.0445355   0.0865864
 -0.00558396   0.00313155  -0.0186365      -0.0545605   0.324701   -0.015848
 -0.00878553   0.0160614    0.00507177     -0.279836   -0.0883647   0.0926142

julia> logdet(E)
-1.5351387062130755

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  1.852550848898623
  1.4391223664161066
  0.0867731698981547
 -8.018030385411285
 -0.8014546424208877
 -2.969110859038955
  1.690578148429183
  5.905786727471093
 -4.435720860092377
  8.257199465018703
  ⋮
 -1.6693294430800305
  1.6417212478725873
 -0.5904621318061591
 -1.861041138936809
 -5.575521758874894
  0.5394523916195695
 -1.9751866417502608
 -0.8862112120695864
 -0.08373137720945424

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.80042   6.544     2.5644    -1.45965   …   5.86976   2.30019   -1.30926
  6.544    19.2867   -2.21391    3.03645      17.2995   -1.98581    2.7236
  2.5644   -2.21391   5.90918   -0.764277     -1.98581   5.30034   -0.685532
 -1.45965   3.03645  -0.764277   3.52148       2.7236   -0.685532   3.15865
  7.02286   6.75805   2.64828   -1.5074        8.99569   3.52515   -2.00651
  6.75805  19.9175   -2.28632    3.13577   …  26.5124   -3.04334    4.17405
  2.64828  -2.28632   6.10246   -0.789276     -3.04334   8.12304   -1.05061
 -1.5074    3.13577  -0.789276   3.63666       4.17405  -1.05061    4.84079
  7.06166   6.79539   2.66292   -1.51572       8.98029   3.51912   -2.00307
  6.79539  20.0276   -2.29896    3.1531       26.467    -3.03813    4.16691
  ⋮                                        ⋱
 -1.55289   3.23041  -0.813096   3.74642       4.0363   -1.01594    4.68103
  6.04827   5.82021   2.28077   -1.29821       6.73013   2.63734   -1.50117
  5.82021  17.1535   -1.96904    2.70061      19.8353   -2.27688    3.12282
  2.28077  -1.96904   5.2556    -0.679745     -2.27688   6.07725   -0.786015
 -1.29821   2.70061  -0.679745   3.13199   …   3.12282  -0.786015   3.62164
  6.09976   5.86976   2.30019   -1.30926       8.24509   3.23101   -1.83908
  5.86976  17.2995   -1.98581    2.7236       24.3002   -2.78941    3.82577
  2.30019  -1.98581   5.30034   -0.685532     -2.78941   7.44525   -0.962948
 -1.30926   2.7236   -0.685532   3.15865       3.82577  -0.962948   4.43687

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.60776  2.50943   0.983373  -0.559734  2.69306  …   0.882055    -0.502063
 0.0      3.60409  -1.29897    1.23223   0.0         -1.16514      1.10527
 0.0      0.0       1.80411    0.768682  0.0          1.61823      0.689483
 0.0      0.0       0.0        1.04829   0.0          0.0          0.940282
 0.0      0.0       0.0       -0.0       2.25111      0.510737    -0.29071
 0.0      0.0      -0.0        0.0       0.0      …  -0.674652     0.639987
 0.0      0.0       0.0        0.0       0.0          0.937006     0.399232
 0.0      0.0       0.0        0.0       0.0          0.0          0.544453
 0.0      0.0       0.0       -0.0       0.0          0.370134    -0.210679
 0.0      0.0      -0.0        0.0       0.0         -0.488924     0.463802
 ⋮                                                ⋱
 0.0      0.0       0.0        0.0       0.0         -0.0         -0.12338
 0.0      0.0       0.0       -0.0       0.0          0.0521662   -0.0296929
 0.0      0.0      -0.0        0.0       0.0         -0.0689084    0.0653677
 0.0      0.0       0.0        0.0       0.0          0.095705     0.0407772
 0.0      0.0       0.0        0.0       0.0      …   0.0          0.0556099
 0.0      0.0       0.0       -0.0       0.0          0.00803799  -0.0045752
 0.0      0.0      -0.0        0.0       0.0         -0.0106177    0.0100721
 0.0      0.0       0.0        0.0       0.0          0.0147466    0.00628313
 0.0      0.0       0.0        0.0       0.0          0.0          0.0085686

julia> logdet(C)
-30.633243106877924

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3808.28   -1971.49   -2024.1    …   3975.16    4081.22   -5724.7
 -1971.49    1245.98    1108.86      -2512.29   -2235.81    3328.72
 -2024.1     1108.86    1729.87      -2235.81   -3487.95    2862.53
  2839.19   -1650.89   -1419.68       3328.72    2862.53   -6718.42
  5519.39   -2857.32   -2933.55       5769.28    5923.21   -8308.46
 -2857.32    1805.82    1607.09   …  -3646.17   -3244.91    4831.09
 -2933.55    1607.09    2507.12      -3244.91   -5062.19    4154.49
  4114.88   -2392.67   -2057.57       4831.09    4154.49   -9750.67
 -3713.37    1922.36    1973.65      -3859.33   -3962.3     5557.91
  1922.36   -1214.93   -1081.23       2439.09    2170.67   -3231.74
     ⋮                            ⋱
 -2330.58    1355.15    1165.36      -2735.97   -2352.79    5522.05
  1485.27    -768.905   -789.42       1548.26    1589.57   -2229.69
  -768.905    485.946    432.468      -978.499   -870.816   1296.49
  -789.42     432.468    674.667      -870.816  -1358.51    1114.91
  1107.32    -643.867   -553.693  …   1296.49    1114.91   -2616.72
 -7678.69    3975.16    4081.22      -8058.74   -8273.75   11605.6
  3975.16   -2512.29   -2235.81       5093.11    4532.61   -6748.24
  4081.22   -2235.81   -3487.95       4532.61    7071.05   -5803.14
 -5724.7     3328.72    2862.53      -6748.24   -5803.14   13620.1

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source