Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  32.387    -12.1209      -9.68086    …  -6.65126    -5.3123     -3.05305
 -12.1209    15.1261      -0.0746638      8.30033    -0.0409712   0.484243
  -9.68086   -0.0746638    7.02939       -0.0409712   3.85733     1.30304
  -5.56372    0.882459     2.37458        0.484243    1.30304     2.24113
  31.8258   -11.9109      -9.51311       -7.44383    -5.94532    -3.41685
 -11.9109    14.864       -0.0733701  …   9.2894     -0.0458534   0.541946
  -9.51311   -0.0733701    6.90758       -0.0458534   4.31697     1.45831
  -5.46731    0.867168     2.33344        0.541946    1.45831     2.50819
  33.4938   -12.5351     -10.0117        -8.89544    -7.10471    -4.08317
 -12.5351    15.643       -0.0772152     11.1009     -0.0547952   0.647631
   ⋮                                  ⋱
  -4.5391     0.719944     1.93728        0.559522    1.5056      2.58953
  27.4028   -10.2555      -8.191         -6.80543    -5.43544    -3.12382
 -10.2555    12.7982      -0.0631733      8.49273    -0.0419209   0.495468
  -8.191     -0.0631733    5.94759       -0.0419209   3.94674     1.33324
  -4.70748    0.746652     2.00914    …   0.495468    1.33324     2.29308
  17.7722    -6.65126     -5.3123        -7.29481    -5.8263     -3.34845
  -6.65126    8.30033     -0.0409712      9.10343    -0.0449354   0.531097
  -5.3123    -0.0409712    3.85733       -0.0449354   4.23054     1.42911
  -3.05305    0.484243     1.30304        0.531097    1.42911     2.45797

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.004048554960788153
   0.005551939802679471
   0.023720474296766395
   0.0824648410960016
   0.058222673958185256
   0.0798430017864818
   0.3411265141489604
   1.1859351305956731
   0.09978096335616388
   0.1368334893245323
   ⋮
  18.899575875407635
   2.0382997156863127
   2.795199134237774
  11.942393393032017
  41.518038852898385
  18.44850408640191
  25.299146270512786
 108.089743435296
 375.77678274815867
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0780294   -0.029203     0.0366782   …  -0.0508906   0.0639174  -0.277382
 -0.0726397   -0.0158357    0.150761       -0.0275961   0.262725    0.127658
 -0.148974     0.0325457   -0.0907799       0.0567158  -0.158198    0.0804071
 -0.008003    -0.177369    -0.0361564      -0.309093   -0.063008    0.0490263
  0.0108564    0.00406308  -0.00510313     -0.0540552   0.067892   -0.294631
  0.0101065    0.00220326  -0.0209758   …  -0.0293122   0.279062    0.135596
  0.0207272   -0.00452816   0.0126304       0.0602426  -0.168035    0.0854071
  0.00111348   0.0246778    0.00503053     -0.328313   -0.0669261   0.0520749
 -0.130711    -0.0489193    0.0614415      -0.0590107   0.074116   -0.321641
 -0.121682    -0.0265271    0.252548       -0.0319993   0.304645    0.148027
  ⋮                                     ⋱
 -0.0155891   -0.345498    -0.0704292      -0.276802   -0.0564256   0.0439046
 -0.0886191   -0.0331662    0.041656       -0.0469507   0.0589689  -0.255907
 -0.082498    -0.0179848    0.171222       -0.0254596   0.242384    0.117774
 -0.169192     0.0369626   -0.1031          0.0523249  -0.14595     0.0741819
 -0.00908913  -0.201441    -0.0410633   …  -0.285162   -0.0581299   0.0452306
  0.18048      0.0675456   -0.0848356      -0.0339562   0.0426482  -0.18508
  0.168013     0.0366275   -0.348706       -0.0184132   0.1753      0.0851782
  0.344573    -0.0752772    0.209971        0.037843   -0.105556    0.0536508
  0.0185107    0.410249     0.0836286      -0.206239   -0.0420414   0.0327123

julia> logdet(E)
9.194235271681082

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -0.5943557260272144
  1.747433854659022
 -2.0108322029536803
  3.4040632626853844
 -0.2285026784211441
  0.6102691282048097
 -0.229130803406691
 -1.1234256745639748
  1.2516109521800105
  1.809055286124922
  ⋮
  1.2410895353064781
  0.34862096168408746
  1.027044485814057
  0.7986540750957328
  3.629886074447538
  2.3345236718263687
  0.14024748669656237
  1.903145942101335
 -2.664177292979527

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  17.3129   -17.3083   -3.2112   -8.11345  …  -17.9722   -3.33439   -8.42469
 -17.3083    22.4052    3.39545  11.9891       23.2647    3.5257    12.449
  -3.2112     3.39545   1.9212    3.23759       3.5257    1.9949     3.36178
  -8.11345   11.9891    3.23759  12.0281       12.449     3.36178   12.4895
  16.8225   -16.8181   -3.12026  -7.88366     -19.1442   -3.55183   -8.97407
 -16.8181    21.7706    3.29929  11.6496   …   24.7818    3.75562   13.2608
  -3.12026    3.29929   1.86679   3.14589       3.75562   2.12499    3.58101
  -7.88366   11.6496    3.14589  11.6875       13.2608    3.58101   13.304
  13.178    -13.1745   -2.44426  -6.17569     -15.3891   -2.85514   -7.21382
 -13.1745    17.0541    2.58451   9.12571      19.9209    3.01896   10.6597
   ⋮                                       ⋱
  -5.49803    8.12435   2.19393   8.1508        8.35491   2.25619    8.3821
  11.0928   -11.0898   -2.0575   -5.19849     -13.9116   -2.58102   -6.52122
 -11.0898    14.3556    2.17555   7.68172      18.0083    2.72911    9.63631
  -2.0575     2.17555   1.23096   2.0744        2.72911   1.54418    2.60223
  -5.19849    7.68172   2.0744    7.70673  …    9.63631   2.60223    9.66767
  17.977    -17.9722   -3.33439  -8.42469     -24.6459   -4.57255  -11.553
 -17.9722    23.2647    3.5257   12.449        31.9036    4.83491   17.0717
  -3.33439    3.5257    1.9949    3.36178       4.83491   2.73567    4.61012
  -8.42469   12.449     3.36178  12.4895       17.0717    4.61012   17.1273

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 4.16087  -4.15977  -0.771762   …  -4.31934   -0.801367   -2.02474
 0.0       2.25865   0.0819517      2.3453     0.0850954   1.78273
 0.0       0.0       1.14842        0.0        1.19247     1.43943
 0.0       0.0       0.0            0.0        0.0         1.90237
 0.0      -0.0      -0.0           -1.07998   -0.200369   -0.506255
 0.0       0.0       0.0        …   0.586405   0.0212768   0.445744
 0.0       0.0       0.0            0.0        0.29816     0.359907
 0.0       0.0       0.0            0.0        0.0         0.475658
 0.0      -0.0      -0.0           -0.901055  -0.167173   -0.42238
 0.0       0.0       0.0            0.489251   0.0177517   0.371894
 ⋮                              ⋱
 0.0       0.0       0.0            0.0        0.0         0.0241887
 0.0      -0.0      -0.0            0.857589   0.159108    0.402005
 0.0       0.0       0.0           -0.46565   -0.0168954  -0.353954
 0.0       0.0       0.0           -0.0       -0.236761   -0.285793
 0.0       0.0       0.0        …  -0.0       -0.0        -0.377708
 0.0      -0.0      -0.0           -1.17891   -0.218724   -0.552629
 0.0       0.0       0.0            0.640121   0.0232258   0.486574
 0.0       0.0       0.0            0.0        0.325472    0.392875
 0.0       0.0       0.0            0.0        0.0         0.519229

julia> logdet(C)
-37.75546194387149

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  270.945    235.308    227.417   …  -16.038    -15.5002      7.7014
  235.308    281.511    185.129      -19.1872   -12.618      11.703
  227.417    185.129    959.302      -12.618    -65.3838     19.7208
 -112.994   -171.705   -289.34        11.703     19.7208    -16.3374
 -138.612   -120.38    -116.343        7.30135    7.0565     -3.50609
 -120.38    -144.017    -94.7093  …    8.73501    5.74437    -5.32782
 -116.343    -94.7093  -490.764        5.74437   29.7661     -8.97794
   57.806     87.8415   148.022       -5.32782   -8.97794     7.43765
  -29.7096   -25.8019   -24.9366       1.09018    1.05362    -0.523502
  -25.8019   -30.8682   -20.2997       1.30425    0.857706   -0.79551
    ⋮                             ⋱
  125.283    190.378    320.808      -13.7675   -23.1997     19.2195
 -227.613   -197.675   -191.046       15.2782    14.7658     -7.33652
 -197.675   -236.489   -155.522       18.2781    12.0202    -11.1485
 -191.046   -155.522   -805.881       12.0202    62.286     -18.7864
   94.9229   144.244    243.066   …  -11.1485   -18.7864     15.5634
  -18.467    -16.038    -15.5002       3.64125    3.51914    -1.74852
  -16.038    -19.1872   -12.618        4.35622    2.86477    -2.65703
  -15.5002   -12.618    -65.3838       2.86477   14.8446     -4.47737
    7.7014    11.703     19.7208      -2.65703   -4.47737     3.70922

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.