Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 14.3805     1.83012   -6.07472   2.76396  …   1.36953   -4.54589   2.06835
  1.83012    2.32701   -3.6012    4.61406      1.74137   -2.69488   3.45283
 -6.07472   -3.6012     7.93328  -7.10073     -2.69488    5.9367   -5.31368
  2.76396    4.61406   -7.10073   9.39265      3.45283   -5.31368   7.02879
 11.3795     1.4482    -4.80701   2.18716      0.951309  -3.15768   1.43672
  1.4482     1.8414    -2.84968   3.65117  …   1.2096    -1.87193   2.39842
 -4.80701   -2.84968    6.27771  -5.61891     -1.87193    4.12377  -3.69101
  2.18716    3.65117   -5.61891   7.43253      2.39842   -3.69101   4.88236
 10.1049     1.286     -4.26861   1.94219      0.922346  -3.06154   1.39298
  1.286      1.63515   -2.53051   3.24222      1.17277   -1.81494   2.3254
  ⋮                                        ⋱
  1.57103    2.62263   -4.03605   5.33877      2.65943   -4.09269   5.4137
  7.4309     0.945689  -3.13902   1.42823      0.763272  -2.53353   1.15274
  0.945689   1.20245   -1.86087   2.38425      0.970505  -1.50192   1.92434
 -3.13902   -1.86087    4.09941  -3.6692      -1.50192    3.30866  -2.96144
  1.42823    2.38425   -3.6692    4.85351  …   1.92434   -2.96144   3.9173
 10.7613     1.36953   -4.54589   2.06835      1.76018   -5.84255   2.65832
  1.36953    1.74137   -2.69488   3.45283      2.23808   -3.46357   4.43772
 -4.54589   -2.69488    5.9367   -5.31368     -3.46357    7.63008  -6.82935
  2.06835    3.45283   -5.31368   7.02879      4.43772   -6.82935   9.03367

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.00013555068432121549
   0.003728734166486713
   0.0394667096897656
   0.08928032869919747
   0.0008817410926683531
   0.024254972630287447
   0.2567262563087778
   0.5807579281154469
   0.0018777930278524593
   0.051654412927579205
   ⋮
  10.584792836092863
   0.025961530669398446
   0.7141509237377387
   7.558915686497411
  17.099537367185498
   0.21786628156860507
   5.993075224063079
  63.43358079610294
 143.4974181673873
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0165603    0.10273     0.327354   …  -0.0870769  -0.277475   -0.225001
  0.387633     0.114799   -0.109191      -0.0973068   0.0925534  -0.0969781
 -0.0107876    0.352842    0.0824028     -0.299079   -0.069847    0.202452
 -0.194424     0.200553   -0.250155      -0.169995    0.212038   -0.185419
  0.0154101   -0.0955949  -0.304618      -0.0738384  -0.23529    -0.190794
 -0.360711    -0.106826    0.101607   …  -0.082513    0.0784822  -0.0822343
  0.0100383   -0.328335   -0.0766795     -0.253609   -0.059228    0.171673
  0.18092     -0.186624    0.23278       -0.14415     0.179801   -0.157229
  0.0112224   -0.0696169  -0.221838      -0.0671239  -0.213894   -0.173444
 -0.262687    -0.0777956   0.0739953     -0.0750097   0.0713455  -0.0747563
  ⋮                                   ⋱
 -0.0604014    0.0623055  -0.0777152     -0.117608    0.146696   -0.128279
 -0.00288715   0.0179101   0.0570716     -0.0517794  -0.164998   -0.133795
  0.0675807    0.0200142  -0.0190365     -0.0578625   0.0550359  -0.057667
 -0.00188072   0.0615151   0.0143662     -0.177844   -0.0415338   0.120386
 -0.0338962    0.0349648  -0.0436124  …  -0.101085    0.126086   -0.110257
  0.00554045  -0.0343697  -0.109521      -0.0810783  -0.25836    -0.209501
 -0.129688    -0.0384075   0.0365313     -0.0906035   0.0861775  -0.0902975
  0.00360912  -0.118048   -0.0275689     -0.278476   -0.0650353   0.188506
  0.0650471   -0.0670977   0.0836925     -0.158284    0.197431   -0.172646

julia> logdet(E)
-49.48902587132086

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
   6.56997456745564
 -11.677031364758282
   8.115415318806361
   7.956519351253912
  -5.129640420451925
   2.842368831187889
  -4.35267704606537
  -0.2863152460942619
  -1.2604261456211623
  -7.5410144757631254
   ⋮
   1.6089599055596073
  -2.6488771263161603
   1.9381464680499823
  -7.551062868863613
  -6.932916400816925
  -4.285669741079694
  -2.590922527026588
  -5.549377708614967
   0.3654663088696453

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.88877   0.547225  -3.82371    4.19008  …  0.41268   -2.88358    3.15987
  0.547225  3.008      0.611095   1.95161     2.26843    0.460846   1.47177
 -3.82371   0.611095   4.06448   -4.56967     0.460846   3.06515   -3.44614
  4.19008   1.95161   -4.56967   19.082       1.47177   -3.44614   14.3904
  3.5925    0.402127  -2.80984    3.07907     0.47311   -3.30584    3.62258
  0.402127  2.21042    0.449061   1.43414  …  2.6006     0.528329   1.68729
 -2.80984   0.449061   2.98677   -3.35801     0.528329   3.51399   -3.95076
  3.07907   1.43414   -3.35801   14.0224      1.68729   -3.95076   16.4976
  3.25497   0.364345  -2.54585    2.78978     0.328992  -2.29882    2.51908
  0.364345  2.00274    0.40687    1.29939     1.80841    0.36739    1.17331
  ⋮                                        ⋱
  3.54875   1.6529    -3.87024   16.1614      1.44222   -3.37694   14.1014
  3.66613   0.410368  -2.86743    3.14217     0.504899  -3.52796    3.86599
  0.410368  2.25572    0.458264   1.46353     2.77534    0.563828   1.80066
 -2.86743   0.458264   3.04798   -3.42683     0.563828   3.7501    -4.21622
  3.14217   1.46353   -3.42683   14.3098   …  1.80066   -4.21622   17.6061
  3.68678   0.41268   -2.88358    3.15987     0.643098  -4.49362    4.92417
  0.41268   2.26843    0.460846   1.47177     3.535      0.718157   2.29353
 -2.88358   0.460846   3.06515   -3.44614     0.718157   4.77656   -5.37027
  3.15987   1.47177   -3.44614   14.3904      2.29353   -5.37027   22.4252

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.21106  0.247495  -1.72936    1.89506   …  -1.30417      1.42912
 0.0      1.71661    0.605323   0.863677      0.456493     0.651327
 0.0      0.0        0.841055  -2.15828       0.634267    -1.62763
 0.0      0.0        0.0        3.17595       0.0          2.39509
 0.0      0.0       -0.0        0.0          -0.698368     0.765281
 0.0      0.0        0.0        0.0       …   0.244447     0.348779
 0.0      0.0        0.0       -0.0           0.339643    -0.87158
 0.0      0.0        0.0        0.0           0.0          1.28254
 0.0      0.0       -0.0        0.0           0.0122909   -0.0134686
 0.0      0.0        0.0        0.0          -0.00430216  -0.00613834
 ⋮                                        ⋱
 0.0      0.0        0.0        0.0          -0.0         -0.23448
 0.0      0.0       -0.0        0.0          -0.40154      0.440014
 0.0      0.0        0.0        0.0           0.14055      0.200537
 0.0      0.0        0.0       -0.0           0.195285    -0.501132
 0.0      0.0        0.0        0.0       …   0.0          0.737424
 0.0      0.0       -0.0        0.0          -0.601952     0.659628
 0.0      0.0        0.0        0.0           0.2107       0.300627
 0.0      0.0        0.0       -0.0           0.292753    -0.751251
 0.0      0.0        0.0        0.0           0.0          1.10548

julia> logdet(C)
-29.223289788273874

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  25.337     -11.9799     28.4085   …   -1.04873     2.48689   0.215771
 -11.9799     13.5313    -16.2887        1.18454    -1.42592  -0.232338
  28.4085    -16.2887     39.2908       -1.42592     3.43954   0.42344
   2.46481    -2.65406     4.83708      -0.232338    0.42344   0.165009
  -5.27636     2.49479    -5.91599      22.674     -53.7677   -4.66506
   2.49479    -2.81786     3.39207  …  -25.6102     30.829     5.02324
  -5.91599     3.39207    -8.18219      30.829     -74.3642   -9.15496
  -0.51329     0.5527     -1.00731       5.02324    -9.15496  -3.56757
  -9.6768      4.57542   -10.8499        6.62534   -15.7109   -1.36313
   4.57542    -5.16793     6.22103      -7.4833      9.00822   1.46779
   ⋮                                ⋱
  -2.76696     2.97941    -5.43003       2.00189    -3.64848  -1.42177
 -21.7588     10.2881    -24.3965       14.9819    -35.5271   -3.08244
  10.2881    -11.6203     13.9883      -16.922      20.3703    3.31911
 -24.3965     13.9883    -33.7419       20.3703    -49.1363   -6.04915
  -2.11672     2.27924    -4.15396  …    3.31911    -6.04915  -2.35728
   2.21801    -1.04873     2.48689      -5.20062    12.3324    1.07
  -1.04873     1.18454    -1.42592       5.87408    -7.07108  -1.15215
   2.48689    -1.42592     3.43954      -7.07108    17.0565    2.09983
   0.215771   -0.232338    0.42344      -1.15215     2.09983   0.818275

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source