Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.75794    0.923155  -6.893     …   0.585921  -4.37495   -3.59663
  0.923155  31.4222    -0.689184     19.9435    -0.437422   5.74276
 -6.893     -0.689184  13.6529       -0.437422   8.66542   -1.38859
 -5.6667     9.04808   -2.18781       5.74276   -1.38859   14.9702
  4.23117    0.445998  -3.33017       0.294503  -2.19899   -1.80778
  0.445998  15.1808    -0.332961  …  10.0242    -0.219862   2.8865
 -3.33017   -0.332961   6.59604      -0.219862   4.35552   -0.697952
 -2.73772    4.37134   -1.05698       2.8865    -0.697952   7.5245
  6.57788    0.693359  -5.17717       0.538328  -4.01958   -3.30448
  0.693359  23.6004    -0.51763      18.3235    -0.40189    5.27629
  ⋮                               ⋱
 -3.62267    5.78435   -1.39865       3.82589   -0.925096   9.9733
  5.63785    0.594273  -4.43731       0.358818  -2.67921   -2.20257
  0.594273  20.2277    -0.443656     12.2134    -0.267876   3.51686
 -4.43731   -0.443656   8.78894      -0.267876   5.3067    -0.850373
 -3.64789    5.82462   -1.40839   …   3.51686   -0.850373   9.16772
  5.55862    0.585921  -4.37495       0.560015  -4.18151   -3.43761
  0.585921  19.9435    -0.437422     19.0617    -0.418081   5.48885
 -4.37495   -0.437422   8.66542      -0.418081   8.28229   -1.3272
 -3.59663    5.74276   -1.38859       5.48885   -1.3272    14.3083
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0009322976663595677
   0.008165011651593699
   0.009874935506664188
   0.017942714761814575
   0.010948843455627282
   0.09588936839856194
   0.11597060348661475
   0.21071808091420569
   0.07453575985618954
   0.6527800826345292
   ⋮
   9.744319761733582
   0.5954252591286864
   5.2147016493376945
   6.306769012834662
  11.459371799398165
   9.224075726756489
  80.78394755425896
  97.70179224622935
 177.52373054073198
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.058548    -0.0198812   0.0400801  …  -0.11979     0.241494    0.0319508
 -0.00784463   0.0260596   0.032398       0.157016    0.195207   -0.364885
  0.0382438    0.0562824  -0.0284803      0.339117   -0.171602    0.0242749
  0.0224871   -0.0348652  -0.0446148     -0.210072   -0.268817   -0.251762
 -0.290502     0.0986463  -0.198868      -0.0603918   0.121748    0.0161079
  0.0389233   -0.129302   -0.160752   …   0.0791594   0.0984132  -0.183956
 -0.189757    -0.279261    0.141313       0.170965   -0.0865127   0.0122381
 -0.111576     0.172993    0.221369      -0.105907   -0.135523   -0.126925
  0.334574    -0.113612    0.229038      -0.0979809   0.197527    0.0261337
 -0.0448283    0.148918    0.185139       0.12843     0.159667   -0.298453
  ⋮                                   ⋱
  0.110158    -0.170794   -0.218555      -0.146236   -0.187129   -0.175257
 -0.298197     0.101259   -0.204136      -0.0765622   0.154347    0.0204209
  0.0399543   -0.132727   -0.16501        0.100355    0.124764   -0.233211
 -0.194783    -0.286658    0.145056       0.216742   -0.109677    0.015515
 -0.114532     0.177575    0.227232   …  -0.134265   -0.171811   -0.16091
 -0.156732     0.0532219  -0.107294      -0.0870974   0.175586    0.0232309
  0.021       -0.0697612  -0.0867292      0.114164    0.141932   -0.265302
 -0.102378    -0.150667    0.0762416      0.246567   -0.124769    0.0176499
 -0.0601978    0.0933337   0.119433      -0.15274    -0.195453   -0.183052
julia> logdet(E)-4.134685440383264
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  3.8164839637966006
 -0.43837320181059547
 -2.4798717446177228
  0.5786531342861745
  1.1390026451389836
 -1.2824064858288207
 10.40375430965448
  0.4537293316761708
 -1.4762528130032866
  2.1485497199741497
  ⋮
  1.2203361679673213
 -1.0925309624644248
 -0.9115651423444597
  9.268287763394797
 -0.775023487312084
  2.6780701771629003
 -2.22545750364652
  3.0187309781460914
 -1.0688715444054244

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 12.322       2.89326   0.73291   …    2.13256   0.540212    3.62548
  2.89326    31.1555   -6.09912       22.964    -4.49553   -14.1639
  0.73291    -6.09912  11.9912        -4.49553   8.83845    -0.319246
  4.91872   -19.2163   -0.433123     -14.1639   -0.319246   13.8241
  9.66366     2.26908   0.574794       2.14392   0.543089    3.64479
  2.26908    24.4341   -4.78331   …   23.0863   -4.51947   -14.2394
  0.574794   -4.78331   9.40425       -4.51947   8.88552    -0.320946
  3.85756   -15.0706   -0.339682     -14.2394   -0.320946   13.8977
 10.2938      2.41705   0.612277       1.83848   0.465717    3.12552
  2.41705    26.0275   -5.09524       19.7973   -3.87559   -12.2107
  ⋮                               ⋱
  3.82666   -14.9499   -0.336961     -11.4363   -0.257767   11.1619
  8.00915     1.88059   0.476384       1.94215   0.491979    3.30178
  1.88059    20.2507   -3.96436       20.9137   -4.09415   -12.8993
  0.476384   -3.96436   7.79415       -4.09415   8.04931    -0.290742
  3.19711   -12.4904   -0.281525  …  -12.8993   -0.290742   12.5898
  9.08226     2.13256   0.540212       2.47009   0.625713    4.19929
  2.13256    22.964    -4.49553       26.5986   -5.20705   -16.4057
  0.540212   -4.49553   8.83845       -5.20705  10.2373     -0.369774
  3.62548   -14.1639   -0.319246     -16.4057   -0.369774   16.0121
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.51027  0.824228   0.20879  …   0.607521    0.153895    1.03282
 0.0      5.52052   -1.13598      4.06906    -0.837308   -2.71989
 0.0      0.0        3.26453      0.0         2.40621    -1.11031
 0.0      0.0        0.0          0.0         0.0         0.701518
 0.0      0.0        0.0          0.269154    0.0681812   0.457578
 0.0      0.0       -0.0      …   1.80274    -0.370958   -1.20501
 0.0      0.0        0.0          0.0         1.06604    -0.491908
 0.0      0.0        0.0          0.0         0.0         0.310798
 0.0      0.0        0.0         -0.0924463  -0.0234182  -0.157164
 0.0      0.0       -0.0         -0.619187    0.127413    0.413885
 ⋮                            ⋱
 0.0      0.0        0.0         -0.0        -0.0        -0.173027
 0.0      0.0        0.0          0.0587708   0.0148876   0.0999139
 0.0      0.0       -0.0          0.393636   -0.0810002  -0.263119
 0.0      0.0        0.0          0.0         0.232774   -0.10741
 0.0      0.0        0.0      …   0.0         0.0         0.067864
 0.0      0.0        0.0          0.162744    0.0412257   0.276675
 0.0      0.0       -0.0          1.09003    -0.2243     -0.728611
 0.0      0.0        0.0          0.0         0.644582   -0.297432
 0.0      0.0        0.0          0.0         0.0         0.187924
julia> logdet(C)2.191438087152118
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  23.5674    -25.0645    -15.3527    …   -0.753536  -0.461563   -0.968537
 -25.0645     32.7585     19.6601         0.984846   0.59106     1.22033
 -15.3527     19.6601     15.8412         0.59106    0.476247    0.737637
 -32.216      40.5912     24.5357         1.22033    0.737637    1.59858
   2.15649    -2.29348    -1.40482        7.05127    4.3191      9.06315
  -2.29348     2.9975      1.79897   …   -9.21577   -5.53089   -11.4193
  -1.40482     1.79897     1.44952       -5.53089   -4.45651    -6.90249
  -2.94786     3.71422     2.24509      -11.4193    -6.90249   -14.9588
   0.535691   -0.56972    -0.34897       -4.14733   -2.54036    -5.33065
  -0.56972     0.744605    0.446878       5.42041    3.25309     6.71646
   ⋮                                 ⋱
   0.211455   -0.266427   -0.161044      13.4468     8.128      17.6147
  -7.08817     7.53844     4.61751        3.39006    2.07651     4.35733
   7.53844    -9.85248    -5.91302       -4.4307    -2.65911    -5.4901
   4.61751    -5.91302    -4.76441       -2.65911   -2.14257    -3.31854
   9.68932   -12.2083     -7.37938   …   -5.4901    -3.31854    -7.19179
   0.708528   -0.753536   -0.461563     -13.3477    -8.17583   -17.1561
  -0.753536    0.984846    0.59106       17.4449    10.4697     21.6161
  -0.461563    0.59106     0.476247      10.4697     8.43593    13.066
  -0.968537    1.22033     0.737637      21.6161    13.066      28.3162

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.