Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 24.9173   14.5124   8.71771  14.1918   …   8.7985    5.28531   8.6041
 14.5124   15.3871   7.46898   4.94101      9.32879   4.52824   2.9956
  8.71771   7.46898  7.58865   8.32032      4.52824   4.6008    5.04438
 14.1918    4.94101  8.32032  27.4749       2.9956    5.04438  16.6573
  6.66787   3.88353  2.33286   3.79773      2.09702   1.25969   2.05069
  3.88353   4.11759  1.9987    1.32222  …   2.22341   1.07925   0.713967
  2.33286   1.9987   2.03072   2.22652      1.07925   1.09655   1.20227
  3.79773   1.32222  2.22652   7.35229      0.713967  1.20227   3.97007
 19.7911   11.5268   6.92422  11.2721       8.87438   5.33089   8.67831
 11.5268   12.2215   5.93239   3.9245       9.40924   4.56729   3.02144
  ⋮                                     ⋱
  9.59308   3.33992  5.6242   18.5719       1.84506   3.10695  10.2596
 18.431    10.7347   6.44839  10.4975       7.74322   4.6514    7.57214
 10.7347   11.3817   5.52472   3.65481      8.20991   3.98513   2.63631
  6.44839   5.52472  5.61324   6.15444      3.98513   4.04898   4.43937
 10.4975    3.65481  6.15444  20.3229   …   2.63631   4.43937  14.6594
 15.1067    8.7985   5.28531   8.6041       9.58695   5.75894   9.37514
  8.7985    9.32879  4.52824   2.9956      10.1648    4.93403   3.26404
  5.28531   4.52824  4.6008    5.04438      4.93403   5.01308   5.49642
  8.6041    2.9956   5.04438  16.6573       3.26404   5.49642  18.15
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0060876311975999855
   0.01727662227399488
   0.06025001303115492
   0.17602241387302517
   0.028261813016638798
   0.08020667685309804
   0.2797105388394521
   0.8171836279402237
   0.037905354542124195
   0.10757492879074496
   ⋮
  16.791849775312688
   0.8539476009667272
   2.4234927617662954
   8.451621396915899
  24.69169257534895
   9.841607778096872
  27.93036152026144
  97.40356759976231
 284.56775735283014
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0575684    0.165852     …  -0.262394    0.124532    0.253137
 -0.140534    -0.114933         0.181835    0.220543    0.159984
  0.18777     -0.145773         0.230628    0.0209246   0.121881
 -0.0655414   -0.0255108        0.0403606  -0.303558    0.228482
  0.15983      0.460461        -0.0811521   0.0385149   0.0782893
 -0.39017     -0.319094     …   0.0562373   0.0682086   0.0494794
  0.521313    -0.404717         0.0713276   0.00647147  0.0376949
 -0.181965    -0.0708268        0.0124826  -0.0938831   0.0706641
  0.0501507    0.144482        -0.25015     0.118722    0.241326
 -0.122426    -0.100124         0.173351    0.210252    0.152519
  ⋮                         ⋱
  0.0685479    0.0266811        0.029703   -0.2234      0.168149
 -0.101941    -0.293686        -0.245162    0.116354    0.236514
  0.248854     0.20352          0.169894    0.20606     0.149478
 -0.332498     0.258132         0.215482    0.0195505   0.113877
  0.116059     0.0451739    …   0.0377102  -0.283623    0.213478
  0.00110129   0.00317275      -0.191522    0.0908967   0.184766
 -0.00268842  -0.00219868       0.132722    0.160975    0.116773
  0.00359205  -0.00278865       0.168336    0.0152729   0.0889614
 -0.00125381  -0.000488023      0.0294594  -0.221568    0.16677
julia> logdet(E)1.102242368486273
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  2.0416179054791876
 -1.7631822751737074
  0.2613487178480247
 -0.958906796803662
  1.8931448254183623
  0.5175293264962215
 -2.6848332994976456
  0.5177589880334741
  1.196329749988877
  0.9260833983363693
  ⋮
 -0.5229543232648437
  1.1854461232625175
  0.2767089036883797
 -5.211606583881839
 -0.3002656889233971
 -0.31844430958745895
  0.37277542083004317
 -0.6475603058587809
 -0.029846955745733716

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 12.1044    -5.0911     -8.07252    …  -2.3489     -3.72445    -0.580861
 -5.0911    61.7375      0.0423067     28.484       0.0195192   6.24045
 -8.07252    0.0423067  11.7692         0.0195192   5.42999    -2.66829
 -1.25898   13.5258     -5.78335        6.24045    -2.66829     4.04435
  8.28979   -3.48668    -5.52852       -2.21379    -3.51022    -0.54745
 -3.48668   42.2813      0.028974   …  26.8456      0.0183964   5.8815
 -5.52852    0.028974    8.0602         0.0183964   5.11766    -2.51481
 -0.862223   9.26324    -3.96077        5.8815     -2.51481     3.81172
  6.64557   -2.79512    -4.43198       -1.81982    -2.88554    -0.450025
 -2.79512   33.8951      0.0232272     22.0681      0.0151226   4.83481
  ⋮                                 ⋱
 -0.950614  10.2129     -4.36681        6.36321    -2.72078     4.12392
  7.1858    -3.02234    -4.79226       -1.62775    -2.58099    -0.402528
 -3.02234   36.6505      0.0251154     19.739       0.0135265   4.32453
 -4.79226    0.0251154   6.98679        0.0135265   3.7629     -1.84908
 -0.747396   8.02961    -3.43329    …   4.32453    -1.84908     2.80267
  5.58465   -2.3489     -3.72445       -2.2345     -3.54305    -0.552571
 -2.3489    28.484       0.0195192     27.0967      0.0185685   5.93651
 -3.72445    0.0195192   5.42999        0.0185685   5.16552    -2.53833
 -0.580861   6.24045    -2.66829        5.93651    -2.53833     3.84737
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.47914  -1.46332  -2.32027   …  -0.675139  -1.07051    -0.166955
 0.0       7.71985  -0.434333      3.56174   -0.20039     0.776716
 0.0       0.0       2.48936       0.0        1.14852    -1.09197
 0.0       0.0       0.0           0.0        0.0         0.205885
 0.0      -0.0      -0.0          -0.314608  -0.498847   -0.0777996
 0.0       0.0      -0.0       …   1.65973   -0.0933797   0.361942
 0.0       0.0       0.0           0.0        0.535201   -0.508849
 0.0       0.0       0.0           0.0        0.0         0.0959405
 0.0      -0.0      -0.0          -0.144558  -0.229212   -0.0357477
 0.0       0.0      -0.0           0.762623  -0.0429066   0.166307
 ⋮                             ⋱
 0.0       0.0       0.0           0.0        0.0         0.031386
 0.0      -0.0      -0.0           0.237847   0.377134    0.0588174
 0.0       0.0      -0.0          -1.25478    0.0705962  -0.273632
 0.0       0.0       0.0          -0.0       -0.404618    0.384696
 0.0       0.0       0.0       …  -0.0       -0.0        -0.0725321
 0.0      -0.0      -0.0          -0.217649  -0.345107   -0.0538225
 0.0       0.0      -0.0           1.14822   -0.064601    0.250395
 0.0       0.0       0.0           0.0        0.370257   -0.352027
 0.0       0.0       0.0           0.0        0.0         0.0663726
julia> logdet(C)-18.31262847115984
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  52.5937    -11.7772      72.1492    …    8.29609    -50.8232    -51.6527
 -11.7772      3.34698    -16.9563        -2.35767     11.9443     12.7098
  72.1492    -16.9563     102.623         11.9443     -72.2892    -73.4228
  73.3268    -18.0429     104.232         12.7098     -73.4228    -77.2251
 -66.3638     14.8607     -91.0394       -33.5653     205.626     208.983
  14.8607     -4.22329     21.3959    …    9.53895    -48.3258    -51.4226
 -91.0394     21.3959    -129.491        -48.3258     292.476     297.062
 -92.5253     22.7669    -131.522        -51.4226     297.062     312.446
  -1.85112     0.414519    -2.53941        2.24067    -13.7267    -13.9508
   0.414519   -0.117803     0.596806      -0.636779     3.22602     3.43275
   ⋮                                  ⋱
  99.3232    -24.4396     141.185         52.5387    -303.51     -319.228
 -51.8138     11.6026     -71.0793       -26.1788     160.375     162.993
  11.6026     -3.29735     16.7049         7.43977    -37.6911    -40.1064
 -71.0793     16.7049    -101.101        -37.6911     228.113     231.69
 -72.2394     17.7754    -102.686     …  -40.1064     231.69      243.688
 -37.0479      8.29609    -50.8232       -24.3858     149.392     151.83
   8.29609    -2.35767     11.9443         6.93023    -35.1097    -37.3595
 -50.8232     11.9443     -72.2892       -35.1097     212.49      215.822
 -51.6527     12.7098     -73.4228       -37.3595     215.822     226.998

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.