Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.93971    1.00128   -0.524252  …   0.917816   -0.480552   -2.65625
  1.00128    4.06576   -1.73733       3.72685    -1.59251    -4.40156
 -0.524252  -1.73733   23.8909       -1.59251    21.8994     -8.38818
 -2.89781   -4.80184   -9.151        -4.40156    -8.38818    20.3757
  1.53681    0.523446  -0.274067      0.637473   -0.333769   -1.84491
  0.523446   2.12548   -0.908237  …   2.5885     -1.10609    -3.05712
 -0.274067  -0.908237  12.4896       -1.10609    15.2103     -5.82605
 -1.51491   -2.51029   -4.78392      -3.05712    -5.82605    14.152
  1.9754     0.672831  -0.352282      0.761418   -0.398664   -2.20362
  0.672831   2.73207   -1.16744       3.09178    -1.32114    -3.65152
  ⋮                               ⋱
 -1.17614   -1.94893   -3.71412      -2.84682    -5.42527    13.1785
  1.94001    0.660778  -0.345971      0.865243   -0.453025   -2.5041
  0.660778   2.68312   -1.14652       3.51337    -1.50129    -4.14944
 -0.345971  -1.14652   15.7664       -1.50129    20.645      -7.9077
 -1.91236   -3.16888   -6.03903   …  -4.14944    -7.9077     19.2085
  2.69466    0.917816  -0.480552      1.14417    -0.599065   -3.31134
  0.917816   3.72685   -1.59251       4.64596    -1.98526    -5.48708
 -0.480552  -1.59251   21.8994       -1.98526    27.3002    -10.4569
 -2.65625   -4.40156   -8.38818      -5.48708   -10.4569     25.4007

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   4.843821072214268e-6
   5.6288170978073595e-6
   3.87510044932722e-5
   7.748048454836101e-5
   0.009312072448156801
   0.010821199179482655
   0.07449741761729509
   0.14895345527344353
   0.03273089247212235
   0.03803530403515837
   ⋮
   8.390885510093563
   1.015856901132749
   1.1804880091096555
   8.126946630239892
  16.24937910755266
  14.663197383460886
  17.039534472898826
 117.30689866966556
 234.5486386390105
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.051121   -0.17888      0.0361516  …  -0.326416     0.0659687   0.019944
  0.16543     0.0619068    0.0680874      0.112966     0.124244    0.0259559
  0.0418445  -0.00508521  -0.125473      -0.00927938  -0.228961    0.248313
  0.0656227  -0.0134699   -0.119797      -0.0245797   -0.218604   -0.239307
 -0.128175    0.448503    -0.0906426     -0.207358     0.0419071   0.0126696
 -0.414781   -0.155218    -0.170715   …   0.0717625    0.0789271   0.0164887
 -0.104916    0.0127501    0.314597      -0.00589479  -0.145449    0.157743
 -0.164535    0.0337731    0.300367      -0.0156144   -0.138869   -0.152022
 -0.0587276   0.205496    -0.0415308     -0.256592     0.0518572   0.0156778
 -0.190045   -0.0711183   -0.0782185      0.0888015    0.0976671   0.0204037
  ⋮                                   ⋱
  0.146527   -0.0300765   -0.267491      -0.0138322   -0.123019   -0.13467
 -0.102626    0.359102    -0.0725746     -0.277365     0.0560555   0.016947
 -0.332101   -0.124278    -0.136686       0.0959907    0.105574    0.0220555
 -0.0840032   0.0102086    0.251888      -0.00788496  -0.194555    0.210999
 -0.131738    0.027041     0.240494   …  -0.0208861   -0.185754   -0.203346
  0.0667206  -0.233465     0.0471833     -0.365216     0.0738103   0.0223147
  0.215911    0.0807977    0.0888643      0.126394     0.139013    0.0290413
  0.0546135  -0.00663697  -0.163761      -0.0103824   -0.256177    0.27783
  0.0856476  -0.0175803   -0.156354      -0.0275014   -0.244589   -0.267753

julia> logdet(E)
-48.507157993063174

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  9.81107238402184
  3.4734949395323174
  0.8491461390668298
 -1.5541134965108063
 -3.2467050430295736
 -3.9343348680674177
  6.196122843725345
  8.350027799692905
 -1.3839748570620074
  1.1606112940520603
  ⋮
 -4.447337523485494
 -1.2550055311962336
 -4.019812255300647
 -7.822981261246539
  5.910538097452851
  6.328404107986243
 -2.0237093510269846
  0.7737962042950464
 -6.139817467207851

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 7.2465      2.98708    0.107753   …   1.23812    0.0446627   0.726942
 2.98708    25.2319    -5.00368       10.4584    -2.07398    -0.945055
 0.107753   -5.00368    7.03595       -2.07398    2.91634     1.04598
 1.75382    -2.28004    2.52353       -0.945055   1.04598     1.05203
 7.33647     3.02417    0.109091       2.57546    0.0929047   1.51214
 3.02417    25.5452    -5.06581    …  21.7549    -4.31417    -1.96585
 0.109091   -5.06581    7.12332       -4.31417    6.06639     2.17578
 1.7756     -2.30835    2.55486       -1.96585    2.17578     2.18837
 4.97584     2.05109    0.0739892      3.24669    0.117118    1.90624
 2.05109    17.3256    -3.4358        27.4248    -5.43855    -2.4782
 ⋮                                 ⋱
 1.35689    -1.76402    1.9524        -1.45621    1.61172     1.62105
 4.17541     1.72115    0.0620871      2.09311    0.0755049   1.22894
 1.72115    14.5385    -2.88311       17.6805    -3.50619    -1.59767
 0.0620871  -2.88311    4.0541        -3.50619    4.93024     1.76829
 1.01055    -1.31375    1.45405    …  -1.59767    1.76829     1.77852
 3.00361     1.23812    0.0446627      3.72289    0.134296    2.18584
 1.23812    10.4584    -2.07398       31.4472    -6.23624    -2.84168
 0.0446627  -2.07398    2.91634       -6.23624    8.76912     3.14515
 0.726942   -0.945055   1.04598       -2.84168    3.14515     3.16335

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.69193  1.10964   0.0400282  …  0.459936   0.0165913    0.270045
 0.0      4.89904  -1.03043       2.03061   -0.427102    -0.254072
 0.0      0.0       2.44388       0.0        1.01297      0.31645
 0.0      0.0       0.0           0.0        0.0          0.445466
 0.0      0.0       0.0           0.71967    0.0259607    0.422543
 0.0      0.0      -0.0        …  3.17732   -0.668294    -0.397551
 0.0      0.0       0.0           0.0        1.58501      0.495154
 0.0      0.0       0.0           0.0        0.0          0.697028
 0.0      0.0       0.0           0.645408   0.0232819    0.378942
 0.0      0.0      -0.0           2.84946   -0.599334    -0.356528
 ⋮                             ⋱
 0.0      0.0       0.0           0.0        0.0          0.228581
 0.0      0.0       0.0           0.146931   0.00530027   0.0862686
 0.0      0.0      -0.0           0.648698  -0.136442    -0.0811659
 0.0      0.0       0.0           0.0        0.323603     0.101093
 0.0      0.0       0.0        …  0.0        0.0          0.142309
 0.0      0.0       0.0           0.161612   0.00582984   0.0948881
 0.0      0.0      -0.0           0.713513  -0.150075    -0.0892756
 0.0      0.0       0.0           0.0        0.355936     0.111194
 0.0      0.0       0.0           0.0        0.0          0.156527

julia> logdet(C)
-0.9370232312795785

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.31345   -1.37991      2.17014    …  -0.711278    1.1186     -4.71213
 -1.37991    2.10196      0.772058       1.08346     0.397958    1.0691
  2.17014    0.772058    10.1582         0.397958    5.23605    -5.62137
 -9.14176    2.07411    -10.9057         1.0691     -5.62137    17.9942
 -0.834387   0.138496    -0.217808      -0.885907    1.39324    -5.86903
  0.138496  -0.210965    -0.0774882  …   1.34946     0.495662    1.33158
 -0.217808  -0.0774882   -1.01954        0.495662    6.52158    -7.00151
  0.917521  -0.20817      1.09456        1.33158    -7.00151    22.4121
 -1.2675     0.210387    -0.330869       0.663853   -1.04402     4.39795
  0.210387  -0.320474    -0.117711      -1.01122    -0.371424   -0.997821
  ⋮                                  ⋱
  7.86046   -1.78341      9.3772         0.638107   -3.35519    10.7401
  4.03824   -0.67029      1.05414        1.96771    -3.09455    13.0359
 -0.67029    1.02102      0.375025      -2.99733    -1.10093    -2.95762
  1.05414    0.375025     4.93432       -1.10093   -14.4853     15.5512
 -4.44059    1.0075      -5.29744    …  -2.95762    15.5512    -49.7801
  4.28518   -0.711278     1.1186        -1.61333     2.53724   -10.6881
 -0.711278   1.08346      0.397958       2.45752     0.902656    2.42496
  1.1186     0.397958     5.23605        0.902656   11.8765    -12.7505
 -4.71213    1.0691      -5.62137        2.42496   -12.7505     40.8149

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source