Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3.81877   -2.93091    0.583447  …  -2.48199    0.494082  -3.3207
 -2.93091    7.06884   -2.09441       5.98612   -1.77361   -0.417021
  0.583447  -2.09441    7.30191      -1.77361    6.18349    0.403342
 -3.92132   -0.492448   0.476295     -0.417021   0.403342   5.7507
  2.39868   -1.84099    0.36648      -2.76155    0.549733  -3.69473
 -1.84099    4.44014   -1.31556   …   6.66037   -1.97339   -0.463992
  0.36648   -1.31556    4.58654      -1.97339    6.87998    0.448773
 -2.46309   -0.309321   0.299175     -0.463992   0.448773   6.39844
  2.53468   -1.94537    0.387259     -2.82526    0.562417  -3.77997
 -1.94537    4.6919    -1.39015       6.81404   -2.01892   -0.474698
  ⋮                               ⋱
 -2.92444   -0.367257   0.355211     -0.493893   0.477692   6.81076
  2.64944   -2.03345    0.404793     -1.83025    0.364342  -2.44872
 -2.03345    4.90432   -1.45309       4.41423   -1.30788   -0.307516
  0.404793  -1.45309    5.06603      -1.30788    4.55978    0.297429
 -2.72059   -0.341658   0.330451  …  -0.307516   0.297429   4.24064
  3.23386   -2.48199    0.494082     -4.22037    0.840137  -5.64651
 -2.48199    5.98612   -1.77361      10.1788    -3.01585   -0.709102
  0.494082  -1.77361    6.18349      -3.01585   10.5144     0.685842
 -3.3207    -0.417021   0.403342     -0.709102   0.685842   9.77849
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
  5.639863945371459e-5
  0.005677032336157018
  0.009617117135987755
  0.011467240880588269
  0.0003915164565503987
  0.03940966671365042
  0.06676153289836237
  0.07960499684859139
  0.004070526035168738
  0.40973520195973356
  ⋮
  6.401317891598471
  0.04140398331911233
  4.16768479568039
  7.060222752415361
  8.418455741751915
  0.37701368041608013
 37.94983133683886
 64.28851407677867
 76.65622310191328
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.112004     0.0146543    0.0401419  …   0.03397     0.0930529  -0.173963
 -0.0520705   -0.085382     0.0489348     -0.197924    0.113436    0.202177
 -0.00150382  -0.103206    -0.0766662     -0.239242   -0.17772    -0.136577
 -0.0688791    0.0429702   -0.100594       0.0996091  -0.233186    0.133022
 -0.429646     0.0562136    0.153984       0.029685    0.0813151  -0.152019
 -0.199742    -0.327524     0.187713   …  -0.172957    0.0991267   0.176674
 -0.00576864  -0.395899    -0.294091      -0.209064   -0.155302   -0.119349
 -0.26422      0.164833    -0.385877       0.0870443  -0.203772    0.116242
 -0.155631     0.0203623    0.0557779      0.0344022   0.0942369  -0.176176
 -0.0723528   -0.11864      0.0679957     -0.200442    0.114879    0.20475
  ⋮                                    ⋱
 -0.0111306    0.00694383  -0.0162556      0.101897   -0.238544    0.136078
  0.436147    -0.0570641   -0.156314       0.0293053   0.0802749  -0.150074
  0.202764     0.33248     -0.190554      -0.170745    0.0978587   0.174414
  0.00585592   0.401889     0.29854       -0.20639    -0.153315   -0.117822
  0.268217    -0.167327     0.391716   …   0.0859309  -0.201165    0.114755
  0.415025    -0.0543006   -0.148744       0.0411864   0.112821   -0.210918
  0.192944     0.316378    -0.181325      -0.23997     0.137533    0.245127
  0.00557232   0.382426     0.284082      -0.290066   -0.215473   -0.165591
  0.255228    -0.159224     0.372745       0.12077    -0.282723    0.16128
julia> logdet(E)-44.107144272044025
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  -1.3088833257767216
  -2.8923207359226697
  -1.3424117709155623
  -1.0417103450688237
   8.00148754883205
  -6.9243469456258
  -5.8480686425873305
   0.6165445779741712
   5.992938984650323
  -2.64439886828469
   ⋮
  -5.691126504293786
   3.210225286290404
  14.270820111779448
   6.235451388087649
   6.681703353840608
 -14.569495993103986
   3.8185638097188015
   5.454014762234588
  -5.566826608995644

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 13.1088     0.982147  0.325568  2.46149   …   0.987717   0.327415  2.47545
  0.982147  11.0335    7.75077   1.31233      11.096      7.79473   1.31977
  0.325568   7.75077   7.11128   1.6024        7.79473    7.15161   1.61149
  2.46149    1.31233   1.6024    0.954741      1.31977    1.61149   0.960156
 13.9031     1.04166   0.345295  2.61065       1.10906    0.367639  2.77957
  1.04166   11.702     8.22042   1.39185   …  12.4592     8.75235   1.48191
  0.345295   8.22042   7.54218   1.6995        8.75235    8.03022   1.80947
  2.61065    1.39185   1.6995    1.01259       1.48191    1.80947   1.07812
  7.99817    0.599245  0.198642  1.50185       0.592756   0.19649   1.48559
  0.599245   6.73194   4.72904   0.800702      6.65905    4.67783   0.792031
  ⋮                                        ⋱
  1.76701    0.94207   1.1503    0.685372      1.44737    1.76729   1.05299
 13.9327     1.04388   0.34603   2.6162        1.5647     0.518675  3.9215
  1.04388   11.7269    8.23792   1.39481      17.5778    12.3481    2.09072
  0.34603    8.23792   7.55823   1.70311      12.3481    11.3292    2.55285
  2.6162     1.39481   1.70311   1.01475   …   2.09072    2.55285   1.52103
 13.1831     0.987717  0.327415  2.47545       1.73158    0.573995  4.33975
  0.987717  11.096     7.79473   1.31977      19.4526    13.6651    2.31371
  0.327415   7.79473   7.15161   1.61149      13.6651    12.5376    2.82512
  2.47545    1.31977   1.61149   0.960156      2.31371    2.82512   1.68326
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.6206  0.271266  0.0899209  0.679857  3.83999  …   0.0904309    0.683713
 0.0     3.31057   2.33385    0.340698  0.0          2.34709      0.342631
 0.0     0.0       1.28699    0.579749  0.0          1.29428      0.583037
 0.0     0.0       0.0        0.200876  0.0          0.0          0.202015
 0.0     0.0       0.0        0.0       3.01015      0.00677208   0.051201
 0.0     0.0       0.0        0.0       0.0      …   0.175766     0.0256585
 0.0     0.0       0.0        0.0       0.0          0.0969247    0.0436617
 0.0     0.0       0.0        0.0       0.0          0.0          0.0151283
 0.0     0.0       0.0        0.0       0.0         -0.0031186   -0.0235785
 0.0     0.0       0.0        0.0       0.0         -0.0809416   -0.011816
 ⋮                                               ⋱
 0.0     0.0       0.0        0.0       0.0          0.0          0.0836845
 0.0     0.0       0.0        0.0       0.0          0.028067     0.212203
 0.0     0.0       0.0        0.0       0.0          0.728464     0.106342
 0.0     0.0       0.0        0.0       0.0          0.401706     0.180957
 0.0     0.0       0.0        0.0       0.0      …   0.0          0.0626994
 0.0     0.0       0.0        0.0       0.0          0.017639     0.133362
 0.0     0.0       0.0        0.0       0.0          0.457812     0.0668319
 0.0     0.0       0.0        0.0       0.0          0.252457     0.113724
 0.0     0.0       0.0        0.0       0.0          0.0          0.0394042
julia> logdet(C)-39.39600602715418
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   39.6406    -41.4872     86.1718  …    21.1325   -43.8936    96.68
  -41.4872     60.9454   -112.176       -31.0439    57.1393  -107.713
   86.1718   -112.176     223.903        57.1393  -114.05     226.043
 -189.802     211.461    -443.767      -107.713    226.043   -501.793
  -24.7009     25.8516    -53.6956      -18.1754    37.7516   -83.1518
   25.8516    -37.9765     69.8993  …    26.7      -49.1439    92.6406
  -53.6956     69.8993   -139.519       -49.1439    98.0914  -194.413
  118.27     -131.766     276.521        92.6406  -194.413    431.578
   -9.46263     9.90345   -20.5701      -11.6498    24.1974   -53.2972
    9.90345   -14.5483     26.7776       17.1137   -31.4994    59.3792
    ⋮                               ⋱
 -135.36      150.806    -316.478      -201.267    422.375   -937.63
   23.5129    -24.6083     51.1131       22.8974   -47.5595   104.755
  -24.6083     36.15      -66.5374      -33.6367    61.9114  -116.709
   51.1131    -66.5374    132.809        61.9114  -123.575    244.921
 -112.582     125.429    -263.222   …  -116.709    244.921   -543.702
  -20.1918     21.1325    -43.8936      -27.1232    56.3368  -124.087
   21.1325    -31.0439     57.1393       39.8445   -73.3375   138.248
  -43.8936     57.1393   -114.05        -73.3375   146.382   -290.123
   96.68     -107.713     226.043       138.248   -290.123    644.044

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.