Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.5968    -0.490433   7.49635  -2.42977   …  -0.278257   4.25321  -1.37858
 -0.490433   8.08434    2.07957   1.3937        4.58682    1.17989   0.790747
  7.49635    2.07957   31.3861   22.0403        1.17989   17.8075   12.505
 -2.42977    1.3937    22.0403   41.0618        0.790747  12.505    23.2973
  3.6248    -0.38673    5.91124  -1.91599      -0.313769   4.79601  -1.55452
 -0.38673    6.3749     1.63984   1.099     …   5.1722     1.33047   0.891664
  5.91124    1.63984   24.7494   17.3799        1.33047   20.0802   14.1009
 -1.91599    1.099     17.3799   32.3792        0.891664  14.1009   26.2705
  2.59356   -0.276708   4.22952  -1.3709       -0.290732   4.44389  -1.44038
 -0.276708   4.56127    1.17332   0.786343      4.79246    1.23279   0.826198
  ⋮                                         ⋱
 -1.12771    0.646849  10.2294   19.0577        0.733394  11.598    21.6075
  3.27761   -0.349689   5.34505  -1.73247      -0.319817   4.88846  -1.58448
 -0.349689   5.7643     1.48278   0.993739      5.2719     1.35611   0.908852
  5.34505    1.48278   22.3789   15.7152        1.35611   20.4672   14.3727
 -1.73247    0.993739  15.7152   29.2779    …   0.908852  14.3727   26.7769
  2.60809   -0.278257   4.25321  -1.37858      -0.381152   5.82597  -1.88835
 -0.278257   4.58682    1.17989   0.790747      6.28295    1.61619   1.08315
  4.25321    1.17989   17.8075   12.505         1.61619   24.3925   17.1292
 -1.37858    0.790747  12.505    23.2973        1.08315   17.1292   31.9122

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0009526044983713343
   0.02788043261678485
   0.06132423236699753
   0.2041059291133309
   0.005668749152475512
   0.16591059462480245
   0.36492761775844407
   1.2145914854009034
   0.01986613397489699
   0.581433754081675
   ⋮
  17.21106614566819
   0.14063572865962806
   4.116068067189371
   9.053471947875664
  30.132742511377852
   1.663151036277494
  48.6764134304931
 107.0659027796618
 356.3482936471423
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.12784     -0.0176382    …  -0.0407665    0.164166   0.0178022
  0.0169352    0.146284         0.3381       0.0117942  0.0158195
 -0.0598251   -0.00083463      -0.00192905   0.225008   0.215597
  0.0393658   -0.00692013      -0.0159942   -0.196252   0.263029
  0.0149346   -0.00206053      -0.0444926    0.179171   0.0194293
  0.00197841   0.0170892    …   0.369003     0.0128722  0.0172654
 -0.0069889   -9.75033e-5      -0.00210537   0.245574   0.235302
  0.0045988   -0.000808425     -0.0174561   -0.21419    0.28707
  0.157011    -0.0216629       -0.0378006    0.152223   0.016507
  0.0207996    0.179663         0.313503     0.0109361  0.0146686
  ⋮                         ⋱
  0.0275083   -0.0048357       -0.0127965   -0.157016   0.210442
 -0.136022     0.018767        -0.0438533    0.176597   0.0191502
 -0.0180191   -0.155646         0.363701     0.0126872  0.0170174
  0.0636538    0.000888045     -0.00207512   0.242046   0.231922
 -0.0418852    0.00736301   …  -0.0172053   -0.211113   0.282946
 -0.528028     0.0728522       -0.037209     0.14984    0.0162487
 -0.0699489   -0.604206         0.308596     0.010765   0.014439
  0.2471       0.00344734      -0.00176071   0.205373   0.196783
 -0.162595     0.0285827       -0.0145985   -0.179127   0.240076

julia> logdet(E)
-3.274483033060534

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -3.7999303000445677
 -0.5555169669807595
  2.5644695549307874
 -1.1195610928465225
  1.1770354698679433
 -0.700976212533873
  0.03043237108532176
 -0.05959517686387853
  8.921921528644383
  2.2834660076200333
  ⋮
 -1.0637891965425863
  8.567872661343214
  2.708824319957192
 -3.802585941473172
  2.0907161478283296
  0.05243468041745338
 -1.5159154002234885
 -3.9823761245350418
  0.8359396348357794

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.5022     7.30479   -3.07078   …    5.80905   -2.442      -6.48794
  7.30479   16.621    -11.3648        13.2177    -9.03769    -2.08126
 -3.07078  -11.3648    13.2735        -9.03769   10.5556      0.713839
 -8.15849   -2.61715    0.897642      -2.08126    0.713839    7.80697
  8.46975    5.8911    -2.47649        8.17706   -3.43746    -9.1327
  5.8911    13.4044    -9.16534   …   18.6058   -12.7218     -2.92966
 -2.47649   -9.16534   10.7047       -12.7218    14.8585      1.00483
 -6.57958   -2.11065    0.723921      -2.92966    1.00483    10.9894
  7.79379    5.42094   -2.27884        8.64334   -3.63347    -9.65347
  5.42094   12.3346    -8.43387       19.6667   -13.4473     -3.09672
  ⋮                               ⋱
 -4.50504   -1.44516    0.495669      -1.90436    0.653166    7.14342
  7.73776    5.38196   -2.26246        7.11085   -2.98925    -7.94188
  5.38196   12.2459    -8.37324       16.1798   -11.063      -2.54766
 -2.26246   -8.37324    9.77955      -11.063     12.9211      0.873809
 -6.01094   -1.92824    0.661357  …   -2.54766    0.873809    9.5565
  8.35178    5.80905   -2.442         11.5366    -4.84973   -12.8848
  5.80905   13.2177    -9.03769       26.2499   -17.9486     -4.13331
 -2.442     -9.03769   10.5556       -17.9486    20.9631      1.41766
 -6.48794   -2.08126    0.713839      -4.13331    1.41766    15.5044

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.24072  2.25407  -0.947561  -2.5175    …   1.79252   -0.753537   -2.00201
 0.0      3.39709  -2.71671    0.900022      2.7015    -2.16043     0.715732
 0.0      0.0       2.23498    0.428306      0.0        1.77734     0.340605
 0.0      0.0       0.0        1.57667       0.0        0.0         1.25382
 0.0      0.0      -0.0       -0.0           1.24878   -0.524961   -1.39472
 0.0      0.0      -0.0        0.0       …   1.88203   -1.50509     0.498623
 0.0      0.0       0.0        0.0           0.0        1.23821     0.237287
 0.0      0.0       0.0        0.0           0.0        0.0         0.873492
 0.0      0.0      -0.0       -0.0           1.08155   -0.454662   -1.20795
 0.0      0.0      -0.0        0.0           1.63      -1.30354     0.431851
 ⋮                                       ⋱
 0.0      0.0       0.0        0.0          -0.0       -0.0        -0.426669
 0.0      0.0      -0.0       -0.0           0.329167  -0.138375   -0.367637
 0.0      0.0      -0.0        0.0           0.496086  -0.396728    0.131433
 0.0      0.0       0.0        0.0           0.0        0.32638     0.0625467
 0.0      0.0       0.0        0.0       …   0.0        0.0         0.230245
 0.0      0.0      -0.0       -0.0           0.15988   -0.0672099  -0.178564
 0.0      0.0      -0.0        0.0           0.240954  -0.192694    0.063838
 0.0      0.0       0.0        0.0           0.0        0.158526    0.0303795
 0.0      0.0       0.0        0.0           0.0        0.0         0.111832

julia> logdet(C)
-4.0188099969659135

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   4.12541   -1.99898  -0.958872  …   -4.16864   -1.99961    6.22103
  -1.99898    2.08939   1.40987        4.35718    2.94012   -2.57158
  -0.958872   1.40987   1.57576        2.94012    3.28605   -1.17843
   2.98316   -1.23315  -0.565089      -2.57158   -1.17843    6.14925
   7.57556   -3.67077  -1.76079      -60.0381   -28.799     89.5973
  -3.67077    3.83679   2.58897   …   62.7535    42.3445   -37.0367
  -1.76079    2.58897   2.89359       42.3445    47.3268   -16.9721
   5.47804   -2.26445  -1.03768      -37.0367   -16.9721    88.5634
  -8.46665    4.10255   1.96791       42.933     20.5941   -64.0706
   4.10255   -4.2881   -2.89351      -44.8747   -30.2804    26.4848
   ⋮                              ⋱
  11.1595    -4.613    -2.1139       -51.4824   -23.5918   123.106
 -17.3089     8.3871    4.02312       62.2289    29.8499   -92.8667
   8.3871    -8.76643  -5.91538      -65.0434   -43.8897    38.3882
   4.02312   -5.91538  -6.61138      -43.8897   -49.0538    17.5914
 -12.5164     5.1739    2.37094   …   38.3882    17.5914   -91.7952
   8.60304   -4.16864  -1.99961      -54.2052   -26.0011    80.8926
  -4.16864    4.35718   2.94012       56.6568    38.2306   -33.4385
  -1.99961    2.94012   3.28605       38.2306    42.7288   -15.3232
   6.22103   -2.57158  -1.17843      -33.4385   -15.3232    79.9592

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source