Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.84637   -3.70168   -0.959902  …  -2.86681   -0.743408  -0.242639
 -3.70168   16.8839     0.523763     13.0759     0.405635   4.7232
 -0.959902   0.523763   4.61079       0.405635   3.57088    1.80113
 -0.3133     6.09869    2.32565       4.7232     1.80113    4.46509
  8.87208   -4.79693   -1.24392      -4.12584   -1.06989   -0.3492
 -4.79693   21.8795     0.678734  …  18.8186     0.583779   6.79752
 -1.24392    0.678734   5.97503       0.583779   5.13912    2.59213
 -0.405999   7.90317    3.01376       6.79752    2.59213    6.42605
  5.61654   -3.03673   -0.787472     -2.46942   -0.640359  -0.209005
 -3.03673   13.851      0.429678     11.2634     0.349407   4.06849
  ⋮                               ⋱
 -0.21816    4.2467     1.61942       3.12538    1.19182    2.95459
  5.1841    -2.80292   -0.726841     -2.88594   -0.748369  -0.244258
 -2.80292   12.7846     0.396595     13.1632     0.408342   4.75473
 -0.726841   0.396595   3.4913        0.408342   3.59471    1.81315
 -0.237232   4.61795    1.76099   …   4.75473    1.81315    4.49489
  5.30225   -2.86681   -0.743408     -3.04493   -0.789596  -0.257714
 -2.86681   13.0759     0.405635     13.8884     0.430837   5.01666
 -0.743408   0.405635   3.57088       0.430837   3.79274    1.91303
 -0.242639   4.7232     1.80113       5.01666    1.91303    4.74251
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   4.017426737582163e-6
   1.7167544105924734e-5
   1.904299164860705e-5
   6.169598603740475e-5
   0.02507936447003971
   0.10717086428988373
   0.11887861543002713
   0.38514607016873065
   0.058788780215695925
   0.2512202569483283
   ⋮
   9.552458675967076
   1.0487411891600116
   4.481552943374922
   4.971134761437796
  16.1056133663829
  10.955803044166183
  46.81709070561508
  51.93156702078464
 168.24925898915893
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0564752  -0.0595355  -0.180232    …  -0.0941986  -0.285168   -0.0817683
 -0.067684   -0.0676314  -0.00827402     -0.107008   -0.0130914   0.285962
 -0.110908    0.168685   -0.0268069       0.266898   -0.0424146   0.0322159
  0.147707    0.0729053  -0.092831        0.115353   -0.14688     0.123963
 -0.10127    -0.106757   -0.323187       -0.13962    -0.422674   -0.121196
 -0.121369   -0.121275   -0.0148367   …  -0.158607   -0.0194039   0.42385
 -0.198876    0.302481   -0.0480694       0.395594   -0.0628666   0.0477501
  0.264863    0.130732   -0.166462        0.170975   -0.217704    0.183737
  0.0710496   0.0748996   0.226744       -0.0941087  -0.284896   -0.0816902
  0.0851511   0.0850848   0.0104093      -0.106906   -0.0130789   0.285689
  ⋮                                   ⋱
 -0.209794   -0.10355     0.131852        0.101198   -0.128856    0.108752
 -0.0553916  -0.0583932  -0.176774       -0.0869593  -0.263253   -0.0754842
 -0.0663854  -0.0663338  -0.00811527     -0.0987845  -0.0120853   0.263985
 -0.10878     0.165449   -0.0262926       0.246387   -0.039155    0.02974
  0.144873    0.0715065  -0.0910499   …   0.106488   -0.135592    0.114436
  0.199702    0.210523    0.637319       -0.082894   -0.250946   -0.0719554
  0.239337    0.239151    0.0292577      -0.0941664  -0.0115203   0.251644
  0.39218    -0.596487    0.0947918       0.234868   -0.0373245   0.0283497
 -0.522305   -0.2578      0.328259        0.101509   -0.129253    0.109086
julia> logdet(E)-37.31770260729532
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -1.9512174674445801
  0.23148141597908561
  0.5467643331303189
  0.4060633798978392
 -4.194971833452597
 -4.5592012761707394
 -4.2815472588377865
 -2.0378287637871475
  3.7384942771361245
  3.6766005630834626
  ⋮
  5.6103298842144245
 -3.3095547304265867
  1.1734933715630755
  2.309551183657847
 -0.26211959396313367
  6.405562496170932
 -1.4273739641162233
 -4.894001891294755
  5.779704205981404

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  18.5213   11.2535    17.668    -14.2578   …   9.42733   14.8009   -11.9441
  11.2535    7.27605   11.6996    -8.79814      6.09533    9.80107   -7.37041
  17.668    11.6996    35.335    -12.8218       9.80107   29.601    -10.7411
 -14.2578   -8.79814  -12.8218    20.3717      -7.37041  -10.7411    17.0658
  20.0024   12.1534    19.0809   -15.3979      12.4386    19.5286   -15.7592
  12.1534    7.8579    12.6352    -9.5017   …   8.04228   12.9317    -9.72465
  19.0809   12.6352    38.1607   -13.8471      12.9317    39.0561   -14.172
 -15.3979   -9.5017   -13.8471    22.0007      -9.72465  -14.172     22.517
  12.1356    7.37357   11.5765    -9.34205      7.58328   11.9058    -9.60774
   7.37357   4.76745    7.66589   -5.76475      4.90304    7.88391   -5.9287
   ⋮                                        ⋱
 -10.2325   -6.31425   -9.20194   14.6204      -6.43359   -9.37585   14.8967
   8.8067    5.35093    8.40096   -6.77944      5.17366    8.12264   -6.55484
   5.35093   3.45969    5.56306   -4.18343      3.34507    5.37876   -4.04483
   8.40096   5.56306   16.8015    -6.09663      5.37876   16.2448    -5.89465
  -6.77944  -4.18343   -6.09663    9.68653  …  -4.04483   -5.89465    9.36562
  15.5157    9.42733   14.8009   -11.9441      10.7686    16.9066   -13.6434
   9.42733   6.09533    9.80107   -7.37041      6.9625    11.1955    -8.41899
  14.8009    9.80107   29.601    -10.7411      11.1955    33.8123   -12.2692
 -11.9441   -7.37041  -10.7411    17.0658      -8.41899  -12.2692    19.4938
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 4.30364  2.61488   4.10537  …   2.19055     3.43916    -2.77535
 0.0      0.662154  1.45675      0.554702    1.22035    -0.170965
 0.0      0.0       4.04461      0.0         3.38827     0.222953
 0.0      0.0       0.0          0.0         0.0         2.55245
 0.0      0.0       0.0          0.703494    1.10449    -0.891302
 0.0      0.0       0.0      …   0.178142    0.391916   -0.0549055
 0.0      0.0       0.0          0.0         1.08814     0.0716012
 0.0      0.0       0.0          0.0         0.0         0.819717
 0.0      0.0       0.0         -0.12006    -0.188494    0.152111
 0.0      0.0       0.0         -0.0304021  -0.0668851   0.00937027
 ⋮                           ⋱
 0.0      0.0       0.0          0.0         0.0         0.152315
 0.0      0.0       0.0          0.482895    0.758145   -0.611811
 0.0      0.0       0.0          0.122281    0.26902    -0.0376884
 0.0      0.0       0.0          0.0         0.746926    0.0491488
 0.0      0.0       0.0      …   0.0         0.0         0.562673
 0.0      0.0       0.0          0.659088    1.03477    -0.835041
 0.0      0.0       0.0          0.166898    0.367177   -0.0514398
 0.0      0.0       0.0          0.0         1.01946     0.0670816
 0.0      0.0       0.0          0.0         0.0         0.767975
julia> logdet(C)-5.901318716704775
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   72.9757    -109.654       1.51248   …   -27.5782     0.380389    1.17419
 -109.654      202.091     -10.7078         50.826     -2.69303     0.954303
    1.51248    -10.7078      4.78995        -2.69303    1.20468    -0.138627
    4.66872      3.79443    -0.551198        0.954303  -0.138627    2.10674
 -178.326      267.956      -3.69594        97.9815    -1.35147    -4.17173
  267.956     -493.837      26.1661    …  -180.578      9.56796    -3.3905
   -3.69594     26.1661    -11.7049          9.56796   -4.28004     0.492521
  -11.4087      -9.27221     1.34693        -3.3905     0.492521   -7.48496
  171.223     -257.282       3.54871       -83.8486     1.15653     3.56999
 -257.282      474.165     -25.1237        154.531     -8.18787     2.90146
    ⋮                                  ⋱
   -1.52327     -1.23801     0.179839       -0.311649   0.0452716  -0.688004
  -37.6683      56.601      -0.780704       18.7243    -0.258266   -0.797218
   56.601     -104.314       5.52713       -34.5085     1.82844    -0.647927
   -0.780704     5.52713    -2.47246         1.82844   -0.817918    0.094121
   -2.40988     -1.95859     0.284515  …    -0.647927   0.094121   -1.43038
   18.3535     -27.5782      0.380389      -22.1953     0.306141    0.945
  -27.5782      50.826      -2.69303        40.9054    -2.16738     0.768034
    0.380389    -2.69303     1.20468        -2.16738    0.969537   -0.111568
    1.17419      0.954303   -0.138627        0.768034  -0.111568    1.69553

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.