Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.3051   -8.46646  -6.38596   …   -8.26383  -6.23312    -9.51596
 -8.46646  22.2658   13.1148        21.733    12.8009     -2.63607
 -6.38596  13.1148   10.2512        12.8009   10.0059     -1.03821
 -9.7493   -2.7007   -1.06367       -2.63607  -1.03821    34.6138
  7.16369  -5.88555  -4.43926       -7.40619  -5.58623    -8.52837
 -5.88555  15.4783    9.1169    …   19.4775   11.4724     -2.36249
 -4.43926   9.1169    7.12624       11.4724    8.96744    -0.930465
 -6.77732  -1.87742  -0.739421      -2.36249  -0.930465   31.0215
  3.29893  -2.71033  -2.04431       -3.07644  -2.32045    -3.54258
 -2.71033   7.12787   4.19839        8.0907    4.76551    -0.98135
  ⋮                             ⋱
 -8.31252  -2.30269  -0.906914      -2.65051  -1.0439     34.8034
  8.59877  -7.06458  -5.32857       -6.93273  -5.22912    -7.98317
 -7.06458  18.5791   10.9433        18.2323   10.739      -2.21146
 -5.32857  10.9433    8.55381       10.739     8.39417    -0.870982
 -8.135    -2.25352  -0.887547  …   -2.21146  -0.870982   29.0384
 10.0585   -8.26383  -6.23312      -10.8868   -8.21157   -12.5364
 -8.26383  21.733    12.8009        28.6312   16.8641     -3.47278
 -6.23312  12.8009   10.0059        16.8641   13.1818     -1.36775
 -9.51596  -2.63607  -1.03821       -3.47278  -1.36775    45.6006

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.018293895867545977
   0.034710652946558826
   0.4088794705207291
   0.45700466779709836
   0.04555879891409935
   0.08644280415823184
   1.0182662956229105
   1.138116446803944
   0.11167601756482642
   0.21189294594287464
   ⋮
  15.558936945108234
   0.801920640389719
   1.521556110343361
  17.923404025923144
  20.0329923442437
  11.443476971278145
  21.712737435913773
 255.76852731697696
 285.87253527448075
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0382158    0.078559    -0.0282454  …   0.264734   -0.0951836   0.123661
 -0.0356836    0.0504424    0.0762295      0.169985    0.256884   -0.0409481
  0.0832277   -0.0179563    0.0491881     -0.0605107   0.165758   -0.0349555
  0.0107578    0.0271655   -0.0273522      0.0915444  -0.0921736  -0.304681
  0.015954     0.032796    -0.0117916      0.23889    -0.0858913   0.111588
 -0.0148968    0.0210582    0.0318236  …   0.15339     0.231806   -0.0369506
  0.0347451   -0.00749624   0.0205346     -0.0546034   0.149576   -0.031543
  0.00449106   0.0113408   -0.0114187      0.0826074  -0.0831752  -0.274937
  0.125058     0.257078    -0.0924308      0.126142   -0.0453535   0.0589224
 -0.116772     0.165069     0.249455       0.0809953   0.122401   -0.0195112
  ⋮                                    ⋱
  0.0503657    0.127183    -0.128057       0.0990941  -0.0997752  -0.329808
 -0.124508    -0.255946     0.092024       0.246584   -0.0886577   0.115182
  0.116258    -0.164342    -0.248357       0.158331    0.239272   -0.0381407
 -0.271157     0.0585021   -0.160256      -0.056362    0.154393   -0.0325589
 -0.0350491   -0.0885056    0.0891139  …   0.085268   -0.0858541  -0.283792
 -0.0479488   -0.0985669    0.0354391      0.307305   -0.11049     0.143546
  0.0447717   -0.0632895   -0.0956442      0.197319    0.298193   -0.0475328
 -0.104425     0.0225296   -0.0617156     -0.0702411   0.192413   -0.0405765
 -0.0134977   -0.0340842    0.0343185      0.106265   -0.106996   -0.353675

julia> logdet(E)
20.29828528857088

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  0.3591765278804128
  2.435347984093497
 -4.13561274516506
  0.714534857040826
  2.2522601346438402
 -2.448557054666883
  6.7776901334491475
 -0.15850221419746913
  0.016568380122839156
  3.0549389802069116
  ⋮
  2.1245862423995163
 -0.759365755274069
 -0.2622097457632763
 -0.44382264323102993
 -1.1924377805091348
 -1.370121632481963
  3.277218629213215
 -7.056502825250268
 -0.4568851508586675

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 37.2695     -7.24541   19.5801    …   -4.87577   13.1763     -1.14685
 -7.24541    15.3598    -0.604838      10.3363    -0.407023   -6.7272
 19.5801     -0.604838  21.0991        -0.407023  14.1986     -2.89447
 -1.70423    -9.99664   -4.3012        -6.7272    -2.89447     6.66241
 20.1061     -3.90875   10.563         -4.86696   13.1525     -1.14478
 -3.90875     8.28629   -0.326298  …   10.3176    -0.406288   -6.71505
 10.563      -0.326298  11.3825        -0.406288  14.1729     -2.88924
 -0.919396   -5.39297   -2.3204        -6.71505   -2.88924     6.65038
 34.9166     -6.78799   18.3439        -6.96792   18.8302     -1.63896
 -6.78799    14.3901    -0.566653      14.7715    -0.581674   -9.61378
  ⋮                                ⋱
 -1.77701   -10.4236    -4.48489       -8.90755   -3.8326      8.82177
 28.8306     -5.60482   15.1465        -5.821     15.7307     -1.36919
 -5.60482    11.8819    -0.467884      12.3402    -0.48593    -8.03135
 15.1465     -0.467884  16.3216        -0.48593   16.9512     -3.45561
 -1.31834    -7.73309   -3.32727   …   -8.03135   -3.45561     7.95401
 25.0804     -4.87577   13.1763        -7.42965   20.0779     -1.74757
 -4.87577    10.3363    -0.407023      15.7504    -0.620218  -10.2508
 13.1763     -0.407023  14.1986        -0.620218  21.6356     -4.41057
 -1.14685    -6.7272    -2.89447      -10.2508    -4.41057    10.1521

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 6.10488  -1.18682  3.20728   -0.279159  …  -0.798668  2.15833   -0.187859
 0.0       3.73514  0.857166  -2.76508       2.51355   0.576827  -1.86075
 0.0       0.0      3.17455   -0.326259      0.0       2.1363    -0.219555
 0.0       0.0      0.0        1.43887       0.0       0.0        0.96828
 0.0      -0.0      0.0       -0.0          -0.524439  1.41725   -0.123356
 0.0       0.0      0.0       -0.0       …   1.6505    0.378769  -1.22185
 0.0       0.0      0.0       -0.0           0.0       1.40279   -0.144169
 0.0       0.0      0.0        0.0           0.0       0.0        0.635814
 0.0      -0.0      0.0       -0.0          -0.370849  1.00219   -0.0872293
 0.0       0.0      0.0       -0.0           1.16713   0.267841  -0.86401
 ⋮                                       ⋱
 0.0       0.0      0.0        0.0           0.0       0.0        0.216141
 0.0      -0.0      0.0       -0.0          -0.351008  0.948566  -0.0825623
 0.0       0.0      0.0       -0.0           1.10468   0.25351   -0.817783
 0.0       0.0      0.0       -0.0           0.0       0.938885  -0.0964924
 0.0       0.0      0.0        0.0       …   0.0       0.0        0.425551
 0.0      -0.0      0.0       -0.0          -0.296185  0.800412  -0.0696672
 0.0       0.0      0.0       -0.0           0.932145  0.213915  -0.690056
 0.0       0.0      0.0       -0.0           0.0       0.792244  -0.0814216
 0.0       0.0      0.0        0.0           0.0       0.0        0.359085

julia> logdet(C)
12.251999984783374

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.98149    7.20179   -4.98664   …   -2.03066    1.40606    -1.7784
  7.20179   32.3806     1.27587       -9.13022   -0.35975    -9.72484
 -4.98664    1.27587   10.3944        -0.35975   -2.93086    -1.39452
  6.30716   34.4895     4.9457        -9.72484   -1.39452   -13.5689
  3.14467    3.24389   -2.24613       -1.24919    0.864961   -1.09401
  3.24389   14.5852     0.574687  …   -5.6166    -0.221306   -5.98239
 -2.24613    0.574687   4.68193       -0.221306  -1.80296    -0.85786
  2.84093   15.5351     2.22769       -5.98239   -0.85786    -8.3471
  1.3715     1.41478   -0.979615      -0.376999   0.261041   -0.330167
  1.41478    6.3611     0.250641      -1.69506   -0.066789   -1.80546
  ⋮                               ⋱
 -3.66817  -20.0587    -2.87636        4.74635    0.680614    6.62248
  9.7671    10.0753    -6.9763        -3.28027    2.27131    -2.87278
 10.0753    45.3005     1.78493      -14.7487    -0.581131  -15.7093
 -6.9763     1.78493   14.5417        -0.581131  -4.73443    -2.25267
  8.82371   48.2507     6.91903   …  -15.7093    -2.25267   -21.9188
 -1.96854   -2.03066    1.40606        1.16064   -0.803647    1.01646
 -2.03066   -9.13022   -0.35975        5.21846    0.205619    5.55832
  1.40606   -0.35975   -2.93086        0.205619   1.67516     0.797049
 -1.7784    -9.72484   -1.39452        5.55832    0.797049    7.75541

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source