Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 37.7423   19.0003     3.67906    5.52113   …  11.905      2.30517    3.45936
 19.0003   16.7274     4.4444     1.7716       10.4808     2.78471    1.11003
  3.67906   4.4444     5.48266   -2.19099       2.78471    3.43525   -1.3728
  5.52113   1.7716    -2.19099    5.03038       1.11003   -1.3728     3.15187
 24.455    12.3112     2.38383    3.5774       10.0879     1.95333    2.93134
 12.3112   10.8385     2.87973    1.1479    …   8.88112    2.35967    0.9406
  2.38383   2.87973    3.55246   -1.41964       2.35967    2.91092   -1.16327
  3.5774    1.1479    -1.41964    3.25941       0.9406    -1.16327    2.67079
 27.3662   13.7768     2.66762    4.00327      14.9641     2.89752    4.34829
 13.7768   12.1288     3.22255    1.28456      13.174      3.50028    1.39526
  ⋮                                         ⋱
  3.19518   1.02526   -1.26796    2.91117       0.714646  -0.883822   2.0292
 16.0721    8.09107    1.56669    2.35111      14.9944     2.90339    4.35709
  8.09107   7.12319    1.89259    0.754417     13.2007     3.50737    1.39809
  1.56669   1.89259    2.33473   -0.933008      3.50737    4.32673   -1.72906
  2.35111   0.754417  -0.933008   2.14213   …   1.39809   -1.72906    3.96981
 23.648    11.905      2.30517    3.45936      16.7359     3.2406     4.86314
 11.905    10.4808     2.78471    1.11003      14.7339     3.91473    1.56047
  2.30517   2.78471    3.43525   -1.3728        3.91473    4.82925   -1.92988
  3.45936   1.11003   -1.3728     3.15187       1.56047   -1.92988    4.43088
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0012869588710595515
   0.003042363435682516
   0.006794293457815466
   0.037936753625635936
   0.02050268322670221
   0.0484682262852312
   0.10824063584888241
   0.604374591703536
   0.0407426710907861
   0.09631544223056293
   ⋮
  22.208063677164397
   1.3291923932325032
   3.142203241373656
   7.0172585811406165
  39.181706173419386
  14.993987736883641
  35.445701546165395
  79.15835935254208
 441.9902076370551
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0125172    0.0730308   0.0486119  …  -0.119307   -0.0794153  -0.232526
  0.0323318   -0.116913   -0.079373       0.190995    0.129668   -0.137639
 -0.116941     0.0533531  -0.105952      -0.0871608   0.17309    -0.0311951
 -0.115044    -0.0791442   0.0906824      0.129295   -0.148144   -0.0322721
  0.0277792    0.162076    0.107883      -0.116208   -0.077352   -0.226485
  0.0717533   -0.259461   -0.176151   …   0.186033    0.1263     -0.134063
 -0.259526     0.118405   -0.235138      -0.0848963   0.168593   -0.0303847
 -0.255314    -0.175643    0.201249       0.125935   -0.144295   -0.0314337
  0.0275666    0.160835    0.107058      -0.145011   -0.0965248  -0.282622
  0.0712041   -0.257475   -0.174802       0.232144    0.157605   -0.167293
  ⋮                                   ⋱
  0.163828     0.112705   -0.129136       0.108947   -0.12483    -0.0271934
 -0.00994611  -0.0580298  -0.0386267     -0.11469    -0.0763419  -0.223527
 -0.0256907    0.092898    0.0630693      0.183604    0.12465    -0.132313
  0.092921    -0.0423941   0.0841892     -0.0837876   0.166391   -0.0299879
  0.0914131    0.0628875  -0.0720557  …   0.124291   -0.142411   -0.0310232
  0.0227322    0.132629    0.0882829     -0.116778   -0.0777319  -0.227597
  0.0587171   -0.212322   -0.144147       0.186947    0.12692    -0.134722
 -0.212374     0.0968932  -0.192418      -0.0853132   0.169421   -0.0305339
 -0.208928    -0.143732    0.164686       0.126554   -0.145004   -0.031588
julia> logdet(E)-6.480214238681441
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  0.4674646183987768
 -1.6071347218039742
 -2.4497924017933466
 -1.284457069868774
  1.497242147065056
 -4.047437138122633
 -1.4961945081144707
  2.3873661696303574
  2.146496464935005
 -1.6992753405434575
  ⋮
 -2.5739019842161333
  1.5087552943921745
 -1.9182732948059271
 -0.6014076316996528
 -3.336253205008935
 -0.8279082881300855
 -1.2864900622409583
  0.14420486015708234
  3.241387074557881

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 12.8507   15.6932     3.6888     7.16727  …  11.1061     2.61057    5.07229
 15.6932   41.6528    -1.30964   17.721       29.4777    -0.926834  12.5412
  3.6888   -1.30964   21.409      6.55943     -0.926834  15.1512     4.64212
  7.16727  17.721      6.55943   10.0735      12.5412     4.64212    7.12904
  9.34651  11.414      2.68293    5.21288      9.57801    2.25138    4.37439
 11.414    30.2948    -0.952525  12.8888   …  25.4219    -0.79931   10.8156
  2.68293  -0.952525  15.5712     4.77079     -0.79931   13.0665     4.00341
  5.21288  12.8888     4.77079    7.32665     10.8156     4.00341    6.14815
  8.39389  10.2506     2.40948    4.68157     10.8607     2.55288    4.9602
 10.2506   27.2071    -0.855441  11.5751      28.8264    -0.906354  12.264
  ⋮                                        ⋱
  4.52406  11.1857     4.14038    6.35851     11.4819     4.25003    6.5269
 11.5357   14.0874     3.31133    6.43385     15.2326     3.58053    6.9569
 14.0874   37.3906    -1.17563   15.9076      40.4303    -1.2712    17.2008
  3.31133  -1.17563   19.2183     5.88822     -1.2712    20.7806     6.36691
  6.43385  15.9076     5.88822    9.04271  …  17.2008     6.36691    9.77785
  9.09443  11.1061     2.61057    5.07229     15.7212     3.69538    7.18005
 11.1061   29.4777    -0.926834  12.5412      41.7271    -1.31198   17.7526
  2.61057  -0.926834  15.1512     4.64212     -1.31198   21.4472     6.57113
  5.07229  12.5412     4.64212    7.12904     17.7526     6.57113   10.0915
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.58478  4.37773   1.02902  1.99936  …   3.09813    0.728236    1.41495
 0.0      4.74218  -1.2261   1.89118      3.35605   -0.867714    1.33839
 0.0      0.0       4.34129  1.57115      0.0        3.07234     1.11191
 0.0      0.0       0.0      0.17605      0.0        0.0         0.124591
 0.0      0.0       0.0      0.0          1.114      0.261854    0.508777
 0.0      0.0      -0.0      0.0      …   1.20674   -0.312006    0.481248
 0.0      0.0       0.0      0.0          0.0        1.10473     0.399811
 0.0      0.0       0.0      0.0          0.0        0.0         0.0447994
 0.0      0.0       0.0      0.0          2.34979    0.552335    1.07318
 0.0      0.0      -0.0      0.0          2.54541   -0.658123    1.01511
 ⋮                                    ⋱
 0.0      0.0       0.0      0.0         -0.0       -0.0        -0.0218861
 0.0      0.0       0.0      0.0         -0.656453  -0.154304   -0.299809
 0.0      0.0      -0.0      0.0         -0.711102   0.183857   -0.283587
 0.0      0.0       0.0      0.0         -0.0       -0.650988   -0.235599
 0.0      0.0       0.0      0.0      …  -0.0       -0.0        -0.0263991
 0.0      0.0       0.0      0.0          0.353344   0.0830559   0.161376
 0.0      0.0      -0.0      0.0          0.38276   -0.0989635   0.152644
 0.0      0.0       0.0      0.0          0.0        0.350403    0.126814
 0.0      0.0       0.0      0.0          0.0        0.0         0.0142097
julia> logdet(C)-60.00295809104969
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  139.802    -388.771   -283.494  …    -358.153   -261.167    708.475
 -388.771    1272.31     931.702       1172.11     858.325  -2366.01
 -283.494     931.702    691.779        858.325    637.297  -1739.1
  769.042   -2568.28   -1887.77       -2366.01   -1739.1     4805.34
  241.843    -672.533   -490.415       -613.341   -447.252   1213.27
 -672.533    2200.97    1611.75   …    2007.25    1469.89   -4051.83
 -490.415    1611.75    1196.71        1469.89    1091.38   -2978.22
 1330.36    -4442.86   -3265.64       -4051.83   -2978.22    8229.2
 -974.856    2710.94    1976.84        2609.38    1902.77   -5161.71
 2710.94    -8871.98   -6496.87       -8539.58   -6253.46   17238.0
    ⋮                             ⋱
 3600.3    -12023.5    -8837.68      -11314.3    -8316.4    22979.2
  301.565    -838.612   -611.522       -797.676   -581.67    1577.91
 -838.612    2744.49    2009.76        2610.52    1911.66   -5269.57
 -611.522    2009.76    1492.23        1911.66    1419.39   -3873.31
 1658.89    -5540.01   -4072.08   …   -5269.57   -3873.31   10702.4
  128.792    -358.153   -261.167       -369.128   -269.17     730.186
 -358.153    1172.11     858.325       1208.03     884.628  -2438.52
 -261.167     858.325    637.297        884.628    656.826  -1792.39
  708.475   -2366.01   -1739.1        -2438.52   -1792.39    4952.59

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.