Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 36.7305   -5.99964  12.6117    4.83294  …  -4.40647   9.26278   3.54959
 -5.99964   7.42202  -5.47194   5.36009      5.45115  -4.0189    3.93675
 12.6117   -5.47194   6.45567  -3.02456     -4.0189    4.74141  -2.22141
  4.83294   5.36009  -3.02456  12.9609       3.93675  -2.22141   9.5192
 17.5443   -2.86571   6.02397   2.30844     -4.07329   8.5624    3.28119
 -2.86571   3.54511  -2.61366   2.56023  …   5.03898  -3.71502   3.63909
  6.02397  -2.61366   3.08354  -1.44467     -3.71502   4.3829   -2.05344
  2.30844   2.56023  -1.44467   6.19073      3.63909  -2.05344   8.79943
 22.6945   -3.70697   7.79236   2.98611     -4.56741   9.60108   3.67923
 -3.70697   4.58581  -3.38092   3.31181      5.65024  -4.16568   4.08053
  ⋮                                      ⋱
  3.41793   3.79074  -2.13901   9.16613      4.00905  -2.2622    9.69401
 31.2795   -5.10925  10.7401    4.1157      -5.7772   12.1442    4.65376
 -5.10925   6.32054  -4.65986   4.56461      7.14685  -5.26906   5.16136
 10.7401   -4.65986   5.4976   -2.57569     -5.26906   6.21633  -2.91242
  4.1157    4.56461  -2.57569  11.0374   …   5.16136  -2.91242  12.4804
 26.977    -4.40647   9.26278   3.54959     -5.74288  12.072     4.62612
 -4.40647   5.45115  -4.0189    3.93675      7.10439  -5.23776   5.1307
  9.26278  -4.0189    4.74141  -2.22141     -5.23776   6.1794   -2.89512
  3.54959   3.93675  -2.22141   9.5192       5.1307   -2.89512  12.4062
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   3.73662309987164e-6
   0.004148096202902357
   0.030352400701427033
   0.07250243307753437
   1.3365560201476202e-5
   0.014837361981547137
   0.10856777041500043
   0.2593345938702175
   7.201418937281007e-5
   0.07994431804018616
   ⋮
  20.37348435460429
   0.001230750209963162
   1.366279160679825
   9.997321789679724
  23.880488437830625
   0.014257290821458397
  15.827289062736293
 115.81125319932549
 276.63701851151154
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0464451   0.0388435    0.0225863  …  -0.0810307  -0.0471168  -0.310493
  0.0461938   0.130672     0.0782572     -0.272593   -0.163251    0.0658429
  0.144953   -0.00986205  -0.0469417      0.020573    0.0979242  -0.114587
  0.0320468  -0.0874539    0.132256       0.182436   -0.275897   -0.0266053
 -0.0834919   0.069827     0.0406022     -0.0734383  -0.042702   -0.2814
  0.0830402   0.234903     0.140679   …  -0.247051   -0.147954    0.0596735
  0.260575   -0.0177285   -0.0843846      0.0186454   0.0887488  -0.10385
  0.0576089  -0.157211     0.23775        0.165342   -0.250046   -0.0241124
  0.184947   -0.154677    -0.0899399     -0.0747538  -0.043467   -0.286441
 -0.183946   -0.520345    -0.311625      -0.251477   -0.150605    0.0607424
  ⋮                                   ⋱
  0.0184394  -0.0503202    0.0760989      0.168336   -0.254573   -0.0245489
  0.104997   -0.0878122   -0.05106       -0.0962991  -0.0559949  -0.368998
 -0.104429   -0.295406    -0.176913      -0.323957   -0.194012    0.0782494
 -0.327691    0.0222948    0.106119       0.0244496   0.116376   -0.136178
 -0.0724471   0.197704    -0.298987   …   0.216812   -0.327883   -0.0316184
 -0.131342    0.109846     0.0638718     -0.0857193  -0.0498431  -0.328459
  0.130631    0.369528     0.221304      -0.288366   -0.172697    0.0696527
  0.409914   -0.0278889   -0.132746       0.0217634   0.10359    -0.121217
  0.0906252  -0.247311     0.374007       0.192992   -0.291861   -0.0281447
julia> logdet(E)-73.99719703354722
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -2.6680551098909895
 -0.19037200887935146
  1.767039096908353
  1.4586643213861028
 -6.578022846479631
  5.9781396490907355
 12.323902994455482
  2.51365434383844
  3.9457679748282093
 -4.334563481254489
  ⋮
  1.9360518014775658
  2.5660069967113825
  4.423269942419144
  5.4878282961723155
 -2.296424879433462
 -4.808499522606864
 -0.43311952933149306
  1.8747351238048195
  0.3631840177834321

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.00657    -1.03579   -0.448008  …   -0.850069  -0.367681    0.137814
 -1.03579    15.4855    -4.88866       12.7089    -4.01213    -8.9147
 -0.448008   -4.88866   16.5212        -4.01213   13.559       4.94416
  0.167923  -10.8623     6.02431       -8.9147     4.94416     8.66506
  1.04296    -0.538374  -0.232863      -0.645082  -0.279018    0.104582
 -0.538374    8.04893   -2.541     …    9.64427   -3.04464    -6.76499
 -0.232863   -2.541      8.58729       -3.04464   10.2893      3.75191
  0.087282   -5.64594    3.13128       -6.76499    3.75191     6.57556
  2.05222    -1.05935   -0.458203      -1.06682   -0.461431    0.172954
 -1.05935    15.8378    -4.9999        15.9494    -5.03513   -11.1878
  ⋮                                ⋱
  0.161845  -10.4692     5.80627      -10.8527     6.01899    10.5488
  1.36715    -0.705722  -0.305246      -0.999791  -0.43244     0.162088
 -0.705722   10.5509    -3.33084       14.9473    -4.71877   -10.4848
 -0.305246   -3.33084   11.2566        -4.71877   15.9471      5.81496
  0.114413   -7.40092    4.1046    …  -10.4848     5.81496    10.1912
  1.64679    -0.850069  -0.367681      -1.35263   -0.585053    0.21929
 -0.850069   12.7089    -4.01213       20.2224    -6.38409   -14.1851
 -0.367681   -4.01213   13.559         -6.38409   21.575       7.86714
  0.137814   -8.9147     4.94416      -14.1851     7.86714    13.7879
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.41653  -0.731212  -0.316271  …  -0.600106   -0.259564    0.09729
 0.0       3.86662   -1.32413       3.17334    -1.08672    -2.28715
 0.0       0.0        3.82986       0.0         3.14317     0.508225
 0.0       0.0        0.0           0.0         0.0         1.26988
 0.0      -0.0       -0.0          -0.244954   -0.10595     0.0397122
 0.0       0.0       -0.0       …   1.29531    -0.44358    -0.933579
 0.0       0.0        0.0           0.0         1.28299     0.20745
 0.0       0.0        0.0           0.0         0.0         0.518345
 0.0      -0.0       -0.0          -0.199303   -0.0862047   0.0323114
 0.0       0.0       -0.0           1.05391    -0.360913   -0.759595
 ⋮                              ⋱
 0.0       0.0        0.0           0.0         0.0         0.431746
 0.0      -0.0       -0.0          -0.239951   -0.103786    0.0389013
 0.0       0.0       -0.0           1.26885    -0.434522   -0.914514
 0.0       0.0        0.0           0.0         1.25679     0.203213
 0.0       0.0        0.0       …   0.0         0.0         0.50776
 0.0      -0.0       -0.0          -0.0943253  -0.0407985   0.0152922
 0.0       0.0       -0.0           0.498789   -0.170811   -0.359497
 0.0       0.0        0.0           0.0         0.494047    0.0798834
 0.0       0.0        0.0           0.0         0.0         0.199601
julia> logdet(C)-11.963670220989755
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   60.5021     11.7061     1.34339   …   -7.40512  -0.849804    -6.52486
   11.7061     28.683     -2.38355      -18.1444    1.50779    -19.4097
    1.34339    -2.38355    8.73274        1.50779  -5.52419      4.71677
   10.3146     30.6831    -7.45635      -19.4097    4.71677    -29.1713
   44.8193      8.67179    0.995166      -6.07128  -0.696734    -5.34957
    8.67179    21.248     -1.7657    …  -14.8761    1.2362     -15.9135
    0.995166   -1.7657     6.46912        1.2362   -4.52915      3.86716
    7.64095    22.7297    -5.52358      -15.9135    3.86716    -23.9168
 -104.597     -20.2377    -2.32246       14.2201    1.63188     12.5297
  -20.2377    -49.5874     4.1207        34.8428   -2.89542     37.2725
    ⋮                                ⋱
   32.3886     96.3472   -23.4135       -71.1142   17.2816    -106.88
   32.7273      6.33219    0.726675      -5.1745   -0.593821    -4.5594
    6.33219    15.5154    -1.28933      -12.6788    1.05361    -13.563
    0.726675   -1.28933    4.72378        1.05361  -3.86016      3.29595
    5.57947    16.5974    -4.03335   …  -13.563     3.29595    -20.3841
  -38.2726     -7.40512   -0.849804       6.37161   0.7312       5.6142
   -7.40512   -18.1444     1.50779       15.612    -1.29735     16.7007
   -0.849804    1.50779   -5.52419       -1.29735   4.7532      -4.05846
   -6.52486   -19.4097     4.71677       16.7007   -4.05846     25.1

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.