Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  17.6275   -11.5504   -9.64097   …  -9.30654  -7.76803   -10.7905
 -11.5504    36.7548    7.01271      29.6145    5.65036    12.1232
  -9.64097    7.01271  13.9532        5.65036  11.2425      0.278026
 -13.3922    15.0463    0.345061     12.1232    0.278026   37.3738
  12.0406    -7.88958  -6.58532      -8.04833  -6.71783    -9.33167
  -7.88958   25.1056    4.79007   …  25.6107    4.88645    10.4842
  -6.58532    4.79007   9.53078       4.88645   9.72256     0.240438
  -9.1476    10.2774    0.235696     10.4842    0.240438   32.321
  12.4449    -8.15456  -6.80649      -6.8646   -5.72978    -7.95919
  -8.15456   25.9488    4.95095      21.844     4.16777     8.94223
   ⋮                              ⋱
  -5.93156    6.66416   0.152831      5.15356   0.118188   15.8875
  12.7655    -8.36461  -6.98182      -6.56158  -5.47686    -7.60785
  -8.36461   26.6172    5.07848      20.8797    3.98379     8.5475
  -6.98182    5.07848  10.1046        3.98379   7.92653     0.196023
  -9.69838   10.8962    0.249887  …   8.5475    0.196023   26.3504
  14.203     -9.30654  -7.76803      -9.98385  -8.33338   -11.5758
  -9.30654   29.6145    5.65036      31.7698    6.06158    13.0055
  -7.76803    5.65036  11.2425        6.06158  12.0607      0.29826
 -10.7905    12.1232    0.278026     13.0055    0.29826    40.0938

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.00045704423995791116
   0.0018587559214928069
   0.003159089979229802
   0.007218707745456753
   0.025536967810513886
   0.10385644536103349
   0.17651153226991242
   0.40333930769199156
   0.05554654546321358
   0.22590257413095421
   ⋮
  20.997800854100877
   1.8194688997900135
   7.399608825106582
  12.576169802241072
  28.73729300415506
  24.048403225348242
  97.80259324935187
 166.2225732298287
 379.8284267725922
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.391872     -0.323545     …  -0.231296    -0.0742052  -0.148036
 -0.00164117   -0.309885        -0.221531     0.242813    0.226634
 -0.379222      0.333374         0.238322     0.159424    0.0607483
 -0.12052       0.00725333       0.00518527  -0.263663    0.287107
 -0.0887687    -0.073291        -0.186582    -0.0598598  -0.119417
 -0.000371766  -0.0701966    …  -0.178704     0.195872    0.182821
 -0.0859032     0.0755174        0.19225      0.128604    0.0490044
 -0.0273007     0.00164306       0.00418285  -0.212691    0.231603
  0.0285196     0.0235469       -0.171028    -0.0548697  -0.109462
  0.000119441   0.0225528       -0.163807     0.179543    0.167581
  ⋮                          ⋱
  0.107231     -0.00645358       0.00238474  -0.12126     0.132042
  0.0177784     0.0146785       -0.16526     -0.0530191  -0.105771
  7.44564e-5    0.0140588       -0.158282     0.173488    0.161929
  0.0172045    -0.0151244        0.17028      0.113907    0.0434042
  0.00546771   -0.000329068  …   0.00370484  -0.188385    0.205136
  0.214634      0.177211        -0.210394    -0.0674993  -0.134658
  0.000898897   0.169729        -0.201511     0.220869    0.206153
  0.207706     -0.182594         0.216785     0.145017    0.0552584
  0.0660105    -0.00397276       0.00471667  -0.239836    0.261161

julia> logdet(E)
8.428651543130142

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -4.038420106208041
 -7.852692343902399
 -3.8833407649950065
 -0.21201113815774605
 -0.39901712269430734
 -0.5909757360526172
 -1.0095077370581034
 -1.2538088718774167
 -0.9796386824815213
  1.2293335305508013
  ⋮
  8.211387812126697
 -0.1814214163423546
  2.143446149217612
 -0.5953402365221909
 -2.394976138958853
  5.498579196747583
  1.1390707667961317
  3.29883708881446
  6.82839987175902

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.69734   -1.01003   -3.78375    …  -0.772891  -2.89538    -2.81243
 -1.01003   18.4047     8.09634       14.0835     6.19544    -4.78257
 -3.78375    8.09634    5.12089        6.19544    3.91858     0.0353986
 -3.67535   -6.24998    0.0462598     -4.78257    0.0353986   7.45578
  8.81838   -1.02409   -3.83641       -0.586293  -2.19635    -2.13343
 -1.02409   18.6608     8.20902    …  10.6834     4.69968    -3.62793
 -3.83641    8.20902    5.19216        4.69968    2.97252     0.0268524
 -3.7265    -6.33696    0.0469036     -3.62793    0.0268524   5.65575
  7.38794   -0.857969  -3.2141        -0.794308  -2.97561    -2.89037
 -0.857969  15.6338     6.87742       14.4738     6.36712    -4.9151
  ⋮                                ⋱
 -3.70939   -6.30785    0.0466882     -4.11027    0.0304226   6.4077
  7.21819   -0.838256  -3.14025       -0.422445  -1.58255    -1.53721
 -0.838256  15.2746     6.7194         7.69774    3.38629    -2.61405
 -3.14025    6.7194     4.24999        3.38629    2.14181     0.0193481
 -3.05029   -5.18705    0.0383924  …  -2.61405    0.0193481   4.07517
  6.65533   -0.772891  -2.89538       -0.925925  -3.46867    -3.3693
 -0.772891  14.0835     6.19544       16.8721     7.42215    -5.72953
 -2.89538    6.19544    3.91858        7.42215    4.69447     0.0424077
 -2.81243   -4.78257    0.0353986     -5.72953    0.0424077   8.93204

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.94913  -0.342485  -1.28301   …  -0.262074   -0.981776   -0.953649
 0.0       4.27638    1.79052       3.27234     1.37013    -1.19475
 0.0       0.0        0.518482      0.0         0.39675     1.83435
 0.0       0.0        0.0           0.0         0.0         0.0598278
 0.0      -0.0       -0.0           0.117613    0.440598    0.427975
 0.0       0.0        0.0       …  -1.46855    -0.614882    0.536174
 0.0       0.0        0.0          -0.0        -0.178052   -0.823212
 0.0       0.0        0.0          -0.0        -0.0        -0.0268493
 0.0      -0.0       -0.0          -0.0217454  -0.0814621  -0.0791283
 0.0       0.0        0.0           0.27152     0.113686   -0.0991331
 ⋮                              ⋱
 0.0       0.0        0.0          -0.0        -0.0        -0.0138373
 0.0      -0.0       -0.0           0.125529    0.470254    0.456782
 0.0       0.0        0.0          -1.5674     -0.65627     0.572264
 0.0       0.0        0.0          -0.0        -0.190037   -0.878622
 0.0       0.0        0.0       …  -0.0        -0.0        -0.0286565
 0.0      -0.0       -0.0          -0.0331629  -0.124234   -0.120675
 0.0       0.0        0.0           0.414082    0.173376   -0.151183
 0.0       0.0        0.0           0.0         0.0502047   0.232118
 0.0       0.0        0.0           0.0         0.0         0.00757061

julia> logdet(C)
-99.04748159940002

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   32732.4         -56921.7        …  -91074.9         19670.3
  -56921.7          99063.2                1.58476e5  -34245.9
       1.14404e5       -1.9907e5          -3.1849e5    68812.8
  -24708.9          43018.3            68812.8        -14883.5
   90529.0        -157430.0               -3.99023e5   86180.4
 -157430.0              2.73982e5  …       6.94322e5      -1.5004e5
       3.16411e5  -550574.0               -1.39539e6       3.01486e5
  -68338.2              1.18977e5          3.01486e5  -65208.3
   36279.7         -63090.4               -1.71648e5   37072.4
  -63090.4              1.09799e5          2.98678e5  -64543.1
       ⋮                           ⋱
   -3884.72          6763.3            15082.9         -3262.26
      -1.03304e5   179645.0                4.34176e5  -93772.9
  179645.0             -3.12643e5         -7.55492e5       1.63259e5
      -3.6106e5         6.28266e5          1.51832e6      -3.28047e5
   77981.4             -1.35766e5  …      -3.28047e5   70953.1
  -26057.6          45314.2                1.06766e5  -23059.2
   45314.2         -78862.1               -1.85779e5   40146.0
  -91074.9              1.58476e5          3.73361e5  -80668.3
   19670.3         -34245.9           -80668.3         17447.7

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.