Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  6.23893  -4.93439  -3.77019   5.75744  …  -3.89854  -2.97873   4.54881
 -4.93439  25.8396   -4.82837   6.41811     20.4152   -3.81478   5.07079
 -3.77019  -4.82837  12.0725    3.72077     -3.81478   9.53818   2.93969
  5.75744   6.41811   3.72077  35.7371       5.07079   2.93969  28.235
  4.43494  -3.50761  -2.68003   4.09267     -3.25276  -2.48532   3.79532
 -3.50761  18.3681   -3.43224   4.56231  …  17.0335   -3.18288   4.23084
 -2.68003  -3.43224   8.58173   2.6449      -3.18288   7.95823   2.45274
  4.09267   4.56231   2.6449   25.4037       4.23084   2.45274  23.558
  4.7515   -3.75797  -2.87133   4.3848      -3.54177  -2.70613   4.13253
 -3.75797  19.6791   -3.67723   4.88796     18.5469   -3.46567   4.60674
  ⋮                                      ⋱
  2.93667   3.27366   1.89783  18.2283       4.40079   2.55127  24.5043
  5.78636  -4.57645  -3.4967    5.3398      -4.4104   -3.36983   5.14606
 -4.57645  23.9652   -4.47812   5.95254     23.0957   -4.31564   5.73656
 -3.4967   -4.47812  11.1968    3.45086     -4.31564  10.7905    3.32566
  5.3398    5.95254   3.45086  33.1448   …   5.73656   3.32566  31.9422
  4.92923  -3.89854  -2.97873   4.54881     -4.61349  -3.525     5.38302
 -3.89854  20.4152   -3.81478   5.07079     24.1592   -4.51437   6.00072
 -2.97873  -3.81478   9.53818   2.93969     -4.51437  11.2874    3.47879
  4.54881   5.07079   2.93969  28.235        6.00072   3.47879  33.413

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0006260673357550224
   0.01711815471831221
   0.033390416972050675
   0.0493561453109746
   0.009377594743702167
   0.25640551509484694
   0.500140768899097
   0.739284582351544
   0.012137654798246187
   0.33187205415096754
   ⋮
   9.67563627259548
   0.36568254969076075
   9.99862172308375
  19.50316222591587
  28.82865792456428
   3.586179079355277
  98.05457787371171
 191.26379537387785
 282.7171545986796
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0057377   -0.00387644   …  -0.17885    -0.0808173  -0.0320884
  0.00217586   0.000927597      0.0427973   0.285061   -0.12829
  0.0033179    0.00596184       0.275066   -0.101631   -0.0142744
 -0.00168404  -0.000262903     -0.0121298  -0.107274   -0.303209
 -0.147556     0.0996901       -0.150449   -0.0679836  -0.0269928
 -0.0559564   -0.0238549    …   0.0360011   0.239793   -0.107918
 -0.0853261   -0.15332          0.231386   -0.085492   -0.0120077
  0.0433082    0.00676105      -0.0102036  -0.0902389  -0.25506
 -0.085887     0.0580261       -0.151499   -0.0684579  -0.0271811
 -0.0325703   -0.0138851        0.0362523   0.241467   -0.108671
  ⋮                         ⋱
  0.055137     0.00860769      -0.0107272  -0.0948702  -0.26815
  0.132353    -0.0894187       -0.198065   -0.0894998  -0.0355358
  0.050191     0.0213971        0.0473952   0.315686   -0.142073
  0.0765346    0.137523         0.304618   -0.11255    -0.015808
 -0.038846    -0.00606443   …  -0.0134329  -0.118799   -0.335785
  0.404899    -0.273554        -0.182108   -0.0822894  -0.0326729
  0.153547     0.0654589        0.0435768   0.290253   -0.130627
  0.234138     0.420717         0.280077   -0.103482   -0.0145345
 -0.11884     -0.0185526       -0.0123507  -0.109228   -0.308733

julia> logdet(E)
-0.7758740183143118

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -5.827593261252895
 -2.013603239958193
 -4.5407726094092595
  2.57483611483257
  3.928644903757189
  0.4192973625219282
  3.099967067418263
 -2.2504920290520074
 -2.5200465043478935
 -1.1180059575090524
  ⋮
  1.0876085542958664
  3.1128858043740117
  1.4312068679003878
  1.7350245497593346
 -1.206108897922201
  2.683192940556562
  3.3975942471122154
  0.30539066602061266
 -1.7597262357989203

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  11.7946     2.98469   -10.8161    …   2.78273   -10.0843     0.360385
   2.98469   29.0131     -0.746889     27.05       -0.696353   5.83798
 -10.8161    -0.746889   10.5685       -0.696353    9.85342   -0.879137
   0.386539   6.26166    -0.942939      5.83798    -0.879137   3.59121
   8.37399    2.11909    -7.67932       2.39377    -8.67473    0.310011
   2.11909   20.5989     -0.530281  …  23.269      -0.599018   5.02196
  -7.67932   -0.530281    7.50351      -0.599018    8.47613   -0.756253
   0.274438   4.4457     -0.669474      5.02196    -0.756253   3.08924
  10.6471     2.69431    -9.76385       3.07769   -11.1532     0.398584
   2.69431   26.1905     -0.674226     29.9172     -0.770162   6.45677
   ⋮                                ⋱
   0.378676   6.13428    -0.923757      7.62571    -1.14835    4.69092
   4.76361    1.20546    -4.36843       1.59097    -5.76546    0.206042
   1.20546   11.7179     -0.301654     15.4652     -0.398124   3.33774
  -4.36843   -0.301654    4.26842      -0.398124    5.63347   -0.502627
   0.156116   2.52897    -0.380835  …   3.33774    -0.502627   2.05319
  10.9965     2.78273   -10.0843        4.40675   -15.9695     0.570707
   2.78273   27.05       -0.696353     42.8365     -1.10275    9.24505
 -10.0843    -0.696353    9.85342      -1.10275    15.6039    -1.39221
   0.360385   5.83798    -0.879137      9.24505    -1.39221    5.68705

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.43432  0.869075  -3.14942    0.112552  …  -2.93633      0.104936
 0.0      5.31581    0.374392   1.15953       0.34906      1.08107
 0.0      0.0        0.71378   -1.43263       0.665484    -1.3357
 0.0      0.0        0.0        0.664993      0.0          0.619998
 0.0      0.0       -0.0        0.0          -0.612263     0.0218806
 0.0      0.0        0.0        0.0       …   0.0727836    0.225419
 0.0      0.0        0.0       -0.0           0.138762    -0.278511
 0.0      0.0        0.0        0.0           0.0          0.129278
 0.0      0.0       -0.0        0.0          -0.505749     0.0180741
 0.0      0.0        0.0        0.0           0.0601215    0.186203
 ⋮                                        ⋱
 0.0      0.0        0.0        0.0          -0.0         -0.00593998
 0.0      0.0       -0.0        0.0          -0.939681     0.0335817
 0.0      0.0        0.0        0.0           0.111706     0.345965
 0.0      0.0        0.0       -0.0           0.212968    -0.427449
 0.0      0.0        0.0        0.0       …   0.0          0.198411
 0.0      0.0       -0.0        0.0          -0.0731949    0.00261579
 0.0      0.0        0.0        0.0           0.00870115   0.0269484
 0.0      0.0        0.0       -0.0           0.0165888   -0.0332955
 0.0      0.0        0.0        0.0           0.0          0.0154549

julia> logdet(C)
-55.94452733451596

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   2771.95    -468.258     2910.47   …  -1149.48     7144.61    2934.8
   -468.258     93.9321    -492.786       230.584   -1209.69    -555.627
   2910.47    -492.786     3082.93      -1209.69     7567.97    3102.19
   1195.54    -226.344     1263.73       -555.627    3102.19    1545.6
  11760.9    -1986.74     12348.6       -5298.7     32934.2    13528.4
  -1986.74     398.538    -2090.81   …   1062.91    -5576.25   -2561.25
  12348.6    -2090.81     13080.4       -5576.25    34885.7    14300.0
   5072.46    -960.338     5361.78      -2561.25    14300.0     7124.69
   3597.64    -607.741     3777.43      -1554.8      9663.87    3969.63
   -607.741    121.912     -639.575       311.891   -1636.24    -751.547
      ⋮                              ⋱
  -4438.52     840.318    -4691.68       2202.18   -12295.3    -6125.86
 -10263.6     1733.8     -10776.5        4623.04   -28734.6   -11803.3
   1733.8     -347.798     1824.62       -927.377    4865.2     2234.65
 -10776.5     1824.62    -11415.0        4865.2    -30437.3   -12476.6
  -4426.66     838.073    -4679.15   …   2234.65   -12476.6    -6216.19
   6804.57   -1149.48      7144.61      -3113.64    19352.9     7949.61
  -1149.48     230.584    -1209.69        624.593   -3276.74   -1505.05
   7144.61   -1209.69      7567.97      -3276.74    20499.7     8403.04
   2934.8     -555.627     3102.19      -1505.05     8403.04    4186.64

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source