Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 3.39159   4.77678  4.14893  3.39674  2.37918  …   4.83581   4.2002   3.43872
 4.77678  13.7879   6.95566  4.1519   3.35089     13.9583    7.04162  4.20321
 4.14893   6.95566  5.93669  3.54206  2.91045      7.04162   6.01006  3.58583
 3.39674   4.1519   3.54206  3.98027  2.3828       4.20321   3.58583  4.02945
 2.37918   3.35089  2.91045  2.3828   2.60047      4.86845   4.22855  3.46193
 3.35089   9.67215  4.87936  2.91254  3.66255  …  14.0525    7.08915  4.23158
 2.91045   4.87936  4.16456  2.48474  3.18115      7.08915   6.05062  3.61003
 2.3828    2.91254  2.48474  2.79214  2.60441      4.23158   3.61003  4.05665
 3.9104    5.50748  4.78359  3.91634  3.54472      6.78688   5.89483  4.82611
 5.50748  15.897    8.01966  4.78702  4.99245     19.59      9.88266  5.89905
 ⋮                                             ⋱
 5.07057   6.19786  5.2875   5.94165  4.59959      8.23571   7.02602  7.89525
 3.38347   4.76534  4.13899  3.3886   2.60411      5.86138   5.09097  4.16799
 4.76534  13.7549   6.939    4.14196  3.66768     16.9186    8.53499  5.09462
 4.13899   6.939    5.92247  3.53357  3.1856       8.53499   7.28465  4.34631
 3.3886    4.14196  3.53357  3.97073  2.60806  …   5.09462   4.34631  4.88401
 3.4335    4.83581  4.2002   3.43872  3.45668      8.29986   7.20894  5.90198
 4.83581  13.9583   7.04162  4.20321  4.86845     23.9571   12.0858   7.21411
 4.2002    7.04162  6.01006  3.58583  4.22855     12.0858   10.3152   6.15448
 3.43872   4.20321  3.58583  4.02945  3.46193      7.21411   6.15448  6.91588
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   4.225090562474745e-5
   0.0018065509621037583
   0.006145392096630355
   0.037192831265360694
   0.0002057245552144006
   0.008796305964463326
   0.029922626201812436
   0.18109620506569477
   0.0011338102160876
   0.04847910136904591
   ⋮
  15.264308822145134
   0.019680809815841034
   0.8415058892125519
   2.8625727971784167
  17.324718318404223
   0.2955200467202801
  12.635753407632684
  42.983375921933224
 260.1418140197459
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.540677     -0.00613426   …   0.00256131  -0.113612   0.0941942
  0.00942212    0.134579        -0.0561923    0.172209   0.199808
 -0.229611     -0.483991         0.202087    -0.080058   0.127721
 -0.268618      0.406081        -0.169556    -0.154206   0.0874299
 -0.00554349    6.28938e-5       0.0022242   -0.0986583  0.0817966
 -9.66037e-5   -0.00137982   …  -0.0487964    0.149544   0.17351
  0.00235417    0.00496229       0.175489    -0.069521   0.110911
  0.00275411   -0.00416349      -0.147239    -0.13391    0.0759226
 -0.058179      0.000660071      0.00348796  -0.154715   0.128272
 -0.00101386   -0.0144812       -0.0765219    0.234512   0.272096
  ⋮                          ⋱
  0.0588753    -0.0890041       -0.304251    -0.276707   0.156884
  0.184904     -0.00209783       0.00311456  -0.138152   0.11454
  0.00322222    0.0460239       -0.06833      0.209407   0.242967
 -0.0785234    -0.165518         0.245738    -0.0973508  0.155309
 -0.0918635     0.138874     …  -0.20618     -0.187515   0.106315
  0.0503815    -0.000571604      0.00330855  -0.146756   0.121674
  0.000877974   0.0125403       -0.0725858    0.22245    0.2581
 -0.0213956    -0.0450993        0.261043    -0.103414   0.164982
 -0.0250305     0.0378395       -0.219022    -0.199193   0.112937
julia> logdet(E)-64.56561460945896
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
   3.32563638888333
   1.123203115873894
   0.11690312697381224
   0.519125413914237
   7.404618971068046
   5.085716704621068
  -7.887776321198285
  -2.0301503049492675
 -12.696518846414056
  -1.106874829693252
   ⋮
  -2.367357268364937
  -0.3265277488342347
   1.670502266030072
  -0.0015933096732503938
   1.6660298782883298
  19.46851313484383
  -1.6145867996384702
  -7.708701183061485
  -8.50323759212588

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  7.23644  -2.03425   6.76191  -3.64168  …  -1.9461    6.46892  -3.48389
 -2.03425  26.9522   -2.93688  11.3069      25.7844   -2.80962  10.817
  6.76191  -2.93688   7.34328  -4.00922     -2.80962   7.02509  -3.8355
 -3.64168  11.3069   -4.00922   5.92169     10.817    -3.8355    5.6651
  4.28758  -1.20529   4.00642  -2.15769     -1.26454   4.20339  -2.26377
 -1.20529  15.9691   -1.74009   6.69934  …  16.7542   -1.82564   7.0287
  4.00642  -1.74009   4.35088  -2.37545     -1.82564   4.56478  -2.49224
 -2.15769   6.69934  -2.37545   3.50859      7.0287   -2.49224   3.68108
  4.30954  -1.21146   4.02694  -2.16874     -1.69944   5.64901  -3.04232
 -1.21146  16.0509   -1.74901   6.73366     22.5163   -2.45351   9.446
  ⋮                                      ⋱
 -2.95664   9.17998  -3.25504   4.80775     10.3194   -3.65907   5.40451
  5.78231  -1.62548   5.40314  -2.9099      -2.26275   7.52147  -4.05075
 -1.62548  21.5363   -2.34672   9.03486     29.9797   -3.26677  12.577
  5.40314  -2.34672   5.86768  -3.20358     -3.26677   8.16814  -4.45957
 -2.9099    9.03486  -3.20358   4.73175  …  12.577    -4.45957   6.58686
  6.92289  -1.9461    6.46892  -3.48389     -2.48728   8.26779  -4.45269
 -1.9461   25.7844   -2.80962  10.817       32.9545   -3.59092  13.825
  6.46892  -2.80962   7.02509  -3.8355      -3.59092   8.97863  -4.90207
 -3.48389  10.817    -3.8355    5.6651      13.825    -4.90207   7.24045
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.69006  -0.756209   2.51366   …  -0.723442    2.40474    -1.29509
 0.0       5.13618   -0.201711      4.91363    -0.192971    1.91536
 0.0       0.0        0.992013      0.0         0.949029   -0.195277
 0.0       0.0        0.0           0.0         0.0         0.188716
 0.0      -0.0        0.0          -0.0762948   0.253606   -0.136582
 0.0       0.0       -0.0       …   0.518196   -0.0203509   0.201996
 0.0       0.0        0.0           0.0         0.100085   -0.020594
 0.0       0.0        0.0           0.0         0.0         0.0199022
 0.0      -0.0        0.0          -0.26832     0.891906   -0.480343
 0.0       0.0       -0.0           1.82244    -0.0715717   0.710396
 ⋮                              ⋱
 0.0       0.0        0.0          -0.0        -0.0        -0.0326891
 0.0      -0.0        0.0          -0.174491    0.580014   -0.312371
 0.0       0.0       -0.0           1.18515    -0.0465437   0.461977
 0.0       0.0        0.0           0.0         0.228902   -0.0470998
 0.0       0.0        0.0       …   0.0         0.0         0.0455176
 0.0      -0.0        0.0          -0.0464512   0.154405   -0.0831562
 0.0       0.0       -0.0           0.315497   -0.0123904   0.122983
 0.0       0.0        0.0           0.0         0.0609358  -0.0125384
 0.0       0.0        0.0           0.0         0.0         0.0121172
julia> logdet(C)-63.57561273339263
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1669.42    -1629.59     111.601   …    909.487    -62.2854   -2351.74
 -1629.59     3046.59   -1592.97       -1700.33     889.048     4407.85
   111.601   -1592.97    1694.12         889.048   -945.499    -2376.0
  4213.77    -7897.86    4257.25        4407.85   -2376.0     -11547.2
  -522.186     509.726    -34.9081      -290.0       19.8604     749.878
   509.726    -952.956    498.271   …    542.168   -283.483    -1405.49
   -34.9081    498.271   -529.909       -283.483    301.483      757.615
 -1318.04     2470.4    -1331.64       -1405.49     757.615     3681.95
   814.363    -794.932     54.4402       431.703    -29.5647   -1116.29
  -794.932    1486.16    -777.067       -807.087    422.001     2092.26
     ⋮                              ⋱
  -494.362     926.58    -499.463       -522.65     281.728     1369.18
   642.54     -627.209     42.9538       356.568    -24.4192    -922.007
  -627.209    1172.59    -613.113       -666.619    348.555     1728.12
    42.9538   -613.113    652.043        348.555   -370.686     -931.521
  1621.83    -3039.78    1638.56    …   1728.12    -931.521    -4527.12
  -931.719     909.487    -62.2854      -536.431     36.737     1387.1
   909.487   -1700.33     889.048       1002.88    -524.376    -2599.83
   -62.2854    889.048   -945.499       -524.376    557.671     1401.41
 -2351.74     4407.85   -2376.0        -2599.83    1401.41      6810.73

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.