Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.86048     0.0638051  -3.54682   …   0.0720297  -4.00401    1.05522
  0.0638051   0.666349   -0.201271      0.752244   -0.227215  -1.16849
 -3.54682    -0.201271    9.94141      -0.227215   11.2229    -0.79548
  0.934732   -1.03507    -0.704649     -1.16849    -0.79548    2.99058
  6.07711     0.0797761  -4.43462       0.110133   -6.12213    1.61343
  0.0797761   0.833144   -0.251651  …   1.15018    -0.347412  -1.78662
 -4.43462    -0.251651   12.4298       -0.347412   17.1598    -1.21629
  1.16871    -1.29416    -0.881029     -1.78662    -1.21629    4.5726
  4.04841     0.0531447  -2.95423       0.0548745  -3.05038    0.803901
  0.0531447   0.555018   -0.167643      0.573083   -0.1731    -0.890193
  ⋮                                 ⋱
  0.896037   -0.99222    -0.675478     -1.23819    -0.842929   3.16897
  3.39527     0.0445708  -2.47761       0.0750177  -4.1701     1.09899
  0.0445708   0.465476   -0.140597      0.783448   -0.236641  -1.21696
 -2.47761    -0.140597    6.94454      -0.236641   11.6884    -0.828478
  0.652954   -0.723043   -0.49223   …  -1.21696    -0.828478   3.11464
  5.48701     0.0720297  -4.00401       0.121619   -6.76062    1.7817
  0.0720297   0.752244   -0.227215      1.27013    -0.383644  -1.97295
 -4.00401    -0.227215   11.2229       -0.383644   18.9494    -1.34314
  1.05522    -1.16849    -0.79548      -1.97295    -1.34314    5.04949
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   6.920748234217583e-5
   0.0009749379421425436
   0.001338804230623754
   0.004544359814835446
   0.0019010219336358747
   0.026780029401782306
   0.03677487059385842
   0.12482649837806603
   0.00440240465540746
   0.062017446524074246
   ⋮
   8.915129355009222
   0.17565070730036395
   2.4744223213410876
   3.3979260924865575
  11.533724225895623
   2.066885502994877
  29.116578594364867
  39.9834665555824
 135.71757133479576
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0310874     0.247008    …  -0.166242   -0.157671   -0.11913
 -0.348624      0.132325       -0.0890576   0.0643312  -0.00207435
 -0.00745103    0.0987944      -0.066491   -0.102979    0.2282
 -0.160106     -0.244768        0.164735   -0.165901   -0.0292344
  0.00194608    0.0154628      -0.262203   -0.248684   -0.187896
 -0.021824      0.00828357  …  -0.140465    0.101465   -0.00327173
 -0.000466437   0.00618456     -0.104872   -0.162422    0.359925
 -0.0100227    -0.0153226       0.259825   -0.261664   -0.0461095
 -0.0149117    -0.118482       -0.167508   -0.158871   -0.120037
  0.167225     -0.0634722      -0.0897355   0.0648208  -0.00209014
  ⋮                         ⋱
  0.0206063     0.0315027       0.192075   -0.193435   -0.0340864
  0.0228737     0.181745       -0.187624   -0.17795    -0.134452
 -0.256513      0.0973628      -0.100512    0.0726052  -0.00234115
 -0.00548237    0.0726917      -0.0750428  -0.116224    0.257551
 -0.117804     -0.180097    …   0.185922   -0.187238   -0.0329944
 -0.0266946    -0.212105       -0.237371   -0.225132   -0.170101
  0.299362     -0.113627       -0.127162    0.091856   -0.00296188
  0.00639817   -0.0848343      -0.09494    -0.14704     0.325839
  0.137482      0.210181        0.235219   -0.236883   -0.0417427
julia> logdet(E)-44.797816758161645
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -1.2054756822062265
 -8.261636174399218
  0.44045746761618665
  1.2416534234821017
  2.9875266488178185
  8.247903888086073
  1.8474210729162766
  1.3550661753948519
  0.915541577710314
  2.0253752631957918
  ⋮
 -0.8140417408764262
  1.9767453217443793
 -0.5750980966946746
  0.9670309001868391
  0.5998335131593627
 -0.15198469536414583
  2.456169955847993
  0.05218334974687659
 -0.18209533415940987

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  42.0706    20.3948   -16.1871    …   21.0355   -16.6957    -1.41173
  20.3948    11.8884    -8.57173       12.2619    -8.84102   -2.7224
 -16.1871    -8.57173    7.18407       -8.84102    7.40977    0.840177
  -1.36873   -2.63948    0.814585      -2.7224     0.840177   4.71625
  25.1593    12.1966    -9.68031       17.0965   -13.5693    -1.14737
  12.1966     7.10956   -5.12611   …    9.96575   -7.18548   -2.21261
  -9.68031   -5.12611    4.29626       -7.18548    6.02224    0.682848
  -0.818535  -1.57847    0.487143      -2.21261    0.682848   3.8331
  34.4697    16.7101   -13.2626        20.836    -16.5373    -1.39834
  16.7101     9.74052   -7.02308       12.1455    -8.75715   -2.69657
   ⋮                               ⋱
  -1.1153    -2.15076    0.663758      -2.53321    0.78179    4.3885
  20.154      9.77014   -7.75445       16.2904   -12.9295    -1.09328
   9.77014    5.69514   -4.10629        9.4959    -6.84671   -2.1083
  -7.75445   -4.10629    3.44154       -6.84671    5.73831    0.650654
  -0.655691  -1.26444    0.390228  …   -2.1083     0.650654   3.65238
  43.3923    21.0355   -16.6957        28.1726   -22.3603    -1.89071
  21.0355    12.2619    -8.84102       16.4221   -11.8407    -3.64607
 -16.6957    -8.84102    7.40977      -11.8407     9.9238     1.12524
  -1.41173   -2.7224     0.840177      -3.64607    1.12524    6.3164
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 6.48618  3.14434  -2.49563   …  3.24313    -2.57404    -0.217652
 0.0      1.41474  -0.512185     1.45919    -0.528276   -1.44057
 0.0      0.0       0.832804     0.0         0.858968   -0.529342
 0.0      0.0       0.0          0.0         0.0         1.56895
 0.0      0.0      -0.0          1.30123    -1.03277    -0.0873277
 0.0      0.0      -0.0       …  0.585465   -0.211958   -0.577994
 0.0      0.0       0.0          0.0         0.344641   -0.212386
 0.0      0.0       0.0          0.0         0.0         0.629505
 0.0      0.0      -0.0          0.447567   -0.355229   -0.030037
 0.0      0.0      -0.0          0.201375   -0.0729046  -0.198805
 ⋮                            ⋱
 0.0      0.0       0.0          0.0         0.0         0.118703
 0.0      0.0      -0.0          0.984253   -0.781191   -0.0660549
 0.0      0.0      -0.0          0.442847   -0.160326   -0.437196
 0.0      0.0       0.0          0.0         0.260687   -0.160649
 0.0      0.0       0.0       …  0.0         0.0         0.476159
 0.0      0.0      -0.0          0.14807    -0.117522   -0.00993724
 0.0      0.0      -0.0          0.0666215  -0.0241193  -0.0657714
 0.0      0.0       0.0          0.0         0.0392176  -0.024168
 0.0      0.0       0.0          0.0         0.0         0.071633
julia> logdet(C)-36.33745093696635
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  127.342   -153.587    111.658   …   140.887   -102.425     64.6065
 -153.587    679.274    433.937      -623.106   -398.056   -246.597
  111.658    433.937    825.362      -398.056   -757.114   -125.556
  -70.4302   268.826    136.874      -246.597   -125.556   -203.741
   29.6508   -35.7619    25.9989       36.5509   -26.5725    16.7611
  -35.7619   158.165    101.04    …  -161.654   -103.269    -63.9754
   25.9989   101.04     192.181      -103.269   -196.421    -32.5734
  -16.3993    62.5944    31.8703      -63.9754   -32.5734   -52.8572
   82.2368   -99.186     72.1083       91.0351   -66.1826    41.7458
  -99.186    438.672    280.235      -402.623   -257.205   -159.34
    ⋮                             ⋱
   19.7659   -75.4446   -38.413        68.4733    34.8635    56.5734
  115.453   -139.248    101.233       130.595    -94.9427    59.8867
 -139.248    615.855    393.424      -577.585   -368.976   -228.582
  101.233    393.424    748.304      -368.976   -701.803   -116.384
  -63.8547   243.727    124.095   …  -228.582   -116.384   -188.857
 -116.812    140.887   -102.425      -134.762     97.9723   -61.7976
  140.887   -623.106   -398.056       596.016    380.75     235.876
 -102.425   -398.056   -757.114       380.75     724.198    120.097
   64.6065  -246.597   -125.556       235.876    120.097    194.883

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.