Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 25.8786    -3.70217    4.15079   …   -2.28399    2.56076      2.79136
 -3.70217   15.6373     0.28448        9.64715    0.175505    -9.60395
  4.15079    0.28448    1.60093        0.175505   0.987665    -0.13221
  4.52457  -15.5672    -0.214303      -9.60395   -0.13221     10.8623
 16.2489    -2.32456    2.60625       -1.67433    1.87722      2.04626
 -2.32456    9.81849    0.178622  …    7.07205    0.128658    -7.04038
  2.60625    0.178622   1.00521        0.128658   0.724029    -0.0969197
  2.84094   -9.77453   -0.134559      -7.04038   -0.0969197    7.96282
 14.4392    -2.06566    2.31597       -1.71549    1.92337      2.09657
 -2.06566    8.72494    0.158728       7.24589    0.13182     -7.21345
  ⋮                               ⋱
  2.53315   -8.71557   -0.119981     -10.0266    -0.138029    11.3403
 18.3683    -2.62775    2.94617       -2.22415    2.49367      2.71822
 -2.62775   11.0991     0.20192        9.39438    0.170907    -9.35231
  2.94617    0.20192    1.13631        0.170907   0.961786    -0.128746
  3.21147  -11.0494    -0.152109  …   -9.35231   -0.128746    10.5777
 15.9654    -2.28399    2.56076       -2.88986    3.24005      3.53182
 -2.28399    9.64715    0.175505      12.2062     0.222061   -12.1516
  2.56076    0.175505   0.987665       0.222061   1.24966     -0.167282
  2.79136   -9.60395   -0.13221      -12.1516    -0.167282    13.7437

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   1.1352374632170206e-5
   1.3527700852764603e-5
   0.00030657757449084925
   0.00047576354101694444
   0.006150082329961833
   0.007328552542992738
   0.16608660167853248
   0.2577421060926553
   0.02771617075641664
   0.03302710480628942
   ⋮
  16.017273279338767
   0.521102415515648
   0.6209553348224136
  14.072678165270293
  21.838737574397086
   4.617830131424221
   5.502692311595036
 124.70722707572916
 193.52737082119486
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0429435     0.00158719   …  -0.00245316   0.318318   -0.199824
  0.00272419   -0.179639         0.27765      0.141371    0.220395
  0.242547      0.0142928       -0.0220908    0.0648559  -0.0209336
  0.0172321    -0.168821         0.260929    -0.141947   -0.23817
 -0.0664175     0.00245479      -0.00196942   0.255549   -0.160421
  0.0042133    -0.277835     …   0.2229       0.113494    0.176936
  0.375129      0.0221055       -0.0177348    0.052067   -0.0168057
  0.0266516    -0.261102         0.209476    -0.113957   -0.191206
  0.00392742   -0.000145158     -0.00177474   0.230287   -0.144563
 -0.000249142   0.016429         0.200866     0.102275    0.159445
  ⋮                          ⋱
  0.0254084    -0.248923         0.211997    -0.115328   -0.193506
  0.0109039    -0.000403008     -0.00223336   0.289796   -0.18192
 -0.000691705   0.0456126        0.252773     0.128704    0.200648
 -0.0615856    -0.00362911      -0.0201115    0.0590448  -0.0190579
 -0.00437544    0.0428656    …   0.237549    -0.129229   -0.21683
  0.00979343   -0.000361966     -0.00215117   0.279132   -0.175225
 -0.000621262   0.0409675        0.243471     0.123968    0.193264
 -0.0553138    -0.00325952      -0.0193714    0.056872   -0.0183566
 -0.00392985    0.0385002        0.228808    -0.124473   -0.208851

julia> logdet(E)
-39.340622928214586

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -2.4974364201622663
  1.3370918617362706
 -6.4475870734749545
  0.15116512390517942
 -5.051808897637635
 10.5356897054321
 -5.788785917998287
  6.56994617823728
  0.8805892806763632
 -2.684496189725421
  ⋮
 -2.167650539132878
  1.82718106607308
  0.7611763996457179
 -6.129393107070806
 -1.3508114834838607
 -1.60770201070371
  3.8784321387180114
  9.448530471175383
  3.0112188310696264

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  31.9239    -14.0148    -1.69398   …  -12.4268    -1.50204   -19.6409
 -14.0148     12.4281     0.779821      11.0199     0.691461   10.3757
  -1.69398     0.779821   9.30292        0.691461   8.24882     4.0721
 -22.1508     11.7016     4.59247       10.3757     4.0721     17.3095
  16.3741     -7.18833   -0.868863     -10.5377    -1.2737    -16.6551
  -7.18833     6.3745     0.399979  …    9.34466    0.586346    8.79843
  -0.868863    0.399979   4.77157        0.586346   6.99486     3.45307
 -11.3614      6.00189    2.35553        8.79843    3.45307    14.6781
  16.1907     -7.1078    -0.859129     -12.3157    -1.48862   -19.4654
  -7.1078      6.30309    0.395498      10.9214     0.685281   10.283
   ⋮                                ⋱
 -18.5106      9.7786     3.83775       10.6907     4.19572    17.835
  24.4418    -10.7301    -1.29696      -13.2303    -1.59916   -20.9109
 -10.7301      9.51528    0.597052      11.7324     0.73617    11.0466
  -1.29696     0.597052   7.12257        0.73617    8.78219     4.3354
 -16.9592      8.95908    3.51612   …   11.0466     4.3354     18.4287
  28.3067    -12.4268    -1.50204      -18.3557    -2.21868   -29.0117
 -12.4268     11.0199     0.691461      16.2776     1.02136    15.3261
  -1.50204     0.691461   8.24882        1.02136   12.1844      6.01494
 -19.6409     10.3757     4.0721        15.3261     6.01494    25.568

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 5.65013  -2.48043  -0.299813   -3.9204    …  -0.265842     -3.47619
 0.0       2.5051    0.0144323   0.789315      0.012797      0.699879
 0.0       0.0       3.03526     1.12204       2.69134       0.994903
 0.0       0.0       0.0         1.50662       0.0           1.3359
 0.0      -0.0      -0.0        -0.0          -0.149014     -1.94853
 0.0       0.0       0.0         0.0       …   0.00717321    0.392308
 0.0       0.0       0.0         0.0           1.5086        0.55768
 0.0       0.0       0.0         0.0           0.0           0.748824
 0.0      -0.0      -0.0        -0.0          -0.071146     -0.930316
 0.0       0.0       0.0         0.0           0.00342481    0.187305
 ⋮                                         ⋱
 0.0       0.0       0.0         0.0          -0.0          -0.321876
 0.0      -0.0      -0.0        -0.0           0.000166283   0.00217435
 0.0       0.0       0.0         0.0          -8.00451e-6   -0.000437772
 0.0       0.0       0.0         0.0          -0.00168343   -0.000622309
 0.0       0.0       0.0         0.0       …  -0.0          -0.000835604
 0.0      -0.0      -0.0        -0.0          -0.074601     -0.975495
 0.0       0.0       0.0         0.0           0.00359113    0.196401
 0.0       0.0       0.0         0.0           0.75525       0.279192
 0.0       0.0       0.0         0.0           0.0           0.374884

julia> logdet(C)
4.826871522488538

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  46.7676      -1.01147    -20.2467    …   3.86256     -11.1481
  -1.01147     50.0327      12.4381       -2.37287       6.49865
 -20.2467      12.4381      41.6932       -7.954         7.67636
  58.4362     -34.0646     -40.2379        7.67636     -20.7656
  24.0058      -0.519186   -10.3926        2.18989      -6.32047
  -0.519186    25.6818       6.38446   …  -1.3453        3.68443
 -10.3926       6.38446     21.4011       -4.50955       4.35214
  29.9953     -17.4853     -20.6541        4.35214     -11.7731
  19.6266      -0.424476    -8.4968        1.63469      -4.71806
  -0.424476    20.9969       5.2198       -1.00423       2.75033
   ⋮                                   ⋱
 -25.9435      15.1234      17.8641       -2.74091       7.41453
   3.41529     -0.0738642   -1.47855      -0.00219683    0.0063405
  -0.0738642    3.65372      0.908312      0.00134957   -0.00369611
  -1.47855      0.908312     3.04472       0.00452384   -0.00436593
   4.2674      -2.48762     -2.93844   …  -0.00436593    0.0118104
  -8.92208      0.192962     3.86256      -1.32354       3.82001
   0.192962    -9.54496     -2.37287       0.813085     -2.22682
   3.86256     -2.37287     -7.954         2.72551      -2.63038
 -11.1481       6.49865      7.67636      -2.63038       7.11551

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source