Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.42973  -2.60257  12.22     -6.65413  …  -1.83515   8.6167   -4.69202
 -2.60257   4.64684  -7.01302   2.57999      3.27662  -4.94508   1.81923
 12.22     -7.01302  21.7145   -9.01501     -4.94508  15.3116   -6.35675
 -6.65413   2.57999  -9.01501  10.7011       1.81923  -6.35675   7.54568
  5.25293  -1.62177   7.61483  -4.14647     -1.13936   5.34973  -2.91306
 -1.62177   2.89565  -4.37011   1.60771  …   2.03431  -3.07018   1.12948
  7.61483  -4.37011  13.5313   -5.61764     -3.07018   9.50626  -3.94662
 -4.14647   1.60771  -5.61764   6.66834      1.12948  -3.94662   4.68478
  5.79633  -1.78954   8.40257  -4.57542     -1.93449   9.08314  -4.94601
 -1.78954   3.19519  -4.82219   1.77402      3.45399  -5.21277   1.91771
  ⋮                                      ⋱
 -4.90577   1.9021   -6.64633   7.88943      1.43375  -5.0098    5.94681
  5.19276  -1.6032    7.52761  -4.09898     -1.56568   7.35144  -4.00305
 -1.6032    2.86248  -4.32006   1.58929      2.79549  -4.21896   1.5521
  7.52761  -4.32006  13.3763   -5.5533      -4.21896  13.0632   -5.42333
 -4.09898   1.58929  -5.5533    6.59196  …   1.5521   -5.42333   6.43769
  5.94405  -1.83515   8.6167   -4.69202     -2.32963  10.9385   -5.95628
 -1.83515   3.27662  -4.94508   1.81923      4.15951  -6.27754   2.30942
  8.6167   -4.94508  15.3116   -6.35675     -6.27754  19.4373   -8.06957
 -4.69202   1.81923  -6.35675   7.54568      2.30942  -8.06957   9.57887
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   1.277109442831036e-5
   0.0001373855170326228
   0.00027594268529595664
   0.0016526787824854459
   0.005058413435140285
   0.054416068180576226
   0.10929620749929934
   0.654597975469022
   0.010126246763953753
   0.10893347121306722
   ⋮
  20.53102425942879
   0.1817235372412435
   1.954897621424008
   3.926467001658747
  23.516436744140588
   1.9413917076783394
  20.884592547603813
  41.94729308715476
 251.23116125903303
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.368082    -0.265685    -0.0444371   …  0.178377    0.0298343   0.155699
 -0.235151    -0.404501    -0.167245       0.271575    0.112286   -0.0820653
 -0.256923    -0.0538521    0.20924        0.0361553  -0.14048     0.259277
  0.0709861   -0.157224     0.433707       0.105557   -0.291184   -0.14078
  0.0912998   -0.065901    -0.0110222      0.121673    0.0203503   0.106204
 -0.0583273   -0.100333    -0.0414838   …  0.185244    0.0765913  -0.0559777
 -0.0637276   -0.0133576    0.0519003      0.024662   -0.0958232   0.176856
  0.0176075   -0.0389981    0.107578       0.0720019  -0.19862    -0.0960279
 -0.0489032    0.0352988    0.00590388     0.155725    0.0260456   0.135927
  0.0312421    0.0537418    0.0222201      0.237088    0.0980266  -0.0716439
  ⋮                                     ⋱
 -0.00997092   0.0220842   -0.0609199      0.0868802  -0.239662   -0.115871
  0.136949    -0.0988512   -0.0165333      0.150925    0.0252429   0.131737
 -0.0874907   -0.150499    -0.0622255      0.229781    0.0950053  -0.0694358
 -0.0955911   -0.0200363    0.0778502      0.0305912  -0.118861    0.219376
  0.0264112   -0.0584969    0.161366    …  0.0893125  -0.246372   -0.119115
  0.0494003   -0.0356576   -0.00596389     0.162047    0.0271031   0.141445
 -0.0315596   -0.054288    -0.022446       0.246714    0.102006   -0.0745526
 -0.0344816   -0.00722747   0.0280821      0.0328455  -0.12762     0.235542
  0.00952702  -0.021101     0.0582078      0.0958941  -0.264527   -0.127893
julia> logdet(E)-40.59226668343824
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
   5.420649072900178
 -10.2802014511694
  -0.04543877210049226
   3.490835683770739
  -5.142580913291581
  -1.9214141298842413
   5.63076248056596
   0.03470600364457515
   2.022063867741087
   1.4206551016866529
   ⋮
  -1.4928945820505923
   4.747417385976083
  -9.830019451729303
  -2.8110816964606267
   0.8191536044661445
  -3.7928546952270006
  -3.3805675882459605
   1.9546051585986923
  -1.1327692262346072

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 15.59      7.30019  -2.6924    -2.18356   …   4.16938  -1.53772   -1.2471
  7.30019   7.00506  -2.47651   -3.67031       4.00082  -1.41442   -2.09623
 -2.6924   -2.47651   4.3829     1.73417      -1.41442   2.50322    0.990443
 -2.18356  -3.67031   1.73417    2.54599      -2.09623   0.990443   1.4541
 13.1285    6.14756  -2.2673    -1.83879       6.43474  -2.37322   -1.92469
  6.14756   5.89903  -2.08549   -3.0908    …   6.1746   -2.18291   -3.23518
 -2.2673   -2.08549   3.69088    1.46036      -2.18291   3.8633     1.52858
 -1.83879  -3.0908    1.46036    2.144        -3.23518   1.52858    2.24416
 10.9309    5.11853  -1.88778   -1.531         8.17512  -3.01509   -2.44526
  5.11853   4.91159  -1.7364    -2.57343       7.84461  -2.77331   -4.11019
  ⋮                                        ⋱
 -1.82288  -3.06405   1.44772    2.12545      -2.77404   1.3107     1.92427
 10.2736    4.81074  -1.77426   -1.43894       4.30282  -1.58693   -1.28701
  4.81074   4.61625  -1.63199   -2.41869       4.12886  -1.45968   -2.16332
 -1.77426  -1.63199   2.88828    1.1428       -1.45968   2.58333    1.02214
 -1.43894  -2.41869   1.1428     1.67778   …  -2.16332   1.02214    1.50064
  8.90396   4.16938  -1.53772   -1.2471        9.73868  -3.59175   -2.91293
  4.16938   4.00082  -1.41442   -2.09623       9.34496  -3.30373   -4.8963
 -1.53772  -1.41442   2.50322    0.990443     -3.30373   5.84692    2.31344
 -1.2471   -2.09623   0.990443   1.4541       -4.8963    2.31344    3.39643
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.94842  1.84889  -0.681895  …  1.05596    -0.389453   -0.315849
 0.0      1.89385  -0.64195      1.08164    -0.366639   -0.798514
 0.0      0.0       1.87238      0.0         1.06938     0.140175
 0.0      0.0       0.0          0.0         0.0         0.271019
 0.0      0.0      -0.0          1.29519    -0.477682   -0.387403
 0.0      0.0      -0.0       …  1.32668    -0.4497     -0.979415
 0.0      0.0       0.0          0.0         1.31165     0.171931
 0.0      0.0       0.0          0.0         0.0         0.332418
 0.0      0.0      -0.0          1.02585    -0.378345   -0.306841
 0.0      0.0      -0.0          1.05079    -0.356182   -0.77574
 ⋮                            ⋱
 0.0      0.0       0.0          0.0         0.0         0.144379
 0.0      0.0      -0.0          0.0723771  -0.0266936  -0.0216487
 0.0      0.0      -0.0          0.074137   -0.0251299  -0.0547312
 0.0      0.0       0.0          0.0         0.0732967   0.00960775
 0.0      0.0       0.0       …  0.0         0.0         0.018576
 0.0      0.0      -0.0          0.143771   -0.0530247  -0.0430034
 0.0      0.0      -0.0          0.147267   -0.0499186  -0.108719
 0.0      0.0       0.0          0.0         0.145598    0.019085
 0.0      0.0       0.0          0.0         0.0         0.0368997
julia> logdet(C)-30.87623841361577
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  148.117    -359.643      58.3877   …  -209.198      33.9631   -250.822
 -359.643    1146.35     -143.774        666.81      -83.6309    838.821
   58.3877   -143.774     169.211        -83.6309     98.4269   -158.476
 -431.201    1442.06     -272.445        838.821    -158.476    1209.0
  -92.9953    225.803     -36.6588       131.123     -21.2876    157.212
  225.803    -719.736      90.2689   …  -417.948      52.4188   -525.762
  -36.6588     90.2689   -106.239         52.4188    -61.6927     99.331
  270.73     -905.4       171.055       -525.762      99.331    -757.788
  -99.2029    240.875     -39.1058       149.111     -24.208     178.779
  240.875    -767.779      96.2944      -475.285      59.6099   -597.89
    ⋮                                ⋱
  191.547    -640.588     121.025       -384.507      72.6439   -554.194
   -6.68976    16.2434     -2.63711       12.4546     -2.02199    14.9326
   16.2434    -51.7753      6.49363      -39.6984      4.97895   -49.9391
   -2.63711     6.49363    -7.64248        4.97895    -5.85982     9.43487
   19.4754    -65.1313     12.3051   …   -49.9391      9.43487   -71.9778
   86.1569   -209.198      33.9631      -127.082      20.6316   -152.367
 -209.198     666.81      -83.6309       405.068     -50.8034    509.56
   33.9631    -83.6309     98.4269       -50.8034     59.7915    -96.2699
 -250.822     838.821    -158.476        509.56      -96.2699    734.435

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.