Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.21285   0.770868  -0.200264  -0.497259  …  0.768777  -0.199721  -0.49591
  0.770868  1.26502    3.24229    1.56871      1.26159    3.23349    1.56446
 -0.200264  3.24229    9.53408    4.25479      3.23349    9.50822    4.24325
 -0.497259  1.56871    4.25479    5.52955      1.56446    4.24325    5.51455
  5.06364   0.748803  -0.194532  -0.483025     0.983691  -0.255554  -0.634543
  0.748803  1.22881    3.14948    1.52381   …  1.61427    4.13742    2.00181
 -0.194532  3.14948    9.26117    4.133        4.13742   12.1663     5.42946
 -0.483025  1.52381    4.133      5.37127      2.00181    5.42946    7.05616
  4.9152    0.726851  -0.188829  -0.468865     1.03988   -0.270151  -0.670788
  0.726851  1.19278    3.05715    1.47914      1.70647    4.37375    2.11615
  ⋮                                         ⋱
 -0.616066  1.94352    5.27136    6.85069      2.22436    6.03308    7.84062
  5.21313   0.770909  -0.200275  -0.497286     0.93664   -0.24333   -0.604192
  0.770909  1.26508    3.24246    1.5688       1.53705    3.93952    1.90606
 -0.200275  3.24246    9.53459    4.25502      3.93952   11.5843     5.16977
 -0.497286  1.5688     4.25502    5.52985   …  1.90606    5.16977    6.71866
  5.19871   0.768777  -0.199721  -0.49591      1.21162   -0.314768  -0.781573
  0.768777  1.26159    3.23349    1.56446      1.98831    5.0961     2.46564
 -0.199721  3.23349    9.50822    4.24325      5.0961    14.9853     6.68752
 -0.49591   1.56446    4.24325    5.51455      2.46564    6.68752    8.69113

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   6.0124543814434004e-5
   0.010952213518025115
   0.021047775933507904
   0.051780525173226
   0.00020762827037775533
   0.03782131231754165
   0.07268434877211766
   0.17881384537653214
   0.00045800209020527895
   0.08342910175105225
   ⋮
   6.1069520080495945
   0.011062353068493283
   2.0151047331328393
   3.8725936848955054
   9.527131770116078
   0.16974146905790272
  30.919898830718545
  59.421330802638266
 146.18493303866134
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0556752    0.0583918   …  -0.0460996  -0.275122   -0.00682576
  0.336127    -0.0310671       0.0245271  -0.0484946   0.0801398
 -0.107888    -0.17601         0.138958   -0.0153871   0.230206
 -0.0173971    0.304416       -0.240333    0.038927    0.142591
  0.0469025   -0.0491911      -0.0534209  -0.318815   -0.00790979
 -0.283164     0.0261719   …   0.0284224  -0.0561963   0.0928673
  0.0908884    0.148277        0.161027   -0.0178308   0.266766
  0.0146559   -0.25645        -0.278501    0.0451092   0.165236
 -0.00615341   0.00645366     -0.0538075  -0.321122   -0.00796703
  0.0371498   -0.00343364      0.0286281  -0.0566029   0.0935392
  ⋮                        ⋱
 -0.0144432    0.252728       -0.327098    0.0529804   0.194069
 -0.0253134    0.0265486      -0.0519494  -0.310033   -0.00769191
  0.152824    -0.0141251       0.0276395  -0.0546483   0.0903091
 -0.0490527   -0.0800253       0.156591   -0.0173396   0.259418
 -0.00790983   0.138406    …  -0.27083     0.0438666   0.160685
 -0.0615732    0.0645776      -0.0577509  -0.344657   -0.00855091
  0.371735    -0.0343583       0.0307261  -0.0607512   0.100394
 -0.119317    -0.194656        0.174078   -0.019276    0.288388
 -0.0192401    0.336665       -0.301075    0.0487654   0.178629

julia> logdet(E)
-64.84622896885982

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  -0.6119099223861002
 -11.76598426598542
   0.0944299504182674
   2.390927560228764
  -3.7804272620036232
   2.289467258995901
   1.446518722033391
  -1.372166487690629
   3.4509535571631322
   4.379825065225917
   ⋮
  -4.945884444651148
   0.8631987649881716
   7.105869056198343
  -2.8709303683362872
  -1.0190983608132318
  -5.470684910072662
  -7.522740550758944
   4.107733967869561
  -9.387673731259296

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.5604      5.54719     2.97327    …   8.6779     4.65132   -0.142691
  5.54719     8.79148    -0.0730644     13.7532    -0.1143    -3.88513
  2.97327    -0.0730644   1.77493       -0.1143     2.77666    0.288876
 -0.0912127  -2.4835      0.184658      -3.88513    0.288876   2.33749
 12.909       6.78085     3.6345        13.1178     7.03109   -0.215697
  6.78085    10.7466     -0.0893135  …  20.7898    -0.172781  -5.8729
  3.6345     -0.0893135   2.16966       -0.172781   4.1973     0.436674
 -0.111498   -3.03581     0.225725      -5.8729     0.436674   3.53344
  8.40497     4.41496     2.3664         9.49164    5.08748   -0.156072
  4.41496     6.99706    -0.0581513     15.0429    -0.125019  -4.24944
  ⋮                                  ⋱
 -0.083942   -2.28553     0.169939      -4.96909    0.369472   2.98966
  9.50665     4.99365     2.67657        9.64077    5.16741   -0.158524
  4.99365     7.9142     -0.0657735     15.2792    -0.126983  -4.31621
  2.67657    -0.0657735   1.59781       -0.126983   3.08475    0.320928
 -0.0821108  -2.23568     0.166232   …  -4.31621    0.320928   2.59685
 16.5205      8.6779      4.65132       20.4439    10.9579    -0.336161
  8.6779     13.7532     -0.1143        32.4006    -0.269276  -9.15283
  4.65132    -0.1143      2.77666       -0.269276   6.54142    0.68055
 -0.142691   -3.88513     0.288876      -9.15283    0.68055    5.50681

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.24968  1.70699   0.914941  …   2.67038    1.43131   -0.0439093
 0.0      2.42439  -0.67434       3.79266   -1.05492   -1.5716
 0.0      0.0       0.695037      0.0        1.0873    -1.05138
 0.0      0.0       0.0           0.0        0.0        0.281899
 0.0      0.0       0.0           0.824294   0.441818  -0.0135539
 0.0      0.0      -0.0       …   1.17072   -0.325634  -0.485123
 0.0      0.0       0.0           0.0        0.335628  -0.324539
 0.0      0.0       0.0           0.0        0.0        0.0870166
 0.0      0.0       0.0           0.582111   0.312009  -0.00957168
 0.0      0.0      -0.0           0.826754  -0.22996   -0.342591
 ⋮                            ⋱
 0.0      0.0       0.0          -0.0       -0.0       -0.0441969
 0.0      0.0       0.0          -0.683269  -0.366229   0.011235
 0.0      0.0      -0.0          -0.970426   0.269922   0.402125
 0.0      0.0       0.0          -0.0       -0.278207   0.269015
 0.0      0.0       0.0       …  -0.0       -0.0       -0.0721293
 0.0      0.0       0.0           0.492909   0.264197  -0.00810492
 0.0      0.0      -0.0           0.700062  -0.194721  -0.290092
 0.0      0.0       0.0           0.0        0.200698  -0.194067
 0.0      0.0       0.0           0.0        0.0        0.0520339

julia> logdet(C)
-53.925603305145536

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1550.3    -1672.38   -2417.57   …   440.647    636.996    628.739
 -1672.38    1830.93    2604.35      -482.424   -686.209   -690.13
 -2417.57    2604.35    3841.53      -686.209  -1012.19    -976.569
 -2386.24    2619.23    3706.35      -690.13    -976.569  -1009.94
   322.92    -348.347   -503.568      102.895    148.744    146.816
  -348.347    381.373    542.473  …  -112.65    -160.236   -161.152
  -503.568    542.473    800.17      -160.236   -236.355   -228.038
  -497.04     545.572    772.012     -161.152   -228.038   -235.83
  -687.688    741.838   1072.39      -191.592   -276.963   -273.373
   741.838   -812.169  -1155.25       209.756    298.361    300.066
     ⋮                            ⋱
 -1505.58    1652.59    2338.5       -382.908   -541.835   -560.35
 -1293.43    1395.28    2017.0       -366.753   -530.174   -523.302
  1395.28   -1527.56   -2172.83       401.524    571.135    574.398
  2017.0    -2172.83   -3205.02       571.135    842.448    812.803
  1990.86   -2185.25   -3092.23   …   574.398    812.803    840.577
  -408.482    440.647    636.996     -161.147   -232.953   -229.933
   440.647   -482.424   -686.209      176.425    250.951    252.385
   636.996   -686.209  -1012.19       250.951    370.163    357.137
   628.739   -690.13    -976.569      252.385    357.137    369.341

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source