Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.49715   4.25467  -2.80142    8.05189   …  2.63639  -1.73589    4.98933
  4.25467   9.90051   3.38309   10.9359       6.13482   2.09632    6.77637
 -2.80142   3.38309   5.23027   -0.597724     2.09632   3.24092   -0.370378
  8.05189  10.9359   -0.597724  24.5804       6.77637  -0.370378  15.2312
  5.96666   2.98761  -1.96714    5.65401      2.03463  -1.33967    3.8505
  2.98761   6.9521    2.37559    7.67912   …  4.73454   1.61783    5.22965
 -1.96714   2.37559   3.67267   -0.419719     1.61783   2.50117   -0.285838
  5.65401   7.67912  -0.419719  17.2603       5.22965  -0.285838  11.7546
  5.53131   2.76962  -1.82361    5.24146      1.84345  -1.21379    3.4887
  2.76962   6.44484   2.20226    7.11881      4.28966   1.46581    4.73825
  ⋮                                        ⋱
  5.85566   7.953    -0.434689  17.8759       6.75675  -0.369305  15.1871
  6.23339   3.12116  -2.05508    5.90676      2.88572  -1.90006    5.46119
  3.12116   7.26288   2.48179    8.0224       6.71501   2.29457    7.41724
 -2.05508   2.48179   3.83685   -0.438482     2.29457   3.54742   -0.405406
  5.90676   8.0224   -0.438482  18.0319    …  7.41724  -0.405406  16.6717
  5.26523   2.63639  -1.73589    4.98933      2.36569  -1.55765    4.47703
  2.63639   6.13482   2.09632    6.77637      5.50491   1.88107    6.08059
 -1.73589   2.09632   3.24092   -0.370378     1.88107   2.90815   -0.332348
  4.98933   6.77637  -0.370378  15.2312       6.08059  -0.332348  13.6673

julia> E = eigen(K)
Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0008046676156485759
   0.00850543013581887
   0.017140677439816992
   0.06145031629748631
   0.005818523395256543
   0.06150246793776225
   0.12394363927984643
   0.4443450886670107
   0.01518726276319659
   0.16053113095967633
   ⋮
  18.9393925170509
   0.3795168723662637
   4.0115374106039505
   8.084302345080754
  28.98268973872171
   3.011451162350852
  31.831388477291497
  64.14861490108889
 229.9764807755255
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.129053     -0.227643     -0.160326     …  -0.166427   -0.115038
 -0.196279     -0.150887      0.156009         0.161946   -0.14982
  0.220559     -0.0414947     0.236841         0.245854   -0.000580946
  0.0513355     0.173643     -0.0180239       -0.0187098  -0.281138
 -0.0644603     0.113704      0.0800808       -0.154908   -0.107076
  0.0980387     0.0753659    -0.0779244    …   0.150737   -0.13945
 -0.110166      0.020726     -0.118299         0.228837   -0.000540737
 -0.0256413    -0.0867323     0.00900269      -0.0174148  -0.261679
  0.0516046    -0.0910275    -0.0641097       -0.167038   -0.11546
 -0.0784862    -0.0603352     0.0623834        0.16254    -0.150369
  ⋮                                        ⋱
 -0.0713313    -0.24128       0.0250445       -0.0170985  -0.256926
 -0.00140551    0.00247923    0.0017461       -0.166334   -0.114974
  0.00213766    0.0016433    -0.00169908       0.161855   -0.149736
 -0.00240208    0.000451914  -0.00257941       0.245716   -0.000580621
 -0.000559089  -0.00189113    0.000196297  …  -0.0186993  -0.28098
  0.110807     -0.195458     -0.137659        -0.123276   -0.0852109
 -0.168529     -0.129554      0.133952         0.119956   -0.110974
  0.189376     -0.035628      0.203355         0.182108   -0.000430317
  0.0440775     0.149093     -0.0154756       -0.0138587  -0.208243

julia> logdet(E)
-10.340100477607878

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
  2.598370862034385
  3.827155105601317
  5.574231435163367
 -3.2654022465326378
 -2.3076138783611193
 -0.7601630424736122
 -7.532633438166987
  2.1675965718968224
 -0.47338511234116526
  3.0120405403163777
  ⋮
  2.8165032481492527
  7.569502018503032
 -9.07851749211959
  7.197436859675919
  2.6605483425940077
 -7.483091109875247
 11.608008878129038
 -1.9834126455663543
 -4.866750105314728

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1.84797    -0.047236    -1.61576   …  -0.0443965   -1.51863     1.12487
 -0.047236    1.18841     -1.53926       1.11697     -1.44673     1.76037
 -1.61576    -1.53926     12.4979       -1.44673     11.7466    -10.0602
  1.19681     1.87296    -10.7037        1.76037    -10.0602      9.11053
  1.00648    -0.0257266   -0.880007     -0.0267286   -0.914282    0.677218
 -0.0257266   0.647254    -0.838339  …   0.672464    -0.870992    1.05982
 -0.880007   -0.838339     6.80683      -0.870992     7.07195    -6.0567
  0.65183     1.02009     -5.82964       1.05982     -6.0567      5.48493
  1.40485    -0.0359094   -1.22832      -0.0545607   -1.86631     1.3824
 -0.0359094   0.903443    -1.17016       1.37269     -1.77794     2.16339
  ⋮                                  ⋱
  0.952089    1.48998     -8.51501       1.71409     -9.7958      8.87105
  2.09167    -0.0534652   -1.82884      -0.0691896   -2.36671     1.75305
 -0.0534652   1.34513     -1.74224       1.74074     -2.25465     2.74344
 -1.82884    -1.74224     14.146        -2.25465     18.3064    -15.6784
  1.35464     2.11995    -12.1152    …   2.74344    -15.6784     14.1983
  1.73688    -0.0443965   -1.51863      -0.0652507   -2.23198     1.65325
 -0.0443965   1.11697     -1.44673       1.64164     -2.12629     2.58726
 -1.51863    -1.44673     11.7466       -2.12629     17.2643    -14.7858
  1.12487     1.76037    -10.0602        2.58726    -14.7858     13.39

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.3594  -0.0347477  -1.18858  …  -0.0326588   -1.11713     0.827471
 0.0      1.08959    -1.4506       1.02409     -1.3634      1.64201
 0.0      0.0         2.99682      0.0          2.81666    -2.23398
 0.0      0.0         0.0          0.0          0.0         0.437356
 0.0     -0.0        -0.0         -0.00339872  -0.116257    0.0861129
 0.0      0.0        -0.0      …   0.106574    -0.141886    0.17088
 0.0      0.0         0.0          0.0          0.293123   -0.232485
 0.0      0.0         0.0          0.0          0.0         0.0455146
 0.0     -0.0        -0.0         -0.0222182   -0.76        0.56294
 0.0      0.0        -0.0          0.696701    -0.927538    1.11708
 ⋮                             ⋱
 0.0      0.0         0.0         -0.0         -0.0        -0.0240474
 0.0     -0.0        -0.0         -0.00196788  -0.0673137   0.0498599
 0.0      0.0        -0.0          0.0617073   -0.0821527   0.0989409
 0.0      0.0         0.0          0.0          0.16972    -0.13461
 0.0      0.0         0.0      …   0.0          0.0         0.0263532
 0.0     -0.0        -0.0         -0.00651461  -0.22284     0.16506
 0.0      0.0        -0.0          0.20428     -0.271963    0.32754
 0.0      0.0         0.0          0.0          0.561853   -0.445622
 0.0      0.0         0.0          0.0          0.0         0.0872415

julia> logdet(C)
-62.3637206639576

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  38.6067       4.57833     13.1724   …   -1.71927    -4.94654    -3.33996
   4.57833    146.62      -112.335       -55.0594     42.1844     57.4329
  13.1724    -112.335      182.465        42.1844    -68.52      -83.203
   8.89412   -152.941      221.565        57.4329    -83.203    -104.905
  -7.38757     -0.876084    -2.52059       0.669427    1.92602     1.30047
  -0.876084   -28.0564      21.4957   …   21.4383    -16.4252    -22.3624
  -2.52059     21.4957     -34.9155      -16.4252     26.6794     32.3965
  -1.70193     29.2659     -42.3975      -22.3624     32.3965     40.8464
 -22.7086      -2.69298     -7.74801      -1.27084    -3.65635    -2.46881
  -2.69298    -86.2423      66.0755      -40.6984     31.1815     42.4528
   ⋮                                  ⋱
  -4.46041     76.6997    -111.115       -22.3077     32.3172     40.7465
  21.9177       2.5992       7.47817      -0.80839    -2.32583    -1.57043
   2.5992      83.2388     -63.7743      -25.8885     19.8348     27.0045
   7.47817    -63.7743     103.588        19.8348    -32.2176    -39.1215
   5.04935    -86.8269     125.786    …   27.0045    -39.1215    -49.3255
 -14.4978      -1.71927     -4.94654       2.1533      6.19527     4.18312
  -1.71927    -55.0594      42.1844       68.9589    -52.8336    -71.9315
  -4.94654     42.1844     -68.52        -52.8336     85.8175    104.207
  -3.33996     57.4329     -83.203       -71.9315    104.207     131.388

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.