Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 14.2515    -0.692247   -1.4722    …  -0.527033  -1.12084   -7.08528
 -0.692247   4.00259    -2.64704       3.04732   -2.01529   -0.102043
 -1.4722    -2.64704     5.27408      -2.01529    4.01534    1.60571
 -9.30637   -0.134031    2.10907      -0.102043   1.60571    9.10549
 10.2482    -0.497793   -1.05865      -0.700825  -1.49044   -9.42169
 -0.497793   2.87825    -1.90348   …   4.05219   -2.67984   -0.135692
 -1.05865   -1.90348     3.79257      -2.67984    5.33943    2.13521
 -6.69219   -0.0963815   1.51663      -0.135692   2.13521   12.1081
 11.6556    -0.566156   -1.20404      -0.737743  -1.56895   -9.91801
 -0.566156   3.27353    -2.16489       4.26565   -2.82101   -0.14284
  ⋮                                ⋱
 -8.89694   -0.128135    2.01628      -0.157144   2.47277   14.0223
  7.12133   -0.345909   -0.735643     -0.71355   -1.5175    -9.59276
 -0.345909   2.00006    -1.3227        4.12577   -2.7285    -0.138156
 -0.735643  -1.3227      2.6354       -2.7285     5.43638    2.17398
 -4.6503    -0.0669741   1.05388   …  -0.138156   2.17398   12.3279
 10.8502    -0.527033   -1.12084      -0.868717  -1.8475   -11.6788
 -0.527033   3.04732    -2.01529       5.02295   -3.32183   -0.168199
 -1.12084   -2.01529     4.01534      -3.32183    6.61856    2.64672
 -7.08528   -0.102043    1.60571      -0.168199   2.64672   15.0087
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   1.0131140198225115e-8
   2.0650579562146173e-8
   4.289349605150521e-8
   1.3302227230823788e-7
   0.018000791883886234
   0.036691505369252614
   0.07621224072397195
   0.2363511108218332
   0.11597149411437196
   0.23638786151329405
   ⋮
  15.479606613600238
   1.4567638818372557
   2.9693615779210547
   6.167686419393216
  19.127367501291708
  18.125148086758994
  36.94498401146114
  76.73874338729875
 237.98391269389262
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.00937555    0.0703705     0.0249336    …  -0.0501303  -0.165801
  0.0855417     0.0138882    -0.0703687        0.14148     0.000603963
  0.0709368    -0.0202136     0.0823099       -0.165488    0.0315707
 -0.00497955    0.0831182     0.0106653       -0.0214432   0.147949
  0.0166628     0.125067      0.0443135       -0.0717078  -0.237166
  0.15203       0.0246829    -0.125064     …   0.202377    0.000863925
  0.126073     -0.0359248     0.146286        -0.236719    0.0451597
 -0.00884997    0.147723      0.0189551       -0.030673    0.21163
 -0.0327266    -0.245638     -0.0870341       -0.0678513  -0.224411
 -0.298595     -0.0484786     0.245632         0.191493    0.000817463
  ⋮                                        ⋱
  0.00656303   -0.109549     -0.0140569       -0.037438    0.258306
  0.00249134    0.0186993     0.00662553      -0.0718139  -0.237517
  0.0227307     0.00369046   -0.0186989        0.202676    0.000865203
  0.0188498    -0.0053713     0.021872        -0.237069    0.0452265
 -0.0013232     0.0220867     0.00283407   …  -0.0307183   0.211943
  0.000297087   0.00222986    0.000790081     -0.0692844  -0.229151
  0.0027106     0.000440081  -0.0022298        0.195537    0.000834728
  0.00224781   -0.000640517   0.00260819      -0.228719    0.0436335
 -0.000157789   0.0026338     0.000337957     -0.0296363   0.204478
julia> logdet(E)-46.46074187507356
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  0.7227221047391346
  3.19114699509971
  0.7327551924719008
  0.6067965658838099
  0.7545264486775628
  1.6099263377243278
 -1.6106604664430515
  3.4799702712723053
 -1.3179133632812587
 -7.362224475764689
  ⋮
 -2.5923715022798586
  0.30512912026488587
  1.497777286245602
  0.7749641655354097
 -0.3058310409167539
 -0.11745634523297596
 -1.1537349098226617
 -1.6535566732432894
 -1.075276420014472

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.643829  -0.138186  -1.32258   1.84086   …  -0.139874  -1.33873   1.86335
 -0.138186   0.289726   1.2016   -0.889297      0.293265   1.21627  -0.900157
 -1.32258    1.2016     6.91177  -4.56363       1.21627    6.99618  -4.61937
  1.84086   -0.889297  -4.56363   8.37475      -0.900157  -4.61937   8.47703
  0.958074  -0.205633  -1.96811   2.73937      -0.207783  -1.98868   2.76801
 -0.205633   0.431138   1.78808  -1.32335   …   0.435646   1.80677  -1.33719
 -1.96811    1.78808   10.2853   -6.79109       1.80677   10.3928   -6.86209
  2.73937   -1.32335   -6.79109  12.4624       -1.33719   -6.86209  12.5927
  0.754435  -0.161926  -1.54979   2.15711      -0.211619  -2.0254    2.81912
 -0.161926   0.339499   1.40802  -1.04207       0.44369    1.84013  -1.36188
  ⋮                                         ⋱
  1.29459   -0.625399  -3.20938   5.88955      -0.44658   -2.29173   4.20556
  0.902212  -0.193643  -1.85335   2.57964      -0.234756  -2.24684   3.12733
 -0.193643   0.406      1.68382  -1.24619       0.492198   2.04132  -1.51077
 -1.85335    1.68382    9.68561  -6.39512       2.04132   11.742    -7.75288
  2.57964   -1.24619   -6.39512  11.7357    …  -1.51077   -7.75288  14.2274
  0.651692  -0.139874  -1.33873   1.86335      -0.254615  -2.43691   3.39189
 -0.139874   0.293265   1.21627  -0.900157      0.533836   2.214    -1.63858
 -1.33873    1.21627    6.99618  -4.61937       2.214     12.7353   -8.40873
  1.86335   -0.900157  -4.61937   8.47703      -1.63858   -8.40873  15.4309
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 0.80239  -0.172218  -1.6483     2.29423   …  -1.66843      2.32225
 0.0       0.509968   1.79958   -0.969059      1.82156     -0.980894
 0.0       0.0        0.977952   0.983517      0.989895     0.995528
 0.0       0.0        0.0        1.09768       0.0          1.11108
 0.0      -0.0       -0.0        0.0           0.00549187  -0.00764402
 0.0       0.0        0.0       -0.0       …  -0.00599594   0.00322876
 0.0       0.0        0.0        0.0          -0.00325839  -0.00327693
 0.0       0.0        0.0        0.0          -0.0         -0.00365729
 0.0      -0.0       -0.0        0.0          -0.749593     1.04334
 0.0       0.0        0.0       -0.0           0.818393    -0.440697
 ⋮                                         ⋱
 0.0       0.0        0.0        0.0          -0.0         -0.215681
 0.0      -0.0       -0.0        0.0          -0.0528179    0.0735161
 0.0       0.0        0.0       -0.0           0.0576657   -0.0310525
 0.0       0.0        0.0        0.0           0.0313374    0.0315158
 0.0       0.0        0.0        0.0       …   0.0          0.0351739
 0.0      -0.0       -0.0        0.0          -0.854368     1.18918
 0.0       0.0        0.0       -0.0           0.932785    -0.502296
 0.0       0.0        0.0        0.0           0.506906     0.50979
 0.0       0.0        0.0        0.0           0.0          0.568963
julia> logdet(C)-64.80849704232809
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   937.448    -1292.16     277.052    …   105.741    -22.6718     15.7364
 -1292.16      2560.91    -508.24        -209.565     41.5905    -22.8325
   277.052     -508.24     116.296         41.5905    -9.51679     4.21396
  -192.301      279.015    -51.495        -22.8325     4.21396    -4.19011
  -320.971      442.422    -94.8592       -41.7603     8.95379    -6.2148
   442.422     -876.824    174.015    …    82.7637   -16.4254      9.01725
   -94.8592     174.015    -39.8184       -16.4254     3.75847    -1.66422
    65.8415     -95.5315    17.6313         9.01725   -1.66422     1.6548
    46.7443     -64.4316    13.8147       -10.1238     2.17063    -1.50663
   -64.4316     127.695    -25.3425        20.0641    -3.98193     2.18601
     ⋮                                ⋱
   -17.1818      24.9296    -4.60101        4.08377   -0.753699    0.749434
    45.843      -63.1893    13.5484        12.021     -2.57742     1.78898
   -63.1893     125.233    -24.8539       -23.8242     4.72817    -2.59568
    13.5484     -24.8539     5.6871         4.72817   -1.08191     0.479059
    -9.40387     13.6444    -2.5182   …    -2.59568    0.479059   -0.476348
   -76.7137     105.741    -22.6718       -77.956     16.7145    -11.6015
   105.741     -209.565     41.5905       154.499    -30.662      16.8329
   -22.6718      41.5905    -9.51679      -30.662      7.01612    -3.10668
    15.7364     -22.8325     4.21396       16.8329    -3.10668     3.0891

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.