Factorization methods

Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factoization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decompositon of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  7.47281   -2.47045  -0.651305  -5.25294  …  -2.51491  -0.663026  -5.34747
 -2.47045    2.52414   1.5858     2.3222       2.56956   1.61434    2.36399
 -0.651305   1.5858    3.33327    2.2174       1.61434   3.39325    2.2573
 -5.25294    2.3222    2.2174     6.34702      2.36399   2.2573     6.46124
  8.44027   -2.79028  -0.735626  -5.93301     -3.16767  -0.83512   -6.73546
 -2.79028    2.85092   1.79111    2.62285  …   3.23651   2.03336    2.97759
 -0.735626   1.79111   3.76481    2.50448      2.03336   4.274      2.84321
 -5.93301    2.62285   2.50448    7.16873      2.97759   2.84321    8.13831
  6.84692   -2.26353  -0.596755  -4.81298     -2.96115  -0.780673  -6.29633
 -2.26353    2.31273   1.45298    2.12771      3.0255    1.90079    2.78346
  ⋮                                        ⋱
 -5.93543    2.62391   2.5055     7.17165      2.86394   2.73469    7.82769
  7.9956    -2.64328  -0.69687   -5.62043     -3.59731  -0.948389  -7.64899
 -2.64328    2.70072   1.69675    2.48466      3.67548   2.30915    3.38144
 -0.69687    1.69675   3.56646    2.37253      2.30915   4.85369    3.22884
 -5.62043    2.48466   2.37253    6.79105  …   3.38144   3.22884    9.24212
  7.60729   -2.51491  -0.663026  -5.34747     -3.41841  -0.901225  -7.26861
 -2.51491    2.56956   1.61434    2.36399      3.4927    2.19431    3.21328
 -0.663026   1.61434   3.39325    2.2573       2.19431   4.61232    3.06827
 -5.34747    2.36399   2.2573     6.46124      3.21328   3.06827    8.78251

julia> E = eigen(K)
Eigen{Number,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
eigenvalues:
40-element Array{Float64,1}:
   0.0028839473505453167
   0.0067972785076672974
   0.01785173390943177
   0.06468748446554191
   0.00903178096487154
   0.02128732711672825
   0.05590703675040066
   0.20258455507200301
   0.03055575836147018
   0.07201795814914026
   ⋮
   9.167297655577643
   0.5502200637471709
   1.2968333187805217
   3.4058812369746616
  12.34154008378293
   7.034413470396719
  16.57966033510694
  43.54326283338697
 157.7832244421677
eigenvectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.261626     0.140376      0.292637    …   0.0668606  0.139381    0.17978
  0.312435    -0.410228      0.141803       -0.19539    0.0675399  -0.0828428
 -0.309693    -0.0265566     0.45154        -0.0126488  0.215066   -0.0597936
  0.233001     0.357163      0.0814292       0.170115   0.0387843  -0.170257
  0.0863072    0.0463084     0.0965372       0.0958387  0.199791    0.257699
  0.103069    -0.135329      0.0467789   …  -0.280074   0.0968125  -0.118748
 -0.102164    -0.00876068    0.148957       -0.0181309  0.308278   -0.0857088
  0.0768641    0.117824      0.0268625       0.243845   0.0555939  -0.244048
 -0.196403    -0.105381     -0.219682        0.0809174  0.168685    0.217578
 -0.234545     0.307958     -0.106451       -0.236468   0.0817395  -0.10026
  ⋮                                      ⋱
 -0.125086    -0.191743     -0.0437152       0.210889   0.0480804  -0.211065
  0.00622077   0.00333777    0.00695811      0.0893872  0.186342    0.240352
  0.00742887  -0.00975412    0.00337168     -0.26122    0.0902954  -0.110754
 -0.00736366  -0.000631443   0.0107364      -0.0169104  0.287526   -0.0799392
  0.00554013   0.00849237    0.00193617  …   0.22743    0.0518515  -0.227619
  0.0497537    0.0266955     0.055651        0.0782458  0.163116    0.210394
  0.0594161   -0.0780134     0.0269667      -0.228661   0.0790408  -0.0969495
 -0.0588945   -0.00505028    0.0858697      -0.0148026  0.251688   -0.0699754
  0.04431      0.0679219     0.0154855       0.199082   0.0453886  -0.199248

julia> logdet(E)
-33.93789420290764

julia> b = randn(40);

julia> (E + 0.1I) \ b  # solve a system
40-element Array{Float64,1}:
 -5.519489476878249
 -1.008502196851448
  3.59097657744834
  5.422255806689549
 -3.5746238303456543
 -0.59524187357826
 -4.53196823987552
  3.404691816918722
  7.090584355035916
 -3.278526704213777
  ⋮
 -5.410584245116199
  1.9705177932760625
 -3.574560736056199
 -3.1784387705410175
  4.781791782294302
 -3.1460045935519485
  8.45765971332609
 -0.7305538680708348
 -2.71295141881196

LinearAlgebra.eigen — Function.

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. Performs Eigenvalue decompositon on the matrices of a AbstractKroneckerProduct instances and returns an Eigen type. The functions size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method.

det(K::Eigen)

Compute the determinant of the eigenvalue decomp of aKronecker product.

source

LinearAlgebra.logdet — Method.

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

source

Base.inv — Method.

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

source

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);

julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite

julia> K = As ⊗ Bs
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.32182   9.69487   -4.61044   12.921    …   5.40101   -2.56847    7.19827
  9.69487  17.671      1.41684   26.6966       9.8445     0.789322  14.8727
 -4.61044   1.41684   16.9495     6.04924      0.789322   9.44259    3.37003
 12.921    26.6966     6.04924   42.5105      14.8727     3.37003   23.6826
  4.82984   5.62674   -2.67582    7.4991       4.21108   -2.00259    5.61236
  5.62674  10.2559     0.82231   15.4943   …   7.67559    0.615421  11.596
 -2.67582   0.82231    9.83722    3.51088      0.615421   7.36222    2.62756
  7.4991   15.4943     3.51088   24.6724      11.596      2.62756   18.4649
  2.09764   2.44373   -1.16213    3.25692      3.13263   -1.48974    4.17506
  2.44373   4.45423    0.357136   6.72928      5.7099     0.457814   8.62629
  ⋮                                        ⋱
  8.37368  17.3013     3.92033   27.5498      10.6788     2.41973   17.0044
  2.27761   2.6534    -1.26184    3.53635      1.8883    -0.897989   2.51666
  2.6534    4.83639    0.387777   7.30663      3.44183    0.275963   5.19979
 -1.26184   0.387777   4.63894    1.65562      0.275963   3.30131    1.17823
  3.53635   7.30663    1.65562   11.6347   …   5.19979    1.17823    8.27991
  4.63608   5.40101   -2.56847    7.19827      5.15828   -2.45304    6.87476
  5.40101   9.8445     0.789322  14.8727       9.40206    0.753848  14.2043
 -2.56847   0.789322   9.44259    3.37003      0.753848   9.01821    3.21857
  7.19827  14.8727     3.37003   23.6826      14.2043     3.21857   22.6182

julia> C = cholesky(K)
40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.88476  3.36072  -1.59821  4.47905   …   1.87226    -0.890361    2.49528
 0.0      2.52517   2.68812  4.61108       1.40678     1.49755     2.56883
 0.0      0.0       2.67755  0.303465      0.0         1.49166     0.16906
 0.0      0.0       0.0      1.04617       0.0         0.0         0.58282
 0.0      0.0      -0.0      0.0           0.77577    -0.368921    1.03392
 0.0      0.0       0.0      0.0       …   0.582897    0.62051     1.0644
 0.0      0.0       0.0      0.0           0.0         0.61807     0.0700501
 0.0      0.0       0.0      0.0           0.0         0.0         0.241491
 0.0      0.0      -0.0      0.0           0.975664   -0.463981    1.30033
 0.0      0.0       0.0      0.0           0.733093    0.780398    1.33866
 ⋮                                     ⋱
 0.0      0.0       0.0      0.0           0.0         0.0         0.171166
 0.0      0.0      -0.0      0.0          -0.0342099   0.0162687  -0.0455936
 0.0      0.0       0.0      0.0          -0.0257046  -0.0273632  -0.0469377
 0.0      0.0       0.0      0.0          -0.0        -0.0272556  -0.00308907
 0.0      0.0       0.0      0.0       …  -0.0        -0.0        -0.0106493
 0.0      0.0      -0.0      0.0           0.559332   -0.265993    0.745457
 0.0      0.0       0.0      0.0           0.42027     0.447389    0.767432
 0.0      0.0       0.0      0.0           0.0         0.445631    0.0505063
 0.0      0.0       0.0      0.0           0.0         0.0         0.174116

julia> logdet(C)
-45.21174080901596

julia> inv(C)
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   30.4018     -34.4095      7.09175   …     7.7973    -1.60701    -2.5741
  -34.4095      99.5159      0.94138       -22.5506    -0.213319   11.8222
    7.09175      0.94138     5.06315        -0.213319  -1.14733     0.785676
   11.3595     -52.1714     -3.46719        11.8222     0.785676   -6.93223
  108.877     -123.23       25.3975         13.866     -2.85776    -4.57755
 -123.23       356.394       3.37134   …   -40.102     -0.379348   21.0235
   25.3975       3.37134    18.1326         -0.379348  -2.0403      1.39718
   40.6816    -186.84      -12.417          21.0235     1.39718   -12.3277
  -12.9453      14.6518     -3.01971        39.5036    -8.14162   -13.0412
   14.6518     -42.3745     -0.400845     -114.248     -1.08074    59.8949
    ⋮                                  ⋱
    0.719307    -3.30359    -0.219549       38.3832     2.55086   -22.5069
   85.8648     -97.184      20.0295         -8.52637    1.75727     2.81479
  -97.184      281.066       2.65877        24.6591     0.233265  -12.9276
   20.0295       2.65877    14.3             0.233265   1.2546     -0.859139
   32.0831    -147.349      -9.79251   …   -12.9276    -0.859139    7.58041
   -6.88913      7.7973     -1.60701       -37.1017     7.6466     12.2483
    7.7973     -22.5506     -0.213319      107.302      1.01503   -56.2532
   -1.60701     -0.213319   -1.14733         1.01503    5.45929    -3.73846
   -2.5741      11.8222      0.785676      -56.2532    -3.73846    32.9855

LinearAlgebra.cholesky — Function.

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.

source