Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 13.1062   -7.27311    9.59199  2.3944    …   -6.97205    9.19494  2.29529
 -7.27311   6.71032   -2.99635  0.226558       6.43256   -2.87232  0.21718
  9.59199  -2.99635   10.1083   3.23187       -2.87232    9.68984  3.09809
  2.3944    0.226558   3.23187  2.20117        0.21718    3.09809  2.11006
 12.4968   -6.93497    9.14605  2.28308       -8.1683    10.7726   2.68911
 -6.93497   6.39835   -2.85705  0.216025  …    7.53624   -3.36515  0.254444
  9.14605  -2.85705    9.63832  3.08161       -3.36515   11.3524   3.62965
  2.28308   0.216025   3.08161  2.09884        0.254444   3.62965  2.4721
 13.0198   -7.22519    9.52879  2.37862      -10.0297    13.2275   3.30192
 -7.22519   6.66611   -2.97661  0.225066       9.25365   -4.13202  0.312428
  ⋮                                       ⋱
  2.09335   0.198073   2.82552  1.92442        0.260187   3.71158  2.5279
 12.6593   -7.02513    9.26495  2.31276       -7.48033    9.86529  2.46262
 -7.02513   6.48153   -2.89419  0.218834       6.90151   -3.08172  0.233013
  9.26495  -2.89419    9.76361  3.12167       -3.08172   10.3963   3.32395
  2.31276   0.218834   3.12167  2.12612   …    0.233013   3.32395  2.26389
 12.5637   -6.97205    9.19494  2.29529      -10.4606    13.7957   3.44375
 -6.97205   6.43256   -2.87232  0.21718        9.65113   -4.30951  0.325848
  9.19494  -2.87232    9.68984  3.09809       -4.30951   14.5382   4.64823
  2.29529   0.21718    3.09809  2.11006        0.325848   4.64823  3.16584
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.00024414959856611197
   0.0007102657518614457
   0.0055177714867066865
   0.0228580818578436
   0.0052480529968699716
   0.015267329250271982
   0.1186057947951671
   0.49133984123374175
   0.019169007349770686
   0.0557653565587809
   ⋮
  20.825810438297403
   0.279436467001701
   0.812920248479235
   6.315253348862361
  26.161753590030536
   2.841418838330099
   8.26608972290944
  64.2159558729875
 266.0229006331173
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0571762    0.0096012    0.0125163   …  -0.0301625  -0.0393203   0.191245
  0.0490566   -0.0107592   -0.0606963       0.0338004   0.19068    -0.100211
 -0.0347033   -0.0390021   -0.0465241       0.122526    0.146157    0.15177
 -0.0185446    0.0741268   -0.0349098      -0.232872    0.10967     0.040538
  0.0986311    0.0165624    0.0215911      -0.0393216  -0.0512603   0.249318
  0.0846245   -0.01856     -0.104703    …   0.0440641   0.248581   -0.130641
 -0.0598644   -0.06728     -0.0802559       0.159732    0.190539    0.197856
 -0.0319901    0.127872    -0.0602207      -0.303585    0.142973    0.0528477
 -0.0980415   -0.0164634   -0.021462       -0.0437452  -0.057027    0.277366
 -0.0841186    0.018449     0.104078        0.0490213   0.276546   -0.145338
  ⋮                                     ⋱
  0.105002    -0.419719     0.197665       -0.29016     0.13665     0.0505107
 -0.0177451   -0.00297981  -0.00388453     -0.0339923  -0.0443128   0.215528
 -0.0152251    0.0033392    0.0188376       0.038092    0.21489    -0.112935
  0.0107704    0.0121046    0.0144391       0.138084    0.164715    0.17104
  0.00575545  -0.0230058    0.0108345   …  -0.26244     0.123595    0.0456851
 -0.274654    -0.0461208   -0.0601238      -0.0381949  -0.0497914   0.242174
 -0.235651     0.0516834    0.291564        0.0428015   0.241458   -0.126897
  0.166702     0.187352     0.223486        0.155155    0.185079    0.192186
  0.0890816   -0.356079     0.167694       -0.294886    0.138876    0.0513334
julia> logdet(E)-29.47030587130535
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 10.107866531807066
  8.106922568584448
 -6.5119254366332155
 -5.402529200019791
  2.2497737400819164
  2.230781269683519
 -3.865245043461855
  4.647832595848161
 -8.299681759459311
 -7.494589580493447
  ⋮
  1.2823339757303134
 -7.907207021013063
 -5.066604330833214
  9.485743913954705
 -3.718963356388381
  0.5606831099142411
 -2.08807237489793
 -2.6447203619737283
 -2.045240599835965

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 20.2152    6.92868   9.67829  -8.90947  …   4.8942     6.83644  -6.29337
  6.92868   4.67121   5.25537  -2.07108      3.29959    3.71223  -1.46295
  9.67829   5.25537   9.91084  -7.97982      3.71223    7.00071  -5.6367
 -8.90947  -2.07108  -7.97982  10.0348      -1.46295   -5.6367    7.08825
 14.5931    5.00173   6.98664  -6.43164      3.74974    5.2378   -4.82173
  5.00173   3.37209   3.79379  -1.49509  …   2.52802    2.84416  -1.12085
  6.98664   3.79379   7.15452  -5.76054      2.84416    5.36366  -4.31861
 -6.43164  -1.49509  -5.76054   7.24399     -1.12085   -4.31861   5.43073
 14.8454    5.0882    7.10742  -6.54283      5.38576    7.52308  -6.92547
  5.0882    3.43038   3.85938  -1.52094      3.631      4.08508  -1.60988
  ⋮                                      ⋱
 -4.58096  -1.06488  -4.10297   5.15956     -1.58684   -6.11405   7.68852
 13.7696    4.71946   6.59235  -6.06867      3.12803    4.36937  -4.02228
  4.71946   3.18178   3.57969  -1.41071      2.10887    2.3726   -0.935013
  6.59235   3.57969   6.75075  -5.43544      2.3726     4.47436  -3.60258
 -6.06867  -1.41071  -5.43544   6.83517  …  -0.935013  -3.60258   4.53031
 14.2794    4.8942    6.83644  -6.29337      6.55974    9.16294  -8.43506
  4.8942    3.29959   3.71223  -1.46295      4.42247    4.97553  -1.9608
  6.83644   3.71223   7.00071  -5.6367       4.97553    9.38311  -7.55491
 -6.29337  -1.46295  -5.6367    7.08825     -1.9608    -7.55491   9.50044
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 4.49613  1.54103  2.15258  -1.98159   …   1.08854     1.52052    -1.39973
 0.0      1.5154   1.27899   0.648411      1.07043     0.903439    0.458017
 0.0      0.0      1.90825  -2.38104       0.0         1.34793    -1.68189
 0.0      0.0      0.0       0.135346      0.0         0.0         0.095604
 0.0      0.0      0.0      -0.0           0.0693935   0.096932   -0.089232
 0.0      0.0      0.0       0.0       …   0.0682393   0.0575937   0.0291983
 0.0      0.0      0.0      -0.0           0.0         0.0859296  -0.107219
 0.0      0.0      0.0       0.0           0.0         0.0         0.0060947
 0.0      0.0      0.0      -0.0           0.640299    0.894399   -0.82335
 0.0      0.0      0.0       0.0           0.629649    0.531421    0.269415
 ⋮                                     ⋱
 0.0      0.0      0.0       0.0           0.0         0.0         0.0380857
 0.0      0.0      0.0      -0.0          -0.23455    -0.32763     0.301604
 0.0      0.0      0.0       0.0          -0.230649   -0.194667   -0.0986904
 0.0      0.0      0.0      -0.0          -0.0        -0.290442    0.362402
 0.0      0.0      0.0       0.0       …  -0.0        -0.0        -0.0206001
 0.0      0.0      0.0      -0.0           0.287286    0.401293   -0.369416
 0.0      0.0      0.0       0.0           0.282507    0.238435    0.12088
 0.0      0.0      0.0      -0.0           0.0         0.355744   -0.443883
 0.0      0.0      0.0       0.0           0.0         0.0         0.0252317
julia> logdet(C)-59.578607718301825
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  181.408    -755.944     633.16    …    546.98    -458.137   -367.969
 -755.944    3194.96    -2683.86       -2311.79    1941.96    1552.79
  633.16    -2683.86     2263.31        1941.96   -1637.67   -1308.26
  508.545   -2146.01     1808.06        1552.79   -1308.26   -1048.49
  -10.3858     43.2788    -36.2493      -594.04     497.553    399.628
   43.2788   -182.916     153.654   …   2510.68   -2109.04   -1686.39
  -36.2493    153.654    -129.577      -2109.04    1778.57    1420.82
  -29.1149    122.862    -103.514      -1686.39    1420.82    1138.69
  -33.2347    138.492    -115.998        508.963   -426.295   -342.394
  138.492    -585.331     491.694      -2151.11    1806.99    1444.87
    ⋮                               ⋱
  415.052   -1751.48     1475.66         954.819   -804.453   -644.718
  -24.1925    100.812     -84.4381      -577.002    483.283    388.166
  100.812    -426.08      357.919       2438.67   -2048.55   -1638.02
  -84.4381    357.919    -301.834      -2048.55    1727.55    1380.07
  -67.8194    286.192    -241.122   …  -1638.02    1380.07    1106.04
 -131.261     546.98     -458.137       -819.434    686.338    551.256
  546.98    -2311.79     1941.96        3463.3    -2909.27   -2326.25
 -458.137    1941.96    -1637.67       -2909.27    2453.4     1959.91
 -367.969    1552.79    -1308.26       -2326.25    1959.91    1570.74

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.