Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 15.273     8.345    10.1051    4.94918  …   6.66815   8.07457   3.95469
  8.345    11.4985    9.93271  13.2454       9.18802   7.93683  10.5838
 10.1051    9.93271  15.9963   16.3361       7.93683  12.782    13.0535
  4.94918  13.2454   16.3361   28.8968      10.5838   13.0535   23.0903
  8.75498   4.78361   5.79255   2.83703      5.57805   6.75455   3.30819
  4.78361   6.59132   5.69374   7.59265  …   7.68597   6.63932   8.8536
  5.79255   5.69374   9.16959   9.36434      6.63932  10.6924   10.9195
  2.83703   7.59265   9.36434  16.5645       8.8536   10.9195   19.3155
  8.8149    4.81635   5.83219   2.85644      4.67529   5.66138   2.77278
  4.81635   6.63643   5.73271   7.64462      6.44206   5.5648    7.42072
  ⋮                                      ⋱
  3.98942  10.6768   13.1681   23.293        9.86776  12.1703   21.528
 12.2807    6.71      8.12524   3.97951      6.67504   8.08291   3.95878
  6.71      9.24568   7.98664  10.6503       9.19751   7.94503  10.5948
  8.12524   7.98664  12.8622   13.1354       7.94503  12.7952   13.067
  3.97951  10.6503   13.1354   23.2352   …  10.5948   13.067    23.1141
 12.2041    6.66815   8.07457   3.95469      7.68699   9.30829   4.55893
  6.66815   9.18802   7.93683  10.5838      10.5919    9.1495   12.2009
  8.07457   7.93683  12.782    13.0535       9.1495   14.735    15.0479
  3.95469  10.5838   13.0535   23.0903      12.2009   15.0479   26.6182
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0020974691972982257
   0.00595459214813944
   0.024859455326798202
   0.08641832561470421
   0.019885152901078
   0.056452783898726114
   0.23568120611623647
   0.8192928985641759
   0.03444058071667492
   0.09777479057954247
   ⋮
  19.550290737606257
   0.5786117335275964
   1.6426448072385667
   6.857775341692125
  23.839519196222426
   7.795962991442084
  22.132282121626922
  92.39868407353346
 321.20331928709345
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.139337   -0.00900641   0.214558    …  -0.0120886    0.287986   0.119192
  0.138436   -0.200914     0.0348829      -0.269672     0.0468207  0.151078
  0.13959     0.181294     0.0405893       0.243337     0.0544799  0.187032
 -0.123905   -0.0101052   -0.156581       -0.0135635   -0.210167   0.245466
  0.024158    0.00156151  -0.0371996      -0.00999945   0.238215   0.0985932
 -0.0240017   0.034834    -0.00604791  …  -0.223067     0.038729   0.124968
 -0.0242017  -0.0314323   -0.00703727      0.201283     0.0450646  0.154709
  0.0214823   0.00175202   0.0271476      -0.0112194   -0.173845   0.203044
 -0.213605   -0.0138069    0.32892        -0.00891472   0.212374   0.087898
  0.212224   -0.308003     0.0534758      -0.198869     0.0345278  0.111412
  ⋮                                    ⋱
  0.184308    0.0150315    0.232912       -0.013303    -0.206131   0.240752
  0.121931    0.00788128  -0.187755       -0.0123382    0.293931   0.121653
 -0.121142    0.175815    -0.0305251      -0.275239     0.0477873  0.154197
 -0.122151   -0.158646    -0.0355186       0.248361     0.0556047  0.190893
  0.108426    0.00884284   0.13702     …  -0.0138435   -0.214505   0.250534
  0.168393    0.0108845   -0.259301       -0.0118353    0.281951   0.116695
 -0.167304    0.242811    -0.042157       -0.264021     0.0458396  0.147912
 -0.168698   -0.219099    -0.0490534       0.238238     0.0533384  0.183113
  0.149743    0.0122125    0.189232       -0.0132793   -0.205763   0.240323
julia> logdet(E)-8.031784933276022
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
 -4.114324173818922
  1.3321241020535401
  3.109188860813215
 -3.785820113777369
 -1.4035511821379014
  4.698838274664763
  2.666672421084676
 -2.2171268085891054
 -3.962476382172178
  0.10938821634825781
  ⋮
  0.12244831296997168
  2.7132004556880216
  0.9220755586208969
  0.9874438199692626
  1.7253936109325758
  3.951753689467561
  0.28502082666059153
 -1.689495202015611
  2.9968891953126695

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  48.4249   -17.4986     4.5363    …  -12.6311     3.27446   -7.04423
 -17.4986    11.8936     1.15884        8.58523    0.836493   1.27503
   4.5363     1.15884   22.7037         0.836493  16.3883    -6.12758
  -9.75879    1.76638   -8.4889         1.27503   -6.12758    3.38257
  29.6896   -10.7285     2.78122       -9.52034    2.46803   -5.30939
 -10.7285     7.29204    0.710492  …    6.47087    0.630483   0.96102
   2.78122    0.710492  13.9197         0.630483  12.3522    -4.61849
  -5.98316    1.08297   -5.20458        0.96102   -4.61849    2.54951
  33.2246   -12.0059     3.11238       -8.71165    2.25838   -4.85839
 -12.0059     8.16028    0.795089       5.92121    0.576927   0.879387
   ⋮                               ⋱
  -6.52456    1.18097   -5.67553        0.923347  -4.43744    2.44957
  29.6036   -10.6974     2.77318       -9.7923     2.53853   -5.46106
 -10.6974     7.27094    0.708436       6.65572    0.648493   0.988473
   2.77318    0.708436  13.8795         0.648493  12.7051    -4.75042
  -5.96585    1.07984   -5.18953   …    0.988473  -4.75042    2.62234
  34.9548   -12.6311     3.27446      -11.0269     2.85859   -6.1496
 -12.6311     8.58523    0.836493       7.49489    0.730256   1.1131
   3.27446    0.836493  16.3883         0.730256  14.3069    -5.34937
  -7.04423    1.27503   -6.12758        1.1131    -5.34937    2.95297
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 6.9588  -2.5146   0.651879  …  -1.81513    0.470549    -1.01228
 0.0      2.36017  1.18553       1.70365    0.855759    -0.538284
 0.0      0.0      4.56872       0.0        3.29786     -1.05709
 0.0      0.0      0.0           0.0        0.0          0.0988119
 0.0     -0.0      0.0          -0.48729    0.126324    -0.271756
 0.0      0.0      0.0       …   0.457363   0.229737    -0.144508
 0.0      0.0      0.0           0.0        0.885346    -0.283787
 0.0      0.0      0.0           0.0        0.0          0.0265271
 0.0     -0.0      0.0          -0.0188943  0.00489811  -0.0105372
 0.0      0.0      0.0           0.0177339  0.00890791  -0.00560319
 ⋮                           ⋱
 0.0      0.0      0.0           0.0        0.0          0.0204172
 0.0     -0.0      0.0          -0.245928   0.0637537   -0.137151
 0.0      0.0      0.0           0.230824   0.115945    -0.0729309
 0.0      0.0      0.0           0.0        0.446821    -0.143223
 0.0      0.0      0.0       …   0.0        0.0          0.0133878
 0.0     -0.0      0.0          -0.0736878  0.0191026   -0.0410949
 0.0      0.0      0.0           0.0691623  0.0347408   -0.0218524
 0.0      0.0      0.0           0.0        0.133882    -0.0429141
 0.0      0.0      0.0           0.0        0.0          0.00401142
julia> logdet(C)-54.81764091890578
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   233.932      162.585    …    389.982      774.568     2424.68
   162.585      122.641         294.172      524.569     1651.52
   322.919      218.694         524.569     1104.66      3416.43
  1010.86       688.524        1651.52      3416.43     10658.4
    -9.45747     -6.57302        46.3631      92.0846     288.258
    -6.57302     -4.95817  …     34.9727      62.3635     196.342
   -13.0551      -8.84143        62.3635     131.328      406.163
   -40.8671     -27.8359        196.342      406.163     1267.13
    14.6773      10.2008         74.2066     147.386      461.372
    10.2008       7.6947         55.9756      99.8159     314.255
     ⋮                     ⋱
 -2047.59     -1394.67        -5924.81    -12256.4     -38237.0
  -225.995     -157.069        -999.31     -1984.79     -6213.12
  -157.069     -118.48         -753.802    -1344.18     -4231.94
  -311.964     -211.275       -1344.18     -2830.64     -8754.44
  -976.56      -665.165    …  -4231.94     -8754.44    -27311.7
   561.119      389.982        2273.82      4516.17     14137.2
   389.982      294.172        1715.19      3058.53      9629.31
   774.568      524.569        3058.53      6440.81     19919.7
  2424.68      1651.52         9629.31     19919.7      62144.7

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.