Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.932349   -2.9814     2.77156  …   -2.15847    2.00655   1.42309
 -2.9814     27.0729   -20.7427       19.6002   -15.0173   -8.10146
  2.77156   -20.7427    18.1337      -15.0173    13.1284    6.93219
  1.96566   -11.1902     9.57513      -8.10146    6.93219  18.5227
  0.693439   -2.21743    2.06136      -1.90297    1.76903   1.25464
 -2.21743    20.1356   -15.4275   …   17.2801   -13.2396   -7.14248
  2.06136   -15.4275    13.487       -13.2396    11.5744    6.11161
  1.46197    -8.32276    7.12154      -7.14248    6.11161  16.3302
  0.520164   -1.66335    1.54627      -1.31236    1.21999   0.865245
 -1.66335    15.1042   -11.5725       11.917     -9.13053  -4.92571
  ⋮                               ⋱
  1.243      -7.07624    6.05493      -6.74238    5.76926  15.4154
  0.899813   -2.87736    2.67484      -2.36144    2.19524   1.55692
 -2.87736    26.1281   -20.0188       21.4433   -16.4294   -8.86328
  2.67484   -20.0188    17.5009      -16.4294    14.363     7.58406
  1.89706   -10.7997     9.24099  …   -8.86328    7.58406  20.2645
  0.675001   -2.15847    2.00655      -2.41492    2.24495   1.59217
 -2.15847    19.6002   -15.0173       21.9289   -16.8014   -9.06398
  2.00655   -15.0173    13.1284      -16.8014    14.6882    7.75579
  1.42309    -8.10146    6.93219      -9.06398    7.75579  20.7234
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.002429744201688581
   0.007764489040636266
   0.0928775492970967
   0.2798068125175123
   0.005191752147973734
   0.01659075989424196
   0.19845595916089817
   0.5978767719235281
   0.007987831610313186
   0.02552590966309215
   ⋮
  20.03699416804177
   0.22386330125694195
   0.7153774245051014
   8.557227869441096
  25.779864695559784
   3.4685856285464207
  11.084210050029363
 132.58750961629482
 399.4387096358297
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.191188     -0.0560803   -0.00296151   …   0.00488134   0.0287432
  0.0226015    -0.123144     0.0782907       -0.129043    -0.222399
  0.0541079    -0.147196    -0.0573957        0.094603     0.181473
  0.0044026     0.00587032   0.174866        -0.288225     0.159623
 -0.0298448    -0.00875426  -0.000462299      0.00454135   0.0267412
  0.00352815   -0.0192231    0.0122214    …  -0.120055    -0.206908
  0.00844636   -0.0229777   -0.00895959       0.0880138    0.168833
  0.000687256   0.00091637   0.027297        -0.26815      0.148505
 -0.00918825   -0.00269515  -0.000142327      0.0037137    0.0218677
  0.0010862    -0.00591815   0.00376256      -0.0981755   -0.1692
  ⋮                                       ⋱
  0.00725299    0.00967096   0.28808         -0.23332      0.129216
 -0.146418     -0.0429482   -0.00226803       0.00558824   0.0329056
  0.017309     -0.0943079    0.0599577       -0.147731    -0.254605
  0.0414377    -0.112728    -0.0439556        0.108303     0.207753
  0.00337166    0.00449569   0.133918     …  -0.329964     0.182739
 -0.16363      -0.0479971   -0.00253465       0.00420721   0.0247736
  0.0193438    -0.105394     0.0670061       -0.111222    -0.191685
  0.0463089    -0.12598     -0.0491228        0.081538     0.156411
  0.00376802    0.00502419   0.149661        -0.24842      0.137578
julia> logdet(E)-21.430972052708256
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  11.838076198717845
   0.3418114421645679
  -1.26589615767087
   0.08586729306273073
   5.65079674418513
  -3.0753257244593843
  -4.458285210849492
   0.19883351035821278
   5.687165128376868
  -1.4365487105007833
   ⋮
   1.3669294168819153
 -14.358218544116058
   0.8972551902396081
  -2.625133089612964
   1.5905719540610543
  -4.656800554584853
  -0.3268559296749387
  -0.9829249111867779
   0.8589820843132252

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 26.3248    2.9039    -1.67594  13.478     …   2.42458   -1.3993    11.2533
  2.9039   27.6471     8.14452   0.846551     23.0836     6.80017    0.706818
 -1.67594   8.14452   10.2908    4.11679       6.80017    8.59215    3.43727
 13.478     0.846551   4.11679  24.6462        0.706818   3.43727   20.5781
 15.9032    1.75429   -1.01246   8.14226       1.31835   -0.760865   6.11891
  1.75429  16.7021     4.92024   0.511415  …  12.5516     3.69756    0.384329
 -1.01246   4.92024    6.21681   2.48702       3.69756    4.67194    1.869
  8.14226   0.511415   2.48702  14.8892        0.384329   1.869     11.1892
 16.683     1.84031   -1.06211   8.5415        2.00953   -1.15977    9.32691
  1.84031  17.521      5.1615    0.536492     19.1321     5.63611    0.585824
  ⋮                                        ⋱
  6.48967   0.407616   1.98224  11.8672        0.44263    2.15252   12.8866
 22.2355    2.45282   -1.4156   11.3843        2.00348   -1.15628    9.29884
  2.45282  23.3525     6.87938   0.715051     19.0745     5.61915    0.584061
 -1.4156    6.87938    8.69222   3.4773        5.61915    7.09989    2.8403
 11.3843    0.715051   3.4773   20.8177    …   0.584061   2.8403    17.0041
 21.9795    2.42458   -1.3993   11.2533        2.46152   -1.42062   11.4247
  2.42458  23.0836     6.80017   0.706818     23.4353     6.90379    0.717587
 -1.3993    6.80017    8.59215   3.43727       6.90379    8.72306    3.48964
 11.2533    0.706818   3.43727  20.5781        0.717587   3.48964   20.8916
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 5.13077  0.565978  -0.326645  …   0.472556   -0.272728    2.19329
 0.0      5.2275     1.59338       4.36464     1.33037    -0.102255
 0.0      0.0        2.765         0.0         2.3086      1.56117
 0.0      0.0        0.0           0.0         0.0         3.15011
 0.0      0.0       -0.0          -0.0491921   0.0283904  -0.228317
 0.0      0.0        0.0       …  -0.45435    -0.138489    0.0106445
 0.0      0.0        0.0          -0.0        -0.24032    -0.162514
 0.0      0.0        0.0          -0.0        -0.0        -0.327919
 0.0      0.0       -0.0           0.172669   -0.0996532   0.801415
 0.0      0.0        0.0           1.59481     0.486111   -0.0373632
 ⋮                             ⋱
 0.0      0.0        0.0           0.0         0.0         0.417752
 0.0      0.0       -0.0           0.0343947  -0.0198503   0.159637
 0.0      0.0        0.0           0.317677    0.0968304  -0.00744253
 0.0      0.0        0.0           0.0         0.16803     0.113629
 0.0      0.0        0.0       …   0.0         0.0         0.229278
 0.0      0.0       -0.0           0.086137   -0.0497125   0.39979
 0.0      0.0        0.0           0.795581    0.242499   -0.0186388
 0.0      0.0        0.0           0.0         0.420809    0.284568
 0.0      0.0        0.0           0.0         0.0         0.574198
julia> logdet(C)29.458995963564497
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  2.79496    -0.762354    1.77835    …   0.320494  -0.747619   0.75643
 -0.762354    2.31462    -2.24075       -0.973069   0.942013  -0.299191
  1.77835    -2.24075     7.19021        0.942013  -3.02277    0.881396
 -1.7993      0.71168    -2.09656       -0.299191   0.881396  -1.30338
 -1.68227     0.458857   -1.07038        0.1279    -0.298353   0.301869
  0.458857   -1.39316     1.3487     …  -0.388324   0.37593   -0.119399
 -1.07038     1.3487     -4.32776        0.37593   -1.2063     0.35174
  1.08299    -0.428357    1.26191       -0.119399   0.35174   -0.52014
 -0.159354    0.0434653  -0.101392       0.491978  -1.14764    1.16117
  0.0434653  -0.131967    0.127755      -1.49372    1.44605   -0.459277
  ⋮                                  ⋱
  0.0276837  -0.0109498   0.0322572     -0.272015   0.801335  -1.18499
 -1.31163     0.357759   -0.834548       0.236446  -0.551559   0.558058
  0.357759   -1.08621     1.05154       -0.717885   0.694973  -0.220729
 -0.834548    1.05154    -3.37424        0.694973  -2.23006    0.650252
  0.844383   -0.333979    0.983879   …  -0.220729   0.650252  -0.961569
 -1.175       0.320494   -0.747619      -0.74581    1.73976   -1.76026
  0.320494   -0.973069    0.942013       2.26439   -2.19212    0.696236
 -0.747619    0.942013   -3.02277       -2.19212    7.03418   -2.05106
  0.75643    -0.299191    0.881396       0.696236  -2.05106    3.03303

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.