Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1.8243     1.85439  -1.23101  …    2.1799    -1.44709    0.973795
  1.85439   10.0647   -6.11606      11.8314    -7.18964   -8.13691
 -1.23101   -6.11606   5.35474      -7.18964    6.29468    2.69786
  0.828385  -6.92188   2.29501      -8.13691    2.69786   11.5576
  2.47516    2.51599  -1.67019       3.28515   -2.18079    1.46753
  2.51599   13.6554   -8.2981   …   17.83     -10.8349   -12.2625
 -1.67019   -8.2981    7.26516     -10.8349     9.48619    4.06572
  1.12393   -9.39142   3.1138      -12.2625     4.06572   17.4175
  2.27595    2.31349  -1.53577       2.45512   -1.62979    1.09674
  2.31349   12.5564   -7.63024      13.3251    -8.09733   -9.1642
  ⋮                             ⋱
  1.0044    -8.39265   2.78265     -12.0736     4.00309   17.1492
  2.3249     2.36324  -1.5688        3.38463   -2.24683    1.51197
  2.36324   12.8264   -7.79432      18.37     -11.163    -12.6338
 -1.5688    -7.79432   6.82409     -11.163      9.77346    4.18884
  1.05569   -8.82126   2.92476  …  -12.6338     4.18884   17.945
  2.14453    2.1799   -1.44709       3.55794   -2.36188    1.58939
  2.1799    11.8314   -7.18964      19.3106   -11.7346   -13.2807
 -1.44709   -7.18964   6.29468     -11.7346    10.2739     4.40333
  0.973795  -8.13691   2.69786     -13.2807     4.40333   18.8639
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0001359220454330783
   0.006748814478220137
   0.030819030936338025
   0.09830101736633068
   0.0005754105408992727
   0.028570339542553957
   0.1304688669491873
   0.4161461916901357
   0.001283550634859824
   0.06373097962490283
   ⋮
  18.852747000417697
   0.03717586503972033
   1.8458596280023647
   8.429273787734232
  26.886185704711036
   0.4250009927820498
  21.102190187077667
  96.36493258169273
 307.36757851945083
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.376692    0.534187     0.282126    …  -0.163234   -0.0862108  -0.0160693
 -0.445568    0.169202     0.181435       -0.0517039  -0.0554421  -0.152646
 -0.306118    0.441036    -0.372849       -0.13477     0.113934    0.0877039
 -0.274732   -0.033399     0.508019        0.0102059  -0.155238    0.127808
  0.0695419   0.0986173    0.0520838      -0.294076   -0.155314   -0.0289498
 -0.0822571   0.0312366    0.0334951   …  -0.0931475  -0.0998822  -0.275
 -0.056513    0.0814206   -0.0688324      -0.242796    0.205258    0.158004
 -0.0507187  -0.00616585   0.0937864       0.0183865  -0.27967     0.230254
 -0.313225   -0.444184    -0.234592       -0.227169   -0.119977   -0.0223632
  0.370496   -0.140694    -0.150866       -0.0719548  -0.0771572  -0.212432
  ⋮                                    ⋱
  0.0179609   0.0021835   -0.0332123       0.0178727  -0.271854    0.223819
 -0.014638   -0.0207582   -0.0109632      -0.265044   -0.139981   -0.0260919
  0.0173145  -0.00657507  -0.00705045     -0.0839518  -0.0900216  -0.247851
  0.0118955  -0.0171384    0.0144887      -0.218826    0.184994    0.142405
  0.0106759   0.00129786  -0.0197413   …   0.0165714  -0.252061    0.207523
 -0.145603   -0.206479    -0.10905        -0.221272   -0.116863   -0.0217827
  0.172225   -0.0654014   -0.07013        -0.070087   -0.0751543  -0.206918
  0.118324   -0.170473     0.144117       -0.182687    0.154442    0.118887
  0.106192    0.0129097   -0.196364        0.0138346  -0.210432    0.17325
julia> logdet(E)-37.0212260847147
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  2.2854663654846212
 11.79382740512271
  7.645999723115072
  2.9602422670726707
  5.816002651188354
 -4.155514395630833
 -3.528647911721028
 -2.653997968098717
 -1.1963430133065391
 -4.9975717084715185
  ⋮
  1.4896777023122063
 -4.26549100506332
  1.2043487669744182
  3.9314303072709316
 -2.169097335999905
 -6.449993445841667
  4.472200292035162
  2.969360143400582
  5.9591125934120734

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.9499    -0.795217   6.60591  …  -0.580206   4.81981   -4.81352
 -0.795217  24.836     27.4         18.1208    19.9916     1.22762
  6.60591   27.4       35.3818      19.9916    25.8153    -1.84573
 -6.59729    1.68255   -2.52971      1.22762   -1.84573    7.06255
  7.57151   -0.549866   4.56777     -0.438168   3.63989   -3.63514
 -0.549866  17.1733    18.9462   …  13.6847    15.0975     0.927094
  4.56777   18.9462    24.4654      15.0975    19.4956    -1.39388
 -4.56181    1.16343   -1.74921      0.927094  -1.39388    5.3336
  9.09556   -0.660547   5.4872      -0.598624   4.9728    -4.96631
 -0.660547  20.63      22.7598      18.6961    20.6262     1.26659
  ⋮                              ⋱
 -5.12005    1.3058    -1.96327      1.08633   -1.63329    6.24966
  4.39351   -0.31907    2.65053     -0.24519    2.03681   -2.03415
 -0.31907    9.9651    10.9939       7.65772    8.44827    0.518784
  2.65053   10.9939    14.1965       8.44827   10.9093    -0.779991
 -2.64707    0.675101  -1.01501  …   0.518784  -0.779991   2.98458
  7.98929   -0.580206   4.81981     -0.61918    5.14356   -5.13685
 -0.580206  18.1208    19.9916      19.3381    21.3344     1.31009
  4.81981   19.9916    25.8153      21.3344    27.5494    -1.96971
 -4.81352    1.22762   -1.84573      1.31009   -1.96971    7.53696
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 3.30907  -0.240314  1.99631   -1.9937    2.28811  …  1.45654     -1.45464
 0.0       4.97777   5.60084    0.241762  0.0         4.08648      0.176395
 0.0       0.0       0.165071   0.583094  0.0         0.120439     0.425437
 0.0       0.0       0.0        2.30358   0.0         0.0          1.68074
 0.0      -0.0       0.0       -0.0       2.44306     0.125724    -0.12556
 0.0       0.0       0.0        0.0       0.0      …  0.352731     0.0152258
 0.0       0.0       0.0        0.0       0.0         0.0103959    0.0367223
 0.0       0.0       0.0        0.0       0.0         0.0          0.145076
 0.0      -0.0       0.0       -0.0       0.0         0.48356     -0.482929
 0.0       0.0       0.0        0.0       0.0         1.35668      0.0585614
 ⋮                                                 ⋱
 0.0       0.0       0.0        0.0       0.0         0.0          0.3243
 0.0      -0.0       0.0       -0.0       0.0         0.294343    -0.293959
 0.0       0.0       0.0        0.0       0.0         0.825808     0.0356463
 0.0       0.0       0.0        0.0       0.0         0.0243387    0.0859735
 0.0       0.0       0.0        0.0       0.0      …  0.0          0.339648
 0.0      -0.0       0.0       -0.0       0.0         0.0473704   -0.0473085
 0.0       0.0       0.0        0.0       0.0         0.132902     0.00573678
 0.0       0.0       0.0        0.0       0.0         0.00391698   0.0138362
 0.0       0.0       0.0        0.0       0.0         0.0          0.0546616
julia> logdet(C)-49.05899785842215
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  24212.1     38835.6    -34543.3    …  -45705.6     40654.0     -851.855
  38835.6     62864.4    -55866.1       -73985.0     65748.7    -1107.9
 -34543.3    -55866.1     49687.8        65748.7    -58477.6      996.832
    723.813     941.371    -846.999      -1107.9       996.832   -282.199
  12517.5     20077.9    -17858.7       -23803.6     21172.7     -443.647
  20077.9     32500.6    -28882.5    …  -38531.5     34242.0     -576.995
 -17858.7    -28882.5     25688.4        34242.0    -30455.2      519.152
    374.208     486.685    -437.895       -576.995     519.152   -146.969
   4303.57     6902.83    -6139.89       -8338.73     7417.09    -155.416
   6902.83    11173.8     -9929.89      -13498.1     11995.5     -202.13
      ⋮                              ⋱
    770.945    1002.67     -902.152      -1181.84     1063.36    -301.033
  13681.7     21945.1    -19519.6       -25902.2     23039.3     -482.761
  21945.1     35523.2    -31568.6       -41928.6     37260.9     -627.865
 -19519.6    -31568.6     28077.4        37260.9    -33140.2      564.922
    409.01      531.946    -478.619  …    -627.865     564.922   -159.927
 -28495.2    -45705.6     40654.0        54206.3    -48215.1     1010.29
 -45705.6    -73985.0     65748.7        87745.2    -77977.1     1313.95
  40654.0     65748.7    -58477.6       -77977.1     69353.6    -1182.23
   -851.855   -1107.9       996.832       1313.95    -1182.23     334.684

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.