Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 19.0284     16.2487     0.358002  …  13.2325     0.291546   0.0961117
 16.2487     14.2746     1.69222      11.6248     1.3781     0.730815
  0.358002    1.69222   22.9477        1.3781    18.6879     7.77912
  0.11802     0.8974     9.55232       0.730815   7.77912    3.98924
 15.8789     13.5593     0.298746     14.5814     0.321268   0.10591
 13.5593     11.9119     1.41213   …  12.8099     1.51858    0.805317
  0.298746    1.41213   19.1494        1.51858   20.593      8.57215
  0.0984852   0.748863   7.97123       0.805317   8.57215    4.39592
 12.0557     10.2946     0.226817     14.9211     0.328751   0.108377
 10.2946      9.04383    1.07213      13.1083     1.55396    0.824077
  ⋮                                ⋱
  0.0876566   0.666525   7.09479       0.816484   8.69102    4.45688
 16.9491     14.4731     0.318881     15.0326     0.331207   0.109186
 14.4731     12.7147     1.5073       13.2062     1.56556    0.830232
  0.318881    1.5073    20.4401        1.56556   21.2301     8.83736
  0.105123    0.799336   8.50848   …   0.830232   8.83736    4.53192
 15.4962     13.2325     0.291546     18.7148     0.412336   0.135931
 13.2325     11.6248     1.3781       16.441      1.94905    1.0336
  0.291546    1.3781    18.6879        1.94905   26.4305    11.0021
  0.0961117   0.730815   7.77912       1.0336    11.0021     5.64201
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   0.0012251912460550453
   0.006813802965454383
   0.22443672824005048
   0.2806299795799183
   0.0033838018087656956
   0.01881874268471693
   0.6198619272031732
   0.7750595962499591
   0.015572063638374486
   0.08660278445430009
   ⋮
  13.562165725180767
   0.09916104086099768
   0.5514761850052045
  18.16482092194151
  22.712830312448204
   1.1322778408653524
   6.297072505740055
 207.4163808146495
 259.3481698338562
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.13545      0.0299144  -0.0439872  …   0.0443293  -0.0651835  0.230175
  0.157526    -0.0340103  -0.0256714     -0.050399   -0.0380418  0.202495
  0.00941763   0.0841814   0.189826       0.124746    0.281298   0.0678063
 -0.04531     -0.19017     0.0817007     -0.281808    0.12107    0.030008
 -0.0549002    0.0121248  -0.0178287      0.0466925  -0.0686583  0.242445
  0.0638479   -0.0137849  -0.0104051  …  -0.0530857  -0.0400698  0.21329
  0.00381712   0.0341201   0.0769396      0.131396    0.296294   0.071421
 -0.0183649   -0.0770792   0.0331147     -0.296831    0.127524   0.0316077
  0.170186    -0.0375859   0.0552676      0.0419629  -0.0617039  0.217887
 -0.197923     0.0427322   0.0322548     -0.0477086  -0.0360111  0.191686
  ⋮                                   ⋱
  0.0375565    0.157628   -0.06772       -0.278334    0.119578   0.029638
  0.275079    -0.0607517   0.0893315      0.0497192  -0.0731089  0.258161
 -0.319912     0.06907     0.0521349     -0.0565269  -0.0426672  0.227116
 -0.0191258   -0.17096    -0.385509       0.139914    0.315501   0.0760506
  0.092018     0.386208   -0.165922   …  -0.316073    0.135791   0.0336566
  0.0877133   -0.0193716   0.0284848      0.0509724  -0.0749517  0.264668
 -0.102009     0.022024    0.016624      -0.0579517  -0.0437427  0.232841
 -0.00609857  -0.0545133  -0.122926       0.14344     0.323453   0.0779676
  0.0293414    0.123149   -0.0529069     -0.32404     0.139214   0.0345049
julia> logdet(E)-19.6732817183562
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  3.6904389924306695
 -5.489391793476282
 -1.8887547243535743
  5.640673307563998
  0.6484484377025468
 -1.2657687696785842
  1.3549867617974871
 -3.877771278404388
 -7.714804428455689
 10.367598466656762
  ⋮
  3.2475234253010874
  0.23417345359557176
  2.3824342887457286
 -0.803628560782781
  2.2009543650366004
  6.08078382138962
 -6.903600155681593
  0.5606868205536651
 -1.4724513917935784

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of aKronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of aKronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  3.86664   -1.57178   -1.6576    …  -0.835355  -0.880962  -1.18158
 -1.57178    8.0467    -1.17314       4.27658   -0.62349    3.34943
 -1.6576    -1.17314   12.0227       -0.62349    6.38967    0.944687
 -2.22324    6.30221    1.7775        3.34943    0.944687   3.08174
  2.26767   -0.921805  -0.972132     -0.831968  -0.87739   -1.17679
 -0.921805   4.71916   -0.688015  …   4.25924   -0.620962   3.33585
 -0.972132  -0.688015   7.05094      -0.620962   6.36377    0.940857
 -1.30387    3.69606    1.04245       3.33585    0.940857   3.06925
  2.59209   -1.05368   -1.11121      -1.06004   -1.11792   -1.4994
 -1.05368    5.39429   -0.786443      5.42686   -0.791191   4.25034
  ⋮                               ⋱
 -1.60748    4.55672    1.28519       4.58661    1.29363    4.22004
  3.73706   -1.51911   -1.60204      -1.54773   -1.63223   -2.18922
 -1.51911    7.77703   -1.13383       7.92357   -1.15519    6.20577
 -1.60204   -1.13383   11.6197       -1.15519   11.8387     1.7503
 -2.14873    6.091      1.71793   …   6.20577    1.7503     5.7098
  2.055     -0.835355  -0.880962     -1.4118    -1.48888   -1.99695
 -0.835355   4.27658   -0.62349       7.22769   -1.05374    5.66076
 -0.880962  -0.62349    6.38967      -1.05374   10.799      1.59658
 -1.18158    3.34943    0.944687      5.66076    1.59658    5.20834
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 1.96638  -0.799329  -0.842969  …  -0.424819   -0.448012   -0.600894
 0.0       2.72172   -0.678597      1.44651    -0.360654    1.05416
 0.0       0.0        3.29417       0.0         1.75075     0.350164
 0.0       0.0        0.0           0.0         0.0         0.207165
 0.0      -0.0       -0.0          -0.408236   -0.430524   -0.577438
 0.0       0.0       -0.0       …   1.39005    -0.346576    1.01301
 0.0       0.0        0.0           0.0         1.68241     0.336495
 0.0       0.0        0.0           0.0         0.0         0.199079
 0.0      -0.0       -0.0          -0.0367893  -0.0387978  -0.0520373
 0.0       0.0       -0.0           0.125268   -0.0312326   0.0912898
 ⋮                              ⋱
 0.0       0.0        0.0           0.0         0.0         0.096442
 0.0      -0.0       -0.0          -0.32388    -0.341562   -0.458118
 0.0       0.0       -0.0           1.10281    -0.27496     0.803683
 0.0       0.0        0.0           0.0         1.33476     0.266963
 0.0       0.0        0.0       …   0.0         0.0         0.157942
 0.0      -0.0       -0.0          -0.0506767  -0.0534435  -0.0716807
 0.0       0.0       -0.0           0.172555   -0.0430224   0.125751
 0.0       0.0        0.0           0.0         0.208847    0.0417711
 0.0       0.0        0.0           0.0         0.0         0.0247128
julia> logdet(C)-49.80516574360102
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  10.393     -16.8974    -3.68738  …     31.7699     6.93287    -44.1469
 -16.8974     85.3197    21.6443       -160.415    -40.6948     199.005
  -3.68738    21.6443     7.34132       -40.6948   -13.8029      51.119
  23.4804   -105.845    -27.1886        199.005     51.119     -255.585
  53.7141    -87.3311   -19.0575        540.528    117.955     -751.108
 -87.3311    440.959    111.864    …  -2729.27    -692.374     3385.84
 -19.0575    111.864     37.9422       -692.374   -234.84       869.731
 121.354    -547.038   -140.519        3385.84     869.731    -4348.49
  -7.1712     11.6593     2.54431       -78.6923   -17.1723     109.349
  11.6593    -58.871    -14.9347        397.339    100.799     -492.925
   ⋮                               ⋱
 -17.8675     80.543     20.6893       -822.199   -211.201     1055.96
  23.8044    -38.7025    -8.44571       256.892     56.0594    -356.973
 -38.7025    195.419     49.5749      -1297.12    -329.059     1609.16
  -8.44571    49.5749    16.8148       -329.059   -111.61       413.35
  53.7803   -242.43     -62.2739   …   1609.16     413.35     -2066.67
 -19.5405     31.7699     6.93287      -203.533    -44.4153     282.826
  31.7699   -160.415    -40.6948       1027.7      260.71     -1274.92
   6.93287   -40.6948   -13.8029        260.71      88.4279    -327.493
 -44.1469    199.005     51.119       -1274.92    -327.493     1637.4

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.