Factorization methods

Many forms of matrix factorization such as eigenvalue decomposition, LU factorization, Cholesky factorization etc., can be computed efficiently. The decomposition of the Kronecker product is the Kronecker product of the decompositions. We have overloaded some of the factorization functions from LinearAlgebra to compute the factorization of instances of AbstractKroneckerProduct.

Eigenvalue decomposition

The function eigen of LinearAlgebra is overloaded to compute the decomposition of AbstractKroneckerProducts. The result is a factorization of the Eigen type, containing a vector of the eigenvalues and a matrix with the eigenvectors. Just like long-time users would expect! The eigenvectors are structured as Kronecker products and can be processed accordingly.

The functions det, logdet, inv and \ are overloaded the make use of this decomposition.

The eigenvalue decomposition of matrices can be used to solve large systems of the form:

\[(A \otimes B + c\cdot I) \mathbf{x} = \mathbf{b}\]

The case where $A$ and $B$ are positive semi-definite frequently occurs in machine learning, for example in ridge regression.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 17.3225    2.41066  -7.75671   4.45768  …    2.68752   -8.64755    4.96964
  2.41066  10.6693   -8.34899  -7.373        11.8947    -9.30786   -8.21978
 -7.75671  -8.34899  12.9812    3.13325      -9.30786   14.4721     3.4931
  4.45768  -7.373     3.13325   8.46977      -8.21978    3.4931     9.44251
 14.778     2.05655  -6.61729   3.80287       2.34204   -7.53592    4.3308
  2.05655   9.10207  -7.12258  -6.28995  …   10.3656    -8.11135   -7.16314
 -6.61729  -7.12258  11.0744    2.673        -8.11135   12.6117     3.04407
  3.80287  -6.28995   2.673     7.22562      -7.16314    3.04407    8.22869
 13.5535    1.88614  -6.06898   3.48777       2.58604   -8.32103    4.78199
  1.88614   8.34787  -6.5324   -5.76877      11.4456    -8.9564    -7.90941
  ⋮                                      ⋱
  5.85564  -9.68523   4.11587  11.126       -11.3384     4.81842   13.0251
 12.8526    1.78861  -5.75516   3.30742       2.45141   -7.88784    4.53304
  1.78861   7.91621  -6.19461  -5.47047      10.8497    -8.49014   -7.49765
 -5.75516  -6.19461   9.63154   2.32475      -8.49014   13.2007     3.18622
  3.30742  -5.47047   2.32475   6.28423  …   -7.49765    3.18622    8.61296
 19.312     2.68752  -8.64755   4.96964       3.9058   -12.5676     7.22243
  2.68752  11.8947   -9.30786  -8.21978      17.2867   -13.5272   -11.9459
 -8.64755  -9.30786  14.4721    3.4931      -13.5272    21.0325     5.07657
  4.96964  -8.21978   3.4931    9.44251     -11.9459     5.07657   13.7229
julia> E = eigen(K)Eigen{Float64,Float64,Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}},Array{Float64,1}}
values:
40-element Array{Float64,1}:
   1.0219992765142298e-5
   5.5720019517677454e-5
   0.0003491493204314003
   0.0004972252736064665
   0.010446451092922835
   0.056954684035925414
   0.3568858984376279
   0.5082429725988954
   0.0173252160600358
   0.09445812724112357
   ⋮
  18.55323097716493
   0.47583241169319684
   2.594267126794713
  16.256035302545612
  23.150300252847618
   5.9058698391866695
  32.19916004546818
 201.76437384000613
 287.33364241602595
vectors:
40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0571398   0.122496    0.175804   …  -0.131341   -0.188498   -0.162079
  0.181224    0.090394   -0.117149      -0.0969209   0.125607   -0.14099
  0.0393418   0.200441    0.0216435     -0.214914   -0.0232062   0.1852
  0.1854     -0.0931385   0.1641         0.0998636  -0.175948    0.0485621
 -0.0633946   0.135905    0.195048      -0.119794   -0.171926   -0.14783
  0.201061    0.100289   -0.129972   …  -0.0884001   0.114565   -0.128595
  0.0436484   0.222383    0.0240127     -0.19602    -0.0211661   0.168918
  0.205695   -0.103334    0.182063       0.0910841  -0.16048     0.0442928
  0.0176133  -0.0377594  -0.0541914     -0.132455   -0.190096   -0.163454
 -0.0558622  -0.0278639   0.0361111     -0.0977429   0.126673   -0.142185
  ⋮                                  ⋱
 -0.0759643   0.0381619  -0.0672369      0.141903   -0.250017    0.0690051
  0.0367461  -0.0787763  -0.113058      -0.119843   -0.171996   -0.147891
 -0.116544   -0.0581316   0.0753374     -0.0884363   0.114612   -0.128647
 -0.0253004  -0.128902   -0.0139187     -0.1961     -0.0211747   0.168987
 -0.119229    0.0598966  -0.105531   …   0.0911214  -0.160546    0.0443109
 -0.104132    0.223237    0.320385      -0.170549   -0.244768   -0.210463
  0.330263    0.164734   -0.213492      -0.125854    0.163104   -0.183078
  0.0716967   0.365285    0.0394431     -0.27907    -0.0301337   0.240485
  0.337873   -0.169736    0.299056       0.129675   -0.228472    0.0630587
julia> logdet(E)-33.016684165277894
julia> b = randn(40);
julia> (E + 0.1I) \ b  # solve a system40-element Array{Float64,1}:
  2.2748717256382003
  2.0067097170060078
 -1.3174800899074406
  3.0452185248119545
  0.2850544767790089
 -0.18754891636007756
 -3.784525652082047
  5.354251262814168
  1.4584663809878078
 -4.04140283823261
  ⋮
 -3.7460163014739396
 -1.0407690996284766
 -2.2396642283571864
 -1.887626614669458
  1.3946463003892975
  1.5675848494639413
  6.762495124608422
  0.24156039468695992
  3.5942605603696616

LinearAlgebra.eigen — Function

eigen(K::AbstractKroneckerProduct)

Wrapper around eigen from the LinearAlgebra package. If the matrices of an AbstractKroneckerProduct instance are square, performs Eigenvalue decompositon on them and returns an Eigen type. Otherwise, it collects the instance and runs eigen on the full matrix. The functions, \, inv, and logdet are overloaded to efficiently work with this type.

Missing docstring.

Missing docstring for +(E::Eigen, B::UniformScaling). Check Documenter's build log for details.

Missing docstring.

Missing docstring for +(::Eigen, ::UniformScaling). Check Documenter's build log for details.

LinearAlgebra.det — Method

det(K::AbstractKroneckerProduct)

Compute the determinant of a Kronecker product.

LinearAlgebra.logdet — Method

logdet(K::Eigen)

Compute the logarithm of the determinant of the eigenvalue decomp of a Kronecker product.

Base.inv — Method

inv(K::Eigen)

Compute the inverse of the eigenvalue decomp of a Kronecker product. Returns another type of Eigen.

Cholesky factorization

Similar to the eigenvalue decomposition, cholesky has been overloaded to allow for efficient Cholesky decomposition of Kronecker products of symmetric and positive definite matrices.

julia> A, B = rand(10, 10), randn(4, 4);
julia> As, Bs = (A, B) .|> X -> X * X';  # make positive definite
julia> K = As ⊗ Bs40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.47986    -5.70245     1.64253   …   -5.2037      1.49887    0.483453
 -5.70245    31.1913    -13.4191        28.4633    -12.2454     0.672532
  1.64253   -13.4191     10.3828       -12.2454      9.4747     0.36787
  0.52979     0.736991    0.403128       0.672532    0.36787   12.719
  3.74814    -4.77104     1.37425       -6.77566     1.95166    0.629497
 -4.77104    26.0967    -11.2273    …   37.0616    -15.9446     0.875694
  1.37425   -11.2273      8.68694      -15.9446     12.3369     0.478997
  0.443257    0.616615    0.337283       0.875694    0.478997  16.5612
  1.92903    -2.45548     0.707275      -4.1379      1.19188    0.384434
 -2.45548    13.431      -5.77827       22.6335     -9.73736    0.534786
  ⋮                                 ⋱
  0.507236    0.705617    0.385967       0.875103    0.478674  16.5501
  3.09532    -3.94006     1.13489       -5.42341     1.56216    0.503866
 -3.94006    21.5514     -9.27181       29.665     -12.7625     0.700928
  1.13489    -9.27181     7.17392      -12.7625      9.87475    0.383402
  0.366054    0.509218    0.278538  …    0.700928    0.383402  13.256
  4.08804    -5.2037      1.49887       -8.06429     2.32284    0.749218
 -5.2037     28.4633    -12.2454        44.1102    -18.977      1.04224
  1.49887   -12.2454      9.4747       -18.977      14.6832     0.570096
  0.483453    0.672532    0.36787        1.04224     0.570096  19.711
julia> C = cholesky(K)40×40 CholeskyKronecker{Cholesky{Float64,Array{Float64,2}},Cholesky{Float64,Array{Float64,2}}}
U factor:
40×40 Kronecker.KroneckerProduct{Float64,UpperTriangular{Float64,Array{Float64,2}},UpperTriangular{Float64,Array{Float64,2}}}:
 2.11657  -2.6942   0.776035  …  -2.45856     0.708161    0.228414
 0.0       4.8921  -2.31563       4.46422    -2.1131      0.263266
 0.0       0.0      2.10201       0.0         1.91816     0.380703
 0.0       0.0      0.0           0.0         0.0         3.36752
 0.0      -0.0      0.0          -1.37807     0.396939    0.128031
 0.0       0.0     -0.0       …   2.50229    -1.18444     0.147566
 0.0       0.0      0.0           0.0         1.07517     0.213392
 0.0       0.0      0.0           0.0         0.0         1.88756
 0.0      -0.0      0.0          -0.510542    0.147056    0.0474322
 0.0       0.0     -0.0           0.927038   -0.438805    0.0546696
 ⋮                            ⋱
 0.0       0.0      0.0           0.0         0.0         0.878655
 0.0      -0.0      0.0           0.0884518  -0.0254776  -0.00821768
 0.0       0.0     -0.0          -0.16061     0.0760234  -0.00947156
 0.0       0.0      0.0          -0.0        -0.0690099  -0.0136966
 0.0       0.0      0.0       …  -0.0        -0.0        -0.121154
 0.0      -0.0      0.0          -0.585248    0.168575    0.0543729
 0.0       0.0     -0.0           1.06269    -0.503014    0.0626693
 0.0       0.0      0.0           0.0         0.456609    0.0906246
 0.0       0.0      0.0           0.0         0.0         0.801622
julia> logdet(C)20.505063609278235
julia> inv(C)40×40 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  7.84139     2.06774     1.44938    …  -1.41373   -0.990951    0.307195
  2.06774     2.40044     2.7864        -1.6412    -1.90509     0.195617
  1.44938     2.7864      5.8404        -1.90509   -3.99313     0.253892
 -0.449308   -0.286112   -0.371346       0.195617   0.253892   -1.27923
  4.89439     1.29063     0.904664      -1.23227   -0.863757    0.267765
  1.29063     1.49829     1.7392     …  -1.43054   -1.66056     0.170508
  0.904664    1.7392      3.64543       -1.66056   -3.48059     0.221304
 -0.280446   -0.178583   -0.231784       0.170508   0.221304   -1.11503
  2.81916     0.743401    0.521085      -0.545138  -0.382113    0.118455
  0.743401    0.863012    1.00178       -0.632849  -0.734604    0.0754302
  ⋮                                  ⋱
 -0.150908   -0.0960955  -0.124723       0.127623   0.165643   -0.834586
 -1.69802    -0.44776    -0.313856       0.178652   0.125226   -0.0388201
 -0.44776    -0.519804   -0.603383       0.207397   0.240744   -0.02472
 -0.313856   -0.603383   -1.26471        0.240744   0.504609   -0.0320842
  0.0972955   0.0619562   0.0804133  …  -0.02472   -0.0320842   0.161655
 -5.36123    -1.41373    -0.990951       1.71981    1.20549    -0.373703
 -1.41373    -1.6412     -1.90509        1.99652    2.31754    -0.237968
 -0.990951   -1.90509    -3.99313        2.31754    4.85764    -0.30886
  0.307195    0.195617    0.253892      -0.237968  -0.30886     1.55618

LinearAlgebra.cholesky — Function

cholesky(K::AbstractKroneckerProduct; check = true)

Wrapper around cholesky from the LinearAlgebra package. Performs Cholesky on the matrices of a AbstractKroneckerProduct instances and returns a CholeskyKronecker type. Similar to Cholesky, size, \, inv, det, and logdet are overloaded to efficiently work with this type.