Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.608086  0.301701
 0.127446  0.324382

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.369769   0.18346    0.18346    0.0910236
 0.077498   0.197252   0.0384505  0.0978664
 0.077498   0.0384505  0.197252   0.0978664
 0.0162424  0.0413411  0.0413411  0.105224

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  4.17257   -3.88082  -3.88082   3.60947
 -1.63935    7.82189   1.52473  -7.27499
 -1.63935    1.52473   7.82189  -7.27499
  0.644081  -3.07313  -3.07313  14.6629

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.00255612   0.00126822   0.00126822   …  1.14633e-6  1.14633e-6  5.68752e-7
 0.000535725  0.00136356   0.000265799     1.23251e-6  2.40254e-7  6.11508e-7
 0.000535725  0.000265799  0.00136356      2.40254e-7  1.23251e-6  6.11508e-7
 0.00011228   0.000285781  0.000285781     2.58316e-7  2.58316e-7  6.57479e-7
 0.000535725  0.000265799  0.000265799     1.23251e-6  1.23251e-6  6.11508e-7
 0.00011228   0.000285781  5.57075e-5   …  1.32517e-6  2.58316e-7  6.57479e-7
 0.00011228   5.57075e-5   0.000285781     2.58316e-7  1.32517e-6  6.57479e-7
 2.35322e-5   5.98953e-5   5.98953e-5      2.77735e-7  2.77735e-7  7.06906e-7
 0.000535725  0.000265799  0.000265799     1.23251e-6  1.23251e-6  6.11508e-7
 0.00011228   0.000285781  5.57075e-5      1.32517e-6  2.58316e-7  6.57479e-7
 ⋮                                      ⋱                          ⋮
 8.76077e-11  2.22984e-10  2.22984e-10     4.95984e-7  4.95984e-7  1.26241e-6
 1.99444e-9   9.89541e-10  9.89541e-10     2.20104e-6  2.20104e-6  1.09204e-6
 4.18006e-10  1.06393e-9   2.07393e-10     2.36651e-6  4.61305e-7  1.17414e-6
 4.18006e-10  2.07393e-10  1.06393e-9   …  4.61305e-7  2.36651e-6  1.17414e-6
 8.76077e-11  2.22984e-10  2.22984e-10     4.95984e-7  4.95984e-7  1.26241e-6
 4.18006e-10  2.07393e-10  2.07393e-10     2.36651e-6  2.36651e-6  1.17414e-6
 8.76077e-11  2.22984e-10  4.34664e-11     2.54441e-6  4.95984e-7  1.26241e-6
 8.76077e-11  4.34664e-11  2.22984e-10     4.95984e-7  2.54441e-6  1.26241e-6
 1.83612e-11  4.67341e-11  4.67341e-11  …  5.33271e-7  5.33271e-7  1.35731e-6

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 2.14152e-5   1.47584e-5   1.47584e-5   …  3.56626e-7   2.45771e-7
 6.2343e-6    7.53715e-6   4.29641e-6      1.03819e-7   1.25516e-7
 6.2343e-6    4.29641e-6   7.53715e-6      1.8213e-7    1.25516e-7
 1.8149e-6    2.19418e-6   2.19418e-6      5.30208e-8   6.41011e-8
 6.2343e-6    4.29641e-6   4.29641e-6      1.8213e-7    1.25516e-7
 1.8149e-6    2.19418e-6   1.25075e-6   …  5.30208e-8   6.41011e-8
 1.8149e-6    1.25075e-6   2.19418e-6      9.3014e-8    6.41011e-8
 5.28347e-7   6.38761e-7   6.38761e-7      2.70778e-8   3.27366e-8
 6.2343e-6    4.29641e-6   4.29641e-6      1.8213e-7    1.25516e-7
 1.8149e-6    2.19418e-6   1.25075e-6      5.30208e-8   6.41011e-8
 ⋮                                      ⋱               ⋮
 2.72549e-11  3.29506e-11  3.29506e-11     1.25301e-10  1.51487e-10
 3.21598e-10  2.21631e-10  2.21631e-10     8.42795e-10  5.80817e-10
 9.36221e-11  1.13187e-10  6.45202e-11     2.45351e-10  2.96624e-10
 9.36221e-11  6.45202e-11  1.13187e-10  …  4.30417e-10  2.96624e-10
 2.72549e-11  3.29506e-11  3.29506e-11     1.25301e-10  1.51487e-10
 9.36221e-11  6.45202e-11  6.45202e-11     4.30417e-10  2.96624e-10
 2.72549e-11  3.29506e-11  1.87828e-11     1.25301e-10  1.51487e-10
 2.72549e-11  1.87828e-11  3.29506e-11     2.19815e-10  1.51487e-10
 7.93432e-12  9.59244e-12  9.59244e-12  …  6.39916e-11  7.73646e-11

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.735882  0.482304
 0.302672  0.0363378

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.0465673   0.0305206    0.0305206    …  0.00103919   0.000681094
 0.0191534   0.00229949   0.0125533       0.000427425  5.13151e-5
 0.0191534   0.0125533    0.00229949      7.82948e-5   5.13151e-5
 0.0078779   0.000945793  0.000945793     3.22031e-5   3.86619e-6
 0.0191534   0.0125533    0.0125533       7.82948e-5   5.13151e-5
 0.0078779   0.000945793  0.00516325   …  3.22031e-5   3.86619e-6
 0.0078779   0.00516325   0.000945793     5.8989e-6    3.86619e-6
 0.00324022  0.00038901   0.00038901      2.42625e-6   2.91287e-7
 0.0191534   0.0125533    0.0125533       7.82948e-5   5.13151e-5
 0.0078779   0.000945793  0.00516325      3.22031e-5   3.86619e-6
 ⋮                                     ⋱
 1.56879e-5  1.88343e-6   1.88343e-6   …  4.43776e-13  5.32782e-14
 9.2733e-5   6.07781e-5   6.07781e-5      1.43206e-11  9.38585e-12
 3.81416e-5  4.57915e-6   2.49984e-5      5.89014e-12  7.0715e-13
 3.81416e-5  2.49984e-5   4.57915e-6      1.07894e-12  7.0715e-13
 1.56879e-5  1.88343e-6   1.88343e-6      4.43776e-13  5.32782e-14
 3.81416e-5  2.49984e-5   2.49984e-5   …  1.07894e-12  7.0715e-13
 1.56879e-5  1.88343e-6   1.0282e-5       4.43776e-13  5.32782e-14
 1.56879e-5  1.0282e-5    1.88343e-6      8.129e-14    5.32782e-14
 6.45251e-6  7.74666e-7   7.74666e-7      3.34351e-14  4.0141e-15

julia> sum(P10)  # expected number of edges
83.83637619799966

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 84 stored entries:
  [146 ,    2]  =  true
  [129 ,    3]  =  true
  [553 ,    3]  =  true
  [866 ,    3]  =  true
  [2   ,    5]  =  true
  [223 ,   10]  =  true
  [129 ,   12]  =  true
  [281 ,   14]  =  true
  [1   ,   19]  =  true
  ⋮
  [3   ,  702]  =  true
  [26  ,  709]  =  true
  [321 ,  760]  =  true
  [103 ,  769]  =  true
  [577 ,  786]  =  true
  [7   ,  829]  =  true
  [129 ,  873]  =  true
  [34  ,  919]  =  true
  [129 ,  997]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source