Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.0169518  0.724345
 0.0578273  0.222504

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.000287363  0.0122789   0.0122789   0.524676
 0.000980275  0.00377184  0.0418869   0.16117
 0.000980275  0.0418869   0.00377184  0.16117
 0.00334399   0.0128668   0.0128668   0.049508

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 34.0786   -110.94      -110.94      361.158
 -8.8568      2.59633     28.8327     -8.45215
 -8.8568     28.8327       2.59633    -8.45215
  2.30182    -0.674768    -0.674768    0.197805

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 5.63098e-22  2.4061e-20   2.4061e-20   …  0.000488224  0.0208617
 1.92088e-21  7.39105e-21  8.20789e-20     0.00166547   0.00640828
 1.92088e-21  8.20789e-20  7.39105e-21     0.000149972  0.00640828
 6.55267e-21  2.52129e-20  2.52129e-20     0.000511597  0.00196849
 1.92088e-21  8.20789e-20  8.20789e-20     0.000149972  0.00640828
 6.55267e-21  2.52129e-20  2.79994e-19  …  0.000511597  0.00196849
 6.55267e-21  2.79994e-19  2.52129e-20     4.60684e-5   0.00196849
 2.2353e-20   8.60084e-20  8.60084e-20     0.000157152  0.000604679
 1.92088e-21  8.20789e-20  8.20789e-20     0.000149972  0.00640828
 6.55267e-21  2.52129e-20  2.79994e-19     0.000511597  0.00196849
 ⋮                                      ⋱               ⋮
 4.09897e-16  1.57718e-15  1.57718e-15     1.24582e-8   4.7936e-8
 3.52241e-17  1.50512e-15  1.50512e-15     1.1889e-8    5.08016e-7
 1.20159e-16  4.62341e-16  5.13438e-15     4.05569e-8   1.56052e-7
 1.20159e-16  5.13438e-15  4.62341e-16  …  3.65207e-9   1.56052e-7
 4.09897e-16  1.57718e-15  1.57718e-15     1.24582e-8   4.7936e-8
 1.20159e-16  5.13438e-15  5.13438e-15     3.65207e-9   1.56052e-7
 4.09897e-16  1.57718e-15  1.75148e-14     1.24582e-8   4.7936e-8
 4.09897e-16  1.75148e-14  1.57718e-15     1.12184e-9   4.7936e-8
 1.39827e-15  5.38019e-15  5.38019e-15  …  3.82691e-9   1.47249e-8

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 3.16645e-17  1.30226e-16  1.30226e-16  …  1.80272e-10  7.414e-10
 1.03964e-17  6.86195e-17  4.27569e-17     5.91888e-11  3.90664e-10
 1.03964e-17  4.27569e-17  6.86195e-17     9.49906e-11  3.90664e-10
 3.41345e-18  2.25298e-17  2.25298e-17     3.11882e-11  2.05852e-10
 1.03964e-17  4.27569e-17  4.27569e-17     9.49906e-11  3.90664e-10
 3.41345e-18  2.25298e-17  1.40384e-17  …  3.11882e-11  2.05852e-10
 3.41345e-18  1.40384e-17  2.25298e-17     5.00532e-11  2.05852e-10
 1.12074e-18  7.39721e-18  7.39721e-18     1.64339e-11  1.08469e-10
 1.03964e-17  4.27569e-17  4.27569e-17     9.49906e-11  3.90664e-10
 3.41345e-18  2.25298e-17  1.40384e-17     3.11882e-11  2.05852e-10
 ⋮                                      ⋱               ⋮
 1.51351e-22  9.98964e-22  9.98964e-22     9.76676e-14  6.44637e-13
 1.40399e-21  5.77415e-21  5.77415e-21     5.64533e-13  2.32174e-12
 4.60973e-22  3.04256e-21  1.89583e-21     1.85353e-13  1.22339e-12
 4.60973e-22  1.89583e-21  3.04256e-21  …  2.97468e-13  1.22339e-12
 1.51351e-22  9.98964e-22  9.98964e-22     9.76676e-14  6.44637e-13
 4.60973e-22  1.89583e-21  1.89583e-21     2.97468e-13  1.22339e-12
 1.51351e-22  9.98964e-22  6.22456e-22     9.76676e-14  6.44637e-13
 1.51351e-22  6.22456e-22  9.98964e-22     1.56744e-13  6.44637e-13
 4.96931e-23  3.27989e-22  3.27989e-22  …  5.14638e-14  3.39677e-13

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.872962  0.27545
 0.302186  0.690245

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.257013     0.0810966    0.0810966    …  7.96854e-6  7.96854e-6  2.51435e-6
 0.088968     0.203219     0.0280725       1.99682e-5  2.7584e-6   6.30067e-6
 0.088968     0.0280725    0.203219        2.7584e-6   1.99682e-5  6.30067e-6
 0.0307973    0.0703464    0.0703464       6.91222e-6  6.91222e-6  1.57887e-5
 0.088968     0.0280725    0.0280725       1.99682e-5  1.99682e-5  6.30067e-6
 0.0307973    0.0703464    0.0097176    …  5.00381e-5  6.91222e-6  1.57887e-5
 0.0307973    0.0097176    0.0703464       6.91222e-6  5.00381e-5  1.57887e-5
 0.0106608    0.0243512    0.0243512       1.73212e-5  1.73212e-5  3.95647e-5
 0.088968     0.0280725    0.0280725       1.99682e-5  1.99682e-5  6.30067e-6
 0.0307973    0.0703464    0.0097176       5.00381e-5  6.91222e-6  1.57887e-5
 ⋮                                      ⋱
 1.83425e-5   4.18975e-5   4.18975e-5   …  0.00428887  0.00428887  0.00979655
 0.000153074  4.83003e-5   4.83003e-5      0.0049443   0.0049443   0.0015601
 5.29884e-5   0.000121035  1.67197e-5      0.0123898   0.00171152  0.00390942
 5.29884e-5   1.67197e-5   0.000121035     0.00171152  0.0123898   0.00390942
 1.83425e-5   4.18975e-5   4.18975e-5      0.00428887  0.00428887  0.00979655
 5.29884e-5   1.67197e-5   1.67197e-5   …  0.0123898   0.0123898   0.00390942
 1.83425e-5   4.18975e-5   5.78769e-6      0.0310475   0.00428887  0.00979655
 1.83425e-5   5.78769e-6   4.18975e-5      0.00428887  0.0310475   0.00979655
 6.34946e-6   1.45033e-5   1.45033e-5      0.0107474   0.0107474   0.024549

julia> sum(P10)  # expected number of edges
2022.31307664358

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 2006 stored entries:
  [7   ,    1]  =  true
  [67  ,    1]  =  true
  [325 ,    1]  =  true
  [729 ,    1]  =  true
  [9   ,    2]  =  true
  [35  ,    2]  =  true
  [74  ,    2]  =  true
  [12  ,    3]  =  true
  [265 ,    3]  =  true
  ⋮
  [1010, 1016]  =  true
  [345 , 1017]  =  true
  [763 , 1017]  =  true
  [1019, 1019]  =  true
  [385 , 1021]  =  true
  [793 , 1021]  =  true
  [831 , 1021]  =  true
  [1001, 1021]  =  true
  [965 , 1023]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source