Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.587946  0.642263
 0.225241  0.204518
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.34568    0.377616   0.377616   0.412502
 0.132429   0.120246   0.144664   0.131355
 0.132429   0.144664   0.120246   0.131355
 0.0507335  0.0460659  0.0460659  0.0418278
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  70.1513  -220.301  -220.301   691.826
 -77.2592   201.67    242.623  -633.317
 -77.2592   242.623   201.67   -633.317
  85.0874  -222.103  -222.103   579.756
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.00170627   0.00186391   0.00186391   …  0.00451006   0.00492672
 0.000653669  0.000593531  0.000714059     0.00172779   0.00156883
 0.000653669  0.000714059  0.000593531     0.00143615   0.00156883
 0.000250419  0.00022738   0.00022738      0.000550188  0.00049957
 0.000653669  0.000714059  0.000714059     0.00143615   0.00156883
 0.000250419  0.00022738   0.000273554  …  0.000550188  0.00049957
 0.000250419  0.000273554  0.00022738      0.00045732   0.00049957
 9.59352e-5   8.7109e-5    8.7109e-5       0.000175199  0.00015908
 0.000653669  0.000714059  0.000714059     0.00143615   0.00156883
 0.000250419  0.00022738   0.000273554     0.000550188  0.00049957
 ⋮                                      ⋱               ⋮
 4.45101e-8   4.04151e-8   4.04151e-8      1.85219e-8   1.68178e-8
 3.03276e-7   3.31294e-7   3.31294e-7      1.51829e-7   1.65856e-7
 1.16184e-7   1.05495e-7   1.26918e-7      5.81655e-8   5.28142e-8
 1.16184e-7   1.26918e-7   1.05495e-7   …  4.83476e-8   5.28142e-8
 4.45101e-8   4.04151e-8   4.04151e-8      1.85219e-8   1.68178e-8
 1.16184e-7   1.26918e-7   1.26918e-7      4.83476e-8   5.28142e-8
 4.45101e-8   4.04151e-8   4.86221e-8      1.85219e-8   1.68178e-8
 4.45101e-8   4.86221e-8   4.04151e-8      1.53955e-8   1.68178e-8
 1.70517e-8   1.54829e-8   1.54829e-8   …  5.89799e-9   5.35536e-9
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000193203  0.000200542  0.000200542  …  0.000291154  0.000302214
 7.03299e-5   7.34806e-5   7.30015e-5      0.000105986  0.000110734
 7.03299e-5   7.30015e-5   7.34806e-5      0.000106682  0.000110734
 2.56015e-5   2.67484e-5   2.67484e-5      3.88343e-5   4.0574e-5
 7.03299e-5   7.30015e-5   7.30015e-5      0.000106682  0.000110734
 2.56015e-5   2.67484e-5   2.6574e-5    …  3.88343e-5   4.0574e-5
 2.56015e-5   2.6574e-5    2.67484e-5      3.90891e-5   4.0574e-5
 9.31948e-6   9.73698e-6   9.73698e-6      1.42292e-5   1.48667e-5
 7.03299e-5   7.30015e-5   7.30015e-5      0.000106682  0.000110734
 2.56015e-5   2.67484e-5   2.6574e-5       3.88343e-5   4.0574e-5
 ⋮                                      ⋱               ⋮
 2.87342e-9   3.00214e-9   3.00214e-9      4.62293e-9   4.83003e-9
 2.16844e-8   2.25081e-8   2.25081e-8      3.46597e-8   3.59763e-8
 7.89356e-9   8.24718e-9   8.1934e-9       1.26168e-8   1.3182e-8
 7.89356e-9   8.1934e-9    8.24718e-9   …  1.26996e-8   1.3182e-8
 2.87342e-9   3.00214e-9   3.00214e-9      4.62293e-9   4.83003e-9
 7.89356e-9   8.1934e-9    8.1934e-9       1.26996e-8   1.3182e-8
 2.87342e-9   3.00214e-9   2.98257e-9      4.62293e-9   4.83003e-9
 2.87342e-9   2.98257e-9   3.00214e-9      4.65327e-9   4.83003e-9
 1.04598e-9   1.09284e-9   1.09284e-9   …  1.69388e-9   1.76977e-9

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.795818  0.266492
 0.956671  0.955062
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.101892  0.0341201  0.0341201  …  5.39483e-6  5.39483e-6  1.80655e-6
 0.122487  0.122281   0.0410166     1.93341e-5  6.48525e-6  6.47435e-6
 0.122487  0.0410166  0.122281      6.48525e-6  1.93341e-5  6.47435e-6
 0.147244  0.146996   0.146996      2.3242e-5   2.3242e-5   2.32029e-5
 0.122487  0.0410166  0.0410166     1.93341e-5  1.93341e-5  6.47435e-6
 0.147244  0.146996   0.0493069  …  6.92902e-5  2.3242e-5   2.32029e-5
 0.147244  0.0493069  0.146996      2.3242e-5   6.92902e-5  2.32029e-5
 0.177005  0.176707   0.176707      8.32953e-5  8.32953e-5  8.31552e-5
 0.122487  0.0410166  0.0410166     1.93341e-5  1.93341e-5  6.47435e-6
 0.147244  0.146996   0.0493069     6.92902e-5  2.3242e-5   2.32029e-5
 ⋮                               ⋱
 0.534166  0.533268   0.533268   …  0.176482    0.176482    0.176186
 0.36964   0.12378    0.12378       0.0409644   0.0409644   0.0137176
 0.444353  0.443606   0.148799      0.146809    0.0492442   0.0491614
 0.444353  0.148799   0.443606      0.0492442   0.146809    0.0491614
 0.534166  0.533268   0.533268      0.176482    0.176482    0.176186
 0.444353  0.148799   0.148799   …  0.146809    0.146809    0.0491614
 0.534166  0.533268   0.178874      0.526138    0.176482    0.176186
 0.534166  0.178874   0.533268      0.176482    0.526138    0.176186
 0.642133  0.641054   0.641054      0.632482    0.632482    0.631419
julia> sum(P10)  # expected number of edges54134.48898402788
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 49246 stored entries:
  [3   ,    1]  =  1
  [8   ,    1]  =  1
  [9   ,    1]  =  1
  [14  ,    1]  =  1
  [16  ,    1]  =  1
  [22  ,    1]  =  1
  [24  ,    1]  =  1
  [31  ,    1]  =  1
  [32  ,    1]  =  1
  ⋮
  [818 , 1023]  =  1
  [860 , 1023]  =  1
  [888 , 1023]  =  1
  [894 , 1023]  =  1
  [959 , 1023]  =  1
  [1022, 1023]  =  1
  [829 , 1024]  =  1
  [830 , 1024]  =  1
  [983 , 1024]  =  1
  [1024, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.