Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.851452   0.769736
 0.0157868  0.102332
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.724971     0.655393   0.655393   0.592493
 0.0134417    0.087131   0.0121517  0.0787688
 0.0134417    0.0121517  0.087131   0.0787688
 0.000249225  0.0016155  0.0016155  0.0104719
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  1.8627    -14.0111   -14.0111    105.39
 -0.287359   15.4985     2.16149  -116.579
 -0.287359    2.16149   15.4985   -116.579
  0.044331   -2.39096   -2.39096   128.955
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.145185     0.131251     0.131251     …  0.0478539    0.0432612
 0.00269189   0.0174492    0.00243354      0.000887262  0.00575134
 0.00269189   0.00243354   0.0174492       0.00636192   0.00575134
 4.99106e-5   0.000323527  0.000323527     0.000117957  0.000764611
 0.00269189   0.00243354   0.00243354      0.00636192   0.00575134
 4.99106e-5   0.000323527  4.51205e-5   …  0.000117957  0.000764611
 4.99106e-5   4.51205e-5   0.000323527     0.000845783  0.000764611
 9.25396e-7   5.99853e-6   5.99853e-6      1.56817e-5   0.000101651
 0.00269189   0.00243354   0.00243354      0.00636192   0.00575134
 4.99106e-5   0.000323527  4.51205e-5      0.000117957  0.000764611
 ⋮                                      ⋱               ⋮
 1.29244e-20  8.37774e-20  8.37774e-20     1.53025e-12  9.9193e-12
 3.75958e-17  3.39876e-17  3.39876e-17     6.20808e-10  5.61227e-10
 6.97067e-19  4.51847e-18  6.30167e-19     1.15105e-11  7.46122e-11
 6.97067e-19  6.30167e-19  4.51847e-18  …  8.25331e-11  7.46122e-11
 1.29244e-20  8.37774e-20  8.37774e-20     1.53025e-12  9.9193e-12
 6.97067e-19  6.30167e-19  6.30167e-19     8.25331e-11  7.46122e-11
 1.29244e-20  8.37774e-20  1.1684e-20      1.53025e-12  9.9193e-12
 1.29244e-20  1.1684e-20   8.37774e-20     1.09723e-11  9.9193e-12
 2.39632e-22  1.55332e-21  1.55332e-21  …  2.03439e-13  1.31872e-12
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.025732     0.0256287    0.0256287    …  0.0246177    0.0245189
 0.00052563   0.000789762  0.000523519     0.000502868  0.000755562
 0.00052563   0.000523519  0.000789762     0.000758609  0.000755562
 1.07371e-5   1.61325e-5   1.61325e-5      1.54962e-5   2.32831e-5
 0.00052563   0.000523519  0.000523519     0.000758609  0.000755562
 1.07371e-5   1.61325e-5   1.0694e-5    …  1.54962e-5   2.32831e-5
 1.07371e-5   1.0694e-5    1.61325e-5      2.3377e-5    2.32831e-5
 2.19327e-7   3.2954e-7    3.2954e-7       4.77523e-7   7.1748e-7
 0.00052563   0.000523519  0.000523519     0.000758609  0.000755562
 1.07371e-5   1.61325e-5   1.0694e-5       1.54962e-5   2.32831e-5
 ⋮                                      ⋱               ⋮
 6.64876e-21  9.9898e-21   9.9898e-21      3.88287e-19  5.83403e-19
 1.59341e-17  1.58701e-17  1.58701e-17     6.16846e-16  6.14369e-16
 3.25488e-19  4.89047e-19  3.24181e-19     1.26004e-17  1.89321e-17
 3.25488e-19  3.24181e-19  4.89047e-19  …  1.90085e-17  1.89321e-17
 6.64876e-21  9.9898e-21   9.9898e-21      3.88287e-19  5.83403e-19
 3.25488e-19  3.24181e-19  3.24181e-19     1.90085e-17  1.89321e-17
 6.64876e-21  9.9898e-21   6.62206e-21     3.88287e-19  5.83403e-19
 6.64876e-21  6.62206e-21  9.9898e-21      5.85756e-19  5.83403e-19
 1.35815e-22  2.04062e-22  2.04062e-22  …  1.19653e-20  1.79779e-20

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.625127  0.894991
 0.224069  0.391094
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.00911348   0.0130477    0.0130477    …  0.230322     0.32975
 0.00326662   0.0057016    0.00467679      0.082556     0.144094
 0.00326662   0.00467679   0.0057016       0.100646     0.144094
 0.00117088   0.00204367   0.00204367      0.0360754    0.0629665
 0.00326662   0.00467679   0.00467679      0.100646     0.144094
 0.00117088   0.00204367   0.00167634   …  0.0360754    0.0629665
 0.00117088   0.00167634   0.00204367      0.0439804    0.0629665
 0.000419688  0.000732529  0.000732529     0.0157643    0.0275152
 0.00326662   0.00467679   0.00467679      0.100646     0.144094
 0.00117088   0.00204367   0.00167634      0.0360754    0.0629665
 ⋮                                      ⋱
 8.90043e-7   1.55349e-6   1.55349e-6   …  0.000109761  0.000191578
 6.9276e-6    9.9182e-6    9.9182e-6       0.000700763  0.00100328
 2.48311e-6   4.33406e-6   3.55506e-6      0.00025118   0.000438413
 2.48311e-6   3.55506e-6   4.33406e-6      0.00030622   0.000438413
 8.90043e-7   1.55349e-6   1.55349e-6      0.000109761  0.000191578
 2.48311e-6   3.55506e-6   3.55506e-6   …  0.00030622   0.000438413
 8.90043e-7   1.55349e-6   1.27427e-6      0.000109761  0.000191578
 8.90043e-7   1.27427e-6   1.55349e-6      0.000133812  0.000191578
 3.19025e-7   5.56831e-7   5.56831e-7      4.79634e-5   8.3716e-5
julia> sum(P10)  # expected number of edges1970.3812140805571
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1957 stored entries:
  [264,    1]  =  1
  [321,    3]  =  1
  [339,    4]  =  1
  [257,    8]  =  1
  [283,   11]  =  1
  [1  ,   13]  =  1
  [434,   13]  =  1
  [577,   14]  =  1
  [106,   15]  =  1
  ⋮
  [717, 1023]  =  1
  [823, 1023]  =  1
  [837, 1023]  =  1
  [75 , 1024]  =  1
  [129, 1024]  =  1
  [132, 1024]  =  1
  [206, 1024]  =  1
  [265, 1024]  =  1
  [398, 1024]  =  1
  [654, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.