Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.333313  0.0035049
 0.803551  0.271053
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.111097  0.00116823  0.00116823  1.22843e-5
 0.267834  0.0903455   0.00281637  0.000950015
 0.267834  0.00281637  0.0903455   0.000950015
 0.645694  0.217805    0.217805    0.0734699
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
   9.58968   -0.124001   -0.124001   0.00160342
 -28.4291    11.7924      0.367607  -0.152483
 -28.4291     0.367607   11.7924    -0.152483
  84.2794   -34.9591    -34.9591    14.501
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.88028e-6  1.97718e-8   1.97718e-8   …  3.26802e-28  3.43643e-30
 4.53298e-6  1.52906e-6   4.76658e-8      7.87854e-28  2.65758e-28
 4.53298e-6  4.76658e-8   1.52906e-6      2.52734e-26  2.65758e-28
 1.09281e-5  3.68626e-6   3.68626e-6      6.09291e-26  2.05526e-26
 4.53298e-6  4.76658e-8   4.76658e-8      2.52734e-26  2.65758e-28
 1.09281e-5  3.68626e-6   1.14913e-7   …  6.09291e-26  2.05526e-26
 1.09281e-5  1.14913e-7   3.68626e-6      1.95453e-24  2.05526e-26
 2.63455e-5  8.88685e-6   8.88685e-6      4.71198e-24  1.58944e-24
 4.53298e-6  4.76658e-8   4.76658e-8      2.52734e-26  2.65758e-28
 1.09281e-5  3.68626e-6   1.14913e-7      6.09291e-26  2.05526e-26
 ⋮                                     ⋱               ⋮
 0.0300606   0.01014      0.01014         6.02887e-9   2.03365e-9
 0.0051722   5.43875e-5   5.43875e-5      3.23367e-11  3.40031e-13
 0.0124691   0.00420609   0.000131117     7.79573e-11  2.62965e-11
 0.0124691   0.000131117  0.00420609   …  2.50077e-9   2.62965e-11
 0.0300606   0.01014      0.01014         6.02887e-9   2.03365e-9
 0.0124691   0.000131117  0.000131117     2.50077e-9   2.62965e-11
 0.0300606   0.01014      0.000316098     6.02887e-9   2.03365e-9
 0.0300606   0.000316098  0.01014         1.93399e-7   2.03365e-9
 0.0724703   0.0244456    0.0244456    …  4.66246e-7   1.57274e-7
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 4.77432e-12  8.87792e-14  8.87792e-14  …  4.38842e-31  8.16033e-33
 2.0354e-11   3.19729e-12  3.78485e-13     1.87088e-30  2.93886e-31
 2.0354e-11   3.78485e-13  3.19729e-12     1.58044e-29  2.93886e-31
 8.67734e-11  1.36307e-11  1.36307e-11     6.73777e-29  1.0584e-29
 2.0354e-11   3.78485e-13  3.78485e-13     1.58044e-29  2.93886e-31
 8.67734e-11  1.36307e-11  1.61356e-12  …  6.73777e-29  1.0584e-29
 8.67734e-11  1.61356e-12  1.36307e-11     5.6918e-28   1.0584e-29
 3.69934e-10  5.81108e-11  5.81108e-11     2.42654e-27  3.81171e-28
 2.0354e-11   3.78485e-13  3.78485e-13     1.58044e-29  2.93886e-31
 8.67734e-11  1.36307e-11  1.61356e-12     6.73777e-29  1.0584e-29
 ⋮                                      ⋱               ⋮
 4.03666e-5   6.34096e-6   6.34096e-6      6.86675e-15  1.07866e-15
 2.221e-6     4.12997e-8   4.12997e-8      4.47243e-17  8.31655e-19
 9.46858e-6   1.48737e-6   1.7607e-7       1.90669e-16  2.99512e-17
 9.46858e-6   1.7607e-7    1.48737e-6   …  1.6107e-15   2.99512e-17
 4.03666e-5   6.34096e-6   6.34096e-6      6.86675e-15  1.07866e-15
 9.46858e-6   1.7607e-7    1.7607e-7       1.6107e-15   2.99512e-17
 4.03666e-5   6.34096e-6   7.50623e-7      6.86675e-15  1.07866e-15
 4.03666e-5   7.50623e-7   6.34096e-6      5.80076e-14  1.07866e-15
 0.000172092  2.70329e-5   2.70329e-5   …  2.47299e-13  3.88468e-14

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.0699036  0.484521
 0.694499   0.762908
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 2.78609e-12  1.93112e-11  1.93112e-11  …  0.000102875  0.000713057
 2.76801e-11  3.04066e-11  1.91858e-10     0.00102208   0.00112275
 2.76801e-11  1.91858e-10  3.04066e-11     0.000161984  0.00112275
 2.75005e-10  3.02093e-10  3.02093e-10     0.00160932   0.00176784
 2.76801e-11  1.91858e-10  1.91858e-10     0.000161984  0.00112275
 2.75005e-10  3.02093e-10  1.90613e-9   …  0.00160932   0.00176784
 2.75005e-10  1.90613e-9   3.02093e-10     0.000255053  0.00176784
 2.7322e-9    3.00132e-9   3.00132e-9      0.00253398   0.00278358
 2.76801e-11  1.91858e-10  1.91858e-10     0.000161984  0.00112275
 2.75005e-10  3.02093e-10  1.90613e-9      0.00160932   0.00176784
 ⋮                                      ⋱
 0.00262752   0.00288633   0.00288633   …  0.0386156    0.0424192
 2.66196e-5   0.000184508  0.000184508     0.00246848   0.0171097
 0.000264468  0.000290519  0.0018331       0.0245246    0.0269403
 0.000264468  0.0018331    0.000290519     0.00388678   0.0269403
 0.00262752   0.00288633   0.00288633      0.0386156    0.0424192
 0.000264468  0.0018331    0.0018331    …  0.00388678   0.0269403
 0.00262752   0.00288633   0.018212        0.0386156    0.0424192
 0.00262752   0.018212     0.00288633      0.00611998   0.0424192
 0.0261047    0.028676     0.028676        0.0608026    0.0667917
julia> sum(P10)  # expected number of edges1086.2176016047729
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1082 stored entries:
  [1007,    4]  =  1
  [1024,   12]  =  1
  [1017,   19]  =  1
  [1022,   23]  =  1
  [998 ,   28]  =  1
  [957 ,   32]  =  1
  [1019,   32]  =  1
  [748 ,   37]  =  1
  [764 ,   37]  =  1
  ⋮
  [752 , 1022]  =  1
  [999 , 1022]  =  1
  [1008, 1022]  =  1
  [1019, 1022]  =  1
  [492 , 1023]  =  1
  [191 , 1024]  =  1
  [469 , 1024]  =  1
  [531 , 1024]  =  1
  [593 , 1024]  =  1
  [973 , 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.