Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.933621  0.689169
 0.676738  0.99972
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.871649  0.643423  0.643423  0.474954
 0.631817  0.93336   0.466387  0.688976
 0.631817  0.466387  0.93336   0.688976
 0.457974  0.676548  0.676548  0.99944
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  4.58325  -3.15952  -3.15952   2.17805
 -3.10253   4.28022   2.13876  -2.95062
 -3.10253   2.13876   4.28022  -2.95062
  2.10018  -2.8974   -2.8974    3.99722
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.43858     0.323745    0.323745    …  0.0155509  0.0155509  0.0114792
 0.317906    0.46963     0.234668       0.0225584  0.0112721  0.0166519
 0.317906    0.234668    0.46963        0.0112721  0.0225584  0.0166519
 0.230435    0.340413    0.340413       0.0163515  0.0163515  0.0241555
 0.317906    0.234668    0.234668       0.0225584  0.0225584  0.0166519
 0.230435    0.340413    0.170099    …  0.0327236  0.0163515  0.0241555
 0.230435    0.170099    0.340413       0.0163515  0.0327236  0.0241555
 0.167031    0.246749    0.246749       0.0237198  0.0237198  0.0350403
 0.317906    0.234668    0.234668       0.0225584  0.0225584  0.0166519
 0.230435    0.340413    0.170099       0.0327236  0.0163515  0.0241555
 ⋮                                   ⋱                        ⋮
 0.012729    0.0188041   0.0188041      0.465082   0.465082   0.687048
 0.0242267   0.0178834   0.0178834      0.442311   0.442311   0.3265
 0.0175608   0.0259419   0.0129628      0.641623   0.32061    0.473625
 0.0175608   0.0129628   0.0259419   …  0.32061    0.641623   0.473625
 0.012729    0.0188041   0.0188041      0.465082   0.465082   0.687048
 0.0175608   0.0129628   0.0129628      0.641623   0.641623   0.473625
 0.012729    0.0188041   0.00939614     0.930748   0.465082   0.687048
 0.012729    0.00939614  0.0188041      0.465082   0.930748   0.687048
 0.00922664  0.0136302   0.0136302   …  0.674655   0.674655   0.996643
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 32.9315  32.7927  32.7927  32.6546  …  31.5703  31.4374  31.4374  31.3049
 32.2012  36.0766  32.0656  35.9247     30.8703  34.5855  30.7402  34.4398
 32.2012  32.0656  36.0766  35.9247     30.8703  30.7402  34.5855  34.4398
 31.4872  35.2766  35.2766  39.5222     30.1857  33.8186  33.8186  37.8887
 32.2012  32.0656  32.0656  31.9305     34.7318  34.5855  34.5855  34.4398
 31.4872  35.2766  31.3545  35.128   …  33.9616  38.0489  33.8186  37.8887
 31.4872  31.3545  35.2766  35.128      33.9616  33.8186  38.0489  37.8887
 30.7889  34.4944  34.4944  38.6458     33.2085  37.2052  37.2052  41.6829
 32.2012  32.0656  32.0656  31.9305     34.7318  34.5855  34.5855  34.4398
 31.4872  35.2766  31.3545  35.128      33.9616  38.0489  33.8186  37.8887
  ⋮                                  ⋱                              ⋮
 25.7326  28.8296  28.8296  32.2992     71.2583  79.8342  79.8342  89.4423
 26.913   26.7996  26.7996  26.6867     74.5269  74.2129  74.2129  73.9004
 26.3162  29.4833  26.2053  29.3592     72.8742  81.6447  72.5673  81.3008
 26.3162  26.2053  29.4833  29.3592  …  72.8742  72.5673  81.6447  81.3008
 25.7326  28.8296  28.8296  32.2992     71.2583  79.8342  79.8342  89.4423
 26.3162  26.2053  26.2053  26.095      81.99    81.6447  81.6447  81.3008
 25.7326  28.8296  25.6242  28.7081     80.1719  89.8206  79.8342  89.4423
 25.7326  25.6242  28.8296  28.7081     80.1719  79.8342  89.8206  89.4423
 25.162   28.1903  28.1903  31.583   …  78.3941  87.8289  87.8289  98.3991

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.325346  0.0320863
 0.467229  0.0366855
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 1.32879e-5   1.31048e-6  1.31048e-6  1.29242e-7  …  1.17279e-14  1.15663e-15
 1.90828e-5   1.49833e-6  1.88198e-6  1.47768e-7     1.68424e-14  1.32242e-15
 1.90828e-5   1.88198e-6  1.49833e-6  1.47768e-7     1.3409e-14   1.32242e-15
 2.74047e-5   2.15174e-6  2.15174e-6  1.68949e-7     1.92566e-14  1.51198e-15
 1.90828e-5   1.88198e-6  1.88198e-6  1.85604e-7     1.3409e-14   1.32242e-15
 2.74047e-5   2.15174e-6  2.7027e-6   2.12209e-7  …  1.92566e-14  1.51198e-15
 2.74047e-5   2.7027e-6   2.15174e-6  2.12209e-7     1.5331e-14   1.51198e-15
 3.93558e-5   3.09011e-6  3.09011e-6  2.42627e-7     2.20168e-14  1.7287e-15
 1.90828e-5   1.88198e-6  1.88198e-6  1.85604e-7     1.3409e-14   1.32242e-15
 2.74047e-5   2.15174e-6  2.7027e-6   2.12209e-7     1.92566e-14  1.51198e-15
 ⋮                                                ⋱
 0.000345233  2.71067e-5  2.71067e-5  2.12835e-6  …  4.91823e-14  3.86166e-15
 0.000167396  1.65089e-5  1.65089e-5  1.62814e-6     2.99536e-14  2.95409e-15
 0.000240396  1.88753e-5  2.37084e-5  1.86152e-6     4.30163e-14  3.37753e-15
 0.000240396  2.37084e-5  1.88753e-5  1.86152e-6     3.42472e-14  3.37753e-15
 0.000345233  2.71067e-5  2.71067e-5  2.12835e-6     4.91823e-14  3.86166e-15
 0.000240396  2.37084e-5  2.37084e-5  2.33817e-6  …  3.42472e-14  3.37753e-15
 0.000345233  2.71067e-5  3.40475e-5  2.67332e-6     4.91823e-14  3.86166e-15
 0.000345233  3.40475e-5  2.71067e-5  2.67332e-6     3.91562e-14  3.86166e-15
 0.000495787  3.89279e-5  3.89279e-5  3.05651e-6     5.62321e-14  4.4152e-15
julia> sum(P10)  # expected number of edges0.22479168611633626
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 0 stored entries

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.