Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.48234   0.209791
 0.797077  0.836187

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.232652  0.101191  0.101191  0.0440123
 0.384462  0.403326  0.16722   0.175425
 0.384462  0.16722   0.403326  0.175425
 0.635332  0.666505  0.666505  0.699209

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  12.5427  -3.14684  -3.14684   0.789511
 -11.956    7.23503   2.99965  -1.8152
 -11.956    2.99965   7.23503  -1.8152
  11.3968  -6.89663  -6.89663   4.1734

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.000158576  6.89719e-5   6.89719e-5   …  1.67114e-8  1.67114e-8  7.26853e-9
 0.00026205   0.000274908  0.000113977     6.66084e-8  2.76159e-8  2.8971e-8
 0.00026205   0.000113977  0.000274908     2.76159e-8  6.66084e-8  2.8971e-8
 0.000433044  0.000454292  0.000454292     1.10072e-7  1.10072e-7  1.15473e-7
 0.00026205   0.000113977  0.000113977     6.66084e-8  6.66084e-8  2.8971e-8
 0.000433044  0.000454292  0.00018835   …  2.65488e-7  1.10072e-7  1.15473e-7
 0.000433044  0.00018835   0.000454292     1.10072e-7  2.65488e-7  1.15473e-7
 0.000715614  0.000750727  0.000750727     4.38725e-7  4.38725e-7  4.60252e-7
 0.00026205   0.000113977  0.000113977     6.66084e-8  6.66084e-8  2.8971e-8
 0.000433044  0.000454292  0.00018835      2.65488e-7  1.10072e-7  1.15473e-7
 ⋮                                      ⋱                          ⋮
 0.0397975    0.0417502    0.0417502       0.0279461   0.0279461   0.0293174
 0.0145734    0.00633863   0.00633863      0.00424286  0.00424286  0.00184541
 0.0240829    0.0252646    0.0104747       0.0169112   0.00701141  0.00735544
 0.0240829    0.0104747    0.0252646    …  0.00701141  0.0169112   0.00735544
 0.0397975    0.0417502    0.0417502       0.0279461   0.0279461   0.0293174
 0.0240829    0.0104747    0.0104747       0.0169112   0.0169112   0.00735544
 0.0397975    0.0417502    0.0173097       0.0674048   0.0279461   0.0293174
 0.0397975    0.0173097    0.0417502       0.0279461   0.0674048   0.0293174
 0.0657662    0.0689931    0.0689931    …  0.111388    0.111388    0.116853

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 1.67126e-5   1.15611e-5   1.15611e-5   …  2.90107e-7  2.90107e-7  2.00685e-7
 4.39252e-5   3.62124e-5   3.03857e-5      9.08689e-7  7.62479e-7  6.28595e-7
 4.39252e-5   3.03857e-5   3.62124e-5      7.62479e-7  9.08689e-7  6.28595e-7
 0.000115447  9.51756e-5   9.51756e-5      2.38827e-6  2.38827e-6  1.96892e-6
 4.39252e-5   3.03857e-5   3.03857e-5      9.08689e-7  9.08689e-7  6.28595e-7
 0.000115447  9.51756e-5   7.98616e-5   …  2.84624e-6  2.38827e-6  1.96892e-6
 0.000115447  7.98616e-5   9.51756e-5      2.38827e-6  2.84624e-6  1.96892e-6
 0.000303425  0.000250147  0.000250147     7.48067e-6  7.48067e-6  6.16715e-6
 4.39252e-5   3.03857e-5   3.03857e-5      9.08689e-7  9.08689e-7  6.28595e-7
 0.000115447  9.51756e-5   7.98616e-5      2.84624e-6  2.38827e-6  1.96892e-6
 ⋮                                      ⋱                          ⋮
 0.690878     0.569567     0.569567        0.0693095   0.0693095   0.0571395
 0.100015     0.0691863    0.0691863       0.00841913  0.00841913  0.00582402
 0.262865     0.216709     0.18184         0.0263708   0.0221277   0.0182423
 0.262865     0.18184      0.216709     …  0.0221277   0.0263708   0.0182423
 0.690878     0.569567     0.569567        0.0693095   0.0693095   0.0571395
 0.262865     0.18184      0.18184         0.0263708   0.0263708   0.0182423
 0.690878     0.569567     0.477923        0.0826001   0.0693095   0.0571395
 0.690878     0.477923     0.569567        0.0693095   0.0826001   0.0571395
 1.81581      1.49697      1.49697      …  0.217095    0.217095    0.178975

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.68763    0.484909
 0.0955901  0.44297

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 0.0236345    0.0166668    0.0166668    …  0.0010193    0.000718798
 0.00328552   0.0152253    0.00231692      0.000141696  0.000656629
 0.00328552   0.00231692   0.0152253       0.000931138  0.000656629
 0.000456733  0.00211653   0.00211653      0.000129441  0.000599838
 0.00328552   0.00231692   0.00231692      0.000931138  0.000656629
 0.000456733  0.00211653   0.000322084  …  0.000129441  0.000599838
 0.000456733  0.000322084  0.00211653      0.000850605  0.000599838
 6.34922e-5   0.000294227  0.000294227     0.000118246  0.000547958
 0.00328552   0.00231692   0.00231692      0.000931138  0.000656629
 0.000456733  0.00211653   0.000322084     0.000129441  0.000599838
 ⋮                                      ⋱
 4.58215e-10  2.12339e-9   2.12339e-9   …  6.87176e-5   0.000318441
 2.37112e-8   1.67209e-8   1.67209e-8      0.000541123  0.000381595
 3.29618e-9   1.52747e-8   2.32443e-9      7.52236e-5   0.000348591
 3.29618e-9   2.32443e-9   1.52747e-8      0.000494322  0.000348591
 4.58215e-10  2.12339e-9   2.12339e-9      6.87176e-5   0.000318441
 3.29618e-9   2.32443e-9   2.32443e-9   …  0.000494322  0.000348591
 4.58215e-10  2.12339e-9   3.23128e-10     6.87176e-5   0.000318441
 4.58215e-10  3.23128e-10  2.12339e-9      0.000451568  0.000318441
 6.36982e-11  2.95181e-10  2.95181e-10     6.27743e-5   0.0002909

julia> sum(P10)  # expected number of edges
215.15530100199902

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 215 stored entries:
  [399 ,   12]  =  true
  [138 ,   16]  =  true
  [1   ,   22]  =  true
  [17  ,   25]  =  true
  [9   ,   27]  =  true
  [11  ,   28]  =  true
  [18  ,   32]  =  true
  [2   ,   38]  =  true
  [1   ,   48]  =  true
  ⋮
  [577 , 1009]  =  true
  [354 , 1012]  =  true
  [66  , 1015]  =  true
  [293 , 1015]  =  true
  [489 , 1019]  =  true
  [350 , 1021]  =  true
  [890 , 1022]  =  true
  [695 , 1023]  =  true
  [883 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source