Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.832508  0.892741
 0.624305  0.755468

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.69307   0.743214  0.743214  0.796987
 0.519739  0.628933  0.557343  0.674438
 0.519739  0.557343  0.628933  0.674438
 0.389757  0.471643  0.471643  0.570732

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 111.359   -131.594  -131.594   155.505
 -92.0251   122.715   108.747  -145.013
 -92.0251   108.747   122.715  -145.013
  76.0479  -101.409  -101.409   135.229

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.110831    0.11885     0.11885     …  0.238985   0.238985   0.256276
 0.0831131   0.100575    0.0891265      0.202237   0.179217   0.216869
 0.0831131   0.0891265   0.100575       0.179217   0.202237   0.216869
 0.0623273   0.0754219   0.0754219      0.151659   0.151659   0.183522
 0.0831131   0.0891265   0.0891265      0.202237   0.202237   0.216869
 0.0623273   0.0754219   0.0668368   …  0.17114    0.151659   0.183522
 0.0623273   0.0668368   0.0754219      0.151659   0.17114    0.183522
 0.0467398   0.0565596   0.0565596      0.128339   0.128339   0.155303
 0.0831131   0.0891265   0.0891265      0.202237   0.202237   0.216869
 0.0623273   0.0754219   0.0668368      0.17114    0.151659   0.183522
 ⋮                                   ⋱                        ⋮
 0.00467473  0.00565686  0.00565686     0.0337509  0.0337509  0.0408417
 0.00831264  0.00891407  0.00891407     0.0531846  0.0531846  0.0570326
 0.00623372  0.00754339  0.00668474     0.0450066  0.0398836  0.0482629
 0.00623372  0.00668474  0.00754339  …  0.0398836  0.0450066  0.0482629
 0.00467473  0.00565686  0.00565686     0.0337509  0.0337509  0.0408417
 0.00623372  0.00668474  0.00668474     0.0450066  0.0450066  0.0482629
 0.00467473  0.00565686  0.00501295     0.0380861  0.0337509  0.0408417
 0.00467473  0.00501295  0.00565686     0.0337509  0.0380861  0.0408417
 0.00350562  0.00424213  0.00424213  …  0.0285611  0.0285611  0.0345617

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 14.6097    16.5637   16.5637   18.779    …  58.1191   58.1191   65.8924
 11.5832    13.1803   13.1324   14.9431      46.2474   46.0794   52.4329
 11.5832    13.1324   13.1803   14.9431      46.0794   46.2474   52.4329
  9.18366   10.4499   10.4499   11.8908      36.667    36.667    41.7227
 11.5832    13.1324   13.1324   14.8888      46.2474   46.2474   52.4329
  9.18366   10.4499   10.412    11.8476   …  36.8007   36.667    41.7227
  9.18366   10.412    10.4499   11.8476      36.667    36.8007   41.7227
  7.28121    8.28516   8.28516   9.42753     29.1772   29.1772   33.2002
 11.5832    13.1324   13.1324   14.8888      46.2474   46.2474   52.4329
  9.18366   10.4499   10.412    11.8476      36.8007   36.667    41.7227
  ⋮                                       ⋱                       ⋮
  1.13685    1.29361   1.29361   1.47197      4.69021   4.69021   5.3369
  1.80855    2.05043   2.05043   2.32467      7.43423   7.43423   8.42854
  1.43389    1.6316    1.62567   1.84983      5.91568   5.89418   6.70688
  1.43389    1.62567   1.6316    1.84983  …   5.89418   5.91568   6.70688
  1.13685    1.29361   1.29361   1.47197      4.69021   4.69021   5.3369
  1.43389    1.62567   1.62567   1.8431       5.91568   5.91568   6.70688
  1.13685    1.29361   1.28891   1.46662      4.70731   4.69021   5.3369
  1.13685    1.28891   1.29361   1.46662      4.69021   4.70731   5.3369
  0.901348   1.02563   1.02563   1.16704  …   3.73216   3.73216   4.24676

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.741264  0.896948
 0.639032  0.839331

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.0500877  0.0606074  0.0606074  0.0733364  …  0.278534  0.278534  0.337033
 0.0431799  0.0567142  0.0522487  0.0686256     0.260642  0.24012   0.315383
 0.0431799  0.0522487  0.0567142  0.0686256     0.24012   0.260642  0.315383
 0.0372247  0.0488924  0.0488924  0.0642173     0.224695  0.224695  0.295124
 0.0431799  0.0522487  0.0522487  0.0632222     0.260642  0.260642  0.315383
 0.0372247  0.0488924  0.0450428  0.059161   …  0.243899  0.224695  0.295124
 0.0372247  0.0450428  0.0488924  0.059161      0.224695  0.243899  0.295124
 0.0320908  0.0421494  0.0421494  0.0553607     0.210262  0.210262  0.276166
 0.0431799  0.0522487  0.0522487  0.0632222     0.260642  0.260642  0.315383
 0.0372247  0.0488924  0.0450428  0.059161      0.243899  0.224695  0.295124
 ⋮                                           ⋱
 0.0131728  0.0173017  0.0173017  0.0227248  …  0.141175  0.141175  0.185425
 0.0177247  0.0214474  0.0214474  0.0259518     0.175002  0.175002  0.211756
 0.0152802  0.0200697  0.0184894  0.0242848     0.16376   0.150866  0.198154
 0.0152802  0.0184894  0.0200697  0.0242848     0.150866  0.16376   0.198154
 0.0131728  0.0173017  0.0173017  0.0227248     0.141175  0.141175  0.185425
 0.0152802  0.0184894  0.0184894  0.0223727  …  0.16376   0.16376   0.198154
 0.0131728  0.0173017  0.0159395  0.0209355     0.153241  0.141175  0.185425
 0.0131728  0.0159395  0.0173017  0.0209355     0.141175  0.153241  0.185425
 0.0113561  0.0149156  0.0149156  0.0195907     0.132107  0.132107  0.173514

julia> sum(P10)  # expected number of edges
86452.47042017106

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 82379 stored entries:
  [2   ,    1]  =  true
  [5   ,    1]  =  true
  [40  ,    1]  =  true
  [57  ,    1]  =  true
  [68  ,    1]  =  true
  [77  ,    1]  =  true
  [99  ,    1]  =  true
  [119 ,    1]  =  true
  [126 ,    1]  =  true
  ⋮
  [972 , 1024]  =  true
  [985 , 1024]  =  true
  [989 , 1024]  =  true
  [990 , 1024]  =  true
  [995 , 1024]  =  true
  [1005, 1024]  =  true
  [1018, 1024]  =  true
  [1020, 1024]  =  true
  [1021, 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source