Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.299236  0.39868
 0.479212  0.159989

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.0895422  0.1193     0.1193     0.158946
 0.143397   0.0478745  0.191052   0.0637845
 0.143397   0.191052   0.0478745  0.0637845
 0.229644   0.0766686  0.0766686  0.0255965

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  1.24862  -3.11146  -3.11146   7.75351
 -3.73995   2.33535   9.31967  -5.81952
 -3.73995   9.31967   2.33535  -5.81952
 11.2022   -6.99503  -6.99503   4.36793

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 5.15426e-7   6.86717e-7   6.86717e-7   …  1.21029e-5   1.6125e-5
 8.2543e-7    2.75577e-7   1.09974e-6      1.93821e-5   6.4709e-6
 8.2543e-7    1.09974e-6   2.75577e-7      4.85684e-6   6.4709e-6
 1.32189e-6   4.41323e-7   4.41323e-7      7.77799e-6   2.59675e-6
 8.2543e-7    1.09974e-6   1.09974e-6      4.85684e-6   6.4709e-6
 1.32189e-6   4.41323e-7   1.76118e-6   …  7.77799e-6   2.59675e-6
 1.32189e-6   1.76118e-6   4.41323e-7      1.94903e-6   2.59675e-6
 2.11693e-6   7.06757e-7   7.06757e-7      3.12128e-6   1.04207e-6
 8.2543e-7    1.09974e-6   1.09974e-6      4.85684e-6   6.4709e-6
 1.32189e-6   4.41323e-7   1.76118e-6      7.77799e-6   2.59675e-6
 ⋮                                      ⋱               ⋮
 9.15832e-5   3.05759e-5   3.05759e-5      2.09921e-9   7.0084e-10
 3.57099e-5   4.75773e-5   4.75773e-5      3.26646e-9   4.35199e-9
 5.71876e-5   1.90926e-5   7.61927e-5      5.23107e-9   1.74644e-9
 5.71876e-5   7.61927e-5   1.90926e-5   …  1.31082e-9   1.74644e-9
 9.15832e-5   3.05759e-5   3.05759e-5      2.09921e-9   7.0084e-10
 5.71876e-5   7.61927e-5   7.61927e-5      1.31082e-9   1.74644e-9
 9.15832e-5   3.05759e-5   0.000122019     2.09921e-9   7.0084e-10
 9.15832e-5   0.000122019  3.05759e-5      5.26027e-10  7.0084e-10
 0.000146666  4.89657e-5   4.89657e-5   …  8.42405e-10  2.81245e-10

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 2.38205e-7  1.55425e-7  1.55425e-7  …  2.17388e-9  2.17388e-9  1.41843e-9
 1.86821e-7  1.8392e-7   1.21898e-7     2.57242e-9  1.70494e-9  1.67847e-9
 1.86821e-7  1.21898e-7  1.8392e-7      1.70494e-9  2.57242e-9  1.67847e-9
 1.46521e-7  1.44245e-7  1.44245e-7     2.01751e-9  2.01751e-9  1.98618e-9
 1.86821e-7  1.21898e-7  1.21898e-7     2.57242e-9  2.57242e-9  1.67847e-9
 1.46521e-7  1.44245e-7  9.56026e-8  …  3.04402e-9  2.01751e-9  1.98618e-9
 1.46521e-7  9.56026e-8  1.44245e-7     2.01751e-9  3.04402e-9  1.98618e-9
 1.14914e-7  1.13129e-7  1.13129e-7     2.38738e-9  2.38738e-9  2.35031e-9
 1.86821e-7  1.21898e-7  1.21898e-7     2.57242e-9  2.57242e-9  1.67847e-9
 1.46521e-7  1.44245e-7  9.56026e-8     3.04402e-9  2.01751e-9  1.98618e-9
 ⋮                                   ⋱                          ⋮
 1.64499e-8  1.61945e-8  1.61945e-8     9.17842e-9  9.17842e-9  9.03589e-9
 2.67433e-8  1.74496e-8  1.74496e-8     9.88981e-9  9.88981e-9  6.45297e-9
 2.09744e-8  2.06487e-8  1.36855e-8     1.17029e-8  7.75643e-9  7.63599e-9
 2.09744e-8  1.36855e-8  2.06487e-8  …  7.75643e-9  1.17029e-8  7.63599e-9
 1.64499e-8  1.61945e-8  1.61945e-8     9.17842e-9  9.17842e-9  9.03589e-9
 2.09744e-8  1.36855e-8  1.36855e-8     1.17029e-8  1.17029e-8  7.63599e-9
 1.64499e-8  1.61945e-8  1.07333e-8     1.38484e-8  9.17842e-9  9.03589e-9
 1.64499e-8  1.07333e-8  1.61945e-8     9.17842e-9  1.38484e-8  9.03589e-9
 1.29014e-8  1.27011e-8  1.27011e-8  …  1.08611e-8  1.08611e-8  1.06924e-8

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.407012   0.958322
 0.0616288  0.590525

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 0.000124758  0.000293746  0.000293746  …  0.277466     0.653302
 1.88906e-5   0.000181009  4.44784e-5      0.0420132    0.402569
 1.88906e-5   4.44784e-5   0.000181009     0.170976     0.402569
 2.86037e-6   2.7408e-5    2.7408e-5       0.0258889    0.248066
 1.88906e-5   4.44784e-5   4.44784e-5      0.170976     0.402569
 2.86037e-6   2.7408e-5    6.73483e-6   …  0.0258889    0.248066
 2.86037e-6   6.73483e-6   2.7408e-5       0.105357     0.248066
 4.33111e-7   4.15005e-6   4.15005e-6      0.0159529    0.15286
 1.88906e-5   4.44784e-5   4.44784e-5      0.170976     0.402569
 2.86037e-6   2.7408e-5    6.73483e-6      0.0258889    0.248066
 ⋮                                      ⋱
 5.2199e-12   5.00169e-11  5.00169e-11  …  0.000873381  0.00836871
 2.27671e-10  5.36059e-10  5.36059e-10     0.00936051   0.0220396
 3.44735e-11  3.30324e-10  8.11689e-11     0.00141735   0.013581
 3.44735e-11  8.11689e-11  3.30324e-10     0.00576802   0.013581
 5.2199e-12   5.00169e-11  5.00169e-11     0.000873381  0.00836871
 3.44735e-11  8.11689e-11  8.11689e-11  …  0.00576802   0.013581
 5.2199e-12   5.00169e-11  1.22904e-11     0.000873381  0.00836871
 5.2199e-12   1.22904e-11  5.00169e-11     0.0035543    0.00836871
 7.90386e-13  7.57345e-12  7.57345e-12     0.000538184  0.00515686

julia> sum(P10)  # expected number of edges
1117.1428110010745

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1102 stored entries:
  [33  ,   14]  =  true
  [4   ,   56]  =  true
  [13  ,   80]  =  true
  [48  ,   88]  =  true
  [129 ,  115]  =  true
  [4   ,  116]  =  true
  [537 ,  116]  =  true
  [25  ,  118]  =  true
  [521 ,  124]  =  true
  ⋮
  [795 , 1024]  =  true
  [801 , 1024]  =  true
  [804 , 1024]  =  true
  [830 , 1024]  =  true
  [885 , 1024]  =  true
  [905 , 1024]  =  true
  [921 , 1024]  =  true
  [957 , 1024]  =  true
  [989 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source