Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.0243958  0.424959
 0.620303   0.565203
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.000595156  0.0103672  0.0103672  0.180591
 0.0151328    0.0137886  0.263603   0.240188
 0.0151328    0.263603   0.0137886  0.240188
 0.384775     0.350597   0.350597   0.319454
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  5.11884  -3.84871   -3.84871    2.89373
 -5.61786   0.220944   4.22391   -0.166122
 -5.61786   4.22391    0.220944  -0.166122
  6.16553  -0.242484  -0.242484   0.00953662
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 4.44413e-20  7.74138e-19  7.74138e-19  …  1.99131e-6   3.46873e-5
 1.12999e-18  1.02962e-18  1.96837e-17     5.06322e-5   4.61346e-5
 1.12999e-18  1.96837e-17  1.02962e-18     2.64847e-6   4.61346e-5
 2.87318e-17  2.61796e-17  2.61796e-17     6.73416e-5   6.13598e-5
 1.12999e-18  1.96837e-17  1.96837e-17     2.64847e-6   4.61346e-5
 2.87318e-17  2.61796e-17  5.00489e-16  …  6.73416e-5   6.13598e-5
 2.87318e-17  5.00489e-16  2.61796e-17     3.52251e-6   6.13598e-5
 7.30551e-16  6.65658e-16  6.65658e-16     8.95653e-5   8.16094e-5
 1.12999e-18  1.96837e-17  1.96837e-17     2.64847e-6   4.61346e-5
 2.87318e-17  2.61796e-17  5.00489e-16     6.73416e-5   6.13598e-5
 ⋮                                      ⋱               ⋮
 0.000127631  0.000116294  0.000116294     0.000876987  0.000799086
 1.97414e-7   3.43883e-6   3.43883e-6      2.59327e-5   0.000451731
 5.01958e-6   4.5737e-6    8.74378e-5      0.000659381  0.00060081
 5.01958e-6   8.74378e-5   4.5737e-6    …  3.44909e-5   0.00060081
 0.000127631  0.000116294  0.000116294     0.000876987  0.000799086
 5.01958e-6   8.74378e-5   8.74378e-5      3.44909e-5   0.00060081
 0.000127631  0.000116294  0.00222324      0.000876987  0.000799086
 0.000127631  0.00222324   0.000116294     4.58735e-5   0.000799086
 0.00324521   0.00295695   0.00295695   …  0.00116641   0.0010628
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.15656e-7  1.09683e-7  1.09683e-7  1.04019e-7  …  6.45466e-8   6.12134e-8
 1.60102e-7  2.55239e-7  1.51834e-7  2.42059e-7     8.93517e-8   1.42447e-7
 1.60102e-7  1.51834e-7  2.55239e-7  2.42059e-7     1.50204e-7   1.42447e-7
 2.21629e-7  3.53327e-7  3.53327e-7  5.63285e-7     2.07927e-7   3.31483e-7
 1.60102e-7  1.51834e-7  1.51834e-7  1.43994e-7     1.50204e-7   1.42447e-7
 2.21629e-7  3.53327e-7  2.10184e-7  3.35081e-7  …  2.07927e-7   3.31483e-7
 2.21629e-7  2.10184e-7  3.53327e-7  3.35081e-7     3.49533e-7   3.31483e-7
 3.068e-7    4.8911e-7   4.8911e-7   7.79755e-7     4.83858e-7   7.71381e-7
 1.60102e-7  1.51834e-7  1.51834e-7  1.43994e-7     1.50204e-7   1.42447e-7
 2.21629e-7  3.53327e-7  2.10184e-7  3.35081e-7     2.07927e-7   3.31483e-7
 ⋮                                               ⋱               ⋮
 4.13704e-6  6.59541e-6  6.59541e-6  1.05146e-5     0.000416081  0.000663329
 2.15889e-6  2.04741e-6  2.04741e-6  1.94168e-6     0.000129164  0.000122494
 2.98855e-6  4.76444e-6  2.83422e-6  4.51841e-6     0.000178801  0.00028505
 2.98855e-6  2.83422e-6  4.76444e-6  4.51841e-6  …  0.000300572  0.00028505
 4.13704e-6  6.59541e-6  6.59541e-6  1.05146e-5     0.000416081  0.000663329
 2.98855e-6  2.83422e-6  2.83422e-6  2.68786e-6     0.000300572  0.00028505
 4.13704e-6  6.59541e-6  3.92341e-6  6.25482e-6     0.000416081  0.000663329
 4.13704e-6  3.92341e-6  6.59541e-6  6.25482e-6     0.000699449  0.000663329
 5.7269e-6   9.13e-6     9.13e-6     1.45553e-5  …  0.000968245  0.00154361

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.689311  0.339063
 0.487053  0.456685
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.0242189    0.011913     0.011913     …  4.08261e-5   2.00818e-5
 0.0171126    0.0160456    0.00841746      2.88469e-5   2.70482e-5
 0.0171126    0.00841746   0.0160456       5.49888e-5   2.70482e-5
 0.0120914    0.0113375    0.0113375       3.88539e-5   3.64313e-5
 0.0171126    0.00841746   0.00841746      5.49888e-5   2.70482e-5
 0.0120914    0.0113375    0.0059476    …  3.88539e-5   3.64313e-5
 0.0120914    0.0059476    0.0113375       7.40645e-5   3.64313e-5
 0.00854355   0.00801085   0.00801085      5.23325e-5   4.90695e-5
 0.0171126    0.00841746   0.00841746      5.49888e-5   2.70482e-5
 0.0120914    0.0113375    0.0059476       3.88539e-5   3.64313e-5
 ⋮                                      ⋱
 0.00106318   0.000996886  0.000996886  …  0.000312457  0.000292975
 0.00212953   0.00104748   0.00104748      0.000328316  0.000161494
 0.00150468   0.00141086   0.000740132     0.000231982  0.000217517
 0.00150468   0.000740132  0.00141086      0.00044221   0.000217517
 0.00106318   0.000996886  0.000996886     0.000312457  0.000292975
 0.00150468   0.000740132  0.000740132  …  0.00044221   0.000217517
 0.00106318   0.000996886  0.000522962     0.000312457  0.000292975
 0.00106318   0.000522962  0.000996886     0.000595615  0.000292975
 0.000751219  0.000704379  0.000704379     0.000420849  0.000394609
julia> sum(P10)  # expected number of edges889.8509629241505
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 889 stored entries:
  [217,    1]  =  1
  [820,    1]  =  1
  [883,    1]  =  1
  [918,    1]  =  1
  [116,    2]  =  1
  [130,    2]  =  1
  [161,    3]  =  1
  [459,    3]  =  1
  [921,    3]  =  1
  ⋮
  [455,  999]  =  1
  [907,  999]  =  1
  [933, 1000]  =  1
  [812, 1001]  =  1
  [338, 1002]  =  1
  [744, 1003]  =  1
  [273, 1007]  =  1
  [913, 1009]  =  1
  [245, 1011]  =  1
  [151, 1015]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.