Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.519381  0.96278
 0.808014  0.53753

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.269757  0.50005   0.50005   0.926945
 0.419668  0.279183  0.77794   0.517523
 0.419668  0.77794   0.279183  0.517523
 0.652887  0.434332  0.434332  0.288939

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  1.16153  -2.08043  -2.08043   3.72629
 -1.746     1.12231   3.12729  -2.01019
 -1.746     3.12729   1.12231  -2.01019
  2.62459  -1.68705  -1.68705   1.08442

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.000385334  0.000714296  0.000714296  …  0.342203     0.634344
 0.000599474  0.000398799  0.00111125      0.532374     0.354161
 0.000599474  0.00111125   0.000398799     0.191056     0.354161
 0.000932615  0.000620421  0.000620421     0.29723      0.197732
 0.000599474  0.00111125   0.00111125      0.191056     0.354161
 0.000932615  0.000620421  0.00172879   …  0.29723      0.197732
 0.000932615  0.00172879   0.000620421     0.106668     0.197732
 0.00145089   0.000965204  0.000965204     0.165947     0.110396
 0.000599474  0.00111125   0.00111125      0.191056     0.354161
 0.000932615  0.000620421  0.00172879      0.29723      0.197732
 ⋮                                      ⋱               ⋮
 0.0497846    0.0331191    0.0331191       0.00156666   0.00104222
 0.0205698    0.0381303    0.0381303       0.00180371   0.00334355
 0.0320009    0.0212886    0.0593203       0.00280608   0.00186674
 0.0320009    0.0593203    0.0212886    …  0.00100703   0.00186674
 0.0497846    0.0331191    0.0331191       0.00156666   0.00104222
 0.0320009    0.0593203    0.0593203       0.00100703   0.00186674
 0.0497846    0.0331191    0.092286        0.00156666   0.00104222
 0.0497846    0.092286     0.0331191       0.000562237  0.00104222
 0.0774511    0.0515242    0.0515242    …  0.000874685  0.000581883

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.74915   1.69886   1.69886   …  1.30656   1.26899  1.26899  1.23251
 1.42577   1.78118   1.38478      1.065     1.33048  1.03438  1.29223
 1.42577   1.38478   1.78118      1.065     1.03438  1.33048  1.29223
 1.16218   1.45188   1.45188      0.868107  1.0845   1.0845   1.35484
 1.42577   1.38478   1.38478      1.36987   1.33048  1.33048  1.29223
 1.16218   1.45188   1.12876   …  1.11661   1.39495  1.0845   1.35484
 1.16218   1.12876   1.45188      1.11661   1.0845   1.39495  1.35484
 0.947315  1.18345   1.18345      0.91017   1.13705  1.13705  1.42049
 1.42577   1.38478   1.38478      1.36987   1.33048  1.33048  1.29223
 1.16218   1.45188   1.12876      1.11661   1.39495  1.0845   1.35484
 ⋮                             ⋱                              ⋮
 0.184617  0.230636  0.230636     1.32897   1.66025  1.66025  2.07411
 0.27786   0.269871  0.269871     2.0002    1.94269  1.94269  1.88683
 0.22649   0.282947  0.219978     1.6304    2.03682  1.58352  1.97825
 0.22649   0.219978  0.282947  …  1.6304    1.58352  2.03682  1.97825
 0.184617  0.230636  0.230636     1.32897   1.66025  1.66025  2.07411
 0.22649   0.219978  0.219978     2.09711   2.03682  2.03682  1.97825
 0.184617  0.230636  0.179308     1.7094    2.13551  1.66025  2.07411
 0.184617  0.179308  0.230636     1.7094    1.66025  2.13551  2.07411
 0.150485  0.187997  0.187997  …  1.39337   1.7407   1.7407   2.17461

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.914416  0.876297
 0.278724  0.565771

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.40873     0.391691    0.391691    …  0.278614    0.278614    0.267
 0.124585    0.252891    0.119392       0.179884    0.0849248   0.172386
 0.124585    0.119392    0.252891       0.0849248   0.179884    0.172386
 0.037975    0.0770841   0.0770841      0.0548308   0.0548308   0.111299
 0.124585    0.119392    0.119392       0.179884    0.179884    0.172386
 0.037975    0.0770841   0.036392    …  0.11614     0.0548308   0.111299
 0.037975    0.036392    0.0770841      0.0548308   0.11614     0.111299
 0.0115752   0.0234961   0.0234961      0.0354009   0.0354009   0.071859
 0.124585    0.119392    0.119392       0.179884    0.179884    0.172386
 0.037975    0.0770841   0.036392       0.11614     0.0548308   0.111299
 ⋮                                   ⋱
 9.28358e-6  1.88444e-5  1.88444e-5  …  0.00256421  0.00256421  0.005205
 9.99202e-5  9.57549e-5  9.57549e-5     0.0130297   0.0130297   0.0124865
 3.04568e-5  6.18231e-5  2.91872e-5     0.00841246  0.00397159  0.00806178
 3.04568e-5  2.91872e-5  6.18231e-5     0.00397159  0.00841246  0.00806178
 9.28358e-6  1.88444e-5  1.88444e-5     0.00256421  0.00256421  0.005205
 3.04568e-5  2.91872e-5  2.91872e-5  …  0.00841246  0.00841246  0.00806178
 9.28358e-6  1.88444e-5  8.89658e-6     0.00543142  0.00256421  0.005205
 9.28358e-6  8.89658e-6  1.88444e-5     0.00256421  0.00543142  0.005205
 2.82974e-6  5.74398e-6  5.74398e-6     0.00165556  0.00165556  0.00336055

julia> sum(P10)  # expected number of edges
16149.11547586631

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 15705 stored entries:
  [3   ,    1]  =  true
  [26  ,    1]  =  true
  [69  ,    1]  =  true
  [385 ,    1]  =  true
  [769 ,    1]  =  true
  [942 ,    1]  =  true
  [1   ,    2]  =  true
  [33  ,    2]  =  true
  [132 ,    2]  =  true
  ⋮
  [728 , 1024]  =  true
  [773 , 1024]  =  true
  [777 , 1024]  =  true
  [835 , 1024]  =  true
  [878 , 1024]  =  true
  [894 , 1024]  =  true
  [915 , 1024]  =  true
  [982 , 1024]  =  true
  [993 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source