Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.316959  0.0823522
 0.515035  0.239575

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.100463  0.0261022  0.0261022  0.00678188
 0.163245  0.0759355  0.0424143  0.0197296
 0.163245  0.0424143  0.0759355  0.0197296
 0.265261  0.12339    0.12339    0.0573964

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
   51.0793   -17.5581   -17.5581    6.03546
 -109.809     67.578     37.7461  -23.2294
 -109.809     37.7461    67.578   -23.2294
  236.066   -145.278   -145.278    89.4058

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 1.02809e-6   2.67119e-7  2.67119e-7  6.94028e-8  …  3.74479e-13  9.72971e-14
 1.67058e-6   7.77091e-7  4.34049e-7  2.01904e-7     6.08502e-13  2.83053e-13
 1.67058e-6   4.34049e-7  7.77091e-7  2.01904e-7     1.08942e-12  2.83053e-13
 2.71457e-6   1.26272e-6  1.26272e-6  5.87369e-7     1.77023e-12  8.23445e-13
 1.67058e-6   4.34049e-7  4.34049e-7  1.12775e-7     1.08942e-12  2.83053e-13
 2.71457e-6   1.26272e-6  7.05299e-7  3.28079e-7  …  1.77023e-12  8.23445e-13
 2.71457e-6   7.05299e-7  1.26272e-6  3.28079e-7     3.16929e-12  8.23445e-13
 4.41098e-6   2.05183e-6  2.05183e-6  9.54434e-7     5.14987e-12  2.39553e-12
 1.67058e-6   4.34049e-7  4.34049e-7  1.12775e-7     1.08942e-12  2.83053e-13
 2.71457e-6   1.26272e-6  7.05299e-7  3.28079e-7     1.77023e-12  8.23445e-13
 ⋮                                                ⋱               ⋮
 0.000214393  9.97277e-5  9.97277e-5  4.63896e-5     2.64201e-8   1.22896e-8
 8.11974e-5   2.10967e-5  2.10967e-5  5.48134e-6     5.58898e-9   1.45213e-9
 0.00013194   6.13736e-5  3.42806e-5  1.59461e-5     9.0817e-9    4.22447e-9
 0.00013194   3.42806e-5  6.13736e-5  1.59461e-5  …  1.62592e-8   4.22447e-9
 0.000214393  9.97277e-5  9.97277e-5  4.63896e-5     2.64201e-8   1.22896e-8
 0.00013194   3.42806e-5  3.42806e-5  8.90678e-6     1.62592e-8   4.22447e-9
 0.000214393  9.97277e-5  5.57036e-5  2.59112e-5     2.64201e-8   1.22896e-8
 0.000214393  5.57036e-5  9.97277e-5  2.59112e-5     4.73006e-8   1.22896e-8
 0.000348374  0.00016205  0.00016205  7.53799e-5  …  7.68602e-8   3.57525e-8

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 7.23689e-11  2.32143e-11  2.32143e-11  …  2.67792e-16  8.59016e-17
 1.45184e-10  5.05552e-11  4.65717e-11     5.37234e-16  1.87073e-16
 1.45184e-10  4.65717e-11  5.05552e-11     5.83186e-16  1.87073e-16
 2.91262e-10  1.01422e-10  1.01422e-10     1.16997e-15  4.07401e-16
 1.45184e-10  4.65717e-11  4.65717e-11     5.83186e-16  1.87073e-16
 2.91262e-10  1.01422e-10  9.34304e-11  …  1.16997e-15  4.07401e-16
 2.91262e-10  9.34304e-11  1.01422e-10     1.27004e-15  4.07401e-16
 5.8432e-10   2.03469e-10  2.03469e-10     2.54791e-15  8.87221e-16
 1.45184e-10  4.65717e-11  4.65717e-11     5.83186e-16  1.87073e-16
 2.91262e-10  1.01422e-10  9.34304e-11     1.16997e-15  4.07401e-16
 ⋮                                      ⋱               ⋮
 1.53313e-7   5.33861e-8   5.33861e-8      1.28903e-12  4.48862e-13
 3.80932e-8   1.22195e-8   1.22195e-8      2.95045e-13  9.46438e-14
 7.64212e-8   2.66111e-8   2.45142e-8      5.91908e-13  2.06112e-13
 7.64212e-8   2.45142e-8   2.66111e-8   …  6.42537e-13  2.06112e-13
 1.53313e-7   5.33861e-8   5.33861e-8      1.28903e-12  4.48862e-13
 7.64212e-8   2.45142e-8   2.45142e-8      6.42537e-13  2.06112e-13
 1.53313e-7   5.33861e-8   4.91795e-8      1.28903e-12  4.48862e-13
 1.53313e-7   4.91795e-8   5.33861e-8      1.39929e-12  4.48862e-13
 3.07572e-7   1.07101e-7   1.07101e-7   …  2.80721e-12  9.77513e-13

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.335511  0.824943
 0.324853  0.615745

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 1.80747e-5  4.44414e-5  4.44414e-5  …  0.0593641   0.0593641   0.145962
 1.75005e-5  3.31715e-5  4.30297e-5     0.0443099   0.0574784   0.108948
 1.75005e-5  4.30297e-5  3.31715e-5     0.0574784   0.0443099   0.108948
 1.69446e-5  3.21177e-5  3.21177e-5     0.0429024   0.0429024   0.0813196
 1.75005e-5  4.30297e-5  4.30297e-5     0.0443099   0.0443099   0.108948
 1.69446e-5  3.21177e-5  4.16628e-5  …  0.0330733   0.0429024   0.0813196
 1.69446e-5  4.16628e-5  3.21177e-5     0.0429024   0.0330733   0.0813196
 1.64063e-5  3.10975e-5  3.10975e-5     0.0320227   0.0320227   0.0606977
 1.75005e-5  4.30297e-5  4.30297e-5     0.0443099   0.0443099   0.108948
 1.69446e-5  3.21177e-5  4.16628e-5     0.0330733   0.0429024   0.0813196
 ⋮                                   ⋱
 1.35174e-5  2.56216e-5  2.56216e-5  …  0.00553758  0.00553758  0.0104962
 1.44189e-5  3.54527e-5  3.54527e-5     0.00766236  0.00766236  0.0188399
 1.39609e-5  2.64622e-5  3.43265e-5     0.00571925  0.00741896  0.0140623
 1.39609e-5  3.43265e-5  2.64622e-5     0.00741896  0.00571925  0.0140623
 1.35174e-5  2.56216e-5  2.56216e-5     0.00553758  0.00553758  0.0104962
 1.39609e-5  3.43265e-5  3.43265e-5  …  0.00571925  0.00571925  0.0140623
 1.35174e-5  2.56216e-5  3.32361e-5     0.0042689   0.00553758  0.0104962
 1.35174e-5  3.32361e-5  2.56216e-5     0.00553758  0.0042689   0.0104962
 1.3088e-5   2.48077e-5  2.48077e-5     0.0041333   0.0041333   0.00783448

julia> sum(P10)  # expected number of edges
1676.3716485717277

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1671 stored entries:
  [525 ,   30]  =  true
  [266 ,   31]  =  true
  [167 ,   48]  =  true
  [563 ,   55]  =  true
  [911 ,   60]  =  true
  [898 ,   61]  =  true
  [194 ,   62]  =  true
  [269 ,   62]  =  true
  [67  ,   63]  =  true
  ⋮
  [770 , 1024]  =  true
  [795 , 1024]  =  true
  [864 , 1024]  =  true
  [867 , 1024]  =  true
  [875 , 1024]  =  true
  [898 , 1024]  =  true
  [900 , 1024]  =  true
  [961 , 1024]  =  true
  [980 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source