Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.408703  0.273932
 0.884337  0.26335
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.167038  0.111957  0.111957  0.0750386
 0.361431  0.107632  0.242248  0.0721399
 0.361431  0.242248  0.107632  0.0721399
 0.782052  0.23289   0.23289   0.0693532
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
   3.82712   -3.9809    -3.9809    4.14086
 -12.8516     5.93945   13.368    -6.17811
 -12.8516    13.368      5.93945  -6.17811
  43.156    -19.9448   -19.9448    9.21766
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.17216e-5   1.45588e-5  1.45588e-5  …  2.66364e-7  2.66364e-7  1.7853e-7
 4.70004e-5   1.39964e-5  3.15019e-5     2.56074e-7  5.76349e-7  1.71633e-7
 4.70004e-5   3.15019e-5  1.39964e-5     5.76349e-7  2.56074e-7  1.71633e-7
 0.000101698  3.0285e-5   3.0285e-5      5.54084e-7  5.54084e-7  1.65003e-7
 4.70004e-5   3.15019e-5  3.15019e-5     2.56074e-7  2.56074e-7  1.71633e-7
 0.000101698  3.0285e-5   6.81627e-5  …  2.46182e-7  5.54084e-7  1.65003e-7
 0.000101698  6.81627e-5  3.0285e-5      5.54084e-7  2.46182e-7  1.65003e-7
 0.00022005   6.55295e-5  6.55295e-5     5.3268e-7   5.3268e-7   1.58629e-7
 4.70004e-5   3.15019e-5  3.15019e-5     2.56074e-7  2.56074e-7  1.71633e-7
 0.000101698  3.0285e-5   6.81627e-5     2.46182e-7  5.54084e-7  1.65003e-7
 ⋮                                    ⋱                          ⋮
 0.105731     0.0314861   0.0314861      3.88679e-7  3.88679e-7  1.15746e-7
 0.022583     0.0151362   0.0151362      1.86849e-7  1.86849e-7  1.25235e-7
 0.0488644    0.0145515   0.0327512      1.79631e-7  4.04297e-7  1.20397e-7
 0.0488644    0.0327512   0.0145515   …  4.04297e-7  1.79631e-7  1.20397e-7
 0.105731     0.0314861   0.0314861      3.88679e-7  3.88679e-7  1.15746e-7
 0.0488644    0.0327512   0.0327512      1.79631e-7  1.79631e-7  1.20397e-7
 0.105731     0.0314861   0.070866       1.72692e-7  3.88679e-7  1.15746e-7
 0.105731     0.070866    0.0314861      3.88679e-7  1.72692e-7  1.15746e-7
 0.228777     0.0681284   0.0681284   …  3.73665e-7  3.73665e-7  1.11275e-7
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.20965e-5   9.93898e-6   9.93898e-6   …  3.36923e-9  3.36923e-9  1.51548e-9
 3.20861e-5   1.68227e-5   1.44323e-5      5.70275e-9  4.89244e-9  2.56509e-9
 3.20861e-5   1.44323e-5   1.68227e-5      4.89244e-9  5.70275e-9  2.56509e-9
 4.6592e-5    2.44281e-5   2.44281e-5      8.28092e-9  8.28092e-9  4.34167e-9
 3.20861e-5   1.44323e-5   1.44323e-5      5.70275e-9  5.70275e-9  2.56509e-9
 4.6592e-5    2.44281e-5   2.0957e-5    …  9.65245e-9  8.28092e-9  4.34167e-9
 4.6592e-5    2.0957e-5    2.44281e-5      8.28092e-9  9.65245e-9  4.34167e-9
 6.76558e-5   3.54718e-5   3.54718e-5      1.40162e-8  1.40162e-8  7.34869e-9
 3.20861e-5   1.44323e-5   1.44323e-5      5.70275e-9  5.70275e-9  2.56509e-9
 4.6592e-5    2.44281e-5   2.0957e-5       9.65245e-9  8.28092e-9  4.34167e-9
 ⋮                                      ⋱                          ⋮
 0.00133739   0.000701192  0.000701192     9.44187e-7  9.44187e-7  4.95035e-7
 0.000634264  0.000285292  0.000285292     3.84159e-7  3.84159e-7  1.72794e-7
 0.00092101   0.000482884  0.00041427      6.50225e-7  5.57834e-7  2.92471e-7
 0.00092101   0.00041427   0.000482884  …  5.57834e-7  6.50225e-7  2.92471e-7
 0.00133739   0.000701192  0.000701192     9.44187e-7  9.44187e-7  4.95035e-7
 0.00092101   0.00041427   0.00041427      6.50225e-7  6.50225e-7  2.92471e-7
 0.00133739   0.000701192  0.000601558     1.10057e-6  9.44187e-7  4.95035e-7
 0.00133739   0.000601558  0.000701192     9.44187e-7  1.10057e-6  4.95035e-7
 0.00194202   0.00101819   0.00101819   …  1.59813e-6  1.59813e-6  8.37894e-7

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.540891  0.583911
 0.945979  0.925297
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.00214337  0.00231385  0.00231385  …  0.00426808  0.00426808  0.00460754
 0.0037486   0.00366665  0.00404675     0.00676343  0.00746456  0.00730136
 0.0037486   0.00404675  0.00366665     0.00746456  0.00676343  0.00730136
 0.00655603  0.0064127   0.0064127      0.0118287   0.0118287   0.0115701
 0.0037486   0.00404675  0.00404675     0.00676343  0.00676343  0.00730136
 0.00655603  0.0064127   0.00707747  …  0.0107177   0.0118287   0.0115701
 0.00655603  0.00707747  0.0064127      0.0118287   0.0107177   0.0115701
 0.011466    0.0112153   0.0112153      0.0187445   0.0187445   0.0183347
 0.0037486   0.00404675  0.00404675     0.00676343  0.00676343  0.00730136
 0.00655603  0.0064127   0.00707747     0.0107177   0.0118287   0.0115701
 ⋮                                   ⋱
 0.328127    0.320953    0.320953    …  0.29681     0.29681     0.290321
 0.107275    0.115807    0.115807       0.107096    0.107096    0.115614
 0.187616    0.183515    0.202539       0.16971     0.187303    0.183208
 0.187616    0.202539    0.183515       0.187303    0.16971     0.183208
 0.328127    0.320953    0.320953       0.29681     0.29681     0.290321
 0.187616    0.202539    0.202539    …  0.16971     0.16971     0.183208
 0.328127    0.320953    0.354225       0.268932    0.29681     0.290321
 0.328127    0.354225    0.320953       0.29681     0.268932    0.290321
 0.57387     0.561324    0.561324       0.470342    0.470342    0.460059
julia> sum(P10)  # expected number of edges58281.6542331809
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 55439 stored entries:
  [248 ,    1]  =  1
  [279 ,    1]  =  1
  [304 ,    1]  =  1
  [344 ,    1]  =  1
  [358 ,    1]  =  1
  [364 ,    1]  =  1
  [379 ,    1]  =  1
  [412 ,    1]  =  1
  [492 ,    1]  =  1
  ⋮
  [984 , 1024]  =  1
  [990 , 1024]  =  1
  [992 , 1024]  =  1
  [1003, 1024]  =  1
  [1005, 1024]  =  1
  [1008, 1024]  =  1
  [1009, 1024]  =  1
  [1011, 1024]  =  1
  [1016, 1024]  =  1
  [1022, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.