Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.275292  0.502094
 0.940133  0.57188

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.0757856  0.138222  0.138222  0.252098
 0.258811   0.157434  0.472035  0.287137
 0.258811   0.472035  0.157434  0.287137
 0.883851   0.537643  0.537643  0.327047

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  3.30437  -2.90114  -2.90114   2.54712
 -5.43217   1.59066   4.76929  -1.39655
 -5.43217   4.76929   1.59066  -1.39655
  8.93013  -2.61494  -2.61494   0.765712

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.8946e-7   3.45549e-7  3.45549e-7  6.30234e-7  …  0.000140743  0.000256695
 6.47016e-7  3.93577e-7  1.18007e-6  7.1783e-7      0.000480643  0.000292374
 6.47016e-7  1.18007e-6  3.93577e-7  7.1783e-7      0.000160305  0.000292374
 2.20959e-6  1.34408e-6  1.34408e-6  8.17602e-7     0.000547448  0.000333011
 6.47016e-7  1.18007e-6  1.18007e-6  2.15228e-6     0.000160305  0.000292374
 2.20959e-6  1.34408e-6  4.02998e-6  2.45142e-6  …  0.000547448  0.000333011
 2.20959e-6  4.02998e-6  1.34408e-6  2.45142e-6     0.000182586  0.000333011
 7.54583e-6  4.59011e-6  4.59011e-6  2.79215e-6     0.000623538  0.000379296
 6.47016e-7  1.18007e-6  1.18007e-6  2.15228e-6     0.000160305  0.000292374
 2.20959e-6  1.34408e-6  4.02998e-6  2.45142e-6     0.000547448  0.000333011
 ⋮                                               ⋱               ⋮
 0.139597    0.0849166   0.0849166   0.0516545      0.00176613   0.00107433
 0.0119697   0.0218311   0.0218311   0.039817       0.000454053  0.000828129
 0.0408772   0.0248655   0.0745543   0.0453511      0.00155061   0.000943231
 0.0408772   0.0745543   0.0248655   0.0453511   …  0.000517162  0.000943231
 0.139597    0.0849166   0.0849166   0.0516545      0.00176613   0.00107433
 0.0408772   0.0745543   0.0745543   0.135977       0.000517162  0.000943231
 0.139597    0.0849166   0.254606    0.154876       0.00176613   0.00107433
 0.139597    0.254606    0.0849166   0.154876       0.000589042  0.00107433
 0.476731    0.289994    0.289994    0.176402    …  0.00201161   0.00122365

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.000730569  0.000567256  0.000567256  …  4.51812e-5   3.50813e-5
 0.00106214   0.00106565   0.00082471      6.56871e-5   6.59037e-5
 0.00106214   0.00082471   0.00106565      8.48774e-5   6.59037e-5
 0.00154421   0.0015493    0.0015493       0.0001234    0.000123807
 0.00106214   0.00082471   0.00082471      8.48774e-5   6.59037e-5
 0.00154421   0.0015493    0.00119901   …  0.0001234    0.000123807
 0.00154421   0.00119901   0.0015493       0.000159451  0.000123807
 0.00224506   0.00225247   0.00225247      0.000231819  0.000232584
 0.00106214   0.00082471   0.00082471      8.48774e-5   6.59037e-5
 0.00154421   0.0015493    0.00119901      0.0001234    0.000123807
 ⋮                                      ⋱               ⋮
 0.0448134    0.0449613    0.0449613       0.0359608    0.0360794
 0.0212013    0.0164619    0.0164619       0.0131665    0.0102233
 0.0308238    0.0309255    0.0239334       0.0191423    0.0192055
 0.0308238    0.0239334    0.0309255    …  0.0247347    0.0192055
 0.0448134    0.0449613    0.0449613       0.0359608    0.0360794
 0.0308238    0.0239334    0.0239334       0.0247347    0.0192055
 0.0448134    0.0449613    0.0347957       0.0359608    0.0360794
 0.0448134    0.0347957    0.0449613       0.0464667    0.0360794
 0.0651525    0.0653674    0.0653674    …  0.0675561    0.0677789

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.960966  0.220688
 0.25439   0.174321

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.671555    0.154224    0.154224    …  1.19322e-6  1.19322e-6  2.74025e-7
 0.177776    0.121821    0.0408268      9.4252e-7   3.15873e-7  2.16452e-7
 0.177776    0.0408268   0.121821       3.15873e-7  9.4252e-7   2.16452e-7
 0.0470616   0.032249    0.032249       2.49507e-7  2.49507e-7  1.70975e-7
 0.177776    0.0408268   0.0408268      9.4252e-7   9.4252e-7   2.16452e-7
 0.0470616   0.032249    0.0108078   …  7.44494e-7  2.49507e-7  1.70975e-7
 0.0470616   0.0108078   0.032249       2.49507e-7  7.44494e-7  1.70975e-7
 0.0124583   0.00853706  0.00853706     1.97085e-7  1.97085e-7  1.35052e-7
 0.177776    0.0408268   0.0408268      9.4252e-7   9.4252e-7   2.16452e-7
 0.0470616   0.032249    0.0108078      7.44494e-7  2.49507e-7  1.70975e-7
 ⋮                                   ⋱
 4.2876e-6   2.93808e-6  2.93808e-6  …  4.78715e-8  4.78715e-8  3.28039e-8
 6.11828e-5  1.40508e-5  1.40508e-5     2.28936e-7  2.28936e-7  5.25756e-8
 1.61965e-5  1.10987e-5  3.71957e-6     1.80836e-7  6.06047e-8  4.15293e-8
 1.61965e-5  3.71957e-6  1.10987e-5     6.06047e-8  1.80836e-7  4.15293e-8
 4.2876e-6   2.93808e-6  2.93808e-6     4.78715e-8  4.78715e-8  3.28039e-8
 1.61965e-5  3.71957e-6  3.71957e-6  …  1.80836e-7  1.80836e-7  4.15293e-8
 4.2876e-6   2.93808e-6  9.84658e-7     1.42842e-7  4.78715e-8  3.28039e-8
 4.2876e-6   9.84658e-7  2.93808e-6     4.78715e-8  1.42842e-7  3.28039e-8
 1.13503e-6  7.77778e-7  7.77778e-7     3.78135e-8  3.78135e-8  2.59117e-8

julia> sum(P10)  # expected number of edges
117.28581759788648

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 117 stored entries:
  [3  ,   1]  =  1
  [41 ,   1]  =  1
  [129,   1]  =  1
  [153,   1]  =  1
  [165,   1]  =  1
  [647,   1]  =  1
  [282,   2]  =  1
  [9  ,   5]  =  1
  [257,   5]  =  1
  ⋮
  [145, 585]  =  1
  [129, 645]  =  1
  [531, 649]  =  1
  [29 , 666]  =  1
  [274, 681]  =  1
  [904, 741]  =  1
  [29 , 781]  =  1
  [513, 801]  =  1
  [537, 861]  =  1
  [771, 865]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.