Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.047152  0.181722
 0.341249  0.876397
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.00222331  0.00856853  0.00856853  0.0330227
 0.0160906   0.0413239   0.0620123   0.15926
 0.0160906   0.0620123   0.0413239   0.15926
 0.116451    0.29907     0.29907     0.768073
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
 1794.51   -372.092   -372.092    77.1536
 -698.741    96.5482   144.884   -20.0193
 -698.741   144.884     96.5482  -20.0193
  272.074   -37.5937   -37.5937    5.19449
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.20781e-16  4.65484e-16  4.65484e-16  …  3.36489e-10  1.29681e-9
 8.74117e-16  2.24491e-15  3.36881e-15     2.43524e-9   6.2542e-9
 8.74117e-16  3.36881e-15  2.24491e-15     1.6228e-9    6.2542e-9
 6.32618e-15  1.62469e-14  1.62469e-14     1.17446e-8   3.01624e-8
 8.74117e-16  3.36881e-15  3.36881e-15     1.6228e-9    6.2542e-9
 6.32618e-15  1.62469e-14  2.43808e-14  …  1.17446e-8   3.01624e-8
 6.32618e-15  2.43808e-14  1.62469e-14     7.82636e-9   3.01624e-8
 4.5784e-14   1.17583e-13  1.17583e-13     5.66411e-8   1.45466e-7
 8.74117e-16  3.36881e-15  3.36881e-15     1.6228e-9    6.2542e-9
 6.32618e-15  1.62469e-14  2.43808e-14     1.17446e-8   3.01624e-8
 ⋮                                      ⋱               ⋮
 3.44579e-7   8.84948e-7   8.84948e-7      0.0165764    0.0425715
 6.57877e-9   2.53543e-8   2.53543e-8      0.000474922  0.00183033
 4.7612e-8    1.22277e-7   1.83495e-7      0.00343712   0.00882722
 4.7612e-8    1.83495e-7   1.22277e-7   …  0.00229043   0.00882722
 3.44579e-7   8.84948e-7   8.84948e-7      0.0165764    0.0425715
 4.7612e-8    1.83495e-7   1.83495e-7      0.00229043   0.00882722
 3.44579e-7   8.84948e-7   1.32799e-6      0.0165764    0.0425715
 3.44579e-7   1.32799e-6   8.84948e-7      0.0110462    0.0425715
 2.49379e-6   6.40457e-6   6.40457e-6   …  0.0799437    0.205312
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 4.9353e-15   1.28945e-14  1.28945e-14  …  1.91121e-10  4.99343e-10
 2.42141e-14  6.37763e-14  6.32644e-14     9.377e-10    2.46976e-9
 2.42141e-14  6.32644e-14  6.37763e-14     9.45288e-10  2.46976e-9
 1.18802e-13  3.12907e-13  3.12907e-13     4.63789e-9   1.22155e-8
 2.42141e-14  6.32644e-14  6.32644e-14     9.45288e-10  2.46976e-9
 1.18802e-13  3.12907e-13  3.10395e-13  …  4.63789e-9   1.22155e-8
 1.18802e-13  3.10395e-13  3.12907e-13     4.67542e-9   1.22155e-8
 5.82882e-13  1.53522e-12  1.53522e-12     2.29391e-8   6.0418e-8
 2.42141e-14  6.32644e-14  6.32644e-14     9.45288e-10  2.46976e-9
 1.18802e-13  3.12907e-13  3.10395e-13     4.63789e-9   1.22155e-8
 ⋮                                      ⋱               ⋮
 1.95716e-7   5.15486e-7   5.15486e-7      0.0082153    0.0216378
 8.13046e-9   2.12425e-8   2.12425e-8      0.000338541  0.000884509
 3.98906e-8   1.05066e-7   1.04222e-7      0.00166099   0.0043748
 3.98906e-8   1.04222e-7   1.05066e-7   …  0.00167443   0.0043748
 1.95716e-7   5.15486e-7   5.15486e-7      0.0082153    0.0216378
 3.98906e-8   1.04222e-7   1.04222e-7      0.00167443   0.0043748
 1.95716e-7   5.15486e-7   5.11348e-7      0.0082153    0.0216378
 1.95716e-7   5.11348e-7   5.15486e-7      0.00828178   0.0216378
 9.60245e-7   2.52914e-6   2.52914e-6   …  0.040633     0.107021

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.125453  0.936465
 0.60341   0.545391
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 9.65649e-10  7.20824e-9   7.20824e-9   …  0.0694876    0.518701
 4.64462e-9   4.19803e-9   3.46705e-8      0.334224     0.302088
 4.64462e-9   3.46705e-8   4.19803e-9      0.0404691    0.302088
 2.23399e-8   2.01918e-8   2.01918e-8      0.19465      0.175934
 4.64462e-9   3.46705e-8   3.46705e-8      0.0404691    0.302088
 2.23399e-8   2.01918e-8   1.66759e-7   …  0.19465      0.175934
 2.23399e-8   1.66759e-7   2.01918e-8      0.0235689    0.175934
 1.07451e-7   9.71195e-8   9.71195e-8      0.113363     0.102463
 4.64462e-9   3.46705e-8   3.46705e-8      0.0404691    0.302088
 2.23399e-8   2.01918e-8   1.66759e-7      0.19465      0.175934
 ⋮                                      ⋱
 0.00133044   0.00120251   0.00120251   …  0.00442352   0.00399819
 5.75086e-5   0.000429282  0.000429282     0.00157914   0.0117878
 0.000276607  0.000250011  0.00206478      0.00759542   0.00686511
 0.000276607  0.00206478   0.000250011     0.000919682  0.00686511
 0.00133044   0.00120251   0.00120251      0.00442352   0.00399819
 0.000276607  0.00206478   0.00206478   …  0.000919682  0.00686511
 0.00133044   0.00120251   0.00993125      0.00442352   0.00399819
 0.00133044   0.00993125   0.00120251      0.000535617  0.00399819
 0.00639919   0.00578389   0.00578389      0.00257623   0.00232852
julia> sum(P10)  # expected number of edges2788.281927431696
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 2760 stored entries:
  [1018,    7]  =  1
  [757 ,   16]  =  1
  [1011,   16]  =  1
  [1014,   20]  =  1
  [742 ,   28]  =  1
  [464 ,   30]  =  1
  [990 ,   36]  =  1
  [840 ,   41]  =  1
  [479 ,   42]  =  1
  ⋮
  [657 , 1024]  =  1
  [659 , 1024]  =  1
  [769 , 1024]  =  1
  [776 , 1024]  =  1
  [777 , 1024]  =  1
  [789 , 1024]  =  1
  [793 , 1024]  =  1
  [802 , 1024]  =  1
  [945 , 1024]  =  1
  [971 , 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.