Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.183846  0.151527
 0.612756  0.0537355

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.0337995  0.0278576   0.0278576   0.0229604
 0.112653   0.00987907  0.0928489   0.00814236
 0.112653   0.0928489   0.00987907  0.00814236
 0.375469   0.0329267   0.0329267   0.0028875

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  0.419451   -1.1828    -1.1828    3.33533
 -4.78308     1.43508   13.4877   -4.04673
 -4.78308    13.4877     1.43508  -4.04673
 54.5424    -16.3645   -16.3645    4.90987

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 1.49094e-9   1.22884e-9   1.22884e-9   …  1.77762e-10  1.46512e-10
 4.96926e-9   4.35778e-10  4.09568e-9      5.92476e-10  5.1957e-11
 4.96926e-9   4.09568e-9   4.35778e-10     6.30391e-11  5.1957e-11
 1.65624e-8   1.45244e-9   1.45244e-9      2.10108e-10  1.84253e-11
 4.96926e-9   4.09568e-9   4.09568e-9      6.30391e-11  5.1957e-11
 1.65624e-8   1.45244e-9   1.36508e-8   …  2.10108e-10  1.84253e-11
 1.65624e-8   1.36508e-8   1.45244e-9      2.23553e-11  1.84253e-11
 5.52022e-8   4.84095e-9   4.84095e-9      7.45098e-11  6.53412e-12
 4.96926e-9   4.09568e-9   4.09568e-9      6.30391e-11  5.1957e-11
 1.65624e-8   1.45244e-9   1.36508e-8      2.10108e-10  1.84253e-11
 ⋮                                      ⋱               ⋮
 0.000840651  7.37207e-5   7.37207e-5      1.86374e-14  1.63441e-15
 7.56747e-5   6.23714e-5   6.23714e-5      1.57682e-14  1.29962e-14
 0.000252222  2.21186e-5   0.000207882     5.25551e-14  4.6088e-15
 0.000252222  0.000207882  2.21186e-5   …  5.59183e-15  4.6088e-15
 0.000840651  7.37207e-5   7.37207e-5      1.86374e-14  1.63441e-15
 0.000252222  0.000207882  0.000207882     5.59183e-15  4.6088e-15
 0.000840651  7.37207e-5   0.000692867     1.86374e-14  1.63441e-15
 0.000840651  0.000692867  7.37207e-5      1.98301e-15  1.63441e-15
 0.00280187   0.000245709  0.000245709  …  6.60933e-15  5.79604e-16

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 1.70292e-11  4.84058e-12  4.84058e-12  …  1.66697e-17  4.73838e-18
 1.95747e-11  1.28728e-11  5.56415e-12     1.91614e-17  1.2601e-17
 1.95747e-11  5.56415e-12  1.28728e-11     4.43304e-17  1.2601e-17
 2.25007e-11  1.4797e-11   1.4797e-11      5.09569e-17  3.35104e-17
 1.95747e-11  5.56415e-12  5.56415e-12     4.43304e-17  1.2601e-17
 2.25007e-11  1.4797e-11   6.39587e-12  …  5.09569e-17  3.35104e-17
 2.25007e-11  6.39587e-12  1.4797e-11      1.1789e-16   3.35104e-17
 2.58641e-11  1.70088e-11  1.70088e-11     1.35512e-16  8.91156e-17
 1.95747e-11  5.56415e-12  5.56415e-12     4.43304e-17  1.2601e-17
 2.25007e-11  1.4797e-11   6.39587e-12     5.09569e-17  3.35104e-17
 ⋮                                      ⋱               ⋮
 7.88324e-11  5.18419e-11  5.18419e-11     3.38979e-13  2.2292e-13
 5.96627e-11  1.69592e-11  1.69592e-11     1.10891e-13  3.15211e-14
 6.8581e-11   4.51004e-11  1.94943e-11     1.27467e-13  8.38254e-14
 6.8581e-11   1.94943e-11  4.51004e-11  …  2.94898e-13  8.38254e-14
 7.88324e-11  5.18419e-11  5.18419e-11     3.38979e-13  2.2292e-13
 6.8581e-11   1.94943e-11  1.94943e-11     2.94898e-13  8.38254e-14
 7.88324e-11  5.18419e-11  2.24082e-11     3.38979e-13  2.2292e-13
 7.88324e-11  2.24082e-11  5.18419e-11     7.84236e-13  2.2292e-13
 9.06161e-11  5.95912e-11  5.95912e-11  …  9.01463e-13  5.92822e-13

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.0505977  0.289571
 0.615929   0.624501

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 1.09979e-13  6.29411e-13  6.29411e-13  …  7.24317e-7   4.14527e-6
 1.33878e-12  1.35741e-12  7.66187e-12     8.81716e-6   8.93986e-6
 1.33878e-12  7.66187e-12  1.35741e-12     1.56209e-6   8.93986e-6
 1.62971e-11  1.65239e-11  1.65239e-11     1.90155e-5   1.92801e-5
 1.33878e-12  7.66187e-12  7.66187e-12     1.56209e-6   8.93986e-6
 1.62971e-11  1.65239e-11  9.32684e-11  …  1.90155e-5   1.92801e-5
 1.62971e-11  9.32684e-11  1.65239e-11     3.36887e-6   1.92801e-5
 1.98386e-10  2.01146e-10  2.01146e-10     4.10095e-5   4.15802e-5
 1.33878e-12  7.66187e-12  7.66187e-12     1.56209e-6   8.93986e-6
 1.62971e-11  1.65239e-11  9.32684e-11     1.90155e-5   1.92801e-5
 ⋮                                      ⋱
 0.00064552   0.000654504  0.000654504  …  0.00412621   0.00418363
 4.35622e-6   2.49307e-5   2.49307e-5      0.000157171  0.000899493
 5.30286e-5   5.37665e-5   0.000303483     0.00191326   0.00193988
 5.30286e-5   0.000303483  5.37665e-5      0.000338962  0.00193988
 0.00064552   0.000654504  0.000654504     0.00412621   0.00418363
 5.30286e-5   0.000303483  0.000303483  …  0.000338962  0.00193988
 0.00064552   0.000654504  0.00369432      0.00412621   0.00418363
 0.00064552   0.00369432   0.000654504     0.00073102   0.00418363
 0.00785796   0.00796732   0.00796732      0.00889876   0.0090226

julia> sum(P10)  # expected number of edges
97.32358563555232

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 97 stored entries:
  [1024,   16]  =  true
  [1021,   26]  =  true
  [956 ,   29]  =  true
  [894 ,  132]  =  true
  [890 ,  151]  =  true
  [1024,  163]  =  true
  [1016,  189]  =  true
  [998 ,  191]  =  true
  [1024,  191]  =  true
  ⋮
  [633 ,  944]  =  true
  [736 ,  947]  =  true
  [479 ,  952]  =  true
  [640 ,  954]  =  true
  [504 ,  955]  =  true
  [766 ,  979]  =  true
  [384 ,  989]  =  true
  [994 , 1018]  =  true
  [759 , 1020]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source