Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.592413  0.632038
 0.331755  0.915183
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.350954  0.374428  0.374428  0.399472
 0.196536  0.542167  0.209682  0.578431
 0.196536  0.209682  0.542167  0.578431
 0.110061  0.303616  0.303616  0.83756
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  7.57655   -5.23247  -5.23247   3.61362
 -2.74651    4.90443   1.89678  -3.38707
 -2.74651    1.89678   4.90443  -3.38707
  0.995611  -1.77786  -1.77786   3.17472
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.00186852   0.0019935    0.0019935    …  0.00380892  0.00380892  0.00406368
 0.00104638   0.00288657   0.00111637      0.00551526  0.00213301  0.00588416
 0.00104638   0.00111637   0.00288657      0.00213301  0.00551526  0.00588416
 0.00058598   0.00161649   0.00161649      0.00308858  0.00308858  0.00852019
 0.00104638   0.00111637   0.00111637      0.00551526  0.00551526  0.00588416
 0.00058598   0.00161649   0.000625175  …  0.00798603  0.00308858  0.00852019
 0.00058598   0.000625175  0.00161649      0.00308858  0.00798603  0.00852019
 0.000328152  0.000905245  0.000905245     0.00447222  0.00447222  0.0123371
 0.00104638   0.00111637   0.00111637      0.00551526  0.00551526  0.00588416
 0.00058598   0.00161649   0.000625175     0.00798603  0.00308858  0.00852019
 ⋮                                      ⋱                          ⋮
 3.17404e-6   8.75595e-6   8.75595e-6      0.0864251   0.0864251   0.238414
 1.01211e-5   1.07981e-5   1.07981e-5      0.106582    0.106582    0.113711
 5.66788e-6   1.56355e-5   6.04698e-6      0.154329    0.0596864   0.164652
 5.66788e-6   6.04698e-6   1.56355e-5   …  0.0596864   0.154329    0.164652
 3.17404e-6   8.75595e-6   8.75595e-6      0.0864251   0.0864251   0.238414
 5.66788e-6   6.04698e-6   6.04698e-6      0.154329    0.154329    0.164652
 3.17404e-6   8.75595e-6   3.38634e-6      0.223467    0.0864251   0.238414
 3.17404e-6   3.38634e-6   8.75595e-6      0.0864251   0.223467    0.238414
 1.77748e-6   4.90338e-6   4.90338e-6   …  0.125142    0.125142    0.34522
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000964196  0.00163875   0.00163875   …  0.329605  0.329605  0.560197
 0.000860176  0.00180108   0.00146196      0.362253  0.294046  0.615686
 0.000860176  0.00146196   0.00180108      0.294046  0.362253  0.615686
 0.000767377  0.00160677   0.00160677      0.323172  0.323172  0.676672
 0.000860176  0.00146196   0.00146196      0.362253  0.362253  0.615686
 0.000767377  0.00160677   0.00130424   …  0.398135  0.323172  0.676672
 0.000767377  0.00130424   0.00160677      0.323172  0.398135  0.676672
 0.00068459   0.00143343   0.00143343      0.355183  0.355183  0.743698
 0.000860176  0.00146196   0.00146196      0.362253  0.362253  0.615686
 0.000767377  0.00160677   0.00130424      0.398135  0.323172  0.676672
 ⋮                                      ⋱                      ⋮
 0.000274666  0.000575108  0.000575108     0.756135  0.756135  1.58323
 0.000345113  0.000586555  0.000586555     0.771186  0.771186  1.31071
 0.000307881  0.000644655  0.000523276     0.847574  0.687988  1.44054
 0.000307881  0.000523276  0.000644655  …  0.687988  0.847574  1.44054
 0.000274666  0.000575108  0.000575108     0.756135  0.756135  1.58323
 0.000307881  0.000523276  0.000523276     0.847574  0.847574  1.44054
 0.000274666  0.000575108  0.000466823     0.931529  0.756135  1.58323
 0.000274666  0.000466823  0.000575108     0.756135  0.931529  1.58323
 0.000245034  0.000513063  0.000513063  …  0.831032  0.831032  1.74005

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.909054  0.508696
 0.572311  0.400933
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.385385    0.215657    0.215657    0.120679    …  0.00207356   0.00116034
 0.242626    0.169972    0.135771    0.0951143      0.00130544   0.00091453
 0.242626    0.135771    0.169972    0.0951143      0.00163429   0.00091453
 0.152749    0.107009    0.107009    0.074965       0.0010289    0.000720794
 0.242626    0.135771    0.135771    0.0759757      0.00163429   0.00091453
 0.152749    0.107009    0.0854768   0.0598809   …  0.0010289    0.000720794
 0.152749    0.0854768   0.107009    0.0598809      0.00128808   0.000720794
 0.096166    0.0673692   0.0673692   0.0471955      0.000810932  0.000568099
 0.242626    0.135771    0.135771    0.0759757      0.00163429   0.00091453
 0.152749    0.107009    0.0854768   0.0598809      0.0010289    0.000720794
 ⋮                                               ⋱
 0.00598792  0.00419484  0.00419484  0.0029387   …  0.000194385  0.000136176
 0.0151074   0.00845396  0.00845396  0.00473074     0.000391748  0.000219218
 0.00951116  0.00666305  0.00532234  0.00372857     0.000246632  0.000172778
 0.00951116  0.00532234  0.00666305  0.00372857     0.000308759  0.000172778
 0.00598792  0.00419484  0.00419484  0.0029387      0.000194385  0.000136176
 0.00951116  0.00532234  0.00532234  0.00297832  …  0.000308759  0.000172778
 0.00598792  0.00419484  0.00335077  0.00234739     0.000194385  0.000136176
 0.00598792  0.00335077  0.00419484  0.00234739     0.000243351  0.000136176
 0.0037698   0.00264094  0.00264094  0.00185011     0.000153206  0.000107329
julia> sum(P10)  # expected number of edges6106.371551625221
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 6052 stored entries:
  [3  ,    1]  =  1
  [10 ,    1]  =  1
  [13 ,    1]  =  1
  [24 ,    1]  =  1
  [33 ,    1]  =  1
  [36 ,    1]  =  1
  [38 ,    1]  =  1
  [50 ,    1]  =  1
  [52 ,    1]  =  1
  ⋮
  [594, 1015]  =  1
  [507, 1016]  =  1
  [64 , 1017]  =  1
  [749, 1017]  =  1
  [628, 1020]  =  1
  [593, 1021]  =  1
  [283, 1022]  =  1
  [406, 1022]  =  1
  [47 , 1023]  =  1
  [358, 1023]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.