Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.388943  0.22538
 0.536     0.681457
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.151277  0.0876601  0.0876601  0.0507962
 0.208474  0.265048   0.120804   0.153587
 0.208474  0.120804   0.265048   0.153587
 0.287296  0.365261   0.365261   0.464384
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  22.3192   -7.38169   -7.38169   2.44137
 -17.5552   12.7387     5.80607  -4.21312
 -17.5552    5.80607   12.7387   -4.21312
  13.808   -10.0197   -10.0197    7.27067
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.1985e-5    6.9449e-6    6.9449e-6    …  2.96456e-8   1.71787e-8
 1.65164e-5   2.09985e-5   9.57072e-6      4.08543e-8   5.19412e-8
 1.65164e-5   9.57072e-6   2.09985e-5      8.9636e-8    5.19412e-8
 2.27611e-5   2.89379e-5   2.89379e-5      1.23527e-7   1.57049e-7
 1.65164e-5   9.57072e-6   9.57072e-6      8.9636e-8    5.19412e-8
 2.27611e-5   2.89379e-5   1.31893e-5   …  1.23527e-7   1.57049e-7
 2.27611e-5   1.31893e-5   2.89379e-5      2.71022e-7   1.57049e-7
 3.13669e-5   3.98791e-5   3.98791e-5      3.73494e-7   4.74851e-7
 1.65164e-5   9.57072e-6   9.57072e-6      8.9636e-8    5.19412e-8
 2.27611e-5   2.89379e-5   1.31893e-5      1.23527e-7   1.57049e-7
 ⋮                                      ⋱               ⋮
 0.000408035  0.000518766  0.000518766     0.00260895   0.00331695
 0.000214853  0.0001245    0.0001245       0.000626129  0.000362822
 0.000296087  0.000376438  0.000171573     0.000862864  0.00109702
 0.000296087  0.000171573  0.000376438  …  0.00189316   0.00109702
 0.000408035  0.000518766  0.000518766     0.00260895   0.00331695
 0.000296087  0.000171573  0.000171573     0.00189316   0.00109702
 0.000408035  0.000518766  0.000236443     0.00260895   0.00331695
 0.000408035  0.000236443  0.000518766     0.00572413   0.00331695
 0.00056231   0.000714907  0.000714907  …  0.00788838   0.0100291
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.64579e-7   1.45928e-7   1.45928e-7   …  4.38309e-8   3.88637e-8
 3.47045e-7   3.53973e-7   3.07716e-7      9.24258e-8   9.42708e-8
 3.47045e-7   3.07716e-7   3.53973e-7      1.0632e-7    9.42708e-8
 7.31812e-7   7.4642e-7    7.4642e-7       2.24195e-7   2.28671e-7
 3.47045e-7   3.07716e-7   3.07716e-7      1.0632e-7    9.42708e-8
 7.31812e-7   7.4642e-7    6.48879e-7   …  2.24195e-7   2.28671e-7
 7.31812e-7   6.48879e-7   7.4642e-7       2.57897e-7   2.28671e-7
 1.54317e-6   1.57397e-6   1.57397e-6      5.43825e-7   5.54681e-7
 3.47045e-7   3.07716e-7   3.07716e-7      1.0632e-7    9.42708e-8
 7.31812e-7   7.4642e-7    6.48879e-7      2.24195e-7   2.28671e-7
 ⋮                                      ⋱               ⋮
 0.000603278  0.000615321  0.000615321     0.000651812  0.000664824
 0.000135672  0.000120297  0.000120297     0.000127431  0.00011299
 0.000286091  0.000291802  0.00025367      0.000268714  0.000274078
 0.000286091  0.00025367   0.000291802  …  0.000309107  0.000274078
 0.000603278  0.000615321  0.000615321     0.000651812  0.000664824
 0.000286091  0.00025367   0.00025367      0.000309107  0.000274078
 0.000603278  0.000615321  0.000534912     0.000651812  0.000664824
 0.000603278  0.000534912  0.000615321     0.000749795  0.000664824
 0.00127213   0.00129752   0.00129752   …  0.00158109   0.00161265

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.534122  0.00561889
 0.978099  0.0434442
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.00188976  1.988e-5     1.988e-5     …  2.98189e-21  3.13691e-23
 0.00346058  0.000153709  3.64048e-5      5.46052e-21  2.4254e-22
 0.00346058  3.64048e-5   0.000153709     2.30555e-20  2.4254e-22
 0.00633711  0.000281475  0.000281475     4.22198e-20  1.87528e-21
 0.00346058  3.64048e-5   3.64048e-5      2.30555e-20  2.4254e-22
 0.00633711  0.000281475  6.66654e-5   …  4.22198e-20  1.87528e-21
 0.00633711  6.66654e-5   0.000281475     1.78261e-19  1.87528e-21
 0.0116047   0.000515445  0.000515445     3.26436e-19  1.44993e-20
 0.00346058  3.64048e-5   3.64048e-5      2.30555e-20  2.4254e-22
 0.00633711  0.000281475  6.66654e-5      4.22198e-20  1.87528e-21
 ⋮                                     ⋱
 0.437608    0.0194373    0.0194373    …  6.97414e-14  3.0977e-15
 0.130497    0.00137281   0.00137281      4.92568e-15  5.18174e-17
 0.23897     0.0106143    0.00251393      9.02004e-15  4.00643e-16
 0.23897     0.00251393   0.0106143       3.80845e-14  4.00643e-16
 0.437608    0.0194373    0.0194373       6.97414e-14  3.0977e-15
 0.23897     0.00251393   0.00251393   …  3.80845e-14  4.00643e-16
 0.437608    0.0194373    0.00460357      6.97414e-14  3.0977e-15
 0.437608    0.00460357   0.0194373       2.94463e-13  3.0977e-15
 0.80136     0.035594     0.035594        5.39228e-13  2.39509e-14
julia> sum(P10)  # expected number of edges86.06368038289102
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 82 stored entries:
  [24 ,   1]  =  1
  [108,   1]  =  1
  [120,   1]  =  1
  [183,   1]  =  1
  [206,   1]  =  1
  [220,   1]  =  1
  [232,   1]  =  1
  [249,   1]  =  1
  [291,   1]  =  1
  ⋮
  [57 ,  33]  =  1
  [203,  65]  =  1
  [719, 129]  =  1
  [724, 129]  =  1
  [392, 257]  =  1
  [722, 257]  =  1
  [884, 257]  =  1
  [348, 273]  =  1
  [693, 514]  =  1
  [966, 522]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.