Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.74927   0.619639
 0.492383  0.996972

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.561406  0.464277  0.464277  0.383952
 0.368928  0.747002  0.3051    0.617763
 0.368928  0.3051    0.747002  0.617763
 0.242441  0.490892  0.490892  0.993953

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  5.08996  -3.16351  -3.16351   1.96619
 -2.51382   3.82534   1.56239  -2.37753
 -2.51382   1.56239   3.82534  -2.37753
  1.24152  -1.88925  -1.88925   2.87492

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.0313086    0.0258919    0.0258919    …  0.00387404  0.00387404  0.00320379
 0.0205744    0.0416589    0.0170149       0.00623317  0.00254583  0.00515477
 0.0205744    0.0170149    0.0416589       0.00254583  0.00623317  0.00515477
 0.0135205    0.0273762    0.0273762       0.00409613  0.00409613  0.00829379
 0.0205744    0.0170149    0.0170149       0.00623317  0.00623317  0.00515477
 0.0135205    0.0273762    0.0111813    …  0.0100289   0.00409613  0.00829379
 0.0135205    0.0111813    0.0273762       0.00409613  0.0100289   0.00829379
 0.008885     0.0179902    0.0179902       0.00659049  0.00659049  0.0133444
 0.0205744    0.0170149    0.0170149       0.00623317  0.00623317  0.00515477
 0.0135205    0.0273762    0.0111813       0.0100289   0.00409613  0.00829379
 ⋮                                      ⋱                          ⋮
 0.000309011  0.000625682  0.000625682     0.295987    0.295987    0.599311
 0.000715558  0.000591759  0.000591759     0.279939    0.279939    0.231507
 0.000470229  0.000952115  0.000388875     0.45041     0.183962    0.372484
 0.000470229  0.000388875  0.000952115  …  0.183962    0.45041     0.372484
 0.000309011  0.000625682  0.000625682     0.295987    0.295987    0.599311
 0.000470229  0.000388875  0.000388875     0.45041     0.45041     0.372484
 0.000309011  0.000625682  0.000255549     0.724689    0.295987    0.599311
 0.000309011  0.000255549  0.000625682     0.295987    0.724689    0.599311
 0.000203067  0.000411167  0.000411167  …  0.47623     0.47623     0.964265

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.179168  0.223734  0.223734  0.279385  …   2.06276   2.06276   2.57585
 0.177785  0.268605  0.222007  0.335418      2.47646   2.04684   3.09245
 0.177785  0.222007  0.268605  0.335418      2.04684   2.47646   3.09245
 0.176413  0.266533  0.266533  0.402689      2.45735   2.45735   3.71268
 0.177785  0.222007  0.222007  0.277229      2.47646   2.47646   3.09245
 0.176413  0.266533  0.220294  0.33283   …   2.97314   2.45735   3.71268
 0.176413  0.220294  0.266533  0.33283       2.45735   2.97314   3.71268
 0.175052  0.264477  0.264477  0.399582      2.9502    2.9502    4.45729
 0.177785  0.222007  0.222007  0.277229      2.47646   2.47646   3.09245
 0.176413  0.266533  0.220294  0.33283       2.97314   2.45735   3.71268
 ⋮                                       ⋱                       ⋮
 0.164535  0.248586  0.248586  0.375574     12.7327   12.7327   19.2371
 0.167103  0.208668  0.208668  0.260572     10.6881   10.6881   13.3467
 0.165814  0.250519  0.207058  0.312833     12.8317   10.6056   16.0235
 0.165814  0.207058  0.250519  0.312833  …  10.6056   12.8317   16.0235
 0.164535  0.248586  0.248586  0.375574     12.7327   12.7327   19.2371
 0.165814  0.207058  0.207058  0.258562     12.8317   12.8317   16.0235
 0.164535  0.248586  0.205461  0.310419     15.4052   12.7327   19.2371
 0.164535  0.205461  0.248586  0.310419     12.7327   15.4052   19.2371
 0.163265  0.246668  0.246668  0.372677  …  15.2864   15.2864   23.0953

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.866729   0.672905
 0.0401951  0.399853

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.23924      0.18574      0.18574      …  0.024517     0.0190343
 0.0110949    0.11037      0.00861381      0.00113699   0.0113106
 0.0110949    0.00861381   0.11037         0.0145685    0.0113106
 0.000514535  0.00511849   0.00511849      0.000675624  0.00672097
 0.0110949    0.00861381   0.00861381      0.0145685    0.0113106
 0.000514535  0.00511849   0.000399471  …  0.000675624  0.00672097
 0.000514535  0.000399471  0.00511849      0.00865688   0.00672097
 2.38619e-5   0.000237374  0.000237374     0.000401469  0.00399373
 0.0110949    0.00861381   0.00861381      0.0145685    0.0113106
 0.000514535  0.00511849   0.000399471     0.000675624  0.00672097
 ⋮                                      ⋱
 2.37381e-13  2.36142e-12  2.36142e-12  …  1.7674e-5    0.000175817
 1.10374e-10  8.56911e-11  8.56911e-11     0.000641352  0.000497928
 5.11865e-12  5.09194e-11  3.97398e-12     2.97431e-5   0.000295879
 5.11865e-12  3.97398e-12  5.09194e-11     0.000381104  0.000295879
 2.37381e-13  2.36142e-12  2.36142e-12     1.7674e-5    0.000175817
 5.11865e-12  3.97398e-12  3.97398e-12  …  0.000381104  0.000295879
 2.37381e-13  2.36142e-12  1.84296e-13     1.7674e-5    0.000175817
 2.37381e-13  1.84296e-13  2.36142e-12     0.00022646   0.000175817
 1.10087e-14  1.09512e-13  1.09512e-13     1.05022e-5   0.000104474

julia> sum(P10)  # expected number of edges
924.6051521991646

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 917 stored entries:
  [2  ,   10]  =  1
  [265,   10]  =  1
  [259,   12]  =  1
  [13 ,   13]  =  1
  [6  ,   14]  =  1
  [1  ,   16]  =  1
  [148,   16]  =  1
  [1  ,   17]  =  1
  [2  ,   22]  =  1
  ⋮
  [378, 1020]  =  1
  [587, 1020]  =  1
  [365, 1022]  =  1
  [101, 1023]  =  1
  [185, 1023]  =  1
  [565, 1023]  =  1
  [67 , 1024]  =  1
  [129, 1024]  =  1
  [361, 1024]  =  1
  [746, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.