Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.109975  0.667577
 0.114048  0.23607
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.0120945  0.0734167  0.0734167  0.44566
 0.0125424  0.0259618  0.076136   0.157595
 0.0125424  0.076136   0.0259618  0.157595
 0.013007   0.0269234  0.0269234  0.0557292
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  22.1371   -62.6011   -62.6011   177.028
 -10.6947    10.3127    30.2433   -29.1631
 -10.6947    30.2433    10.3127   -29.1631
   5.16673   -4.98219   -4.98219    4.80425
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 3.12984e-12  1.8999e-11   1.8999e-11   …  0.00129066   0.00783465
 3.24576e-12  6.71846e-12  1.97027e-11     0.00133846   0.00277051
 3.24576e-12  1.97027e-11  6.71846e-12     0.000456406  0.00277051
 3.36598e-12  6.9673e-12   6.9673e-12      0.000473311  0.000979714
 3.24576e-12  1.97027e-11  1.97027e-11     0.000456406  0.00277051
 3.36598e-12  6.9673e-12   2.04324e-11  …  0.000473311  0.000979714
 3.36598e-12  2.04324e-11  6.9673e-12      0.000161395  0.000979714
 3.49065e-12  7.22536e-12  7.22536e-12     0.000167373  0.000346449
 3.24576e-12  1.97027e-11  1.97027e-11     0.000456406  0.00277051
 3.36598e-12  6.9673e-12   2.04324e-11     0.000473311  0.000979714
 ⋮                                      ⋱               ⋮
 4.66947e-12  9.66541e-12  9.66541e-12     4.09264e-8   8.47142e-8
 4.34187e-12  2.63564e-11  2.63564e-11     1.11601e-7   6.77449e-7
 4.50269e-12  9.3202e-12   2.73326e-11     1.15735e-7   2.39561e-7
 4.50269e-12  2.73326e-11  9.3202e-12   …  3.94647e-8   2.39561e-7
 4.66947e-12  9.66541e-12  9.66541e-12     4.09264e-8   8.47142e-8
 4.50269e-12  2.73326e-11  2.73326e-11     3.94647e-8   2.39561e-7
 4.66947e-12  9.66541e-12  2.83449e-11     4.09264e-8   8.47142e-8
 4.66947e-12  2.83449e-11  9.66541e-12     1.39556e-8   8.47142e-8
 4.84242e-12  1.00234e-11  1.00234e-11  …  1.44725e-8   2.99568e-8
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 2.22548e-13  5.82694e-13  5.82694e-13  …  8.82262e-9   2.31001e-8
 9.95468e-14  3.32611e-13  2.60641e-13     3.94639e-9   1.31859e-8
 9.95468e-14  2.60641e-13  3.32611e-13     5.03609e-9   1.31859e-8
 4.45277e-14  1.48778e-13  1.48778e-13     2.25266e-9   7.5267e-9
 9.95468e-14  2.60641e-13  2.60641e-13     5.03609e-9   1.31859e-8
 4.45277e-14  1.48778e-13  1.16586e-13  …  2.25266e-9   7.5267e-9
 4.45277e-14  1.16586e-13  1.48778e-13     2.87468e-9   7.5267e-9
 1.99174e-14  6.6549e-14   6.6549e-14      1.28585e-9   4.29636e-9
 9.95468e-14  2.60641e-13  2.60641e-13     5.03609e-9   1.31859e-8
 4.45277e-14  1.48778e-13  1.16586e-13     2.25266e-9   7.5267e-9
 ⋮                                      ⋱               ⋮
 3.19193e-17  1.0665e-16   1.0665e-16      1.4493e-11   4.84247e-11
 1.59532e-16  4.17699e-16  4.17699e-16     5.67622e-11  1.48619e-10
 7.13593e-17  2.38429e-16  1.86838e-16     2.539e-11    8.48342e-11
 7.13593e-17  1.86838e-16  2.38429e-16  …  3.24008e-11  8.48342e-11
 3.19193e-17  1.0665e-16   1.0665e-16      1.4493e-11   4.84247e-11
 7.13593e-17  1.86838e-16  1.86838e-16     3.24008e-11  8.48342e-11
 3.19193e-17  1.0665e-16   8.35736e-17     1.4493e-11   4.84247e-11
 3.19193e-17  8.35736e-17  1.0665e-16      1.84949e-11  4.84247e-11
 1.42776e-17  4.77051e-17  4.77051e-17  …  8.27282e-12  2.76416e-11

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.526401  0.313931
 0.768846  0.900857
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.00163367  0.000974275  0.000974275  …  1.55892e-5   9.29696e-6
 0.00238609  0.00279578   0.001423        2.27691e-5   2.66786e-5
 0.00238609  0.001423     0.00279578      4.47347e-5   2.66786e-5
 0.00348506  0.00408344   0.00408344      6.53383e-5   7.65569e-5
 0.00238609  0.001423     0.001423        4.47347e-5   2.66786e-5
 0.00348506  0.00408344   0.00207839   …  6.53383e-5   7.65569e-5
 0.00348506  0.00207839   0.00408344      0.000128371  7.65569e-5
 0.00509018  0.00596416   0.00596416      0.000187495  0.000219688
 0.00238609  0.001423     0.001423        4.47347e-5   2.66786e-5
 0.00348506  0.00408344   0.00207839      6.53383e-5   7.65569e-5
 ⋮                                     ⋱
 0.0494164   0.0579011    0.0579011    …  0.104693     0.122669
 0.0231646   0.0138147    0.0138147       0.024979     0.0148968
 0.0338336   0.0396428    0.0201774       0.0364836    0.0427478
 0.0338336   0.0201774    0.0396428       0.0716796    0.0427478
 0.0494164   0.0579011    0.0579011       0.104693     0.122669
 0.0338336   0.0201774    0.0201774    …  0.0716796    0.0427478
 0.0494164   0.0579011    0.0294706       0.104693     0.122669
 0.0494164   0.0294706    0.0579011       0.205692     0.122669
 0.0721761   0.0845688    0.0845688       0.300428     0.352011
julia> sum(P10)  # expected number of edges9926.504634345421
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 9778 stored entries:
  [91  ,    1]  =  1
  [198 ,    1]  =  1
  [199 ,    1]  =  1
  [202 ,    1]  =  1
  [311 ,    1]  =  1
  [435 ,    1]  =  1
  [439 ,    1]  =  1
  [458 ,    1]  =  1
  [551 ,    1]  =  1
  ⋮
  [640 , 1024]  =  1
  [750 , 1024]  =  1
  [948 , 1024]  =  1
  [993 , 1024]  =  1
  [1002, 1024]  =  1
  [1004, 1024]  =  1
  [1008, 1024]  =  1
  [1014, 1024]  =  1
  [1022, 1024]  =  1
  [1024, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.