Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.590951  0.0563896
 0.94023   0.256955

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.349223  0.0333235  0.0333235  0.00317978
 0.55563   0.151848   0.0530192  0.0144896
 0.55563   0.0530192  0.151848   0.0144896
 0.884033  0.241597   0.241597   0.0660261

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
   6.76003   -1.48351   -1.48351   0.325559
 -24.7357    15.5468     5.42832  -3.41179
 -24.7357     5.42832   15.5468   -3.41179
  90.511    -56.8877   -56.8877   35.7549

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.00181391  0.000173086  0.000173086  …  1.08327e-14  1.03368e-15
 0.00288601  0.000788718  0.000275388     1.72354e-14  4.71025e-15
 0.00288601  0.000275388  0.000788718     4.93624e-14  4.71025e-15
 0.00459178  0.00125489   0.00125489      7.8538e-14   2.14636e-14
 0.00288601  0.000275388  0.000275388     4.93624e-14  4.71025e-15
 0.00459178  0.00125489   0.000438156  …  7.8538e-14   2.14636e-14
 0.00459178  0.000438156  0.00125489      2.24934e-13  2.14636e-14
 0.00730574  0.00199658   0.00199658      3.57881e-13  9.78053e-14
 0.00288601  0.000275388  0.000275388     4.93624e-14  4.71025e-15
 0.00459178  0.00125489   0.000438156     7.8538e-14   2.14636e-14
 ⋮                                     ⋱               ⋮
 0.300005    0.0819883    0.0819883       6.65291e-8   1.81817e-8
 0.118512    0.0113086    0.0113086       9.17635e-9   8.75623e-10
 0.188558    0.051531     0.0179926       1.46e-8      3.99003e-9
 0.188558    0.0179926    0.051531     …  4.18147e-8   3.99003e-9
 0.300005    0.0819883    0.0819883       6.65291e-8   1.81817e-8
 0.188558    0.0179926    0.0179926       4.18147e-8   3.99003e-9
 0.300005    0.0819883    0.028627        6.65291e-8   1.81817e-8
 0.300005    0.028627     0.0819883       1.90541e-7   1.81817e-8
 0.477322    0.130447     0.130447     …  3.03159e-7   8.28504e-8

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.7941e-5    2.13259e-6  2.13259e-6  …  1.20088e-15  1.42744e-16
 3.55584e-5   5.30973e-6  4.2267e-6      2.38009e-15  3.55406e-16
 3.55584e-5   4.2267e-6   5.30973e-6     2.98996e-15  3.55406e-16
 7.04754e-5   1.05237e-5  1.05237e-5     5.92598e-15  8.84892e-16
 3.55584e-5   4.2267e-6   4.2267e-6      2.98996e-15  3.55406e-16
 7.04754e-5   1.05237e-5  8.37715e-6  …  5.92598e-15  8.84892e-16
 7.04754e-5   8.37715e-6  1.05237e-5     7.44442e-15  8.84892e-16
 0.000139679  2.08575e-5  2.08575e-5     1.47545e-14  2.20321e-15
 3.55584e-5   4.2267e-6   4.2267e-6      2.98996e-15  3.55406e-16
 7.04754e-5   1.05237e-5  8.37715e-6     5.92598e-15  8.84892e-16
 ⋮                                    ⋱               ⋮
 0.0332575    0.00496616  0.00496616     2.17897e-11  3.25373e-12
 0.00846644   0.00100638  0.00100638     4.41561e-12  5.24868e-13
 0.0167801    0.00250568  0.00199459     8.75156e-12  1.30682e-12
 0.0167801    0.00199459  0.00250568  …  1.0994e-11   1.30682e-12
 0.0332575    0.00496616  0.00496616     2.17897e-11  3.25373e-12
 0.0167801    0.00199459  0.00199459     1.0994e-11   1.30682e-12
 0.0332575    0.00496616  0.00395321     2.17897e-11  3.25373e-12
 0.0332575    0.00395321  0.00496616     2.7373e-11   3.25373e-12
 0.0659151    0.00984272  0.00984272  …  5.42522e-11  8.10116e-12

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.238365  0.503397
 0.54735   0.526477

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 5.92148e-7   1.25054e-6   1.25054e-6   …  0.000494814  0.00104498
 1.35973e-6   1.30788e-6   2.87158e-6      0.00113622   0.00109289
 1.35973e-6   2.87158e-6   1.30788e-6      0.0005175    0.00109289
 3.1223e-6    3.00323e-6   3.00323e-6      0.00118832   0.001143
 1.35973e-6   2.87158e-6   2.87158e-6      0.0005175    0.00109289
 3.1223e-6    3.00323e-6   6.59391e-6   …  0.00118832   0.001143
 3.1223e-6    6.59391e-6   3.00323e-6      0.000541226  0.001143
 7.16963e-6   6.89622e-6   6.89622e-6      0.0012428    0.00119541
 1.35973e-6   2.87158e-6   2.87158e-6      0.0005175    0.00109289
 3.1223e-6    3.00323e-6   6.59391e-6      0.00118832   0.001143
 ⋮                                      ⋱
 0.00105107   0.00101098   0.00101098   …  0.00162634   0.00156432
 0.000199336  0.000420972  0.000420972     0.000677207  0.00143018
 0.000457728  0.000440273  0.000966665     0.00155505   0.00149575
 0.000457728  0.000966665  0.000440273     0.000708256  0.00149575
 0.00105107   0.00101098   0.00101098      0.00162634   0.00156432
 0.000457728  0.000966665  0.000966665  …  0.000708256  0.00149575
 0.00105107   0.00101098   0.00221972      0.00162634   0.00156432
 0.00105107   0.00221972   0.00101098      0.000740728  0.00156432
 0.00241353   0.00232149   0.00232149      0.00170091   0.00163605

julia> sum(P10)  # expected number of edges
389.20445255860415

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 389 stored entries:
  [751 ,    2]  =  true
  [497 ,   12]  =  true
  [660 ,   20]  =  true
  [1004,   31]  =  true
  [988 ,   36]  =  true
  [360 ,   44]  =  true
  [479 ,   46]  =  true
  [408 ,   47]  =  true
  [902 ,   48]  =  true
  ⋮
  [302 , 1018]  =  true
  [223 , 1020]  =  true
  [976 , 1020]  =  true
  [707 , 1021]  =  true
  [780 , 1021]  =  true
  [491 , 1022]  =  true
  [1022, 1022]  =  true
  [114 , 1023]  =  true
  [244 , 1023]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source