Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.941376  0.587607
 0.789486  0.236149

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.886189  0.553159  0.553159  0.345282
 0.743204  0.222305  0.463908  0.138763
 0.743204  0.463908  0.222305  0.138763
 0.623289  0.186436  0.186436  0.0557663

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  0.955362   -2.37722   -2.37722   5.91521
 -3.19394     3.80842    7.94745  -9.47646
 -3.19394     7.94745    3.80842  -9.47646
 10.6779    -12.7322   -12.7322   15.1818

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.484349   0.302331   0.302331   0.188715    …  0.0027147    0.00169451
 0.4062     0.121501   0.25355    0.0758412      0.00227668   0.000680995
 0.4062     0.25355    0.121501   0.0758412      0.00109099   0.000680995
 0.34066    0.101897   0.101897   0.0304793      0.000914959  0.00027368
 0.4062     0.25355    0.25355    0.158266       0.00109099   0.000680995
 0.34066    0.101897   0.21264    0.0636044   …  0.000914959  0.00027368
 0.34066    0.21264    0.101897   0.0636044      0.000438449  0.00027368
 0.285695   0.0854564  0.0854564  0.0255615      0.000367706  0.000109987
 0.4062     0.25355    0.25355    0.158266       0.00109099   0.000680995
 0.34066    0.101897   0.21264    0.0636044      0.000914959  0.00027368
 ⋮                                            ⋱               ⋮
 0.0699124  0.020912   0.020912   0.00625514     2.50202e-7   7.48398e-8
 0.0994011  0.0620462  0.0620462  0.0387292      7.42354e-7   4.63378e-7
 0.0833629  0.0249353  0.0520351  0.0155646      6.22576e-7   1.86223e-7
 0.0833629  0.0520351  0.0249353  0.0155646   …  2.98339e-7   1.86223e-7
 0.0699124  0.020912   0.020912   0.00625514     2.50202e-7   7.48398e-8
 0.0833629  0.0520351  0.0520351  0.0324803      2.98339e-7   1.86223e-7
 0.0699124  0.020912   0.0436393  0.0130533      2.50202e-7   7.48398e-8
 0.0699124  0.0436393  0.020912   0.0130533      1.19897e-7   7.48398e-8
 0.0586321  0.0175379  0.0175379  0.00524588  …  1.00552e-7   3.00768e-8

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 36.6758    18.7963    18.7963    9.63307   …  0.0234961    0.0120417
 25.254     14.1171    12.9426    7.235        0.0161788    0.00904402
 25.254     12.9426    14.1171    7.235        0.0176469    0.00904402
 17.3892     9.72067    9.72067   5.43391      0.0121512    0.00679259
 25.254     12.9426    12.9426    6.63307      0.0176469    0.00904402
 17.3892     9.72067    8.91193   4.98182   …  0.0121512    0.00679259
 17.3892     8.91193    9.72067   4.98182      0.0132539    0.00679259
 11.9737     6.69339    6.69339   3.74164      0.00912626   0.00510163
 25.254     12.9426    12.9426    6.63307      0.0176469    0.00904402
 17.3892     9.72067    8.91193   4.98182      0.0121512    0.00679259
  ⋮                                         ⋱               ⋮
  0.605101   0.338255   0.338255  0.189087     0.000924024  0.000516535
  1.27623    0.654065   0.654065  0.335207     0.00178673   0.000915698
  0.878776   0.491241   0.450371  0.25176      0.0012303    0.000687743
  0.878776   0.450371   0.491241  0.25176   …  0.00134194   0.000687743
  0.605101   0.338255   0.338255  0.189087     0.000924024  0.000516535
  0.878776   0.450371   0.450371  0.230814     0.00134194   0.000687743
  0.605101   0.338255   0.310113  0.173355     0.000924024  0.000516535
  0.605101   0.310113   0.338255  0.173355     0.00100788   0.000516535
  0.416656   0.232913   0.232913  0.1302    …  0.000693997  0.000387948

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.301293  0.24497
 0.5841    0.582208

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 6.16443e-6  5.01206e-6   5.01206e-6   …  9.57202e-7   7.78264e-7
 1.19506e-5  1.19119e-5   9.7166e-6       1.85567e-6   1.84966e-6
 1.19506e-5  9.7166e-6    1.19119e-5      2.27493e-6   1.84966e-6
 2.3168e-5   2.3093e-5    2.3093e-5       4.41028e-6   4.396e-6
 1.19506e-5  9.7166e-6    9.7166e-6       2.27493e-6   1.84966e-6
 2.3168e-5   2.3093e-5    1.8837e-5    …  4.41028e-6   4.396e-6
 2.3168e-5   1.8837e-5    2.3093e-5       5.40672e-6   4.396e-6
 4.49145e-5  4.4769e-5    4.4769e-5       1.04817e-5   1.04478e-5
 1.19506e-5  9.7166e-6    9.7166e-6       2.27493e-6   1.84966e-6
 2.3168e-5   2.3093e-5    1.8837e-5       4.41028e-6   4.396e-6
 ⋮                                     ⋱
 0.00238436  0.00237664   0.00237664   …  0.00188896   0.00188285
 0.00063442  0.000515823  0.000515823     0.000409978  0.000333338
 0.00122991  0.00122593   0.000999996     0.000794801  0.000792227
 0.00122991  0.000999996  0.00122593      0.000974375  0.000792227
 0.00238436  0.00237664   0.00237664      0.00188896   0.00188285
 0.00122991  0.000999996  0.000999996  …  0.000974375  0.000792227
 0.00238436  0.00237664   0.00193863      0.00188896   0.00188285
 0.00238436  0.00193863   0.00237664      0.00231575   0.00188285
 0.00462243  0.00460746   0.00460746      0.00448941   0.00447487

julia> sum(P10)  # expected number of edges
217.0125883627108

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 217 stored entries:
  [606 ,    2]  =  true
  [996 ,    4]  =  true
  [520 ,    5]  =  true
  [381 ,   11]  =  true
  [231 ,   12]  =  true
  [16  ,   13]  =  true
  [767 ,   16]  =  true
  [892 ,   18]  =  true
  [509 ,   22]  =  true
  ⋮
  [1019,  974]  =  true
  [40  ,  976]  =  true
  [331 ,  976]  =  true
  [767 ,  979]  =  true
  [1016,  987]  =  true
  [935 , 1004]  =  true
  [228 , 1010]  =  true
  [996 , 1014]  =  true
  [380 , 1019]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source