Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.8002    0.0785269
 0.618146  0.445876
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.640319  0.0628372  0.0628372  0.00616647
 0.494641  0.35679    0.0485411  0.0350132
 0.494641  0.0485411  0.35679    0.0350132
 0.382105  0.275616   0.275616   0.198805
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  2.09231  -0.368494  -0.368494   0.0648986
 -2.9007    3.755      0.510867  -0.661325
 -2.9007    0.510867   3.755     -0.661325
  4.02144  -5.20581   -5.20581    6.73899
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0689256   0.00676395   0.00676395   …  5.60274e-13  5.4982e-14
 0.0532443   0.0384057    0.00522508      4.32807e-13  3.12188e-13
 0.0532443   0.00522508   0.0384057       3.18124e-12  3.12188e-13
 0.0411307   0.029668     0.029668        2.45748e-12  1.7726e-12
 0.0532443   0.00522508   0.00522508      3.18124e-12  3.12188e-13
 0.0411307   0.029668     0.00403633   …  2.45748e-12  1.7726e-12
 0.0411307   0.00403633   0.029668        1.80631e-11  1.7726e-12
 0.0317731   0.0229183    0.0229183       1.39535e-11  1.00648e-11
 0.0532443   0.00522508   0.00522508      3.18124e-12  3.12188e-13
 0.0411307   0.029668     0.00403633      2.45748e-12  1.7726e-12
 ⋮                                     ⋱               ⋮
 0.00402906  0.0029062    0.0029062       1.50746e-5   1.08735e-5
 0.00675177  0.000662579  0.000662579     3.43684e-6   3.37271e-7
 0.00521568  0.00376212   0.000511836     2.65492e-6   1.91502e-6
 0.00521568  0.000511836  0.00376212   …  1.95144e-5   1.91502e-6
 0.00402906  0.0029062    0.0029062       1.50746e-5   1.08735e-5
 0.00521568  0.000511836  0.000511836     1.95144e-5   1.91502e-6
 0.00402906  0.0029062    0.000395388     1.50746e-5   1.08735e-5
 0.00402906  0.000395388  0.0029062       0.000110802  1.08735e-5
 0.00311241  0.00224501   0.00224501   …  8.55938e-5   6.17397e-5
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0114176  0.00162184  0.00162184  0.000230377  …  5.42401e-12  7.70464e-13
 0.0127668  0.00409968  0.00181348  0.000582346     6.06492e-12  1.94758e-12
 0.0127668  0.00181348  0.00409968  0.000582346     1.37108e-11  1.94758e-12
 0.0142753  0.0045841   0.0045841   0.00147205      1.53309e-11  4.92308e-12
 0.0127668  0.00181348  0.00181348  0.000257598     1.37108e-11  1.94758e-12
 0.0142753  0.0045841   0.00202776  0.000651157  …  1.53309e-11  4.92308e-12
 0.0142753  0.00202776  0.0045841   0.000651157     3.46582e-11  4.92308e-12
 0.0159621  0.00512576  0.00512576  0.00164599      3.87534e-11  1.24446e-11
 0.0127668  0.00181348  0.00181348  0.000257598     1.37108e-11  1.94758e-12
 0.0142753  0.0045841   0.00202776  0.000651157     1.53309e-11  4.92308e-12
 ⋮                                               ⋱               ⋮
 0.0390053  0.0125254   0.0125254   0.00402219      6.46027e-8   2.07453e-8
 0.0311972  0.00443146  0.00443146  0.000629474     2.28562e-8   3.24665e-9
 0.0348834  0.0112018   0.00495508  0.00159118      2.55569e-8   8.20687e-9
 0.0348834  0.00495508  0.0112018   0.00159118   …  5.77759e-8   8.20687e-9
 0.0390053  0.0125254   0.0125254   0.00402219      6.46027e-8   2.07453e-8
 0.0348834  0.00495508  0.00495508  0.000703853     5.77759e-8   8.20687e-9
 0.0390053  0.0125254   0.00554058  0.0017792       6.46027e-8   2.07453e-8
 0.0390053  0.00554058  0.0125254   0.0017792       1.46046e-7   2.07453e-8
 0.0436142  0.0140055   0.0140055   0.00449746   …  1.63303e-7   5.244e-8

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.712021  0.492299
 0.579843  0.312088
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.0334911   0.0231561   0.0231561   0.0160104   …  0.00120935   0.00083616
 0.0272739   0.0146796   0.0188574   0.0101496      0.000984852  0.000530075
 0.0272739   0.0188574   0.0146796   0.0101496      0.000766657  0.000530075
 0.0222108   0.0119545   0.0119545   0.00643423     0.000624337  0.000336035
 0.0272739   0.0188574   0.0188574   0.0130382      0.000766657  0.000530075
 0.0222108   0.0119545   0.0153568   0.00826545  …  0.000624337  0.000336035
 0.0222108   0.0153568   0.0119545   0.00826545     0.000486014  0.000336035
 0.0180876   0.00973527  0.00973527  0.0052398      0.000395792  0.000213026
 0.0272739   0.0188574   0.0188574   0.0130382      0.000766657  0.000530075
 0.0222108   0.0119545   0.0153568   0.00826545     0.000624337  0.000336035
 ⋮                                               ⋱
 0.00527579  0.00283958  0.00283958  0.00152834  …  2.56894e-5   1.38267e-5
 0.00795523  0.00550033  0.00550033  0.00380298     4.97609e-5   3.44052e-5
 0.00647844  0.00348688  0.00447926  0.00241086     4.05234e-5   2.18108e-5
 0.00647844  0.00447926  0.00348688  0.00241086     3.15454e-5   2.18108e-5
 0.00527579  0.00283958  0.00283958  0.00152834     2.56894e-5   1.38267e-5
 0.00647844  0.00447926  0.00447926  0.00309701  …  3.15454e-5   2.18108e-5
 0.00527579  0.00283958  0.00364774  0.00196332     2.56894e-5   1.38267e-5
 0.00527579  0.00364774  0.00283958  0.00196332     1.99979e-5   1.38267e-5
 0.00429641  0.00231245  0.00231245  0.00124462     1.62855e-5   8.76532e-6
julia> sum(P10)  # expected number of edges1638.4494696619959
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1636 stored entries:
  [7  ,    1]  =  1
  [32 ,    1]  =  1
  [33 ,    1]  =  1
  [58 ,    1]  =  1
  [73 ,    1]  =  1
  [154,    1]  =  1
  [189,    1]  =  1
  [327,    1]  =  1
  [386,    1]  =  1
  ⋮
  [713, 1006]  =  1
  [999, 1007]  =  1
  [805, 1011]  =  1
  [738, 1012]  =  1
  [6  , 1017]  =  1
  [284, 1018]  =  1
  [516, 1018]  =  1
  [94 , 1020]  =  1
  [134, 1022]  =  1
  [233, 1023]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.