Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.975895  0.814048
 0.753715  0.564601
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.952371  0.794425  0.794425  0.662674
 0.735547  0.550992  0.61356   0.459613
 0.735547  0.61356   0.550992  0.459613
 0.568087  0.425549  0.425549  0.318775
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
   81.427  -117.402  -117.402   169.272
 -108.701   140.744   156.726  -202.926
 -108.701   156.726   140.744  -202.926
  145.111  -187.886  -187.886   243.271
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.74617    0.622422   0.622422   …  0.10152     0.10152     0.0846836
 0.576292   0.431695   0.480716      0.0704117   0.0784074   0.0587342
 0.576292   0.480716   0.431695      0.0784074   0.0704117   0.0587342
 0.445089   0.333412   0.333412      0.0543812   0.0543812   0.0407365
 0.576292   0.480716   0.480716      0.0704117   0.0704117   0.0587342
 0.445089   0.333412   0.371273   …  0.0488356   0.0543812   0.0407365
 0.445089   0.371273   0.333412      0.0543812   0.0488356   0.0407365
 0.343756   0.257505   0.257505      0.0377173   0.0377173   0.0282537
 0.576292   0.480716   0.480716      0.0704117   0.0704117   0.0587342
 0.445089   0.333412   0.371273      0.0488356   0.0543812   0.0407365
 ⋮                                ⋱                          ⋮
 0.0435196  0.0326001  0.0326001     0.00201965  0.00201965  0.0015129
 0.0729586  0.0608588  0.0608588     0.00377034  0.00377034  0.00314505
 0.0563483  0.04221    0.0470032     0.002615    0.00291195  0.00218132
 0.0563483  0.0470032  0.04221    …  0.00291195  0.002615    0.00218132
 0.0435196  0.0326001  0.0326001     0.00201965  0.00201965  0.0015129
 0.0563483  0.0470032  0.0470032     0.002615    0.002615    0.00218132
 0.0435196  0.0326001  0.0363021     0.00181369  0.00201965  0.0015129
 0.0435196  0.0363021  0.0326001     0.00201965  0.00181369  0.0015129
 0.0336116  0.0251781  0.0251781  …  0.00140077  0.00140077  0.00104931
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 217.407    174.105    174.105    …  18.8882    18.8882    15.1261
 161.202    129.441    129.094       14.0427    14.0051    11.2458
 161.202    129.094    129.441       14.0051    14.0427    11.2458
 119.527     95.9773    95.9773      10.4123    10.4123     8.36086
 161.202    129.094    129.094       14.0427    14.0427    11.2458
 119.527     95.9773    95.72     …  10.4403    10.4123     8.36086
 119.527     95.72      95.9773      10.4123    10.4403     8.36086
  88.6258    71.1646    71.1646       7.7412     7.7412     6.21602
 161.202    129.094    129.094       14.0427    14.0427    11.2458
 119.527     95.9773    95.72        10.4403    10.4123     8.36086
   ⋮                              ⋱                         ⋮
   8.09697    6.50169    6.50169      0.722602   0.722602   0.580234
  14.7276    11.7942    11.7942       1.31082    1.31082    1.04974
  10.9201     8.76862    8.7451       0.97455    0.971937   0.780444
  10.9201     8.7451     8.76862  …   0.971937   0.97455    0.780444
   8.09697    6.50169    6.50169      0.722602   0.722602   0.580234
  10.9201     8.7451     8.7451       0.97455    0.97455    0.780444
   8.09697    6.50169    6.48426      0.724545   0.722602   0.580234
   8.09697    6.48426    6.50169      0.722602   0.724545   0.580234
   6.00368    4.82083    4.82083  …   0.537231   0.537231   0.431385

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.550108  0.977626
 0.462843  0.712621
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.00253792   0.00451028   0.00451028   …  0.448747   0.448747   0.797491
 0.00213533   0.00328768   0.0037948       0.327105   0.377561   0.581316
 0.00213533   0.0037948    0.00328768      0.377561   0.327105   0.581316
 0.00179659   0.00276615   0.00276615      0.275216   0.275216   0.423739
 0.00213533   0.0037948    0.0037948       0.327105   0.327105   0.581316
 0.00179659   0.00276615   0.00319282   …  0.238437   0.275216   0.423739
 0.00179659   0.00319282   0.00276615      0.275216   0.238437   0.423739
 0.00151159   0.00232734   0.00232734      0.200613   0.200613   0.308876
 0.00213533   0.0037948    0.0037948       0.327105   0.327105   0.581316
 0.00179659   0.00276615   0.00319282      0.238437   0.275216   0.423739
 ⋮                                      ⋱
 0.000536227  0.000825609  0.000825609  …  0.0300937  0.0300937  0.0463341
 0.000757492  0.00134618   0.00134618      0.0490686  0.0490686  0.0872024
 0.000637329  0.00098127   0.00113263      0.0357676  0.0412847  0.0635645
 0.000637329  0.00113263   0.00098127      0.0412847  0.0357676  0.0635645
 0.000536227  0.000825609  0.000825609     0.0300937  0.0300937  0.0463341
 0.000637329  0.00113263   0.00113263   …  0.0357676  0.0357676  0.0635645
 0.000536227  0.000825609  0.000952958     0.0260721  0.0300937  0.0463341
 0.000536227  0.000952958  0.000825609     0.0300937  0.0260721  0.0463341
 0.000451164  0.00069464   0.00069464      0.0219362  0.0219362  0.0337743
julia> sum(P10)  # expected number of edges20834.239362892662
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 20403 stored entries:
  [292 ,    2]  =  1
  [545 ,    3]  =  1
  [130 ,    4]  =  1
  [283 ,    4]  =  1
  [294 ,    4]  =  1
  [379 ,    4]  =  1
  [442 ,    4]  =  1
  [596 ,    4]  =  1
  [708 ,    4]  =  1
  ⋮
  [941 , 1024]  =  1
  [947 , 1024]  =  1
  [951 , 1024]  =  1
  [957 , 1024]  =  1
  [963 , 1024]  =  1
  [996 , 1024]  =  1
  [1004, 1024]  =  1
  [1010, 1024]  =  1
  [1013, 1024]  =  1
  [1016, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.