Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.791491   0.3736
 0.0852226  0.172771
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.626458    0.295701   0.295701   0.139577
 0.0674529   0.136747   0.0318392  0.0645474
 0.0674529   0.0318392  0.136747   0.0645474
 0.00726289  0.014724   0.014724   0.0298499
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  2.71224   -5.86495  -5.86495   12.6824
 -1.33786   12.4252    2.893    -26.8682
 -1.33786    2.893    12.4252   -26.8682
  0.659926  -6.12896  -6.12896   56.9217
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0604437    0.0285307    0.0285307    …  1.56648e-5   7.39412e-6
 0.00650819   0.013194     0.003072        1.68669e-6   3.41941e-6
 0.00650819   0.003072     0.013194        7.24418e-6   3.41941e-6
 0.00070076   0.00142065   0.00142065      7.80007e-7   1.5813e-6
 0.00650819   0.003072     0.003072        7.24418e-6   3.41941e-6
 0.00070076   0.00142065   0.000330773  …  7.80007e-7   1.5813e-6
 0.00070076   0.000330773  0.00142065      3.35007e-6   1.5813e-6
 7.54533e-5   0.000152966  0.000152966     3.60714e-7   7.31272e-7
 0.00650819   0.003072     0.003072        7.24418e-6   3.41941e-6
 0.00070076   0.00142065   0.000330773     7.80007e-7   1.5813e-6
 ⋮                                      ⋱               ⋮
 1.36317e-12  2.76354e-12  2.76354e-12     7.54529e-10  1.52965e-9
 1.1758e-10   5.55e-11     5.55e-11        1.51531e-8   7.1526e-9
 1.26602e-11  2.56659e-11  5.97588e-12     1.63159e-9   3.30771e-9
 1.26602e-11  5.97588e-12  2.56659e-11  …  7.00756e-9   3.30771e-9
 1.36317e-12  2.76354e-12  2.76354e-12     7.54529e-10  1.52965e-9
 1.26602e-11  5.97588e-12  5.97588e-12     7.00756e-9   3.30771e-9
 1.36317e-12  2.76354e-12  6.43444e-13     7.54529e-10  1.52965e-9
 1.36317e-12  6.43444e-13  2.76354e-12     3.24064e-9   1.52965e-9
 1.46777e-13  2.9756e-13   2.9756e-13   …  3.48931e-10  7.07386e-10
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.00662313   0.00362446   0.00362446   …  8.73067e-6   4.7778e-6
 0.000826782  0.000620654  0.000452451     1.08987e-6   8.18152e-7
 0.000826782  0.000452451  0.000620654     1.49504e-6   8.18152e-7
 0.000103209  7.74779e-5   7.74779e-5      1.8663e-7    1.40101e-7
 0.000826782  0.000452451  0.000452451     1.49504e-6   8.18152e-7
 0.000103209  7.74779e-5   5.64807e-5   …  1.8663e-7    1.40101e-7
 0.000103209  5.64807e-5   7.74779e-5      2.56012e-7   1.40101e-7
 1.28839e-5   9.67176e-6   9.67176e-6      3.19586e-8   2.39909e-8
 0.000826782  0.000452451  0.000452451     1.49504e-6   8.18152e-7
 0.000103209  7.74779e-5   5.64807e-5      1.8663e-7    1.40101e-7
 ⋮                                      ⋱               ⋮
 7.59752e-13  5.70335e-13  5.70335e-13     2.36285e-14  1.77376e-14
 4.87546e-11  2.66806e-11  2.66806e-11     1.10535e-12  6.04898e-13
 6.08617e-12  4.5688e-12   3.33061e-12     1.37984e-13  1.03583e-13
 6.08617e-12  3.33061e-12  4.5688e-12   …  1.89281e-13  1.03583e-13
 7.59752e-13  5.70335e-13  5.70335e-13     2.36285e-14  1.77376e-14
 6.08617e-12  3.33061e-12  3.33061e-12     1.89281e-13  1.03583e-13
 7.59752e-13  5.70335e-13  4.15769e-13     2.36285e-14  1.77376e-14
 7.59752e-13  4.15769e-13  5.70335e-13     3.24126e-14  1.77376e-14
 9.48418e-14  7.11964e-14  7.11964e-14  …  4.04615e-15  3.03739e-15

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.901864  0.631761
 0.238749  0.472494
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.355966    0.249356    0.249356    0.174676    …  0.0144583    0.0101281
 0.0942345   0.186494    0.0660118   0.13064        0.00382751   0.0075748
 0.0942345   0.0660118   0.186494    0.13064        0.0108133    0.0075748
 0.0249466   0.0493702   0.0493702   0.0977055      0.0028626    0.00566519
 0.0942345   0.0660118   0.0660118   0.0462417      0.0108133    0.0075748
 0.0249466   0.0493702   0.0174752   0.0345841   …  0.0028626    0.00566519
 0.0249466   0.0174752   0.0493702   0.0345841      0.00808728   0.00566519
 0.00660407  0.0130697   0.0130697   0.0258654      0.00214093   0.00423699
 0.0942345   0.0660118   0.0660118   0.0462417      0.0108133    0.0075748
 0.0249466   0.0493702   0.0174752   0.0345841      0.0028626    0.00566519
 ⋮                                               ⋱
 2.27309e-6  4.49854e-6  4.49854e-6  8.90278e-6  …  0.000374682  0.000741509
 3.24351e-5  2.2721e-5   2.2721e-5   1.59162e-5     0.00189242   0.00132565
 8.58651e-6  1.6993e-5   6.0149e-6   1.19037e-5     0.000500979  0.000991456
 8.58651e-6  6.0149e-6   1.6993e-5   1.19037e-5     0.00141534   0.000991456
 2.27309e-6  4.49854e-6  4.49854e-6  8.90278e-6     0.000374682  0.000741509
 8.58651e-6  6.0149e-6   6.0149e-6   4.21347e-6  …  0.00141534   0.000991456
 2.27309e-6  4.49854e-6  1.59232e-6  3.15125e-6     0.000374682  0.000741509
 2.27309e-6  1.59232e-6  4.49854e-6  3.15125e-6     0.00105853   0.000741509
 6.01753e-7  1.19089e-6  1.19089e-6  2.35682e-6     0.000280224  0.000554574
julia> sum(P10)  # expected number of edges3250.176073344362
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 3225 stored entries:
  [2  ,    1]  =  1
  [19 ,    1]  =  1
  [74 ,    1]  =  1
  [913,    1]  =  1
  [17 ,    2]  =  1
  [659,    2]  =  1
  [65 ,    4]  =  1
  [68 ,    4]  =  1
  [714,    4]  =  1
  ⋮
  [977, 1021]  =  1
  [1  , 1022]  =  1
  [205, 1023]  =  1
  [217, 1023]  =  1
  [332, 1023]  =  1
  [933, 1023]  =  1
  [4  , 1024]  =  1
  [43 , 1024]  =  1
  [449, 1024]  =  1
  [838, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.