Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.111109  0.919459
 0.488357  0.456562

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.0123452  0.10216    0.10216    0.845404
 0.0542609  0.0507281  0.449024   0.41979
 0.0542609  0.449024   0.0507281  0.41979
 0.238493   0.222965   0.222965   0.208449

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  1.31397  -2.64618   -2.64618    5.32907
 -1.40548   0.319769   2.83046   -0.643974
 -1.40548   2.83046    0.319769  -0.643974
  1.50336  -0.342038  -0.342038   0.0778189

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 3.53987e-12  2.92935e-11  2.92935e-11  …  0.0441166    0.365078
 1.55588e-11  1.45458e-11  1.28754e-10     0.193906     0.181281
 1.55588e-11  1.28754e-10  1.45458e-11     0.0219064    0.181281
 6.83857e-11  6.39333e-11  6.39333e-11     0.0962851    0.0900163
 1.55588e-11  1.28754e-10  1.28754e-10     0.0219064    0.181281
 6.83857e-11  6.39333e-11  5.65912e-10  …  0.0962851    0.0900163
 6.83857e-11  5.65912e-10  6.39333e-11     0.0108777    0.0900163
 3.00576e-10  2.81006e-10  2.81006e-10     0.0478109    0.0446981
 1.55588e-11  1.28754e-10  1.28754e-10     0.0219064    0.181281
 6.83857e-11  6.39333e-11  5.65912e-10     0.0962851    0.0900163
 ⋮                                      ⋱               ⋮
 4.18664e-5   3.91406e-5   3.91406e-5      0.000176713  0.000165208
 2.16714e-6   1.79337e-5   1.79337e-5      8.09676e-5   0.000670031
 9.52525e-6   8.90509e-6   7.88242e-5      0.000355877  0.000332707
 9.52525e-6   7.88242e-5   8.90509e-6   …  4.02049e-5   0.000332707
 4.18664e-5   3.91406e-5   3.91406e-5      0.000176713  0.000165208
 9.52525e-6   7.88242e-5   7.88242e-5      4.02049e-5   0.000332707
 4.18664e-5   3.91406e-5   0.000346456     0.000176713  0.000165208
 4.18664e-5   0.000346456  3.91406e-5      1.9964e-5    0.000165208
 0.000184015  0.000172035  0.000172035  …  8.77477e-5   8.20347e-5

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 9.30226e-5  0.000105237  0.000105237  …  0.000361382  0.000408833
 5.58951e-5  0.000132561  6.32344e-5      0.000217146  0.000514986
 5.58951e-5  6.32344e-5   0.000132561     0.000455214  0.000514986
 3.35861e-5  7.96531e-5   7.96531e-5      0.000273527  0.000648701
 5.58951e-5  6.32344e-5   6.32344e-5      0.000455214  0.000514986
 3.35861e-5  7.96531e-5   3.79961e-5   …  0.000273527  0.000648701
 3.35861e-5  3.79961e-5   7.96531e-5      0.000573409  0.000648701
 2.01811e-5  4.78617e-5   4.78617e-5      0.000344548  0.000817135
 5.58951e-5  6.32344e-5   6.32344e-5      0.000455214  0.000514986
 3.35861e-5  7.96531e-5   3.79961e-5      0.000273527  0.000648701
 ⋮                                     ⋱               ⋮
 3.42949e-7  8.13341e-7   8.13341e-7      0.00218396   0.00517951
 9.49856e-7  1.07458e-6   1.07458e-6      0.00288543   0.0032643
 5.70747e-7  1.35359e-6   6.45689e-7      0.00173379   0.00411187
 5.70747e-7  6.45689e-7   1.35359e-6   …  0.00363462   0.00411187
 3.42949e-7  8.13341e-7   8.13341e-7      0.00218396   0.00517951
 5.70747e-7  6.45689e-7   6.45689e-7      0.00363462   0.00411187
 3.42949e-7  8.13341e-7   3.87979e-7      0.00218396   0.00517951
 3.42949e-7  3.87979e-7   8.13341e-7      0.00457835   0.00517951
 2.0607e-7   4.88718e-7   4.88718e-7   …  0.00275102   0.00652436

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.101125  0.690052
 0.316942  0.431591

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 1.11838e-10  7.63155e-10  7.63155e-10  …  0.00358754   0.0244805
 3.50519e-10  4.77313e-10  2.39185e-9      0.0112439    0.0153112
 3.50519e-10  2.39185e-9   4.77313e-10     0.00224382   0.0153112
 1.09858e-9   1.49598e-9   1.49598e-9      0.00703249   0.00957637
 3.50519e-10  2.39185e-9   2.39185e-9      0.00224382   0.0153112
 1.09858e-9   1.49598e-9   7.49645e-9   …  0.00703249   0.00957637
 1.09858e-9   7.49645e-9   1.49598e-9      0.00140339   0.00957637
 3.44313e-9   4.68863e-9   4.68863e-9      0.00439845   0.00598952
 3.50519e-10  2.39185e-9   2.39185e-9      0.00224382   0.0153112
 1.09858e-9   1.49598e-9   7.49645e-9      0.00703249   0.00957637
 ⋮                                      ⋱
 3.26349e-6   4.44401e-6   4.44401e-6   …  0.000263296  0.000358539
 3.32231e-7   2.26706e-6   2.26706e-6      0.000134317  0.000916547
 1.04126e-6   1.41793e-6   7.10533e-6      0.000420972  0.000573252
 1.04126e-6   7.10533e-6   1.41793e-6      8.40083e-5   0.000573252
 3.26349e-6   4.44401e-6   4.44401e-6      0.000263296  0.000358539
 1.04126e-6   7.10533e-6   7.10533e-6   …  8.40083e-5   0.000573252
 3.26349e-6   4.44401e-6   2.22693e-5      0.000263296  0.000358539
 3.26349e-6   2.22693e-5   4.44401e-6      5.25428e-5   0.000358539
 1.02283e-5   1.39282e-5   1.39282e-5      0.000164678  0.000224247

julia> sum(P10)  # expected number of edges
74.88457400695323

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 75 stored entries:
  [411,  110]  =  1
  [624,  158]  =  1
  [867,  178]  =  1
  [488,  182]  =  1
  [360,  215]  =  1
  [366,  215]  =  1
  [642,  375]  =  1
  [719,  376]  =  1
  [715,  414]  =  1
  ⋮
  [155, 1016]  =  1
  [173, 1016]  =  1
  [148, 1022]  =  1
  [188, 1022]  =  1
  [927, 1022]  =  1
  [17 , 1023]  =  1
  [526, 1024]  =  1
  [613, 1024]  =  1
  [666, 1024]  =  1
  [917, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.