Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.787402  0.331032
 0.668119  0.880242

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.620003  0.260655  0.260655  0.109582
 0.526079  0.693105  0.221169  0.291388
 0.526079  0.221169  0.693105  0.291388
 0.446383  0.588107  0.588107  0.774826

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  3.47887  -1.30829   -1.30829    0.492009
 -2.64052   3.11195    0.993019  -1.17031
 -2.64052   0.993019   3.11195   -1.17031
  2.0042   -2.36203   -2.36203    2.78373

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.0568017   0.02388     0.02388     …  4.11873e-6  4.11873e-6  1.73156e-6
 0.0481968   0.063499    0.0202624      1.09521e-5  3.49479e-6  4.60436e-6
 0.0481968   0.0202624   0.063499       3.49479e-6  1.09521e-5  4.60436e-6
 0.0408955   0.0538796   0.0538796      9.29295e-6  9.29295e-6  1.22434e-5
 0.0481968   0.0202624   0.0202624      1.09521e-5  1.09521e-5  4.60436e-6
 0.0408955   0.0538796   0.0171929   …  2.91225e-5  9.29295e-6  1.22434e-5
 0.0408955   0.0171929   0.0538796      9.29295e-6  2.91225e-5  1.22434e-5
 0.0347003   0.0457174   0.0457174      2.47107e-5  2.47107e-5  3.25562e-5
 0.0481968   0.0202624   0.0202624      1.09521e-5  1.09521e-5  4.60436e-6
 0.0408955   0.0538796   0.0171929      2.91225e-5  9.29295e-6  1.22434e-5
 ⋮                                   ⋱                          ⋮
 0.00932377  0.012284    0.012284       0.0617654   0.0617654   0.0813754
 0.0129502   0.0054444   0.0054444      0.0273751   0.0273751   0.0115088
 0.0109884   0.0144771   0.00461963     0.0727927   0.023228    0.0306028
 0.0109884   0.00461963  0.0144771   …  0.023228    0.0727927   0.0306028
 0.00932377  0.012284    0.012284       0.0617654   0.0617654   0.0813754
 0.0109884   0.00461963  0.00461963     0.0727927   0.0727927   0.0306028
 0.00932377  0.012284    0.0039198      0.193562    0.0617654   0.0813754
 0.00932377  0.0039198   0.012284       0.0617654   0.193562    0.0813754
 0.00791132  0.0104231   0.0104231   …  0.164239    0.164239    0.216384

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.125491  0.0823574  0.0823574  …  0.00122065  0.00122065  0.000801085
 0.166222  0.148589   0.109088      0.00220229  0.00161682  0.00144531
 0.166222  0.109088   0.148589      0.00161682  0.00220229  0.00144531
 0.220171  0.196816   0.196816      0.00291707  0.00291707  0.00260763
 0.166222  0.109088   0.109088      0.00220229  0.00220229  0.00144531
 0.220171  0.196816   0.144494   …  0.00397336  0.00291707  0.00260763
 0.220171  0.144494   0.196816      0.00291707  0.00397336  0.00260763
 0.291631  0.260695   0.260695      0.00526297  0.00526297  0.00470468
 0.166222  0.109088   0.109088      0.00220229  0.00220229  0.00144531
 0.220171  0.196816   0.144494      0.00397336  0.00291707  0.00260763
 ⋮                               ⋱                          ⋮
 2.76323   2.47011    2.47011       0.590883    0.590883    0.528204
 1.57497   1.03362    1.03362       0.247255    0.247255    0.162268
 2.08615   1.86485    1.36909       0.446096    0.327505    0.292764
 2.08615   1.36909    1.86485    …  0.327505    0.446096    0.292764
 2.76323   2.47011    2.47011       0.590883    0.590883    0.528204
 2.08615   1.36909    1.36909       0.446096    0.446096    0.292764
 2.76323   2.47011    1.81345       0.804846    0.590883    0.528204
 2.76323   1.81345    2.47011       0.590883    0.804846    0.528204
 3.66008   3.27183    3.27183    …  1.06607     1.06607     0.952983

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.947363  0.136673
 0.462161  0.897518

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.582324     0.08401      0.08401      …  1.57638e-8  1.57638e-8  2.27418e-9
 0.284081     0.551686     0.0409834       1.03519e-7  7.69018e-9  1.49344e-8
 0.284081     0.0409834    0.551686        7.69018e-9  1.03519e-7  1.49344e-8
 0.138586     0.269134     0.269134        5.05007e-8  5.05007e-8  9.80725e-8
 0.284081     0.0409834    0.0409834       1.03519e-7  1.03519e-7  1.49344e-8
 0.138586     0.269134     0.0199933    …  6.798e-7    5.05007e-8  9.80725e-8
 0.138586     0.0199933    0.269134        5.05007e-8  6.798e-7    9.80725e-8
 0.0676075    0.131294     0.131294        3.31633e-7  3.31633e-7  6.44033e-7
 0.284081     0.0409834    0.0409834       1.03519e-7  1.03519e-7  1.49344e-8
 0.138586     0.269134     0.0199933       6.798e-7    5.05007e-8  9.80725e-8
 ⋮                                      ⋱
 0.000911287  0.00176972   0.00176972   …  0.0265964   0.0265964   0.0516502
 0.00382915   0.000552418  0.000552418     0.00830204  0.00830204  0.00119771
 0.00186801   0.00362768   0.000269491     0.0545187   0.00405006  0.00786523
 0.00186801   0.000269491  0.00362768      0.00405006  0.0545187   0.00786523
 0.000911287  0.00176972   0.00176972      0.0265964   0.0265964   0.0516502
 0.00186801   0.000269491  0.000269491  …  0.0545187   0.0545187   0.00786523
 0.000911287  0.00176972   0.000131468     0.358019    0.0265964   0.0516502
 0.000911287  0.000131468  0.00176972      0.0265964   0.358019    0.0516502
 0.000444562  0.00086334   0.00086334      0.174656    0.174656    0.339182

julia> sum(P10)  # expected number of edges
7594.61927655387

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 7251 stored entries:
  [1   ,    1]  =  1
  [27  ,    1]  =  1
  [46  ,    1]  =  1
  [95  ,    1]  =  1
  [106 ,    1]  =  1
  [113 ,    1]  =  1
  [131 ,    1]  =  1
  [141 ,    1]  =  1
  [155 ,    1]  =  1
  ⋮
  [376 , 1020]  =  1
  [956 , 1020]  =  1
  [1020, 1020]  =  1
  [509 , 1021]  =  1
  [863 , 1021]  =  1
  [1021, 1021]  =  1
  [757 , 1022]  =  1
  [1022, 1022]  =  1
  [1023, 1023]  =  1
  [1024, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.