Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.217204   0.904296
 0.0907582  0.0419883
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0471777   0.196417    0.196417    0.817752
 0.0197131   0.00912004  0.0820723   0.0379699
 0.0197131   0.0820723   0.00912004  0.0379699
 0.00823705  0.00381078  0.00381078  0.00176302
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  0.331268  -7.13447  -7.13447  153.654
 -0.716039   1.71364  15.4212   -36.9064
 -0.716039  15.4212    1.71364  -36.9064
  1.54773   -3.70405  -3.70405    8.86461
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.10261e-8   4.59053e-8   4.59053e-8   …  0.0718267    0.299039
 4.60721e-9   2.13147e-9   1.91814e-8      0.0300126    0.013885
 4.60721e-9   1.91814e-8   2.13147e-9      0.00333506   0.013885
 1.92511e-9   8.90631e-10  8.90631e-10     0.00139354   0.000644708
 4.60721e-9   1.91814e-8   1.91814e-8      0.00333506   0.013885
 1.92511e-9   8.90631e-10  8.01489e-9   …  0.00139354   0.000644708
 1.92511e-9   8.01489e-9   8.90631e-10     0.000154854  0.000644708
 8.04401e-10  3.72147e-10  3.72147e-10     6.47051e-5   2.99351e-5
 4.60721e-9   1.91814e-8   1.91814e-8      0.00333506   0.013885
 1.92511e-9   8.90631e-10  8.01489e-9      0.00139354   0.000644708
 ⋮                                      ⋱               ⋮
 7.47501e-13  3.45823e-13  3.45823e-13     1.39791e-15  6.46728e-16
 4.28131e-12  1.78246e-11  1.78246e-11     7.20517e-14  2.99976e-13
 1.78893e-12  8.27631e-13  7.44795e-12     3.01066e-14  1.39285e-14
 1.78893e-12  7.44795e-12  8.27631e-13  …  3.34551e-15  1.39285e-14
 7.47501e-13  3.45823e-13  3.45823e-13     1.39791e-15  6.46728e-16
 1.78893e-12  7.44795e-12  7.44795e-12     3.34551e-15  1.39285e-14
 7.47501e-13  3.45823e-13  3.1121e-12      1.39791e-15  6.46728e-16
 7.47501e-13  3.1121e-12   3.45823e-13     1.55339e-16  6.46728e-16
 3.12341e-13  1.44501e-13  1.44501e-13  …  6.49078e-17  3.00289e-17
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.17353e-11  3.94156e-11  3.94156e-11  …  1.51599e-8   2.74915e-8
 3.95588e-12  1.40981e-11  7.17375e-12     2.75914e-9   9.83314e-9
 3.95588e-12  7.17375e-12  1.40981e-11     5.42238e-9   9.83314e-9
 7.19981e-13  2.5659e-12   2.5659e-12      9.86887e-10  3.51711e-9
 3.95588e-12  7.17375e-12  7.17375e-12     5.42238e-9   9.83314e-9
 7.19981e-13  2.5659e-12   1.30564e-12  …  9.86887e-10  3.51711e-9
 7.19981e-13  1.30564e-12  2.5659e-12      1.93947e-9   3.51711e-9
 1.31039e-13  4.67001e-13  4.67001e-13     3.52989e-10  1.258e-9
 3.95588e-12  7.17375e-12  7.17375e-12     5.42238e-9   9.83314e-9
 7.19981e-13  2.5659e-12   1.30564e-12     9.86887e-10  3.51711e-9
 ⋮                                      ⋱               ⋮
 1.57769e-19  5.62264e-19  5.62264e-19     9.45606e-14  3.36999e-13
 4.76284e-18  8.63712e-18  8.63712e-18     1.45258e-12  2.63416e-12
 8.6685e-19   3.08932e-18  1.57198e-18     2.64373e-13  9.42183e-13
 8.6685e-19   1.57198e-18  3.08932e-18  …  5.19556e-13  9.42183e-13
 1.57769e-19  5.62264e-19  5.62264e-19     9.45606e-14  3.36999e-13
 8.6685e-19   1.57198e-18  1.57198e-18     5.19556e-13  9.42183e-13
 1.57769e-19  5.62264e-19  2.86105e-19     9.45606e-14  3.36999e-13
 1.57769e-19  2.86105e-19  5.62264e-19     1.85834e-13  3.36999e-13
 2.87144e-20  1.02334e-19  1.02334e-19  …  3.38224e-14  1.20538e-13

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.119736  0.518944
 0.272774  0.196529
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 6.05665e-10  2.625e-9     2.625e-9     …  0.000326819  0.00141646
 1.37979e-9   9.94114e-10  5.98011e-9      0.000744537  0.000536426
 1.37979e-9   5.98011e-9   9.94114e-10     0.000123769  0.000536426
 3.14334e-9   2.26473e-9   2.26473e-9      0.000281963  0.00020315
 1.37979e-9   5.98011e-9   5.98011e-9      0.000123769  0.000536426
 3.14334e-9   2.26473e-9   1.36235e-8   …  0.000281963  0.00020315
 3.14334e-9   1.36235e-8   2.26473e-9      4.68726e-5   0.00020315
 7.16096e-9   5.15935e-9   5.15935e-9      0.000106782  7.69348e-5
 1.37979e-9   5.98011e-9   5.98011e-9      0.000123769  0.000536426
 3.14334e-9   2.26473e-9   1.36235e-8      0.000281963  0.00020315
 ⋮                                      ⋱
 1.00103e-6   7.21228e-7   7.21228e-7   …  3.15019e-7   2.26966e-7
 1.92881e-7   8.35962e-7   8.35962e-7      3.65133e-7   1.58252e-6
 4.39409e-7   3.16587e-7   1.90444e-6      8.31823e-7   5.99315e-7
 4.39409e-7   1.90444e-6   3.16587e-7      1.3828e-7    5.99315e-7
 1.00103e-6   7.21228e-7   7.21228e-7      3.15019e-7   2.26966e-7
 4.39409e-7   1.90444e-6   1.90444e-6   …  1.3828e-7    5.99315e-7
 1.00103e-6   7.21228e-7   4.33856e-6      3.15019e-7   2.26966e-7
 1.00103e-6   4.33856e-6   7.21228e-7      5.23678e-8   2.26966e-7
 2.28049e-6   1.64306e-6   1.64306e-6      1.19301e-7   8.59543e-8
julia> sum(P10)  # expected number of edges2.7882305867584773
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 3 stored entries:
  [909,  112]  =  1
  [185,  712]  =  1
  [285, 1019]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.