Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.271949  0.5083
 0.508028  0.948076

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.073956  0.138232  0.138232  0.258369
 0.138158  0.257828  0.258231  0.481907
 0.138158  0.258231  0.257828  0.481907
 0.258093  0.481649  0.481649  0.898847

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  5.53227e6  -2.96607e6  -2.96607e6   1.59022e6
 -2.96448e6   1.58689e6   1.58937e6  -8.50795e5
 -2.96448e6   1.58937e6   1.58689e6  -8.50795e5
  1.58852e6  -8.5034e5   -8.5034e5    4.55189e5

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.63622e-7   3.05827e-7   3.05827e-7   …  0.000159151  0.00029747
 3.05663e-7   5.70424e-7   5.71316e-7      0.000297311  0.000554838
 3.05663e-7   5.71316e-7   5.70424e-7      0.000296847  0.000554838
 5.7101e-7    1.06561e-6   1.06561e-6      0.000554541  0.00103488
 3.05663e-7   5.71316e-7   5.71316e-7      0.000296847  0.000554838
 5.7101e-7    1.06561e-6   1.06728e-6   …  0.000554541  0.00103488
 5.7101e-7    1.06728e-6   1.06561e-6      0.000553675  0.00103488
 1.06671e-6   1.99067e-6   1.99067e-6      0.00103432   0.00193024
 3.05663e-7   5.71316e-7   5.71316e-7      0.000296847  0.000554838
 5.7101e-7    1.06561e-6   1.06728e-6      0.000554541  0.00103488
 ⋮                                      ⋱               ⋮
 0.000158217  0.000295263  0.000295263     0.151509     0.282744
 4.53369e-5   8.47394e-5   8.47394e-5      0.0434825    0.0812732
 8.46941e-5   0.000158055  0.000158302     0.0812298    0.15159
 8.46941e-5   0.000158302  0.000158055  …  0.081103     0.15159
 0.000158217  0.000295263  0.000295263     0.151509     0.282744
 8.46941e-5   0.000158302  0.000158302     0.081103     0.15159
 0.000158217  0.000295263  0.000295725     0.151509     0.282744
 0.000158217  0.000295725  0.000295263     0.151272     0.282744
 0.000295567  0.000551582  0.000551582  …  0.282592     0.52737

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.80547e-6   3.37052e-6   3.37052e-6   …  0.00173284  0.00173284  0.00323493
 3.36872e-6   6.28884e-6   6.28884e-6      0.0032332   0.0032332   0.00603585
 3.36872e-6   6.28884e-6   6.28884e-6      0.0032332   0.0032332   0.00603585
 6.28548e-6   1.1734e-5    1.1734e-5       0.00603262  0.00603262  0.0112619
 3.36872e-6   6.28884e-6   6.28884e-6      0.0032332   0.0032332   0.00603585
 6.28548e-6   1.1734e-5    1.1734e-5    …  0.00603263  0.00603262  0.0112619
 6.28548e-6   1.1734e-5    1.1734e-5       0.00603262  0.00603263  0.0112619
 1.17277e-5   2.18937e-5   2.18937e-5      0.0112559   0.0112559   0.0210129
 3.36872e-6   6.28884e-6   6.28884e-6      0.0032332   0.0032332   0.00603585
 6.28548e-6   1.1734e-5    1.1734e-5       0.00603263  0.00603262  0.0112619
 ⋮                                      ⋱                          ⋮
 0.00172267   0.00321595   0.00321595      1.65338     1.65338     3.08659
 0.000494829  0.000923765  0.000923765     0.474924    0.474924    0.886606
 0.00092327   0.0017236    0.00172359      0.886132    0.886132    1.65426
 0.00092327   0.00172359   0.0017236    …  0.886132    0.886132    1.65426
 0.00172267   0.00321595   0.00321595      1.65338     1.65338     3.08659
 0.00092327   0.00172359   0.00172359      0.886132    0.886132    1.65426
 0.00172267   0.00321595   0.00321595      1.65338     1.65338     3.08659
 0.00172267   0.00321595   0.00321595      1.65338     1.65338     3.08659
 0.00321423   0.00600044   0.00600044   …  3.08494     3.08494     5.75907

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.490373  0.502057
 0.687462  0.782026

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.000804023  0.00082318  0.00082318  0.000842793  …  0.000993814  0.00101749
 0.00112717   0.00128222  0.00115403  0.00131277      0.00139324   0.00158489
 0.00112717   0.00115403  0.00128222  0.00131277      0.00154801   0.00158489
 0.0015802    0.00179757  0.00179757  0.00204483      0.00217018   0.0024687
 0.00112717   0.00115403  0.00115403  0.00118153      0.00154801   0.00158489
 0.0015802    0.00179757  0.00161785  0.0018404    …  0.00217018   0.0024687
 0.0015802    0.00161785  0.00179757  0.0018404       0.00241125   0.0024687
 0.00221531   0.00252004  0.00252004  0.00286668      0.00338036   0.00384535
 0.00112717   0.00115403  0.00115403  0.00118153      0.00154801   0.00158489
 0.0015802    0.00179757  0.00161785  0.0018404       0.00217018   0.0024687
 ⋮                                                 ⋱
 0.0168177    0.0191311   0.0191311   0.0217627    …  0.0482806    0.0549218
 0.00855704   0.00876092  0.00876092  0.00896966      0.0221097    0.0226365
 0.0119963    0.0136464   0.0122821   0.0139715       0.0309959    0.0352596
 0.0119963    0.0122821   0.0136464   0.0139715       0.034439     0.0352596
 0.0168177    0.0191311   0.0191311   0.0217627       0.0482806    0.0549218
 0.0119963    0.0122821   0.0122821   0.0125747    …  0.034439     0.0352596
 0.0168177    0.0191311   0.0172184   0.0195869       0.0482806    0.0549218
 0.0168177    0.0172184   0.0191311   0.0195869       0.0536437    0.0549218
 0.0235771    0.0268202   0.0268202   0.0305095       0.075204     0.0855487

julia> sum(P10)  # expected number of edges
8179.709106628832

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 8129 stored entries:
  [188 ,    1]  =  true
  [367 ,    1]  =  true
  [477 ,    1]  =  true
  [620 ,    1]  =  true
  [761 ,    1]  =  true
  [844 ,    1]  =  true
  [955 ,    1]  =  true
  [1001,    1]  =  true
  [37  ,    2]  =  true
  ⋮
  [703 , 1024]  =  true
  [719 , 1024]  =  true
  [742 , 1024]  =  true
  [778 , 1024]  =  true
  [784 , 1024]  =  true
  [799 , 1024]  =  true
  [841 , 1024]  =  true
  [860 , 1024]  =  true
  [876 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source