Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.721065  0.902065
 0.137618  0.203185

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.519935   0.650447   0.650447   0.81372
 0.0992316  0.146509   0.12414    0.183286
 0.0992316  0.12414    0.146509   0.183286
 0.0189387  0.0279619  0.0279619  0.041284

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  82.5068  -366.3    -366.3     1626.23
 -55.8823   292.801   248.097  -1299.93
 -55.8823   248.097   292.801  -1299.93
  37.8494  -198.316  -198.316   1039.1

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.0197557    0.0247147    0.0247147    …  0.232053    0.232053    0.290303
 0.00377045   0.00556684   0.0047169       0.0522687   0.0442883   0.065389
 0.00377045   0.0047169    0.00556684      0.0442883   0.0522687   0.065389
 0.000719606  0.00106245   0.00106245      0.00997569  0.00997569  0.0147285
 0.00377045   0.0047169    0.0047169       0.0522687   0.0522687   0.065389
 0.000719606  0.00106245   0.000900239  …  0.0117732   0.00997569  0.0147285
 0.000719606  0.000900239  0.00106245      0.00997569  0.0117732   0.0147285
 0.00013734   0.000202774  0.000202774     0.00224696  0.00224696  0.00331751
 0.00377045   0.0047169    0.0047169       0.0522687   0.0522687   0.065389
 0.000719606  0.00106245   0.000900239     0.0117732   0.00997569  0.0147285
 ⋮                                      ⋱                          ⋮
 2.41772e-10  3.56961e-10  3.56961e-10     1.48876e-8  1.48876e-8  2.19806e-8
 6.63748e-9   8.3036e-9    8.3036e-9       3.46313e-7  3.46313e-7  4.33244e-7
 1.26679e-9   1.87034e-9   1.58477e-9      7.80051e-8  6.60953e-8  9.75857e-8
 1.26679e-9   1.58477e-9   1.87034e-9   …  6.60953e-8  7.80051e-8  9.75857e-8
 2.41772e-10  3.56961e-10  3.56961e-10     1.48876e-8  1.48876e-8  2.19806e-8
 1.26679e-9   1.58477e-9   1.58477e-9      7.80051e-8  7.80051e-8  9.75857e-8
 2.41772e-10  3.56961e-10  3.0246e-10      1.75702e-8  1.48876e-8  2.19806e-8
 2.41772e-10  3.0246e-10   3.56961e-10     1.48876e-8  1.75702e-8  2.19806e-8
 4.61431e-11  6.81274e-11  6.81274e-11  …  3.35334e-9  3.35334e-9  4.951e-9

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.00509609   0.00659671   0.00659671   …  0.0871429    0.112804
 0.00100639   0.00130888   0.00130273      0.0172092    0.0223818
 0.00100639   0.00130273   0.00130888      0.0172904    0.0223818
 0.000198744  0.000258481  0.000258481     0.00341455   0.00444087
 0.00100639   0.00130273   0.00130273      0.0172904    0.0223818
 0.000198744  0.000258481  0.000257267  …  0.00341455   0.00444087
 0.000198744  0.000257267  0.000258481     0.00343066   0.00444087
 3.92484e-5   5.10455e-5   5.10455e-5      0.000677495  0.000881132
 0.00100639   0.00130273   0.00130273      0.0172904    0.0223818
 0.000198744  0.000258481  0.000257267     0.00341455   0.00444087
 ⋮                                      ⋱               ⋮
 9.07924e-11  1.18082e-10  1.18082e-10     1.62738e-9   2.11653e-9
 2.32805e-9   3.01358e-9   3.01358e-9      4.15324e-8   5.37623e-8
 4.59749e-10  5.97938e-10  5.9513e-10      8.20192e-9   1.06672e-8
 4.59749e-10  5.9513e-10   5.97938e-10  …  8.24063e-9   1.06672e-8
 9.07924e-11  1.18082e-10  1.18082e-10     1.62738e-9   2.11653e-9
 4.59749e-10  5.9513e-10   5.9513e-10      8.24063e-9   1.06672e-8
 9.07924e-11  1.18082e-10  1.17528e-10     1.62738e-9   2.11653e-9
 9.07924e-11  1.17528e-10  1.18082e-10     1.63506e-9   2.11653e-9
 1.79299e-11  2.33192e-11  2.33192e-11  …  3.22895e-10  4.19949e-10

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.747713  0.290982
 0.735744  0.681465

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 0.0546196  0.0212559  0.0212559  …  1.11822e-5  1.11822e-5  4.35168e-6
 0.0537453  0.0497803  0.0209157     2.61881e-5  1.10032e-5  1.01914e-5
 0.0537453  0.0209157  0.0497803     1.10032e-5  2.61881e-5  1.01914e-5
 0.0528851  0.0489835  0.0489835     2.57689e-5  2.57689e-5  2.38678e-5
 0.0537453  0.0209157  0.0209157     2.61881e-5  2.61881e-5  1.01914e-5
 0.0528851  0.0489835  0.0205809  …  6.13312e-5  2.57689e-5  2.38678e-5
 0.0528851  0.0205809  0.0489835     2.57689e-5  6.13312e-5  2.38678e-5
 0.0520385  0.0481994  0.0481994     6.03495e-5  6.03495e-5  5.58972e-5
 0.0537453  0.0209157  0.0209157     2.61881e-5  2.61881e-5  1.01914e-5
 0.0528851  0.0489835  0.0205809     6.13312e-5  2.57689e-5  2.38678e-5
 ⋮                                ⋱
 0.0472365  0.0437517  0.0437517  …  0.0099573   0.0099573   0.0092227
 0.0487858  0.0189856  0.0189856     0.00432087  0.00432087  0.00168152
 0.0480049  0.0444634  0.0186817     0.0101193   0.00425171  0.00393804
 0.0480049  0.0186817  0.0444634     0.00425171  0.0101193   0.00393804
 0.0472365  0.0437517  0.0437517     0.0099573   0.0099573   0.0092227
 0.0480049  0.0186817  0.0186817  …  0.0101193   0.0101193   0.00393804
 0.0472365  0.0437517  0.0183827     0.0236989   0.0099573   0.0092227
 0.0472365  0.0183827  0.0437517     0.0099573   0.0236989   0.0092227
 0.0464804  0.0430514  0.0430514     0.0233195   0.0233195   0.0215991

julia> sum(P10)  # expected number of edges
7982.0444387616735

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 7899 stored entries:
  [32  ,    1]  =  true
  [33  ,    1]  =  true
  [48  ,    1]  =  true
  [55  ,    1]  =  true
  [78  ,    1]  =  true
  [93  ,    1]  =  true
  [135 ,    1]  =  true
  [163 ,    1]  =  true
  [186 ,    1]  =  true
  ⋮
  [765 , 1017]  =  true
  [934 , 1017]  =  true
  [935 , 1017]  =  true
  [987 , 1017]  =  true
  [425 , 1018]  =  true
  [492 , 1019]  =  true
  [937 , 1019]  =  true
  [512 , 1020]  =  true
  [704 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source