Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.880499  0.594185
 0.247715  0.740498
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.775278   0.523179  0.523179  0.353056
 0.218112   0.652008  0.147188  0.439993
 0.218112   0.147188  0.652008  0.439993
 0.0613626  0.183432  0.183432  0.548337
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  2.15167   -1.72653   -1.72653    1.38539
 -0.719787   2.55847    0.577566  -2.05295
 -0.719787   0.577566   2.55847   -2.05295
  0.240786  -0.855872  -0.855872   3.04218
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.217142    0.146534    0.146534    0.0988849   …  0.0028699    0.00193669
 0.0610896   0.182616    0.041225    0.123234       0.000807401  0.00241358
 0.0610896   0.041225    0.182616    0.123234       0.00357658   0.00241358
 0.0171866   0.0513763   0.0513763   0.15358        0.00100622   0.0030079
 0.0610896   0.041225    0.041225    0.0278197      0.00357658   0.00241358
 0.0171866   0.0513763   0.011598    0.0346701   …  0.00100622   0.0030079
 0.0171866   0.011598    0.0513763   0.0346701      0.00445729   0.0030079
 0.00483519  0.0144539   0.0144539   0.0432074      0.00125399   0.00374857
 0.0610896   0.041225    0.041225    0.0278197      0.00357658   0.00241358
 0.0171866   0.0513763   0.011598    0.0346701      0.00100622   0.0030079
 ⋮                                               ⋱               ⋮
 1.89758e-7  5.67246e-7  5.67246e-7  1.69568e-6     0.00729647   0.0218115
 2.39747e-6  1.61788e-6  1.61788e-6  1.09179e-6     0.0208107    0.0140437
 6.74491e-7  2.01627e-6  4.55165e-7  1.36063e-6     0.00585478   0.0175018
 6.74491e-7  4.55165e-7  2.01627e-6  1.36063e-6  …  0.0259352    0.0175018
 1.89758e-7  5.67246e-7  5.67246e-7  1.69568e-6     0.00729647   0.0218115
 6.74491e-7  4.55165e-7  4.55165e-7  3.07158e-7     0.0259352    0.0175018
 1.89758e-7  5.67246e-7  1.28054e-7  3.82793e-7     0.00729647   0.0218115
 1.89758e-7  1.28054e-7  5.67246e-7  3.82793e-7     0.0323216    0.0218115
 5.33854e-8  1.59586e-7  1.59586e-7  4.77053e-7  …  0.00909317   0.0271824
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.379669     0.396423     0.396423     …  0.610508    0.610508    0.637448
 0.165268     0.286265     0.172561        0.44086     0.265751    0.460314
 0.165268     0.172561     0.286265        0.265751    0.44086     0.460314
 0.0719403    0.12461      0.12461         0.191904    0.191904    0.332402
 0.165268     0.172561     0.172561        0.44086     0.44086     0.460314
 0.0719403    0.12461      0.0751148    …  0.318354    0.191904    0.332402
 0.0719403    0.0751148    0.12461         0.191904    0.318354    0.332402
 0.0313153    0.0542419    0.0542419       0.138578    0.138578    0.240034
 0.165268     0.172561     0.172561        0.44086     0.44086     0.460314
 0.0719403    0.12461      0.0751148       0.318354    0.191904    0.332402
 ⋮                                      ⋱                          ⋮
 4.03668e-5   6.99204e-5   6.99204e-5      0.0102464   0.0102464   0.017748
 0.000213038  0.000222439  0.000222439     0.032597    0.032597    0.0340354
 9.27344e-5   0.000160628  9.68265e-5      0.0235389   0.0141893   0.0245776
 9.27344e-5   9.68265e-5   0.000160628  …  0.0141893   0.0235389   0.0245776
 4.03668e-5   6.99204e-5   6.99204e-5      0.0102464   0.0102464   0.017748
 9.27344e-5   9.68265e-5   9.68265e-5      0.0235389   0.0235389   0.0245776
 4.03668e-5   6.99204e-5   4.21481e-5      0.0169979   0.0102464   0.017748
 4.03668e-5   4.21481e-5   6.99204e-5      0.0102464   0.0169979   0.017748
 1.75715e-5   3.0436e-5    3.0436e-5    …  0.00739912  0.00739912  0.0128162

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.824866  0.134601
 0.542107  0.695195
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.145825    0.0237956    0.0237956    …  1.19621e-8   1.95197e-9
 0.0958373   0.122901     0.0156386       7.86158e-9   1.00817e-8
 0.0958373   0.0156386    0.122901        6.17829e-8   1.00817e-8
 0.0629848   0.0807715    0.0807715       4.06041e-8   5.20705e-8
 0.0958373   0.0156386    0.0156386       6.17829e-8   1.00817e-8
 0.0629848   0.0807715    0.0102778    …  4.06041e-8   5.20705e-8
 0.0629848   0.0102778    0.0807715       3.191e-7     5.20705e-8
 0.041394    0.0530835    0.0530835       2.09715e-7   2.68937e-7
 0.0958373   0.0156386    0.0156386       6.17829e-8   1.00817e-8
 0.0629848   0.0807715    0.0102778       4.06041e-8   5.20705e-8
 ⋮                                     ⋱
 0.00333538  0.00427728   0.00427728   …  0.00398094   0.00510514
 0.00772223  0.00126011   0.00126011      0.0011728    0.000191377
 0.0050751   0.00650828   0.000828149     0.000770772  0.000988435
 0.0050751   0.000828149  0.00650828      0.00605737   0.000988435
 0.00333538  0.00427728   0.00427728      0.00398094   0.00510514
 0.0050751   0.000828149  0.000828149  …  0.00605737   0.000988435
 0.00333538  0.00427728   0.000544264     0.00398094   0.00510514
 0.00333538  0.000544264  0.00427728      0.0312855    0.00510514
 0.00219203  0.00281105   0.00281105      0.020561     0.0263674
julia> sum(P10)  # expected number of edges2617.233652479691
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 2597 stored entries:
  [29  ,    1]  =  1
  [31  ,    1]  =  1
  [39  ,    1]  =  1
  [51  ,    1]  =  1
  [59  ,    1]  =  1
  [73  ,    1]  =  1
  [77  ,    1]  =  1
  [129 ,    1]  =  1
  [194 ,    1]  =  1
  ⋮
  [960 , 1010]  =  1
  [1011, 1012]  =  1
  [984 , 1016]  =  1
  [766 , 1018]  =  1
  [1010, 1018]  =  1
  [1024, 1018]  =  1
  [1005, 1021]  =  1
  [1022, 1021]  =  1
  [1014, 1022]  =  1
  [1022, 1022]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.