Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.980117   0.825258
 0.0149316  0.76819
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.96063      0.80885    0.80885    0.681051
 0.0146347    0.752916   0.0123224  0.633955
 0.0146347    0.0123224  0.752916   0.633955
 0.000222953  0.0114703  0.0114703  0.590115
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  1.07591      -1.15584    -1.15584     1.24171
 -0.0209129     1.37273     0.0224666  -1.47471
 -0.0209129     0.0224666   1.37273    -1.47471
  0.000406493  -0.0266824  -0.0266824   1.75144
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.785845     0.661681     …  0.118513     0.118513     0.0997875
 0.011972     0.615924        0.110317     0.00180548   0.092887
 0.011972     0.0100804       0.00180548   0.110317     0.092887
 0.000182387  0.0093833       0.00168063   0.00168063   0.0864637
 0.011972     0.0100804       0.110317     0.110317     0.092887
 0.000182387  0.0093833    …  0.102689     0.00168063   0.0864637
 0.000182387  0.00015357      0.00168063   0.102689     0.0864637
 2.77858e-6   0.00014295      0.00156441   0.00156441   0.0804846
 0.011972     0.0100804       0.110317     0.110317     0.092887
 0.000182387  0.0093833       0.102689     0.00168063   0.0864637
 ⋮                         ⋱                            ⋮
 8.06213e-21  4.14774e-19     0.000881821  0.000881821  0.0453672
 3.47371e-17  2.92486e-17     0.0621832    0.0621832    0.0523582
 5.29202e-19  2.7226e-17      0.0578831    0.000947331  0.0487376
 5.29202e-19  4.45588e-19  …  0.000947331  0.0578831    0.0487376
 8.06213e-21  4.14774e-19     0.000881821  0.000881821  0.0453672
 5.29202e-19  4.45588e-19     0.0578831    0.0578831    0.0487376
 8.06213e-21  4.14774e-19     0.0538804    0.000881821  0.0453672
 8.06213e-21  6.78831e-21     0.000881821  0.0538804    0.0453672
 1.22823e-22  6.31888e-21  …  0.000820841  0.000820841  0.04223
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.719613     1.06712      1.06712      …  54.8748       81.3745
 0.0193078    0.445574     0.0286317        1.47233      33.9776
 0.0193078    0.0286317    0.445574        22.9128       33.9776
 0.000518041  0.0119551    0.0119551        0.614766     14.1872
 0.0193078    0.0286317    0.0286317       22.9128       33.9776
 0.000518041  0.0119551    0.00076821   …   0.614766     14.1872
 0.000518041  0.00076821   0.0119551        9.56714      14.1872
 1.38994e-5   0.000320763  0.000320763      0.256694      5.92383
 0.0193078    0.0286317    0.0286317       22.9128       33.9776
 0.000518041  0.0119551    0.00076821       0.614766     14.1872
 ⋮                                      ⋱                 ⋮
 3.733e-18    8.61481e-17  8.61481e-17      0.000237166   0.0054732
 5.18552e-15  7.68968e-15  7.68968e-15      0.0211697     0.0313929
 1.39131e-16  3.2108e-15   2.0632e-16       0.000568      0.013108
 1.39131e-16  2.0632e-16   3.2108e-15   …   0.00883935    0.013108
 3.733e-18    8.61481e-17  8.61481e-17      0.000237166   0.0054732
 1.39131e-16  2.0632e-16   2.0632e-16       0.00883935    0.013108
 3.733e-18    8.61481e-17  5.53571e-18      0.000237166   0.0054732
 3.733e-18    5.53571e-18  8.61481e-17      0.00369084    0.0054732
 1.00159e-19  2.31142e-18  2.31142e-18  …   9.9028e-5     0.00228531

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.969471  0.0343839
 0.805322  0.435111
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.733415  0.0260118   0.0260118   0.000922551  …  6.51227e-14  2.30968e-15
 0.609235  0.329166    0.0216075   0.0116744       5.40962e-14  2.92279e-14
 0.609235  0.0216075   0.329166    0.0116744       8.24095e-13  2.92279e-14
 0.50608   0.273432    0.273432    0.147734        6.8456e-13   3.69864e-13
 0.609235  0.0216075   0.0216075   0.000766346     8.24095e-13  2.92279e-14
 0.50608   0.273432    0.017949    0.00969773   …  6.8456e-13   3.69864e-13
 0.50608   0.017949    0.273432    0.00969773      1.04285e-11  3.69864e-13
 0.420391  0.227135    0.227135    0.12272         8.66277e-12  4.68045e-12
 0.609235  0.0216075   0.0216075   0.000766346     8.24095e-13  2.92279e-14
 0.50608   0.273432    0.017949    0.00969773      6.8456e-13   3.69864e-13
 ⋮                                              ⋱
 0.138122  0.0746264   0.0746264   0.0403202    …  3.55735e-5   1.92202e-5
 0.200167  0.00709925  0.00709925  0.000251787     3.38413e-6   1.20024e-7
 0.166275  0.0898375   0.00589722  0.00318624      2.81114e-6   1.51884e-6
 0.166275  0.00589722  0.0898375   0.00318624      4.28245e-5   1.51884e-6
 0.138122  0.0746264   0.0746264   0.0403202       3.55735e-5   1.92202e-5
 0.166275  0.00589722  0.00589722  0.000209155  …  4.28245e-5   1.51884e-6
 0.138122  0.0746264   0.00489871  0.00264675      3.55735e-5   1.92202e-5
 0.138122  0.00489871  0.0746264   0.00264675      0.000541923  1.92202e-5
 0.114735  0.0619908   0.0619908   0.0334933       0.000450165  0.000243222
julia> sum(P10)  # expected number of edges3241.806322545346
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 3091 stored entries:
  [2   ,   1]  =  1
  [3   ,   1]  =  1
  [4   ,   1]  =  1
  [6   ,   1]  =  1
  [10  ,   1]  =  1
  [13  ,   1]  =  1
  [16  ,   1]  =  1
  [18  ,   1]  =  1
  [21  ,   1]  =  1
  ⋮
  [910 , 910]  =  1
  [921 , 921]  =  1
  [1014, 930]  =  1
  [933 , 933]  =  1
  [1006, 961]  =  1
  [979 , 963]  =  1
  [986 , 970]  =  1
  [945 , 977]  =  1
  [980 , 977]  =  1
  [867 , 993]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.