Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.281237  0.463544
 0.923525  0.0578416
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0790945  0.130366   0.130366   0.214873
 0.25973    0.0162672  0.428094   0.0268121
 0.25973    0.428094   0.0162672  0.0268121
 0.852898   0.0534182  0.0534182  0.00334565
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  0.0197265  -0.158089   -0.158089    1.26693
 -0.314963    0.0959144   2.52412    -0.76866
 -0.314963    2.52412     0.0959144  -0.76866
  5.02884    -1.53141    -1.53141     0.466355
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.44837e-7  4.03547e-7  4.03547e-7  6.65138e-7   …  5.97135e-5   9.84215e-5
 8.03993e-7  5.03551e-8  1.32516e-6  8.29968e-8      0.000196087  1.22812e-5
 8.03993e-7  1.32516e-6  5.03551e-8  8.29968e-8      7.45112e-6   1.22812e-5
 2.64014e-6  1.65356e-7  1.65356e-7  1.03565e-8      2.44679e-5   1.53246e-6
 8.03993e-7  1.32516e-6  1.32516e-6  2.18418e-6      7.45112e-6   1.22812e-5
 2.64014e-6  1.65356e-7  4.35156e-6  2.72544e-7   …  2.44679e-5   1.53246e-6
 2.64014e-6  4.35156e-6  1.65356e-7  2.72544e-7      9.29761e-7   1.53246e-6
 8.66968e-6  5.42994e-7  5.42994e-7  3.40084e-8      3.05314e-6   1.91222e-7
 8.03993e-7  1.32516e-6  1.32516e-6  2.18418e-6      7.45112e-6   1.22812e-5
 2.64014e-6  1.65356e-7  4.35156e-6  2.72544e-7      2.44679e-5   1.53246e-6
 ⋮                                                ⋱               ⋮
 0.117222    0.00734175  0.00734175  0.000459824     1.7945e-13   1.12392e-14
 0.0108707   0.0179174   0.0179174   0.029532        4.37944e-13  7.21832e-13
 0.035697    0.00223575  0.0588369   0.00368504      1.43812e-12  9.00711e-14
 0.035697    0.0588369   0.00223575  0.00368504   …  5.46472e-14  9.00711e-14
 0.117222    0.00734175  0.00734175  0.000459824     1.7945e-13   1.12392e-14
 0.035697    0.0588369   0.0588369   0.0969768       5.46472e-14  9.00711e-14
 0.117222    0.00734175  0.193208    0.0121009       1.7945e-13   1.12392e-14
 0.117222    0.193208    0.00734175  0.0121009       6.81895e-15  1.12392e-14
 0.384932    0.0241088   0.0241088   0.00150997   …  2.2392e-14   1.40244e-15
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000289758  8.97961e-5   8.97961e-5   …  7.33632e-10  2.27353e-10
 0.000178902  0.000246483  5.54418e-5      4.52958e-10  6.24064e-10
 0.000178902  5.54418e-5   0.000246483     2.01376e-9   6.24064e-10
 0.000110458  0.000152183  0.000152183     1.24333e-9   1.713e-9
 0.000178902  5.54418e-5   5.54418e-5      2.01376e-9   6.24064e-10
 0.000110458  0.000152183  3.42308e-5   …  1.24333e-9   1.713e-9
 0.000110458  3.42308e-5   0.000152183     5.52759e-9   1.713e-9
 6.81986e-5   9.39607e-5   9.39607e-5      3.41284e-9   4.70204e-9
 0.000178902  5.54418e-5   5.54418e-5      2.01376e-9   6.24064e-10
 0.000110458  0.000152183  3.42308e-5      1.24333e-9   1.713e-9
 ⋮                                      ⋱               ⋮
 1.44017e-6   1.9842e-6    1.9842e-6       1.09988e-5   1.51536e-5
 3.77793e-6   1.17078e-6   1.17078e-6      6.48988e-6   2.01121e-6
 2.33257e-6   3.2137e-6    7.22864e-7      4.00697e-6   5.52061e-6
 2.33257e-6   7.22864e-7   3.2137e-6    …  1.78141e-5   5.52061e-6
 1.44017e-6   1.9842e-6    1.9842e-6       1.09988e-5   1.51536e-5
 2.33257e-6   7.22864e-7   7.22864e-7      1.78141e-5   5.52061e-6
 1.44017e-6   1.9842e-6    4.4631e-7       1.09988e-5   1.51536e-5
 1.44017e-6   4.4631e-7    1.9842e-6       4.88983e-5   1.51536e-5
 8.89189e-7   1.22508e-6   1.22508e-6   …  3.01907e-5   4.15953e-5

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.100373  0.0851605
 0.118678  0.528473
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 1.03791e-10  8.80609e-11  8.80609e-11  …  2.36462e-11  2.00625e-11
 1.2272e-10   5.46471e-10  1.04121e-10     2.79587e-11  1.245e-10
 1.2272e-10   1.04121e-10  5.46471e-10     1.46739e-10  1.245e-10
 1.45101e-10  6.46133e-10  6.46133e-10     1.73501e-10  7.72597e-10
 1.2272e-10   1.04121e-10  1.04121e-10     1.46739e-10  1.245e-10
 1.45101e-10  6.46133e-10  1.2311e-10   …  1.73501e-10  7.72597e-10
 1.45101e-10  1.2311e-10   6.46133e-10     9.10606e-10  7.72597e-10
 1.71563e-10  7.6397e-10   7.6397e-10      1.07668e-9   4.79443e-9
 1.2272e-10   1.04121e-10  1.04121e-10     1.46739e-10  1.245e-10
 1.45101e-10  6.46133e-10  1.2311e-10      1.73501e-10  7.72597e-10
 ⋮                                      ⋱
 4.68761e-10  2.08739e-9   2.08739e-9   …  6.14881e-5   0.000273806
 3.35307e-10  2.84488e-10  2.84488e-10     8.38016e-6   7.11008e-6
 3.96458e-10  1.76542e-9   3.36371e-10     9.90847e-6   4.41224e-5
 3.96458e-10  3.36371e-10  1.76542e-9      5.2004e-5    4.41224e-5
 4.68761e-10  2.08739e-9   2.08739e-9      6.14881e-5   0.000273806
 3.96458e-10  3.36371e-10  3.36371e-10  …  5.2004e-5    4.41224e-5
 4.68761e-10  2.08739e-9   3.97717e-10     6.14881e-5   0.000273806
 4.68761e-10  3.97717e-10  2.08739e-9      0.000322716  0.000273806
 5.5425e-10   2.46807e-9   2.46807e-9      0.000381571  0.00169913
julia> sum(P10)  # expected number of edges0.1602523006173758
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 0 stored entries

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.