Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.279064  0.725276
 0.959029  0.791329
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0778767  0.202398  0.202398  0.526025
 0.26763    0.220831  0.69556   0.573932
 0.26763    0.69556   0.220831  0.573932
 0.919736   0.758908  0.758908  0.626202
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  2.77858  -2.54665   -2.54665    2.33408
 -3.36742   0.979872   3.08634   -0.898081
 -3.36742   3.08634    0.979872  -0.898081
  4.08105  -1.18753   -1.18753    0.345554
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.23071e-7  5.79753e-7  5.79753e-7  …  0.00815153  0.00815153  0.0211855
 7.66605e-7  6.32554e-7  1.99238e-6     0.00889392  0.0280135   0.0231149
 7.66605e-7  1.99238e-6  6.32554e-7     0.0280135   0.00889392  0.0231149
 2.63451e-6  2.17383e-6  2.17383e-6     0.0305648   0.0305648   0.0252201
 7.66605e-7  1.99238e-6  1.99238e-6     0.00889392  0.00889392  0.0231149
 2.63451e-6  2.17383e-6  6.84698e-6  …  0.00970392  0.0305648   0.0252201
 2.63451e-6  6.84698e-6  2.17383e-6     0.0305648   0.00970392  0.0252201
 9.05373e-6  7.47056e-6  7.47056e-6     0.0333484   0.0333484   0.027517
 7.66605e-7  1.99238e-6  1.99238e-6     0.00889392  0.00889392  0.0231149
 2.63451e-6  2.17383e-6  6.84698e-6     0.00970392  0.0305648   0.0252201
 ⋮                                   ⋱                          ⋮
 0.176138    0.145338    0.145338       0.0669743   0.0669743   0.0552629
 0.0149141   0.038761    0.038761       0.0178618   0.0178618   0.0464222
 0.0512536   0.0422912   0.133206       0.0194886   0.0613839   0.0506501
 0.0512536   0.133206    0.0422912   …  0.0613839   0.0194886   0.0506501
 0.176138    0.145338    0.145338       0.0669743   0.0669743   0.0552629
 0.0512536   0.133206    0.133206       0.0194886   0.0194886   0.0506501
 0.176138    0.145338    0.457775       0.0212635   0.0669743   0.0552629
 0.176138    0.457775    0.145338       0.0669743   0.0212635   0.0552629
 0.605313    0.499466    0.499466    …  0.0730739   0.0730739   0.060296
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0458247  0.0459962  0.0459962  …   0.0477461   0.0477461   0.0479247
 0.0608205  0.078312   0.061048       0.0812914   0.0633706   0.0815955
 0.0608205  0.061048   0.078312       0.0633706   0.0812914   0.0815955
 0.0807235  0.103939   0.103939       0.107893    0.107893    0.138923
 0.0608205  0.061048   0.061048       0.0812914   0.0812914   0.0815955
 0.0807235  0.103939   0.0810255  …   0.138405    0.107893    0.138923
 0.0807235  0.0810255  0.103939       0.107893    0.138405    0.138923
 0.10714    0.137952   0.137952       0.183697    0.183697    0.236527
 0.0608205  0.061048   0.061048       0.0812914   0.0812914   0.0815955
 0.0807235  0.103939   0.0810255      0.138405    0.107893    0.138923
 ⋮                                ⋱                           ⋮
 1.03169    1.3284     1.3284        12.9705     12.9705     16.7007
 0.585667   0.587858   0.587858       5.73984     5.73984     5.76131
 0.777322   1.00087    0.78023        9.77253     7.61816     9.80909
 0.777322   0.78023    1.00087    …   7.61816     9.77253     9.80909
 1.03169    1.3284     1.3284        12.9705     12.9705     16.7007
 0.777322   0.78023    0.78023        9.77253     9.77253     9.80909
 1.03169    1.3284     1.03555       16.6385     12.9705     16.7007
 1.03169    1.03555    1.3284        12.9705     16.6385     16.7007
 1.36931    1.76311    1.76311    …  22.0833     22.0833     28.4343

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.365833  0.764002
 0.19895   0.0785674
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 4.29365e-5  8.96683e-5  8.96683e-5  0.000187263  …  0.0324441    0.067756
 2.33501e-5  9.22118e-6  4.87641e-5  1.92574e-5      0.017644     0.00696779
 2.33501e-5  4.87641e-5  9.22118e-6  1.92574e-5      0.00333644   0.00696779
 1.26984e-5  5.01474e-6  5.01474e-6  1.98037e-6      0.00181445   0.000716544
 2.33501e-5  4.87641e-5  4.87641e-5  0.000101839     0.00333644   0.00696779
 1.26984e-5  5.01474e-6  2.65193e-5  1.04727e-5   …  0.00181445   0.000716544
 1.26984e-5  2.65193e-5  5.01474e-6  1.04727e-5      0.000343108  0.000716544
 6.90576e-6  2.72715e-6  2.72715e-6  1.07698e-6      0.000186592  7.3687e-5
 2.33501e-5  4.87641e-5  4.87641e-5  0.000101839     0.00333644   0.00696779
 1.26984e-5  5.01474e-6  2.65193e-5  1.04727e-5      0.00181445   0.000716544
 ⋮                                                ⋱
 1.78641e-7  7.0547e-8   7.0547e-8   2.78597e-8   …  2.20688e-10  8.71519e-11
 6.04029e-7  1.26145e-6  1.26145e-6  2.6344e-6       3.94611e-9   8.24103e-9
 3.28488e-7  1.29723e-7  6.86011e-7  2.70913e-7      2.14601e-9   8.4748e-10
 3.28488e-7  6.86011e-7  1.29723e-7  2.70913e-7      4.05805e-10  8.4748e-10
 1.78641e-7  7.0547e-8   7.0547e-8   2.78597e-8      2.20688e-10  8.71519e-11
 3.28488e-7  6.86011e-7  6.86011e-7  1.43266e-6   …  4.05805e-10  8.4748e-10
 1.78641e-7  7.0547e-8   3.73072e-7  1.4733e-7       2.20688e-10  8.71519e-11
 1.78641e-7  3.73072e-7  7.0547e-8   1.4733e-7       4.17316e-11  8.71519e-11
 9.71498e-8  3.83655e-8  3.83655e-8  1.51509e-8      2.26948e-11  8.96241e-12
julia> sum(P10)  # expected number of edges30.481050711193852
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 30 stored entries:
  [2  ,  55]  =  1
  [2  , 253]  =  1
  [257, 256]  =  1
  [547, 332]  =  1
  [9  , 350]  =  1
  [609, 386]  =  1
  [73 , 440]  =  1
  [49 , 462]  =  1
  [33 , 480]  =  1
  ⋮
  [581, 810]  =  1
  [146, 811]  =  1
  [145, 868]  =  1
  [137, 872]  =  1
  [257, 899]  =  1
  [5  , 939]  =  1
  [1  , 978]  =  1
  [54 , 979]  =  1
  [17 , 992]  =  1
  [513, 992]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.