Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.485842  0.360152
 0.327279  0.891145
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.236043  0.174977  0.174977  0.129709
 0.159006  0.432956  0.11787   0.320947
 0.159006  0.11787   0.432956  0.320947
 0.107112  0.291653  0.291653  0.794139
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  7.99908  -3.23279  -3.23279   1.30652
 -2.93772   4.36101   1.18726  -1.76248
 -2.93772   1.18726   4.36101  -1.76248
  1.0789   -1.60161  -1.60161   2.37757
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000172959  0.000128213  0.000128213  …  6.42446e-6  6.42446e-6  4.76241e-6
 0.000116511  0.000317245  8.63685e-5      1.58964e-5  4.32773e-6  1.17839e-5
 0.000116511  8.63685e-5   0.000317245     4.32773e-6  1.58964e-5  1.17839e-5
 7.84853e-5   0.000213707  0.000213707     1.07083e-5  1.07083e-5  2.91576e-5
 0.000116511  8.63685e-5   8.63685e-5      1.58964e-5  1.58964e-5  1.17839e-5
 7.84853e-5   0.000213707  5.81807e-5   …  3.93334e-5  1.07083e-5  2.91576e-5
 7.84853e-5   5.81807e-5   0.000213707     1.07083e-5  3.93334e-5  2.91576e-5
 5.28703e-5   0.00014396   0.00014396      2.64963e-5  2.64963e-5  7.21464e-5
 0.000116511  8.63685e-5   8.63685e-5      1.58964e-5  1.58964e-5  1.17839e-5
 7.84853e-5   0.000213707  5.81807e-5      3.93334e-5  1.07083e-5  2.91576e-5
 ⋮                                      ⋱                          ⋮
 2.2418e-6    6.10417e-6   6.10417e-6      0.0372293   0.0372293   0.101371
 4.94027e-6   3.66219e-6   3.66219e-6      0.0223357   0.0223357   0.0165573
 3.32792e-6   9.06157e-6   2.46697e-6      0.0552665   0.0150461   0.0409687
 3.32792e-6   2.46697e-6   9.06157e-6   …  0.0150461   0.0552665   0.0409687
 2.2418e-6    6.10417e-6   6.10417e-6      0.0372293   0.0372293   0.101371
 3.32792e-6   2.46697e-6   2.46697e-6      0.0552665   0.0552665   0.0409687
 2.2418e-6    6.10417e-6   1.66183e-6      0.136749    0.0372293   0.101371
 2.2418e-6    1.66183e-6   6.10417e-6      0.0372293   0.136749    0.101371
 1.51015e-6   4.11197e-6   4.11197e-6   …  0.0921187   0.0921187   0.250829
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 3.86148e-6  5.41095e-6   5.41095e-6   …  0.00015793   0.000221301
 4.91707e-6  9.95078e-6   6.8901e-6       0.000201102  0.000406975
 4.91707e-6  6.8901e-6    9.95078e-6      0.000290435  0.000406975
 6.26121e-6  1.26709e-5   1.26709e-5      0.000369828  0.00074843
 4.91707e-6  6.8901e-6    6.8901e-6       0.000290435  0.000406975
 6.26121e-6  1.26709e-5   8.77359e-6   …  0.000369828  0.00074843
 6.26121e-6  8.77359e-6   1.26709e-5      0.000534112  0.00074843
 7.97278e-6  1.61347e-5   1.61347e-5      0.000680118  0.00137637
 4.91707e-6  6.8901e-6    6.8901e-6       0.000290435  0.000406975
 6.26121e-6  1.26709e-5   8.77359e-6      0.000369828  0.00074843
 ⋮                                     ⋱               ⋮
 5.51093e-5  0.000111526  0.000111526     0.088972     0.180055
 3.39876e-5  4.76256e-5   4.76256e-5      0.0379942    0.0532399
 4.32786e-5  8.75838e-5   6.06446e-5      0.0483804    0.0979086
 4.32786e-5  6.06446e-5   8.75838e-5   …  0.0698717    0.0979086
 5.51093e-5  0.000111526  0.000111526     0.088972     0.180055
 4.32786e-5  6.06446e-5   6.06446e-5      0.0698717    0.0979086
 5.51093e-5  0.000111526  7.72225e-5      0.088972     0.180055
 5.51093e-5  7.72225e-5   0.000111526     0.128495     0.180055
 7.0174e-5   0.000142013  0.000142013  …  0.16362      0.331122

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.143322  0.12477
 0.890919  0.899773
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 3.65703e-9  3.18365e-9  3.18365e-9  …  1.05029e-9  1.05029e-9  9.14335e-10
 2.27328e-8  2.29587e-8  1.97902e-8     7.57408e-9  6.5288e-9   6.59368e-9
 2.27328e-8  1.97902e-8  2.29587e-8     6.5288e-9   7.57408e-9  6.59368e-9
 1.41312e-7  1.42716e-7  1.42716e-7     4.70821e-8  4.70821e-8  4.755e-8
 2.27328e-8  1.97902e-8  1.97902e-8     7.57408e-9  7.57408e-9  6.59368e-9
 1.41312e-7  1.42716e-7  1.2302e-7   …  5.46201e-8  4.70821e-8  4.755e-8
 1.41312e-7  1.2302e-7   1.42716e-7     4.70821e-8  5.46201e-8  4.755e-8
 8.78424e-7  8.87154e-7  8.87154e-7     3.3953e-7   3.3953e-7   3.42904e-7
 2.27328e-8  1.97902e-8  1.97902e-8     7.57408e-9  7.57408e-9  6.59368e-9
 1.41312e-7  1.42716e-7  1.2302e-7      5.46201e-8  4.70821e-8  4.755e-8
 ⋮                                   ⋱
 0.0506823   0.051186    0.051186    …  0.0477543   0.0477543   0.0482289
 0.00131161  0.00114183  0.00114183     0.00106528  0.00106528  0.00092739
 0.00815325  0.00823428  0.00709787     0.00768223  0.00662201  0.00668783
 0.00815325  0.00709787  0.00823428     0.00662201  0.00768223  0.00668783
 0.0506823   0.051186    0.051186       0.0477543   0.0477543   0.0482289
 0.00815325  0.00709787  0.00709787  …  0.00768223  0.00768223  0.00668783
 0.0506823   0.051186    0.0441218      0.0554      0.0477543   0.0482289
 0.0506823   0.0441218   0.051186       0.0477543   0.0554      0.0482289
 0.315051    0.318182    0.318182       0.344378    0.344378    0.3478
julia> sum(P10)  # expected number of edges1368.0681437281582
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1313 stored entries:
  [511 ,    1]  =  1
  [768 ,    1]  =  1
  [1024,    1]  =  1
  [1020,    2]  =  1
  [896 ,    3]  =  1
  [959 ,    3]  =  1
  [1024,    4]  =  1
  [896 ,    5]  =  1
  [1023,    5]  =  1
  ⋮
  [894 , 1019]  =  1
  [1008, 1019]  =  1
  [1016, 1019]  =  1
  [960 , 1020]  =  1
  [880 , 1021]  =  1
  [1023, 1021]  =  1
  [736 , 1022]  =  1
  [991 , 1022]  =  1
  [1022, 1022]  =  1
  [1023, 1023]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.