Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.252248  0.840928
 0.443044  0.382487
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0636291  0.212123   0.212123   0.70716
 0.111757   0.0964817  0.372568   0.321644
 0.111757   0.372568   0.0964817  0.321644
 0.196288   0.169459   0.169459   0.146296
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  1.91931  -4.21974  -4.21974   9.27744
 -2.22318   1.26577   4.88783  -2.7829
 -2.22318   4.88783   1.26577  -2.7829
  2.57516  -1.46617  -1.46617   0.834768
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 6.63643e-8  2.21241e-7  2.21241e-7  …  0.0375125   0.0375125   0.125057
 1.16561e-7  1.00629e-7  3.88583e-7     0.0170622   0.0658863   0.0568807
 1.16561e-7  3.88583e-7  1.00629e-7     0.0658863   0.0170622   0.0568807
 2.04725e-7  1.76743e-7  1.76743e-7     0.0299677   0.0299677   0.0258716
 1.16561e-7  3.88583e-7  3.88583e-7     0.0170622   0.0170622   0.0568807
 2.04725e-7  1.76743e-7  6.825e-7    …  0.00776054  0.0299677   0.0258716
 2.04725e-7  6.825e-7    1.76743e-7     0.0299677   0.00776054  0.0258716
 3.59576e-7  3.10428e-7  3.10428e-7     0.0136305   0.0136305   0.0117674
 1.16561e-7  3.88583e-7  3.88583e-7     0.0170622   0.0170622   0.0568807
 2.04725e-7  1.76743e-7  6.825e-7       0.00776054  0.0299677   0.0258716
 ⋮                                   ⋱                          ⋮
 3.25642e-5  2.81132e-5  2.81132e-5     2.49675e-5  2.49675e-5  2.15548e-5
 1.05561e-5  3.51912e-5  3.51912e-5     3.12534e-5  3.12534e-5  0.000104191
 1.85405e-5  1.60063e-5  6.18091e-5     1.42153e-5  5.4893e-5   4.739e-5
 1.85405e-5  6.18091e-5  1.60063e-5  …  5.4893e-5   1.42153e-5  4.739e-5
 3.25642e-5  2.81132e-5  2.81132e-5     2.49675e-5  2.49675e-5  2.15548e-5
 1.85405e-5  6.18091e-5  6.18091e-5     1.42153e-5  1.42153e-5  4.739e-5
 3.25642e-5  2.81132e-5  0.00010856     6.46567e-6  2.49675e-5  2.15548e-5
 3.25642e-5  0.00010856  2.81132e-5     2.49675e-5  6.46567e-6  2.15548e-5
 5.71951e-5  4.93775e-5  4.93775e-5  …  1.13562e-5  1.13562e-5  9.80399e-6
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 4.74453e-5  5.80579e-5  5.80579e-5  7.10445e-5  …  0.000437069  0.000534834
 3.05879e-5  5.6437e-5   3.74298e-5  6.9061e-5      0.000281778  0.000519902
 3.05879e-5  3.74298e-5  5.6437e-5   6.9061e-5      0.000424866  0.000519902
 1.97199e-5  3.63848e-5  3.63848e-5  6.71328e-5     0.000273911  0.000505386
 3.05879e-5  3.74298e-5  3.74298e-5  4.58023e-5     0.000424866  0.000519902
 1.97199e-5  3.63848e-5  2.4131e-5   4.45235e-5  …  0.000273911  0.000505386
 1.97199e-5  2.4131e-5   3.63848e-5  4.45235e-5     0.000413004  0.000505386
 1.27134e-5  2.34572e-5  2.34572e-5  4.32804e-5     0.000266263  0.000491276
 3.05879e-5  3.74298e-5  3.74298e-5  4.58023e-5     0.000424866  0.000519902
 1.97199e-5  3.63848e-5  2.4131e-5   4.45235e-5     0.000273911  0.000505386
 ⋮                                               ⋱               ⋮
 3.79415e-7  7.00049e-7  7.00049e-7  1.29165e-6     0.00021229   0.000391692
 9.12853e-7  1.11704e-6  1.11704e-6  1.36691e-6     0.000338743  0.000414515
 5.88515e-7  1.08586e-6  7.20156e-7  1.32874e-6     0.000218387  0.000402942
 5.88515e-7  7.20156e-7  1.08586e-6  1.32874e-6  …  0.000329286  0.000402942
 3.79415e-7  7.00049e-7  7.00049e-7  1.29165e-6     0.00021229   0.000391692
 5.88515e-7  7.20156e-7  7.20156e-7  8.81242e-7     0.000329286  0.000402942
 3.79415e-7  7.00049e-7  4.64283e-7  8.56638e-7     0.00021229   0.000391692
 3.79415e-7  4.64283e-7  7.00049e-7  8.56638e-7     0.000320093  0.000391692
 2.44608e-7  4.51321e-7  4.51321e-7  8.32722e-7  …  0.000206363  0.000380756

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.982372  0.862058
 0.680307  0.460417
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.837068   0.734549   0.734549   0.644587    …  0.258288     0.226655
 0.579681   0.392316   0.508686   0.344268       0.178868     0.121054
 0.579681   0.508686   0.392316   0.344268       0.137949     0.121054
 0.401437   0.271684   0.271684   0.18387        0.0955319    0.064654
 0.579681   0.508686   0.508686   0.446385       0.137949     0.121054
 0.401437   0.271684   0.352272   0.23841     …  0.0955319    0.064654
 0.401437   0.352272   0.271684   0.23841        0.0736775    0.064654
 0.278001   0.188145   0.188145   0.127333       0.0510227    0.0345311
 0.579681   0.508686   0.508686   0.446385       0.137949     0.121054
 0.401437   0.271684   0.352272   0.23841        0.0955319    0.064654
 ⋮                                            ⋱
 0.0306633  0.0207523  0.0207523  0.0140447   …  0.00118428   0.000801497
 0.0639383  0.0561076  0.0561076  0.0492359      0.00320192   0.00280977
 0.0442782  0.0299666  0.0388553  0.0262965      0.00221738   0.00150067
 0.0442782  0.0388553  0.0299666  0.0262965      0.00171012   0.00150067
 0.0306633  0.0207523  0.0207523  0.0140447      0.00118428   0.000801497
 0.0442782  0.0388553  0.0388553  0.0340966   …  0.00171012   0.00150067
 0.0306633  0.0207523  0.0269079  0.0182107      0.00118428   0.000801497
 0.0306633  0.0269079  0.0207523  0.0182107      0.000913359  0.000801497
 0.0212348  0.0143712  0.0143712  0.00972616     0.000632514  0.000428072
julia> sum(P10)  # expected number of edges56191.126631374864
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 53423 stored entries:
  [2  ,    1]  =  1
  [3  ,    1]  =  1
  [5  ,    1]  =  1
  [6  ,    1]  =  1
  [8  ,    1]  =  1
  [9  ,    1]  =  1
  [11 ,    1]  =  1
  [12 ,    1]  =  1
  [15 ,    1]  =  1
  ⋮
  [273, 1024]  =  1
  [513, 1024]  =  1
  [539, 1024]  =  1
  [593, 1024]  =  1
  [646, 1024]  =  1
  [673, 1024]  =  1
  [706, 1024]  =  1
  [753, 1024]  =  1
  [969, 1024]  =  1
  [993, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.