Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.605419  0.390528
 0.457441  0.141987
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.366532  0.236433   0.236433   0.152512
 0.276943  0.0859615  0.178644   0.0554498
 0.276943  0.178644   0.0859615  0.0554498
 0.209252  0.0649506  0.0649506  0.0201602
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  2.34695   -6.45519   -6.45519   17.7547
 -7.56122   10.0072    20.7968   -27.5243
 -7.56122   20.7968    10.0072   -27.5243
 24.3601   -32.2404   -32.2404    42.6698
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.00242478   0.00156411   0.00156411   …  1.95088e-5   1.25842e-5
 0.00183211   0.000568675  0.00118181      1.47404e-5   4.57533e-6
 0.00183211   0.00118181   0.000568675     7.09293e-6   4.57533e-6
 0.0013843    0.000429678  0.000429678     5.35926e-6   1.66348e-6
 0.00183211   0.00118181   0.00118181      7.09293e-6   4.57533e-6
 0.0013843    0.000429678  0.000892949  …  5.35926e-6   1.66348e-6
 0.0013843    0.000892949  0.000429678     2.57882e-6   1.66348e-6
 0.00104595   0.000324655  0.000324655     1.9485e-6    6.04803e-7
 0.00183211   0.00118181   0.00118181      7.09293e-6   4.57533e-6
 0.0013843    0.000429678  0.000892949     5.35926e-6   1.66348e-6
 ⋮                                      ⋱               ⋮
 0.000111107  3.4487e-5    3.4487e-5       5.94931e-10  1.84663e-10
 0.000194619  0.000125539  0.000125539     2.16567e-9   1.39697e-9
 0.000147049  4.56432e-5   9.48548e-5      1.63633e-9   5.07907e-10
 0.000147049  9.48548e-5   4.56432e-5   …  7.87386e-10  5.07907e-10
 0.000111107  3.4487e-5    3.4487e-5       5.94931e-10  1.84663e-10
 0.000147049  9.48548e-5   9.48548e-5      7.87386e-10  5.07907e-10
 0.000111107  3.4487e-5    7.16702e-5      5.94931e-10  1.84663e-10
 0.000111107  7.16702e-5   3.4487e-5       2.86275e-10  1.84663e-10
 8.39501e-5   2.60576e-5   2.60576e-5   …  2.16303e-10  6.71391e-11
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000689346  0.000369071  0.000369071  …  7.14219e-7  7.14219e-7  3.82387e-7
 0.000432307  0.000251377  0.000231454     4.8646e-7   4.47906e-7  2.60447e-7
 0.000432307  0.000231454  0.000251377     4.47906e-7  4.8646e-7   2.60447e-7
 0.000271111  0.000157645  0.000157645     3.05072e-7  3.05072e-7  1.77393e-7
 0.000432307  0.000231454  0.000231454     4.8646e-7   4.8646e-7   2.60447e-7
 0.000271111  0.000157645  0.000145151  …  3.31332e-7  3.05072e-7  1.77393e-7
 0.000271111  0.000145151  0.000157645     3.05072e-7  3.31332e-7  1.77393e-7
 0.000170021  9.88635e-5   9.88635e-5      2.07787e-7  2.07787e-7  1.20824e-7
 0.000432307  0.000231454  0.000231454     4.8646e-7   4.8646e-7   2.60447e-7
 0.000271111  0.000157645  0.000145151     3.31332e-7  3.05072e-7  1.77393e-7
 ⋮                                      ⋱                          ⋮
 4.06765e-6   2.36525e-6   2.36525e-6      9.62381e-9  9.62381e-9  5.59603e-9
 1.03427e-5   5.5374e-6    5.5374e-6       2.25308e-8  2.25308e-8  1.20628e-8
 6.48617e-6   3.77157e-6   3.47265e-6      1.53459e-8  1.41296e-8  8.21607e-9
 6.48617e-6   3.47265e-6   3.77157e-6   …  1.41296e-8  1.53459e-8  8.21607e-9
 4.06765e-6   2.36525e-6   2.36525e-6      9.62381e-9  9.62381e-9  5.59603e-9
 6.48617e-6   3.47265e-6   3.47265e-6      1.53459e-8  1.53459e-8  8.21607e-9
 4.06765e-6   2.36525e-6   2.17779e-6      1.04522e-8  9.62381e-9  5.59603e-9
 4.06765e-6   2.17779e-6   2.36525e-6      9.62381e-9  1.04522e-8  5.59603e-9
 2.55093e-6   1.48331e-6   1.48331e-6   …  6.55485e-9  6.55485e-9  3.8115e-9

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.0325072  0.255175
 0.145822   0.825344
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 1.31764e-15  1.03432e-14  1.03432e-14  …  1.49115e-7   1.17052e-6
 5.91072e-15  3.34543e-14  4.63979e-14     6.68904e-7   3.78596e-6
 5.91072e-15  4.63979e-14  3.34543e-14     4.823e-7     3.78596e-6
 2.65145e-14  1.5007e-13   1.5007e-13      2.16352e-6   1.22454e-5
 5.91072e-15  4.63979e-14  4.63979e-14     4.823e-7     3.78596e-6
 2.65145e-14  1.5007e-13   2.08133e-13  …  2.16352e-6   1.22454e-5
 2.65145e-14  2.08133e-13  1.5007e-13      1.55996e-6   1.22454e-5
 1.1894e-13   6.73192e-13  6.73192e-13     6.99774e-6   3.96068e-5
 5.91072e-15  4.63979e-14  4.63979e-14     4.823e-7     3.78596e-6
 2.65145e-14  1.5007e-13   2.08133e-13     2.16352e-6   1.22454e-5
 ⋮                                      ⋱
 9.69142e-10  5.48528e-9   5.48528e-9   …  0.00801197   0.0453473
 4.81616e-11  3.78058e-10  3.78058e-10     0.000552203  0.00433468
 2.16045e-10  1.2228e-9    1.69591e-9      0.00247709   0.0140202
 2.16045e-10  1.69591e-9   1.2228e-9       0.00178606   0.0140202
 9.69142e-10  5.48528e-9   5.48528e-9      0.00801197   0.0453473
 2.16045e-10  1.69591e-9   1.69591e-9   …  0.00178606   0.0140202
 9.69142e-10  5.48528e-9   7.60756e-9      0.00801197   0.0453473
 9.69142e-10  7.60756e-9   5.48528e-9      0.00577688   0.0453473
 4.34741e-9   2.46061e-8   2.46061e-8      0.0259141    0.146672
julia> sum(P10)  # expected number of edges9.993831149887372
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 10 stored entries:
  [508 ,  448]  =  1
  [959 ,  476]  =  1
  [640 ,  767]  =  1
  [1024,  768]  =  1
  [960 ,  960]  =  1
  [892 , 1004]  =  1
  [655 , 1008]  =  1
  [956 , 1022]  =  1
  [64  , 1024]  =  1
  [832 , 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.