Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.769828   0.848068
 0.0483359  0.187587
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.592635    0.652866    0.652866    0.719219
 0.0372103   0.14441     0.0409921   0.159087
 0.0372103   0.0409921   0.14441     0.159087
 0.00233636  0.00906719  0.00906719  0.0351889
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  3.29016   -14.8746   -14.8746    67.2469
 -0.847781   13.5023     3.83276  -61.0429
 -0.847781    3.83276   13.5023   -61.0429
  0.218449   -3.47916   -3.47916   55.4113
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0433236    0.0477267    0.0477267    …  0.125641     0.13841
 0.0027202    0.0105568    0.00299666      0.00788872   0.0306154
 0.0027202    0.00299666   0.0105568       0.0277909    0.0306154
 0.000170795  0.000662841  0.000662841     0.00174493   0.00677192
 0.0027202    0.00299666   0.00299666      0.0277909    0.0306154
 0.000170795  0.000662841  0.000188154  …  0.00174493   0.00677192
 0.000170795  0.000188154  0.000662841     0.00614716   0.00677192
 1.07239e-5   4.16184e-5   4.16184e-5      0.000385968  0.0014979
 0.0027202    0.00299666   0.00299666      0.0277909    0.0306154
 0.000170795  0.000662841  0.000188154     0.00174493   0.00677192
 ⋮                                      ⋱               ⋮
 2.59037e-15  1.0053e-14   1.0053e-14      2.21171e-9   8.58345e-9
 6.57066e-13  7.23846e-13  7.23846e-13     1.5925e-7    1.75435e-7
 4.12558e-14  1.6011e-13   4.54488e-14     9.999e-9     3.88052e-8
 4.12558e-14  4.54488e-14  1.6011e-13   …  3.52251e-8   3.88052e-8
 2.59037e-15  1.0053e-14   1.0053e-14      2.21171e-9   8.58345e-9
 4.12558e-14  4.54488e-14  4.54488e-14     3.52251e-8   3.88052e-8
 2.59037e-15  1.0053e-14   2.85363e-15     2.21171e-9   8.58345e-9
 2.59037e-15  2.85363e-15  1.0053e-14      7.79157e-9   8.58345e-9
 1.62644e-16  6.31205e-16  6.31205e-16  …  4.89216e-10  1.8986e-9
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.00418799   0.00536664   0.00536664   …  0.0640727    0.0821051
 0.000305874  0.000503522  0.000391957     0.0046796    0.00770346
 0.000305874  0.000391957  0.000503522     0.00601158   0.00770346
 2.23397e-5   3.67752e-5   3.67752e-5      0.000439061  0.000722773
 0.000305874  0.000391957  0.000391957     0.00601158   0.00770346
 2.23397e-5   3.67752e-5   2.8627e-5    …  0.000439061  0.000722773
 2.23397e-5   2.8627e-5    3.67752e-5      0.000564033  0.000722773
 1.6316e-6    2.68591e-6   2.68591e-6      4.11947e-5   6.78138e-5
 0.000305874  0.000391957  0.000391957     0.00601158   0.00770346
 2.23397e-5   3.67752e-5   2.8627e-5       0.000439061  0.000722773
 ⋮                                      ⋱               ⋮
 1.321e-15    2.17461e-15  2.17461e-15     2.47382e-13  4.07235e-13
 2.47646e-13  3.17343e-13  3.17343e-13     3.61008e-11  4.62608e-11
 1.8087e-14   2.97745e-14  2.31774e-14     2.63665e-12  4.3404e-12
 1.8087e-14   2.31774e-14  2.97745e-14  …  3.38713e-12  4.3404e-12
 1.321e-15    2.17461e-15  2.17461e-15     2.47382e-13  4.07235e-13
 1.8087e-14   2.31774e-14  2.31774e-14     3.38713e-12  4.3404e-12
 1.321e-15    2.17461e-15  1.69278e-15     2.47382e-13  4.07235e-13
 1.321e-15    1.69278e-15  2.17461e-15     3.17796e-13  4.07235e-13
 9.64806e-17  1.58824e-16  1.58824e-16  …  2.32105e-14  3.82086e-14

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.891768  0.582524
 0.590821  0.622352
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.318068    0.207769    0.207769    …  0.0068877   0.0068877   0.0044992
 0.210729    0.221975    0.137653       0.00735862  0.00456329  0.00480682
 0.210729    0.137653    0.221975       0.00456329  0.00735862  0.00480682
 0.139614    0.147064    0.147064       0.00487529  0.00487529  0.00513547
 0.210729    0.137653    0.137653       0.00735862  0.00735862  0.00480682
 0.139614    0.147064    0.0911988   …  0.00786174  0.00487529  0.00513547
 0.139614    0.0911988   0.147064       0.00487529  0.00786174  0.00513547
 0.0924978   0.0974342   0.0974342      0.00520862  0.00520862  0.00548659
 0.210729    0.137653    0.137653       0.00735862  0.00735862  0.00480682
 0.139614    0.147064    0.0911988      0.00786174  0.00487529  0.00513547
 ⋮                                   ⋱
 0.00782268  0.00824015  0.00824015  …  0.00774561  0.00774561  0.00815898
 0.0178216   0.0116415   0.0116415      0.0109428   0.0109428   0.00714811
 0.0118073   0.0124375   0.00771282     0.011691    0.00724993  0.00763684
 0.0118073   0.00771282  0.0124375      0.00724993  0.011691    0.00763684
 0.00782268  0.00824015  0.00824015     0.00774561  0.00774561  0.00815898
 0.0118073   0.00771282  0.00771282  …  0.011691    0.011691    0.00763684
 0.00782268  0.00824015  0.00510996     0.0124903   0.00774561  0.00815898
 0.00782268  0.00510996  0.00824015     0.00774561  0.0124903   0.00815898
 0.00518274  0.00545933  0.00545933     0.00827519  0.00827519  0.00871682
julia> sum(P10)  # expected number of edges19652.926327469417
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 19389 stored entries:
  [27 ,    1]  =  1
  [34 ,    1]  =  1
  [35 ,    1]  =  1
  [41 ,    1]  =  1
  [50 ,    1]  =  1
  [54 ,    1]  =  1
  [61 ,    1]  =  1
  [75 ,    1]  =  1
  [81 ,    1]  =  1
  ⋮
  [107, 1024]  =  1
  [141, 1024]  =  1
  [171, 1024]  =  1
  [206, 1024]  =  1
  [442, 1024]  =  1
  [474, 1024]  =  1
  [488, 1024]  =  1
  [609, 1024]  =  1
  [721, 1024]  =  1
  [911, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.