Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.315101  0.972131
 0.706432  0.158518

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.0992886  0.306319  0.306319  0.945039
 0.222597   0.049949  0.686744  0.1541
 0.222597   0.686744  0.049949  0.1541
 0.499046   0.111982  0.111982  0.0251278

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  0.0619663  -0.380017  -0.380017   2.3305
 -0.276152    0.123176   1.69354   -0.755396
 -0.276152    1.69354    0.123176  -0.755396
  1.23067    -0.548934  -0.548934   0.24485

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 9.58066e-7  2.95577e-6   2.95577e-6   …  0.2309       0.712358
 2.14791e-6  4.81974e-7   6.62661e-6      0.517659     0.116158
 2.14791e-6  6.62661e-6   4.81974e-7      0.0376509    0.116158
 4.81545e-6  1.08055e-6   1.08055e-6      0.0844105    0.018941
 2.14791e-6  6.62661e-6   6.62661e-6      0.0376509    0.116158
 4.81545e-6  1.08055e-6   1.48563e-5   …  0.0844105    0.018941
 4.81545e-6  1.48563e-5   1.08055e-6      0.00613943   0.018941
 1.07959e-5  2.4225e-6    2.4225e-6       0.0137641    0.00308856
 2.14791e-6  6.62661e-6   6.62661e-6      0.0376509    0.116158
 4.81545e-6  1.08055e-6   1.48563e-5      0.0844105    0.018941
 ⋮                                     ⋱               ⋮
 0.00689005  0.00154607   0.00154607      6.87971e-9   1.54375e-9
 0.00137082  0.00422918   0.00422918      1.8819e-8    5.80594e-8
 0.00307328  0.000689619  0.0094815       4.21908e-8   9.46727e-9
 0.00307328  0.0094815    0.000689619  …  3.06867e-9   9.46727e-9
 0.00689005  0.00154607   0.00154607      6.87971e-9   1.54375e-9
 0.00307328  0.0094815    0.0094815       3.06867e-9   9.46727e-9
 0.00689005  0.00154607   0.0212568       6.87971e-9   1.54375e-9
 0.00689005  0.0212568    0.00154607      5.00383e-10  1.54375e-9
 0.015447    0.00346617   0.00346617   …  1.12182e-9   2.51727e-10

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.0556274   0.0325838   0.0325838   0.019086    …  0.000154928  9.07491e-5
 0.0236781   0.0503791   0.0138695   0.0295096      6.59459e-5   0.000140311
 0.0236781   0.0138695   0.0503791   0.0295096      0.00023954   0.000140311
 0.0100787   0.0214441   0.0214441   0.0456259      0.000101961  0.00021694
 0.0236781   0.0138695   0.0138695   0.00812405     0.00023954   0.000140311
 0.0100787   0.0214441   0.00590361  0.0125609   …  0.000101961  0.00021694
 0.0100787   0.00590361  0.0214441   0.0125609      0.000370362  0.00021694
 0.00429006  0.00912781  0.00912781  0.0194209      0.000157647  0.000335419
 0.0236781   0.0138695   0.0138695   0.00812405     0.00023954   0.000140311
 0.0100787   0.0214441   0.00590361  0.0125609      0.000101961  0.00021694
 ⋮                                               ⋱               ⋮
 4.62305e-6  9.8363e-6   9.8363e-6   2.09283e-5     0.0051484    0.0109541
 2.5516e-5   1.4946e-5   1.4946e-5   8.75463e-6     0.00782287   0.00458225
 1.0861e-5   2.31086e-5  6.36185e-6  1.35359e-5     0.00332985   0.0070848
 1.0861e-5   6.36185e-6  2.31086e-5  1.35359e-5  …  0.0120952    0.0070848
 4.62305e-6  9.8363e-6   9.8363e-6   2.09283e-5     0.0051484    0.0109541
 1.0861e-5   6.36185e-6  6.36185e-6  3.72646e-6     0.0120952    0.0070848
 4.62305e-6  9.8363e-6   2.70795e-6  5.76162e-6     0.0051484    0.0109541
 4.62305e-6  2.70795e-6  9.8363e-6   5.76162e-6     0.0187009    0.0109541
 1.96783e-6  4.18687e-6  4.18687e-6  8.90826e-6  …  0.00796015   0.0169365

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.455211  0.00599985
 0.101961  0.659656

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.000382054  5.03562e-6   5.03562e-6   …  4.58642e-21  6.04508e-23
 8.55746e-5   0.000553644  1.12791e-6      1.02729e-21  6.64629e-21
 8.55746e-5   1.12791e-6   0.000553644     5.04257e-19  6.64629e-21
 1.91675e-5   0.000124008  0.000124008     1.12946e-19  7.3073e-19
 8.55746e-5   1.12791e-6   1.12791e-6      5.04257e-19  6.64629e-21
 1.91675e-5   0.000124008  2.52634e-7   …  1.12946e-19  7.3073e-19
 1.91675e-5   2.52634e-7   0.000124008     5.54408e-17  7.3073e-19
 4.29324e-6   2.7776e-5    2.7776e-5       1.24179e-17  8.03405e-17
 8.55746e-5   1.12791e-6   1.12791e-6      5.04257e-19  6.64629e-21
 1.91675e-5   0.000124008  2.52634e-7      1.12946e-19  7.3073e-19
 ⋮                                      ⋱
 5.42131e-10  3.50743e-9   3.50743e-9   …  2.19339e-5   0.000141906
 1.0806e-8    1.42427e-10  1.42427e-10     8.90671e-7   1.17394e-8
 2.42038e-9   1.56592e-8   3.19016e-11     1.99497e-7   1.29069e-6
 2.42038e-9   3.19016e-11  1.56592e-8      9.79253e-5   1.29069e-6
 5.42131e-10  3.50743e-9   3.50743e-9      2.19339e-5   0.000141906
 2.42038e-9   3.19016e-11  3.19016e-11  …  9.79253e-5   1.29069e-6
 5.42131e-10  3.50743e-9   7.14549e-12     2.19339e-5   0.000141906
 5.42131e-10  7.14549e-12  3.50743e-9      0.0107665    0.000141906
 1.21429e-10  7.85614e-10  7.85614e-10     0.00241153   0.0156019

julia> sum(P10)  # expected number of edges
7.475709470161718

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 7 stored entries:
  [192 ,   60]  =  true
  [337 ,   81]  =  true
  [312 ,  120]  =  true
  [439 ,  435]  =  true
  [712 ,  712]  =  true
  [896 ,  886]  =  true
  [1024, 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source