Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.302625  0.204563
 0.435312  0.593065
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0915821  0.061906   0.061906   0.0418461
 0.131737   0.179477   0.0890489  0.121319
 0.131737   0.0890489  0.179477   0.121319
 0.189497   0.258169   0.258169   0.351727
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  43.0132  -14.8363  -14.8363   5.11743
 -31.5719   21.9485   10.8899  -7.57059
 -31.5719   10.8899   21.9485  -7.57059
  23.1739  -16.1103  -16.1103  11.1997
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 5.90015e-7  3.98828e-7  3.98828e-7  2.69593e-7  …  7.94344e-9   5.36947e-9
 8.48709e-7  1.15627e-6  5.73695e-7  7.81597e-7     1.14263e-8   1.5567e-8
 8.48709e-7  5.73695e-7  1.15627e-6  7.81597e-7     2.30295e-8   1.5567e-8
 1.22083e-6  1.66325e-6  1.66325e-6  2.26599e-6     3.31268e-8   4.51317e-8
 8.48709e-7  5.73695e-7  5.73695e-7  3.87796e-7     2.30295e-8   1.5567e-8
 1.22083e-6  1.66325e-6  8.25234e-7  1.12429e-6  …  3.31268e-8   4.51317e-8
 1.22083e-6  8.25234e-7  1.66325e-6  1.12429e-6     6.67666e-8   4.51317e-8
 1.7561e-6   2.3925e-6   2.3925e-6   3.25952e-6     9.60406e-8   1.30845e-7
 8.48709e-7  5.73695e-7  5.73695e-7  3.87796e-7     2.30295e-8   1.5567e-8
 1.22083e-6  1.66325e-6  8.25234e-7  1.12429e-6     3.31268e-8   4.51317e-8
 ⋮                                               ⋱               ⋮
 3.21896e-5  4.38548e-5  4.38548e-5  5.97474e-5     0.000479353  0.000653065
 1.55569e-5  1.05159e-5  1.05159e-5  7.10835e-6     0.000114943  7.76974e-5
 2.23779e-5  3.04875e-5  1.51266e-5  2.06084e-5     0.000165341  0.000225259
 2.23779e-5  1.51266e-5  3.04875e-5  2.06084e-5  …  0.000333242  0.000225259
 3.21896e-5  4.38548e-5  4.38548e-5  5.97474e-5     0.000479353  0.000653065
 2.23779e-5  1.51266e-5  1.51266e-5  1.0225e-5      0.000333242  0.000225259
 3.21896e-5  4.38548e-5  2.17589e-5  2.96442e-5     0.000479353  0.000653065
 3.21896e-5  2.17589e-5  4.38548e-5  2.96442e-5     0.000966127  0.000653065
 4.63032e-5  6.30831e-5  6.30831e-5  8.59438e-5  …  0.00138973   0.00189335
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.20644e-9  1.22377e-9  1.22377e-9  …  1.41135e-9  1.41135e-9  1.43162e-9
 2.6042e-9   2.94396e-9  2.6416e-9      3.39521e-9  3.04651e-9  3.44398e-9
 2.6042e-9   2.6416e-9   2.94396e-9     3.04651e-9  3.39521e-9  3.44398e-9
 5.62135e-9  6.35476e-9  6.35476e-9     7.32881e-9  7.32881e-9  8.28499e-9
 2.6042e-9   2.6416e-9   2.6416e-9      3.39521e-9  3.39521e-9  3.44398e-9
 5.62135e-9  6.35476e-9  5.70209e-9  …  8.16768e-9  7.32881e-9  8.28499e-9
 5.62135e-9  5.70209e-9  6.35476e-9     7.32881e-9  8.16768e-9  8.28499e-9
 1.21341e-8  1.37172e-8  1.37172e-8     1.76305e-8  1.76305e-8  1.99308e-8
 2.6042e-9   2.6416e-9   2.6416e-9      3.39521e-9  3.39521e-9  3.44398e-9
 5.62135e-9  6.35476e-9  5.70209e-9     8.16768e-9  7.32881e-9  8.28499e-9
 ⋮                                   ⋱                          ⋮
 5.71929e-6  6.46547e-6  6.46547e-6     1.97752e-5  1.97752e-5  2.23552e-5
 1.22746e-6  1.24509e-6  1.24509e-6     3.80821e-6  3.80821e-6  3.86291e-6
 2.64957e-6  2.99525e-6  2.68762e-6     9.16121e-6  8.22031e-6  9.2928e-6
 2.64957e-6  2.68762e-6  2.99525e-6  …  8.22031e-6  9.16121e-6  9.2928e-6
 5.71929e-6  6.46547e-6  6.46547e-6     1.97752e-5  1.97752e-5  2.23552e-5
 2.64957e-6  2.68762e-6  2.68762e-6     9.16121e-6  9.16121e-6  9.2928e-6
 5.71929e-6  6.46547e-6  5.80143e-6     2.20386e-5  1.97752e-5  2.23552e-5
 5.71929e-6  5.80143e-6  6.46547e-6     1.97752e-5  2.20386e-5  2.23552e-5
 1.23455e-5  1.39562e-5  1.39562e-5  …  4.7572e-5   4.7572e-5   5.37787e-5

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.730523   0.495759
 0.0363537  0.350816
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.0432853    0.0293749    0.0293749    …  0.0013215    0.000896818
 0.00215405   0.0207867    0.00146181      6.57631e-5   0.000634619
 0.00215405   0.00146181   0.0207867       0.00093514   0.000634619
 0.000107194  0.00103443   0.00103443      4.65362e-5   0.000449078
 0.00215405   0.00146181   0.00146181      0.00093514   0.000634619
 0.000107194  0.00103443   7.27454e-5   …  4.65362e-5   0.000449078
 0.000107194  7.27454e-5   0.00103443      0.000661737  0.000449078
 5.33437e-6   5.14771e-5   5.14771e-5      3.29306e-5   0.000317783
 0.00215405   0.00146181   0.00146181      0.00093514   0.000634619
 0.000107194  0.00103443   7.27454e-5      4.65362e-5   0.000449078
 ⋮                                      ⋱
 8.10157e-14  7.81808e-13  7.81808e-13  …  4.13478e-6   3.9901e-5
 3.27145e-11  2.22012e-11  2.22012e-11     0.000117417  7.96831e-5
 1.628e-12    1.57103e-11  1.10482e-12     5.84311e-6   5.63865e-5
 1.628e-12    1.10482e-12  1.57103e-11     8.3088e-5    5.63865e-5
 8.10157e-14  7.81808e-13  7.81808e-13     4.13478e-6   3.9901e-5
 1.628e-12    1.10482e-12  1.10482e-12  …  8.3088e-5    5.63865e-5
 8.10157e-14  7.81808e-13  5.49801e-14     4.13478e-6   3.9901e-5
 8.10157e-14  5.49801e-14  7.81808e-13     5.87959e-5   3.9901e-5
 4.03165e-15  3.89058e-14  3.89058e-14     2.92591e-6   2.82353e-5
julia> sum(P10)  # expected number of edges119.55271732163564
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 118 stored entries:
  [3  ,   16]  =  1
  [1  ,   27]  =  1
  [37 ,   37]  =  1
  [35 ,   39]  =  1
  [557,   48]  =  1
  [820,   56]  =  1
  [5  ,   69]  =  1
  [323,   79]  =  1
  [577,  114]  =  1
  ⋮
  [217,  990]  =  1
  [919,  991]  =  1
  [9  ,  992]  =  1
  [1  ,  997]  =  1
  [653, 1006]  =  1
  [385, 1016]  =  1
  [83 , 1017]  =  1
  [971, 1019]  =  1
  [218, 1020]  =  1
  [905, 1023]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.