Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.151797  0.876804
 0.534381  0.129817
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0230423  0.133096   0.133096   0.768786
 0.0811173  0.0197059  0.468547   0.113824
 0.0811173  0.468547   0.0197059  0.113824
 0.285563   0.0693718  0.0693718  0.0168525
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  0.0836527  -0.565002  -0.565002   3.8161
 -0.344348    0.097816   2.32578   -0.660663
 -0.344348    2.32578    0.097816  -0.660663
  1.41748    -0.40265   -0.40265    0.114377
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.49676e-10  8.64553e-10  8.64553e-10  …  0.035743     0.206458
 5.26914e-10  1.28003e-10  3.04355e-9      0.125829     0.0305676
 5.26914e-10  3.04355e-9   1.28003e-10     0.00529202   0.0305676
 1.85493e-9   4.50619e-10  4.50619e-10     0.0186299    0.00452576
 5.26914e-10  3.04355e-9   3.04355e-9      0.00529202   0.0305676
 1.85493e-9   4.50619e-10  1.07144e-8   …  0.0186299    0.00452576
 1.85493e-9   1.07144e-8   4.50619e-10     0.000783522  0.00452576
 6.53004e-9   1.58635e-9   1.58635e-9      0.00275829   0.000670072
 5.26914e-10  3.04355e-9   3.04355e-9      0.00529202   0.0305676
 1.85493e-9   4.50619e-10  1.07144e-8      0.0186299    0.00452576
 ⋮                                      ⋱               ⋮
 0.000154035  3.74198e-5   3.74198e-5      6.36912e-10  1.54725e-10
 1.24292e-5   7.17933e-5   7.17933e-5      1.22197e-9   7.05832e-9
 4.37554e-5   1.06295e-5   0.000252739     4.30179e-9   1.04504e-9
 4.37554e-5   0.000252739  1.06295e-5   …  1.80922e-10  1.04504e-9
 0.000154035  3.74198e-5   3.74198e-5      6.36912e-10  1.54725e-10
 4.37554e-5   0.000252739  0.000252739     1.80922e-10  1.04504e-9
 0.000154035  3.74198e-5   0.000889733     6.36912e-10  1.54725e-10
 0.000154035  0.000889733  3.74198e-5      2.67868e-11  1.54725e-10
 0.00054226   0.000131732  0.000131732  …  9.42995e-11  2.29082e-11
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.000199173  0.000100042  0.000100042  …  1.02266e-7   5.1367e-8
 6.09723e-5   0.000196665  3.06258e-5      3.13064e-8   1.00978e-7
 6.09723e-5   3.06258e-5   0.000196665     2.01035e-7   1.00978e-7
 1.86653e-5   6.02046e-5   6.02046e-5      6.15425e-8   1.98504e-7
 6.09723e-5   3.06258e-5   3.06258e-5      2.01035e-7   1.00978e-7
 1.86653e-5   6.02046e-5   9.37539e-6   …  6.15425e-8   1.98504e-7
 1.86653e-5   9.37539e-6   6.02046e-5      3.95198e-7   1.98504e-7
 5.71397e-6   1.84303e-5   1.84303e-5      1.20981e-7   3.90222e-7
 6.09723e-5   3.06258e-5   3.06258e-5      2.01035e-7   1.00978e-7
 1.86653e-5   6.02046e-5   9.37539e-6      6.15425e-8   1.98504e-7
 ⋮                                      ⋱               ⋮
 4.40716e-10  1.42152e-9   1.42152e-9      2.69809e-5   8.70264e-5
 4.70276e-9   2.36215e-9   2.36215e-9      4.48344e-5   2.25199e-5
 1.43965e-9   4.64355e-9   7.2312e-10      1.37251e-5   4.42699e-5
 1.43965e-9   7.2312e-10   4.64355e-9   …  8.81361e-5   4.42699e-5
 4.40716e-10  1.42152e-9   1.42152e-9      2.69809e-5   8.70264e-5
 1.43965e-9   7.2312e-10   7.2312e-10      8.81361e-5   4.42699e-5
 4.40716e-10  1.42152e-9   2.21367e-10     2.69809e-5   8.70264e-5
 4.40716e-10  2.21367e-10  1.42152e-9      0.000173259  8.70264e-5
 1.34915e-10  4.35166e-10  4.35166e-10  …  5.30394e-5   0.000171078

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.918616   0.154137
 0.0705963  0.965215
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.4279       0.0717984    0.0717984    …  4.51121e-8  4.51121e-8  7.56946e-9
 0.0328844    0.449606     0.00551775      2.82495e-7  3.46689e-9  4.74005e-8
 0.0328844    0.00551775   0.449606        3.46689e-9  2.82495e-7  4.74005e-8
 0.00252719   0.0345525    0.0345525       2.17099e-8  2.17099e-8  2.96825e-7
 0.0328844    0.00551775   0.00551775      2.82495e-7  2.82495e-7  4.74005e-8
 0.00252719   0.0345525    0.000424043  …  1.769e-6    2.17099e-8  2.96825e-7
 0.00252719   0.000424043  0.0345525       2.17099e-8  1.769e-6    2.96825e-7
 0.000194216  0.00265539   0.00265539      1.35949e-7  1.35949e-7  1.85874e-6
 0.0328844    0.00551775   0.00551775      2.82495e-7  2.82495e-7  4.74005e-8
 0.00252719   0.0345525    0.000424043     1.769e-6    2.17099e-8  2.96825e-7
 ⋮                                      ⋱
 4.00102e-11  5.47033e-10  5.47033e-10  …  0.00819752  0.00819752  0.112079
 6.77448e-9   1.13671e-9   1.13671e-9      0.017034    0.017034    0.00285818
 5.20623e-10  7.11813e-9   8.73566e-11     0.106668    0.00130908  0.0178981
 5.20623e-10  8.73566e-11  7.11813e-9      0.00130908  0.106668    0.0178981
 4.00102e-11  5.47033e-10  5.47033e-10     0.00819752  0.00819752  0.112079
 5.20623e-10  8.73566e-11  8.73566e-11  …  0.106668    0.106668    0.0178981
 4.00102e-11  5.47033e-10  6.71341e-12     0.667963    0.00819752  0.112079
 4.00102e-11  6.71341e-12  5.47033e-10     0.00819752  0.667963    0.112079
 3.07481e-12  4.20399e-11  4.20399e-11     0.0513334   0.0513334   0.701847
julia> sum(P10)  # expected number of edges1737.2809502907717
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1608 stored entries:
  [65  ,    1]  =  1
  [18  ,    4]  =  1
  [514 ,    4]  =  1
  [6   ,    5]  =  1
  [6   ,    6]  =  1
  [167 ,    6]  =  1
  [8   ,    8]  =  1
  [16  ,    8]  =  1
  [9   ,    9]  =  1
  ⋮
  [867 , 1019]  =  1
  [1003, 1019]  =  1
  [1017, 1019]  =  1
  [1019, 1019]  =  1
  [748 , 1020]  =  1
  [509 , 1021]  =  1
  [1021, 1021]  =  1
  [1022, 1021]  =  1
  [975 , 1023]  =  1
  [1024, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.