Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.10323   0.79655
 0.893407  0.290711

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.0106564  0.0822277  0.0822277  0.634491
 0.0922262  0.0300101  0.711643   0.231566
 0.0922262  0.711643   0.0300101  0.231566
 0.798176   0.259724   0.259724   0.0845131

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
  0.181896  -0.498395   -0.498395    1.3656
 -0.558998   0.0645901   1.53166    -0.176977
 -0.558998   1.53166     0.0645901  -0.176977
  1.7179    -0.198497   -0.198497    0.0229356

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.4644e-12   1.12997e-11  1.12997e-11  …  0.00845564  0.00845564  0.0652461
 1.26737e-11  4.12399e-12  9.77942e-11     0.003086    0.0731797   0.0238124
 1.26737e-11  9.77942e-11  4.12399e-12     0.0731797   0.003086    0.0238124
 1.09686e-10  3.56913e-11  3.56913e-11     0.0267079   0.0267079   0.00869066
 1.26737e-11  9.77942e-11  9.77942e-11     0.003086    0.003086    0.0238124
 1.09686e-10  3.56913e-11  8.46364e-10  …  0.00112628  0.0267079   0.00869066
 1.09686e-10  8.46364e-10  3.56913e-11     0.0267079   0.00112628  0.00869066
 9.49278e-10  3.08892e-10  3.08892e-10     0.00974741  0.00974741  0.00317177
 1.26737e-11  9.77942e-11  9.77942e-11     0.003086    0.003086    0.0238124
 1.09686e-10  3.56913e-11  8.46364e-10     0.00112628  0.0267079   0.00869066
 ⋮                                      ⋱                          ⋮
 0.0298777    0.00972209   0.00972209      3.06819e-6  3.06819e-6  9.9838e-7
 0.000398895  0.00307798   0.00307798      9.71381e-7  9.71381e-7  7.49544e-6
 0.00345225   0.00112335   0.0266385       3.54518e-7  8.40686e-6  2.73556e-6
 0.00345225   0.0266385    0.00112335   …  8.40686e-6  3.54518e-7  2.73556e-6
 0.0298777    0.00972209   0.00972209      3.06819e-6  3.06819e-6  9.9838e-7
 0.00345225   0.0266385    0.0266385       3.54518e-7  3.54518e-7  2.73556e-6
 0.0298777    0.00972209   0.230544        1.29386e-7  3.06819e-6  9.9838e-7
 0.0298777    0.230544     0.00972209      3.06819e-6  1.29386e-7  9.9838e-7
 0.258578     0.0841402    0.0841402    …  1.11978e-6  1.11978e-6  3.64372e-7

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.0201654   0.00876059  0.00876059  …  2.09799e-6  2.09799e-6  9.11447e-7
 0.00982584  0.0222273   0.00426871     5.32302e-6  1.02228e-6  2.31252e-6
 0.00982584  0.00426871  0.0222273      1.02228e-6  5.32302e-6  2.31252e-6
 0.00478777  0.0108306   0.0108306      2.59371e-6  2.59371e-6  5.86731e-6
 0.00982584  0.00426871  0.00426871     5.32302e-6  5.32302e-6  2.31252e-6
 0.00478777  0.0108306   0.00207999  …  1.35055e-5  2.59371e-6  5.86731e-6
 0.00478777  0.00207999  0.0108306      2.59371e-6  1.35055e-5  5.86731e-6
 0.0023329   0.00527733  0.00527733     6.58075e-6  6.58075e-6  1.48865e-5
 0.00982584  0.00426871  0.00426871     5.32302e-6  5.32302e-6  2.31252e-6
 0.00478777  0.0108306   0.00207999     1.35055e-5  2.59371e-6  5.86731e-6
 ⋮                                   ⋱                          ⋮
 7.41318e-6  1.67696e-5  1.67696e-5     0.0113007   0.0113007   0.0255637
 3.12232e-5  1.35645e-5  1.35645e-5     0.00914091  0.00914091  0.00397115
 1.52139e-5  3.44158e-5  6.60949e-6     0.0231923   0.00445403  0.0100756
 1.52139e-5  6.60949e-6  3.44158e-5  …  0.00445403  0.0231923   0.0100756
 7.41318e-6  1.67696e-5  1.67696e-5     0.0113007   0.0113007   0.0255637
 1.52139e-5  6.60949e-6  6.60949e-6     0.0231923   0.0231923   0.0100756
 7.41318e-6  1.67696e-5  3.22056e-6     0.0588433   0.0113007   0.0255637
 7.41318e-6  3.22056e-6  1.67696e-5     0.0113007   0.0588433   0.0255637
 3.61217e-6  8.17119e-6  8.17119e-6  …  0.0286722   0.0286722   0.0648602

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.107212  0.511534
 0.941943  0.496935

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 2.00646e-10  9.5733e-10   9.5733e-10   …  0.000257105  0.00122671
 1.76284e-9   9.30008e-10  8.41091e-9      0.00225888   0.0011917
 1.76284e-9   8.41091e-9   9.30008e-10     0.000249768  0.0011917
 1.54879e-8   8.17087e-9   8.17087e-9      0.00219441   0.00115769
 1.76284e-9   8.41091e-9   8.41091e-9      0.000249768  0.0011917
 1.54879e-8   8.17087e-9   7.38967e-8   …  0.00219441   0.00115769
 1.54879e-8   7.38967e-8   8.17087e-9      0.000242639  0.00115769
 1.36074e-7   7.17877e-8   7.17877e-8      0.00213178   0.00112465
 1.76284e-9   8.41091e-9   8.41091e-9      0.000249768  0.0011917
 1.54879e-8   8.17087e-9   7.38967e-8      0.00219441   0.00115769
 ⋮                                      ⋱
 0.0625843    0.0330172    0.0330172    …  0.00179182   0.000945298
 0.000810778  0.00386841   0.00386841      0.000209936  0.00100166
 0.00712334   0.00375801   0.0339871       0.00184446   0.000973068
 0.00712334   0.0339871    0.00375801      0.000203945  0.000973068
 0.0625843    0.0330172    0.0330172       0.00179182   0.000945298
 0.00712334   0.0339871    0.0339871    …  0.000203945  0.000973068
 0.0625843    0.0330172    0.298604        0.00179182   0.000945298
 0.0625843    0.298604     0.0330172       0.000198124  0.000945298
 0.549853     0.290082     0.290082        0.00174068   0.000918319

julia> sum(P10)  # expected number of edges
1360.3760883474474

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1346 stored entries:
  [1022,    1]  =  true
  [1024,    1]  =  true
  [1007,    2]  =  true
  [1024,    3]  =  true
  [1013,    4]  =  true
  [1021,    4]  =  true
  [512 ,    5]  =  true
  [1012,    5]  =  true
  [1024,    5]  =  true
  ⋮
  [390 , 1016]  =  true
  [718 , 1016]  =  true
  [232 , 1017]  =  true
  [956 , 1018]  =  true
  [12  , 1019]  =  true
  [860 , 1021]  =  true
  [191 , 1022]  =  true
  [110 , 1024]  =  true
  [536 , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source