Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.166274  0.175394
 0.692419  0.797082
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.0276469  0.0291634  0.0291634  0.0307632
 0.115131   0.132534   0.121446   0.139804
 0.115131   0.121446   0.132534   0.139804
 0.479444   0.551915   0.551915   0.635339
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  5168.5   -1137.31   -1137.31    250.259
 -4489.84   1078.16     987.969  -237.245
 -4489.84    987.969   1078.16   -237.245
  3900.29   -936.593   -936.593   224.908
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 4.46558e-10  4.71053e-10  4.71053e-10  …  8.03517e-10  8.47594e-10
 1.85962e-9   2.14071e-9   1.96162e-9      3.34612e-9   3.8519e-9
 1.85962e-9   1.96162e-9   2.14071e-9      3.65159e-9   3.8519e-9
 7.74407e-9   8.91463e-9   8.91463e-9      1.52065e-8   1.7505e-8
 1.85962e-9   1.96162e-9   1.96162e-9      3.65159e-9   3.8519e-9
 7.74407e-9   8.91463e-9   8.16887e-9   …  1.52065e-8   1.7505e-8
 7.74407e-9   8.16887e-9   8.91463e-9      1.65947e-8   1.7505e-8
 3.22489e-8   3.71235e-8   3.71235e-8      6.9106e-8    7.95517e-8
 1.85962e-9   1.96162e-9   1.96162e-9      3.65159e-9   3.8519e-9
 7.74407e-9   8.91463e-9   8.16887e-9      1.52065e-8   1.7505e-8
 ⋮                                      ⋱               ⋮
 0.00291664   0.00335751   0.00335751      0.0125723    0.0144726
 0.000168187  0.000177412  0.000177412     0.000664324  0.000700765
 0.000700385  0.000806252  0.000738805     0.00276647   0.00318464
 0.000700385  0.000738805  0.000806252  …  0.00301903   0.00318464
 0.00291664   0.00335751   0.00335751      0.0125723    0.0144726
 0.000700385  0.000738805  0.000738805     0.00301903   0.00318464
 0.00291664   0.00335751   0.00307663      0.0125723    0.0144726
 0.00291664   0.00307663   0.00335751      0.01372      0.0144726
 0.0121459    0.0139818    0.0139818    …  0.0571348    0.065771
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.20642e-10  1.36723e-10  1.36723e-10  …  4.77839e-10  5.41534e-10
 5.39753e-10  6.12368e-10  6.11701e-10     2.13786e-9   2.42547e-9
 5.39753e-10  6.11701e-10  6.12368e-10     2.14019e-9   2.42547e-9
 2.41486e-9   2.73974e-9   2.73974e-9      9.57524e-9   1.08634e-8
 5.39753e-10  6.11701e-10  6.11701e-10     2.14019e-9   2.42547e-9
 2.41486e-9   2.73974e-9   2.73676e-9   …  9.57524e-9   1.08634e-8
 2.41486e-9   2.73676e-9   2.73974e-9      9.58568e-9   1.08634e-8
 1.08041e-8   1.22577e-8   1.22577e-8      4.28865e-8   4.86562e-8
 5.39753e-10  6.11701e-10  6.11701e-10     2.14019e-9   2.42547e-9
 2.41486e-9   2.73974e-9   2.73676e-9      9.57524e-9   1.08634e-8
 ⋮                                      ⋱               ⋮
 0.00173448   0.00196783   0.00196783      0.00694523   0.0078796
 8.66512e-5   9.82016e-5   9.82016e-5      0.000346592  0.000392792
 0.000387679  0.000439835  0.000439355     0.00155066   0.00175927
 0.000387679  0.000439355  0.000439835  …  0.00155235   0.00175927
 0.00173448   0.00196783   0.00196783      0.00694523   0.0078796
 0.000387679  0.000439355  0.000439355     0.00155235   0.00175927
 0.00173448   0.00196783   0.00196568      0.00694523   0.0078796
 0.00173448   0.00196568   0.00196783      0.00695281   0.0078796
 0.00776009   0.00880408   0.00880408   …  0.031107     0.0352919

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.319918  0.520466
 0.615423  0.472317
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 1.12304e-5  1.82703e-5  1.82703e-5  2.97235e-5  …  0.000896545  0.00145856
 2.16037e-5  1.65801e-5  3.51464e-5  2.69737e-5     0.00172467   0.00132363
 2.16037e-5  3.51464e-5  1.65801e-5  2.69737e-5     0.000813604  0.00132363
 4.15587e-5  3.18949e-5  3.18949e-5  2.44783e-5     0.00156512   0.00120118
 2.16037e-5  3.51464e-5  3.51464e-5  5.71787e-5     0.000813604  0.00132363
 4.15587e-5  3.18949e-5  6.76106e-5  5.18889e-5  …  0.00156512   0.00120118
 4.15587e-5  6.76106e-5  3.18949e-5  5.18889e-5     0.000738335  0.00120118
 7.99459e-5  6.13558e-5  6.13558e-5  4.70886e-5     0.00142033   0.00109005
 2.16037e-5  3.51464e-5  3.51464e-5  5.71787e-5     0.000813604  0.00132363
 4.15587e-5  3.18949e-5  6.76106e-5  5.18889e-5     0.00156512   0.00120118
 ⋮                                               ⋱
 0.00405136  0.00310929  0.00310929  0.00238628  …  0.000793291  0.000608825
 0.00109479  0.00178109  0.00178109  0.0028976      0.00045442   0.000739283
 0.00210604  0.00161632  0.00342626  0.00262954     0.000874162  0.00067089
 0.00210604  0.00342626  0.00161632  0.00262954     0.000412381  0.00067089
 0.00405136  0.00310929  0.00310929  0.00238628     0.000793291  0.000608825
 0.00210604  0.00342626  0.00342626  0.00557408  …  0.000412381  0.00067089
 0.00405136  0.00310929  0.00659105  0.00505841     0.000793291  0.000608825
 0.00405136  0.00659105  0.00310929  0.00505841     0.00037423   0.000608825
 0.00779355  0.00598129  0.00598129  0.00459045     0.000719902  0.000552501
julia> sum(P10)  # expected number of edges710.1496266054398
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 709 stored entries:
  [1016,    1]  =  1
  [452 ,    4]  =  1
  [732 ,    5]  =  1
  [462 ,    6]  =  1
  [179 ,    8]  =  1
  [227 ,   11]  =  1
  [981 ,   11]  =  1
  [20  ,   12]  =  1
  [256 ,   13]  =  1
  ⋮
  [919 , 1015]  =  1
  [72  , 1017]  =  1
  [4   , 1018]  =  1
  [21  , 1019]  =  1
  [438 , 1019]  =  1
  [28  , 1020]  =  1
  [608 , 1021]  =  1
  [2   , 1023]  =  1
  [803 , 1023]  =  1
  [1000, 1023]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.