Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.0697135  0.632435
 0.0551743  0.758474

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.00485997  0.0440892  0.0440892  0.399974
 0.00384639  0.0528759  0.0348941  0.479685
 0.00384639  0.0348941  0.0528759  0.479685
 0.0030442   0.0418482  0.0418482  0.575283

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 1779.17     -1483.51    -1483.51    1236.99
 -129.423      163.528     107.916   -136.354
 -129.423      107.916     163.528   -136.354
    9.41473    -11.8956    -11.8956    15.0303

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.31766e-14  1.19537e-13  1.19537e-13  …  0.000451326  0.00409439
 1.04285e-14  1.4336e-13   9.46067e-14     0.000357198  0.00491036
 1.04285e-14  9.46067e-14  1.4336e-13      0.000541271  0.00491036
 8.25358e-15  1.13461e-13  1.13461e-13     0.000428385  0.00588896
 1.04285e-14  9.46067e-14  9.46067e-14     0.000541271  0.00491036
 8.25358e-15  1.13461e-13  7.48758e-14  …  0.000428385  0.00588896
 8.25358e-15  7.48758e-14  1.13461e-13     0.000649142  0.00588896
 6.53224e-15  8.97979e-14  8.97979e-14     0.000513759  0.00706259
 1.04285e-14  9.46067e-14  9.46067e-14     0.000541271  0.00491036
 8.25358e-15  1.13461e-13  7.48758e-14     0.000428385  0.00588896
 ⋮                                      ⋱               ⋮
 1.00558e-15  1.38236e-14  1.38236e-14     0.00219867   0.0302248
 1.60538e-15  1.45639e-14  1.45639e-14     0.00231641   0.0210142
 1.27057e-15  1.74664e-14  1.15265e-14     0.0018333    0.0252022
 1.27057e-15  1.15265e-14  1.74664e-14  …  0.00277805   0.0252022
 1.00558e-15  1.38236e-14  1.38236e-14     0.00219867   0.0302248
 1.27057e-15  1.15265e-14  1.15265e-14     0.00277805   0.0252022
 1.00558e-15  1.38236e-14  9.12256e-15     0.00219867   0.0302248
 1.00558e-15  9.12256e-15  1.38236e-14     0.00333169   0.0302248
 7.95861e-16  1.09406e-14  1.09406e-14  …  0.00263685   0.0362484

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 1.55805e-17  2.05279e-16  2.05279e-16  …  3.23572e-5   0.000426318
 1.79087e-17  2.39142e-16  2.35954e-16     3.71924e-5   0.000496643
 1.79087e-17  2.35954e-16  2.39142e-16     3.76949e-5   0.000496643
 2.05849e-17  2.74877e-16  2.74877e-16     4.33277e-5   0.00057857
 1.79087e-17  2.35954e-16  2.35954e-16     3.76949e-5   0.000496643
 2.05849e-17  2.74877e-16  2.71213e-16  …  4.33277e-5   0.00057857
 2.05849e-17  2.71213e-16  2.74877e-16     4.3913e-5    0.00057857
 2.36609e-17  3.15953e-16  3.15953e-16     5.04751e-5   0.000674012
 1.79087e-17  2.35954e-16  2.35954e-16     3.76949e-5   0.000496643
 2.05849e-17  2.74877e-16  2.71213e-16     4.33277e-5   0.00057857
 ⋮                                      ⋱               ⋮
 7.20939e-17  9.62696e-16  9.62696e-16     0.000171225  0.00228644
 5.45672e-17  7.18943e-16  7.18943e-16     0.000127871  0.00168475
 6.27213e-17  8.37541e-16  8.26376e-16     0.00014698   0.00196267
 6.27213e-17  8.26376e-16  8.37541e-16  …  0.000148965  0.00196267
 7.20939e-17  9.62696e-16  9.62696e-16     0.000171225  0.00228644
 6.27213e-17  8.26376e-16  8.26376e-16     0.000148965  0.00196267
 7.20939e-17  9.62696e-16  9.49864e-16     0.000171225  0.00228644
 7.20939e-17  9.49864e-16  9.62696e-16     0.000173539  0.00228644
 8.28671e-17  1.10655e-15  1.10655e-15  …  0.000199471  0.00266361

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.49703   0.567396
 0.202524  0.595979

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.000920081  0.00105034   0.00105034   …  0.0030294   0.0030294   0.00345828
 0.000374905  0.00110325   0.000427981     0.00318201  0.00123439  0.0036325
 0.000374905  0.000427981  0.00110325      0.00123439  0.00318201  0.0036325
 0.000152762  0.000449541  0.000449541     0.00129657  0.00129657  0.00381549
 0.000374905  0.000427981  0.000427981     0.00318201  0.00318201  0.0036325
 0.000152762  0.000449541  0.000174389  …  0.00334231  0.00129657  0.00381549
 0.000152762  0.000174389  0.000449541     0.00129657  0.00334231  0.00381549
 6.22459e-5   0.000183174  0.000183174     0.00136189  0.00136189  0.00400771
 0.000374905  0.000427981  0.000427981     0.00318201  0.00318201  0.0036325
 0.000152762  0.000449541  0.000174389     0.00334231  0.00129657  0.00381549
 ⋮                                      ⋱
 2.84892e-7   8.38366e-7   8.38366e-7   …  0.001829    0.001829    0.00538229
 1.71589e-6   1.95882e-6   1.95882e-6      0.0042734   0.0042734   0.00487839
 6.99174e-7   2.0575e-6    7.98157e-7      0.00448868  0.00174128  0.00512415
 6.99174e-7   7.98157e-7   2.0575e-6       0.00174128  0.00448868  0.00512415
 2.84892e-7   8.38366e-7   8.38366e-7      0.001829    0.001829    0.00538229
 6.99174e-7   7.98157e-7   7.98157e-7   …  0.00448868  0.00448868  0.00512415
 2.84892e-7   8.38366e-7   3.25225e-7      0.00471481  0.001829    0.00538229
 2.84892e-7   3.25225e-7   8.38366e-7      0.001829    0.00471481  0.00538229
 1.16085e-7   3.41608e-7   3.41608e-7      0.00192114  0.00192114  0.00565344

julia> sum(P10)  # expected number of edges
503.45926261893817

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 502 stored entries:
  [529,   20]  =  1
  [401,   23]  =  1
  [963,   40]  =  1
  [490,   42]  =  1
  [10 ,   58]  =  1
  [254,   63]  =  1
  [482,   68]  =  1
  [85 ,   80]  =  1
  [13 ,   96]  =  1
  ⋮
  [458, 1022]  =  1
  [514, 1022]  =  1
  [811, 1022]  =  1
  [41 , 1023]  =  1
  [659, 1023]  =  1
  [22 , 1024]  =  1
  [142, 1024]  =  1
  [340, 1024]  =  1
  [442, 1024]  =  1
  [717, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.