Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.865029  0.0579756
 0.99784   0.0143398

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
 0.748275  0.0501506  0.0501506  0.00336117
 0.86316   0.0124044  0.0578503  0.00083136
 0.86316   0.0578503  0.0124044  0.00083136
 0.995684  0.0143088  0.0143088  0.000205631

julia> inv(K2)
4×4 KroneckerPower{Float64,Array{Float64,2},2}:
   0.0995628    -0.40253    -0.40253    1.62742
  -6.92809       6.00597    28.0101   -24.282
  -6.92809      28.0101      6.00597  -24.282
 482.093      -417.927    -417.927    362.302

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.175536  0.0117647   0.0117647   0.000788489  …  2.15143e-14  1.44193e-15
 0.202487  0.00290991  0.013571    0.000195027     2.48175e-14  3.56649e-16
 0.202487  0.013571    0.00290991  0.000195027     5.32141e-15  3.56649e-16
 0.233575  0.00335668  0.00335668  4.82384e-5      6.13842e-15  8.82145e-17
 0.202487  0.013571    0.013571    0.000909549     5.32141e-15  3.56649e-16
 0.233575  0.00335668  0.0156546   0.00022497   …  6.13842e-15  8.82145e-17
 0.233575  0.0156546   0.00335668  0.00022497      1.31621e-15  8.82145e-17
 0.269437  0.00387204  0.00387204  5.56446e-5      1.51829e-15  2.18192e-17
 0.202487  0.013571    0.013571    0.000909549     5.32141e-15  3.56649e-16
 0.233575  0.00335668  0.0156546   0.00022497      6.13842e-15  8.82145e-17
 ⋮                                              ⋱               ⋮
 0.844692  0.012139    0.012139    0.000174447     2.12689e-20  3.05653e-22
 0.634802  0.0425454   0.0425454   0.00285146      7.45446e-20  4.9961e-21
 0.732265  0.0105233   0.0490775   0.000705287     8.59897e-20  1.23575e-21
 0.732265  0.0490775   0.0105233   0.000705287  …  1.8438e-20   1.23575e-21
 0.844692  0.012139    0.012139    0.000174447     2.12689e-20  3.05653e-22
 0.732265  0.0490775   0.0490775   0.00328925      1.8438e-20   1.23575e-21
 0.844692  0.012139    0.0566126   0.000813572     2.12689e-20  3.05653e-22
 0.844692  0.0566126   0.012139    0.000813572     4.56051e-21  3.05653e-22
 0.97438   0.0140027   0.0140027   0.000201231  …  5.2607e-21   7.56009e-23

julia> K12^2  # example
4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2},2},6}:
 0.0753059  0.00476259  0.00476259  0.000301201  …  4.87506e-15  3.08315e-16
 0.0819707  0.00542343  0.00518409  0.000342995     5.30652e-15  3.51095e-16
 0.0819707  0.00518409  0.00542343  0.000342995     5.55151e-15  3.51095e-16
 0.0892252  0.00590341  0.00590341  0.000390587     6.04283e-15  3.99812e-16
 0.0819707  0.00518409  0.00518409  0.000327858     5.55151e-15  3.51095e-16
 0.0892252  0.00590341  0.00564289  0.00037335   …  6.04283e-15  3.99812e-16
 0.0892252  0.00564289  0.00590341  0.00037335      6.32181e-15  3.99812e-16
 0.0971218  0.00642587  0.00642587  0.000425155     6.88131e-15  4.55288e-16
 0.0819707  0.00518409  0.00518409  0.000327858     5.55151e-15  3.51095e-16
 0.0892252  0.00590341  0.00564289  0.00037335      6.04283e-15  3.99812e-16
 ⋮                                               ⋱               ⋮
 0.191404   0.0126639   0.0126639   0.000837879     1.94588e-14  1.28746e-15
 0.161544   0.0102166   0.0102166   0.00064613      1.56985e-14  9.9282e-16
 0.175841   0.0116342   0.0111208   0.000735784     1.70878e-14  1.13058e-15
 0.175841   0.0111208   0.0116342   0.000735784  …  1.78767e-14  1.13058e-15
 0.191404   0.0126639   0.0126639   0.000837879     1.94588e-14  1.28746e-15
 0.175841   0.0111208   0.0111208   0.000703313     1.78767e-14  1.13058e-15
 0.191404   0.0126639   0.012105    0.000800902     1.94588e-14  1.28746e-15
 0.191404   0.012105    0.0126639   0.000800902     2.03572e-14  1.28746e-15
 0.208343   0.0137846   0.0137846   0.000912032  …  2.21589e-14  1.4661e-15

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.577175  0.415509
 0.499838  0.220329

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Float64,Array{Float64,2},10}:
 0.00410276   0.00295358   0.00295358   …  0.000213075  0.000153393
 0.00355303   0.00156618   0.00255783      0.000184525  8.13387e-5
 0.00355303   0.00255783   0.00156618      0.000112986  8.13387e-5
 0.00307695   0.00135632   0.00135632      9.78467e-5   4.3131e-5
 0.00355303   0.00255783   0.00255783      0.000112986  8.13387e-5
 0.00307695   0.00135632   0.0022151    …  9.78467e-5   4.3131e-5
 0.00307695   0.0022151    0.00135632      5.99124e-5   4.3131e-5
 0.00266466   0.00117459   0.00117459      5.18846e-5   2.28708e-5
 0.00355303   0.00255783   0.00255783      0.000112986  8.13387e-5
 0.00307695   0.00135632   0.0022151       9.78467e-5   4.3131e-5
 ⋮                                      ⋱
 0.00112402   0.00049547   0.00049547   …  1.15343e-6   5.08433e-7
 0.00149875   0.00107896   0.00107896      2.51175e-6   1.80821e-6
 0.00129793   0.000572131  0.000934384     2.1752e-6    9.58831e-7
 0.00129793   0.000934384  0.000572131     1.33189e-6   9.58831e-7
 0.00112402   0.00049547   0.00049547      1.15343e-6   5.08433e-7
 0.00129793   0.000934384  0.000934384  …  1.33189e-6   9.58831e-7
 0.00112402   0.00049547   0.000809184     1.15343e-6   5.08433e-7
 0.00112402   0.000809184  0.00049547      7.06254e-7   5.08433e-7
 0.000973412  0.000429081  0.000429081     6.11622e-7   2.69604e-7

julia> sum(P10)  # expected number of edges
217.36960119222329

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 217 stored entries:
  [539,   1]  =  1
  [699,   1]  =  1
  [807,   1]  =  1
  [24 ,   2]  =  1
  [305,   2]  =  1
  [459,   2]  =  1
  [37 ,   5]  =  1
  [337,   5]  =  1
  [758,   5]  =  1
  ⋮
  [33 , 921]  =  1
  [106, 925]  =  1
  [70 , 931]  =  1
  [961, 950]  =  1
  [261, 953]  =  1
  [140, 959]  =  1
  [217, 968]  =  1
  [737, 976]  =  1
  [592, 977]  =  1
  [519, 997]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.