Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.844867  0.636124
 0.482167  0.461843
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.7138    0.53744   0.53744   0.404654
 0.407367  0.390196  0.306718  0.293789
 0.407367  0.306718  0.390196  0.293789
 0.232485  0.222685  0.222685  0.213299
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  30.6088  -42.1594  -42.1594   58.0687
 -31.9558   55.994    44.0147  -77.1238
 -31.9558   44.0147   55.994   -77.1238
  33.3621  -58.4581  -58.4581  102.432
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.13227      0.0995896    0.0995896    …  0.00583106   0.00439037
 0.0754866    0.0723047    0.056836        0.00332779   0.00318752
 0.0754866    0.056836     0.0723047       0.0042335    0.00318752
 0.0430803    0.0412644    0.0412644       0.00241607   0.00231423
 0.0754866    0.056836     0.056836        0.0042335    0.00318752
 0.0430803    0.0412644    0.0324364    …  0.00241607   0.00231423
 0.0430803    0.0324364    0.0412644       0.00307363   0.00231423
 0.024586     0.0235497    0.0235497       0.00175413   0.00168019
 0.0754866    0.056836     0.056836        0.0042335    0.00318752
 0.0430803    0.0412644    0.0324364       0.00241607   0.00231423
 ⋮                                      ⋱               ⋮
 0.00027667   0.000265008  0.000265008     0.00013542   0.000129711
 0.000849461  0.000639583  0.000639583     0.000326828  0.000246078
 0.000484789  0.000464354  0.000365011     0.000186521  0.000178659
 0.000484789  0.000365011  0.000464354  …  0.000237286  0.000178659
 0.00027667   0.000265008  0.000265008     0.00013542   0.000129711
 0.000484789  0.000365011  0.000365011     0.000237286  0.000178659
 0.00027667   0.000265008  0.000208312     0.00013542   0.000129711
 0.00027667   0.000208312  0.000265008     0.000172276  0.000129711
 0.000157896  0.00015124   0.00015124   …  9.83181e-5   9.41738e-5
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 1.276       1.03932     1.03932     0.846545    …  0.133583     0.108805
 0.787782    0.6502      0.641662    0.529598       0.082472     0.0680687
 0.787782    0.641662    0.6502      0.529598       0.0835694    0.0680687
 0.486365    0.401423    0.401423    0.331317       0.0515945    0.0425837
 0.787782    0.641662    0.641662    0.522644       0.0835694    0.0680687
 0.486365    0.401423    0.396152    0.326966    …  0.0515945    0.0425837
 0.486365    0.396152    0.401423    0.326966       0.052281     0.0425837
 0.300274    0.247833    0.247833    0.20455        0.0322775    0.0266404
 0.787782    0.641662    0.641662    0.522644       0.0835694    0.0680687
 0.486365    0.401423    0.396152    0.326966       0.0515945    0.0425837
 ⋮                                               ⋱               ⋮
 0.00633819  0.00523126  0.00523126  0.00431764     0.000757306  0.000625046
 0.0166285   0.0135442   0.0135442   0.011032       0.00196074   0.00159705
 0.0102662   0.00847325  0.00836199  0.00690161     0.00121053   0.000999116
 0.0102662   0.00836199  0.00847325  0.00690161  …  0.00122664   0.000999116
 0.00633819  0.00523126  0.00523126  0.00431764     0.000757306  0.000625046
 0.0102662   0.00836199  0.00836199  0.00681098     0.00122664   0.000999116
 0.00633819  0.00523126  0.00516256  0.00426095     0.000757306  0.000625046
 0.00633819  0.00516256  0.00523126  0.00426095     0.000767383  0.000625046
 0.0039131   0.0032297   0.0032297   0.00266565  …  0.000473771  0.000391029

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.472261   0.313652
 0.0299383  0.531924
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.000551853  0.000366512  0.000366512  …  1.38743e-5   9.21464e-6
 3.49839e-5   0.000621571  2.32345e-5      8.79544e-7   1.56272e-5
 3.49839e-5   2.32345e-5   0.000621571     2.35296e-5   1.56272e-5
 2.21775e-6   3.94036e-5   3.94036e-5      1.49162e-6   2.65022e-5
 3.49839e-5   2.32345e-5   2.32345e-5      2.35296e-5   1.56272e-5
 2.21775e-6   3.94036e-5   1.47292e-6   …  1.49162e-6   2.65022e-5
 2.21775e-6   1.47292e-6   3.94036e-5      3.9904e-5    2.65022e-5
 1.40591e-7   2.49793e-6   2.49793e-6      2.52966e-6   4.49453e-5
 3.49839e-5   2.32345e-5   2.32345e-5      2.35296e-5   1.56272e-5
 2.21775e-6   3.94036e-5   1.47292e-6      1.49162e-6   2.65022e-5
 ⋮                                      ⋱
 9.12488e-15  1.62125e-13  1.62125e-13  …  6.0183e-5    0.00106929
 2.27058e-12  1.50801e-12  1.50801e-12     0.000559792  0.000371785
 1.4394e-13   2.55744e-12  9.55978e-14     3.54872e-5   0.000630513
 1.4394e-13   9.55978e-14  2.55744e-12     0.000949355  0.000630513
 9.12488e-15  1.62125e-13  1.62125e-13     6.0183e-5    0.00106929
 1.4394e-13   9.55978e-14  9.55978e-14  …  0.000949355  0.000630513
 9.12488e-15  1.62125e-13  6.06028e-15     6.0183e-5    0.00106929
 9.12488e-15  6.06028e-15  1.62125e-13     0.00161002   0.00106929
 5.78458e-16  1.02777e-14  1.02777e-14     0.000102065  0.00181342
julia> sum(P10)  # expected number of edges19.77771282477469
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 20 stored entries:
  [9  ,  14]  =  1
  [31 ,  96]  =  1
  [98 ,  98]  =  1
  [122, 122]  =  1
  [336, 352]  =  1
  [307, 379]  =  1
  [278, 434]  =  1
  [259, 452]  =  1
  [260, 468]  =  1
  [387, 475]  =  1
  [23 , 663]  =  1
  [533, 693]  =  1
  [674, 700]  =  1
  [676, 744]  =  1
  [67 , 839]  =  1
  [274, 852]  =  1
  [261, 872]  =  1
  [919, 925]  =  1
  [1  , 951]  =  1
  [647, 999]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.