Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.198876  0.5898
 0.547157  0.675269

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.0395515  0.117297  0.117297  0.347865
 0.108816   0.134295  0.322714  0.398274
 0.108816   0.322714  0.134295  0.398274
 0.299381   0.369479  0.369479  0.455989

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  12.8441   -11.2184   -11.2184    9.79853
 -10.4073     3.78276    9.09008  -3.30398
 -10.4073     9.09008    3.78276  -3.30398
   8.43286   -3.0651    -3.0651    1.11407

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 3.82804e-9   1.13527e-8   1.13527e-8   …  0.000597499  0.00177199
 1.05319e-8   1.29979e-8   3.12342e-8      0.00164387   0.00202877
 1.05319e-8   3.12342e-8   1.29979e-8      0.000684083  0.00202877
 2.89759e-8   3.57604e-8   3.57604e-8      0.00188209   0.00232276
 1.05319e-8   3.12342e-8   3.12342e-8      0.000684083  0.00202877
 2.89759e-8   3.57604e-8   8.59333e-8   …  0.00188209   0.00232276
 2.89759e-8   8.59333e-8   3.57604e-8      0.000783215  0.00232276
 7.97202e-8   9.8386e-8    9.8386e-8       0.00215482   0.00265936
 1.05319e-8   3.12342e-8   3.12342e-8      0.000684083  0.00202877
 2.89759e-8   3.57604e-8   8.59333e-8      0.00188209   0.00232276
 ⋮                                      ⋱               ⋮
 0.000261707  0.000322983  0.000322983     0.00636192   0.00785151
 3.45743e-5   0.000102536  0.000102536     0.00201969   0.00598976
 9.51227e-5   0.000117395  0.000282103     0.00555669   0.00685774
 9.51227e-5   0.000282103  0.000117395  …  0.00231237   0.00685774
 0.000261707  0.000322983  0.000322983     0.00636192   0.00785151
 9.51227e-5   0.000282103  0.000282103     0.00231237   0.00685774
 0.000261707  0.000322983  0.000776138     0.00636192   0.00785151
 0.000261707  0.000776138  0.000322983     0.00264746   0.00785151
 0.000720022  0.000888609  0.000888609  …  0.00728384   0.00898929

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 5.10878e-6   7.27075e-6   7.27075e-6   …  0.000247856  0.000352746
 6.74507e-6   1.09815e-5   9.59949e-6      0.000327242  0.000532777
 6.74507e-6   9.59949e-6   1.09815e-5      0.000374355  0.000532777
 8.90544e-6   1.44988e-5   1.44988e-5      0.000494256  0.000804689
 6.74507e-6   9.59949e-6   9.59949e-6      0.000374355  0.000532777
 8.90544e-6   1.44988e-5   1.26741e-5   …  0.000494256  0.000804689
 8.90544e-6   1.26741e-5   1.44988e-5      0.000565413  0.000804689
 1.17578e-5   1.91426e-5   1.91426e-5      0.000746509  0.00121538
 6.74507e-6   9.59949e-6   9.59949e-6      0.000374355  0.000532777
 8.90544e-6   1.44988e-5   1.26741e-5      0.000494256  0.000804689
 ⋮                                      ⋱               ⋮
 0.000108562  0.000176748  0.000176748     0.0202162    0.0329136
 6.22787e-5   8.86343e-5   8.86343e-5      0.0101379    0.0144281
 8.2226e-5    0.00013387   0.000117023     0.0133849    0.0217917
 8.2226e-5    0.000117023  0.00013387   …  0.0153119    0.0217917
 0.000108562  0.000176748  0.000176748     0.0202162    0.0329136
 8.2226e-5    0.000117023  0.000117023     0.0153119    0.0217917
 0.000108562  0.000176748  0.000154504     0.0202162    0.0329136
 0.000108562  0.000154504  0.000176748     0.0231267    0.0329136
 0.000143333  0.000233358  0.000233358  …  0.0305339    0.0497116

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.610449  0.651346
 0.618195  0.160394

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 0.00718613  0.00766755  0.00766755  0.00818124   …  0.0128811    0.0137441
 0.00727731  0.00188813  0.00776485  0.00201463      0.0130446    0.00338447
 0.00727731  0.00776485  0.00188813  0.00201463      0.00317197   0.00338447
 0.00736965  0.00191209  0.00191209  0.0004961       0.00321222   0.000833424
 0.00727731  0.00776485  0.00776485  0.00828505      0.00317197   0.00338447
 0.00736965  0.00191209  0.00786338  0.00204019   …  0.00321222   0.000833424
 0.00736965  0.00786338  0.00191209  0.00204019      0.000781095  0.000833424
 0.00746316  0.00193635  0.00193635  0.000502395     0.000791006  0.00020523
 0.00727731  0.00776485  0.00776485  0.00828505      0.00317197   0.00338447
 0.00736965  0.00191209  0.00786338  0.00204019      0.00321222   0.000833424
 ⋮                                                ⋱
 0.0080497   0.00208853  0.00208853  0.000541879  …  1.76373e-7   4.57607e-8
 0.00784923  0.00837509  0.00837509  0.00893617      7.07262e-7   7.54644e-7
 0.00794883  0.00206236  0.00848136  0.00220053      7.16236e-7   1.85831e-7
 0.00794883  0.00848136  0.00206236  0.00220053      1.74163e-7   1.85831e-7
 0.0080497   0.00208853  0.00208853  0.000541879     1.76373e-7   4.57607e-8
 0.00794883  0.00848136  0.00848136  0.00904956   …  1.74163e-7   1.85831e-7
 0.0080497   0.00208853  0.00858898  0.00222845      1.76373e-7   4.57607e-8
 0.0080497   0.00858898  0.00208853  0.00222845      4.28875e-8   4.57607e-8
 0.00815184  0.00211503  0.00211503  0.000548754     4.34317e-8   1.12686e-8

julia> sum(P10)  # expected number of edges
1250.5984653865205

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1249 stored entries:
  [172 ,    1]  =  true
  [299 ,    1]  =  true
  [356 ,    1]  =  true
  [397 ,    1]  =  true
  [447 ,    1]  =  true
  [565 ,    1]  =  true
  [613 ,    1]  =  true
  [782 ,    1]  =  true
  [836 ,    1]  =  true
  ⋮
  [132 ,  993]  =  true
  [22  , 1001]  =  true
  [17  , 1011]  =  true
  [955 , 1014]  =  true
  [3   , 1018]  =  true
  [55  , 1019]  =  true
  [533 , 1020]  =  true
  [97  , 1021]  =  true
  [9   , 1024]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source