Kronecker powers and graphs

Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker

julia> A = rand(2, 2)
2×2 Array{Float64,2}:
 0.592142  0.0216687
 0.678368  0.624902

julia> K2 = A ⊗ 2
4×4 KroneckerPower{Array{Float64,2},2}:
 0.350633  0.0128309  0.0128309  0.000469532
 0.401691  0.370031   0.0146993  0.0135408
 0.401691  0.0146993  0.370031   0.0135408
 0.460184  0.423914   0.423914   0.390502

julia> inv(K2)
4×4 KroneckerPower{Array{Float64,2},2}:
  3.09283  -0.107245  -0.107245   0.00371875
 -3.35745   2.93069    0.116421  -0.101623
 -3.35745   0.116421   2.93069   -0.101623
  3.64472  -3.18144   -3.18144    2.77706

julia> K12 = K2 ⊗ 6  # works recursively
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 0.00185829  6.80016e-5   6.80016e-5   …  2.92808e-19  1.07149e-20
 0.00212889  0.0019611    7.79038e-5      3.35446e-19  3.09007e-19
 0.00212889  7.79038e-5   0.0019611       8.44427e-18  3.09007e-19
 0.00243889  0.00224666   0.00224666      9.6739e-18   8.91144e-18
 0.00212889  7.79038e-5   7.79038e-5      8.44427e-18  3.09007e-19
 0.00243889  0.00224666   8.92479e-5   …  9.6739e-18   8.91144e-18
 0.00243889  8.92479e-5   0.00224666      2.43524e-16  8.91144e-18
 0.00279403  0.00257382   0.00257382      2.78985e-16  2.56996e-16
 0.00212889  7.79038e-5   7.79038e-5      8.44427e-18  3.09007e-19
 0.00243889  0.00224666   8.92479e-5      9.6739e-18   8.91144e-18
 ⋮                                     ⋱               ⋮
 0.00828986  0.00763648   0.00763648      0.00013348   0.000122959
 0.00631638  0.00023114   0.00023114      4.04014e-6   1.47844e-7
 0.00723615  0.00666582   0.000264797     4.62846e-6   4.26366e-6
 0.00723615  0.000264797  0.00666582   …  0.000116514  4.26366e-6
 0.00828986  0.00763648   0.00763648      0.00013348   0.000122959
 0.00723615  0.000264797  0.000264797     0.000116514  4.26366e-6
 0.00828986  0.00763648   0.000303356     0.00013348   0.000122959
 0.00828986  0.000303356  0.00763648      0.00336013   0.000122959
 0.009497    0.00874848   0.00874848   …  0.00384942   0.00354602

julia> K12^2  # example
4096×4096 KroneckerPower{KroneckerPower{Array{Float64,2},2},6}:
 5.65263e-6  4.08039e-7   4.08039e-7   …  1.56752e-18  1.13153e-19
 1.27742e-5  6.26952e-6   9.22117e-7      3.5424e-18   1.73859e-18
 1.27742e-5  9.22117e-7   6.26952e-6      2.4085e-17   1.73859e-18
 2.88682e-5  1.41683e-5   1.41683e-5      5.4429e-17   2.67134e-17
 1.27742e-5  9.22117e-7   9.22117e-7      2.4085e-17   1.73859e-18
 2.88682e-5  1.41683e-5   2.08387e-6   …  5.4429e-17   2.67134e-17
 2.88682e-5  2.08387e-6   1.41683e-5      3.70065e-16  2.67134e-17
 6.52384e-5  3.20186e-5   3.20186e-5      8.36301e-16  4.10452e-16
 1.27742e-5  9.22117e-7   9.22117e-7      2.4085e-17   1.73859e-18
 2.88682e-5  1.41683e-5   2.08387e-6      5.4429e-17   2.67134e-17
 ⋮                                     ⋱               ⋮
 0.044379    0.021781     0.021781        2.5979e-6    1.27503e-6
 0.00868979  0.000627279  0.000627279     7.48179e-8   5.40078e-9
 0.0196378   0.00963813   0.00141757      1.69079e-7   8.2983e-8
 0.0196378   0.00141757   0.00963813   …  1.14958e-6   8.2983e-8
 0.044379    0.021781     0.021781        2.5979e-6    1.27503e-6
 0.0196378   0.00141757   0.00141757      1.14958e-6   8.2983e-8
 0.044379    0.021781     0.00320353      2.5979e-6    1.27503e-6
 0.044379    0.00320353   0.021781        1.76632e-5   1.27503e-6
 0.100291    0.0492223    0.0492223    …  3.99167e-5   1.95909e-5

Kronecker.kronecker — Method.

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker.:⊗ — Method.

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

source

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker

julia> P = rand(2, 2)
2×2 Array{Float64,2}:
 0.269806   0.506971
 0.0415727  0.474922

julia> P10 = P ⊗ 10
1024×1024 KroneckerPower{Array{Float64,2},10}:
 2.04414e-6   3.84098e-6   3.84098e-6   …  0.000596894  0.00112158
 3.14969e-7   3.59817e-6   5.91833e-7      9.19716e-5   0.00105067
 3.14969e-7   5.91833e-7   3.59817e-6      0.00055916   0.00105067
 4.85316e-8   5.5442e-7    5.5442e-7       8.61575e-5   0.000984254
 3.14969e-7   5.91833e-7   5.91833e-7      0.00055916   0.00105067
 4.85316e-8   5.5442e-7    9.11919e-8   …  8.61575e-5   0.000984254
 4.85316e-8   9.11919e-8   5.5442e-7       0.000523812  0.000984254
 7.47793e-9   8.54271e-8   8.54271e-8      8.0711e-5    0.000922034
 3.14969e-7   5.91833e-7   5.91833e-7      0.00055916   0.00105067
 4.85316e-8   5.5442e-7    9.11919e-8      8.61575e-5   0.000984254
 ⋮                                      ⋱
 1.00074e-13  1.14324e-12  1.14324e-12  …  5.45467e-5   0.000623136
 4.21511e-12  7.92029e-12  7.92029e-12     0.000377896  0.000710075
 6.49481e-13  7.4196e-12   1.22039e-12     5.82277e-5   0.000665187
 6.49481e-13  1.22039e-12  7.4196e-12      0.000354007  0.000665187
 1.00074e-13  1.14324e-12  1.14324e-12     5.45467e-5   0.000623136
 6.49481e-13  1.22039e-12  1.22039e-12  …  0.000354007  0.000665187
 1.00074e-13  1.14324e-12  1.88042e-13     5.45467e-5   0.000623136
 1.00074e-13  1.88042e-13  1.14324e-12     0.000331628  0.000623136
 1.54198e-14  1.76155e-13  1.76155e-13     5.10985e-5   0.000583744

julia> sum(P10)  # expected number of edges
13.088664958215537

julia> A = fastsample(P10)
1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 13 stored entries:
  [129 ,  487]  =  true
  [317 ,  496]  =  true
  [458 ,  512]  =  true
  [664 ,  724]  =  true
  [241 ,  767]  =  true
  [81  ,  800]  =  true
  [129 ,  919]  =  true
  [679 ,  999]  =  true
  [593 , 1000]  =  true
  [515 , 1004]  =  true
  [333 , 1007]  =  true
  [194 , 1008]  =  true
  [969 , 1019]  =  true

Kronecker.isprob — Function.

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

source

Kronecker.naivesample — Function.

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

source

Kronecker.fastsample — Function.

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.

source