Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.858535  0.317476
 0.298127  0.0402255
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.737083   0.272565   0.272565   0.100791
 0.255953   0.034535   0.0946483  0.0127707
 0.255953   0.0946483  0.034535   0.0127707
 0.0888798  0.0119923  0.0119923  0.00161809
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  0.447778   -3.53405   -3.53405   27.8921
 -3.31866     9.55694   26.1922   -75.4273
 -3.31866    26.1922     9.55694  -75.4273
 24.5958    -70.8302   -70.8302   203.974
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.16036     0.0592994    0.0592994    …  2.83521e-6   1.04843e-6
 0.0556852   0.00751347   0.0205918       9.84528e-7   1.3284e-7
 0.0556852   0.0205918    0.00751347      3.59232e-7   1.3284e-7
 0.0193367   0.00260906   0.00260906      1.24744e-7   1.68313e-8
 0.0556852   0.0205918    0.0205918       3.59232e-7   1.3284e-7
 0.0193367   0.00260906   0.00715051   …  1.24744e-7   1.68313e-8
 0.0193367   0.00715051   0.00260906      4.55161e-8   1.68313e-8
 0.0067147   0.000905998  0.000905998     1.58055e-8   2.1326e-9
 0.0556852   0.0205918    0.0205918       3.59232e-7   1.3284e-7
 0.0193367   0.00260906   0.00715051      1.24744e-7   1.68313e-8
 ⋮                                     ⋱               ⋮
 1.41963e-6  1.91546e-7   1.91546e-7      1.04986e-15  1.41655e-16
 1.1773e-5   4.35352e-6   4.35352e-6      2.38615e-14  8.8237e-15
 4.08818e-6  5.51608e-7   1.51176e-6      8.28591e-15  1.118e-15
 4.08818e-6  1.51176e-6   5.51608e-7   …  3.02334e-15  1.118e-15
 1.41963e-6  1.91546e-7   1.91546e-7      1.04986e-15  1.41655e-16
 4.08818e-6  1.51176e-6   1.51176e-6      3.02334e-15  1.118e-15
 1.41963e-6  1.91546e-7   5.24961e-7      1.04986e-15  1.41655e-16
 1.41963e-6  5.24961e-7   1.91546e-7      3.83069e-16  1.41655e-16
 4.92966e-7  6.65147e-8   6.65147e-8   …  1.33021e-16  1.79482e-17
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.109596    0.0375981   0.0375981   0.0128985    …  8.48973e-7   2.9125e-7
 0.0353066   0.0126849   0.0121124   0.00435169      2.73499e-7   9.82621e-8
 0.0353066   0.0121124   0.0126849   0.00435169      2.86426e-7   9.82621e-8
 0.0113742   0.00408647  0.00408647  0.00146817      9.22732e-8   3.31516e-8
 0.0353066   0.0121124   0.0121124   0.00415529      2.86426e-7   9.82621e-8
 0.0113742   0.00408647  0.00390204  0.00140191   …  9.22732e-8   3.31516e-8
 0.0113742   0.00390204  0.00408647  0.00140191      9.66345e-8   3.31516e-8
 0.00366422  0.00131647  0.00131647  0.000472977     3.11311e-8   1.11847e-8
 0.0353066   0.0121124   0.0121124   0.00415529      2.86426e-7   9.82621e-8
 0.0113742   0.00408647  0.00390204  0.00140191      9.22732e-8   3.31516e-8
 ⋮                                                ⋱               ⋮
 4.25092e-7  1.52726e-7  1.52726e-7  5.48708e-8      5.22576e-12  1.87749e-12
 4.09598e-6  1.40517e-6  1.40517e-6  4.82062e-7      4.80803e-11  1.64945e-11
 1.31953e-6  4.74078e-7  4.52682e-7  1.62638e-7      1.54892e-11  5.56492e-12
 1.31953e-6  4.52682e-7  4.74078e-7  1.62638e-7   …  1.62213e-11  5.56492e-12
 4.25092e-7  1.52726e-7  1.52726e-7  5.48708e-8      5.22576e-12  1.87749e-12
 1.31953e-6  4.52682e-7  4.52682e-7  1.55298e-7      1.62213e-11  5.56492e-12
 4.25092e-7  1.52726e-7  1.45833e-7  5.23944e-8      5.22576e-12  1.87749e-12
 4.25092e-7  1.45833e-7  1.52726e-7  5.23944e-8      5.47275e-12  1.87749e-12
 1.36945e-7  4.92011e-8  4.92011e-8  1.76768e-8   …  1.76307e-12  6.33429e-13

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.52989   0.482908
 0.277388  0.703947
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.00174526   0.00159051   0.00159051   …  0.000756772  0.000689674
 0.00091361   0.00231853   0.000832605     0.000396156  0.00100536
 0.00091361   0.000832605  0.00231853      0.00110317   0.00100536
 0.000478258  0.00121371   0.00121371      0.000577487  0.00146553
 0.00091361   0.000832605  0.000832605     0.00110317   0.00100536
 0.000478258  0.00121371   0.000435854  …  0.000577487  0.00146553
 0.000478258  0.000435854  0.00121371      0.00160811   0.00146553
 0.000250359  0.000635355  0.000635355     0.000841818  0.00213634
 0.00091361   0.000832605  0.000832605     0.00110317   0.00100536
 0.000478258  0.00121371   0.000435854     0.000577487  0.00146553
 ⋮                                      ⋱
 5.15197e-6   1.30745e-5   1.30745e-5   …  0.00807743   0.0204987
 1.88005e-5   1.71336e-5   1.71336e-5      0.0105851    0.00964661
 9.84173e-6   2.49761e-5   8.96912e-6      0.00554112   0.0140621
 9.84173e-6   8.96912e-6   2.49761e-5      0.0154302    0.0140621
 5.15197e-6   1.30745e-5   1.30745e-5      0.00807743   0.0204987
 9.84173e-6   8.96912e-6   8.96912e-6   …  0.0154302    0.0140621
 5.15197e-6   1.30745e-5   4.69517e-6      0.00807743   0.0204987
 5.15197e-6   4.69517e-6   1.30745e-5      0.022493     0.0204987
 2.69696e-6   6.84427e-6   6.84427e-6      0.0117747    0.0298815
julia> sum(P10)  # expected number of edges994.3566598985201
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 992 stored entries:
  [611 ,    6]  =  1
  [547 ,    7]  =  1
  [253 ,    9]  =  1
  [85  ,   14]  =  1
  [5   ,   15]  =  1
  [467 ,   23]  =  1
  [831 ,   32]  =  1
  [86  ,   36]  =  1
  [74  ,   41]  =  1
  ⋮
  [138 , 1023]  =  1
  [441 , 1023]  =  1
  [599 , 1023]  =  1
  [746 , 1023]  =  1
  [8   , 1024]  =  1
  [243 , 1024]  =  1
  [367 , 1024]  =  1
  [853 , 1024]  =  1
  [854 , 1024]  =  1
  [1000, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.