Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.826362   0.693322
 0.0202607  0.823575
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.682875     0.572935   0.572935   0.480696
 0.0167427    0.680572   0.0140472  0.571003
 0.0167427    0.0140472  0.680572   0.571003
 0.000410495  0.0166862  0.0166862  0.678276
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  1.52677      -1.2853     -1.2853      1.08203
 -0.0375599     1.53194     0.0316196  -1.28965
 -0.0375599     0.0316196   1.53194    -1.28965
  0.000924006  -0.037687   -0.037687    1.53712
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.101402     0.0850769    …  0.0147047    0.0147047    0.0123373
 0.00248617   0.10106         0.0174673    0.000360529  0.0146551
 0.00248617   0.00208591      0.000360529  0.0174673    0.0146551
 6.09555e-5   0.00247778      0.00042826   0.00042826   0.0174083
 0.00248617   0.00208591      0.0174673    0.0174673    0.0146551
 6.09555e-5   0.00247778   …  0.0207488    0.00042826   0.0174083
 6.09555e-5   5.1142e-5       0.00042826   0.0207488    0.0174083
 1.4945e-6    6.075e-5        0.000508717  0.000508717  0.0206788
 0.00248617   0.00208591      0.0174673    0.0174673    0.0146551
 6.09555e-5   0.00247778      0.0207488    0.00042826   0.0174083
 ⋮                         ⋱                            ⋮
 1.95147e-19  7.93251e-18     0.00201661   0.00201661   0.0819732
 3.24635e-16  2.7237e-16      0.0692422    0.0692422    0.0580946
 7.95936e-18  3.2354e-16      0.0822506    0.00169767   0.0690087
 7.95936e-18  6.67794e-18  …  0.00169767   0.0822506    0.0690087
 1.95147e-19  7.93251e-18     0.00201661   0.00201661   0.0819732
 7.95936e-18  6.67794e-18     0.0822506    0.0822506    0.0690087
 1.95147e-19  7.93251e-18     0.0977029    0.00201661   0.0819732
 1.95147e-19  1.63729e-19     0.00201661   0.0977029    0.0819732
 4.78459e-21  1.94489e-19  …  0.00239547   0.00239547   0.0973734
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 0.0131284    0.0215491    …  3.05924      3.05924      5.02148
 0.000629721  0.0130417       1.85148      0.146741     3.03905
 0.000629721  0.00103363      0.146741     1.85148      3.03905
 3.02054e-5   0.000625565     0.088809     0.088809     1.83927
 0.000629721  0.00103363      1.85148      1.85148      3.03905
 3.02054e-5   0.000625565  …  1.12054      0.088809     1.83927
 3.02054e-5   4.95796e-5      0.088809     1.12054      1.83927
 1.44885e-6   3.00061e-5      0.0537481    0.0537481    1.11314
 0.000629721  0.00103363      1.85148      1.85148      3.03905
 3.02054e-5   0.000625565     1.12054      0.088809     1.83927
 ⋮                         ⋱                            ⋮
 4.05993e-17  8.40826e-16     0.00096742   0.00096742   0.0200356
 1.76459e-14  2.89643e-14     0.0333251    0.0333251    0.0547003
 8.46411e-16  1.75295e-14     0.0201687    0.00159849   0.0331052
 8.46411e-16  1.38931e-15  …  0.00159849   0.0201687    0.0331052
 4.05993e-17  8.40826e-16     0.00096742   0.00096742   0.0200356
 8.46411e-16  1.38931e-15     0.0201687    0.0201687    0.0331052
 4.05993e-17  8.40826e-16     0.0122063    0.00096742   0.0200356
 4.05993e-17  6.66403e-17     0.00096742   0.0122063    0.0200356
 1.9474e-18   4.03314e-17  …  0.000585493  0.000585493  0.0121258

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.479193  0.654404
 0.374714  0.712345
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.000638418  0.000871847  0.000871847  …  0.0105468   0.0105468   0.0144031
 0.000499223  0.00094904   0.000681758     0.0114806   0.0082473   0.0156784
 0.000499223  0.000681758  0.00094904      0.0082473   0.0114806   0.0156784
 0.000390378  0.000742121  0.000742121     0.00897751  0.00897751  0.0170666
 0.000499223  0.000681758  0.000681758     0.0114806   0.0114806   0.0156784
 0.000390378  0.000742121  0.000533114  …  0.0124971   0.00897751  0.0170666
 0.000390378  0.000533114  0.000742121     0.00897751  0.0124971   0.0170666
 0.000305263  0.000580316  0.000580316     0.00977238  0.00977238  0.0185776
 0.000499223  0.000681758  0.000681758     0.0114806   0.0114806   0.0156784
 0.000390378  0.000742121  0.000533114     0.0124971   0.00897751  0.0170666
 ⋮                                      ⋱
 6.97934e-5   0.00013268   0.00013268   …  0.016258    0.016258    0.030907
 0.000114139  0.000155873  0.000155873     0.0191      0.0191      0.0260836
 8.92533e-5   0.000169674  0.000121888     0.0207911   0.0149356   0.0283931
 8.92533e-5   0.000121888  0.000169674     0.0149356   0.0207911   0.0283931
 6.97934e-5   0.00013268   0.00013268      0.016258    0.016258    0.030907
 8.92533e-5   0.000121888  0.000121888  …  0.0207911   0.0207911   0.0283931
 6.97934e-5   0.00013268   9.53125e-5      0.0226319   0.016258    0.030907
 6.97934e-5   9.53125e-5   0.00013268      0.016258    0.0226319   0.030907
 5.45763e-5   0.000103751  0.000103751     0.0176975   0.0176975   0.0336435
julia> sum(P10)  # expected number of edges2916.166214092922
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 2908 stored entries:
  [540 ,    3]  =  1
  [420 ,   11]  =  1
  [251 ,   16]  =  1
  [547 ,   16]  =  1
  [7   ,   17]  =  1
  [722 ,   18]  =  1
  [178 ,   21]  =  1
  [315 ,   21]  =  1
  [889 ,   21]  =  1
  ⋮
  [548 , 1024]  =  1
  [645 , 1024]  =  1
  [669 , 1024]  =  1
  [740 , 1024]  =  1
  [748 , 1024]  =  1
  [818 , 1024]  =  1
  [830 , 1024]  =  1
  [851 , 1024]  =  1
  [942 , 1024]  =  1
  [1013, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.