Kronecker powers and graphs

Kronecker powers

Repeated Kronecker multiplications of the same matrix, i.e.

\[A^{\otimes n}=\otimes_{i=1}^n A = \underbrace{A \otimes A \otimes \ldots \otimes A}_{n\text{ times}}\,.\]

Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.

julia> using Kronecker
julia> A = rand(2, 2)2×2 Array{Float64,2}:
 0.371962  0.590668
 0.46572   0.112158
julia> K2 = A ⊗ 24×4 KroneckerPower{Float64,Array{Float64,2}}:
 0.138356  0.219706   0.219706   0.348888
 0.17323   0.0417186  0.275086   0.0662483
 0.17323   0.275086   0.0417186  0.0662483
 0.216896  0.0522344  0.0522344  0.0125795
julia> inv(K2)4×4 KroneckerPower{Float64,Array{Float64,2}}:
  0.230984  -1.21645   -1.21645    6.40628
 -0.959127   0.766036   5.05112   -4.03423
 -0.959127   5.05112    0.766036  -4.03423
  3.98263   -3.18085   -3.18085    2.54048
julia> K12 = K2 ⊗ 6  # works recursively4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 7.01426e-6   1.11385e-5   1.11385e-5   …  0.00113573   0.00180352
 8.78231e-6   2.11502e-6   1.39461e-5      0.00142201   0.000342459
 8.78231e-6   1.39461e-5   2.11502e-6      0.000215657  0.000342459
 1.0996e-5    2.64814e-6   2.64814e-6      0.000270016  6.50274e-5
 8.78231e-6   1.39461e-5   1.39461e-5      0.000215657  0.000342459
 1.0996e-5    2.64814e-6   1.74615e-5   …  0.000270016  6.50274e-5
 1.0996e-5    1.74615e-5   2.64814e-6      4.09498e-5   6.50274e-5
 1.37677e-5   3.31565e-6   3.31565e-6      5.12717e-5   1.23476e-5
 8.78231e-6   1.39461e-5   1.39461e-5      0.000215657  0.000342459
 1.0996e-5    2.64814e-6   1.74615e-5      0.000270016  6.50274e-5
 ⋮                                      ⋱               ⋮
 8.31528e-5   2.00255e-5   2.00255e-5      8.66529e-11  2.08684e-11
 5.30424e-5   8.42302e-5   8.42302e-5      3.64475e-10  5.78779e-10
 6.64125e-5   1.5994e-5    0.000105462     4.56347e-10  1.09901e-10
 6.64125e-5   0.000105462  1.5994e-5    …  6.9208e-11   1.09901e-10
 8.31528e-5   2.00255e-5   2.00255e-5      8.66529e-11  2.08684e-11
 6.64125e-5   0.000105462  0.000105462     6.9208e-11   1.09901e-10
 8.31528e-5   2.00255e-5   0.000132045     8.66529e-11  2.08684e-11
 8.31528e-5   0.000132045  2.00255e-5      1.31415e-11  2.08684e-11
 0.000104113  2.50732e-5   2.50732e-5   …  1.6454e-11   3.96257e-12
julia> K12^2  # example4096×4096 KroneckerPower{Float64,KroneckerPower{Float64,Array{Float64,2}}}:
 2.49443e-5  1.72526e-5  1.72526e-5  …  4.32192e-7  4.32192e-7  2.98923e-7
 1.36031e-5  1.73558e-5  9.40846e-6     4.34778e-7  2.3569e-7   3.00711e-7
 1.36031e-5  9.40846e-6  1.73558e-5     2.3569e-7   4.34778e-7  3.00711e-7
 7.41823e-6  9.46477e-6  9.46477e-6     2.371e-7    2.371e-7    3.02511e-7
 1.36031e-5  9.40846e-6  9.40846e-6     4.34778e-7  4.34778e-7  3.00711e-7
 7.41823e-6  9.46477e-6  5.13077e-6  …  4.3738e-7   2.371e-7    3.02511e-7
 7.41823e-6  5.13077e-6  9.46477e-6     2.371e-7    4.3738e-7   3.02511e-7
 4.04543e-6  5.16148e-6  5.16148e-6     2.38519e-7  2.38519e-7  3.04322e-7
 1.36031e-5  9.40846e-6  9.40846e-6     4.34778e-7  4.34778e-7  3.00711e-7
 7.41823e-6  9.46477e-6  5.13077e-6     4.3738e-7   2.371e-7    3.02511e-7
 ⋮                                   ⋱                          ⋮
 3.1643e-8   4.03727e-8  4.03727e-8     2.50181e-7  2.50181e-7  3.192e-7
 1.06402e-7  7.35922e-8  7.35922e-8     4.56035e-7  4.56035e-7  3.15414e-7
 5.80248e-8  7.40327e-8  4.01325e-8     4.58765e-7  2.48692e-7  3.17301e-7
 5.80248e-8  4.01325e-8  7.40327e-8  …  2.48692e-7  4.58765e-7  3.17301e-7
 3.1643e-8   4.03727e-8  4.03727e-8     2.50181e-7  2.50181e-7  3.192e-7
 5.80248e-8  4.01325e-8  4.01325e-8     4.58765e-7  4.58765e-7  3.17301e-7
 3.1643e-8   4.03727e-8  2.18857e-8     4.6151e-7   2.50181e-7  3.192e-7
 3.1643e-8   2.18857e-8  4.03727e-8     2.50181e-7  4.6151e-7   3.192e-7
 1.72561e-8  2.20167e-8  2.20167e-8  …  2.51678e-7  2.51678e-7  3.21111e-7

Kronecker.kronecker — Method

kronecker(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker.:⊗ — Method

⊗(A::AbstractMatrix, pow::Int)

Kronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.

Kronecker graphs

An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).

If we use an initial matrix $P$ with values in $[0,1]$ as a seed, then $P_n=P^{\otimes n}$ can be seen as a probability distribution for an adjacency matrix of a graph. The elements $(P_n)_{i,j}$ give the probabilities that there is an edge from node $i$ to node $j$. Leskovec and co-authors give two algorithms for sampling adjacency matrices from this distribution, both provided by Kronecker.jl:

naivesample gives exact samples, but has a computational time proportional to the number of elements and hence prohibitive for large graphs;

fastsample a recursive heuristic, has a computational time proportional to the expected number of edges in the graph.

The latter can easily scale to generate graphs of millions of nodes. Both return the adjacency matrix as a sparse array.

julia> using Kronecker
julia> P = rand(2, 2)2×2 Array{Float64,2}:
 0.767077  0.859322
 0.915816  0.281305
julia> P10 = P ⊗ 101024×1024 KroneckerPower{Float64,Array{Float64,2}}:
 0.0705328  0.0790147  0.0790147  …  0.195993    0.195993    0.219562
 0.0842093  0.025866   0.0943359     0.0641597   0.233997    0.0718752
 0.0842093  0.0943359  0.025866      0.233997    0.0641597   0.0718752
 0.100538   0.0308815  0.0308815     0.0766005   0.0766005   0.0235289
 0.0842093  0.0943359  0.0943359     0.0641597   0.0641597   0.0718752
 0.100538   0.0308815  0.112628   …  0.0210032   0.0766005   0.0235289
 0.100538   0.112628   0.0308815     0.0766005   0.0210032   0.0235289
 0.120032   0.0368696  0.0368696     0.0250757   0.0250757   0.00770236
 0.0842093  0.0943359  0.0943359     0.0641597   0.0641597   0.0718752
 0.100538   0.0308815  0.112628      0.0210032   0.0766005   0.0235289
 ⋮                                ⋱
 0.347627   0.106778   0.106778   …  3.08594e-5  3.08594e-5  9.47888e-6
 0.24388    0.273207   0.273207      7.89579e-5  7.89579e-5  8.8453e-5
 0.291169   0.0894364  0.326183      2.58475e-5  9.42682e-5  2.89557e-5
 0.291169   0.326183   0.0894364     9.42682e-5  2.58475e-5  2.89557e-5
 0.347627   0.106778   0.106778      3.08594e-5  3.08594e-5  9.47888e-6
 0.291169   0.326183   0.326183   …  2.58475e-5  2.58475e-5  2.89557e-5
 0.347627   0.106778   0.389431      8.46136e-6  3.08594e-5  9.47888e-6
 0.347627   0.389431   0.106778      3.08594e-5  8.46136e-6  9.47888e-6
 0.415033   0.127483   0.127483      1.0102e-5   1.0102e-5   3.10298e-6
julia> sum(P10)  # expected number of edges32203.93973422168
julia> A = fastsample(P10)1024×1024 SparseArrays.SparseMatrixCSC{Bool,Int64} with 30670 stored entries:
  [18 ,    1]  =  1
  [32 ,    1]  =  1
  [40 ,    1]  =  1
  [47 ,    1]  =  1
  [57 ,    1]  =  1
  [63 ,    1]  =  1
  [66 ,    1]  =  1
  [80 ,    1]  =  1
  [95 ,    1]  =  1
  ⋮
  [34 , 1023]  =  1
  [263, 1023]  =  1
  [297, 1023]  =  1
  [394, 1023]  =  1
  [609, 1023]  =  1
  [3  , 1024]  =  1
  [65 , 1024]  =  1
  [267, 1024]  =  1
  [269, 1024]  =  1
  [330, 1024]  =  1

Kronecker.isprob — Function

isprob(A::AbstractArray)

Test if a matrix can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

isprob(K::AbstractKroneckerProduct)

Test if a Kronecker product can be interpeted as a probability matrix, i.e., all elements are between 0 and 1.

Kronecker.naivesample — Function

naivesample(P::AbstractKroneckerProduct)

Sample a Kronecker graph from a probabilistic Kronecker product P using the naive method. This method has a time complexity in the size of the Kronecker product (but is still light in memory use). Consider using fastsample.

Kronecker.fastsample — Function

fastsample(P::AbstractKroneckerProduct)

Uses the heuristic sampling from Leskovec et al. (2008) to sample a large Kronecker graph.