Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.85414   0.512764  0.947678   0.194803  …  0.0       0.0        0.0
 0.47661   0.449491  0.682439   0.382124     0.0       0.0        0.0
 0.502726  0.703252  0.563821   0.064931     0.0       0.0        0.0
 0.248631  0.237131  0.0880059  0.664142     0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       1.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       0.0        1.0
 0.0       0.0       0.0        0.0          0.512764  0.947678   0.194803
 0.0       0.0       0.0        0.0          1.44949   0.682439   0.382124
 0.0       0.0       0.0        0.0          0.703252  1.56382    0.064931
 0.0       0.0       0.0        0.0          0.237131  0.0880059  1.66414

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 3.47594   2.01298   2.80333   0.865908  …   2.88076    4.01183   1.23919
 1.63722   2.54217   1.98851   0.969585      3.63808    2.84574   1.38757
 1.66916   1.78742   2.92919   0.579088      2.55796    4.19194   0.828729
 0.835056  0.755356  0.690924  2.14513       1.08098    0.988776  3.06988
 0.0       0.0       0.0       0.0           0.80393    1.11958   0.34582
 0.0       0.0       0.0       0.0       …   1.01527    0.794156  0.387226
 0.0       0.0       0.0       0.0           0.713846   1.16984   0.231272
 0.0       0.0       0.0       0.0           0.301669   0.275936  0.856706
 0.0       0.0       0.0       0.0           2.41235    3.35951   1.0377
 0.0       0.0       0.0       0.0           3.04653    2.38302   1.16195
 0.0       0.0       0.0       0.0       …   2.14204    3.51033   0.693978
 0.0       0.0       0.0       0.0           0.905216   0.828002  2.57072
 0.0       0.0       0.0       0.0           6.64022    9.24736   2.85637
 0.0       0.0       0.0       0.0           8.38586    6.55949   3.19837
 0.0       0.0       0.0       0.0           5.89615    9.66252   1.91024
 0.0       0.0       0.0       0.0       …   2.49169    2.27915   7.07614
 0.0       0.0       0.0       0.0           8.48544   11.8171    3.65011
 0.0       0.0       0.0       0.0          10.7162     8.38228   4.08715
 0.0       0.0       0.0       0.0           7.53461   12.3476    2.44107
 0.0       0.0       0.0       0.0           3.1841     2.9125    9.04249

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 112 stored entries:
  [1 ,  1]  =  0.85414
  [2 ,  1]  =  0.47661
  [3 ,  1]  =  0.502726
  [4 ,  1]  =  0.248631
  [1 ,  2]  =  0.512764
  [2 ,  2]  =  0.449491
  [3 ,  2]  =  0.703252
  [4 ,  2]  =  0.237131
  [1 ,  3]  =  0.947678
  ⋮
  [17, 19]  =  0.947678
  [18, 19]  =  0.682439
  [19, 19]  =  1.56382
  [20, 19]  =  0.0880059
  [16, 20]  =  1.0
  [17, 20]  =  0.194803
  [18, 20]  =  0.382124
  [19, 20]  =  0.064931
  [20, 20]  =  1.66414

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.