Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.43332   0.94288   0.0907667  0.82888   …  0.0       0.0        0.0
 0.847203  1.678     0.0480094  0.338243     1.0       0.0        0.0
 0.967467  0.176285  1.63612    0.786784     0.0       1.0        0.0
 0.530199  0.243416  0.956805   1.78034      0.0       0.0        1.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       1.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        1.0          0.0       0.0        0.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       1.0        0.0       …  0.0       1.0        0.0
 0.0       0.0       0.0        1.0          0.0       0.0        1.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.94288   0.0907667  0.82888
 0.0       0.0       0.0        0.0          0.677996  0.0480094  0.338243
 0.0       0.0       0.0        0.0          0.176285  0.636123   0.786784
 0.0       0.0       0.0        0.0          0.243416  0.956805   0.780342

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 15.5572   11.7117    6.04343   12.6856   …   8.79926   4.54054   9.53092
 10.9965   14.6663    3.98932    8.85688     11.0191    2.99725   6.65435
 14.7269    8.96603  15.0945    15.1239       6.73635  11.3408   11.3629
 13.0173    8.49948  12.3131    19.6528       6.38582   9.25108  14.7655
 11.6884    8.79926   4.54054    9.53092      6.02295   3.10793   6.52376
  8.2619   11.0191    2.99725    6.65435  …   7.54239   2.05157   4.55479
 11.0646    6.73635  11.3408    11.3629       4.61092   7.76257   7.77773
  9.78016   6.38582   9.25108   14.7655       4.37099   6.33221  10.1068
 14.9472   11.2525    5.80645   12.1881      10.2362    5.28202  11.0873
 10.5653   14.0912    3.83289    8.50958     12.8185    3.4867    7.741
 14.1494    8.61445  14.5026    14.5309   …   7.8364   13.1927   13.2185
 12.5069    8.16619  11.8303    18.8822       7.42863  10.7618   17.1768
 10.9799    8.26589   4.26532    8.9532       6.16558   3.18153   6.67824
  7.7611   10.3512    2.81557    6.25099      7.72099   2.10015   4.66265
 10.3939    6.32802  10.6533    10.6741       4.72011   7.94639   7.96191
  9.18734   5.99874   8.69032   13.8705   …   4.4745    6.48216  10.3461
  2.70004   2.03263   1.04887    2.20165      3.9288    2.02732   4.25548
  1.9085    2.54541   0.692366   1.53716      4.91994   1.33825   2.97111
  2.55592   1.5561    2.61972    2.62484      3.00773   5.06356   5.07345
  2.25922   1.47513   2.137      3.41085      2.85122   4.13053   6.5927

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  1.43332
  [2 ,  1]  =  0.847203
  [3 ,  1]  =  0.967467
  [4 ,  1]  =  0.530199
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [1 ,  2]  =  0.94288
  [2 ,  2]  =  1.678
  [3 ,  2]  =  0.176285
  ⋮
  [18, 19]  =  0.0480094
  [19, 19]  =  0.636123
  [20, 19]  =  0.956805
  [4 , 20]  =  1.0
  [12, 20]  =  1.0
  [17, 20]  =  0.82888
  [18, 20]  =  0.338243
  [19, 20]  =  0.786784
  [20, 20]  =  0.780342

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.