Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.403844  0.709011  0.238025  0.399104  …  0.0       0.0       0.0
 0.892223  0.552259  0.835953  0.201028     1.0       0.0       0.0
 0.482021  0.269272  0.409361  0.141171     0.0       1.0       0.0
 0.797439  0.255671  0.996711  0.456311     0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.709011  0.238025  0.399104
 0.0       1.0       0.0       0.0          1.55226   0.835953  0.201028
 0.0       0.0       1.0       0.0          0.269272  1.40936   0.141171
 0.0       0.0       0.0       1.0          0.255671  0.996711  1.45631
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  8.89174    5.48968   4.71589   3.22046   …   6.49014    5.57534    3.80738
  8.01873    9.30152   7.2467    3.04966      10.9967     8.56738    3.60545
  4.27134    3.07235   7.0682    1.7492        3.63227    8.35635    2.06798
  7.61016    4.72597   7.82016   7.28739       5.58726    9.24535    8.61549
  7.74988    4.7847    4.11028   2.80689       6.37306    5.47476    3.73869
  6.98898    8.10703   6.31609   2.65803   …  10.7983     8.41282    3.54041
  3.72282    2.67781   6.16051   1.52457       3.56675    8.2056     2.03067
  6.63287    4.11907   6.81591   6.35156       5.48646    9.07856    8.46006
  3.46065    2.13657   1.83541   1.2534        2.68478    2.30635    1.575
  3.12087    3.62013   2.8204    1.18692       4.549      3.54407    1.49147
  1.6624     1.19575   2.75093   0.680783  …   1.50256    3.45677    0.855461
  2.96186    1.83934   3.04359   2.83624       2.31128    3.82453    3.56397
  0.887948   0.54821   0.470939  0.321602      0.665296   0.57152    0.390289
  0.800768   0.928869  0.723671  0.304546      1.12726    0.87823    0.36959
  0.426545   0.306812  0.705845  0.174678      0.37234    0.856598   0.211986
  0.759967   0.471945  0.780938  0.727734  …   0.572742   0.947729   0.883162
 11.2105     6.92127   5.94569   4.06029      10.7055     9.19653    6.28027
 10.1099    11.7272    9.13649   3.84495      18.139     14.1319     5.9472
  5.38522    3.87356   8.91144   2.20535       5.99144   13.7838     3.41113
  9.59473    5.9584    9.8595    9.18779       9.21619   15.2502    14.2113
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  0.403844
  [2 ,  1]  =  0.892223
  [3 ,  1]  =  0.482021
  [4 ,  1]  =  0.797439
  [5 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.709011
  [2 ,  2]  =  0.552259
  [3 ,  2]  =  0.269272
  ⋮
  [17, 19]  =  0.238025
  [18, 19]  =  0.835953
  [19, 19]  =  1.40936
  [20, 19]  =  0.996711
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [17, 20]  =  0.399104
  [18, 20]  =  0.201028
  [19, 20]  =  0.141171
  [20, 20]  =  1.45631

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.