Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.370933  0.12606   0.197689    0.820028  …  0.0       0.0         0.0
 0.460282  0.922013  0.00709703  0.148476     1.0       0.0         0.0
 0.821266  0.432017  0.354663    0.945202     0.0       1.0         0.0
 0.440222  0.645712  0.238561    0.109714     0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       1.0       0.0         0.0       …  1.0       0.0         0.0
 0.0       0.0       1.0         0.0          0.0       1.0         0.0
 0.0       0.0       0.0         1.0          0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       1.0       0.0         0.0          1.0       0.0         0.0
 0.0       0.0       1.0         0.0       …  0.0       1.0         0.0
 0.0       0.0       0.0         1.0          0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       1.0       0.0         0.0          1.0       0.0         0.0
 0.0       0.0       1.0         0.0          0.0       1.0         0.0
 0.0       0.0       0.0         1.0       …  0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.12606   0.197689    0.820028
 0.0       1.0       0.0         0.0          1.92201   0.00709703  0.148476
 0.0       0.0       1.0         0.0          0.432017  1.35466     0.945202
 0.0       0.0       0.0         1.0          0.645712  0.238561    1.10971
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.8459    5.08609   2.75989    8.14969  …   6.10468   3.31261    9.78184
  6.22789  16.223     0.930343   3.79449     19.472     1.11666    4.55441
 11.3109    9.76235  10.4479    11.6423      11.7175   12.5403    13.9739
  6.30976   8.26291   2.55414    9.96189      9.91773   3.06566   11.957
  7.37425   3.16617   1.71808    5.07332      4.12613   2.23898    6.61151
  3.87697  10.0991    0.579154   2.36213  …  13.161     0.754749   3.07831
  7.04121   6.07723   6.504      7.24752      7.91979   8.47596    9.4449
  3.92793   5.1438    1.59       6.20144      6.70336   2.07207    8.08167
 14.2183    6.10468   3.31261    9.78184      7.77002   4.21628   12.4503
  7.47516  19.472     1.11666    4.55441     24.7839    1.42129    5.79684
 13.5761   11.7175   12.5403    13.9739   …  14.914    15.9613    17.7859
  7.57343   9.91773   3.06566   11.957       12.6233    3.90197   15.2188
 10.3396    4.43934   2.40894    7.11338      5.71251   3.09981    9.15344
  5.43596  14.1601    0.812041   3.31198     18.2211    1.04493    4.26183
  9.87259   8.52097   9.11937   10.1619      10.9647   11.7347    13.0762
  5.50742   7.2122    2.22936    8.69514  …   9.28061   2.86872   11.1888
 10.3396    4.43934   2.40894    7.11338      6.58692   3.57429   10.5546
  5.43596  14.1601    0.812041   3.31198     21.0102    1.20487    4.91418
  9.87259   8.52097   9.11937   10.1619      12.6431   13.531     15.0778
  5.50742   7.2122    2.22936    8.69514     10.7012    3.30783   12.9015
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 128 stored entries:
  [1 ,  1]  =  0.370933
  [2 ,  1]  =  0.460282
  [3 ,  1]  =  0.821266
  [4 ,  1]  =  0.440222
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.12606
  ⋮
  [19, 19]  =  1.35466
  [20, 19]  =  0.238561
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.820028
  [18, 20]  =  0.148476
  [19, 20]  =  0.945202
  [20, 20]  =  1.10971

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.