Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.66821   0.0935565  0.228663  0.992723   …  0.0        0.0       0.0
 0.473251  1.92391    0.807948  0.438733      0.0        0.0       0.0
 0.654597  0.195782   1.80197   0.0123259     0.0        0.0       0.0
 0.498509  0.326874   0.178778  1.39185       0.0        0.0       0.0
 1.0       0.0        0.0       0.0           0.0        0.0       0.0
 0.0       1.0        0.0       0.0        …  1.0        0.0       0.0
 0.0       0.0        1.0       0.0           0.0        1.0       0.0
 0.0       0.0        0.0       1.0           0.0        0.0       1.0
 1.0       0.0        0.0       0.0           0.0        0.0       0.0
 0.0       1.0        0.0       0.0           0.0        0.0       0.0
 0.0       0.0        1.0       0.0        …  0.0        0.0       0.0
 0.0       0.0        0.0       1.0           0.0        0.0       0.0
 0.0       0.0        0.0       0.0           0.0        0.0       0.0
 0.0       0.0        0.0       0.0           1.0        0.0       0.0
 0.0       0.0        0.0       0.0           0.0        1.0       0.0
 0.0       0.0        0.0       0.0        …  0.0        0.0       1.0
 1.0       0.0        0.0       0.0           0.0935565  0.228663  0.992723
 0.0       1.0        0.0       0.0           0.923914   0.807948  0.438733
 0.0       0.0        1.0       0.0           0.195782   0.801973  0.0123259
 0.0       0.0        0.0       1.0           0.326874   0.178778  0.391855
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.9757    2.8236     3.96027    8.75081  …  0.207587  0.291153  0.643346
  9.15943  13.1393    10.1885     7.28019     0.965979  0.749041  0.535228
  7.38682   2.85373   11.5612     3.55752     0.209802  0.849961  0.261543
  5.7871    3.62272    3.54287    9.33115     0.266336  0.260466  0.686011
 14.5215    3.42383    4.80213   10.611       1.95916   2.74785   6.07178
 11.1065   15.9324    12.3543     8.82779  …  9.11672   7.06931   5.05138
  8.95709   3.46037   14.0188     4.31376     1.98007   8.02177   2.46839
  7.0173    4.39282    4.296     11.3147      2.51363   2.45823   6.47445
 10.0703    2.37434    3.33015    7.35847     0.57958   0.812897  1.79622
  7.70207  11.0487     8.56739    6.12184     2.697     2.09132   1.49435
  6.21151   2.39968    9.72169    2.99148  …  0.585766  2.37308   0.730226
  4.86632   3.04631    2.97916    7.84647     0.743609  0.727219  1.91534
  3.3386    0.787166   1.10405    2.43956     0.854744  1.19883   2.649
  2.55348   3.66298    2.84036    2.02958     3.97745   3.0842    2.20382
  2.05931   0.795568   3.22305    0.99177     0.863866  3.49974   1.07691
  1.61334   1.00995    0.987685   2.60135  …  1.09665   1.07248   2.82467
 10.3024    2.42908    3.40693    7.52812     1.38927   1.94854   4.3056
  7.87965  11.3034     8.76491    6.26298     6.46482   5.01296   3.58202
  6.35472   2.455      9.94583    3.06045     1.4041    5.68837   1.75038
  4.97851   3.11654    3.04785    8.02737     1.78246   1.74317   4.59114
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 116 stored entries:
  [1 ,  1]  =  1.66821
  [2 ,  1]  =  0.473251
  [3 ,  1]  =  0.654597
  [4 ,  1]  =  0.498509
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.0935565
  [2 ,  2]  =  1.92391
  ⋮
  [17, 19]  =  0.228663
  [18, 19]  =  0.807948
  [19, 19]  =  0.801973
  [20, 19]  =  0.178778
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.992723
  [18, 20]  =  0.438733
  [19, 20]  =  0.0123259
  [20, 20]  =  0.391855

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.