Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.409248  0.576835   0.852094   0.543245  …  0.0        0.0        0.0
 0.534774  0.983527   0.0831757  0.690091     1.0        0.0        0.0
 0.238148  0.0547912  0.0264389  0.382985     0.0        1.0        0.0
 0.918272  0.657692   0.127067   0.110456     0.0        0.0        1.0
 1.0       0.0        0.0        0.0          0.0        0.0        0.0
 0.0       1.0        0.0        0.0       …  0.0        0.0        0.0
 0.0       0.0        1.0        0.0          0.0        0.0        0.0
 0.0       0.0        0.0        1.0          0.0        0.0        0.0
 0.0       0.0        0.0        0.0          0.0        0.0        0.0
 0.0       0.0        0.0        0.0          1.0        0.0        0.0
 0.0       0.0        0.0        0.0       …  0.0        1.0        0.0
 0.0       0.0        0.0        0.0          0.0        0.0        1.0
 1.0       0.0        0.0        0.0          0.0        0.0        0.0
 0.0       1.0        0.0        0.0          0.0        0.0        0.0
 0.0       0.0        1.0        0.0          0.0        0.0        0.0
 0.0       0.0        0.0        1.0       …  0.0        0.0        0.0
 1.0       0.0        0.0        0.0          0.576835   0.852094   0.543245
 0.0       1.0        0.0        0.0          1.98353    0.0831757  0.690091
 0.0       0.0        1.0        0.0          0.0547912  1.02644    0.382985
 0.0       0.0        0.0        1.0          0.657692   0.127067   1.11046

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.2073    7.49455  5.63888   6.11318  …   6.82858    5.13781   5.56996
  7.82948  14.7561   3.2274    7.40071     13.4448     2.94061   6.74308
  2.6923    2.083    4.9892    2.67414      1.89791    4.54586   2.43651
  7.64652   7.79519  3.36154   8.28797      7.10251    3.06284   7.5515
  7.35643   5.40136  4.06397   4.4058       3.49427    2.62908   2.85022
  5.64274  10.6348   2.326     5.33373  …   6.87989    1.50475   3.45052
  1.94035   1.50123  3.59575   1.92727      0.971184   2.32618   1.2468
  5.51089   5.61804  2.42268   5.97318      3.63445    1.56729   3.8642
  9.92044   7.28395  5.48042   5.9414       9.60924    7.22997   7.8381
  7.60946  14.3414   3.13671   7.19274     18.9197     4.13806   9.48892
  2.61665   2.02447  4.84901   2.59899  …   2.67075    6.39698   3.42869
  7.43165   7.57615  3.26708   8.05507      9.99472    4.31005  10.6265
  9.30025   6.82858  5.13781   5.56996      5.70902    4.29546   4.65676
  7.13375  13.4448   2.94061   6.74308     11.2405     2.45849   5.63754
  2.45306   1.89791  4.54586   2.43651      1.58674    3.80056   2.03704
  6.96705   7.10251  3.06284   7.5515   …   5.93804    2.56068   6.31342
 16.8745   12.3899   9.32213  10.1062      14.4708    10.8878   11.8036
 12.9436   24.3945   5.3355   12.2348      28.4917     6.23162  14.2896
  4.45088   3.4436   8.2481    4.42086      4.02196    9.63339   5.16336
 12.6411   12.8869   5.55727  13.7016      15.0513     6.49063  16.0028

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 132 stored entries:
  [1 ,  1]  =  0.409248
  [2 ,  1]  =  0.534774
  [3 ,  1]  =  0.238148
  [4 ,  1]  =  0.918272
  [5 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.576835
  [2 ,  2]  =  0.983527
  ⋮
  [17, 19]  =  0.852094
  [18, 19]  =  0.0831757
  [19, 19]  =  1.02644
  [20, 19]  =  0.127067
  [4 , 20]  =  1.0
  [12, 20]  =  1.0
  [17, 20]  =  0.543245
  [18, 20]  =  0.690091
  [19, 20]  =  0.382985
  [20, 20]  =  1.11046

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.