Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.43632   0.26796   0.339815  0.545852    …  0.0       0.0       0.0
 0.611084  1.69582   0.393797  0.00293739     1.0       0.0       0.0
 0.70851   0.880933  1.44733   0.433131       0.0       1.0       0.0
 0.624493  0.182695  0.976575  1.13929        0.0       0.0       1.0
 0.0       0.0       0.0       0.0            0.0       0.0       0.0
 0.0       0.0       0.0       0.0         …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0            0.0       0.0       0.0
 0.0       0.0       0.0       0.0            0.0       0.0       0.0
 1.0       0.0       0.0       0.0            0.0       0.0       0.0
 0.0       1.0       0.0       0.0            1.0       0.0       0.0
 0.0       0.0       1.0       0.0         …  0.0       1.0       0.0
 0.0       0.0       0.0       1.0            0.0       0.0       1.0
 1.0       0.0       0.0       0.0            0.0       0.0       0.0
 0.0       1.0       0.0       0.0            0.0       0.0       0.0
 0.0       0.0       1.0       0.0            0.0       0.0       0.0
 0.0       0.0       0.0       1.0         …  0.0       0.0       0.0
 1.0       0.0       0.0       0.0            0.26796   0.339815  0.545852
 0.0       1.0       0.0       0.0            1.69582   0.393797  0.00293739
 0.0       0.0       1.0       0.0            0.880933  1.44733   0.433131
 0.0       0.0       0.0       1.0            0.182695  0.976575  1.13929

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 19.549     8.82375   9.98102   8.5019   …   6.57664   7.4392    6.33676
 13.0636   21.4895    9.33172   4.18073     16.0168    6.95525   3.11604
 17.7091   17.9292   21.6348    9.31833     13.3633   16.1252    6.94527
 15.1227   11.0343   15.8983   15.0367       8.22421  11.8496   11.2074
  8.54342   3.85621   4.36197   3.71555      2.55129   2.8859    2.45823
  5.70913   9.39147   4.07821   1.82709  …   6.21344   2.69816   1.20881
  7.73933   7.83555   9.45498   4.07235      5.18404   6.25547   2.69429
  6.60902   4.82226   6.94799   6.57144      3.19043   4.59682   4.3477
 11.6795    5.27172   5.96313   5.07944      4.59224   5.19453   4.42473
  7.8048   12.8388    5.57521   2.49777     11.184     4.85661   2.17582
 10.5802   10.7118   12.9257    5.5672   …   9.33111  11.2596    4.84963
  9.03501   6.59238   9.49841   8.98364      5.74267   8.27414   7.82571
 10.1741    4.59224   5.19453   4.42473      2.43868   2.75853   2.34973
  6.79882  11.184     4.85661   2.17582      5.9392    2.57908   1.15546
  9.21652   9.33111  11.2596    4.84963      4.95524   5.97937   2.57537
  7.87047   5.74267   8.27414   7.82571  …   3.04962   4.39394   4.15581
 16.8329    7.5978    8.59428   7.32067      6.83934   7.73635   6.58988
 11.2486   18.5038    8.0352    3.59987     16.6566    7.23307   3.24051
 15.2486   15.4382   18.6289    8.02366     13.8971   16.7693    7.22269
 13.0216    9.50119  13.6895   12.9476       8.55272  12.3229   11.655

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 124 stored entries:
  [1 ,  1]  =  1.43632
  [2 ,  1]  =  0.611084
  [3 ,  1]  =  0.70851
  [4 ,  1]  =  0.624493
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.26796
  [2 ,  2]  =  1.69582
  ⋮
  [17, 19]  =  0.339815
  [18, 19]  =  0.393797
  [19, 19]  =  1.44733
  [20, 19]  =  0.976575
  [4 , 20]  =  1.0
  [12, 20]  =  1.0
  [17, 20]  =  0.545852
  [18, 20]  =  0.00293739
  [19, 20]  =  0.433131
  [20, 20]  =  1.13929

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.