Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.171165   0.926718   0.892444  0.68366   …  0.0       0.0       0.0
 0.350571   0.0793394  0.615859  0.775006     0.0       0.0       0.0
 0.0855815  0.264787   0.763303  0.803108     0.0       0.0       0.0
 0.467412   0.46128    0.203186  0.348551     0.0       0.0       0.0
 0.0        0.0        0.0       0.0          0.0       0.0       0.0
 0.0        0.0        0.0       0.0       …  1.0       0.0       0.0
 0.0        0.0        0.0       0.0          0.0       1.0       0.0
 0.0        0.0        0.0       0.0          0.0       0.0       1.0
 1.0        0.0        0.0       0.0          0.0       0.0       0.0
 0.0        1.0        0.0       0.0          1.0       0.0       0.0
 0.0        0.0        1.0       0.0       …  0.0       1.0       0.0
 0.0        0.0        0.0       1.0          0.0       0.0       1.0
 1.0        0.0        0.0       0.0          0.0       0.0       0.0
 0.0        1.0        0.0       0.0          1.0       0.0       0.0
 0.0        0.0        1.0       0.0          0.0       1.0       0.0
 0.0        0.0        0.0       1.0       …  0.0       0.0       1.0
 0.0        0.0        0.0       0.0          0.926718  0.892444  0.68366
 0.0        0.0        0.0       0.0          1.07934   0.615859  0.775006
 0.0        0.0        0.0       0.0          0.264787  1.7633    0.803108
 0.0        0.0        0.0       0.0          0.46128   0.203186  1.34855
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 5.40145  5.22773  6.79763  6.79524  …   5.75928   7.48881   7.48618
 2.44661  5.23877  4.63537  5.36072      5.77144   5.10669   5.9058
 1.76175  2.74278  7.79729  5.52282      3.02166   8.59011   6.08438
 2.56128  3.15564  3.20561  6.60008      3.4765    3.53156   7.27118
 5.80188  5.61528  7.30156  7.299       15.1703   19.726    19.7191
 2.62798  5.62713  4.97901  5.75813  …  15.2023   13.4514   15.5562
 1.89236  2.94611  8.37533  5.93225      7.95926  22.6269   16.0266
 2.75115  3.38957  3.44326  7.08937      9.15732   9.30235  19.1527
 5.45511  5.27967  6.86517  6.86275      9.69904  12.6117   12.6073
 2.47091  5.29081  4.68142  5.41398      9.71952   8.60004   9.94578
 1.77925  2.77003  7.87475  5.57769  …   5.0887   14.4664   10.2465
 2.58672  3.18699  3.23746  6.66565      5.85467   5.9474   12.2452
 5.21522  5.04749  6.56326  6.56095     10.8068   14.0521   14.0471
 2.36225  5.05814  4.47555  5.17589     10.8296    9.58224  11.0817
 1.70101  2.64821  7.52845  5.3324       5.66988  16.1186   11.4168
 2.47297  3.04683  3.09509  6.37252  …   6.52333   6.62665  13.6437
 3.99397  3.86552  5.02634  5.02458     14.2847   18.5745   18.568
 1.80908  3.87368  3.42751  3.96385     14.3149   12.6661   14.6481
 1.30268  2.02808  5.76551  4.08371      7.49463  21.306    15.0911
 1.89387  2.33336  2.37031  4.88027      8.62275   8.75931  18.0347
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 132 stored entries:
  [1 ,  1]  =  0.171165
  [2 ,  1]  =  0.350571
  [3 ,  1]  =  0.0855815
  [4 ,  1]  =  0.467412
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [1 ,  2]  =  0.926718
  [2 ,  2]  =  0.0793394
  [3 ,  2]  =  0.264787
  ⋮
  [18, 19]  =  0.615859
  [19, 19]  =  1.7633
  [20, 19]  =  0.203186
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.68366
  [18, 20]  =  0.775006
  [19, 20]  =  0.803108
  [20, 20]  =  1.34855

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.