Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.28214   0.44215   0.761776  0.74277   …  0.0       0.0       0.0
 0.5448    1.80709   0.961627  0.479709     0.0       0.0       0.0
 0.336859  0.348409  1.87939   0.559582     0.0       0.0       0.0
 0.489526  0.139303  0.223871  1.17995      0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.44215   0.761776  0.74277
 0.0       0.0       0.0       0.0          0.807093  0.961627  0.479709
 0.0       0.0       0.0       0.0          0.348409  0.879385  0.559582
 0.0       0.0       0.0       0.0          0.139303  0.223871  0.179953

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 6.16636   3.78779   6.56704  5.0655   …  0.0      0.0       0.0      0.0
 5.0459    8.70577   9.05928  5.63134     0.0      0.0       0.0      0.0
 3.45151   3.4405    9.55781  4.55613     0.0      0.0       0.0      0.0
 2.55143   1.66724   2.84003  4.83829     0.0      0.0       0.0      0.0
 6.16636   3.78779   6.56704  5.0655      0.0      0.0       0.0      0.0
 5.0459    8.70577   9.05928  5.63134  …  0.0      0.0       0.0      0.0
 3.45151   3.4405    9.55781  4.55613     0.0      0.0       0.0      0.0
 2.55143   1.66724   2.84003  4.83829     0.0      0.0       0.0      0.0
 6.16636   3.78779   6.56704  5.0655      0.0      0.0       0.0      0.0
 5.0459    8.70577   9.05928  5.63134     0.0      0.0       0.0      0.0
 3.45151   3.4405    9.55781  4.55613  …  0.0      0.0       0.0      0.0
 2.55143   1.66724   2.84003  4.83829     0.0      0.0       0.0      0.0
 8.1658    5.01598   8.6964   6.70799     2.66592  1.63758   2.83914  2.18998
 6.68202  11.5286   11.9967   7.4573      2.1815   3.76379   3.91662  2.43461
 4.57066   4.55607  12.6569   6.03345     1.4922   1.48744   4.13215  1.96976
 3.37873   2.20784   3.76091  6.4071   …  1.10306  0.720801  1.22784  2.09175
 7.2501    4.45349   7.7212   5.95577     3.50044  2.1502    3.7279   2.87552
 5.93272  10.2358   10.6515   6.62106     2.86439  4.94198   5.14266  3.19673
 4.05812   4.04517  11.2376   5.35687     1.95931  1.95306   5.42566  2.58637
 2.99984   1.96026   3.33917  5.68862     1.44836  0.946437  1.6122   2.74654

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 116 stored entries:
  [1 ,  1]  =  1.28214
  [2 ,  1]  =  0.5448
  [3 ,  1]  =  0.336859
  [4 ,  1]  =  0.489526
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [1 ,  2]  =  0.44215
  [2 ,  2]  =  1.80709
  ⋮
  [17, 19]  =  0.761776
  [18, 19]  =  0.961627
  [19, 19]  =  0.879385
  [20, 19]  =  0.223871
  [16, 20]  =  1.0
  [17, 20]  =  0.74277
  [18, 20]  =  0.479709
  [19, 20]  =  0.559582
  [20, 20]  =  0.179953

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.