Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.52206   0.509573  0.50802    0.0371066  …  0.0       0.0        0.0
 0.911296  1.3193    0.0379055  0.394588      1.0       0.0        0.0
 0.666446  0.411673  1.97978    0.808195      0.0       1.0        0.0
 0.97169   0.167857  0.532868   1.18764       0.0       0.0        1.0
 1.0       0.0       0.0        0.0           0.0       0.0        0.0
 0.0       1.0       0.0        0.0        …  1.0       0.0        0.0
 0.0       0.0       1.0        0.0           0.0       1.0        0.0
 0.0       0.0       0.0        1.0           0.0       0.0        1.0
 1.0       0.0       0.0        0.0           0.0       0.0        0.0
 0.0       1.0       0.0        0.0           1.0       0.0        0.0
 0.0       0.0       1.0        0.0        …  0.0       1.0        0.0
 0.0       0.0       0.0        1.0           0.0       0.0        1.0
 0.0       0.0       0.0        0.0           0.0       0.0        0.0
 0.0       0.0       0.0        0.0           0.0       0.0        0.0
 0.0       0.0       0.0        0.0           0.0       0.0        0.0
 0.0       0.0       0.0        0.0        …  0.0       0.0        0.0
 1.0       0.0       0.0        0.0           0.509573  0.50802    0.0371066
 0.0       1.0       0.0        0.0           0.319305  0.0379055  0.394588
 0.0       0.0       1.0        0.0           0.411673  0.979784   0.808195
 0.0       0.0       0.0        1.0           0.167857  0.532868   0.187637

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 16.6469    7.20182   8.58364   4.0734    …  4.55313    5.42675   2.57529
 12.5528   11.9032    5.06952   4.86241      7.52543    3.20505   3.07411
 18.3369    9.74727  23.3541   12.0827       6.16242   14.7649    7.63895
 14.2694    5.99941  10.2896   11.2763       3.79295    6.50529   7.12912
 10.5245    4.55313   5.42675   2.57529      3.41035    4.06469   1.92892
  7.93616   7.52543   3.20505   3.07411   …  5.63663    2.40062   2.30254
 11.593     6.16242  14.7649    7.63895      4.61572   11.0591    5.72166
  9.02139   3.79295   6.50529   7.12912      2.84096    4.87253   5.33979
 10.5245    4.55313   5.42675   2.57529      3.41035    4.06469   1.92892
  7.93616   7.52543   3.20505   3.07411      5.63663    2.40062   2.30254
 11.593     6.16242  14.7649    7.63895   …  4.61572   11.0591    5.72166
  9.02139   3.79295   6.50529   7.12912      2.84096    4.87253   5.33979
  3.36133   1.45419   1.73321   0.822501     1.14279    1.36206   0.64637
  2.53467   2.40349   1.02364   0.981818     1.8888     0.804435  0.77157
  3.70259   1.96817   4.71565   2.43975      1.5467     3.70584   1.9173
  2.88127   1.2114    2.07768   2.27692   …  0.951989   1.63276   1.78933
 11.2443    4.86454   5.7979    2.75142      4.10288    4.8901    2.32062
  8.47893   8.04012   3.42426   3.28436      6.78124    2.88811   2.77012
 12.3858    6.58389  15.7747    8.1614       5.55302   13.3048    6.88354
  9.63839   4.05236   6.95021   7.6167       3.41786    5.86199   6.42413

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  1.52206
  [2 ,  1]  =  0.911296
  [3 ,  1]  =  0.666446
  [4 ,  1]  =  0.97169
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.509573
  [2 ,  2]  =  1.3193
  ⋮
  [19, 19]  =  0.979784
  [20, 19]  =  0.532868
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [17, 20]  =  0.0371066
  [18, 20]  =  0.394588
  [19, 20]  =  0.808195
  [20, 20]  =  0.187637

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.