Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.76429   0.0417006  0.867948   …  0.0        0.0        0.0
 0.356727  1.594      0.610827      0.0        0.0        0.0
 0.845176  0.487878   1.39915       0.0        0.0        0.0
 0.833863  0.99624    0.0568398     0.0        0.0        0.0
 0.0       0.0        0.0           0.0        0.0        0.0
 0.0       0.0        0.0        …  1.0        0.0        0.0
 0.0       0.0        0.0           0.0        1.0        0.0
 0.0       0.0        0.0           0.0        0.0        1.0
 1.0       0.0        0.0           0.0        0.0        0.0
 0.0       1.0        0.0           1.0        0.0        0.0
 0.0       0.0        1.0        …  0.0        1.0        0.0
 0.0       0.0        0.0           0.0        0.0        1.0
 1.0       0.0        0.0           0.0        0.0        0.0
 0.0       1.0        0.0           1.0        0.0        0.0
 0.0       0.0        1.0           0.0        1.0        0.0
 0.0       0.0        0.0        …  0.0        0.0        1.0
 0.0       0.0        0.0           0.0417006  0.867948   0.888342
 0.0       0.0        0.0           1.594      0.610827   0.0898998
 0.0       0.0        0.0           0.487878   1.39915    0.0917984
 0.0       0.0        0.0           0.99624    0.0568398  1.9612
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 16.2806    6.65828   9.38667  10.9784   …   4.13302   5.82663   6.81466
  6.29613   9.44604   6.26074   3.41731      5.86348   3.88626   2.12124
  9.23922   5.47222  10.1895    4.85636      3.3968    6.32497   3.01451
 12.5867   11.1649    7.07983  15.4525       6.93041   4.3947    9.59191
 16.3776    6.69795   9.44259  11.0438      11.6229   16.3857   19.1642
  6.33364   9.50232   6.29804   3.43766  …  16.4893   10.9289    5.96535
  9.29427   5.50482  10.2502    4.8853       9.55247  17.7871    8.47742
 12.6617   11.2314    7.12201  15.5446      19.4897   12.3588   26.9744
 28.4199   11.6229   16.3857   19.1642      15.7559   22.2123   25.9789
 10.9907   16.4893   10.9289    5.96535     22.3528   14.8152    8.08659
 16.1283    9.55247  17.7871    8.47742  …  12.9493   24.1121   11.4919
 21.9718   19.4897   12.3588   26.9744      26.4201   16.7535   36.5663
 22.1482    9.05797  12.7697   14.9351      12.4069   17.491    20.457
  8.5653   12.8505    8.51716   4.64893     17.6016   11.6662    6.36776
 12.5691    7.44445  13.8619    6.60663     10.1969   18.987     9.04928
 17.1231   15.1887    9.63146  21.0217   …  20.8044   13.1925   28.794
 10.1059    4.13302   5.82663   6.81466     10.0073   14.108    16.5003
  3.90822   5.86348   3.88626   2.12124     14.1972    9.40977   5.13614
  5.7351    3.3968    6.32497   3.01451      8.22463  15.3146    7.29901
  7.81302   6.93041   4.3947    9.59191     16.7805   10.6408   23.2248
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 132 stored entries:
  [1 ,  1]  =  1.76429
  [2 ,  1]  =  0.356727
  [3 ,  1]  =  0.845176
  [4 ,  1]  =  0.833863
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [1 ,  2]  =  0.0417006
  [2 ,  2]  =  1.594
  [3 ,  2]  =  0.487878
  ⋮
  [18, 19]  =  0.610827
  [19, 19]  =  1.39915
  [20, 19]  =  0.0568398
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.888342
  [18, 20]  =  0.0898998
  [19, 20]  =  0.0917984
  [20, 20]  =  1.9612

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.