Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.14967   0.846022  0.789996  0.984202   …  0.0       0.0       0.0
 0.643106  1.74918   0.223333  0.381665      1.0       0.0       0.0
 0.395254  0.783082  1.90693   0.0129285     0.0       1.0       0.0
 0.989789  0.035469  0.767173  1.0852        0.0       0.0       1.0
 1.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       1.0       0.0       0.0        …  1.0       0.0       0.0
 0.0       0.0       1.0       0.0           0.0       1.0       0.0
 0.0       0.0       0.0       1.0           0.0       0.0       1.0
 1.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       1.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       1.0       0.0        …  0.0       0.0       0.0
 0.0       0.0       0.0       1.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           1.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       1.0       0.0
 0.0       0.0       0.0       0.0        …  0.0       0.0       1.0
 0.0       0.0       0.0       0.0           0.846022  0.789996  0.984202
 0.0       0.0       0.0       0.0           0.749177  0.223333  0.381665
 0.0       0.0       0.0       0.0           0.783082  0.906932  0.0129285
 0.0       0.0       0.0       0.0           0.035469  0.767173  0.0851995
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 22.9427   21.8027   22.456    15.3914   …  17.5822   18.1091   12.412
 14.8541   25.7838   13.5129   10.716       20.7927   10.8972    8.64165
 12.5771   19.9991   27.1344    6.91377     16.1278   21.8819    5.57545
 15.8796   13.2076   19.2426   15.9234      10.6509   15.5177   12.8411
 21.6162   20.542    21.1576   14.5014      20.542    21.1576   14.5014
 13.9953   24.293    12.7316   10.0964   …  24.293    12.7316   10.0964
 11.8499   18.8427   25.5654    6.51402     18.8427   25.5654    6.51402
 14.9614   12.4439   18.13     15.0027      12.4439   18.13     15.0027
 15.8008   15.0156   15.4656   10.6001      13.1654   13.5599    9.29395
 10.2301   17.7574    9.30644   7.38016     15.5693    8.15967   6.47075
  8.66189  13.7735   18.6876    4.76155  …  12.0763   16.3848    4.17482
 10.9364    9.09612  13.2525   10.9665       7.97527  11.6195    9.61521
  7.16583   6.80974   7.0138    4.80727      9.7072    9.99809   6.8527
  4.63946   8.05319   4.22057   3.34698     11.4797    6.01637   4.77108
  3.92826   6.24641   8.47501   2.15941      8.90418  12.081     3.07822
  4.95975   4.12518   6.01014   4.97344  …   5.8804    8.56738   7.08958
  6.04482   5.74444   5.91658   4.05523      8.91766   9.18489   6.29533
  3.91367   6.79336   3.56031   2.82338     10.546     5.52702   4.38302
  3.31373   5.26923   7.1492    1.8216       8.17996  11.0984    2.82785
  4.18386   3.47985   5.06992   4.1954       5.40211   7.87054   6.51294
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 136 stored entries:
  [1 ,  1]  =  1.14967
  [2 ,  1]  =  0.643106
  [3 ,  1]  =  0.395254
  [4 ,  1]  =  0.989789
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [1 ,  2]  =  0.846022
  [2 ,  2]  =  1.74918
  [3 ,  2]  =  0.783082
  ⋮
  [18, 19]  =  0.223333
  [19, 19]  =  0.906932
  [20, 19]  =  0.767173
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.984202
  [18, 20]  =  0.381665
  [19, 20]  =  0.0129285
  [20, 20]  =  0.0851995

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.