Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.572874  0.559249  0.00779768  …  0.0       0.0         0.0
 0.305633  0.915452  0.901764       0.0       0.0         0.0
 0.477841  0.536617  0.563648       0.0       0.0         0.0
 0.564552  0.348919  0.324748       0.0       0.0         0.0
 1.0       0.0       0.0            0.0       0.0         0.0
 0.0       1.0       0.0         …  1.0       0.0         0.0
 0.0       0.0       1.0            0.0       1.0         0.0
 0.0       0.0       0.0            0.0       0.0         1.0
 1.0       0.0       0.0            0.0       0.0         0.0
 0.0       1.0       0.0            1.0       0.0         0.0
 0.0       0.0       1.0         …  0.0       1.0         0.0
 0.0       0.0       0.0            0.0       0.0         1.0
 0.0       0.0       0.0            0.0       0.0         0.0
 0.0       0.0       0.0            1.0       0.0         0.0
 0.0       0.0       0.0            0.0       1.0         0.0
 0.0       0.0       0.0         …  0.0       0.0         1.0
 0.0       0.0       0.0            0.559249  0.00779768  0.517944
 0.0       0.0       0.0            1.91545   0.901764    0.905422
 0.0       0.0       0.0            0.536617  1.56365     0.0157113
 0.0       0.0       0.0            0.348919  0.324748    1.04173

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 3.75001  2.58158  1.26628   2.0666    …   1.94937   0.956177   1.56051
 2.80136  6.03639  3.97909   3.31957       4.55812   3.00464    2.50663
 2.00047  2.60321  3.79787   1.23281       1.9657    2.8678     0.930906
 1.88929  1.90371  1.47072   2.57915       1.43751   1.11055    1.94754
 4.73099  3.25691  1.59753   2.60721       8.47281   4.15596    6.78264
 3.53418  7.61548  5.01999   4.18795   …  19.8116   13.0594    10.8949
 2.52378  3.28419  4.79137   1.55531       8.54379  12.4647     4.04612
 2.38352  2.40171  1.85545   3.25385       6.24803   4.82693    8.46484
 2.85598  1.96612  0.964391  1.57391       5.86486   2.87675    4.69493
 2.1335   4.59728  3.03045   2.52817      13.7135    9.03972    7.54143
 1.52354  1.98259  2.89243   0.938903  …   5.91399   8.62802    2.80071
 1.43887  1.44986  1.12009   1.96427       4.32487   3.34119    5.85934
 0.0      0.0      0.0       0.0           5.34446   2.62149    4.27834
 0.0      0.0      0.0       0.0          12.4967    8.23761    6.87227
 0.0      0.0      0.0       0.0           5.38923   7.86245    2.5522
 0.0      0.0      0.0       0.0       …   3.94112   3.04472    5.33944
 0.0      0.0      0.0       0.0           7.01747   3.44211    5.61761
 0.0      0.0      0.0       0.0          16.4086   10.8163     9.02353
 0.0      0.0      0.0       0.0           7.07625  10.3237     3.35113
 0.0      0.0      0.0       0.0           5.17483   3.99782    7.01087

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 116 stored entries:
  [1 ,  1]  =  0.572874
  [2 ,  1]  =  0.305633
  [3 ,  1]  =  0.477841
  [4 ,  1]  =  0.564552
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [1 ,  2]  =  0.559249
  [2 ,  2]  =  0.915452
  [3 ,  2]  =  0.536617
  ⋮
  [19, 19]  =  1.56365
  [20, 19]  =  0.324748
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.517944
  [18, 20]  =  0.905422
  [19, 20]  =  0.0157113
  [20, 20]  =  1.04173

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.