Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.75853   0.678104  0.677818  0.493941  …  0.0       0.0       0.0
 0.861629  1.54312   0.963897  0.505276     0.0       0.0       0.0
 0.989594  0.193727  1.3328    0.909029     0.0       0.0       0.0
 0.504015  0.572695  0.838269  1.2505       0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       1.0       0.0       …  0.0       1.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.678104  0.677818  0.493941
 0.0       0.0       0.0       0.0          0.543121  0.963897  0.505276
 0.0       0.0       0.0       0.0          0.193727  0.332797  0.909029
 0.0       0.0       0.0       0.0          0.572695  0.838269  0.250501
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 27.6038  14.844    18.8656  15.0359   …   7.07669  3.8055   4.83652  3.85471
 23.9644  20.8075   21.5761  16.4567       6.14369  5.33435  5.5314   4.21894
 20.8566  11.3642   21.4953  15.7453       5.34694  2.9134   5.51069  4.03658
 17.5128  12.0943   16.9858  17.8438       4.48969  3.10059  4.35458  4.57457
 17.3402   9.32474  11.8511   9.4453       5.60146  3.0122   3.82828  3.05114
 15.0541  13.0709   13.5538  10.3378   …   4.86295  4.22234  4.3783   3.33945
 13.1018   7.13879  13.503    9.89096      4.2323   2.30606  4.36192  3.1951
 11.0012   7.59747  10.6702  11.2092       3.55375  2.45423  3.44681  3.62094
 17.3402   9.32474  11.8511   9.4453      10.2635   5.51924  7.01455  5.5906
 15.0541  13.0709   13.5538  10.3378       8.91037  7.73657  8.02235  6.11886
 13.1018   7.13879  13.503    9.89096  …   7.75482  4.22539  7.99232  5.85438
 11.0012   7.59747  10.6702  11.2092       6.51153  4.49688  6.31558  6.63463
 24.8057  13.3393   16.9533  13.5118       4.11678  2.21381  2.81359  2.24243
 21.5353  18.6983   19.389   14.7885       3.57401  3.1032   3.21782  2.45432
 18.7424  10.2122   19.3164  14.1493       3.11052  1.69484  3.20578  2.34823
 15.7376  10.8684   15.264   16.0351   …   2.61182  1.80373  2.53323  2.6612
  0.0      0.0       0.0      0.0          4.66208  2.50704  3.18627  2.53945
  0.0      0.0       0.0      0.0          4.04742  3.51424  3.64405  2.77941
  0.0      0.0       0.0      0.0          3.52253  1.91933  3.63041  2.65927
  0.0      0.0       0.0      0.0          2.95778  2.04265  2.86877  3.01369
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 112 stored entries:
  [1 ,  1]  =  1.75853
  [2 ,  1]  =  0.861629
  [3 ,  1]  =  0.989594
  [4 ,  1]  =  0.504015
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [1 ,  2]  =  0.678104
  [2 ,  2]  =  1.54312
  ⋮
  [11, 19]  =  1.0
  [17, 19]  =  0.677818
  [18, 19]  =  0.963897
  [19, 19]  =  0.332797
  [20, 19]  =  0.838269
  [12, 20]  =  1.0
  [17, 20]  =  0.493941
  [18, 20]  =  0.505276
  [19, 20]  =  0.909029
  [20, 20]  =  0.250501

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.