Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.14166   0.209189  0.142302  0.0992423  …  0.0       0.0       0.0
 0.519466  1.97654   0.521728  0.901644      0.0       0.0       0.0
 0.953527  0.121552  1.65309   0.634943      0.0       0.0       0.0
 0.585588  0.468667  0.627448  1.30311       0.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0        …  1.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       1.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       1.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           1.0       0.0       0.0
 0.0       0.0       0.0       0.0        …  0.0       1.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       1.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           1.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       1.0       0.0
 0.0       0.0       0.0       0.0        …  0.0       0.0       1.0
 1.0       0.0       0.0       0.0           0.209189  0.142302  0.0992423
 0.0       1.0       0.0       0.0           0.976537  0.521728  0.901644
 0.0       0.0       1.0       0.0           0.121552  0.653086  0.634943
 0.0       0.0       0.0       1.0           0.468667  0.627448  0.303106

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 4.81885   1.6555    1.46713  1.38018   …   0.835369   0.74032  0.696442
 7.40488  11.403     6.86263  8.13379       5.75399    3.4629   4.10434
 6.80528   2.85989   8.55183  4.87101       1.44311    4.31528  2.45793
 5.53914   4.2624    5.17161  7.36196       2.15082    2.60961  3.71486
 2.4316    0.835369  0.74032  0.696442      1.9772     1.75224  1.64838
 3.73652   5.75399   3.4629   4.10434   …  13.6189     8.19621  9.71439
 3.43396   1.44311   4.31528  2.45793       3.41564   10.2137   5.81757
 2.79506   2.15082   2.60961  3.71486       5.09069    6.17658  8.79257
 2.4316    0.835369  0.74032  0.696442      1.9772     1.75224  1.64838
 3.73652   5.75399   3.4629   4.10434      13.6189     8.19621  9.71439
 3.43396   1.44311   4.31528  2.45793   …   3.41564   10.2137   5.81757
 2.79506   2.15082   2.60961  3.71486       5.09069    6.17658  8.79257
 2.4316    0.835369  0.74032  0.696442      1.9772     1.75224  1.64838
 3.73652   5.75399   3.4629   4.10434      13.6189     8.19621  9.71439
 3.43396   1.44311   4.31528  2.45793       3.41564   10.2137   5.81757
 2.79506   2.15082   2.60961  3.71486   …   5.09069    6.17658  8.79257
 3.32367   1.14183   1.01191  0.95194       1.34903    1.19554  1.12468
 5.1073    7.86491   4.7333   5.61006       9.2921     5.59223  6.62807
 4.69375   1.97253   5.89838  3.35964       2.33047    6.96872  3.96929
 3.82046   2.93987   3.56697  5.0777        3.47335    4.21425  5.99912

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  1.14166
  [2 ,  1]  =  0.519466
  [3 ,  1]  =  0.953527
  [4 ,  1]  =  0.585588
  [17,  1]  =  1.0
  [1 ,  2]  =  0.209189
  [2 ,  2]  =  1.97654
  [3 ,  2]  =  0.121552
  [4 ,  2]  =  0.468667
  ⋮
  [19, 19]  =  0.653086
  [20, 19]  =  0.627448
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.0992423
  [18, 20]  =  0.901644
  [19, 20]  =  0.634943
  [20, 20]  =  0.303106

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.