Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.575157  0.54387   0.790935  0.982167  …  0.0       0.0       0.0
 0.332011  0.792098  0.252533  0.952376     0.0       0.0       0.0
 0.819024  0.767923  0.676057  0.754732     0.0       0.0       0.0
 0.319723  0.993973  0.908454  0.880428     0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.54387   0.790935  0.982167
 0.0       1.0       0.0       0.0          1.7921    0.252533  0.952376
 0.0       0.0       1.0       0.0          0.767923  1.67606   0.754732
 0.0       0.0       0.0       1.0          0.993973  0.908454  1.88043
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  9.45354  10.5956   9.54968  12.5767  …   8.14317   7.33934   9.66573
  4.95706  11.7257   6.44858  10.3933      9.0117    4.95602   7.98768
  7.76944  11.2799  11.7658   12.3898      8.6691    9.04255   9.52209
  6.63757  12.1929   9.90521  15.2424      9.37079   7.61259  11.7145
 23.5181   26.3592  23.7573   31.2877     29.6944   26.7633   35.2465
 12.332    29.1707  16.0425   25.8559  …  32.8616   18.0724   29.1274
 19.3285   28.0617  29.2705   30.8228     31.6123   32.9741   34.7227
 16.5127   30.333   24.6418   37.9194     34.171    27.7597   42.7173
 11.6144   13.0175  11.7325   15.4514      8.5338    7.69141  10.1294
  6.09013  14.4059   7.92257  12.7689      9.444     5.19376   8.37085
  9.54534  13.8582  14.4552   15.2218  …   9.08496   9.47632   9.97886
  8.15475  14.9799  12.1693   18.7264      9.82031   7.97777  12.2764
 23.0504   25.835   23.2848   30.6654     23.481    21.1632   27.8713
 12.0867   28.5905  15.7234   25.3417     25.9854   14.2908   23.0327
 18.944    27.5036  28.6883   30.2097     24.9976   26.0744   27.4571
 16.1842   29.7297  24.1517   37.1652  …  27.0209   21.9511   33.7789
 28.5641   32.0148  28.8546   38.0007     36.0399   32.4824   42.7784
 14.9779   35.4294  19.4845   31.4035     39.8839   21.9343   35.3518
 23.4755   34.0825  35.5507   37.436      38.3676   40.0204   42.1427
 20.0556   36.8412  29.9288   46.0553     41.4731   33.6917   51.8457
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 132 stored entries:
  [1 ,  1]  =  0.575157
  [2 ,  1]  =  0.332011
  [3 ,  1]  =  0.819024
  [4 ,  1]  =  0.319723
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.54387
  [2 ,  2]  =  0.792098
  ⋮
  [17, 19]  =  0.790935
  [18, 19]  =  0.252533
  [19, 19]  =  1.67606
  [20, 19]  =  0.908454
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.982167
  [18, 20]  =  0.952376
  [19, 20]  =  0.754732
  [20, 20]  =  1.88043

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.