Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.23121    0.966608  0.219123  0.239673  …  0.0       0.0       0.0
 0.0785746  1.40839   0.11199   0.567271     0.0       0.0       0.0
 0.649846   0.457876  1.61005   0.474761     0.0       0.0       0.0
 0.0094873  0.497343  0.872956  1.4877       0.0       0.0       0.0
 1.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        1.0       0.0       0.0       …  0.0       0.0       0.0
 0.0        0.0       1.0       0.0          0.0       0.0       0.0
 0.0        0.0       0.0       1.0          0.0       0.0       0.0
 1.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        1.0       0.0       0.0          1.0       0.0       0.0
 0.0        0.0       1.0       0.0       …  0.0       1.0       0.0
 0.0        0.0       0.0       1.0          0.0       0.0       1.0
 0.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0        0.0       0.0       0.0          0.966608  0.219123  0.239673
 0.0        0.0       0.0       0.0          1.40839   0.11199   0.567271
 0.0        0.0       0.0       0.0          0.457876  1.61005   0.474761
 0.0        0.0       0.0       0.0          0.497343  0.872956  1.4877
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 5.6611     6.47524   2.96316   3.6405   …   1.0744     0.491662   0.60405
 1.03298    7.37182   2.58697   4.17528      1.22317    0.429242   0.692782
 4.44659    6.26467   9.65718   5.38583      1.03946    1.60237    0.893644
 2.06985    5.37516   6.88223   8.98959      0.891874   1.14193    1.4916
 5.15062    5.89134   2.69596   3.31223      2.4286     1.11136    1.36541
 0.939832   6.70708   2.35369   3.79878  …   2.76487    0.970266   1.56598
 4.04562    5.69976   8.78636   4.90018      2.34962    3.62202    2.02001
 1.88321    4.89047   6.26164   8.17897      2.01601    2.58125    3.37163
 7.97508    9.12199   4.17435   5.12856      7.32884    3.35378    4.12042
 1.45521   10.3851    3.64439   5.88192      8.34361    2.92799    4.72569
 6.26413    8.82536  13.6046    7.5873   …   7.09051   10.9303     6.09583
 2.9159     7.57226   9.69535  12.6641       6.08375    7.78949   10.1747
 5.19232    5.93904   2.71779   3.33904      3.68827    1.6878     2.07362
 0.947441   6.76138   2.37275   3.82953      4.19896    1.47353    2.37822
 4.07838    5.74591   8.8575    4.93985      3.56833    5.50069    3.06775
 1.89845    4.93006   6.31233   8.24519  …   3.06167    3.92009    5.12044
 6.76002    7.73219   3.53836   4.34719      9.44476    4.32205    5.31002
 1.2335     8.80282   3.08914   4.98577     10.7525     3.77334    6.09004
 5.30975    7.48075  11.5318    6.43132      9.13762   14.0859     7.85576
 2.47165    6.41858   8.21819  10.7346       7.8402    10.0384    13.1122
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 124 stored entries:
  [1 ,  1]  =  1.23121
  [2 ,  1]  =  0.0785746
  [3 ,  1]  =  0.649846
  [4 ,  1]  =  0.0094873
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [1 ,  2]  =  0.966608
  [2 ,  2]  =  1.40839
  [3 ,  2]  =  0.457876
  ⋮
  [11, 19]  =  1.0
  [17, 19]  =  0.219123
  [18, 19]  =  0.11199
  [19, 19]  =  1.61005
  [20, 19]  =  0.872956
  [12, 20]  =  1.0
  [17, 20]  =  0.239673
  [18, 20]  =  0.567271
  [19, 20]  =  0.474761
  [20, 20]  =  1.4877

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.