Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.39336   0.605707  0.589447  0.193267  …  0.0       0.0       0.0
 0.733358  1.75829   0.31608   0.645106     1.0       0.0       0.0
 0.174081  0.197038  1.20887   0.397362     0.0       1.0       0.0
 0.636504  0.047669  0.818756  1.90577      0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.605707  0.589447  0.193267
 0.0       1.0       0.0       0.0          1.75829   0.31608   0.645106
 0.0       0.0       1.0       0.0          0.197038  1.20887   0.397362
 0.0       0.0       0.0       1.0          0.047669  0.818756  1.90577

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 11.2865    6.90148   7.03421   5.6941   …   5.39664    5.50042   4.45252
 10.3563   14.2814    8.31343  10.431       11.1674     6.50072   8.15655
  3.41345   2.74729   8.76614   4.99021      2.14825    6.85472   3.90211
  8.40908   3.84205  10.3271   15.9715       3.00431    8.07529  12.489
 12.0029    7.33953   7.48068   6.05551      6.44604    6.57      5.31833
 11.0136   15.1879    8.8411   11.0931   …  13.339      7.76481   9.74263
  3.63011   2.92166   9.32254   5.30694      2.56599    8.18764   4.66089
  8.94282   4.08591  10.9825   16.9853       3.58851    9.64556  14.9175
  4.69738   2.87236   2.9276    2.36985      2.52756    2.57616   2.08537
  4.31021   5.94385   3.46      4.34132      5.23035    3.04466   3.82018
  1.42066   1.1434    3.64842   2.0769   …   1.00615    3.21046   1.82758
  3.49981   1.59904   4.29807   6.64726      1.40709    3.78212   5.84931
  5.08429   3.10895   3.16873   2.56505      1.70303    1.73578   1.40509
  4.66523   6.43343   3.74499   4.6989       3.52412    2.05144   2.57398
  1.53767   1.23758   3.94893   2.24796      0.677926   2.16315   1.23139
  3.78808   1.73075   4.65209   7.19477  …   0.948074   2.54833   3.94117
 13.719     8.38893   8.55026   6.92132      9.30911    9.48813   7.68052
 12.5883   17.3595   10.1052   12.6791      19.2636    11.2136   14.0699
  4.14914   3.3394   10.6555    6.06572      3.70569   11.8243    6.73107
 10.2215    4.67011  12.5528   19.4138       5.18238   13.9297   21.5433

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  1.39336
  [2 ,  1]  =  0.733358
  [3 ,  1]  =  0.174081
  [4 ,  1]  =  0.636504
  [5 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.605707
  [2 ,  2]  =  1.75829
  ⋮
  [18, 19]  =  0.31608
  [19, 19]  =  1.20887
  [20, 19]  =  0.818756
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [17, 20]  =  0.193267
  [18, 20]  =  0.645106
  [19, 20]  =  0.397362
  [20, 20]  =  1.90577

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.