Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.24594   0.19427     0.879178  0.795239  …  0.0         0.0       0.0
 0.924853  1.61901     0.151403  0.772802     1.0         0.0       0.0
 0.547456  0.572063    1.30609   0.210873     0.0         1.0       0.0
 0.593932  0.00809558  0.456115  1.77941      0.0         0.0       1.0
 0.0       0.0         0.0       0.0          0.0         0.0       0.0
 0.0       0.0         0.0       0.0       …  1.0         0.0       0.0
 0.0       0.0         0.0       0.0          0.0         1.0       0.0
 0.0       0.0         0.0       0.0          0.0         0.0       1.0
 1.0       0.0         0.0       0.0          0.0         0.0       0.0
 0.0       1.0         0.0       0.0          0.0         0.0       0.0
 0.0       0.0         1.0       0.0       …  0.0         0.0       0.0
 0.0       0.0         0.0       1.0          0.0         0.0       0.0
 0.0       0.0         0.0       0.0          0.0         0.0       0.0
 0.0       0.0         0.0       0.0          1.0         0.0       0.0
 0.0       0.0         0.0       0.0          0.0         1.0       0.0
 0.0       0.0         0.0       0.0       …  0.0         0.0       1.0
 1.0       0.0         0.0       0.0          0.19427     0.879178  0.795239
 0.0       1.0         0.0       0.0          1.61901     0.151403  0.772802
 0.0       0.0         1.0       0.0          0.572063    1.30609   0.210873
 0.0       0.0         0.0       1.0          0.00809558  0.456115  1.77941
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 14.7161    5.224    11.2899   12.6737   …   5.16159   11.155    12.5223
 14.3589   14.1892    9.22849  15.9104      14.0197     9.11825  15.7203
  9.45687   7.1173   13.0545    8.55664      7.03227   12.8985    8.45442
  9.35028   2.9614    8.70036  17.9418       2.92602    8.59642  17.7274
  8.35093   2.96445   6.40663   7.19193      4.0406     8.73236   9.80273
  8.14822   8.05191   5.23687   9.02864  …  10.9749     7.13795  12.3062
  5.36646   4.03883   7.40798   4.85561      5.50501   10.0972    6.6183
  5.30598   1.6805    4.93717  10.1814       2.29055    6.72946  13.8774
  6.01176   2.13408   4.61208   5.17741      1.48222    3.20332   3.59596
  5.86583   5.7965    3.76998   6.49964      4.02595    2.61844   4.51432
  3.86327   2.90752   5.33294   3.49552  …   2.01942    3.70399   2.42781
  3.81973   1.20978   3.55423   7.32947      0.840249   2.46859   5.09068
  6.67099   2.3681    5.11782   5.74514      3.09167    6.68157   7.50057
  6.50906   6.43212   4.18338   7.21237      8.39745    5.46161   9.4161
  4.2869    3.22635   5.91773   3.87882      4.21216    7.72589   5.06399
  4.23859   1.34244   3.94397   8.1332   …   1.75262    5.14905  10.6183
 16.3979    5.821    12.5801   14.1221       7.5921    16.4077   18.4189
 15.9999   15.8107   10.2831   17.7287      20.6213    13.4119   23.1228
 10.5376    7.93067  14.5463    9.5345      10.3436    18.9722   12.4355
 10.4188    3.29983   9.69464  19.9922       4.30384   12.6443   26.075
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 128 stored entries:
  [1 ,  1]  =  1.24594
  [2 ,  1]  =  0.924853
  [3 ,  1]  =  0.547456
  [4 ,  1]  =  0.593932
  [9 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.19427
  [2 ,  2]  =  1.61901
  [3 ,  2]  =  0.572063
  ⋮
  [18, 19]  =  0.151403
  [19, 19]  =  1.30609
  [20, 19]  =  0.456115
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.795239
  [18, 20]  =  0.772802
  [19, 20]  =  0.210873
  [20, 20]  =  1.77941

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.