Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.72016   0.37461   0.251125    0.828631  …  0.0       0.0         0.0
 0.406277  0.356279  0.00653784  0.681121     1.0       0.0         0.0
 0.363849  0.708981  0.475599    0.439088     0.0       1.0         0.0
 0.621386  0.164273  0.864861    0.994368     0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       1.0       0.0         0.0       …  1.0       0.0         0.0
 0.0       0.0       1.0         0.0          0.0       1.0         0.0
 0.0       0.0       0.0         1.0          0.0       0.0         1.0
 0.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       0.0       0.0         0.0          1.0       0.0         0.0
 0.0       0.0       0.0         0.0       …  0.0       1.0         0.0
 0.0       0.0       0.0         0.0          0.0       0.0         1.0
 0.0       0.0       0.0         0.0          0.0       0.0         0.0
 0.0       0.0       0.0         0.0          1.0       0.0         0.0
 0.0       0.0       0.0         0.0          0.0       1.0         0.0
 0.0       0.0       0.0         0.0       …  0.0       0.0         1.0
 1.0       0.0       0.0         0.0          0.37461   0.251125    0.828631
 0.0       1.0       0.0         0.0          0.356279  0.00653784  0.681121
 0.0       0.0       1.0         0.0          0.708981  0.475599    0.439088
 0.0       0.0       0.0         1.0          0.164273  0.864861    0.994368
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 12.1841   4.91624  5.70302  11.0692   …   7.73386   8.97156  17.4132
  5.36698  7.3315   3.28381   8.03817     11.5334    5.16584  12.645
  5.89124  5.90443  9.15804   8.07854      9.2884   14.4067   12.7085
  9.00362  5.31944  9.34131  16.8279       8.36814  14.695    26.4724
  9.36693  3.77953  4.38439   8.50984      6.262     7.26415  14.0993
  4.12606  5.63635  2.52454   6.17963  …   9.33841   4.18271  10.2385
  4.52909  4.53924  7.04056   6.21066      7.52069  11.6649   10.2899
  6.92185  4.0895   7.18146  12.9371       6.77557  11.8984   21.4344
  4.74375  1.91409  2.22041   4.30969      4.50595   5.22706  10.1454
  2.08958  2.85445  1.27852   3.12958      6.71964   3.00975   7.36733
  2.2937   2.29883  3.56559   3.1453   …   5.41167   8.39374   7.40433
  3.50547  2.07107  3.63695   6.55179      4.8755    8.56171  15.4235
  7.80399  3.14889  3.65283   7.08991      6.34102   7.35581  14.2772
  3.4376   4.69589  2.1033    5.14851      9.45625   4.23549  10.3677
  3.77338  3.78183  5.8658    5.17437      7.61559  11.8121   10.4198
  5.76689  3.40714  5.98318  10.7784   …   6.86107  12.0485   21.7048
  8.61546  3.47631  4.03265   7.82713      6.75131   7.83177  15.201
  3.79504  5.18417  2.32201   5.68386     10.0681    4.50954  11.0386
  4.16574  4.17507  6.47572   5.7124       8.10836  12.5764   11.094
  6.36653  3.76142  6.60532  11.8992       7.30501  12.8281   23.1092
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 136 stored entries:
  [1 ,  1]  =  0.72016
  [2 ,  1]  =  0.406277
  [3 ,  1]  =  0.363849
  [4 ,  1]  =  0.621386
  [5 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.37461
  [2 ,  2]  =  0.356279
  [3 ,  2]  =  0.708981
  ⋮
  [19, 19]  =  0.475599
  [20, 19]  =  0.864861
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.828631
  [18, 20]  =  0.681121
  [19, 20]  =  0.439088
  [20, 20]  =  0.994368

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.