Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.188092  0.919837  0.937726   0.60346   …  0.0       0.0        0.0
 0.330664  0.788291  0.0335427  0.449955     1.0       0.0        0.0
 0.204648  0.309638  0.494606   0.864804     0.0       1.0        0.0
 0.97381   0.295614  0.896921   0.745486     0.0       0.0        1.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       1.0        0.0       …  0.0       1.0        0.0
 0.0       0.0       0.0        1.0          0.0       0.0        1.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       1.0        0.0          0.0       1.0        0.0
 0.0       0.0       0.0        1.0       …  0.0       0.0        1.0
 1.0       0.0       0.0        0.0          0.919837  0.937726   0.60346
 0.0       1.0       0.0        0.0          1.78829   0.0335427  0.449955
 0.0       0.0       1.0        0.0          0.309638  1.49461    0.864804
 0.0       0.0       0.0        1.0          0.295614  0.896921   1.74549
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.73745   5.72634   5.79626   5.99534  …   6.54922   6.62918   6.85688
  2.7088    6.55468   2.34324   3.67025      7.4966    2.67996   4.19767
  3.14516   3.38491   6.6623    5.62444      3.87132   7.61968   6.43267
  5.76187   5.10084   7.16436   9.51695      5.83384   8.19389  10.8845
  6.77963   6.7665    6.84911   7.08436      8.49962   8.6034    8.8989
  3.20083   7.7453    2.76887   4.33693  …   9.72914   3.47808   5.44777
  3.71646   3.99975   7.87247   6.64608      5.02423   9.88888   8.34837
  6.80847   6.02738   8.46572  11.2456       7.5712   10.6341   14.126
 11.8735   11.8505   11.9951   12.4071      15.4186   15.6068   16.1429
  5.60576  13.5647    4.84925   7.59546     17.6489    6.30933   9.88241
  6.5088    7.00494  13.7874   11.6396   …   9.11409  17.9387   15.1442
 11.924    10.556    14.8264   19.695       13.7344   19.2905   25.625
 10.137    10.1173   10.2409   10.5926      13.3157   13.4783   13.9412
  4.78592  11.5809    4.14005   6.48462     15.2419    5.44884   8.5346
  5.55689   5.98047  11.771     9.93729      7.87108  15.4922   13.0788
 10.1801    9.0122   12.658    16.8146   …  11.8612   16.6596   22.1302
  9.76653   9.74761   9.86662  10.2055      15.4739   15.6629   16.2008
  4.61103  11.1576    3.98875   6.24765     17.7123    6.33199   9.9179
  5.35382   5.76193  11.3408    9.57414      9.14683  18.0031   15.1986
  9.80808   8.68287  12.1955   16.2001      13.7837   19.3598   25.7171
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 124 stored entries:
  [1 ,  1]  =  0.188092
  [2 ,  1]  =  0.330664
  [3 ,  1]  =  0.204648
  [4 ,  1]  =  0.97381
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.919837
  [2 ,  2]  =  0.788291
  ⋮
  [18, 19]  =  0.0335427
  [19, 19]  =  1.49461
  [20, 19]  =  0.896921
  [4 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.60346
  [18, 20]  =  0.449955
  [19, 20]  =  0.864804
  [20, 20]  =  1.74549

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.