Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.484545  0.300917  0.161533  0.401983  …  0.0       0.0       0.0
 0.316467  0.422895  0.679462  0.153759     0.0       0.0       0.0
 0.68005   0.469295  0.288124  0.814981     0.0       0.0       0.0
 0.866179  0.145478  0.427113  0.306031     0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.300917  0.161533  0.401983
 0.0       1.0       0.0       0.0          0.422895  0.679462  0.153759
 0.0       0.0       1.0       0.0          0.469295  0.288124  0.814981
 0.0       0.0       0.0       1.0          0.145478  0.427113  0.306031

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.9876    3.43783   3.05451   4.44458  …  1.49789   1.33088   1.93654
  6.021     9.78217   6.23806   4.30896     4.26217   2.71798   1.87745
  9.67179   5.54484  10.2016    8.18204     2.41593   4.44491   3.56498
  9.05724   3.32147   4.68693  10.0081      1.44719   2.04213   4.36059
 14.2387    4.45505   3.95832   5.75969     2.47358   2.19778   3.19796
  7.80256  12.6766    8.08385   5.58394  …  7.03846   4.4884    3.10038
 12.5336    7.18551  13.2201   10.603       3.98962   7.34023   5.88714
 11.7372    4.30426   6.07375  12.9693      2.38986   3.37234   7.20099
 13.3358    4.17254   3.7073    5.39445     2.08649   1.85385   2.69751
  7.30777  11.8727    7.57122   5.22984     5.93701   3.78602   2.6152
 11.7388    6.72985  12.3818    9.93065  …  3.36529   6.19156   4.96586
 10.9929    4.03131   5.68859  12.1469      2.01587   2.8446    6.07411
 19.3618    6.05796   5.38251   7.83201     3.46431   3.07805   4.47883
 10.6099   17.2376   10.9924    7.59303     9.85754   6.28613   4.34216
 17.0431    9.77084  17.9767   14.418       5.58756  10.2802    8.24509
 15.9602    5.85293   8.25907  17.6357   …  3.34706   4.72305  10.0852
  7.90577   2.47358   2.19778   3.19796     1.5524    1.37931   2.00701
  4.33222   7.03846   4.4884    3.10038     4.41728   2.81689   1.94577
  6.95904   3.98962   7.34023   5.88714     2.50385   4.60667   3.69472
  6.51685   2.38986   3.37234   7.20099     1.49986   2.11645   4.51928

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 136 stored entries:
  [1 ,  1]  =  0.484545
  [2 ,  1]  =  0.316467
  [3 ,  1]  =  0.68005
  [4 ,  1]  =  0.866179
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.300917
  ⋮
  [18, 19]  =  0.679462
  [19, 19]  =  0.288124
  [20, 19]  =  0.427113
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.401983
  [18, 20]  =  0.153759
  [19, 20]  =  0.814981
  [20, 20]  =  0.306031

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.