Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.815087  0.253549  0.669621  0.16277   …  0.0       0.0       0.0
 0.219681  0.670235  0.295331  0.496841     1.0       0.0       0.0
 0.590489  0.944356  0.323667  0.914705     0.0       1.0       0.0
 0.91023   0.454978  0.644261  0.292574     0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       1.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       1.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.253549  0.669621  0.16277
 0.0       0.0       0.0       0.0          0.670235  0.295331  0.496841
 0.0       0.0       0.0       0.0          0.944356  0.323667  0.914705
 0.0       0.0       0.0       0.0          0.454978  0.644261  0.292574
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 6.43099   3.01082   3.50049    2.37763    …   9.24225  10.7454    7.29856
 2.69771   5.48912   2.36344    2.71359       16.8498    7.25499   8.32985
 5.03078   5.31902   5.64266    4.49583       16.3277   17.3212   13.8008
 5.15996   3.86092   3.86698    4.93775       11.8518   11.8704   15.1573
 3.15787   1.47843   1.71888    1.16751        9.24225  10.7454    7.29856
 1.32468   2.69537   1.16054    1.33248    …  16.8498    7.25499   8.32985
 2.47031   2.61185   2.77076    2.20762       16.3277   17.3212   13.8008
 2.53374   1.89586   1.89884    2.42463       11.8518   11.8704   15.1573
 0.24028   0.112493  0.130788   0.0888348      3.92551   4.56395   3.09996
 0.100794  0.205089  0.0883046  0.101387       7.15672   3.08145   3.53798
 0.187964  0.198734  0.210826   0.167977   …   6.93495   7.35691   5.86167
 0.192791  0.144255  0.144481   0.184488       5.03387   5.04178   6.43785
 5.22688   2.44708   2.84507    1.93245        9.24225  10.7454    7.29856
 2.1926    4.46135   1.92091    2.2055        16.8498    7.25499   8.32985
 4.08884   4.32311   4.58615    3.65404       16.3277   17.3212   13.8008
 4.19383   3.13801   3.14294    4.01323    …  11.8518   11.8704   15.1573
 0.899735  0.421232  0.48974    0.332644       5.55669   6.46042   4.38809
 0.377426  0.767961  0.330659   0.379647      10.1306    4.3619    5.00813
 0.703837  0.744164  0.789442   0.628994       9.81666  10.414     8.29739
 0.72191   0.540166  0.541014   0.690822       7.12562   7.13681   9.113
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 136 stored entries:
  [1 ,  1]  =  0.815087
  [2 ,  1]  =  0.219681
  [3 ,  1]  =  0.590489
  [4 ,  1]  =  0.91023
  [13,  1]  =  1.0
  [1 ,  2]  =  0.253549
  [2 ,  2]  =  0.670235
  [3 ,  2]  =  0.944356
  [4 ,  2]  =  0.454978
  ⋮
  [19, 19]  =  0.323667
  [20, 19]  =  0.644261
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.16277
  [18, 20]  =  0.496841
  [19, 20]  =  0.914705
  [20, 20]  =  0.292574

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.