Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.252657  0.376185  0.801748  0.73047   …  0.0       0.0       0.0
 0.194375  0.497266  0.889912  0.616075     1.0       0.0       0.0
 0.506287  0.988306  0.670215  0.526077     0.0       1.0       0.0
 0.315261  0.284923  0.725776  0.941292     0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  1.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          1.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       1.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       1.0
 0.0       0.0       0.0       0.0          0.376185  0.801748  0.73047
 0.0       0.0       0.0       0.0          0.497266  0.889912  0.616075
 0.0       0.0       0.0       0.0          0.988306  0.670215  0.526077
 0.0       0.0       0.0       0.0          0.284923  0.725776  0.941292

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 2.23177  2.02331  2.94005  2.7252   …   4.02948   5.85518   5.42731
 1.19434  3.21634  3.07877  2.62018      6.40541   6.13144   5.21815
 1.64348  2.97018  4.34242  2.92505      5.91518   8.64805   5.82531
 1.33338  1.95127  2.95111  4.0649       3.88601   5.87721   8.09535
 0.0      0.0      0.0      0.0          8.65428  12.5754   11.6565
 0.0      0.0      0.0      0.0      …  13.7572   13.1688   11.2072
 0.0      0.0      0.0      0.0         12.7043   18.5738   12.5113
 0.0      0.0      0.0      0.0          8.34614  12.6227   17.3867
 0.0      0.0      0.0      0.0          4.6248    6.72023   6.22916
 0.0      0.0      0.0      0.0          7.35176   7.03731   5.98909
 0.0      0.0      0.0      0.0      …   6.7891    9.92572   6.68596
 0.0      0.0      0.0      0.0          4.46013   6.74552   9.29137
 0.0      0.0      0.0      0.0          6.00318   8.72314   8.0857
 0.0      0.0      0.0      0.0          9.54289   9.13472   7.77409
 0.0      0.0      0.0      0.0          8.81253  12.884     8.67865
 0.0      0.0      0.0      0.0      …   5.78944   8.75596  12.0606
 0.0      0.0      0.0      0.0          5.20746   7.56689   7.01394
 0.0      0.0      0.0      0.0          8.27798   7.92392   6.74364
 0.0      0.0      0.0      0.0          7.64443  11.1762    7.52829
 0.0      0.0      0.0      0.0          5.02205   7.59536  10.4619

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 124 stored entries:
  [1 ,  1]  =  0.252657
  [2 ,  1]  =  0.194375
  [3 ,  1]  =  0.506287
  [4 ,  1]  =  0.315261
  [1 ,  2]  =  0.376185
  [2 ,  2]  =  0.497266
  [3 ,  2]  =  0.988306
  [4 ,  2]  =  0.284923
  [1 ,  3]  =  0.801748
  ⋮
  [19, 19]  =  0.670215
  [20, 19]  =  0.725776
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.73047
  [18, 20]  =  0.616075
  [19, 20]  =  0.526077
  [20, 20]  =  0.941292

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.