Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.70767    0.579378   0.858349  0.317026  …  0.0        0.0       0.0
 0.722222   1.98693    0.4561    0.895403     0.0        0.0       0.0
 0.15386    0.0294572  1.6387    0.035222     0.0        0.0       0.0
 0.0495901  0.0330001  0.583192  1.75593      0.0        0.0       0.0
 1.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        1.0        0.0       0.0       …  0.0        0.0       0.0
 0.0        0.0        1.0       0.0          0.0        0.0       0.0
 0.0        0.0        0.0       1.0          0.0        0.0       0.0
 1.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        1.0        0.0       0.0          0.0        0.0       0.0
 0.0        0.0        1.0       0.0       …  0.0        0.0       0.0
 0.0        0.0        0.0       1.0          0.0        0.0       0.0
 1.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        1.0        0.0       0.0          0.0        0.0       0.0
 0.0        0.0        1.0       0.0          0.0        0.0       0.0
 0.0        0.0        0.0       1.0       …  0.0        0.0       0.0
 0.0        0.0        0.0       0.0          0.579378   0.858349  0.317026
 0.0        0.0        0.0       0.0          0.986927   0.4561    0.895403
 0.0        0.0        0.0       0.0          0.0294572  0.638699  0.035222
 0.0        0.0        0.0       0.0          0.0330001  0.583192  0.755926
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 30.8903   17.5332   28.2852   15.8722   …  0.0       0.0      0.0
 23.0956   37.6788   28.7609   29.9926      0.0       0.0      0.0
  4.12581   2.03115  23.8728    2.14232     0.0       0.0      0.0
  2.77408   1.81904  14.7072   25.3561      0.0       0.0      0.0
 28.1635   15.9854   25.7883   14.4711      0.0       0.0      0.0
 21.0568   34.3527   26.222    27.345    …  0.0       0.0      0.0
  3.76161   1.85186  21.7654    1.95321     0.0       0.0      0.0
  2.5292    1.65846  13.409    23.1178      0.0       0.0      0.0
 22.3952   12.7114   20.5065   11.5072      0.0       0.0      0.0
 16.7441   27.3168   20.8513   21.7443      0.0       0.0      0.0
  2.99117   1.47257  17.3075    1.55316  …  0.0       0.0      0.0
  2.01119   1.31879  10.6626   18.383       0.0       0.0      0.0
 21.0843   11.9674   19.3062   10.8336      0.0       0.0      0.0
 15.764    25.7178   19.6309   20.4716      0.0       0.0      0.0
  2.81609   1.38638  16.2945    1.46225     0.0       0.0      0.0
  1.89347   1.24159  10.0385   17.307    …  0.0       0.0      0.0
 18.3575   10.4196   16.8093    9.43253     1.54773   2.49686  1.40111
 13.7252   22.3918   17.092    17.824       3.32607   2.53885  2.64758
  2.45189   1.20708  14.1871    1.27314     0.179299  2.10736  0.189112
  1.64859   1.08102   8.74022  15.0687      0.160574  1.29827  2.2383
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 136 stored entries:
  [1 ,  1]  =  1.70767
  [2 ,  1]  =  0.722222
  [3 ,  1]  =  0.15386
  [4 ,  1]  =  0.0495901
  [5 ,  1]  =  1.0
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [1 ,  2]  =  0.579378
  [2 ,  2]  =  1.98693
  ⋮
  [19, 18]  =  0.0294572
  [20, 18]  =  0.0330001
  [17, 19]  =  0.858349
  [18, 19]  =  0.4561
  [19, 19]  =  0.638699
  [20, 19]  =  0.583192
  [17, 20]  =  0.317026
  [18, 20]  =  0.895403
  [19, 20]  =  0.035222
  [20, 20]  =  0.755926

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.