Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.4694     0.382655   0.731175  0.336572  …  0.0        0.0       0.0
 0.236515   1.31829    0.96673   0.205427     1.0        0.0       0.0
 0.0907095  0.0983458  1.65078   0.106217     0.0        1.0       0.0
 0.617492   0.618893   0.84243   1.7159       0.0        0.0       1.0
 0.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        0.0        0.0       0.0       …  1.0        0.0       0.0
 0.0        0.0        0.0       0.0          0.0        1.0       0.0
 0.0        0.0        0.0       0.0          0.0        0.0       1.0
 1.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        1.0        0.0       0.0          1.0        0.0       0.0
 0.0        0.0        1.0       0.0       …  0.0        1.0       0.0
 0.0        0.0        0.0       1.0          0.0        0.0       1.0
 0.0        0.0        0.0       0.0          0.0        0.0       0.0
 0.0        0.0        0.0       0.0          1.0        0.0       0.0
 0.0        0.0        0.0       0.0          0.0        1.0       0.0
 0.0        0.0        0.0       0.0       …  0.0        0.0       1.0
 0.0        0.0        0.0       0.0          0.382655   0.731175  0.336572
 0.0        0.0        0.0       0.0          0.318286   0.96673   0.205427
 0.0        0.0        0.0       0.0          0.0983458  0.650782  0.106217
 0.0        0.0        0.0       0.0          0.618893   0.84243   0.715896
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.3177    4.69225   10.8578     4.34988   …   6.32416   14.634     5.86271
  3.13408   8.78372   10.809      2.94633      11.8386    14.5683    3.97102
  1.39239   1.45181   11.5356     1.48324       1.95674   15.5475    1.9991
  7.61605   7.65461   15.2818    13.222        10.3168    20.5966   17.8204
  3.79354   1.72521    3.99211    1.59933       4.59894   10.6419    4.26338
  1.15232   3.22954    3.97419    1.08328   …   8.60906   10.5941    2.88774
  0.511943  0.533792   4.24131    0.545348      1.42294   11.3062    1.45375
  2.80022   2.81439    5.6187     4.86136       7.5024    14.9779   12.9591
  9.2192    4.19267    9.70177    3.88675       7.6431    17.686     7.08541
  2.8004    7.84853    9.65823    2.63264      14.3076    17.6066    4.7992
  1.24414   1.29724   10.3074     1.32532   …   2.36482   18.79      2.41602
  6.80518   6.83964   13.6548    11.8142       12.4684    24.8922   21.537
  3.79354   1.72521    3.99211    1.59933       4.59894   10.6419    4.26338
  1.15232   3.22954    3.97419    1.08328       8.60906   10.5941    2.88774
  0.511943  0.533792   4.24131    0.545348      1.42294   11.3062    1.45375
  2.80022   2.81439    5.6187     4.86136   …   7.5024    14.9779   12.9591
  0.893344  0.406272   0.940105   0.376628      2.22479    5.14812   2.06246
  0.27136   0.760526   0.935886   0.255103      4.16472    5.12502   1.39697
  0.120558  0.125703   0.99879    0.128424      0.688364   5.46949   0.703266
  0.659425  0.662763   1.32315    1.14481       3.62937    7.24572   6.26908
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 128 stored entries:
  [1 ,  1]  =  1.4694
  [2 ,  1]  =  0.236515
  [3 ,  1]  =  0.0907095
  [4 ,  1]  =  0.617492
  [9 ,  1]  =  1.0
  [1 ,  2]  =  0.382655
  [2 ,  2]  =  1.31829
  [3 ,  2]  =  0.0983458
  [4 ,  2]  =  0.618893
  ⋮
  [19, 19]  =  0.650782
  [20, 19]  =  0.84243
  [4 , 20]  =  1.0
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.336572
  [18, 20]  =  0.205427
  [19, 20]  =  0.106217
  [20, 20]  =  0.715896

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.