Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.83364    0.674994  0.434652  0.415617  …  0.0       0.0       0.0
 0.081665   1.83171   0.70869   0.252326     0.0       0.0       0.0
 0.388902   0.667269  1.44325   0.381686     0.0       0.0       0.0
 0.0667492  0.539269  0.283382  1.39794      0.0       0.0       0.0
 0.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        0.0       0.0       0.0       …  1.0       0.0       0.0
 0.0        0.0       0.0       0.0          0.0       1.0       0.0
 0.0        0.0       0.0       0.0          0.0       0.0       1.0
 1.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        1.0       0.0       0.0          1.0       0.0       0.0
 0.0        0.0       1.0       0.0       …  0.0       1.0       0.0
 0.0        0.0       0.0       1.0          0.0       0.0       1.0
 1.0        0.0       0.0       0.0          0.0       0.0       0.0
 0.0        1.0       0.0       0.0          1.0       0.0       0.0
 0.0        0.0       1.0       0.0          0.0       1.0       0.0
 0.0        0.0       0.0       1.0       …  0.0       0.0       1.0
 1.0        0.0       0.0       0.0          0.674994  0.434652  0.415617
 0.0        1.0       0.0       0.0          0.831706  0.70869   0.252326
 0.0        0.0       1.0       0.0          0.667269  0.443245  0.381686
 0.0        0.0       0.0       1.0          0.539269  0.283382  0.39794
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 13.6946    12.2543    8.50056  …  6.5735    5.88218  4.08033  3.13806
  2.81581   16.1978    8.68611     1.35161   7.77509  4.1694   2.15627
  4.86084   10.2844   12.2998      2.33324   4.93658  5.90399  2.54459
  1.75543    7.25217   4.72014     0.842621  3.4811   2.2657   4.43908
  5.86967    5.25237   3.64345     6.30037   5.63778  3.9108   3.00768
  1.20689    6.94261   3.72298  …  1.29545   7.45204  3.99617  2.06668
  2.08342    4.40802   5.27184     2.2363    4.73147  5.65868  2.43887
  0.752401   3.10837   2.02311     0.807611  3.33646  2.17156  4.25464
 10.291      9.20868   6.38785     7.70803   6.8974   4.78457  3.67967
  2.11597   12.1721    6.52729     1.58489   9.11701  4.88901  2.52843
  3.65274    7.72833   9.24282  …  2.73594   5.7886   6.92297  2.98377
  1.31914    5.44973   3.54701     0.988051  4.08191  2.65674  5.20523
 10.291      9.20868   6.38785     7.70803   6.8974   4.78457  3.67967
  2.11597   12.1721    6.52729     1.58489   9.11701  4.88901  2.52843
  3.65274    7.72833   9.24282     2.73594   5.7886   6.92297  2.98377
  1.31914    5.44973   3.54701  …  0.988051  4.08191  2.65674  5.20523
  7.70803    6.8974    4.78457     6.69036   5.98676  4.15288  3.19385
  1.58489    9.11701   4.88901     1.37564   7.91332  4.24353  2.19461
  2.73594    5.7886    6.92297     2.37472   5.02435  6.00895  2.58983
  0.988051   4.08191   2.65674     0.857601  3.54299  2.30598  4.518
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 128 stored entries:
  [1 ,  1]  =  1.83364
  [2 ,  1]  =  0.081665
  [3 ,  1]  =  0.388902
  [4 ,  1]  =  0.0667492
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.674994
  [2 ,  2]  =  1.83171
  ⋮
  [18, 19]  =  0.70869
  [19, 19]  =  0.443245
  [20, 19]  =  0.283382
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.415617
  [18, 20]  =  0.252326
  [19, 20]  =  0.381686
  [20, 20]  =  0.39794

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.