Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);
julia> K = A ⊕ B20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.51242   0.427376  0.970357  0.490807  …  0.0       0.0       0.0
 0.806231  1.34606   0.514799  0.759273     0.0       0.0       0.0
 0.799554  0.799146  1.89578   0.397296     0.0       0.0       0.0
 0.632475  0.978118  0.469087  1.4162       0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       0.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0       …  0.0       0.0       0.0
 0.0       0.0       0.0       1.0          0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.0       0.0       0.0
 0.0       1.0       0.0       0.0          0.0       0.0       0.0
 0.0       0.0       1.0       0.0          0.0       0.0       0.0
 0.0       0.0       0.0       1.0       …  0.0       0.0       0.0
 1.0       0.0       0.0       0.0          0.427376  0.970357  0.490807
 0.0       1.0       0.0       0.0          1.34606   0.514799  0.759273
 0.0       0.0       1.0       0.0          0.799146  1.89578   0.397296
 0.0       0.0       0.0       1.0          0.978118  0.469087  1.4162
julia> exp(K)20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 16.5538   11.1744   14.9889    9.47421  …  0.0       0.0      0.0
 13.2246   14.8992   12.8146   10.5237      0.0       0.0      0.0
 15.4464   14.1001   20.9274   10.6904      0.0       0.0      0.0
 12.9285   13.0628   12.6775   13.8169      0.0       0.0      0.0
  0.0       0.0       0.0       0.0         0.0       0.0      0.0
  0.0       0.0       0.0       0.0      …  0.0       0.0      0.0
  0.0       0.0       0.0       0.0         0.0       0.0      0.0
  0.0       0.0       0.0       0.0         0.0       0.0      0.0
 12.6072    8.51032  11.4155    7.21551     0.0       0.0      0.0
 10.0718   11.3472    9.75953   8.01476     0.0       0.0      0.0
 11.7639   10.7385   15.9382    8.14175  …  0.0       0.0      0.0
  9.84627   9.94853   9.65508  10.5229      0.0       0.0      0.0
 12.6072    8.51032  11.4155    7.21551     0.0       0.0      0.0
 10.0718   11.3472    9.75953   8.01476     0.0       0.0      0.0
 11.7639   10.7385   15.9382    8.14175     0.0       0.0      0.0
  9.84627   9.94853   9.65508  10.5229   …  0.0       0.0      0.0
 18.4333   12.4431   16.6908   10.5499      7.24159   9.71363  6.1398
 14.7262   16.5909   14.2696   11.7185      9.6555    8.30456  6.81991
 17.2001   15.701    23.3035   11.9042      9.1376   13.5621   6.92797
 14.3964   14.5459   14.1169   15.3857      8.46539   8.21569  8.95411
julia> collect(K)20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 112 stored entries:
  [1 ,  1]  =  1.51242
  [2 ,  1]  =  0.806231
  [3 ,  1]  =  0.799554
  [4 ,  1]  =  0.632475
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.427376
  [2 ,  2]  =  1.34606
  ⋮
  [19, 18]  =  0.799146
  [20, 18]  =  0.978118
  [17, 19]  =  0.970357
  [18, 19]  =  0.514799
  [19, 19]  =  1.89578
  [20, 19]  =  0.469087
  [17, 20]  =  0.490807
  [18, 20]  =  0.759273
  [19, 20]  =  0.397296
  [20, 20]  =  1.4162

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.