Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.85263    0.161536  0.512146  0.867158   …  0.0       0.0       0.0
 0.459272   1.34066   0.275932  0.0910299     0.0       0.0       0.0
 0.0202177  0.703763  1.73355   0.497508      0.0       0.0       0.0
 0.315653   0.991921  0.526826  1.15774       0.0       0.0       0.0
 0.0        0.0       0.0       0.0           0.0       0.0       0.0
 0.0        0.0       0.0       0.0        …  1.0       0.0       0.0
 0.0        0.0       0.0       0.0           0.0       1.0       0.0
 0.0        0.0       0.0       0.0           0.0       0.0       1.0
 1.0        0.0       0.0       0.0           0.0       0.0       0.0
 0.0        1.0       0.0       0.0           0.0       0.0       0.0
 0.0        0.0       1.0       0.0        …  0.0       0.0       0.0
 0.0        0.0       0.0       1.0           0.0       0.0       0.0
 0.0        0.0       0.0       0.0           0.0       0.0       0.0
 0.0        0.0       0.0       0.0           0.0       0.0       0.0
 0.0        0.0       0.0       0.0           0.0       0.0       0.0
 0.0        0.0       0.0       0.0        …  0.0       0.0       0.0
 1.0        0.0       0.0       0.0           0.161536  0.512146  0.867158
 0.0        1.0       0.0       0.0           1.34066   0.275932  0.0910299
 0.0        0.0       1.0       0.0           0.703763  1.73355   0.497508
 0.0        0.0       0.0       1.0           0.991921  0.526826  1.15774

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 10.6258    5.87757   6.77078  7.13099   …  1.85555   2.13754   2.25126
  3.62975   6.95718   3.22508  2.30293      2.19639   1.01816   0.727036
  2.21939   6.53507   9.69004  3.92162      2.06313   3.05915   1.23806
  3.87959   7.00559   5.22466  6.66789      2.21167   1.64943   2.10506
  4.4918    2.48459   2.86218  3.01445      3.12163   3.59602   3.78733
  1.53439   2.94097   1.36332  0.973506  …  3.69502   1.71287   1.22311
  0.93819   2.76254   4.09622  1.65777      3.47083   5.14646   2.08281
  1.64      2.96144   2.20859  2.81868      3.72073   2.77486   3.54138
 11.2725    6.23526   7.18283  7.56497      2.48459   2.86218   3.01445
  3.85066   7.38058   3.42135  2.44308      2.94097   1.36332   0.973506
  2.35445   6.93278  10.2798   4.16028   …  2.76254   4.09622   1.65777
  4.1157    7.43194   5.54263  7.07369      2.96144   2.20859   2.81868
  3.59398   1.98797   2.29008  2.41192      0.629042  0.724637  0.763189
  1.22769   2.35313   1.09082  0.778921     0.744587  0.345162  0.246469
  0.750665  2.21036   3.27747  1.32641      0.699411  1.03707   0.419708
  1.3122    2.36951   1.76714  2.25529   …  0.749768  0.559165  0.713626
 14.6271    8.09082   9.32037  9.81623      6.00999   6.92332   7.29165
  4.99657   9.57697   4.43951  3.17012      7.11393   3.29774   2.35481
  3.05512   8.99591  13.3389   5.39834      6.68231   9.90835   4.00997
  5.34049   9.64361   7.19205  9.17875      7.16343   5.34237   6.81812

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  1.85263
  [2 ,  1]  =  0.459272
  [3 ,  1]  =  0.0202177
  [4 ,  1]  =  0.315653
  [9 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.161536
  [2 ,  2]  =  1.34066
  [3 ,  2]  =  0.703763
  ⋮
  [17, 19]  =  0.512146
  [18, 19]  =  0.275932
  [19, 19]  =  1.73355
  [20, 19]  =  0.526826
  [8 , 20]  =  1.0
  [17, 20]  =  0.867158
  [18, 20]  =  0.0910299
  [19, 20]  =  0.497508
  [20, 20]  =  1.15774

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.