Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.97475    0.994779  0.0412408  0.731204  …  0.0       0.0        0.0
 0.0360309  1.65832   0.451736   0.832248     1.0       0.0        0.0
 0.439064   0.108364  1.46092    0.663992     0.0       1.0        0.0
 0.384809   0.976756  0.808891   1.64447      0.0       0.0        1.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0       …  0.0       0.0        0.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 1.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        1.0       0.0        0.0          1.0       0.0        0.0
 0.0        0.0       1.0        0.0       …  0.0       1.0        0.0
 0.0        0.0       0.0        1.0          0.0       0.0        1.0
 1.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        1.0       0.0        0.0          1.0       0.0        0.0
 0.0        0.0       1.0        0.0          0.0       1.0        0.0
 0.0        0.0       0.0        1.0       …  0.0       0.0        1.0
 1.0        0.0       0.0        0.0          0.994779  0.0412408  0.731204
 0.0        1.0       0.0        0.0          1.65832   0.451736   0.832248
 0.0        0.0       1.0        0.0          0.108364  1.46092    0.663992
 0.0        0.0       0.0        1.0          0.976756  0.808891   1.64447

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 14.6275   15.8338    7.11476  14.2014   …  12.0589    5.41856  10.8157
  3.72785  13.9253    7.50279  10.8935      10.6054    5.71408   8.29644
  6.0189    6.99881  10.5535    9.28304      5.33025   8.03746   7.06991
  7.11504  13.7334   10.5151   17.3516      10.4593    8.00824  13.2149
  9.61336  10.4061    4.6759    9.33331     10.4061    4.6759    9.33331
  2.44999   9.15185   4.93092   7.15935  …   9.15185   4.93092   7.15935
  3.95569   4.5997    6.93586   6.10092      4.5997    6.93586   6.10092
  4.67609   9.02576   6.91064  11.4037       9.02576   6.91064  11.4037
 19.4018   21.0018    9.43694  18.8366      21.0018    9.43694  18.8366
  4.94459  18.4703    9.95163  14.4491      18.4703    9.95163  14.4491
  7.9834    9.28315  13.998    12.3129   …   9.28315  13.998    12.3129
  9.43731  18.2159   13.9471   23.015       18.2159   13.9471   23.015
 11.1402   12.0589    5.41856  10.8157      12.0589    5.41856  10.8157
  2.83911  10.6054    5.71408   8.29644     10.6054    5.71408   8.29644
  4.58396   5.33025   8.03746   7.06991      5.33025   8.03746   7.06991
  5.41877  10.4593    8.00824  13.2149   …  10.4593    8.00824  13.2149
 11.1402   12.0589    5.41856  10.8157      15.8338    7.11476  14.2014
  2.83911  10.6054    5.71408   8.29644     13.9253    7.50279  10.8935
  4.58396   5.33025   8.03746   7.06991      6.99881  10.5535    9.28304
  5.41877  10.4593    8.00824  13.2149      13.7334   10.5151   17.3516

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 116 stored entries:
  [1 ,  1]  =  1.97475
  [2 ,  1]  =  0.0360309
  [3 ,  1]  =  0.439064
  [4 ,  1]  =  0.384809
  [9 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.994779
  [2 ,  2]  =  1.65832
  ⋮
  [19, 19]  =  1.46092
  [20, 19]  =  0.808891
  [4 , 20]  =  1.0
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.731204
  [18, 20]  =  0.832248
  [19, 20]  =  0.663992
  [20, 20]  =  1.64447

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.