Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 0.613292   0.766114  0.70069    0.148242  …  0.0       0.0        0.0
 0.5391     0.109561  0.689604   0.885145     0.0       0.0        0.0
 0.0975801  0.23626   0.531553   0.165263     0.0       0.0        0.0
 0.516393   0.192489  0.0167909  0.695952     0.0       0.0        0.0
 1.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        1.0       0.0        0.0       …  1.0       0.0        0.0
 0.0        0.0       1.0        0.0          0.0       1.0        0.0
 0.0        0.0       0.0        1.0          0.0       0.0        1.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0          1.0       0.0        0.0
 0.0        0.0       0.0        0.0       …  0.0       1.0        0.0
 0.0        0.0       0.0        0.0          0.0       0.0        1.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0          0.0       0.0        0.0
 0.0        0.0       0.0        0.0       …  0.0       0.0        0.0
 1.0        0.0       0.0        0.0          0.766114  0.70069    0.148242
 0.0        1.0       0.0        0.0          1.10956   0.689604   0.885145
 0.0        0.0       1.0        0.0          0.23626   1.53155    0.165263
 0.0        0.0       0.0        1.0          0.192489  0.0167909  1.69595

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 2.5176    1.48835   1.86884   1.10259   …   0.0       0.0       0.0
 1.4027    1.80119   1.54466   1.76005       0.0       0.0       0.0
 0.424406  0.494599  1.97333   0.554211      0.0       0.0       0.0
 1.22839   0.717298  0.607857  2.38719       0.0       0.0       0.0
 5.54248   3.27661   4.11425   2.42735       3.74757   4.70561   2.77624
 3.08804   3.9653    3.40056   3.87475   …   4.53525   3.88934   4.43169
 0.934327  1.08886   4.34427   1.22009       1.24536   4.9687    1.39546
 2.70429   1.57913   1.33819   5.25539       1.8061    1.53054   6.01078
 6.84641   4.04746   5.08217   2.99841       7.32407   9.19642   5.42575
 3.81453   4.89818   4.20058   4.78633       8.86348   7.60114   8.66107
 1.15414   1.34502   5.36632   1.50713   …   2.43388   9.71059   2.72722
 3.3405    1.95063   1.65302   6.49179       3.52976   2.99121  11.7472
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0           0.0       0.0       0.0
 0.0       0.0       0.0       0.0       …   0.0       0.0       0.0
 9.36401   5.53582   6.95102   4.101         8.81243  11.0653    6.52834
 5.21723   6.69937   5.74524   6.54638      10.6647    9.1458   10.4211
 1.57854   1.83962   7.33964   2.06134       2.92848  11.6839    3.28143
 4.56889   2.66793   2.26087   8.87898       4.24706   3.59907  14.1344

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 120 stored entries:
  [1 ,  1]  =  0.613292
  [2 ,  1]  =  0.5391
  [3 ,  1]  =  0.0975801
  [4 ,  1]  =  0.516393
  [5 ,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.766114
  [2 ,  2]  =  0.109561
  [3 ,  2]  =  0.23626
  ⋮
  [17, 19]  =  0.70069
  [18, 19]  =  0.689604
  [19, 19]  =  1.53155
  [20, 19]  =  0.0167909
  [8 , 20]  =  1.0
  [12, 20]  =  1.0
  [17, 20]  =  0.148242
  [18, 20]  =  0.885145
  [19, 20]  =  0.165263
  [20, 20]  =  1.69595

Kronecker.kroneckersum — Function

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

Kronecker.:⊕ — Function

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

Base.collect — Method

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

Base.exp — Function

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.