Kronecker sums

Kronecker sums

A Kronecker sum between two square matrices of the same size is defined as

\[A \oplus B = A ⊗ I + I \oplus B\,.\]

To construct objects of the KroneckerSum type, one can either use kroneckersum or the binary operator ⊕. Lazy Kronecker sums work like lazy Kronecker products, though there are far fewer methods to process these constructs efficiently. The most important property for Kronecker sums relates to matrix exponentiation:

\[\exp(A \oplus B) = \exp(A) \otimes \exp(B)\,.\]

The function collect can be used to transform a KroneckerSum struct into a sparse array. It is recommended to make it 'dense' this way before doing operations such as multiplication with a vector.

julia> A, B = rand(Bool, 5, 5), rand(4, 4);

julia> K = A ⊕ B
20×20 KroneckerSum{Float64,Array{Bool,2},Array{Float64,2}}:
 1.60089   0.63326   0.599643   0.280141  …  0.0       0.0        0.0
 0.965409  1.62088   0.612175   0.53721      0.0       0.0        0.0
 0.778787  0.299403  1.76775    0.827059     0.0       0.0        0.0
 0.930511  0.337855  0.0228764  1.54808      0.0       0.0        0.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0       …  0.0       0.0        0.0
 0.0       0.0       1.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        1.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       0.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       0.0        0.0       …  0.0       1.0        0.0
 0.0       0.0       0.0        0.0          0.0       0.0        1.0
 1.0       0.0       0.0        0.0          0.0       0.0        0.0
 0.0       1.0       0.0        0.0          1.0       0.0        0.0
 0.0       0.0       1.0        0.0          0.0       1.0        0.0
 0.0       0.0       0.0        1.0       …  0.0       0.0        1.0
 1.0       0.0       0.0        0.0          0.63326   0.599643   0.280141
 0.0       1.0       0.0        0.0          0.62088   0.612175   0.53721
 0.0       0.0       1.0        0.0          0.299403  0.767751   0.827059
 0.0       0.0       0.0        1.0          0.337855  0.0228764  0.548078

julia> exp(K)
20×20 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 30.822    16.1222   16.8459   14.1433   …   7.00791   7.32249    6.14774
 30.1741   27.0935   19.602    18.9346      11.7768    8.5205     8.23039
 27.7433   15.1964   27.6103   20.9605       6.60547  12.0015     9.11097
 21.776    12.1121    8.98379  21.1386       5.2648    3.90502    9.18841
 11.1028    5.8076    6.0683    5.09476      1.65506   1.72935    1.45191
 10.8694    9.75971   7.06111   6.8207   …   2.78134   2.01229    1.94377
  9.99379   5.47409   9.94589   7.55045      1.56002   2.83439    2.15174
  7.84422   4.36305   3.23617   7.61462      1.24339   0.922249   2.17003
 15.1268    7.91243   8.26762   6.94125      6.55813   6.85253    5.75317
 14.8088   13.2969    9.62026   9.29271     11.021     7.97364    7.70216
 13.6158    7.45805  13.5506   10.2869   …   6.18152  11.2312     8.52622
 10.6872    5.94434   4.40905  10.3744       4.9269    3.65439    8.59868
 33.3514   17.4453   18.2284   15.304       10.5913   11.0668     9.29134
 32.6503   29.3169   21.2107   20.4885      17.7988   12.8774    12.4389
 30.02     16.4434   29.8761   22.6806       9.98313  18.1384    13.7698
 23.563    13.106     9.72103  22.8733   …   7.95691   5.90182   13.8868
 23.4124   12.2464   12.7961   10.7433       8.65956   9.04828    7.59667
 22.9202   20.5802   14.8897   14.3827      14.5524   10.5286    10.1702
 21.0738   11.5431   20.9728   15.9215       8.16227  14.8301    11.2583
 16.541     9.2003    6.82407  16.0568       6.50562   4.82537   11.354

julia> collect(K)
20×20 SparseArrays.SparseMatrixCSC{Float64,Int64} with 132 stored entries:
  [1 ,  1]  =  1.60089
  [2 ,  1]  =  0.965409
  [3 ,  1]  =  0.778787
  [4 ,  1]  =  0.930511
  [5 ,  1]  =  1.0
  [13,  1]  =  1.0
  [17,  1]  =  1.0
  [1 ,  2]  =  0.63326
  [2 ,  2]  =  1.62088
  ⋮
  [18, 19]  =  0.612175
  [19, 19]  =  0.767751
  [20, 19]  =  0.0228764
  [12, 20]  =  1.0
  [16, 20]  =  1.0
  [17, 20]  =  0.280141
  [18, 20]  =  0.53721
  [19, 20]  =  0.827059
  [20, 20]  =  0.548078

Kronecker.kroneckersum — Function.

kroneckersum(A::AbstractMatrix, B::AbstractMatrix)

Construct a sum of Kronecker products between two square matrices and their respective identity matrices. Does not evaluate the Kronecker products explicitly.

source

kroneckersum(A::AbstractMatrix, B::AbstractMatrix...)

Higher-order lazy kronecker sum, e.g.

kroneckersum(A,B,C,D)

source

kroneckersum(A::AbstractMatrix, pow::Int)

Kronecker-sum power, computes A ⊕ A ⊕ ... ⊕ A = (A ⊗ I ⊗ ... ⊗ I) + (I ⊗ A ⊗ ... ⊗ I) + ... (I ⊗ I ⊗ ... A).

source

Kronecker.:⊕ — Function.

⊕(A::AbstractMatrix, B::AbstractMatrix)

Binary operator for kroneckersum, computes as Lazy Kronecker sum. See kroneckersum for documentation.

source

Base.collect — Method.

collect(K::AbstractKroneckerSum)

Collects a lazy instance of the AbstractKroneckerSum type into a full, native matrix. Returns the result as a sparse matrix.

source

Base.exp — Function.

exp(K::AbstractKroneckerSum)

Computes the matrix exponential of an AbstractKroneckerSum K. Returns an instance of KroneckerProduct.

source

Missing docstring.

Missing docstring for mul!(x::AbstractVector, K::AbstractKroneckerSum, v::AbstractVector). Check Documenter's build log for details.