Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.536278  -0.12981    -0.40275   …   0.360581   1.11875    4.29909 
 -1.31841    2.23495    -0.861521     -6.20817    2.3931    -7.75624 
  0.27632   -0.770015   -0.591583      2.13892    1.64328   -0.460447
 -0.316997   1.966       1.23604      -5.46108   -3.43342    2.94928 
 -0.240999   0.0583354   0.180992     -0.25854   -0.802149  -3.08248 
  0.592481  -1.00437     0.38716   …   4.45131   -1.71587    5.56128 
 -0.124176   0.346038    0.265852     -1.53362   -1.17824    0.330144
  0.142456  -0.883503   -0.555463      3.91564    2.46178   -2.11466 
 -0.485842   0.117601    0.364872      0.45002    1.39624    5.36544 
  1.19441   -2.02476     0.780496     -7.74805    2.98669   -9.6801  
 -0.250332   0.697597    0.535946  …   2.66946    2.05088   -0.574656
  0.287184  -1.7811     -1.11979      -6.81565   -4.28504    3.68082 
 -0.901961   0.218326    0.677382      0.125557   0.389556   1.49698 
  2.21742   -3.75894     1.44898      -2.16173    0.833296  -2.70078 
 -0.464739   1.29508     0.994978      0.744789   0.572202  -0.160331
  0.533155  -3.30659    -2.07888   …  -1.90159   -1.19554    1.02696 
  0.442022  -0.106994   -0.331963     -0.339042  -1.05192   -4.04229 
 -1.08668    1.84214    -0.7101        5.83733   -2.25015    7.29293 
  0.227754  -0.634677   -0.487606     -2.01116   -1.54512    0.432943
 -0.261282   1.62045     1.01879       5.13487    3.22832   -2.77311 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 0.21875274211849985
 1.7583322556469567 
 1.030119927923502  
 1.5167418884656745 
 0.9725809132667437 
 2.5363548813635286 
 0.3706379234569356 
 0.7775328608539047 
 1.0477423930294558 
 2.581408798134644  
 ⋮                  
 0.4378155937277111 
 0.2628150528290325 
 0.3404223053347122 
 1.6148534773053456 
 2.2933143433564913 
 2.2912253880653966 
 1.3884057611776892 
 1.508869954297279  
 2.1458474197735873 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 -0.0168656  -4.47231
  0.0246984  -3.80011
 -0.26113    -4.37362
 -0.936049   -4.24909
 -0.025409   -2.21201
  0.633632   -8.53755
  4.02199    -5.88845
  2.51985    -6.56062
 -0.701679   -6.95097
  0.211622   -4.81613
  ⋮                  
  2.093      -6.07479
  1.9437     -6.18561
 -1.0845     -6.31554
 -0.70351    -3.98486
  0.781818   -6.69977
  4.80152    -5.63845
  4.41655    -5.24277
  0.658854   -5.8449 
 -1.14562    -4.98605

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.