Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.569036  -1.01389    -1.88619   …   3.16413    5.8864     7.94726 
 -0.885994  -0.12382     2.65535       0.386418  -8.28681    1.50395 
  0.894521   2.58937    -0.298447     -8.08089    0.931391  -2.48125 
 -1.24041   -3.48231     1.64621      10.8676    -5.13747    4.58565 
  0.445489  -0.793756   -1.47667      -1.61442   -3.00339   -4.05489 
 -0.69363   -0.0969369   2.07883   …  -0.19716    4.22814   -0.767353
  0.700306   2.02718    -0.233649      4.12307   -0.475219   1.26599 
 -0.971095  -2.72624     1.28879      -5.5449     2.62127   -2.33971 
  0.408211  -0.727334   -1.3531        3.60707    6.71043    9.05978 
 -0.635587  -0.0888252   1.90488       0.440511  -9.44687    1.71449 
  0.641704   1.85754    -0.214097  …  -9.21211    1.06177   -2.82859 
 -0.889833  -2.49811     1.18094      12.3889    -5.85666    5.22758 
  0.615827  -1.09726    -2.04129       2.37735    4.42271    5.97112 
 -0.958848  -0.134002    2.8737        0.290332  -6.22624    1.12998 
  0.968076   2.80229    -0.322988     -6.07152    0.699794  -1.86427 
 -1.3424    -3.76865     1.78157   …   8.16527   -3.86       3.44539 
  0.969408  -1.72725    -3.2133       -1.23515   -2.29781   -3.10228 
 -1.50938   -0.21094     4.52365      -0.150841   3.23483   -0.58708 
  1.5239     4.41124    -0.508433      3.15445   -0.363577   0.968578
 -2.11315   -5.93245     2.80447      -4.24225    2.00546   -1.79005 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -2.6467519061548566 
 -2.954856069116216  
 -2.302014994625593  
 -3.724934616468932  
 -3.3352512648800867 
 -2.4363965577920594 
 -4.843017516294058  
 -2.2262328358212664 
 -3.064084788672773  
 -2.84338272249973   
  ⋮                  
 -2.1848676793728052 
  0.25195104627831444
 -0.5422209283325448 
 -1.802352367511193  
 -1.9378661566117386 
 -3.8969687775211552 
 -3.324246229868269  
 -4.052931327247235  
 -2.000341117387351  

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  2.09735    -1.1895   
 -0.0436764  -1.09598  
  2.31993    -3.25161  
  1.89835    -2.37438  
  2.77219    -0.612084 
  3.476       0.0593155
 -1.41483     2.1376   
  2.89579    -0.496763 
  3.36099     0.955445 
  2.6152      1.33557  
  ⋮                    
 -0.414402    0.993258 
  3.33724    -1.94908  
  2.95336     0.0596613
  3.9764      0.239981 
  3.74579    -1.64733  
  0.509257   -0.476092 
  4.21802    -2.49444  
  4.78589    -1.62854  
  2.70304    -1.34243  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.