Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   2.39964    2.73896  -0.940397  …   0.886663  -0.304428  -0.996384
  -8.94379   -7.98303   2.13281      -2.58429    0.690439   0.729919
  -9.31587   -8.32561   2.49464      -2.69519    0.807572  -0.12688
 -12.598     -7.98596  -0.66969      -2.58524   -0.216794   2.28095
   0.639572   0.73001  -0.250643     -1.72271    0.591478   1.93589
  -2.38378   -2.1277    0.568455  …   5.02106   -1.34147   -1.41817
  -2.48295   -2.21901   0.664893      5.23653   -1.56905    0.246517
  -3.35772   -2.12849  -0.178491      5.0229     0.421213  -4.43171
  -2.05593   -2.34664   0.8057       -0.441332   0.151528   0.495946
   7.66273    6.83958  -1.82732       1.28632   -0.343663  -0.363314
   7.98152    7.1331   -2.13732   …   1.34152   -0.401965   0.0631538
  10.7935     6.84209   0.573767      1.28679    0.107908  -1.13533
  -1.09786   -1.2531    0.43024       1.73691   -0.596352  -1.95184
   4.09186    3.6523   -0.975779     -5.06243    1.35252    1.42986
   4.26209    3.80904  -1.14132      -5.27968    1.58197   -0.248548
   5.76369    3.65365   0.306389  …  -5.06429   -0.424683   4.46822
  -0.662728  -0.75644   0.259717      2.53138   -0.869126  -2.84462
   2.47008    2.20474  -0.589035     -7.37801    1.97117    2.08388
   2.57284    2.29935  -0.688965     -7.69463    2.30558   -0.362235
   3.47929    2.20555   0.184954     -7.38072   -0.618935   6.51201

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  2.978144277306038
  1.6318227583491522
  0.4924904979489355
  2.8743118978463045
  2.9607225428252377
  1.3093325494632995
  1.3588962014692003
  0.06912148798520146
  2.1074465726895655
  1.9199628426290194
  ⋮
 -1.5697093469260193
 -0.9817888195683426
 -1.1035522719112552
 -1.4346004908750967
  0.91687312819215
  1.1142576095549657
 -0.11549658440600641
  1.7865043895827357
  1.5820861496759855

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.585974  -2.09938
 -2.13943   -2.63728
 -1.27587   -0.760507
 -1.99219   -5.13194
 -2.22331   -1.29709
 -1.31096   -5.74755
 -1.76039   -7.85327
 -0.975421  -4.63116
 -0.599354  -7.83151
 -1.53795   -4.22709
  ⋮         
 -4.04346   -4.4167
 -2.48718   -2.92411
 -2.25439   -5.51191
 -2.48087   -0.551472
 -4.05857   -2.06608
 -4.37733   -2.92682
 -2.93642   -1.68978
 -4.56151   -2.42347
 -4.24243   -0.913479

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.