Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  -0.67022    0.922807  -1.04838      …   0.397977     -0.072758  
  -2.21445   -1.0269     1.75102         -0.664708      1.10904   
   0.432243  -0.620322   0.000949426     -0.000360413  -0.0112148 
  -3.06913    0.358355   1.26373         -0.479727      2.00449   
   2.55377   -3.51622    3.9947           0.226845     -0.0414716 
   8.43782    3.91283   -6.67202      …  -0.37888       0.632148  
  -1.647      2.36364   -0.00361764      -0.000205433  -0.00639239
  11.6944    -1.36546   -4.81526         -0.273441      1.14255   
   1.51678   -2.08841    2.3726           0.423406     -0.0774068 
   5.01153    2.32397   -3.96275         -0.707179      1.1799    
  -0.978211   1.40385   -0.00214865   …  -0.000383441  -0.0119314 
   6.94576   -0.810996  -2.85996         -0.510378      2.13256   
   3.3742    -4.64583    5.27803          0.294325     -0.0538084 
  11.1485     5.16987   -8.81546         -0.491587      0.820195  
  -2.17611    3.12299   -0.00477984      -0.000266544  -0.00829396
  15.4514    -1.80413   -6.36221      …  -0.354783      1.48243   
  -3.28843    4.52775   -5.14388          3.00527      -0.549422  
 -10.8652    -5.03846    8.5914          -5.01945       8.37477   
   2.1208    -3.04361    0.00465835      -0.0027216    -0.0846872 
 -15.0587     1.75827    6.2005          -3.62259      15.1366    

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 6.0709270745327135
 3.3700019146467035
 8.864160078193262 
 3.366424133951044 
 7.386976186229767 
 2.764005931293559 
 3.7270579965008435
 5.176573500504931 
 2.992344748161224 
 4.85696435383488  
 ⋮                 
 3.3527296301903196
 6.5827957469661085
 3.230802979186094 
 6.183703292433524 
 4.727315075977817 
 3.2814037799690494
 7.643502499220579 
 3.0827177276825126
 5.79770015170459  

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.880645   1.35694  
 -0.405139   0.0107769
  0.912147  -2.38549  
  1.56001    1.68351  
 -1.10103   -1.37829  
 -0.401188   2.2329   
 -1.3501     0.564381 
 -1.30353   -0.0292461
 -0.145469   1.51665  
 -2.01506   -1.20997  
  ⋮                   
 -2.42151   -0.830805 
 -3.57452   -1.65845  
 -1.70093   -0.530505 
 -2.32465   -0.809038 
 -3.8961     0.869657 
 -4.68662   -0.180995 
 -5.00755   -0.302109 
 -3.79552    0.51402  
 -4.18794   -1.32276  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.