Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -9.24381    4.06325    11.104     -5.37502    …    1.3604     -0.65852
 -5.25874   -0.249361    1.46946    2.4372          0.180031    0.298593
  4.61314   -5.34589     0.834115  -0.0978483       0.102192   -0.0119879
 -1.45592   -2.44516     3.0748     1.07871         0.376709    0.132158
  4.7449    -2.08569    -5.69975    2.75903        10.4251     -5.04641
  2.69934    0.127998   -0.754284  -1.25103    …    1.37962     2.2882
 -2.36795    2.74408    -0.428157   0.0502262       0.783121   -0.0918662
  0.747334   1.25512    -1.57831   -0.553708        2.88682     1.01276
 -1.90141    0.835794    2.28405   -1.10562       -16.2605      7.87109
 -1.0817    -0.0512925   0.302262   0.501322       -2.15186    -3.569
  0.948903  -1.09963     0.171574  -0.020127   …   -1.22147     0.143288
 -0.299477  -0.502959    0.632474   0.221886       -4.50269    -1.57964
 -3.76905    1.65674     4.52752   -2.1916         26.4678    -12.8121
 -2.14419   -0.101674    0.599155   0.993739        3.50265     5.80938
  1.88095   -2.17972     0.340101  -0.0398965       1.98822    -0.233234
 -0.593635  -0.996985    1.25371    0.439831   …    7.32919     2.57124
  7.54617   -3.31703    -9.06473    4.38789       -12.6723      6.13417
  4.29296    0.203565   -1.19959   -1.9896         -1.677      -2.78142
 -3.76593    4.36411    -0.680929   0.0798783      -0.951923    0.111668
  1.18854    1.9961     -2.51011   -0.880602       -3.50907    -1.23106

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.0928261117060676
  0.6778040474912997
 -1.5928810185862354
  1.3015918418271044
  0.06375518616022635
  2.094531950732758
  1.3549739232542066
 -0.40313756698407743
  2.8116566496837
  1.2657975339636305
  ⋮
  1.7878461526801532
  0.17296432918344723
  2.8399945545692624
  1.0311299802463478
  0.8249770703387005
 -0.47860741234956006
 -2.7944252559592204
  1.0374990722838027
 -0.982014655587125

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -1.69303    6.79988
 -0.818863   2.45344
  1.52544    1.08476
 -1.22707    1.24954
 -0.801208   4.18389
 -3.89014    4.00587
 -3.48322    0.892939
 -0.988079  -1.8496
 -2.97543    1.15458
 -2.73691    1.8756
  ⋮         
 -4.27178    3.2305
 -1.47869    1.07979
 -3.45131    2.02826
 -3.85751    4.27008
 -3.97132    6.7377
 -3.50212    2.1727
 -0.149062   0.0128486
 -3.20162    1.29447
 -3.11871    3.91217

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.