Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.352771  -0.701949     0.984309  …  -0.406077     0.569422    0.104842 
  0.819302   0.0066509   -0.863677      0.00384754  -0.499636    0.114298 
  1.74638   -0.0145062    0.244244     -0.00839183   0.141295   -0.833656 
  2.44315    0.154903    -1.35677       0.0896116   -0.78489    -0.860519 
  0.476703   0.948551    -1.33011      -3.53097      4.95131     0.911633 
 -1.10713   -0.00898744   1.1671    …   0.0334557   -4.3445      0.993855 
 -2.3599     0.0196024   -0.330049     -0.0729697    1.22861    -7.24891  
 -3.30146   -0.209323     1.83342       0.779202    -6.82488    -7.48249  
 -0.17413   -0.346487     0.485861     -1.32177      1.85346     0.341258 
  0.404413   0.00328293  -0.426316      0.0125237   -1.62631     0.372037 
  0.862024  -0.00716036   0.12056   …  -0.0273153    0.459912   -2.71353  
  1.20596    0.0764614   -0.66971       0.291684    -2.5548     -2.80097  
 -1.40112   -2.78797      3.90943       3.6509      -5.11947    -0.942594 
  3.25406    0.0264157   -3.43031      -0.0345919    4.49205    -1.02761  
  6.93618   -0.0576151    0.970076      0.075448    -1.27033     7.4951   
  9.7036     0.615238    -5.38875   …  -0.805665     7.05667     7.73661  
  0.641267   1.276       -1.78928       0.226064    -0.316999   -0.0583656
 -1.48933   -0.01209      1.56999      -0.00214194   0.278149   -0.0636297
 -3.17457    0.0263694   -0.443987      0.00467175  -0.0786592   0.464098 
 -4.44117   -0.281584     2.46634      -0.0498869    0.43695     0.479052 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.03242853675200716
  0.34763289861248   
 -2.218573097125198  
 -1.1438537021923747 
  0.2580705023274274 
  1.3605607583425847 
  1.0579915672073528 
 -0.9075222545777475 
  0.06025971933005106
  0.956612207839726  
  ⋮                  
  1.0791053848178271 
 -2.1814084208443085 
 -0.9027524607930615 
  0.7805062799099387 
 -0.1419226560841125 
  1.7148458633047872 
 -2.353664915486115  
 -0.47661153458076183
  0.6235564819485686 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.245311  -4.75826
 -1.80492   -5.59233
  0.666005  -7.36385
  0.775691  -6.12049
 -0.587277  -6.28871
  3.13969   -3.84861
  0.535447  -3.36941
  3.66485   -5.35568
  3.45358   -4.0319 
  2.35617   -4.79723
  ⋮                 
  0.565388  -3.25274
  5.30299   -4.8996 
  5.08639   -2.99679
  2.55926   -6.02712
  4.1782    -3.15422
  1.27128   -3.22872
  5.10201   -5.64307
  3.83504   -4.42073
  4.05349   -4.62691

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.