Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  -6.8987    -2.88542    0.564448  -10.7073   …   0.560875    -10.6395
  19.3824    -2.43434   -0.86995    15.0419      -0.864443     14.9466
  30.6274     6.34687   -4.80228     7.06681     -4.77188       7.02207
  -2.79741   14.5642    -2.28821     5.55829     -2.27372       5.5231
   6.22457    2.60346   -0.509291    9.66099     -0.0727367     1.37978
 -17.4884     2.19646    0.78494   -13.572    …   0.112105     -1.93835
 -27.6345    -5.72666    4.333      -6.37625      0.618838     -0.910654
   2.52405  -13.141      2.06461    -5.01514      0.294866     -0.716261
  -2.83599   -1.18617    0.232039   -4.40166     -0.411866      7.8129
   7.96791   -1.00073   -0.357628    6.18356      0.634785    -10.9757
  12.5906     2.60913   -1.97417     2.9051   …   3.50413      -5.15651
  -1.14999    5.98721   -0.940659    2.28496      1.66966      -4.05577
   2.50563    1.04799   -0.205009    3.88892     -0.00805252    0.152752
  -7.03975    0.884161   0.315968   -5.46325      0.0124109    -0.21459
 -11.124     -2.3052     1.7442     -2.56669      0.0685102    -0.100816
   1.01603   -5.28977    0.831084   -2.01879  …   0.032644     -0.0792956
  -2.98355   -1.24788    0.244112   -4.63069     -0.0460805     0.874124
   8.3825    -1.05281   -0.376236    6.5053       0.0710211    -1.22799
  13.2457     2.74489   -2.07689     3.05625      0.392049     -0.576921
  -1.20982    6.29874   -0.989604    2.40385      0.186805     -0.453768

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -3.4108675028447615
 -1.0041523473916336
 -0.684723844784575
 -1.9977945951691778
 -0.6197070761768322
 -2.527322798386345
  0.7706992177647941
 -0.06893986827206167
 -1.5073529132427532
 -0.8203124622284933
  ⋮
  1.045226593240434
  0.7650637086342584
 -0.3453367271648143
  0.0008679889498210836
 -1.0081829635477884
 -0.8071507464942446
  1.8655507834928589
  1.5243282832795992
  1.4606326931943372

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.56151     4.78058
  1.16682     5.75787
  1.71355     1.47018
 -0.99234     0.683375
  1.52848     1.89991
 -0.109374    4.01418
  0.266127    4.37837
  1.42535     0.790597
 -0.0324145   1.11679
  1.26278     0.309433
  ⋮          
 -0.107385   -0.480716
  0.771656   -2.62382
  0.751412   -1.78174
  0.606382   -3.161
  1.22316     3.36399
  1.81491     3.21089
  2.73193    -2.57739
  0.878295   -1.06016
  1.88463    -0.503109

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.