Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.655012   0.214567   -0.810068   …  -0.356036    1.34416      1.18092
  0.740459  -0.147975    0.0136518      0.245539   -0.0226527    0.105777
  0.228008   0.421715   -1.65077       -0.699761    2.73916      1.67707
 -0.182293  -0.440028    2.51077        0.730149   -4.16617     -1.77256
 -4.80727   -1.57475     5.94526        0.0946432  -0.357312    -0.313918
 -5.43438    1.08602    -0.100193   …  -0.0652704   0.00602164  -0.0281183
 -1.6734    -3.09505    12.1154         0.186014   -0.728138    -0.445807
  1.33789    3.22946   -18.427         -0.194092    1.10747      0.471189
  3.85711    1.2635     -4.77017        0.258103   -0.974432    -0.856091
  4.36027   -0.871368    0.0803897     -0.178       0.0164217   -0.0766817
  1.34265    2.48331    -9.72073    …   0.507282   -1.98572     -1.21577
 -1.07345   -2.59115    14.7849        -0.529311    3.0202       1.28499
  4.2772     1.40111    -5.28971       -0.417357    1.57567      1.38431
  4.83516   -0.966273    0.0891454      0.287829   -0.0265542    0.123996
  1.48888    2.75378   -10.7795        -0.820283    3.21094      1.96592
 -1.19037   -2.87337    16.3952     …   0.855905   -4.88372     -2.07785
  0.51952    0.170183   -0.642501      -0.395069    1.49153      1.31039
  0.587291  -0.117366    0.0108278      0.272458   -0.0251361    0.117374
  0.180844   0.334481   -1.3093        -0.776478    3.03947      1.86093
 -0.144585  -0.349006    1.9914         0.810198   -4.62292     -1.96689

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.5705697838463685
  0.21975311607885997
 -2.308128078534804
  0.020830784205403158
  2.0654349688329905
 -2.4438987583476623
 -2.683823574607665
 -3.440831166382678
 -1.590775042289526
 -1.697004018189341
  ⋮
 -1.7409924055912775
 -3.248791269398555
 -2.5868475746811144
 -2.7142198603845427
 -2.7177969727198916
 -0.6012263318272095
 -3.001076711685657
 -1.1677483113156184
 -1.0149226436043957

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  2.93025    1.43874
 -0.237046  -1.5915
  5.96608    0.2479
  2.30026    1.65585
  4.06237   -0.372558
  3.21539    1.41609
 -0.411595  -0.70509
  6.36742   -0.244712
  2.5243     2.15008
  4.17251   -0.527188
  ⋮         
 -1.72324   -1.60629
  1.54907   -0.670064
 -0.027173   0.727816
 -0.386874  -1.65546
 -0.685238   1.35047
 -2.22363    0.817775
  0.148516  -1.11996
 -0.502315   1.65871
 -1.29878   -0.0226201

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.