Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -1.65179    1.38114    -0.353525   …    2.88431    -0.738287  -0.0507832
 -1.43519   -2.29996    -5.23725        -4.80314   -10.9373    -1.00839  
  2.90843    0.179865   -2.10051         0.375623   -4.38663    2.12077  
 -1.52055   -6.42666    -1.23946       -13.4212     -2.58845   -0.494704 
  1.33762   -1.11845     0.286285       -3.74195     0.957813   0.0658833
  1.16222    1.86251     4.24114    …    6.23133    14.1894     1.30824  
 -2.35525   -0.145655    1.701          -0.487313    5.69097   -2.75137  
  1.23135    5.20432     1.00372        17.4119      3.35811    0.641802 
  1.20367   -1.00644     0.257615       -3.75085     0.960093   0.0660402
  1.04583    1.67599     3.81642         6.24617    14.2232     1.31135  
 -2.11939   -0.131069    1.53066    …   -0.488473    5.70452   -2.75792  
  1.10804    4.68314     0.903205       17.4533      3.3661     0.64333  
  0.368646  -0.308243    0.0788997      -2.10504     0.538819   0.0370628
  0.320307   0.513305    1.16885         3.50545     7.98228    0.735951 
 -0.649103  -0.0401424   0.468793       -0.274139    3.20146   -1.54779  
  0.339357   1.4343      0.276624   …    9.79509     1.88911    0.361047 
 -1.75052    1.46369    -0.374657        3.8615     -0.988415  -0.0679884
 -1.52098   -2.43744    -5.55031        -6.43043   -14.6428    -1.35003  
  3.08228    0.190617   -2.22607         0.502883   -5.8728     2.83928  
 -1.61144   -6.81081    -1.31355       -17.9682     -3.4654    -0.662308 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -1.1506638690827025  
 -0.6672029836397494  
  0.42632793513577744 
  0.8936324365221049  
  0.4297023423743743  
 -0.5515963321830679  
  0.5320611542609653  
  1.182538994098486   
 -0.08804784563604813 
 -0.5072320806599177  
  ⋮                   
 -0.6698758562053209  
  0.005008867004514631
 -0.9219315211668597  
 -0.5478321979960251  
  0.42338349142070947 
 -0.07597948310398048 
  0.5687721246529572  
  1.2815529150264466  
  1.0573626228218183  

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 2.62303   -2.75403 
 1.03223   -2.52981 
 5.01858   -2.81316 
 3.22836   -2.58468 
 0.99201   -0.773495
 3.37617   -0.91814 
 1.31896   -0.305981
 4.61752   -1.38938 
 4.02943   -0.508914
 1.77123   -0.306621
 ⋮                  
 0.843889  -0.948777
 2.40617   -2.2891  
 1.53907   -1.86702 
 0.792518  -0.827515
 5.24966    0.743231
 2.47951    0.405652
 6.18465    0.396093
 5.70323    0.551855
 2.6857     0.145351

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.