Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.991015   -1.23449    -2.04851    …   2.38826     3.96308      2.57699
 -0.572005   -0.622045    0.0226158      1.20342    -0.0437528    0.60641
  0.0496976  -0.115748   -0.968272       0.223928    1.87323      0.614992
  0.967203    1.19631     3.075         -2.31439    -5.94893     -3.36576
  2.8829      3.59118     5.95921        8.26477    13.7146       8.91788
  1.66399     1.80956    -0.0657903  …   4.16452    -0.15141      2.09853
 -0.144572    0.336717    2.81674        0.774922    6.48247      2.12823
 -2.81363    -3.48011    -8.94531       -8.00914   -20.5868     -11.6475
  7.3393      9.14243    15.171          4.59457     7.62424      4.95765
  4.23617     4.60677    -0.167489       2.31515    -0.0841723    1.16662
 -0.368053    0.857214    7.17087    …   0.430797    3.60375      1.18313
 -7.16295    -8.85966   -22.773         -4.45246   -11.4447      -6.4751
 -6.98419    -8.70008   -14.4369        -4.33504    -7.19356     -4.67761
 -4.03121    -4.38387     0.159385      -2.18437     0.0794176   -1.10072
  0.350244   -0.815738   -6.82391       -0.406462   -3.40019     -1.1163
  6.81638     8.43099    21.6711     …   4.20095    10.7982       6.10934
  4.76503     5.93571     9.84971        2.62327     4.35305      2.83057
  2.75033     2.99093    -0.108742       1.32183    -0.0480581    0.666082
 -0.238957    0.556544    4.65567        0.245963    2.05756      0.675508
 -4.65053    -5.75211   -14.7853        -2.54213    -6.53432     -3.69695

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.7487836676360882
  1.5834833843222589
  1.1855697523440567
  0.9902787469907327
  1.1539649219726564
  2.441539228362886
  1.3564340774123118
  0.6597064973992341
  0.4317469760377357
  0.6503356931032401
  ⋮
  1.3837667472772426
  1.5202577593085351
  0.8633487567369783
  0.6809406119137908
 -0.47201536868470745
 -0.4531445150111843
  1.0148173062862222
  0.509913132173543
 -0.11062785457094627

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.402274   -0.710207
 -0.509848   -2.40883
 -0.923716   -0.711234
 -0.293342   -0.478496
  0.061194   -3.28729
 -0.0686735   1.2381
  0.597423   -1.19745
  0.608826   -0.234865
  0.618375    0.337284
  0.212067   -1.95474
  ⋮          
 -0.224719   -1.15157
  0.260639   -0.623155
  0.554314    0.603793
  0.0553475  -2.58979
 -2.51682     1.53001
 -1.35258    -3.52669
 -2.78094     1.50469
 -2.06676     0.623576
 -2.164      -4.19695

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.