Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -8.48425    2.21995    -2.00596   …   -0.382046    0.345219   -0.77973 
 -3.98049   -6.25955    -1.88651        1.07725     0.324662    0.717099
  4.18255   -5.41522   -24.3552         0.931941    4.19145    -1.55194 
  4.81927   -3.27558    26.1287         0.563715   -4.49666     3.01021 
  1.70926   -0.447237    0.404126       4.07062    -3.67824     8.30787 
  0.80192    1.26107     0.380061  …  -11.4779     -3.45921    -7.64056 
 -0.842627   1.09096     4.90667       -9.92965   -44.6591     16.5356  
 -0.970904   0.659906   -5.26396       -6.00628    47.9111    -32.0732  
 -8.47587    2.21776    -2.00398       -0.876972    0.792438   -1.78984 
 -3.97656   -6.25337    -1.88465        2.47278     0.745249    1.64608 
  4.17842   -5.40987   -24.3312    …    2.13924     9.62133    -3.56243 
  4.81451   -3.27234    26.1029         1.29399   -10.3219      6.90983 
 -6.70776    1.75512    -1.58594        0.720978   -0.651481    1.47147 
 -3.14703   -4.94888    -1.4915        -2.03293    -0.612686   -1.35328 
  3.30678   -4.28134   -19.2556        -1.75871    -7.90991     2.92875 
  3.81018   -2.58971    20.6577    …   -1.06382     8.48588    -5.68072 
 -3.93358    1.02924    -0.930029      -2.21238     1.99912    -4.51531 
 -1.84548   -2.90213    -0.874647       6.2382      1.88007     4.15263 
  1.93917   -2.51067   -11.2919         5.39674    24.2721     -8.98709 
  2.23437   -1.51866    12.1141         3.2644    -26.0396     17.4317  

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.9655468227313275
  0.2580382018612704
  0.0542780728759079
  1.2973768531873833
 -0.7246369859489424
 -1.2758621247914508
 -3.340622556107296 
 -6.794613183510699 
 -1.6839891378999885
 -6.4545883671494755
  ⋮                 
 -1.2768291457979473
 -2.9766152578158627
 -0.4004226650891244
 -3.0849868612963633
 -0.4876595518399683
 -1.8014461671792483
 -4.095650601745633 
 -1.0301377328907186
 -3.559502663835164 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -2.83531    2.06887 
 -2.08109    1.153   
 -2.97583    0.555156
 -2.74236    3.30796 
 -1.06594   -0.336435
 -0.567324   0.551083
 -1.66229    1.0569  
 -1.4188    -0.798309
 -1.21149    1.75817 
 -0.383054  -1.52571 
  ⋮                  
 -4.66069   -1.15705 
 -5.11733    0.596952
 -6.06311    3.3747  
 -4.64443   -1.23273 
 -2.06505    2.67079 
 -2.62      -0.468342
 -2.95499    0.179072
 -2.88569    2.54362 
 -1.63418   -1.75632 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.