Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   0.831762   1.30944     5.53222     …   1.15689     4.88768       2.08996
  -5.38024   -3.53603    -0.0142459      -3.12406    -0.0125862    -4.32009
  -2.266     -1.10191     3.51313        -0.973529    3.10383      -2.31701
   6.02897   -0.2766     -8.58545        -0.244375   -7.5852       10.0763
  -2.65714   -4.18314   -17.6732         -1.964      -8.29763      -3.54805
  17.1877    11.2962      0.0455097   …   5.3036      0.021367      7.33405
   7.23895    3.52014   -11.223           1.65272    -5.26925       3.93349
 -19.2601     0.883623   27.427           0.414865   12.8771      -17.1061
   1.37708    2.16794     9.15923         0.403827    1.70611       0.729529
  -8.90762   -5.85431    -0.0235857      -1.0905     -0.00439336   -1.50799
  -3.75163   -1.82434     5.8164      …  -0.339823    1.08343      -0.808782
   9.98167   -0.457943  -14.2142         -0.0853022  -2.64771       3.51725
  -0.396431  -0.624103   -2.63674        -1.45962    -6.1667       -2.63686
   2.56431    1.68533     0.00678981      3.94157     0.0158797     5.45058
   1.08001    0.525187   -1.67441         1.22828    -3.91605       2.92332
  -2.87351    0.131832    4.09196     …   0.308323    9.5701      -12.713
  -0.544931  -0.857887   -3.62445         0.28946     1.22293       0.52292
   3.52488    2.31664     0.00933322     -0.781658   -0.00314913   -1.08091
   1.48458    0.721918   -2.30164        -0.243583    0.776597     -0.579728
  -3.9499     0.181215    5.62478        -0.0611439  -1.89786       2.52114

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.19648695052773338
  0.7793689732497765
  1.4183125617642913
  1.2132453862503625
 -0.41508769827705166
  1.3466968016296035
  1.8472357491007032
  1.7551931783420978
  2.4402981707300406
  1.703470190161588
  ⋮
  1.9122259066941445
  1.679623499225174
  2.0285190310051417
  0.19216357129971565
  4.183623231491225
  5.239030869417361
  5.754954512129259
  3.8509069621006087
  4.014952367497227

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.22153    0.0497235
  1.78945    0.317344
  1.33512    0.20576
  0.650747   0.299912
  1.28288    0.303276
 -2.58414    0.51373
 -3.16844    0.560298
 -3.06621    0.2838
 -2.51338    0.236565
 -1.79035    0.475312
  ⋮          
  0.0819194  2.52675
 -0.519455   2.31255
 -0.359217   1.82655
  0.527224   1.36574
 -1.0851     0.392059
 -0.307349   1.66491
 -0.350879   0.828812
 -0.604047   0.600558
 -0.0557056  1.21755

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.