Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.391596   -0.226614    0.680976   …   0.752227  -2.26044   -3.1183
 -1.62522    -0.425861   -5.67318        1.41361   18.8317     1.88119
 -1.32846     2.74276     0.506073      -9.10436   -1.67987   -2.61566
  2.77382    -2.5433      0.655191       8.44226   -2.17485   -5.55056
 -0.108645    0.0628722  -0.188931      -0.36252    1.08937    1.5028
  0.450905    0.118152    1.57398    …  -0.68126   -9.07553   -0.906603
  0.368571   -0.760956   -0.140406       4.38766    0.809577   1.26056
 -0.769573    0.705616   -0.181777      -4.06857    1.04813    2.67498
  0.362892   -0.210003    0.63106       -0.228092   0.685417   0.945541
 -1.5061     -0.394645   -5.25734       -0.428638  -5.71018   -0.570421
 -1.23109     2.54172     0.468978   …   2.76065    0.509373   0.793127
  2.5705     -2.35687     0.607166      -2.55988    0.659464   1.68306
  0.0533653  -0.0308822   0.092801      -0.237912   0.714925   0.986247
 -0.22148    -0.0580348  -0.773121      -0.447091  -5.95601   -0.594978
 -0.181038    0.373774    0.0689658      2.8795     0.531302   0.827272
  0.378006   -0.346591    0.0892871  …  -2.67009    0.687855   1.75551
 -1.23541     0.714926   -2.14835        0.400946  -1.20484   -1.66209
  5.12728     1.34351    17.8978         0.753471  10.0375     1.0027
  4.19106    -8.6529     -1.59657       -4.85273   -0.895389  -1.39418
 -8.75089     8.02363    -2.06701        4.49983   -1.15922   -2.95852

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.6889231133538217
  0.06541941175585242
  0.8652687750191512
  1.7216584841596108
 -2.753732306272377
 -2.703675013056814
 -1.6889573220838594
 -0.2121317041657319
 -0.09157753848237249
 -5.291539565956233
  ⋮
 -1.1885894738451184
  0.16386537111411698
  1.034288819478228
 -6.1351806106793685
 -0.6503339929115176
  0.14389103550849133
  0.24006565225831492
  0.6346355919359088
 -2.242226335278257

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.815653    0.0025975
  2.6248     -0.29813
 -2.17387     0.396144
  0.0436894   1.56107
 -0.771466    0.17443
  1.27318     2.00246
  3.25317     0.92336
 -1.5158      3.61272
  1.2672      3.71974
 -0.820569    1.72575
  ⋮          
  3.96885    -2.34737
 -1.8073      1.02088
  0.951775    0.992549
 -1.93687    -1.91405
  0.759268   -2.74017
  2.62275    -1.65279
 -2.10267     0.306021
  0.857015    0.30658
 -0.90937    -2.16839

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.