Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -1.01964     1.34466      6.37178   …  -0.0156594   -0.0742035  -0.0119657
 -0.85731    -0.37544      5.14372       0.00437224  -0.059902    0.0019486
 -2.72321     3.2031       3.14088      -0.0373022   -0.0365776   0.0549885
  0.629701   -4.43724    -11.2714        0.0516745    0.131263    0.02341
  0.0635564  -0.0838154   -0.397167     -0.863656    -4.09252    -0.659938
  0.0534381   0.023402    -0.32062   …   0.24114     -3.30375     0.10747
  0.169744   -0.199657    -0.195778     -2.05732     -2.01735     3.03276
 -0.0392507   0.276583     0.702571      2.84999      7.23948     1.29112
 -0.255466    0.336897     1.59642      -0.198814    -0.942098   -0.151918
 -0.214795   -0.0940646    1.28873       0.0555105   -0.760524    0.0247397
 -0.682288    0.802522     0.786933  …  -0.473594    -0.464394    0.698142
  0.157768   -1.11173     -2.82399       0.656067     1.66653     0.297217
 -0.146885    0.193705     0.917892      0.347171     1.6451      0.265281
 -0.1235     -0.0540843    0.740983     -0.0969333    1.32804    -0.0432008
 -0.392295    0.461426     0.452462      0.826997     0.810932   -1.21911
  0.0907121  -0.63921     -1.62371   …  -1.14563     -2.91012    -0.519004
  0.648571   -0.855308    -4.05296      -0.105707    -0.500902   -0.0807729
  0.545317    0.238809    -3.27182       0.0295143   -0.404361    0.0131538
  1.73218    -2.03743     -1.99785      -0.251804    -0.246913    0.371194
 -0.40054     2.82244      7.1695        0.348823     0.886074    0.158027

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.5699055771186955
  0.6215410894935687
 -1.6065434353332142
 -1.4922822740450472
 -2.246031430245142
 -0.8222252983794579
 -0.7544204482910302
 -1.7215880155955647
 -2.6494324608712256
 -2.1595428434008817
  ⋮
 -0.15918364058097492
 -2.3223403792567376
 -2.2953677841230973
 -2.4641540048018635
 -1.7538143952229144
 -0.6155918074021381
 -1.2730647994123803
 -2.0514651417678684
 -0.5063504841150723

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  -2.3721   -1.31597
  -5.922    -0.579996
  -2.57848  -0.531288
  -2.13242  -2.42023
  -3.51852   0.548109
  -7.27175  -1.1195
  -9.77445   0.903902
  -5.51024   0.335571
  -7.54503  -0.955574
  -4.3796   -1.43544
   ⋮        
 -12.3009   -1.93233
  -6.48422  -0.60961
  -8.56152  -2.73402
  -6.25568  -1.01242
  -6.16318  -2.40527
  -8.783    -0.97902
  -4.22891  -1.2016
  -6.02178  -3.10852
  -2.78535  -1.6235

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.