Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.580047   3.98447    3.62698   …   0.364978   0.332231   -0.280679
  7.49242   -6.15416   -0.715553     -0.563721  -0.0655446  -0.19624
 -8.7977    12.1582     5.93734       1.11369    0.543861   -0.436007
  5.90359   -8.03594   -2.55925      -0.736092  -0.234427    0.325856
 -0.125098  -0.859326  -0.782226     -1.14634   -1.04349     0.881571
 -1.61588    1.32726    0.154322  …   1.77056    0.205866    0.616361
  1.89739   -2.62213   -1.2805       -3.49792   -1.70818     1.36943
 -1.27322    1.7331     0.551949      2.31195    0.736299   -1.02346
  0.160144   1.10006    1.00136       1.98958    1.81108    -1.53005
  2.06856   -1.69909   -0.197555     -3.07298   -0.3573     -1.06976
 -2.42894    3.35672    1.63923   …   6.07099    2.96472    -2.37678
  1.62991   -2.21862   -0.706576     -4.01262   -1.27792     1.77632
  0.299704   2.05873    1.87402      -2.33489   -2.1254      1.7956
  3.87125   -3.17979   -0.369718      3.60632    0.419312    1.25542
 -4.54568    6.28199    3.06776      -7.12464   -3.47927     2.78929
  3.05032   -4.15208   -1.32233   …   4.70903    1.49971    -2.08461
  0.132989   0.913531   0.831568     -0.764587  -0.695987    0.587991
  1.71781   -1.41098   -0.164057      1.18093    0.137309    0.411101
 -2.01707    2.78753    1.36127      -2.33305   -1.13933     0.913385
  1.35353   -1.84242   -0.586765      1.54203    0.491098   -0.682631

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.29057614716673724
 -1.6167857928042995
 -0.5663719312271394
 -0.07216895243383274
 -0.6366487755601118
 -2.228092775165037
 -2.543361074883098
 -2.5686725475428664
 -2.188754948901498
 -2.577511625711109
  ⋮
 -1.696259494325312
 -0.7853483238477411
  0.11917194814277111
 -1.07221347135521
 -1.3935500010860202
 -3.2057181159809693
 -1.8991023661544368
  0.40276513847483897
 -0.9453131703675542

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.40478   2.77981
 -3.99677   6.03802
 -0.581908  4.72827
 -6.07042   6.15521
 -5.43275   6.17829
 -0.191638  0.979045
 -2.33009   2.70703
 -0.798759  2.9539
 -2.9691    2.63396
 -3.26406   2.5188
  ⋮         
 -1.17888   6.65893
  0.361743  5.78155
 -2.15117   6.64117
 -2.06368   6.52285
 -0.253267  2.99365
 -3.58094   4.56915
 -0.748785  4.44276
 -5.42136   4.77082
 -5.04261   4.52142

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.