Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  5.59361   -2.75515     3.30407     …  -3.49256     4.18839     -4.18867 
 -4.87796   -1.78383    -0.0442826      -2.26127    -0.0561347    2.10727 
  5.74247    0.56742     1.51066         0.719289    1.91499     -2.24311 
  2.12138   -0.549403    1.98288        -0.696449    2.51359     -1.06959 
  2.64411   -1.30237     1.56184        -4.56201     5.4709      -5.47126 
 -2.30582   -0.843222   -0.0209325   …  -2.95369    -0.0733235    2.75253 
  2.71448    0.268221    0.714093        0.939539    2.50137     -2.92996 
  1.00278   -0.259704    0.937312       -0.909706    3.28327     -1.3971  
  6.88733   -3.39238     4.06825        -5.72915     6.87058     -6.87103 
 -6.00615   -2.19641    -0.0545245      -3.70936    -0.0920826    3.45674 
  7.07061    0.698656    1.86005     …   1.17991     3.14132     -3.67957 
  2.61203   -0.676471    2.44149        -1.14245     4.12326     -1.75453 
 -0.459645   0.2264     -0.271506        0.458633   -0.550007     0.550043
  0.400837   0.146583    0.00363884      0.296943    0.00737143  -0.276721
 -0.471877  -0.0466267  -0.124136       -0.0944548  -0.25147      0.294558
 -0.174321   0.0451462  -0.16294     …   0.0914555  -0.330077     0.140455
  6.93089   -3.41384     4.09398        -8.11401     9.73057     -9.73121 
 -6.04414   -2.2103     -0.0548693      -5.25344    -0.130413     4.89567 
  7.11534    0.703075    1.87182         1.67107     4.44894     -5.21125 
  2.62855   -0.68075     2.45693        -1.61801     5.83964     -2.48489 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.1111444283931613 
 -1.0600668290873103 
 -1.078488437947408  
 -1.1971208704638177 
 -0.5259143368464864 
 -0.3041883071175995 
 -0.7228904141793485 
 -0.5189359497898218 
 -0.8795465565903229 
 -0.4001568117051992 
  ⋮                  
  0.0657297513578148 
  0.6662879696270854 
 -0.48554410103350676
 -0.15704241860133328
 -0.18252124514694343
 -1.5545846642970922 
 -0.8826553389796821 
 -1.9904596568770907 
 -0.8219458959786533 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -1.28819    0.900614 
  1.25519    0.585504 
  1.98086    1.17477  
 -1.09544   -0.570043 
  0.968606   1.3147   
 -1.76403   -0.362463 
  0.656388  -0.874224 
  1.17272   -0.220851 
 -1.22638   -1.50013  
 -0.219631   0.182463 
  ⋮                   
 -0.774519  -2.02913  
 -0.193637  -1.14045  
 -1.48779   -1.84548  
 -1.12158   -0.55838  
 -0.511507   0.104132 
  2.66109    1.77454  
  3.53418    1.38948  
 -0.832095  -0.0896104
  2.89405    0.484417 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.