Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  -0.436655    1.65602    0.9884    …   22.5176    13.4397    -10.9674
  -2.39322    -1.55789   -0.596147     -21.1832    -8.10605    27.265
  -1.35358     0.334207  -0.293899       4.54435   -3.99626     1.04449
   0.906103   -0.88253   -0.479467     -12.0001    -6.51952     5.74758
  -1.99247     7.55647    4.5101         1.65129    0.985576   -0.804274
 -10.9203     -7.10868   -2.72023   …   -1.55344   -0.594443    1.99943
  -6.17644     1.525     -1.34107        0.333252  -0.293059    0.0765961
   4.13458    -4.02701   -2.18782       -0.880009  -0.478097    0.421489
  -3.35158    12.7109     7.58654      -16.249     -9.69826     7.91421
 -18.3693    -11.9577    -4.57577       15.2861     5.84944   -19.6748
 -10.3895      2.56523   -2.25584   …   -3.27926    2.88375    -0.75372
   6.95487    -6.77393   -3.68019        8.65945    4.70457    -4.14753
   2.92488   -11.0926    -6.62067       -3.91993   -2.33962     1.90924
  16.0307     10.4353     3.99321        3.68764    1.41113    -4.74637
   9.06681    -2.23864    1.96864       -0.791094   0.695681   -0.181829
  -6.06942     5.91152    3.21165   …    2.08902    1.13494    -1.00056
  -3.20788    12.1659     7.26127        5.99827    3.58008    -2.92151
 -17.5817    -11.445     -4.37958       -5.64282   -2.1593      7.26288
  -9.94409     2.45525   -2.15912        1.21053   -1.06453     0.278234
   6.65667    -6.48349   -3.5224        -3.19661   -1.73668     1.53105

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.9168256246582644
  2.141250593454517
  0.4517548202729908
  0.4341514218534309
  0.04829552598481873
  2.411861227146091
  2.2738136958326116
  1.2319591109960537
  0.7138106943695943
  0.546496096379151
  ⋮
  0.1409046970204398
  0.20520784214229007
 -1.3269069430299076
 -0.049414645770609104
  0.12007311357003603
 -0.5599141997642317
 -0.01775850333836783
 -1.0719205491176107
  0.5982150004982026

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.0589486   0.30846
  1.29059    -0.158216
  1.755       0.930589
  0.92999     1.09677
  1.63972    -0.71715
 -0.141085    1.10831
  2.22755    -1.60117
  2.72649    -0.278181
  1.00269     1.70107
  2.83951    -2.55876
  ⋮          
  0.670341   -1.46243
  1.49371     0.201551
  1.60455     2.245
  0.964061   -2.40311
  0.537921    0.664741
  1.49676    -1.4274
  2.16506    -0.577972
  1.46272     1.19133
  1.84587    -2.37842

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.