Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -9.45865    -0.218049    0.439427   -4.32433    …   0.00317204  -0.0312156
  8.01084    -2.26772     8.81112     5.06652        0.0636039    0.0365732
  8.98729    -1.00379    -2.5035      0.96507       -0.0180718    0.00696644
 -0.197993    4.78422    -1.67348    -1.10341       -0.0120802   -0.00796508
 -8.24639    -0.190103    0.383108   -3.77011        0.211208    -2.07847
  6.98414    -1.97708     7.68185     4.41718    …   4.23502      2.4352
  7.83544    -0.875139   -2.18264     0.841382      -1.2033       0.463856
 -0.172617    4.17105    -1.459      -0.961994      -0.804351    -0.53035
  4.70387     0.108438   -0.218531    2.15053       -0.124333     1.22354
 -3.98386     1.12776    -4.38185    -2.51963       -2.49304     -1.43354
 -4.46946     0.499193    1.24501    -0.479937   …   0.708349    -0.27306
  0.0984637  -2.37923     0.832237    0.548736       0.4735       0.312203
  0.940362    0.021678   -0.043687    0.429917      -0.117327     1.1546
 -0.796423    0.225452   -0.875985   -0.503704      -2.35257     -1.35276
 -0.8935      0.0997949   0.248894   -0.0959454      0.668435    -0.257673
  0.0196841  -0.475638    0.166375    0.109699   …   0.446819     0.294611
 -0.990484   -0.0228335   0.0460155  -0.452832       0.285994    -2.81442
  0.838873   -0.237469    0.922676    0.530552       5.73458      3.29747
  0.941124   -0.105114   -0.26216     0.101059      -1.62937      0.6281
 -0.0207333   0.50099    -0.175242   -0.115546      -1.08916     -0.718138

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.2605128371040331
  0.08320185179165192
 -0.9473165627819703
 -0.1598080629324541
  0.43311604984160607
  0.03649105391653346
 -0.2543014883065758
 -1.1466155345409965
 -0.4933362566796839
 -0.5359107651339474
  ⋮
  0.14915962335011515
  0.34881923964603123
  0.5246221445629987
  0.37934548775547283
 -1.722668863032084
 -1.755285980448964
 -1.3318426542963913
 -0.30279151167829504
 -1.6216726398891432

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  5.7284     -2.47268
  1.53005    -1.27648
  6.20465    -2.38791
  8.08579    -3.97376
  2.04592    -0.298313
  3.6306     -2.18717
 -0.0959859  -0.560205
  3.72941    -1.01238
  3.98219    -3.21933
  0.609885    0.858008
  ⋮          
  2.06495    -1.72768
  7.94139    -2.70412
  9.64086    -4.70045
  2.82073    -0.0185201
  4.69578    -1.97616
  1.77412    -1.07702
  6.32358    -1.38736
  7.55631    -2.49738
  2.83297     0.278241

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.