Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  -1.43321    -1.71637   -1.16808   …   1.36404    0.928306   -2.35108
   0.31633     1.03443    0.364975     -0.822085  -0.290054    0.609585
   0.631643    1.03774    0.394076     -0.824716  -0.313182    0.754595
  -3.2347     -4.85085    0.937405      3.85509   -0.744979   -1.5619
  -1.77934    -2.13089   -1.45018       0.847827   0.576992   -1.46132
   0.392725    1.28425    0.453118  …  -0.510969  -0.180284    0.37889
   0.784188    1.28836    0.489247     -0.512605  -0.194659    0.469021
  -4.0159     -6.02236    1.16379       2.39615   -0.463044   -0.970805
  -0.731001   -0.875425  -0.595774      1.86791    1.27121    -3.21956
   0.161342    0.527603   0.186153     -1.12576   -0.397198    0.834761
   0.322165    0.529291   0.200996  …  -1.12936   -0.428868    1.03334
  -1.64984    -2.47415    0.478117      5.27913   -1.02017    -2.13886
  -7.65655    -9.16927   -6.24018       7.91712    5.38802   -13.646
   1.68991     5.52615    1.94978      -4.7715    -1.68352     3.53812
   3.37438     5.54383    2.10524      -4.78677   -1.81775     4.37978
 -17.2805    -25.9144     5.00783   …  22.3755    -4.32397    -9.06551
   2.95976     3.54453    2.41224      -1.96494   -1.33725     3.3868
  -0.65326    -2.13622   -0.753718      1.18423    0.41783    -0.878123
  -1.30442    -2.14306   -0.813815      1.18802    0.451146   -1.08701
   6.68006    10.0176    -1.93586      -5.55336    1.07316     2.24996

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  2.042645143272722
  0.1491211059115482
  3.6419846068983004
 -0.32355442739622536
  3.6999487368513853
 -0.3608700418924929
 -0.9271967695689073
  2.4535294222820636
 -0.388879463657218
  0.6262204324923435
  ⋮
  0.7472118044426003
  4.157208278430003
  0.5310298692575658
  3.5905092926785893
 -1.168235424004566
 -1.2343042537301836
  3.580328406114107
  0.2041268529621573
  0.13544210350894628

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  4.68722  2.30309
  3.62869  2.96934
  4.50143  0.903993
  6.31028  3.27021
  9.0379   2.42305
  5.5921   1.2066
  4.83098  1.46214
  6.52831  1.75529
  8.34416  2.85996
 11.697    2.66776
  ⋮        
  3.57252  3.2543
  4.99256  1.99379
  6.79763  4.54941
  9.56357  4.36805
  6.64526  1.75868
  5.39297  2.61244
  7.10901  1.58992
  9.36295  4.11478
 13.4374   2.7046

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.