Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.230898    0.094564   -0.336913   …   0.351297   -1.2516      1.08982
 -0.199751    0.108875    0.348589       0.404461    1.29498    -7.77738
 -0.0916309  -1.34077     0.426442      -4.98083     1.58419     9.9553
 -0.312592   -1.27913     0.621535      -4.75187     2.30895     5.52997
 -0.53045     0.217245   -0.774002      -0.286836    1.02194    -0.889847
 -0.458895    0.250122    0.800826   …  -0.330244   -1.05736     6.35027
 -0.210507   -3.08019     0.979679       4.06687    -1.2935     -8.12855
 -0.718128   -2.9386      1.42787        3.87993    -1.88527    -4.51525
  2.21262    -0.906176    3.22853       -0.832822    2.96719    -2.58365
  1.91415    -1.04331    -3.34042       -0.958858   -3.07002    18.4379
  0.878068   12.8481     -4.08645    …  11.8081     -3.75566   -23.6011
  2.99547    12.2575     -5.95597       11.2653     -5.47384   -13.1099
  0.0928518  -0.0380273   0.135484      -0.0673461   0.239941   -0.208927
  0.0803266  -0.0437823  -0.140179      -0.0775381  -0.248257    1.49098
  0.0368478   0.539167   -0.171487       0.954862   -0.303701   -1.9085
  0.125704    0.514383   -0.24994    …   0.910968   -0.442642   -1.06014
 -0.0381226   0.0156131  -0.0556263     -0.793169    2.82591    -2.46064
 -0.0329801   0.0179759   0.0575541     -0.913204   -2.92384    17.56
 -0.0151288  -0.221368    0.070408      11.2459     -3.57684   -22.4774
 -0.0516107  -0.211192    0.102619      10.7289     -5.21322   -12.4857

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.6750319213116673
  1.280451616079594
  1.4094685580846384
  1.087706068343921
  2.7623672890243762
  1.3220919632102235
  1.3956178925435963
  1.4950340536546465
  1.6965932544130589
  2.3152658946767692
  ⋮
 -1.8901286471103282
  0.433812271744597
 -1.391713493422136
  0.9446980142411984
  2.8317654636529457
  2.540668471499964
  1.7342730134495146
  2.9153437507690056
  3.3777417409544137

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.444825     1.81482
  0.684587     5.17404
  1.04701      2.18354
  0.00186363   2.84596
  1.09652      2.35063
 -0.231892     1.99233
  1.9275       5.46003
  1.069        2.81861
 -1.59638      4.74446
  1.50391      2.79368
  ⋮           
 -1.43867      1.95323
  0.104808    -0.217508
 -0.701086     1.77111
 -1.3523       2.35974
  1.13766      1.76703
  2.88861      5.14889
  1.72298      1.00133
  0.888948     2.00632
  1.21442      2.62343

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.