Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.145059   -0.0312393   0.571679  …  -0.271232    4.96353    1.08777
  0.120456    0.148261   -0.609326      1.28725    -5.2904    -4.81872
  0.347921   -0.738887    0.103581     -6.4153      0.899328  -4.8757
  0.0188666   0.201864    0.480707      1.75266     4.17368    2.64164
  0.21615    -0.0465493   0.85185      -0.0931733   1.70507    0.373671
  0.179489    0.220921   -0.907947  …   0.442196   -1.81735   -1.65532
  0.518432   -1.101       0.154344     -2.20378     0.308936  -1.6749
  0.0281128   0.300795    0.716294      0.602072    1.43374    0.907452
 -0.199385    0.0429388  -0.785779     -0.318231    5.82361    1.27626
 -0.165568   -0.203786    0.837526      1.51031    -6.20712   -5.65371
 -0.478221    1.01561    -0.142373  …  -7.52694     1.05516   -5.72056
 -0.0259323  -0.277465   -0.660737      2.05636     4.8969     3.09938
 -1.37875     0.296921   -5.43365      -0.240255    4.39666    0.963542
 -1.1449     -1.40918     5.79148       1.14024    -4.6862    -4.26839
 -3.30689     7.02292    -0.984508     -5.68263     0.796619  -4.31887
 -0.179322   -1.91866    -4.56899   …   1.5525      3.69702    2.33994
 -0.426216    0.0917883  -1.67972       0.378572   -6.92786   -1.51826
 -0.353926   -0.435623    1.79034      -1.79669     7.38409    6.72574
 -1.02227     2.17102    -0.304344      8.95416    -1.25524    6.80527
 -0.0554343  -0.593123   -1.41243      -2.44628    -5.82542   -3.68707

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -3.104178167193847
 -1.5615820216094016
 -1.2852496264057087
 -2.3954033269067074
 -1.625425281979636
 -4.892255185992109
 -2.967128509894668
 -2.197633351636965
 -3.0496315762277195
 -4.660562058409884
  ⋮
 -3.61272384649194
 -1.7051418377490266
 -2.6483137456175427
 -3.408199614808233
 -2.502666770819107
 -1.6956999206918488
 -1.3612624649303935
 -1.811665490680913
 -1.6422226531918338

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 1.10393    2.71824
 3.12192    1.46274
 2.99458    0.190286
 1.22927   -1.1127
 3.45429    0.705344
 1.65831   -0.103392
 3.06404    0.961622
 2.73364   -0.0342393
 1.66325   -0.190465
 3.1845    -1.28531
 ⋮         
 3.0517     0.850126
 1.88742   -0.655697
 0.429361  -0.389358
 1.83726   -0.586197
 0.431786   1.72065
 1.54857    3.13397
 1.16536    1.01362
 0.379303   1.40517
 1.21594    1.97937

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.