Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -2.0887      0.926029    -2.126     …  -0.9648     2.21501    1.06283   
  6.26092    -4.23081     -6.35323       4.40794    6.61922   -0.00238898
  1.11647    -2.34584     -5.3756        2.44406    5.60067    0.801393  
 -0.204458    1.98609      2.88751      -2.06924   -3.0084    -0.77164   
 -0.206279    0.0914544   -0.209963     -0.650154   1.49264    0.716215  
  0.618327   -0.417833    -0.627443  …   2.9704     4.46052   -0.00160987
  0.110262   -0.231675    -0.530893      1.64699    3.77415    0.540038  
 -0.0201922   0.196146     0.285169     -1.39441   -2.02728   -0.519988  
 -4.10432     1.81966     -4.17762       0.430619  -0.988627  -0.474374  
 12.3028     -8.31358    -12.4842       -1.9674    -2.95436    0.00106628
  2.19387    -4.6096     -10.5631    …  -1.09086   -2.49975   -0.357686  
 -0.401763    3.90268      5.67398       0.923565   1.34274    0.344406  
 -2.3544      1.04383     -2.39645       1.18856   -2.72873   -1.30933   
  7.05738    -4.76901     -7.16143      -5.43026   -8.15439    0.00294305
  1.2585     -2.64426     -6.05944      -3.01089   -6.8996    -0.987256  
 -0.230468    2.23874      3.25483   …   2.54915    3.70613    0.950603  
  0.635876   -0.281917     0.647233      0.447627  -1.02767   -0.49311   
 -1.90605     1.28801      1.93415      -2.0451    -3.07104    0.00110839
 -0.339894    0.714159     1.63653      -1.13394   -2.59848   -0.371813  
  0.0622445  -0.604638    -0.879062      0.960042   1.39577    0.358009  

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.8336053618882439
 -1.1287550595152023
 -1.2713967767249978
 -1.1205127079181816
 -1.2354760508419504
 -1.211256934464334 
 -1.0198509737702182
 -0.7190155758707992
 -1.1745137955461364
 -2.5207355622637637
  ⋮                 
 -1.8011257339245672
 -1.5156366465783717
 -1.2255161129426717
 -3.596295174009166 
  1.7262158315938851
  0.2195333122606017
  0.8176106486345752
  0.9024047086422506
 -1.2922563370727984

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.148482  6.99819 
  1.86417   5.64884 
  0.781298  2.17336 
 -0.239633  6.33585 
 -0.492305  4.40793 
  0.152654  4.62561 
  2.00895   1.72504 
  1.73201   0.538821
  0.841423  3.91377 
  0.868978  3.04226 
  ⋮                 
  1.36616   4.22528 
  0.473871  1.53926 
  0.222258  4.9514  
 -0.834387  4.00214 
 -0.544512  3.62388 
  1.8432    3.793   
  0.812973  1.99921 
  0.839917  5.30409 
 -1.41345   2.98092 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.