Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   0.867531   0.342212  -0.101681   …   0.147563   -0.0438452    0.854349
  -1.66156   -0.400119  -0.283986      -0.172532   -0.122455     1.47781
   3.19979    0.159999   0.0212395      0.0689918   0.0091585   -0.343302
   2.7347     0.200266  -1.04776        0.086355   -0.451797     0.466425
   0.850526   0.335504  -0.0996882      0.748812   -0.222494     4.33543
  -1.62899   -0.392276  -0.278419   …  -0.875521   -0.621404     7.4992
   3.13707    0.156863   0.0208231      0.350102    0.0464751   -1.7421
   2.68109    0.19634   -1.02722        0.438212   -2.29266      2.36689
  -4.89982   -1.93282    0.574297      -0.664418    0.197418    -3.84681
   9.38452    2.25988    1.60395        0.776847    0.551369    -6.65402
 -18.0724    -0.903675  -0.119961   …  -0.310644   -0.0412373    1.54576
 -15.4456    -1.1311     5.91777       -0.388824    2.03427     -2.10014
   1.86928    0.737368  -0.219094       0.300832   -0.0893861    1.74174
  -3.58019   -0.862141  -0.611907      -0.351737   -0.249646     3.01277
   6.89462    0.344751   0.0457649      0.140652    0.0186712   -0.69988
   5.89247    0.431514  -2.25762    …   0.17605    -0.921067     0.950889
  -2.91199   -1.14868    0.341307      -1.57515     0.468023    -9.1197
   5.57727    1.34306    0.953237       1.84169     1.30714    -15.7748
 -10.7405    -0.537058  -0.0712932     -0.736449   -0.0977618    3.66455
  -9.17938   -0.672219   3.51695       -0.921791    4.82268     -4.97883

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -3.12608725484655
 -3.2617985598743804
 -1.91505715325115
 -2.3482925074613448
 -2.9371336605119596
 -1.2235628177850053
 -1.47941765145807
 -0.09995280976112994
  0.21641229570314405
 -0.45183394081839434
  ⋮
 -1.7504945648422385
 -1.0820940387279987
 -1.7762606564684318
 -1.5688245611905083
 -4.145435818215507
 -3.221732994877371
 -1.5384909154713167
 -2.3154183055525133
 -2.568326825867578

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -3.32483  -2.11328
 -5.95518  -2.72595
 -3.40413  -2.43468
 -3.61688   0.284823
 -4.80702   0.0311315
 -2.38916  -0.0869095
 -3.47079  -0.0235541
 -2.39891  -0.81843
 -2.49709   1.97981
 -2.88803   2.05965
  ⋮        
 -4.36941  -1.20304
 -3.41031  -1.62946
 -4.9007    1.36452
 -4.58659   1.30536
 -2.53657  -1.79776
 -4.3578   -1.84803
 -2.41654  -1.1562
 -3.39029   0.126174
 -4.25216  -0.523175

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.