Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -1.41597   -0.634721   -0.82631    …  -4.01667   -5.22909    -7.58473
  0.786877   0.913958    0.359631       5.78375    2.27584    -0.0398716
 -0.995304  -0.395844    0.0658113     -2.505      0.41647    -5.01377
 -1.15645   -0.323985    0.919331      -2.05026    5.81775     3.95641
 12.105      5.42616     7.06403        1.74613    2.2732      3.29724
 -6.72692   -7.81332    -3.07445    …  -2.51432   -0.989355    0.0173331
  8.50874    3.38402    -0.562613       1.08898   -0.181049    2.1796
  9.88636    2.76971    -7.85926        0.891291  -2.5291     -1.71994
  8.4957     3.80827     4.95779       -7.35561   -9.57587   -13.8897
 -4.72119   -5.48367    -2.15776       10.5916     4.16767    -0.0730157
  5.97174    2.37503    -0.394862   …  -4.58732    0.762669   -9.18157
  6.9386     1.94388    -5.51591       -3.75457   10.6539      7.24525
 -7.94352   -3.56075    -4.63555       -3.00544   -3.91263    -5.67522
  4.41434    5.12725     2.01751        4.32765    1.70288    -0.0298337
 -5.5836    -2.22066     0.369198      -1.87435    0.311621   -3.75152
 -6.48762   -1.81754     5.1574     …  -1.53409    4.35309     2.96035
  0.357149   0.160095    0.208419       2.03164    2.64489     3.83638
 -0.198473  -0.230527   -0.0907095     -2.92544   -1.15112     0.0201672
  0.251045   0.0998433  -0.0165995      1.26703   -0.210652    2.53598
  0.29169    0.0817185  -0.231882       1.03703   -2.94264    -2.00116

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.8732028865534095
 -0.4713619754955599
  1.151994188869568
  0.4530859295184606
 -1.0289208685903914
 -2.557131243968723
 -5.606978585939877
 -4.5408861723006435
 -3.215197450946855
 -2.1736822640065254
  ⋮
 -2.2775682510587627
 -2.513796579857731
 -0.49281054900392074
 -0.19563486071303227
 -0.2633446226701133
 -3.274578954963514
 -2.717018017998703
 -0.7185585428910832
  1.1692304644213616

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.65095     0.673186
  1.08421    -0.118595
  0.826557    0.48378
  2.38658    -0.144422
 -0.0628015   0.0195526
  5.58909     1.95851
  1.23669     2.05705
  3.78343     2.01431
  3.51855     0.53174
  0.914412    2.33533
  ⋮          
  2.54741     2.73608
  4.09452     3.2244
  3.87428     2.3917
  2.08825     2.44305
  3.97916    -1.24857
 -0.216333    0.0153396
  2.29598    -0.471994
  2.02098    -0.469581
 -1.06236     0.186135

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.