Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   0.837087   -0.416975  -1.69647   …  -0.852643  -3.46898    6.10006
  -2.3028      1.78934    0.673207      3.6589     1.37659   -3.26494
  -0.43595    -0.158458   0.431168     -0.324019   0.881664  -3.6746
   2.16405    -1.40276    0.439047     -2.8684     0.897774   1.28537
  -2.09152     1.04184    4.23873      -1.65187   -6.72066   11.818
   5.75369    -4.47079   -1.68205   …   7.08859    2.66695   -6.32537
   1.08925     0.395918  -1.0773       -0.627741   1.7081    -7.11901
  -5.40702     3.50488   -1.09699      -5.55711    1.73931    2.49022
  -5.12448     2.55264   10.3854       -0.770722  -3.13569    5.51398
  14.0973    -10.954     -4.12124       3.30736    1.24433   -2.95126
   2.6688      0.970049  -2.63953   …  -0.292888   0.796955  -3.32155
 -13.2479      8.58741   -2.68776      -2.59281    0.811518   1.16187
  -1.54651     0.770357   3.1342        0.283611   1.15387   -2.02904
   4.25439    -3.30579   -1.24374      -1.21704   -0.457891   1.08601
   0.805412    0.292749  -0.796577      0.107777  -0.293265   1.22227
  -3.99805     2.59158   -0.811133  …   0.954104  -0.298623  -0.427548
  -5.70919     2.8439    11.5704       -0.59562   -2.42328    4.26124
  15.7058    -12.2039    -4.59148       2.55595    0.961629  -2.28075
   2.97331     1.08073   -2.9407       -0.226346   0.615893  -2.56692
 -14.7595      9.56723   -2.99443      -2.00374    0.627147   0.897905

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -1.2158884822659572
 -1.749941727466687
 -1.4963135390320563
 -0.4205871675452365
 -0.6190084350282317
 -0.33754493619957515
 -0.7223062946980549
 -0.3805903491258332
  1.4880059292229513
  1.304431645163581
  ⋮
  2.6602720363856056
  2.873206569484494
  4.15342787518232
  3.8428029848412
 -1.822043395537644
 -1.9883033387778382
 -1.9863231000030912
 -0.06047111212706824
  0.0031555819060270227

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.821099   -2.62807
 -0.206928   -1.9483
  4.03146    -3.54204
  0.825418   -3.3452
  5.02223    -3.47822
  1.75312    -5.41152
  0.301891   -2.61362
  4.64673    -7.17145
  2.19862    -6.56609
  6.59272    -7.39925
  ⋮          
  2.05587    -4.08658
  7.4672     -6.74488
  2.42843    -6.6219
  6.40728    -8.79336
  0.0282269  -2.46874
 -0.481578   -1.78345
  1.86219    -3.60868
  0.115308   -3.03377
  3.62324    -3.27771

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.