Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.58673    1.97768   -0.75247   …  -1.01774    0.38723     0.32145
  8.75876   -5.45516   -0.60694       2.80729    0.312338   -1.22237
  4.98914   -2.73514   -0.937492      1.40753    0.482445   -0.542252
 -6.69693    3.88521    2.21507      -1.99937   -1.1399      0.386932
 -0.447677   1.50898   -0.574137     -3.56633    1.35692     1.12642
  6.68296   -4.1623    -0.463097  …   9.83724    1.09449    -4.28341
  3.80673   -2.08692   -0.71531       4.93225    1.69057    -1.90015
 -5.10978    2.96443    1.69011      -7.00616   -3.99442     1.35588
  0.198814  -0.670137   0.254975     -1.02314    0.389284    0.323155
 -2.96791    1.84848    0.205662      2.82218    0.313995   -1.22886
 -1.69057    0.926801   0.317669  …   1.415      0.485004   -0.545129
  2.26926   -1.3165    -0.750578     -2.00998   -1.14595     0.388985
  0.140413  -0.473288   0.180077     -1.38741    0.527882    0.438209
 -2.0961     1.3055     0.14525       3.82697    0.425788   -1.66637
 -1.19397    0.654559   0.224356      1.91879    0.657681   -0.739212
  1.60268   -0.929787  -0.5301    …  -2.7256    -1.55395     0.527476
  0.273309  -0.921235   0.350513     -0.158927   0.0604688   0.0501968
 -4.07997    2.5411     0.282722      0.438379   0.0487739  -0.190883
 -2.32402    1.27407    0.436699      0.219797   0.0753373  -0.0846767
  3.11954   -1.80979   -1.03182      -0.312217  -0.178004    0.0604223

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -1.6845984584798337
 -0.2584479040718469
 -0.5717868942349338
 -1.4216315410622404
 -1.726030249322083
 -2.469954384364041
 -2.3415659606621855
  0.11234398236136943
 -3.544181187334436
 -2.5398492165083844
  ⋮
 -3.7928381315259974
 -2.7091597774612617
 -4.535581706139472
 -3.7581508020523615
 -1.9338535171472209
 -1.2831243802034056
  0.1500519365545948
 -2.986004767381813
 -2.3892667944957298

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.914872   1.51443
 -0.246197   1.10717
 -1.12689    0.319427
 -0.411874  -0.608949
  2.52531    1.0281
  4.18575    4.15021
  0.134791   3.06842
 -1.67942    1.39377
 -1.2402     2.3196
  3.04032    2.3701
  ⋮         
 -1.24048    0.0187278
 -2.72715   -0.847731
 -3.09204   -1.68296
  2.85325   -0.817635
  1.9089     1.5756
 -0.357808   1.76759
 -1.34848    0.387798
 -0.217243   0.46021
  2.47286   -0.104542

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.