Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -2.06959   -2.41921   4.52519    7.82276  …  -0.371843   0.69554    1.20239
  2.40081    2.60143  -2.34282   -4.74883      0.39985   -0.3601    -0.729915
 -0.580333   5.02584  -2.27577   -8.67416      0.772493  -0.349795  -1.33325
  1.63054   -2.25564   3.70355   -5.38964     -0.346701   0.56925   -0.828411
 -2.3975    -2.80252   5.24217    9.06223      0.898365  -1.68041   -2.90495
  2.7812     3.01361  -2.71402   -5.50125  …  -0.96603    0.869995   1.76346
 -0.672284   5.82216  -2.63635  -10.0485      -1.86633    0.845097   3.22111
  1.88888   -2.61303   4.29035   -6.2436       0.837622  -1.3753     2.00142
  1.83698    2.14731  -4.01658   -6.94353      1.84474   -3.45062   -5.96514
 -2.13097   -2.30905   2.0795     4.21509     -1.98369    1.78648    3.62116
  0.515108  -4.46097   2.01999    7.69924  …  -3.83239    1.73536    6.61437
 -1.44727    2.00212  -3.28729    4.78388      1.72001   -2.82409    4.1098
  3.88183    4.53761  -8.48769  -14.6728      -0.252383   0.472088   0.816106
 -4.50309   -4.87939   4.39432    8.90717      0.271393  -0.244413  -0.49542
  1.08851   -9.42675   4.26856   16.2697       0.524319  -0.237419  -0.904928
 -3.05832    4.2308   -6.94658   10.1091   …  -0.235318   0.386371  -0.562272
 -2.45514   -2.8699    5.36821    9.28011      0.370807  -0.693601  -1.19904
  2.84807    3.08607  -2.77927   -5.63352     -0.398736   0.359097   0.72788
 -0.688448   5.96214  -2.69974  -10.2901      -0.77034    0.34882    1.32954
  1.9343    -2.67585   4.3935    -6.39372      0.345734  -0.567664   0.826102

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 1.2727529498082948
 2.7567865013573014
 1.595063509316577
 3.248424320633604
 1.7218030837307259
 0.27131781445747344
 1.053441398181662
 1.8188714575068066
 2.5709182021053567
 1.4346775091279602
 ⋮
 3.21693238430478
 3.630093944464997
 4.694249336689703
 2.7804608251984204
 1.9506771523924338
 2.5956239652954975
 2.958774807956696
 3.8573271457683656
 2.692934015167177

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -2.68683   -1.46062
 -1.16147   -1.19591
 -1.05683   -0.544734
 -0.312054  -2.50589
 -2.91967   -2.39344
 -2.10241   -3.02694
 -1.06189   -1.35992
 -1.02235   -1.90374
  0.190874  -1.75499
 -1.06521   -3.40007
  ⋮         
  0.960207   0.704403
 -0.165662   2.76176
  0.651906  -0.883369
 -0.402908   2.16917
 -1.42531    1.75553
  1.28419    0.746964
 -0.685181   2.81013
  1.46478   -1.97312
 -0.762536   2.1274

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.