Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -2.10145     1.15411     2.08992   …   -7.46437    -13.5168    16.7779
 -0.20103    -0.0177185   2.77605        0.114597   -17.9545   -16.3579
 -0.589527    0.670187    0.797503      -4.33452     -5.15795    3.78076
 -0.0315092   2.6061     -1.30228      -16.8553       8.42267  -10.2542
 -7.23735     3.97474     7.19763       -4.70632     -8.5224    10.5786
 -0.692343   -0.0610222   9.56067   …    0.0722537  -11.3204   -10.3137
 -2.03032     2.30811     2.74658       -2.73293     -3.25211    2.38378
 -0.108517    8.97535    -4.48503      -10.6273       5.31052   -6.46532
 -3.18813     1.75092     3.17063        2.76459      5.00624   -6.21407
 -0.304984   -0.0268809   4.21158       -0.0424434    6.64983    6.05848
 -0.894376    1.01675     1.2099    …    1.60538      1.91036   -1.40029
 -0.0478029   3.95374    -1.9757         6.24271     -3.11952    3.79786
  5.12729    -2.8159     -5.09915       -2.99596     -5.42521    6.73413
  0.490489    0.0432311  -6.77324        0.0459955   -7.20635   -6.56552
  1.43837    -1.63518    -1.94581       -1.73974     -2.07024    1.51748
  0.0768787  -6.35857     3.17741   …   -6.76517      3.38059   -4.11571
 -5.40943     2.97085     5.37975        3.09562      5.60569   -6.95814
 -0.51748    -0.04561     7.14596       -0.0475256    7.44608    6.78393
 -1.51753     1.72516     2.05289        1.79761      2.1391    -1.56796
 -0.0811092   6.70847    -3.35226        6.99021     -3.49305    4.25262

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -2.7930937180071607
 -0.6382133331099901
 -2.090993785023745
 -2.163013468760112
 -2.422751785460485
 -1.0191009273470466
  0.01912519338709876
 -0.4305010658166905
 -1.5433456288709688
 -0.47072002263205615
  ⋮
  0.4046054364248945
 -2.2315193160647557
 -1.5909022633244783
 -2.1871612581921216
 -3.6507183505062177
 -1.7973853814967964
 -2.9235699769484924
 -2.9459777094624027
 -3.3376483800490493

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.150568   -1.2392
 -0.0959898  -2.42881
 -0.345757   -1.85889
  1.73919    -1.2237
 -0.478921   -0.098044
 -0.469131   -0.543897
 -1.33223    -1.32161
 -1.30375    -1.75328
 -0.352215    0.830075
 -0.890117   -0.179052
  ⋮          
 -3.11643    -2.57303
 -3.24242    -2.85866
 -1.33976    -0.617187
 -2.19146    -1.43258
 -1.59272    -2.61211
 -3.83735    -3.2348
 -3.58369    -3.12809
 -2.19947    -0.558737
 -2.8432     -2.31042

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.