Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -4.73924    6.2573   -1.90329   …  -12.5399    3.81426     7.23568
 -1.46505    3.50731  -2.01426       -7.02879   4.03665    -1.94572
  2.25539    4.671    -2.57678       -9.36085   5.16397    10.1393 
 -1.00101    3.96747   0.334525      -7.95094  -0.6704     -8.1154 
 -5.80637    7.66624  -2.33185      -10.4284    3.17201     6.01733
 -1.79493    4.29705  -2.46781   …   -5.84527   3.35696    -1.6181 
  2.76323    5.72276  -3.15699       -7.78466   4.29445     8.43203
 -1.22641    4.86081   0.409849      -6.61216  -0.557518   -6.74892
  3.34967   -4.42262   1.34523       13.5494   -4.12132    -7.81818
  1.03548   -2.47895   1.42367        7.59463  -4.36162     2.10236
 -1.59409   -3.30143   1.82125   …   10.1144   -5.57968   -10.9555 
  0.707509  -2.80418  -0.23644        8.59103   0.72437     8.76872
 -5.63877    7.44496  -2.26454      -16.8347    5.12065     9.7139 
 -1.74312    4.17302  -2.39658       -9.43614   5.4192     -2.61213
  2.68347    5.55758  -3.06587      -12.5669    6.93262    13.612  
 -1.19101    4.72051   0.398019  …  -10.6741   -0.900012  -10.8949 
 -9.63572   12.7222   -3.86972      -20.5498    6.25067    11.8576 
 -2.9787     7.13099  -4.09535      -11.5185    6.61511    -3.18858
  4.5856     9.49696  -5.23905      -15.3402    8.46251    16.6159 
 -2.03524    8.06655   0.680148     -13.0297   -1.09863   -13.2992 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.0693267543699707 
  4.837614607531128  
 -1.5635531299916094 
  0.3804345374534193 
  2.1015784133157225 
  0.6603971702639554 
  2.8765964689591175 
 -1.649603313184651  
  0.18368096535912823
  1.1409135084739659 
  ⋮                  
  5.216979065701077  
 -0.32118491215602973
  0.29180234054668336
  2.6846470374941545 
 -0.626194322687149  
  1.7115411071471989 
 -1.630928983743076  
 -0.9262969932194053 
 -0.6404529861418932 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.15967    3.10517 
 -0.107954  -0.373246
  0.377947   3.59113 
  0.840226   2.6769  
  0.89757    2.28899 
 -0.884955   3.84945 
  0.685017   0.011604
 -0.856923   2.85559 
 -0.509479   2.55666 
 -1.02235    1.37246 
  ⋮                  
  0.517756  -1.82919 
 -1.80593    1.45378 
 -0.875304  -0.142969
 -0.649509   0.986066
  2.96943    2.01334 
  1.33638   -0.234272
  3.1535     2.87303 
  3.69143    1.75913 
  2.11671    1.29698 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.