Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.117091  -0.336058  -0.0939191  …    0.634697   0.17738     1.39149
 -0.641169   0.182232   0.0548647      -0.344173  -0.10362    -1.36641
 -0.164168  -1.63843    0.586696        3.09442   -1.10806     0.443909
 -0.429748   0.305643   0.150783       -0.577253  -0.284777   -0.693336
  0.761131   2.18449    0.610506       -0.253539  -0.0708573  -0.555851
  4.16782   -1.18457   -0.356639   …    0.137485   0.0413927   0.545833
  1.06715   10.6504    -3.81373        -1.23611    0.442633   -0.177326
  2.79351   -1.98679   -0.980145        0.230593   0.113759    0.276964
 -0.239887  -0.68849   -0.192414        2.38416    0.666307    5.22695
 -1.31358    0.373343   0.112402       -1.29284   -0.389237   -5.13275
 -0.336334  -3.35669    1.20198    …   11.6238    -4.16231     1.66749
 -0.880434   0.626178   0.308914       -2.16838   -1.06973    -2.60443
  0.969037   2.7812     0.777268       -1.09497   -0.306013   -2.40056
  5.30627   -1.50814   -0.454056        0.59376    0.178763    2.3573
  1.35864   13.5595    -4.85546        -5.33843    1.91161    -0.765822
  3.55656   -2.52948   -1.24787    …    0.995865   0.491292    1.19613
  0.671065   1.926      0.538264       -2.91359   -0.814269   -6.38765
  3.67463   -1.0444    -0.314437        1.57993    0.475671    6.27253
  0.94087    9.39008   -3.36244       -14.205      5.0866     -2.03777
  2.46295   -1.75168   -0.864162        2.64989    1.30728     3.18277

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.22477199768999023
  1.4283650681111273
  1.2072036125502483
  0.0830011605043085
  1.2834976571948218
 -3.6751020601931774
 -1.3862198804414703
  0.06339879307343199
 -2.76577508596344
 -1.500095727002012
  ⋮
  2.2619042707031736
  1.4114279710177318
  1.4222285833596955
  2.7075758089378463
 -1.387803652913691
 -0.23503952544555737
  0.49178190461762705
 -1.2513865987338357
 -0.3995594779454419

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.35091    5.19781
 -5.97251    3.12291
 -1.46077    3.10518
 -0.0727027  3.60691
 -0.999188   1.56666
 -2.10011    2.74992
 -6.42298    4.05272
 -3.21138    2.80709
 -3.00225    2.69195
 -2.85555    3.36745
  ⋮          
 -5.14573    2.68504
 -2.98069    2.23943
 -1.71358    2.20555
 -2.04069    3.00891
  0.364062   5.59597
 -6.04303    4.52247
 -2.43217    4.17217
 -0.389337   4.50142
 -0.17312    4.35599

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.