Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1.16761    1.30626      0.0543726    0.124667   …   0.128671     0.295021 
  0.989375   0.0108808   -1.12074     -1.52884       -2.6522      -3.61797  
  2.29041    2.53522      2.22825      5.73391        5.27311     13.5692   
 -2.17995   -3.36207     -2.0382      -3.03003       -4.82335     -7.17049  
 -1.64978   -1.8457      -0.0768262   -0.176149      -0.026537    -0.0608445
 -1.39795   -0.0153741    1.58355      2.16019    …   0.546985     0.746163 
 -3.23626   -3.58216     -3.14843     -8.10177       -1.08752     -2.79848  
  3.08019    4.75047      2.87989      4.2813         0.99476      1.47883  
  0.385017   0.430738     0.0179293    0.0411086      0.0862578    0.197774 
  0.326245   0.00358792  -0.369561    -0.504133      -1.77796     -2.42538  
  0.75526    0.835985     0.734763     1.89075    …   3.53494      9.09638  
 -0.718836  -1.10864     -0.672093    -0.999147      -3.23344     -4.8069   
 -2.10635   -2.35647     -0.0980871   -0.224896      -0.0991694   -0.227378 
 -1.78481   -0.0196287    2.02179      2.758          2.04409      2.78843  
 -4.13186   -4.57349     -4.01973    -10.3439        -4.06408    -10.458    
  3.9326     6.06511      3.67687      5.46611    …   3.71744      5.52642  
  2.62846    2.94058      0.1224       0.280642       0.0108734    0.0249307
  2.22722    0.0244942   -2.52294     -3.44164       -0.224124    -0.305736 
  5.15605    5.70714      5.01611     12.9078         0.445604     1.14666  
 -4.90739   -7.5685      -4.58828     -6.82102       -0.407597    -0.605942 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 2.3888802543755885 
 1.2235386449851866 
 3.1612415815392265 
 5.430400930337248  
 2.293324788430263  
 1.6171865488800838 
 0.1593067410342659 
 1.289241650566329  
 4.09348656389828   
 0.18243059790700505
 ⋮                  
 3.0562745390326196 
 3.033392285964365  
 5.852612758922352  
 2.828544228834731  
 3.29140659575109   
 1.3906277811068541 
 2.624831434740144  
 4.924720920224827  
 1.1474280524496014 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -3.26572     3.19003  
  0.0982949   5.1866   
 -0.452321   -0.165572 
  0.507088    1.28567  
 -2.40821     1.13467  
  0.263199   -0.0524052
  1.19902     0.765216 
  1.27385     0.0505747
  0.891338    0.275587 
  0.0209582   0.0805058
  ⋮                    
  2.62729     5.70427  
  2.26038     0.793574 
  2.23236     1.90083  
 -1.61907     1.6769   
 -4.55537     0.848029 
  0.0125541   2.57054  
 -1.1963     -0.338672 
  0.203189    0.662615 
 -4.04912     0.246847 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.