Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.0528612    -0.0774866    -0.0154106    …   2.53561    0.504285  -0.955431
 -0.0392039    -0.0505843    -0.0134761        1.65528    0.44098   -0.835501
 -0.00217799    0.0574599    -0.00966873      -1.88027    0.316391   0.259178
  0.0664947     0.00898034    0.0326245       -0.293865  -1.06758    1.14682
  5.20789       7.63398       1.51826         -1.6488    -0.327915   0.621276
  3.86237       4.98356       1.32766      …  -1.07636   -0.286751   0.54329
  0.214575     -5.66094       0.952563         1.22266   -0.205736  -0.168532
 -6.55106      -0.884743     -3.21416          0.191088   0.694199  -0.745731
  7.0828       10.3823        2.06485         -5.99137   -1.19157    2.25758
  5.25288       6.77771       1.80564         -3.91124   -1.04199    1.9742
  0.291826     -7.69897       1.2955       …   4.44288   -0.747599  -0.612409
 -8.90953      -1.20326      -4.37131          0.694372   2.52257   -2.70983
  0.000810072   0.00118744    0.00023616       1.16651    0.231996  -0.439546
  0.000600781   0.000775179   0.000206514      0.76151    0.202873  -0.384372
  3.33766e-5   -0.000880544   0.000148168     -0.865018   0.145556   0.119235
 -0.001019     -0.000137619  -0.000499954  …  -0.135193  -0.491138   0.527597
  1.89473       2.77739       0.552371         6.11994    1.21714   -2.30603
  1.40521       1.81312       0.483031         3.99518    1.06435   -2.01656
  0.0780668    -2.05956       0.346561        -4.53822    0.763642   0.625551
 -2.3834       -0.321887     -1.16938         -0.709273  -2.5767     2.76798

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -1.766976649113092
  1.9509980302281356
 -0.6535106047833009
  1.245065171432769
  1.8862303247934906
  0.9440957625748694
  1.393086523239269
  1.226579350057631
  1.4086101762566305
  1.6821346700726247
  ⋮
  1.8794306231585265
 -1.3254325550957793
  0.7687758296456293
  2.145410222511767
 -2.2698524261350843
  1.9656537332317938
 -1.8994551056176563
 -0.0034072900155329377
  0.8150908695262411

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 4.48526  -2.44319
 2.88041  -1.03897
 3.18084  -1.889
 4.58701   0.0813265
 5.16826  -0.716213
 2.44639  -4.65566
 2.22774  -2.6908
 1.82416  -2.59458
 4.13575  -3.95826
 4.9647   -2.80651
 ⋮        
 3.36282  -1.9201
 3.22723  -2.23192
 4.65219  -0.934258
 5.58573  -1.38059
 6.97194  -3.57456
 3.96553  -2.18169
 4.69394  -2.56678
 7.22607  -0.0912381
 7.092    -1.16298

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.