Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.620584    -0.177697    -0.19407     …  -0.285033   -0.311296     2.87407
  0.107753     0.693244    -2.51665         1.11199    -4.03681     12.471
  0.137651     1.53468     -3.40946         2.4617     -5.46891      9.78475
  2.82034     -1.21421      0.097498       -1.94765     0.156391    -2.67822
 -0.243917    -0.0698427   -0.0762779       0.0857786   0.0936821   -0.864929
  0.0423516    0.272475    -0.989153    …  -0.334645    1.21485     -3.75305
  0.0541028    0.603198    -1.34007        -0.740828    1.64583     -2.94464
  1.10852     -0.477238     0.038321        0.586129   -0.0470646    0.805988
 -0.00719498  -0.0020602   -0.00225002      0.0866718   0.0946576   -0.873936
  0.00124928   0.00803739  -0.0291777      -0.33813     1.2275      -3.79213
  0.00159591   0.0177929   -0.0395289   …  -0.748542    1.66296     -2.97531
  0.0326988   -0.0140774    0.00113038      0.592232   -0.0475547    0.814381
  0.608125     0.17413      0.190174        0.529735    0.578544    -5.34147
 -0.10559     -0.679327     2.46612        -2.06664     7.50241    -23.1774
 -0.134887    -1.50387      3.34101        -4.57507    10.164      -18.185
 -2.76372      1.18984     -0.0955407   …   3.6197     -0.290653     4.97747
  0.261591     0.0749034    0.0818049       0.348996    0.381152    -3.51902
 -0.0454204   -0.292219     1.06083        -1.36153     4.94268    -15.2695
 -0.058023    -0.646905     1.43717        -3.01411     6.69615    -11.9805
 -1.18884      0.511819    -0.0410977       2.3847     -0.191486     3.27922

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.4158184179264054
  1.7001788008034668
 -2.7537280464902474
 -6.100285157058824
  0.5333175378096926
 -1.0950369810118883
  0.2646064767358502
 -4.401909169918004
 -5.5170346824417615
  0.11485845126396568
  ⋮
  1.4992670545240643
 -2.6377098780971653
 -4.082087658508274
  1.0173205741707052
  1.0605098115711051
  1.6310079623988596
 -1.51684751660229
 -3.512356996217878
  0.5680519270766307

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 3.96639    0.45848
 3.47057    0.101862
 3.09193    0.902826
 1.25775    0.613282
 2.73738   -2.4395
 3.28029   -1.58239
 2.128     -2.46077
 2.38167   -1.6317
 0.67498   -1.85934
 2.49413   -2.91634
 ⋮         
 2.17628   -5.40287
 1.46974   -3.68284
 0.556236  -6.32675
 1.92717   -5.07402
 3.30631    0.948525
 2.14847    0.0344951
 2.45993    0.800893
 1.65157    0.700303
 2.99314   -1.44339

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.