Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.0447024   -0.339836  -0.0401076  …    5.13674    0.60624    -3.95011
 -0.0308059    0.195005  -0.511569       -2.94757    7.73255     5.56725
  0.222842    -0.436203   0.170727        6.59336   -2.5806     -1.04631
 -0.372787     0.718295   0.100016      -10.8573    -1.51177     2.00299
 -0.277552     2.11001    0.249024       -3.43396   -0.405277    2.64068
  0.19127     -1.21076    3.17628    …    1.97047   -5.16927    -3.72175
 -1.3836       2.70834   -1.06003        -4.40772    1.72515     0.699465
  2.31459     -4.45982   -0.620987        7.25819    1.01063    -1.33901
 -1.07654      8.18407    0.965887       -0.348126  -0.0410859   0.267706
  0.741878    -4.69619   12.3198          0.199762  -0.524048   -0.377302
 -5.36657     10.5048    -4.11152    …   -0.446843   0.174892    0.07091
  8.97761    -17.2983    -2.40862         0.735816   0.102456   -0.135746
 -0.43709      3.32284    0.392163        7.43098    0.877007   -5.71436
  0.301213    -1.90672    5.00201        -4.26405   11.1862      8.05377
 -2.1789       4.2651    -1.66933         9.53817   -3.73318    -1.51362
  3.64503     -7.02333   -0.977933   …  -15.7065    -2.18698     2.89759
  0.224883    -1.7096    -0.201768       -2.5881    -0.305449    1.99023
 -0.154974     0.981005  -2.57354         1.48511   -3.89598    -2.80501
  1.12105     -2.19439    0.858872       -3.32201    1.30022     0.527173
 -1.87537      3.61351    0.503147        5.47035    0.761695   -1.00919

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 0.9416724151765857
 2.262095752080416
 2.207881817982319
 0.525667513273906
 1.3712158166985404
 0.4858270426659605
 1.7736416444173653
 2.2523040798455245
 0.2537313441212176
 1.0853251398267323
 ⋮
 4.22893525357895
 5.6264158203383365
 2.0048144062115516
 1.1759535434679256
 1.229083333498832
 2.6785350473741056
 2.067373958411385
 1.235876177221237
 1.6007523070426508

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 1.31629  -1.87628
 3.18673  -3.45597
 3.52909  -3.37774
 2.10258  -1.74637
 3.21231  -1.68588
 5.2171   -5.16138
 7.13788  -9.3005
 7.43668  -8.76161
 4.85784  -5.4136
 6.73343  -5.9803
 ⋮        
 5.94573  -5.56965
 6.02009  -5.19654
 5.07848  -2.94709
 7.39478  -3.1329
 6.1099   -3.15128
 7.43344  -9.03455
 7.79849  -7.79678
 5.98188  -5.149
 9.10739  -6.39729

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.