Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.48473   -4.24246   -0.0928026   …   2.03975     0.0446189  -2.34314
  5.18749    5.87006    4.98374        -2.82229    -2.39615    -0.929625
  5.32595   -4.29712   10.315           2.06603    -4.95939    -2.47298
  3.65275    6.30322   -1.98935        -3.03055     0.956467    0.69936
  0.124955  -0.118205  -0.00258569      2.07878     0.0454727  -2.38798
  0.144535   0.163553   0.138858    …  -2.87629    -2.442      -0.947414
  0.148393  -0.119727   0.287399        2.10556    -5.05429    -2.5203
  0.101774   0.175622  -0.0554278      -3.08854     0.974769    0.712743
 -1.1244     1.06366    0.0232672       0.690349    0.0151012  -0.793032
 -1.30059   -1.47172   -1.24951        -0.955198   -0.810973   -0.31463
 -1.33531    1.07736   -2.58615     …   0.699243   -1.6785     -0.836975
 -0.915806  -1.58032    0.498764       -1.02568     0.323715    0.236697
  3.62269   -3.42698   -0.0749642       0.596477    0.0130478  -0.685197
  4.19036    4.74173    4.02577        -0.825313   -0.700699   -0.271847
  4.3022    -3.47113    8.33228         0.604162   -1.45026    -0.723165
  2.95062    5.09162   -1.60696     …  -0.886213    0.279697    0.204512
  2.90438   -2.74748   -0.0601003       6.27516     0.137267   -7.20853
  3.35949    3.80154    3.22754        -8.68259    -7.37161    -2.85994
  3.44916   -2.78288    6.68015         6.356     -15.2573     -7.60796
  2.36557    4.08205   -1.28833        -9.32329     2.94251     2.15154

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.30787772764009474
 -0.5802760903357944
 -0.7418049573863217
 -0.32912576243779046
 -1.1808441505694967
 -0.6958515258763985
 -1.4599708348618334
 -0.02801975039022141
 -0.6884106013398186
 -1.2085692843555191
  ⋮
 -2.460672859502844
 -0.549522399083895
 -1.436070920348797
 -2.2236347022480443
  1.2767977243075366
  1.4798798890006508
  1.5907546100627465
  0.6355337218422429
  1.6074429825300096

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 3.91974   -3.38438
 3.54771   -1.70892
 4.7591    -2.67973
 1.35613   -2.28249
 3.95847   -2.05798
 6.63835   -4.50454
 5.3628    -2.68265
 7.03751   -4.33718
 2.44612   -2.67003
 6.6481    -2.47385
 ⋮         
 3.04797   -1.85272
 4.12456   -1.98324
 0.918501  -0.934659
 3.33101   -1.80941
 2.9943    -3.00732
 2.65014   -1.81238
 3.16876   -1.83534
 1.06038   -2.27519
 3.14637   -2.11979

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.