Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  1.86198      3.74713       1.33242      …  -5.82755   -2.07218   11.778
 -4.62444     -5.92075      -0.76809          9.20796    1.19453   -2.42567
  1.02114      4.33475       0.107848        -6.74141   -0.167726   3.87178
 -1.45858     -0.347367     -0.952778         0.540226   1.48176   -2.33466
  0.0423732    0.0852736     0.0303219       -3.57874   -1.27254    7.23297
 -0.105239    -0.134739     -0.0174794    …   5.65468    0.733573  -1.48962
  0.023238     0.0986459     0.00245431      -4.13995   -0.103002   2.37769
 -0.033193    -0.00790504   -0.0216824        0.331757   0.909962  -1.43373
  0.00412502   0.00830138    0.00295183      -3.63842   -1.29376    7.35358
 -0.010245    -0.0131168    -0.00170162       5.74897    0.745805  -1.51446
  0.00226222   0.00960317    0.000238927  …  -4.20898   -0.104719   2.41734
 -0.00323134  -0.000769555  -0.00211078       0.337289   0.925135  -1.45764
 -1.428       -2.87378      -1.02187         -4.16603   -1.48137    8.41993
  3.54661      4.54078       0.589069         6.58263    0.853955  -1.73408
 -0.783136    -3.32443      -0.0827118       -4.81933   -0.119905   2.76787
  1.11863      0.266405      0.730711     …   0.386199   1.05929   -1.66901
  0.19923      0.400939      0.142567        -7.33122   -2.60686   14.8171
 -0.49481     -0.633513     -0.0821847       11.5839     1.50276   -3.05156
  0.10926      0.463813      0.0115396       -8.48088   -0.211004   4.8708
 -0.156067    -0.0371679    -0.101946         0.679619   1.8641    -2.93706

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  1.90931304979435
  0.4359473169405017
 -0.2731520035021953
  0.815934848595326
  1.799021644363672
  2.6913109092752534
  0.7222716546203719
  1.3104734128968694
  0.4056482280487875
  1.545906445524692
  ⋮
  1.3257032635236712
  1.7980979910802357
  0.05390766029065936
  1.8531295297092902
  2.3664282927762548
  0.1114217911243153
  0.6794477698071683
  0.024048931668027386
  0.4938366612950031

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -2.34207    1.52469
 -2.02188    3.05419
 -2.96268    2.00044
 -1.46985    0.864823
 -2.18373   -0.0479105
 -1.61389    1.33282
 -1.44142    2.52863
 -2.40646    1.61848
 -0.933837   0.25209
 -2.25168   -0.908151
  ⋮         
 -2.444      3.9125
 -3.12858    2.83522
 -0.865837   0.0464207
 -2.79192    0.452583
 -0.891385   0.897604
 -0.806448   1.83118
 -1.44673    1.08604
 -0.403699   0.191052
 -1.47887   -0.580475

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.