Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -2.02051     0.285258    -2.05922     1.56591    …   2.94895     -2.24249   
  0.19587    -0.00974904   0.223146    0.0677022     -0.319561    -0.0969542 
 -1.12007     0.231408    -1.21602     0.291557       1.74143     -0.417529  
 -3.01362     0.717216    -3.25213     1.48986        4.65728     -2.13358   
  4.55578    -0.64319      4.64307    -3.53077       -2.25571      1.71532   
 -0.441643    0.0219818   -0.503143   -0.152653   …   0.244438     0.0741621 
  2.52551    -0.521771     2.74185    -0.657393      -1.33205      0.319376  
  6.79502    -1.61716      7.33281    -3.35929       -3.56244      1.63202   
 -2.89327     0.408475    -2.94871     2.24231        3.20785     -2.43937   
  0.280477   -0.0139601    0.319534    0.0969462     -0.347616    -0.105466  
 -1.60389     0.331364    -1.74129     0.417495   …   1.89432     -0.454186  
 -4.31536     1.02702     -4.65689     2.13341        5.06616     -2.3209    
  0.554484   -0.0782826    0.565108   -0.429729      -5.82826      4.43202   
 -0.0537523   0.00267541  -0.0612376  -0.0185794      0.631575     0.191619  
  0.307379   -0.0635047    0.333711   -0.0800112     -3.44173      0.825198  
  0.827022   -0.196824     0.892476   -0.408859   …  -9.20457      4.21678   
 -1.81198     0.255818    -1.8467      1.4043         0.0559523   -0.0425482 
  0.175656   -0.00874289   0.200117    0.060715      -0.00606323  -0.00183957
 -1.00448     0.207525    -1.09052     0.261467       0.0330413   -0.00792205
 -2.7026      0.643196    -2.9165      1.3361         0.0883655   -0.0404818 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 2.4104829187221153
 1.7353699032765264
 1.6025983504731325
 4.66523566531526  
 2.753324203968247 
 2.792638558000614 
 4.245449777797649 
 1.9764551425929178
 5.211924031113675 
 1.995067689424603 
 ⋮                 
 3.7299447184226557
 1.563536069798958 
 3.8064888192442843
 3.084317077291842 
 3.641734347753203 
 4.4189238830517   
 2.9860804574326325
 6.661097601495388 
 3.626072754028025 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 -0.367987   2.18154 
 -0.15483    2.52016 
  0.206867   2.24139 
  0.249927   1.5249  
  0.661399   4.79688 
 -1.1706     3.41403 
 -1.58768    3.54684 
  0.791568   3.6664  
  0.954168   2.47304 
  0.903579   4.02899 
  ⋮                  
 -1.38814   -1.6208  
 -1.68181   -0.540336
 -1.14158   -0.244436
 -1.03515    0.600151
 -2.31571    0.864453
 -3.13278    0.190456
 -2.38225    1.01113 
 -1.55333    0.801451
 -1.98609   -0.142672

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.