Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.7902     -1.60675     0.435886    0.235907  …   -0.651993    -0.352867 
 -0.459979   -3.03399    -1.94184     3.0405         2.90458     -4.54794  
  0.339439   -0.767591   -0.137144   -1.00472        0.205138     1.50285  
 -0.975191   -1.92391    -3.17405     7.76022        4.7477     -11.6076   
 -1.27488    -2.59227     0.703242    0.380603       1.67809      0.908201 
 -0.742113   -4.89492    -3.13289     4.90543   …   -7.47575     11.7054   
  0.547639   -1.2384     -0.221263   -1.62098       -0.527981    -3.868    
 -1.57334    -3.10396    -5.12089    12.52         -12.2195      29.8755   
 -0.75512    -1.53542     0.416535    0.225434      -0.104364    -0.0564832
 -0.439559   -2.8993     -1.85563     2.90552        0.464934    -0.727987 
  0.32437    -0.733514   -0.131055   -0.960116  …    0.0328364    0.24056  
 -0.931898   -1.8385     -3.03314     7.41571        0.759962    -1.85803  
 -3.60981    -7.34        1.99123     1.07768       -0.224179    -0.121328 
 -2.10129   -13.8599     -8.87076    13.8897         0.998699    -1.56375  
  1.55064    -3.50653    -0.626504   -4.58978        0.0705339    0.516733 
 -4.45489    -8.78883   -14.4998     35.4504    …    1.63243     -3.99112  
  1.88832     3.83962    -1.04163    -0.563742      -0.948483    -0.513331 
  1.0992      7.25026     4.64037    -7.26582        4.22542     -6.61609  
 -0.811152    1.8343      0.32773     2.40096        0.298424     2.18626  
  2.33039     4.59752     7.58496   -18.5444         6.90669    -16.8861   

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -2.2548321467704038
 -2.9069002623479467
 -2.6214198035291876
 -3.0738888417534724
 -2.4372702193726146
 -3.3572599581106264
 -3.590289300392384 
 -3.9972417463302796
 -2.8448108901640987
 -2.4031851139949776
  ⋮                 
 -3.9504112067656725
 -3.853478797643875 
 -2.709377785721527 
 -3.124095585137189 
 -0.7883493355100029
 -2.034003834483565 
 -1.5097049215313896
 -2.6303687484686953
 -2.0829421549123377

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.700432  -1.89856  
  2.0852     1.24316  
  2.27518   -4.22667  
  2.91916   -2.05122  
 -1.15039    1.11862  
 -1.23096   -2.36146  
 -0.237295  -0.387737 
 -1.90507   -2.55622  
  0.229482  -0.0269712
 -1.87386    2.3174   
  ⋮                   
 -1.06098    0.475929 
 -2.83468   -3.7844   
 -0.818877  -1.03301  
 -2.98042    2.01858  
  0.666879  -2.53772  
  1.86573    0.300815 
  1.28156   -2.72166  
  2.19738    0.119311 
 -0.921603   3.51545  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.