Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.664725    5.26512     1.63294    …  -5.51139   -1.70931     0.48786
 10.5988      3.43568    -6.20223       -3.59637    6.49232     8.15064
 -0.700541   -3.26419     1.5609         3.41687   -1.6339     -1.73118
  1.83907     3.00906     0.820084      -3.14981   -0.858442    0.911412
  1.63818    12.9756      4.02429       -4.59412   -1.42483     0.406665
 26.1202      8.46705   -15.2851     …  -2.99782    5.4118      6.79412
 -1.72645    -8.04443     3.84675        2.84819   -1.36197    -1.44306
  4.5323      7.41568     2.02106       -2.62558   -0.715571    0.759725
 -0.615929   -4.87862    -1.51307        4.01482    1.24516    -0.355386
 -9.82077    -3.18347     5.74694        2.61981   -4.72939    -5.9374
  0.649115    3.02457    -1.44631    …  -2.48904    1.19023     1.26109
 -1.70407    -2.78817    -0.759884       2.2945     0.625339   -0.663926
  0.868653    6.88039     2.1339        -5.07695   -1.57458     0.449404
 13.8504      4.48969    -8.10499       -3.31289    5.98056     7.50816
 -0.915457   -4.2656      2.03976        3.14753   -1.50511    -1.59472
  2.40328     3.9322      1.07168    …  -2.90152   -0.790775    0.83957
 -0.026758   -0.211944   -0.0657327     11.7019     3.62927    -1.03584
 -0.426647   -0.138301    0.249666       7.63593  -13.7847    -17.3057
  0.0281997   0.131398   -0.0628327     -7.25479    3.46915     3.67569
 -0.0740305  -0.121128   -0.0330119      6.68776    1.82267    -1.93514

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -3.377613574208314
 -3.292850075625937
 -4.7242245222277255
 -2.352575355319686
 -3.252577010668371
 -0.845853922016982
 -1.4607786969656904
 -1.0303646668050748
 -0.8131599461080278
  0.544195672610797
  ⋮
 -2.0454189204505737
 -3.0057119344638545
 -2.611128212529639
 -1.8687460565004432
 -1.1062198989521819
 -1.0646106558943458
 -1.1902540933124357
 -2.0778480543285376
  0.26383727137620244

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -3.2704      1.50492
 -2.09622     1.21631
 -1.18561     0.312018
 -2.91096     0.810631
 -2.444       1.48306
  0.0272299   0.0908083
  0.327056    0.473349
  0.954348   -0.273766
  0.232308   -0.251962
  0.347988   -0.0582928
  ⋮          
 -1.80146     2.06687
 -1.19385     0.517104
 -1.62484    -0.263431
 -2.27435     2.06435
 -3.80932     2.68481
 -4.43635     2.22521
 -2.95828     1.34589
 -2.24156     1.16004
 -4.51551     2.96254

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.