Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.695278    10.0376    -0.434972  …  -2.90545    0.125905   -0.0723655
  5.30439     -2.15156   -4.02142       0.622782   1.16403     1.77173  
 -8.21874     -1.4201    -6.92832       0.411057   2.00545     1.37903  
  0.0050255   -5.32464    4.69817       1.54125   -1.35992    -1.23872  
  0.430971     6.22185   -0.26962       1.29303   -0.0560326   0.0322053
  3.28795     -1.33365   -2.4927    …  -0.277161  -0.518035   -0.788484 
 -5.09443     -0.880255  -4.29456      -0.182935  -0.892498   -0.613719 
  0.00311508  -3.30051    2.91219      -0.685914   0.605214    0.551274 
 -0.642273    -9.27237    0.401812     -3.79499    0.164453   -0.0945212
 -4.90001      1.98753    3.71485       0.813455   1.52041     2.31416  
  7.59219      1.31184    6.40014   …   0.536907   2.61944     1.80124  
 -0.00464238   4.91872   -4.34001       2.01313   -1.77627    -1.61796  
 -0.236268    -3.41096    0.147812     11.9409    -0.51745     0.29741  
 -1.80253      0.731139   1.36655      -2.55953   -4.78395    -7.28149  
  2.79288      0.482576   2.35437      -1.68937   -8.24205    -5.66757  
 -0.00170776   1.80941   -1.59653   …  -6.33429    5.58903     5.09091  
  0.662965     9.57111   -0.414757     -5.32321    0.230678   -0.132584 
  5.05787     -2.05157   -3.83453       1.14103    2.13267     3.24607  
 -7.83679     -1.3541    -6.60634       0.753117   3.67428     2.52659  
  0.00479194  -5.07719    4.47983       2.82381   -2.49157    -2.26951  

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -4.431248592379231 
 -7.736350400656786 
 -3.8187639261331765
 -2.9042176380097193
 -5.5599673763050985
 -1.2220084076201345
 -4.934921518256974 
 -2.3190144284240466
 -2.737775368099024 
 -3.287615782687242 
  ⋮                 
 -5.183551402333079 
 -3.404308881471121 
 -2.9728209795572518
 -4.5832444085604465
 -2.713429872059529 
 -4.876676593084815 
 -1.6589504949293026
 -1.5194891583275487
 -3.7922957764872285

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 0.997658    3.64893  
 1.25129     1.85116  
 2.78549    -0.0810436
 1.80778     1.29811  
 1.94548     2.02049  
 0.730344    2.19086  
 0.960395    0.943463 
 1.81068    -1.25017  
 1.63117     0.898636 
 1.63603     1.05116  
 ⋮                    
 2.60277    -0.696057 
 2.99427    -4.22447  
 1.27471     0.430204 
 1.66538     0.422984 
 0.0899486   1.82165  
 0.436551    1.62509  
 2.05037    -1.88022  
 1.58886     2.25734  
 2.10357     1.03552  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.