Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  4.99646   -22.9302    -16.0425    …   4.59896   3.21754    6.06348
  3.91506    11.7356    -11.0524       -2.35374   2.21672   -4.65118
  6.08127   -14.4126    -19.0008        2.89065   3.81086    4.34348
  5.59582   -20.2506    -24.3307        4.06153   4.87985    5.37981
  2.66793   -12.2438     -8.5661       -5.12639  -3.58655   -6.75887
  2.0905      6.26639    -5.90158   …   2.62368  -2.47094    5.1846 
  3.24717    -7.69581   -10.1457       -3.22217  -4.24792   -4.84161
  2.98796   -10.8131    -12.9917       -4.52733  -5.4395    -5.9968 
  1.19295    -5.47478    -3.83029       7.92916   5.54744   10.4542 
  0.934756    2.80198    -2.63886      -4.05814   3.82189   -8.01919
  1.45196    -3.44114    -4.5366    …   4.98384   6.57039    7.48868
  1.33605    -4.83501    -5.80916       7.00258   8.41345    9.27545
 -0.308492    1.41575     0.990497      2.89991   2.02885    3.82337
 -0.241724   -0.724581    0.682399     -1.48417   1.39777   -2.93284
 -0.37547     0.889865    1.17315       1.82272   2.40297    2.73881
 -0.345497    1.25031     1.50222   …   2.56103   3.07703    3.39228
 -2.23345    10.2499      7.17111       9.95729   6.96637   13.1282 
 -1.75006    -5.2459      4.94051      -5.09613   4.79946  -10.0704 
 -2.71837     6.44254     8.49347       6.25861   8.25098    9.40415
 -2.50137     9.05215    10.876         8.7937   10.5655    11.6479 

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -2.7043045056455464
 -2.5926991265555674
 -1.0121404830407215
 -2.5377105862938096
 -4.3710125416925765
 -3.9100633521225645
 -4.266069081769226 
 -2.4513481938222124
 -4.538179251566596 
 -6.4979903813143265
  ⋮                 
  2.116757814789008 
  2.0888405445165783
  2.3208506338462556
  2.960661524168731 
 -1.559606763375851 
 -2.05333402161854  
 -1.2327710952282946
 -3.491309836025582 
 -2.0879651568163395

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 3.52479   -0.0901938
 2.9055    -1.80744  
 4.26447   -0.751981 
 1.41203   -1.53698  
 0.281097  -1.65172  
 3.29184    0.396768 
 3.09103   -1.19877  
 5.08174    0.0742486
 3.73538   -0.520952 
 2.29179   -1.01393  
 ⋮                   
 2.82393   -2.10349  
 4.34178   -1.3374   
 3.58176   -1.41532  
 2.19413   -1.89999  
 4.10619   -3.68736  
 4.63562   -4.97025  
 6.05396   -5.24286  
 3.4032    -5.20905  
 2.32786   -4.18872  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.