Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -3.9283     -5.73463    2.69639   …   12.2872     -5.77739   -7.78798
 -0.948563    0.844614  -5.34047       -1.80971    11.4427     8.65642
 -1.95022     4.59224   -4.29804       -9.83952     9.20915   29.3173 
 -6.05809    -5.89359   -2.19573       12.6278      4.70466   -5.76825
  1.1045      1.61238   -0.758132       2.94774    -1.38601   -1.86836
  0.266703   -0.237476   1.50156   …   -0.434153    2.74514    2.0767 
  0.548333   -1.29118    1.20846       -2.36053     2.2093     7.03329
  1.70333     1.65707    0.617364       3.02945     1.12866   -1.38382
 -1.85871    -2.71338    1.27582        4.80171    -2.25774   -3.04345
 -0.44882     0.399636  -2.52689       -0.707211    4.47168    3.38282
 -0.92276     2.17286   -2.03365   …   -3.84517     3.59882   11.4568 
 -2.86643    -2.7886    -1.03893        4.93481     1.83853   -2.25416
 -2.27692    -3.3239     1.56288       11.9036     -5.59699   -7.54479
 -0.549806    0.489555  -3.09545       -1.75319    11.0854     8.38611
 -1.13038     2.66175   -2.49123       -9.53227     8.92158   28.4018 
 -3.51139    -3.41604   -1.27269   …   12.2335      4.55775   -5.58813
  8.81596    12.8697    -6.05128      -19.8087      9.31394   12.5553 
  2.12878    -1.8955    11.9852         2.91749   -18.4472   -13.9553 
  4.37671   -10.306      9.64572       15.8626    -14.8464   -47.2634 
 13.5957     13.2265     4.92769      -20.3578     -7.58455    9.29919

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.011550756106361786
 -2.2150584807327713  
 -3.1713688417036865  
 -0.8654324768078996  
 -1.785118023754348   
 -2.7577663507710595  
 -6.210956295311244   
 -7.1056312966576956  
 -5.488192797429015   
 -5.487099635397862   
  ⋮                   
 -2.029075734255949   
 -3.386036817461553   
 -1.6589124395177077  
 -1.1746639227195377  
 -0.5221461783522936  
 -2.5186024815026435  
 -3.3607336832134544  
 -3.272659520240698   
 -1.8126257876996443  

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.742089    1.78798 
  0.255996    2.9931  
 -0.202715    1.07628 
  1.10204     3.1793  
 -0.379003    1.64894 
 -0.657494    1.38105 
 -0.019033    4.79642 
  0.00497355  2.36787 
 -0.794195    3.23453 
 -0.678338    3.07748 
  ⋮                   
  0.883762    0.318679
  0.533217    0.215495
  1.02227     0.25223 
  0.165124    1.02925 
  0.656608    1.40093 
  0.81928     2.42288 
  0.139059    0.952242
  0.157252    2.59965 
 -0.163868    1.76314 

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.