Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.893111   -9.34184  -2.54951   …   12.9698     3.53963     1.04846
  4.44754     4.95479  -1.28035       -6.87903    1.77759     3.59941
  0.0379553   5.26937   4.56074       -7.31577   -6.33195    -8.94293
  1.03545    -3.71972  -1.76524        5.1643     2.45078     4.57274
  0.75232    -7.86918  -2.1476        -0.434393  -0.118552   -0.0351158
  3.74643     4.17372  -1.07852   …    0.230397  -0.0595361  -0.120554
  0.031972    4.4387    3.84179        0.245024   0.212074    0.299522
  0.872221   -3.13334  -1.48696       -0.172966  -0.082083   -0.153153
 -0.836846    8.75331   2.38889       -1.22791   -0.335112   -0.0992624
 -4.16735    -4.64265   1.19969        0.651267  -0.168292   -0.340771
 -0.0355642  -4.93741  -4.27342   …    0.692615   0.599472    0.846666
 -0.970218    3.48538   1.65403       -0.488926  -0.232026   -0.432921
  0.590232   -6.17376  -1.6849       -12.039     -3.28559    -0.973215
  2.93926     3.27449  -0.846149       6.38532   -1.65001    -3.34108
  0.0250836   3.48238   3.01407        6.79072    5.8775      8.3011
  0.684301   -2.45826  -1.1666    …   -4.79366   -2.27489    -4.24456
 -0.687824    7.19456   1.96349        3.03044    0.827047    0.244977
 -3.42525    -3.81591   0.986056      -1.60731    0.415339    0.841014
 -0.0292311  -4.05817  -3.51243       -1.70935   -1.47948    -2.08955
 -0.797446    2.86472   1.35949        1.20666    0.572633    1.06844

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.6512957045875211
 -3.2369042889220214
 -2.094090177162031
 -0.42776918826453214
 -2.703395641369087
  0.6606031614643572
 -0.17546370119594035
  0.22191321535000294
  0.35048368322691903
  0.07931628065707935
  ⋮
 -1.4167240445804392
 -0.34592924543836817
 -0.4626325162510308
 -1.0120356556827184
  0.816301077105639
 -0.9615048841261219
 -0.001239155343391786
  0.32895797488932094
 -0.7338324484584222

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  0.411409   2.13755
 -1.28674    4.73048
 -0.811256   0.392895
  0.787902   3.51723
  0.933399   0.641751
 -1.21028    0.678815
 -2.74626    2.46452
 -2.89195   -0.468505
 -0.843628   1.26603
 -1.44456   -0.629847
  ⋮         
 -1.93497    0.11954
 -2.12601   -0.850163
 -1.26204   -0.546615
 -1.81738   -1.04148
 -0.3977     0.255639
 -2.69995    0.825374
 -0.727789  -0.160521
 -1.25757   -0.29032
  0.421988  -0.596135

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.