Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.606409     2.39614    -2.09715    …  -1.27981    1.12012     1.29956 
  1.085       -0.818152    1.56524        0.436987  -0.836015   -1.08747 
  2.84216     -3.86316     0.744969       2.06337   -0.397899   -1.0501  
 -0.823188     1.10518    -0.335378      -0.590293   0.17913     0.38089 
  1.24611      4.92385    -4.30945       -7.20125    6.30267     7.31238 
  2.22957     -1.68123     3.21641    …   2.45883   -4.70408    -6.11894 
  5.84037     -7.93843     1.53084       11.6101    -2.23889    -5.90869 
 -1.69158      2.27104    -0.689171      -3.32145    1.00793     2.14319 
 -0.877146    -3.46592     3.03344        0.692307  -0.60592    -0.702991
 -1.56941      1.18342    -2.26405       -0.236385   0.452237    0.588257
 -4.11107      5.5879     -1.07757    …  -1.11617    0.215241    0.568044
  1.19071     -1.5986      0.485111       0.319315  -0.0968994  -0.20604 
  1.7993       7.10969    -6.22253       -1.37527    1.20367     1.3965  
  3.21934     -2.42757     4.64427        0.469581  -0.898373   -1.16858 
  8.43308    -11.4625      2.21043        2.21727   -0.427578   -1.12842 
 -2.44251      3.27922    -0.995114   …  -0.634322   0.192492    0.4093  
 -0.0548854   -0.216872    0.189811      -1.00631    0.880741    1.02184 
 -0.0982021    0.0740501  -0.141668       0.3436    -0.657354   -0.855067
 -0.257241     0.34965    -0.0674263      1.62241   -0.312865   -0.825686
  0.0745059   -0.100029    0.0303547     -0.464143   0.140849    0.299491

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -1.1864021044312318 
  0.9284819154225573 
  0.13339208320809257
 -3.6871307490168763 
  1.4290659085013344 
 -3.8682312116423385 
 -2.1866779503263705 
 -1.94039495190321   
 -5.846243392816956  
 -1.3063206359149786 
  ⋮                  
  1.7994364873131579 
  0.6098089785705338 
 -2.356172004135997  
  2.189265590524323  
 -4.538469195992417  
 -2.1521539644297794 
 -2.0779461163999233 
 -5.575534054745836  
 -1.4428486111863843 

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -2.51287   4.54447
  0.192802  4.99234
 -1.62503   3.38729
  0.927137  2.6903 
 -0.88901   4.32899
 -3.53814   1.51528
 -1.32065   2.70906
 -2.85521   1.28818
  0.587104  1.21541
 -1.97335   1.68116
  ⋮                
 -2.6188    3.22813
 -2.02983   1.62952
  0.339364  1.08893
 -2.83672   2.28449
 -3.79581   3.0123 
 -1.4768    3.98184
 -2.03673   3.12722
  0.923328  2.24454
 -1.91775   3.38761

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.