Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
   1.24609    -0.25598    -0.191649   …  -0.402833  -0.301597    4.93435
  -5.76549    -6.21314     7.71935       -9.77758   12.1479     -7.47822
  -5.75327    -3.84462     5.09723       -6.05024    8.02147     6.76863
   1.62579     1.76967     1.43883        2.78492    2.26427   -18.1466
   0.483595   -0.0993433  -0.0743772      0.168697   0.126301   -2.06639
  -2.23753    -2.41126     2.99581    …   4.09461   -5.08723     3.13169
  -2.23279    -1.49206     1.97819        2.53369   -3.35919    -2.83454
   0.630954    0.686792    0.558395      -1.16625   -0.948221    7.59933
   0.693333   -0.142429   -0.106635      -0.211691  -0.158491    2.59303
  -3.20796    -3.45704     4.29511       -5.13817    6.38378    -3.92984
  -3.20116    -2.13917     2.83614    …  -3.17944    4.21532     3.55695
   0.904603    0.984658    0.800574       1.46349    1.18989    -9.53611
   2.17864    -0.447551   -0.335077       0.219985   0.1647     -2.69462
 -10.0803    -10.863      13.4964         5.33948   -6.63389     4.08381
 -10.0589     -6.72187     8.91192        3.30401   -4.38048    -3.69631
   2.84251     3.09406     2.51562    …  -1.52083   -1.23651     9.90974
  -1.29619     0.266272    0.199355      -0.345677  -0.258805    4.23424
   5.99732     6.46297    -8.02974       -8.39028   10.4243     -6.41717
   5.9846      3.99921    -5.30218       -5.1918     6.88334     5.80826
  -1.69116    -1.84083    -1.49668        2.38978    1.943     -15.5718

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.4229003001098065
 -0.3395975780473628
 -0.20540141859604322
 -0.0977526537982335
  0.5945546985708285
 -0.6411542928761051
  0.7461447079267325
  1.2012807744977458
  0.39440454082084847
  0.5317248821440045
  ⋮
 -1.468442790677161
 -2.218161397402582
 -1.6679207099152615
 -0.21296757178003664
 -0.43274749177690297
 -1.039825834058574
 -2.277198911767897
 -1.924238503419779
 -0.5541048050465359

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -3.86047   -3.01281
 -0.994241  -5.37465
 -4.51707   -5.82719
 -2.05166   -2.76444
 -2.04857   -2.32305
 -2.0502    -3.54936
 -0.199125  -2.91999
 -1.63468   -5.78979
 -0.888345  -2.85374
 -1.08806   -3.89503
  ⋮         
  1.52695   -3.78604
  0.285955  -3.8829
  1.33952   -2.16553
  2.79537   -2.67114
 -1.92669   -3.22486
  0.714926  -3.09057
 -2.01061   -4.56173
 -1.2736    -2.69105
 -1.37737   -3.59699

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.