Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.958449     0.0839658     0.986025    …  -0.113984   -1.33854    -0.91919
  3.12701     -0.881379      2.14835         1.19648    -2.91641     1.03746
  1.47711     -0.423642     -2.7046          0.575097    3.67151     4.4238
 -3.65901      1.47291      -0.288391       -1.99948     0.391493   -3.76274
  0.245839     0.021537      0.252912       -0.083897   -0.985217   -0.676561
  0.802068    -0.226071      0.551046    …   0.880657   -2.14659     0.763613
  0.378874    -0.108663     -0.693721        0.423295    2.70238     3.25609
 -0.938524     0.377796     -0.0739714      -1.4717      0.288155   -2.76953
  0.0693742    0.00607759    0.0713702       0.0716394   0.841273    0.577713
  0.226338    -0.0637957     0.155502       -0.75199     1.83297    -0.652047
  0.106916    -0.0306639    -0.195764    …  -0.36145    -2.30756    -2.78037
 -0.264845     0.106612     -0.0208742       1.25668    -0.246054    2.36489
  0.588819     0.0515841     0.605761       -0.0136774  -0.160616   -0.110297
  1.92106     -0.541472      1.31983         0.14357    -0.34995     0.124489
  0.907456    -0.260263     -1.66156         0.0690079   0.440558    0.530827
 -2.2479       0.904874     -0.177172    …  -0.239925    0.0469766  -0.451505
 -0.00826409  -0.000723983  -0.00850186     -0.0344009  -0.403975   -0.277415
 -0.0269622    0.00759957   -0.0185239       0.361102   -0.880182    0.31311
 -0.0127362    0.00365279    0.02332         0.173566    1.10808     1.33512
  0.0315493   -0.0126999     0.00248661     -0.603451    0.118154   -1.13561

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.3199834880188398
 -1.4896236562232086
 -1.7458330189281595
 -0.6748194812578624
 -1.0206569613718648
 -0.08641849766994242
 -2.0131223649740524
 -2.329000172270585
 -0.7603537511568338
 -0.8910182174158937
  ⋮
 -1.190544385814399
 -2.458543799163629
 -0.29734816665982183
 -0.5601272004778244
 -0.510499722513312
 -1.5262069270316878
 -1.9059682188755522
 -1.0593835496957378
 -1.1951629746449297

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -0.435309     5.62342
 -0.874128     5.53875
  0.652252     0.575836
  0.832286     0.788715
  2.1603       3.84225
  0.443291     4.3878
  0.0903196    5.59065
  0.826896    -0.496646
  0.789257     0.948229
  1.40785      3.7453
  ⋮           
 -2.57382      4.7131
 -0.0242041    1.28607
  0.00822822   1.24451
  0.219676     3.39634
  0.462049     3.56097
 -0.806846     3.99923
  1.77629     -0.0887079
  1.40153      0.169666
  3.06689      3.32845

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.