Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.246858    0.392293    0.359545  …  -2.50848     -2.29908     0.740547
  0.0120946   0.0132961  -0.145004     -0.0850207    0.927214   -0.284635
  0.0906222  -0.243531   -0.30138       1.55723      1.92714    -0.969936
 -0.460537    0.142106    0.553197     -0.908684    -3.53736     0.812687
 -0.269549    0.428354    0.392595     -4.49963     -4.12401     1.32837
  0.0132063   0.0145183  -0.158333  …  -0.152507     1.66321    -0.51057
  0.0989523  -0.265917   -0.329083      2.79332      3.45685    -1.73984
 -0.50287     0.155169    0.604048     -1.62997     -6.3452      1.45777
  0.621365   -0.98744    -0.905011      0.521761     0.478206   -0.154033
 -0.0304432  -0.0334676   0.36499       0.0176842   -0.19286     0.0592039
 -0.228105    0.612991    0.758602  …  -0.323903    -0.400844    0.201746
  1.15922    -0.357695   -1.39245       0.189005     0.735768   -0.169038
  4.09588    -6.50896    -5.96561      -0.141865    -0.130022    0.0418809
 -0.200674   -0.22061     2.40592      -0.00480827   0.0524378  -0.0160973
 -1.50361     4.04069     5.00052       0.088068     0.108988   -0.0548538
  7.64126    -2.35784    -9.17869   …  -0.0513898   -0.200052    0.0459608
  1.68958    -2.685      -2.46086      -3.6727      -3.36611     1.08424
 -0.0827796  -0.0910035   0.992461     -0.12448      1.35755    -0.416738
 -0.620251    1.66681     2.06275       2.27997      2.82155    -1.4201
  3.15208    -0.972627   -3.78628      -1.33042     -5.1791      1.18987

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.518000390009965
  0.7616730787107122
  0.6459845814761314
  0.6063625772211914
  0.5451425756759342
  1.8533504036762782
  2.286442376464571
  2.3515514667077504
  1.4237305392413047
  2.664799908971173
  ⋮
 -0.44610081952065944
 -0.06645905241107115
 -0.2426058598005473
 -0.9881822909358252
 -1.2599056727740372
 -0.041968306728018996
  0.5848505468681513
 -0.846788765927054
 -1.301633225177717

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
  1.55857    2.07105
  0.477798   1.72473
  1.82758    0.107485
  1.61837    3.39273
  1.71234    2.89134
  2.88001    0.936045
  0.898276  -1.45866
  3.76803   -0.331576
  2.86427    1.21535
  1.20312    2.029
  ⋮         
  0.434644   0.0292051
  2.1995     0.251545
  2.04765    2.51631
  1.82695    2.0766
  1.13817    0.0852276
 -0.309349  -0.206735
  1.19921   -0.865335
  1.0955     0.677541
  1.03627    0.432608

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.