Multiplication

Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -3.05572     -1.42521     2.64178     …   11.4805        5.29323  
 -0.879231     2.5203      2.62597         11.4118       -8.40433  
 -0.80926     -5.53074    -1.10833         -4.81651      21.5314   
  1.03941      1.94302    -0.0632371       -0.274813     -3.73015  
  8.35665      3.89759    -7.22463          4.03946       1.86244  
  2.40448     -6.8924     -7.18139     …    4.01529      -2.95709  
  2.21313     15.1252      3.031           -1.69471       7.57589  
 -2.84253     -5.31368     0.172938        -0.0966938    -1.31247  
 -7.31194     -3.41033     6.32144        -15.0006       -6.91618  
 -2.10388      6.03074     6.28361        -14.9108       10.9812   
 -1.93645    -13.2343     -2.65208     …    6.2933      -28.1331   
  2.48717      4.64939    -0.151318         0.359073      4.87386  
  1.56708      0.730893   -1.35479         -0.282589     -0.130291 
  0.450898    -1.29249    -1.34669         -0.280898      0.20687  
  0.415015     2.83635     0.568387         0.118557     -0.529987 
 -0.533044    -0.996445    0.0324301   …    0.00676441    0.0918164
  0.125625     0.0585924  -0.108608        -6.76797      -3.12045  
  0.0361465   -0.103613   -0.107958        -6.72747       4.9545   
  0.0332699    0.227377    0.045565         2.83942     -12.6931   
 -0.0427317   -0.0798805   0.00259977       0.162007      2.19899  

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -2.534416489628819  
 -1.2334380990276526 
 -3.0302149323622043 
 -2.6532195638359375 
 -2.486037787801678  
 -0.08531756899905356
 -0.919832789534674  
 -1.238757640103192  
 -0.7324763263145229 
 -0.731223035782658  
  ⋮                  
 -1.519696447502433  
 -2.355618889930786  
 -1.8080508756881284 
 -0.8319401699372965 
 -1.5263050838440277 
 -2.209855165094204  
 -2.828724714439554  
 -1.2689754396588364 
 -1.301754724423938  

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Transpose{Float64,Array{Float64,2}}:
 -1.47519     0.27368  
 -0.38534    -0.0869836
 -3.23171     0.683696 
 -5.38342     1.33869  
 -4.61449     1.06855  
 -0.736721   -0.333805 
 -0.0557117   0.161205 
 -1.79083     0.318381 
 -3.85206     2.19674  
 -3.18954     0.944983 
  ⋮                    
 -1.03609     0.220877 
 -2.6162      0.967654 
 -3.95358     0.806083 
 -3.63599     1.50307  
 -1.75311     1.98238  
  0.610125    2.75194  
 -2.55675     1.97906  
 -4.27322     4.36479  
 -4.7053      4.05225  

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

LinearAlgebra.mul! — Function.

mul!(x::AbstractVector, K::AbstractKroneckerProduct, v::AbstractVector)

Calculates the vector-matrix multiplication K * v and stores the result in x, overwriting its existing value.

source

LinearAlgebra.lmul! — Function.

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

source

LinearAlgebra.rmul! — Function.

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

source

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.