Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -0.276536     -0.164581   -0.42871   …   2.10006     5.47036   -1.64701
  0.0162282    -0.0446942  -0.299384      0.5703      3.82016   -2.42673
  0.104534     -0.0417674  -0.503723      0.532954    6.42754   -4.61605
  0.000362309   0.160252    0.139119     -2.04483    -1.77517    3.078
  0.310371      0.184717    0.481164      0.32901     0.857025  -0.258031
 -0.0182137     0.0501626   0.336014  …   0.0893472   0.598492  -0.380188
 -0.117324      0.0468777   0.565355      0.0834963   1.00698   -0.723183
 -0.000406638  -0.17986    -0.156141     -0.320357   -0.27811    0.482221
  0.681975      0.405878    1.05726      -0.637428   -1.66041    0.499914
 -0.0400209     0.110222    0.73832      -0.173103   -1.15953    0.736582
 -0.257795      0.103004    1.24225   …  -0.161767   -1.95094    1.40111
 -0.000893502  -0.395203   -0.343086      0.620664    0.538815  -0.934263
 -0.355338     -0.21148    -0.550876      0.321659    0.837878  -0.252266
  0.0208526    -0.0574303  -0.384697      0.087351    0.585121  -0.371694
  0.134322     -0.0536695  -0.647265      0.0816308   0.984485  -0.707026
  0.000465553   0.205918    0.178763  …  -0.3132     -0.271897   0.471448
 -2.11895      -1.26109    -3.28498      -3.09239    -8.05524    2.42525
  0.124348     -0.342468   -2.29402      -0.839781   -5.62527    3.57341
  0.80099      -0.320041   -3.85976      -0.784787   -9.4647     6.79725
  0.00277618    1.22793     1.066         3.01106     2.61398   -4.53243

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  0.9109463567679742
  1.1757098304826048
 -0.4558769495595977
  0.6006655398409986
 -1.5204740396004677
  0.6510607283342043
 -0.3450874314674396
 -1.9216626847534772
 -1.3489229898498007
 -3.5395430250807256
  ⋮
  0.635255555614508
  0.4357127906073817
  0.8371626667977564
 -0.8013731781225494
  0.6454116036782747
 -0.73051029358993
 -1.782173794288489
 -1.4925160685630536
 -2.940027339223085

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -1.55365    1.30328
 -6.2698    -1.41281
 -8.38418   -2.40689
 -6.3944    -4.0131
 -8.45739   -2.89419
 -0.672889   0.746968
 -4.99337   -0.320875
 -6.26548   -0.624097
 -4.83147   -2.36285
 -5.94932   -0.489973
  ⋮         
 -7.08646   -1.53305
 -8.85017   -2.30398
 -6.84774   -3.2697
 -8.82192   -1.84292
  0.540563   1.22674
 -5.50967   -1.36327
 -6.85532   -2.67858
 -5.39438   -2.85223
 -7.14561   -3.84893

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.