Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
  0.609821  -0.28743    -0.0248171   …  -0.686899   -0.0593078    2.16274
 -0.534885  -0.162187   -0.487687       -0.387594   -1.16547     -0.271519
  0.866842  -0.548997   -0.545287       -1.31199    -1.30312      0.18743
  1.01416   -0.557493   -0.441225       -1.33229    -1.05444      1.45517
  1.01687   -0.479287   -0.0413823      -0.175382   -0.0151427    0.5522
 -0.891914  -0.270445   -0.813213    …  -0.0989621  -0.297573    -0.0693255
  1.44545   -0.915446   -0.90926        -0.334983   -0.332719     0.0478555
  1.6911    -0.929613   -0.735738       -0.340167   -0.269223     0.37154
  0.140854  -0.0663892  -0.00573214      0.646552    0.0558242   -2.03571
 -0.123545  -0.0374612  -0.112644        0.364827    1.09701      0.255571
  0.200219  -0.126805   -0.125948    …   1.23493     1.22658     -0.176421
  0.234246  -0.128767   -0.101912        1.25404     0.992501    -1.3697
  0.931871  -0.439224   -0.0379232      -0.0905677  -0.00781974   0.285158
 -0.81736   -0.247839   -0.745237       -0.0511043  -0.153668    -0.0357999
  1.32463   -0.838925   -0.833256       -0.172986   -0.171817     0.0247127
  1.54975   -0.851908   -0.674238    …  -0.175663   -0.139028     0.191864
 -1.63409    0.770202    0.0665003       1.01954     0.0880282   -3.21007
  1.43328    0.434599    1.30681         0.57529     1.72986      0.403006
 -2.3228     1.4711      1.46116         1.94733     1.93417     -0.278195
 -2.71756    1.49387     1.18231         1.97747     1.56506     -2.15985

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
 -0.12066422950120856
 -0.6260470271394115
 -0.8239056638036688
 -1.7068329330939978
  1.8151704915881068
 -0.34284256869773805
  0.8116409580852598
  0.339442158763333
 -0.4402632898782122
  1.8816851033300022
  ⋮
 -3.2989108615261706
 -1.9375674341629399
 -2.4516217556008364
 -0.6396619483422727
 -0.21766373601455882
 -1.9769484462506774
 -0.18731919910512076
 -1.9150059903332723
  0.5243060355491364

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 -1.28401    -4.37059
 -1.13127    -7.94111
 -1.23924    -6.82884
 -3.77202    -8.62027
 -2.55463    -6.63794
  0.252045   -4.67157
 -2.36309    -6.09555
 -3.33036    -5.02119
 -5.23001    -7.18029
 -2.60969    -5.79039
  ⋮          
 -4.52822    -6.16526
 -5.10236    -5.5676
 -7.45754    -7.63312
 -5.21479    -6.53654
 -0.0437659  -3.18539
  0.309234   -6.51289
 -1.55834    -5.19856
 -2.578      -6.61609
 -1.73514    -4.85611

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.