Multiplication

Kronecker.jl allows for efficient multiplication of large Kronecker systems by overloading the multiplication function *. We distinguish three cases:

Kronecker-Kronecker multiplications yield again a type of AbstractKroneckerProduct;

Kronecker-vector multiplications use the 'vec trick' and yield a vector;

sampled Kronecker-vector multiplications use the sampled-vec trick to yield a vector.

Kronecker-kronecker multiplications

Multiplying two conformable Kronecker products of the same order yield a new Kronecker product, based on the mixed-product property:

\[(A \otimes B)(C \otimes D) = (AC) \otimes (BD),\]

A, B, C, D = randn(5, 5), randn(4, 4), randn(5, 4), randn(4, 4);
(A ⊗ B) * (C ⊗ D)

20×16 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
 -3.25953   -0.711538   -0.699824   …   0.48552     0.477527    0.440727
 11.218      6.55944     1.85047       -4.47585    -1.26267     4.26668
  2.24379    1.73027     3.33443       -1.18066    -2.27526     0.996686
  2.97182    2.07233     2.40151       -1.41406    -1.63868     1.68712
  1.17535    0.256574    0.25235       -0.197602   -0.194349   -0.179372
 -4.04508   -2.36527    -0.667261   …   1.82163     0.513896   -1.7365
 -0.80909   -0.623919   -1.20236        0.480516    0.926008   -0.405642
 -1.07161   -0.747261   -0.865961       0.575508    0.666926   -0.686643
  2.8564     0.623537    0.613271      -0.371552   -0.365435   -0.337274
 -9.83054   -5.74819    -1.62161        3.42522     0.966281   -3.26514
 -1.96629   -1.51628    -2.92203    …   0.903516    1.74118    -0.76273
 -2.60428   -1.81603    -2.1045         1.08213     1.25402    -1.2911
 -1.24948   -0.272756   -0.268265       0.0408062   0.0401344   0.0370415
  4.30021    2.51445     0.709346      -0.37618    -0.106123    0.358599
  0.86012    0.66327     1.2782        -0.0992299  -0.191227    0.0837679
  1.1392     0.794391    0.920578   …  -0.118847   -0.137725    0.141797
  0.295516   0.0645096   0.0634476     -0.111107   -0.109278   -0.100856
 -1.01704   -0.594694   -0.167768       1.02426     0.288952   -0.976391
 -0.203427  -0.15687    -0.302307       0.270183    0.520672   -0.228083
 -0.269432  -0.187882   -0.217726       0.323595    0.374997   -0.386083

The Vec trick

Reshaping allows computing a product between a Kronecker product and vector as two matrix multiplications. This is the so-called vec trick which holds for any set of conformable matrices:

\[(A \otimes B) \text{vec}(X) = \text{vec}(B^\intercal X A).\]

Here, $\text{vec}(\cdot)$ is the vectorization operator, which stacks all columns of a matrix into a vector.

A, B = rand(10, 10), rand(5, 6);
x = randn(60);
(A ⊗ B) * x

50-element Array{Float64,1}:
  3.8796926522576722
  3.5896772674926285
  2.7164372095046296
  2.1036882099828675
  4.2991140475956975
  3.4116017793305162
  3.2075713760010367
  2.98618058052144
  0.5626921082962257
  2.31580983007547
  ⋮
  3.1853103920227626
  2.193998882678826
  1.4244867007279458
  4.697551206166595
  1.482343076113596
  1.3679946665076106
  0.7502350590236009
 -0.5118081328996372
  2.8336721892455983

Note that this trick is extended to also work with matrices:

A, B = rand(10, 10), rand(5, 6);
x = randn(60, 2);
(A ⊗ B) * x

50×2 Array{Float64,2}:
 2.76246   -0.65064
 3.15213    1.11703
 2.10706    1.83825
 0.477349  -1.54551
 2.9162     0.073147
 3.19433   -0.0416993
 2.10163    1.31944
 1.97628    2.04501
 1.7998     1.54612
 2.96594    1.23013
 ⋮         
 5.04333    1.48499
 4.62481    2.63622
 2.60775    0.0517745
 4.44043    0.666074
 3.43686   -0.318497
 3.07226    0.510789
 3.36442    1.54987
 1.53671    3.36993
 4.18497    0.791179

The vec trick works with higher-order Kronecker products. However, at the moment this has a substantial overhead and likely be relatively slow.

Docstrings

Missing docstring.

Missing docstring for mul!. Check Documenter's build log for details.

LinearAlgebra.lmul! — Function

lmul!(a::Number, K::AbstractKroneckerProduct)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the left matrix.

lmul!(a::Number, K::KroneckerPower)

Scale an KroneckerPower K inplace by a factor a by rescaling the matrix the base matrix with a factor a^(1/N).

It is recommended to rewrite your Kronecker product rather as copy(A) ⊗ (A ⊗ n - 1) (note the copy) for numerical stability. This will only modify the first matrix, leaving the chain of Kronecker products alone.

LinearAlgebra.rmul! — Function

rmul!(K::AbstractKroneckerProduct, a::Number)

Scale an AbstractKroneckerProduct K inplace by a factor a by rescaling the right matrix.

Sampled Kronecker-vector multiplications

See Indexed Kronecker products for the specifics.